Generate EJScreen Multisite (EJAM) Summary Report in HTML within shiny app

Creates a 2 page report on overall results or for one site, with demographic and environmental indicators, and EJ Indexes if needed.

Usage

build_community_report(
  output_df,
  analysis_title,
  totalpop,
  locationstr,
  include_ejindexes = FALSE,
  in_shiny = FALSE,
  filename = NULL
)

Arguments

output_df,

single row of results table from doaggregate - either results_overall or one row of bysite

analysis_title,

title to use in header of report

totalpop,

total population included in location(s) analyzed

locationstr,

description of the location(s) analyzed

include_ejindexes,

whether to build tables for EJ and EJ supp. indexes

in_shiny,

whether the function is being called in or outside of shiny - affects location of header

filename,

path to file to save HTML content to; if null, returns as string (used in Shiny app)

Details

For the same function for use in RStudio, see ejam2report() which relies on build_community_report()

This function gets called by app_server, ejam2report(), and community_report_template.Rmd