utility - what type of formula is used to aggregate this variable?

utility - what type of formula is used to aggregate this variable?

Usage

calctype(varnames)

Arguments

varnames

vector like names_d

Value

vector same length as varnames, like c("sum of counts", "wtdmean")

Examples

 calctype("pop")
#> [1] "sum of counts"
 calctype(names_d)
#> [1] "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean"
#> [8] "wtdmean" "wtdmean"

 x = names_these
 cbind(indicator = x, calctype = calctype(x), calcweight = calcweight(x))
#>       indicator          calctype  calcweight      
#>  [1,] "Demog.Index"      "wtdmean" "pop"           
#>  [2,] "Demog.Index.Supp" "wtdmean" "pop"           
#>  [3,] "pctlowinc"        "wtdmean" "povknownratio" 
#>  [4,] "pctlingiso"       "wtdmean" "hhlds"         
#>  [5,] "pctunemployed"    "wtdmean" "unemployedbase"
#>  [6,] "pctlths"          "wtdmean" "age25up"       
#>  [7,] "pctunder5"        "wtdmean" "pop"           
#>  [8,] "pctover64"        "wtdmean" "pop"           
#>  [9,] "pctmin"           "wtdmean" "pop"           
#> [10,] "pcthisp"          "wtdmean" "pop"           
#> [11,] "pctnhba"          "wtdmean" "pop"           
#> [12,] "pctnhaa"          "wtdmean" "pop"           
#> [13,] "pctnhaiana"       "wtdmean" "pop"           
#> [14,] "pctnhnhpia"       "wtdmean" "pop"           
#> [15,] "pctnhotheralone"  "wtdmean" "pop"           
#> [16,] "pctnhmulti"       "wtdmean" "pop"           
#> [17,] "pctnhwa"          "wtdmean" "pop"           
#> [18,] "pm"               "wtdmean" "pop"           
#> [19,] "o3"               "wtdmean" "pop"           
#> [20,] "no2"              "wtdmean" "pop"           
#> [21,] "dpm"              "wtdmean" "pop"           
#> [22,] "rsei"             "wtdmean" "pop"           
#> [23,] "traffic.score"    "wtdmean" "pop"           
#> [24,] "pctpre1960"       "wtdmean" "builtunits"    
#> [25,] "proximity.npl"    "wtdmean" "pop"           
#> [26,] "proximity.rmp"    "wtdmean" "pop"           
#> [27,] "proximity.tsdf"   "wtdmean" "pop"           
#> [28,] "ust"              "wtdmean" "pop"           
#> [29,] "proximity.npdes"  "wtdmean" "pop"           
#> [30,] "drinking"         "wtdmean" "pop"           

 x = names(testoutput_ejamit_10pts_1miles$results_overall)
 cbind(indicator = x, calctype = calctype(x), calcweight = calcweight(x))
#>        indicator                                        calctype             
#>   [1,] "EJAM Report"                                    NA                   
#>   [2,] "EJSCREEN Map"                                   NA                   
#>   [3,] "ejam_uniq_id"                                   NA                   
#>   [4,] "valid"                                          NA                   
#>   [5,] "invalid_msg"                                    NA                   
#>   [6,] "pop"                                            "sum of counts"      
#>   [7,] "lon"                                            "input"              
#>   [8,] "lat"                                            "input"              
#>   [9,] "ST"                                             "mode"               
#>  [10,] "statename"                                      "mode"               
#>  [11,] "in_how_many_states"                             NA                   
#>  [12,] "REGION"                                         "mode"               
#>  [13,] "ratio.to.avg.Demog.Index"                       "ratio to avg"       
#>  [14,] "ratio.to.avg.Demog.Index.Supp"                  "ratio to avg"       
#>  [15,] "ratio.to.avg.pctlowinc"                         "ratio to avg"       
#>  [16,] "ratio.to.avg.pctlingiso"                        "ratio to avg"       
#>  [17,] "ratio.to.avg.pctunemployed"                     "ratio to avg"       
#>  [18,] "ratio.to.avg.pctlths"                           "ratio to avg"       
#>  [19,] "ratio.to.avg.pctunder5"                         "ratio to avg"       
#>  [20,] "ratio.to.avg.pctover64"                         "ratio to avg"       
#>  [21,] "ratio.to.avg.pctmin"                            "ratio to avg"       
#>  [22,] "ratio.to.avg.pcthisp"                           "ratio to avg"       
#>  [23,] "ratio.to.avg.pctnhba"                           "ratio to avg"       
#>  [24,] "ratio.to.avg.pctnhaa"                           "ratio to avg"       
#>  [25,] "ratio.to.avg.pctnhaiana"                        "ratio to avg"       
#>  [26,] "ratio.to.avg.pctnhnhpia"                        "ratio to avg"       
#>  [27,] "ratio.to.avg.pctnhotheralone"                   "ratio to avg"       
#>  [28,] "ratio.to.avg.pctnhmulti"                        "ratio to avg"       
#>  [29,] "ratio.to.avg.pctnhwa"                           "ratio to avg"       
#>  [30,] "ratio.to.avg.pm"                                "ratio to avg"       
#>  [31,] "ratio.to.avg.o3"                                "ratio to avg"       
#>  [32,] "ratio.to.avg.no2"                               "ratio to avg"       
#>  [33,] "ratio.to.avg.dpm"                               "ratio to avg"       
#>  [34,] "ratio.to.avg.rsei"                              "ratio to avg"       
#>  [35,] "ratio.to.avg.traffic.score"                     "ratio to avg"       
#>  [36,] "ratio.to.avg.pctpre1960"                        "ratio to avg"       
#>  [37,] "ratio.to.avg.proximity.npl"                     "ratio to avg"       
#>  [38,] "ratio.to.avg.proximity.rmp"                     "ratio to avg"       
#>  [39,] "ratio.to.avg.proximity.tsdf"                    "ratio to avg"       
#>  [40,] "ratio.to.avg.ust"                               "ratio to avg"       
#>  [41,] "ratio.to.avg.proximity.npdes"                   "ratio to avg"       
#>  [42,] "ratio.to.avg.drinking"                          "ratio to avg"       
#>  [43,] "ratio.to.avg.pctdisability"                     "ratio to avg"       
#>  [44,] "ratio.to.avg.lowlifex"                          "ratio to avg"       
#>  [45,] "ratio.to.state.avg.Demog.Index"                 "ratio to avg"       
#>  [46,] "ratio.to.state.avg.Demog.Index.Supp"            "ratio to avg"       
#>  [47,] "ratio.to.state.avg.pctlowinc"                   "ratio to avg"       
#>  [48,] "ratio.to.state.avg.pctlingiso"                  "ratio to avg"       
#>  [49,] "ratio.to.state.avg.pctunemployed"               "ratio to avg"       
#>  [50,] "ratio.to.state.avg.pctlths"                     "ratio to avg"       
#>  [51,] "ratio.to.state.avg.pctunder5"                   "ratio to avg"       
#>  [52,] "ratio.to.state.avg.pctover64"                   "ratio to avg"       
#>  [53,] "ratio.to.state.avg.pctmin"                      "ratio to avg"       
#>  [54,] "ratio.to.state.avg.pcthisp"                     "ratio to avg"       
#>  [55,] "ratio.to.state.avg.pctnhba"                     "ratio to avg"       
#>  [56,] "ratio.to.state.avg.pctnhaa"                     "ratio to avg"       
#>  [57,] "ratio.to.state.avg.pctnhaiana"                  "ratio to avg"       
#>  [58,] "ratio.to.state.avg.pctnhnhpia"                  "ratio to avg"       
#>  [59,] "ratio.to.state.avg.pctnhotheralone"             "ratio to avg"       
#>  [60,] "ratio.to.state.avg.pctnhmulti"                  "ratio to avg"       
#>  [61,] "ratio.to.state.avg.pctnhwa"                     "ratio to avg"       
#>  [62,] "ratio.to.state.avg.pm"                          "ratio to avg"       
#>  [63,] "ratio.to.state.avg.o3"                          "ratio to avg"       
#>  [64,] "ratio.to.state.avg.no2"                         "ratio to avg"       
#>  [65,] "ratio.to.state.avg.dpm"                         "ratio to avg"       
#>  [66,] "ratio.to.state.avg.rsei"                        "ratio to avg"       
#>  [67,] "ratio.to.state.avg.traffic.score"               "ratio to avg"       
#>  [68,] "ratio.to.state.avg.pctpre1960"                  "ratio to avg"       
#>  [69,] "ratio.to.state.avg.proximity.npl"               "ratio to avg"       
#>  [70,] "ratio.to.state.avg.proximity.rmp"               "ratio to avg"       
#>  [71,] "ratio.to.state.avg.proximity.tsdf"              "ratio to avg"       
#>  [72,] "ratio.to.state.avg.ust"                         "ratio to avg"       
#>  [73,] "ratio.to.state.avg.proximity.npdes"             "ratio to avg"       
#>  [74,] "ratio.to.state.avg.drinking"                    "ratio to avg"       
#>  [75,] "ratio.to.state.avg.pctdisability"               "ratio to avg"       
#>  [76,] "ratio.to.state.avg.lowlifex"                    "ratio to avg"       
#>  [77,] "Demog.Index"                                    "wtdmean"            
#>  [78,] "Demog.Index.Supp"                               "wtdmean"            
#>  [79,] "pctlowinc"                                      "wtdmean"            
#>  [80,] "pctlingiso"                                     "wtdmean"            
#>  [81,] "pctunemployed"                                  "wtdmean"            
#>  [82,] "pctlths"                                        "wtdmean"            
#>  [83,] "pctunder5"                                      "wtdmean"            
#>  [84,] "pctover64"                                      "wtdmean"            
#>  [85,] "pctmin"                                         "wtdmean"            
#>  [86,] "pcthisp"                                        "wtdmean"            
#>  [87,] "pctnhba"                                        "wtdmean"            
#>  [88,] "pctnhaa"                                        "wtdmean"            
#>  [89,] "pctnhaiana"                                     "wtdmean"            
#>  [90,] "pctnhnhpia"                                     "wtdmean"            
#>  [91,] "pctnhotheralone"                                "wtdmean"            
#>  [92,] "pctnhmulti"                                     "wtdmean"            
#>  [93,] "pctnhwa"                                        "wtdmean"            
#>  [94,] "Demog.Index.State"                              "wtdmean"            
#>  [95,] "Demog.Index.Supp.State"                         "wtdmean"            
#>  [96,] "pctpoor"                                        "wtdmean"            
#>  [97,] "occupiedunits"                                  "sum of counts"      
#>  [98,] "pctmale"                                        "wtdmean"            
#>  [99,] "pctfemale"                                      "wtdmean"            
#> [100,] "lifexyears"                                     "wtdmean"            
#> [101,] "percapincome"                                   "wtdmean"            
#> [102,] "pctownedunits"                                  "wtdmean"            
#> [103,] "pctunder18"                                     "wtdmean"            
#> [104,] "pctover17"                                      "wtdmean"            
#> [105,] "pctlan_nonenglish"                              "wtdmean"            
#> [106,] "pctlan_ie"                                      "wtdmean"            
#> [107,] "pctlan_api"                                     "wtdmean"            
#> [108,] "pctlan_other"                                   "wtdmean"            
#> [109,] "pctlan_english"                                 "wtdmean"            
#> [110,] "pctlan_spanish"                                 "wtdmean"            
#> [111,] "pctlan_french"                                  "wtdmean"            
#> [112,] "pctlan_rus_pol_slav"                            "wtdmean"            
#> [113,] "pctlan_other_ie"                                "wtdmean"            
#> [114,] "pctlan_vietnamese"                              "wtdmean"            
#> [115,] "pctlan_other_asian"                             "wtdmean"            
#> [116,] "pctlan_arabic"                                  "wtdmean"            
#> [117,] "pctspanish_li"                                  "wtdmean"            
#> [118,] "pctie_li"                                       "wtdmean"            
#> [119,] "pctapi_li"                                      "wtdmean"            
#> [120,] "pctother_li"                                    "wtdmean"            
#> [121,] "pctdisability"                                  "wtdmean"            
#> [122,] "lowlifex"                                       "wtdmean"            
#> [123,] "rateheartdisease"                               "wtdmean"            
#> [124,] "rateasthma"                                     "wtdmean"            
#> [125,] "ratecancer"                                     "wtdmean"            
#> [126,] "yesno_houseburden"                              "flag"               
#> [127,] "yesno_fooddesert"                               "flag"               
#> [128,] "pctnobroadband"                                 "wtdmean"            
#> [129,] "pctnohealthinsurance"                           "wtdmean"            
#> [130,] "yesno_transdis"                                 "flag"               
#> [131,] "pctfire30"                                      "wtdmean"            
#> [132,] "pctflood30"                                     "wtdmean"            
#> [133,] "pm"                                             "wtdmean"            
#> [134,] "o3"                                             "wtdmean"            
#> [135,] "no2"                                            "wtdmean"            
#> [136,] "dpm"                                            "wtdmean"            
#> [137,] "rsei"                                           "wtdmean"            
#> [138,] "traffic.score"                                  "wtdmean"            
#> [139,] "pctpre1960"                                     "wtdmean"            
#> [140,] "proximity.npl"                                  "wtdmean"            
#> [141,] "proximity.rmp"                                  "wtdmean"            
#> [142,] "proximity.tsdf"                                 "wtdmean"            
#> [143,] "ust"                                            "wtdmean"            
#> [144,] "proximity.npdes"                                "wtdmean"            
#> [145,] "drinking"                                       "wtdmean"            
#> [146,] "num_school"                                     "sum of counts"      
#> [147,] "num_church"                                     "sum of counts"      
#> [148,] "num_hospital"                                   "sum of counts"      
#> [149,] "yesno_tribal"                                   "flag"               
#> [150,] "yesno_impwaters"                                "flag"               
#> [151,] "yesno_airnonatt"                                "flag"               
#> [152,] "yesno_cejstdis"                                 "flag"               
#> [153,] "yesno_iradis"                                   "flag"               
#> [154,] "count.NPL"                                      "sum of counts"      
#> [155,] "count.TSDF"                                     "sum of counts"      
#> [156,] "num_airpoll"                                    "sum of counts"      
#> [157,] "num_tri"                                        "sum of counts"      
#> [158,] "num_waterdis"                                   "sum of counts"      
#> [159,] "num_brownfield"                                 "sum of counts"      
#> [160,] "sitecount_max"                                  "max"                
#> [161,] "sitecount_unique"                               "count of unique ids"
#> [162,] "sitecount_avg"                                  "wtdmean"            
#> [163,] "distance_min"                                   "min"                
#> [164,] "distance_min_avgperson"                         "min"                
#> [165,] "pctile.Demog.Index"                             "lookedup"           
#> [166,] "pctile.Demog.Index.Supp"                        "lookedup"           
#> [167,] "pctile.pctlowinc"                               "lookedup"           
#> [168,] "pctile.pctlingiso"                              "lookedup"           
#> [169,] "pctile.pctunemployed"                           "lookedup"           
#> [170,] "pctile.pctlths"                                 "lookedup"           
#> [171,] "pctile.pctunder5"                               "lookedup"           
#> [172,] "pctile.pctover64"                               "lookedup"           
#> [173,] "pctile.pctmin"                                  "lookedup"           
#> [174,] "pctile.pcthisp"                                 "lookedup"           
#> [175,] "pctile.pctnhba"                                 "lookedup"           
#> [176,] "pctile.pctnhaa"                                 "lookedup"           
#> [177,] "pctile.pctnhaiana"                              "lookedup"           
#> [178,] "pctile.pctnhnhpia"                              "lookedup"           
#> [179,] "pctile.pctnhotheralone"                         "lookedup"           
#> [180,] "pctile.pctnhmulti"                              "lookedup"           
#> [181,] "pctile.pctnhwa"                                 "lookedup"           
#> [182,] "pctile.pctdisability"                           "lookedup"           
#> [183,] "pctile.lowlifex"                                "lookedup"           
#> [184,] "pctile.rateheartdisease"                        "lookedup"           
#> [185,] "pctile.rateasthma"                              "lookedup"           
#> [186,] "pctile.ratecancer"                              "lookedup"           
#> [187,] "pctile.pctfire30"                               "lookedup"           
#> [188,] "pctile.pctflood30"                              "lookedup"           
#> [189,] "pctile.pm"                                      "lookedup"           
#> [190,] "pctile.o3"                                      "lookedup"           
#> [191,] "pctile.no2"                                     "lookedup"           
#> [192,] "pctile.dpm"                                     "lookedup"           
#> [193,] "pctile.rsei"                                    "lookedup"           
#> [194,] "pctile.traffic.score"                           "lookedup"           
#> [195,] "pctile.pctpre1960"                              "lookedup"           
#> [196,] "pctile.proximity.npl"                           "lookedup"           
#> [197,] "pctile.proximity.rmp"                           "lookedup"           
#> [198,] "pctile.proximity.tsdf"                          "lookedup"           
#> [199,] "pctile.ust"                                     "lookedup"           
#> [200,] "pctile.proximity.npdes"                         "lookedup"           
#> [201,] "pctile.drinking"                                "lookedup"           
#> [202,] "state.pctile.Demog.Index"                       "lookedup"           
#> [203,] "state.pctile.Demog.Index.Supp"                  "lookedup"           
#> [204,] "state.pctile.pctlowinc"                         "lookedup"           
#> [205,] "state.pctile.pctlingiso"                        "lookedup"           
#> [206,] "state.pctile.pctunemployed"                     "lookedup"           
#> [207,] "state.pctile.pctlths"                           "lookedup"           
#> [208,] "state.pctile.pctunder5"                         "lookedup"           
#> [209,] "state.pctile.pctover64"                         "lookedup"           
#> [210,] "state.pctile.pctmin"                            "lookedup"           
#> [211,] "state.pctile.pcthisp"                           "lookedup"           
#> [212,] "state.pctile.pctnhba"                           "lookedup"           
#> [213,] "state.pctile.pctnhaa"                           "lookedup"           
#> [214,] "state.pctile.pctnhaiana"                        "lookedup"           
#> [215,] "state.pctile.pctnhnhpia"                        "lookedup"           
#> [216,] "state.pctile.pctnhotheralone"                   "lookedup"           
#> [217,] "state.pctile.pctnhmulti"                        "lookedup"           
#> [218,] "state.pctile.pctnhwa"                           "lookedup"           
#> [219,] "state.pctile.pctdisability"                     "lookedup"           
#> [220,] "state.pctile.lowlifex"                          "lookedup"           
#> [221,] "state.pctile.rateheartdisease"                  "lookedup"           
#> [222,] "state.pctile.rateasthma"                        "lookedup"           
#> [223,] "state.pctile.ratecancer"                        "lookedup"           
#> [224,] "state.pctile.pctfire30"                         "lookedup"           
#> [225,] "state.pctile.pctflood30"                        "lookedup"           
#> [226,] "state.pctile.pm"                                "lookedup"           
#> [227,] "state.pctile.o3"                                "lookedup"           
#> [228,] "state.pctile.no2"                               "lookedup"           
#> [229,] "state.pctile.dpm"                               "lookedup"           
#> [230,] "state.pctile.rsei"                              "lookedup"           
#> [231,] "state.pctile.traffic.score"                     "lookedup"           
#> [232,] "state.pctile.pctpre1960"                        "lookedup"           
#> [233,] "state.pctile.proximity.npl"                     "lookedup"           
#> [234,] "state.pctile.proximity.rmp"                     "lookedup"           
#> [235,] "state.pctile.proximity.tsdf"                    "lookedup"           
#> [236,] "state.pctile.ust"                               "lookedup"           
#> [237,] "state.pctile.proximity.npdes"                   "lookedup"           
#> [238,] "state.pctile.drinking"                          "lookedup"           
#> [239,] "avg.Demog.Index"                                "constant"           
#> [240,] "avg.Demog.Index.Supp"                           "constant"           
#> [241,] "avg.pctlowinc"                                  "constant"           
#> [242,] "avg.pctlingiso"                                 "constant"           
#> [243,] "avg.pctunemployed"                              "constant"           
#> [244,] "avg.pctlths"                                    "constant"           
#> [245,] "avg.pctunder5"                                  "constant"           
#> [246,] "avg.pctover64"                                  "constant"           
#> [247,] "avg.pctmin"                                     "constant"           
#> [248,] "avg.pcthisp"                                    "constant"           
#> [249,] "avg.pctnhba"                                    "constant"           
#> [250,] "avg.pctnhaa"                                    "constant"           
#> [251,] "avg.pctnhaiana"                                 "constant"           
#> [252,] "avg.pctnhnhpia"                                 "constant"           
#> [253,] "avg.pctnhotheralone"                            "constant"           
#> [254,] "avg.pctnhmulti"                                 "constant"           
#> [255,] "avg.pctnhwa"                                    "constant"           
#> [256,] "avg.pctdisability"                              "constant"           
#> [257,] "avg.lowlifex"                                   "constant"           
#> [258,] "avg.pm"                                         "constant"           
#> [259,] "avg.o3"                                         "constant"           
#> [260,] "avg.no2"                                        "constant"           
#> [261,] "avg.dpm"                                        "constant"           
#> [262,] "avg.rsei"                                       "constant"           
#> [263,] "avg.traffic.score"                              "constant"           
#> [264,] "avg.pctpre1960"                                 "constant"           
#> [265,] "avg.proximity.npl"                              "constant"           
#> [266,] "avg.proximity.rmp"                              "constant"           
#> [267,] "avg.proximity.tsdf"                             "constant"           
#> [268,] "avg.ust"                                        "constant"           
#> [269,] "avg.proximity.npdes"                            "constant"           
#> [270,] "avg.drinking"                                   "constant"           
#> [271,] "state.avg.Demog.Index"                          "constant"           
#> [272,] "state.avg.Demog.Index.Supp"                     "constant"           
#> [273,] "state.avg.pctlowinc"                            "constant"           
#> [274,] "state.avg.pctlingiso"                           "constant"           
#> [275,] "state.avg.pctunemployed"                        "constant"           
#> [276,] "state.avg.pctlths"                              "constant"           
#> [277,] "state.avg.pctunder5"                            "constant"           
#> [278,] "state.avg.pctover64"                            "constant"           
#> [279,] "state.avg.pctmin"                               "constant"           
#> [280,] "state.avg.pcthisp"                              "constant"           
#> [281,] "state.avg.pctnhba"                              "constant"           
#> [282,] "state.avg.pctnhaa"                              "constant"           
#> [283,] "state.avg.pctnhaiana"                           "constant"           
#> [284,] "state.avg.pctnhnhpia"                           "constant"           
#> [285,] "state.avg.pctnhotheralone"                      "constant"           
#> [286,] "state.avg.pctnhmulti"                           "constant"           
#> [287,] "state.avg.pctnhwa"                              "constant"           
#> [288,] "state.avg.pctdisability"                        "constant"           
#> [289,] "state.avg.lowlifex"                             "constant"           
#> [290,] "state.avg.pm"                                   "constant"           
#> [291,] "state.avg.o3"                                   "constant"           
#> [292,] "state.avg.no2"                                  "constant"           
#> [293,] "state.avg.dpm"                                  "constant"           
#> [294,] "state.avg.rsei"                                 "constant"           
#> [295,] "state.avg.traffic.score"                        "constant"           
#> [296,] "state.avg.pctpre1960"                           "constant"           
#> [297,] "state.avg.proximity.npl"                        "constant"           
#> [298,] "state.avg.proximity.rmp"                        "constant"           
#> [299,] "state.avg.proximity.tsdf"                       "constant"           
#> [300,] "state.avg.ust"                                  "constant"           
#> [301,] "state.avg.proximity.npdes"                      "constant"           
#> [302,] "state.avg.drinking"                             "constant"           
#> [303,] "lowinc"                                         "sum of counts"      
#> [304,] "lingiso"                                        "sum of counts"      
#> [305,] "unemployed"                                     "sum of counts"      
#> [306,] "lths"                                           "sum of counts"      
#> [307,] "under5"                                         "sum of counts"      
#> [308,] "over64"                                         "sum of counts"      
#> [309,] "mins"                                           "sum of counts"      
#> [310,] "hisp"                                           "sum of counts"      
#> [311,] "nhba"                                           "sum of counts"      
#> [312,] "nhaa"                                           "sum of counts"      
#> [313,] "nhaiana"                                        "sum of counts"      
#> [314,] "nhnhpia"                                        "sum of counts"      
#> [315,] "nhotheralone"                                   "sum of counts"      
#> [316,] "nhmulti"                                        "sum of counts"      
#> [317,] "nhwa"                                           "sum of counts"      
#> [318,] "ba"                                             "sum of counts"      
#> [319,] "aa"                                             "sum of counts"      
#> [320,] "aiana"                                          "sum of counts"      
#> [321,] "nhpia"                                          "sum of counts"      
#> [322,] "otheralone"                                     "sum of counts"      
#> [323,] "multi"                                          "sum of counts"      
#> [324,] "wa"                                             "sum of counts"      
#> [325,] "age25up"                                        "sum of counts"      
#> [326,] "hhlds"                                          "sum of counts"      
#> [327,] "unemployedbase"                                 "sum of counts"      
#> [328,] "pre1960"                                        "sum of counts"      
#> [329,] "builtunits"                                     "sum of counts"      
#> [330,] "povknownratio"                                  "sum of counts"      
#> [331,] "ownedunits"                                     "sum of counts"      
#> [332,] "poor"                                           "sum of counts"      
#> [333,] "female"                                         "sum of counts"      
#> [334,] "male"                                           "sum of counts"      
#> [335,] "over17"                                         "sum of counts"      
#> [336,] "under18"                                        "sum of counts"      
#> [337,] "lan_nonenglish"                                 "sum of counts"      
#> [338,] "lan_ie"                                         "sum of counts"      
#> [339,] "lan_api"                                        "sum of counts"      
#> [340,] "lan_other"                                      "sum of counts"      
#> [341,] "lan_universe"                                   "sum of counts"      
#> [342,] "lan_eng_na"                                     "sum of counts"      
#> [343,] "lan_spanish"                                    "sum of counts"      
#> [344,] "spanish_li"                                     "sum of counts"      
#> [345,] "ie_li"                                          "sum of counts"      
#> [346,] "api_li"                                         "sum of counts"      
#> [347,] "other_li"                                       "sum of counts"      
#> [348,] "pctile.EJ.DISPARITY.pm.eo"                      "lookedup"           
#> [349,] "pctile.EJ.DISPARITY.o3.eo"                      "lookedup"           
#> [350,] "pctile.EJ.DISPARITY.no2.eo"                     "lookedup"           
#> [351,] "pctile.EJ.DISPARITY.dpm.eo"                     "lookedup"           
#> [352,] "pctile.EJ.DISPARITY.rsei.eo"                    "lookedup"           
#> [353,] "pctile.EJ.DISPARITY.traffic.score.eo"           "lookedup"           
#> [354,] "pctile.EJ.DISPARITY.pctpre1960.eo"              "lookedup"           
#> [355,] "pctile.EJ.DISPARITY.proximity.npl.eo"           "lookedup"           
#> [356,] "pctile.EJ.DISPARITY.proximity.rmp.eo"           "lookedup"           
#> [357,] "pctile.EJ.DISPARITY.proximity.tsdf.eo"          "lookedup"           
#> [358,] "pctile.EJ.DISPARITY.ust.eo"                     "lookedup"           
#> [359,] "pctile.EJ.DISPARITY.proximity.npdes.eo"         "lookedup"           
#> [360,] "pctile.EJ.DISPARITY.drinking.eo"                "lookedup"           
#> [361,] "pctile.EJ.DISPARITY.pm.supp"                    "lookedup"           
#> [362,] "pctile.EJ.DISPARITY.o3.supp"                    "lookedup"           
#> [363,] "pctile.EJ.DISPARITY.no2.supp"                   "lookedup"           
#> [364,] "pctile.EJ.DISPARITY.dpm.supp"                   "lookedup"           
#> [365,] "pctile.EJ.DISPARITY.rsei.supp"                  "lookedup"           
#> [366,] "pctile.EJ.DISPARITY.traffic.score.supp"         "lookedup"           
#> [367,] "pctile.EJ.DISPARITY.pctpre1960.supp"            "lookedup"           
#> [368,] "pctile.EJ.DISPARITY.proximity.npl.supp"         "lookedup"           
#> [369,] "pctile.EJ.DISPARITY.proximity.rmp.supp"         "lookedup"           
#> [370,] "pctile.EJ.DISPARITY.proximity.tsdf.supp"        "lookedup"           
#> [371,] "pctile.EJ.DISPARITY.ust.supp"                   "lookedup"           
#> [372,] "pctile.EJ.DISPARITY.proximity.npdes.supp"       "lookedup"           
#> [373,] "pctile.EJ.DISPARITY.drinking.supp"              "lookedup"           
#> [374,] "state.pctile.EJ.DISPARITY.pm.eo"                "lookedup"           
#> [375,] "state.pctile.EJ.DISPARITY.o3.eo"                "lookedup"           
#> [376,] "state.pctile.EJ.DISPARITY.no2.eo"               "lookedup"           
#> [377,] "state.pctile.EJ.DISPARITY.dpm.eo"               "lookedup"           
#> [378,] "state.pctile.EJ.DISPARITY.rsei.eo"              "lookedup"           
#> [379,] "state.pctile.EJ.DISPARITY.traffic.score.eo"     "lookedup"           
#> [380,] "state.pctile.EJ.DISPARITY.pctpre1960.eo"        "lookedup"           
#> [381,] "state.pctile.EJ.DISPARITY.proximity.npl.eo"     "lookedup"           
#> [382,] "state.pctile.EJ.DISPARITY.proximity.rmp.eo"     "lookedup"           
#> [383,] "state.pctile.EJ.DISPARITY.proximity.tsdf.eo"    "lookedup"           
#> [384,] "state.pctile.EJ.DISPARITY.ust.eo"               "lookedup"           
#> [385,] "state.pctile.EJ.DISPARITY.proximity.npdes.eo"   "lookedup"           
#> [386,] "state.pctile.EJ.DISPARITY.drinking.eo"          "lookedup"           
#> [387,] "state.pctile.EJ.DISPARITY.pm.supp"              "lookedup"           
#> [388,] "state.pctile.EJ.DISPARITY.o3.supp"              "lookedup"           
#> [389,] "state.pctile.EJ.DISPARITY.no2.supp"             "lookedup"           
#> [390,] "state.pctile.EJ.DISPARITY.dpm.supp"             "lookedup"           
#> [391,] "state.pctile.EJ.DISPARITY.rsei.supp"            "lookedup"           
#> [392,] "state.pctile.EJ.DISPARITY.traffic.score.supp"   "lookedup"           
#> [393,] "state.pctile.EJ.DISPARITY.pctpre1960.supp"      "lookedup"           
#> [394,] "state.pctile.EJ.DISPARITY.proximity.npl.supp"   "lookedup"           
#> [395,] "state.pctile.EJ.DISPARITY.proximity.rmp.supp"   "lookedup"           
#> [396,] "state.pctile.EJ.DISPARITY.proximity.tsdf.supp"  "lookedup"           
#> [397,] "state.pctile.EJ.DISPARITY.ust.supp"             "lookedup"           
#> [398,] "state.pctile.EJ.DISPARITY.proximity.npdes.supp" "lookedup"           
#> [399,] "state.pctile.EJ.DISPARITY.drinking.supp"        "lookedup"           
#> [400,] "bgcount_near_site"                              "count of unique ids"
#> [401,] "blockcount_near_site"                           "count of unique ids"
#> [402,] "radius.miles"                                   "input"              
#> [403,] "disability"                                     "sum of counts"      
#> [404,] "disab_universe"                                 "sum of counts"      
#> [405,] "areawater"                                      "sum of counts"      
#> [406,] "arealand"                                       "sum of counts"      
#> [407,] "area"                                           "sum of counts"      
#> [408,] "pctba"                                          "wtdmean"            
#> [409,] "pctaa"                                          "wtdmean"            
#> [410,] "pctaiana"                                       "wtdmean"            
#> [411,] "pctnhpia"                                       "wtdmean"            
#> [412,] "pctotheralone"                                  "wtdmean"            
#> [413,] "pctmulti"                                       "wtdmean"            
#> [414,] "pctwa"                                          "wtdmean"            
#> [415,] "EJ.DISPARITY.pm.eo"                             "wtdmean"            
#> [416,] "EJ.DISPARITY.o3.eo"                             "wtdmean"            
#> [417,] "EJ.DISPARITY.no2.eo"                            "wtdmean"            
#> [418,] "EJ.DISPARITY.dpm.eo"                            "wtdmean"            
#> [419,] "EJ.DISPARITY.rsei.eo"                           "wtdmean"            
#> [420,] "EJ.DISPARITY.traffic.score.eo"                  "wtdmean"            
#> [421,] "EJ.DISPARITY.pctpre1960.eo"                     "wtdmean"            
#> [422,] "EJ.DISPARITY.proximity.npl.eo"                  "wtdmean"            
#> [423,] "EJ.DISPARITY.proximity.rmp.eo"                  "wtdmean"            
#> [424,] "EJ.DISPARITY.proximity.tsdf.eo"                 "wtdmean"            
#> [425,] "EJ.DISPARITY.ust.eo"                            "wtdmean"            
#> [426,] "EJ.DISPARITY.proximity.npdes.eo"                "wtdmean"            
#> [427,] "EJ.DISPARITY.drinking.eo"                       "wtdmean"            
#> [428,] "state.EJ.DISPARITY.pm.eo"                       "wtdmean"            
#> [429,] "state.EJ.DISPARITY.o3.eo"                       "wtdmean"            
#> [430,] "state.EJ.DISPARITY.no2.eo"                      "wtdmean"            
#> [431,] "state.EJ.DISPARITY.dpm.eo"                      "wtdmean"            
#> [432,] "state.EJ.DISPARITY.rsei.eo"                     "wtdmean"            
#> [433,] "state.EJ.DISPARITY.traffic.score.eo"            "wtdmean"            
#> [434,] "state.EJ.DISPARITY.pctpre1960.eo"               "wtdmean"            
#> [435,] "state.EJ.DISPARITY.proximity.npl.eo"            "wtdmean"            
#> [436,] "state.EJ.DISPARITY.proximity.rmp.eo"            "wtdmean"            
#> [437,] "state.EJ.DISPARITY.proximity.tsdf.eo"           "wtdmean"            
#> [438,] "state.EJ.DISPARITY.ust.eo"                      "wtdmean"            
#> [439,] "state.EJ.DISPARITY.proximity.npdes.eo"          "wtdmean"            
#> [440,] "state.EJ.DISPARITY.drinking.eo"                 "wtdmean"            
#> [441,] "EJ.DISPARITY.pm.supp"                           "wtdmean"            
#> [442,] "EJ.DISPARITY.o3.supp"                           "wtdmean"            
#> [443,] "EJ.DISPARITY.no2.supp"                          "wtdmean"            
#> [444,] "EJ.DISPARITY.dpm.supp"                          "wtdmean"            
#> [445,] "EJ.DISPARITY.rsei.supp"                         "wtdmean"            
#> [446,] "EJ.DISPARITY.traffic.score.supp"                "wtdmean"            
#> [447,] "EJ.DISPARITY.pctpre1960.supp"                   "wtdmean"            
#> [448,] "EJ.DISPARITY.proximity.npl.supp"                "wtdmean"            
#> [449,] "EJ.DISPARITY.proximity.rmp.supp"                "wtdmean"            
#> [450,] "EJ.DISPARITY.proximity.tsdf.supp"               "wtdmean"            
#> [451,] "EJ.DISPARITY.ust.supp"                          "wtdmean"            
#> [452,] "EJ.DISPARITY.proximity.npdes.supp"              "wtdmean"            
#> [453,] "EJ.DISPARITY.drinking.supp"                     "wtdmean"            
#> [454,] "state.EJ.DISPARITY.pm.supp"                     "wtdmean"            
#> [455,] "state.EJ.DISPARITY.o3.supp"                     "wtdmean"            
#> [456,] "state.EJ.DISPARITY.no2.supp"                    "wtdmean"            
#> [457,] "state.EJ.DISPARITY.dpm.supp"                    "wtdmean"            
#> [458,] "state.EJ.DISPARITY.rsei.supp"                   "wtdmean"            
#> [459,] "state.EJ.DISPARITY.traffic.score.supp"          "wtdmean"            
#> [460,] "state.EJ.DISPARITY.pctpre1960.supp"             "wtdmean"            
#> [461,] "state.EJ.DISPARITY.proximity.npl.supp"          "wtdmean"            
#> [462,] "state.EJ.DISPARITY.proximity.rmp.supp"          "wtdmean"            
#> [463,] "state.EJ.DISPARITY.proximity.tsdf.supp"         "wtdmean"            
#> [464,] "state.EJ.DISPARITY.ust.supp"                    "wtdmean"            
#> [465,] "state.EJ.DISPARITY.proximity.npdes.supp"        "wtdmean"            
#> [466,] "state.EJ.DISPARITY.drinking.supp"               "wtdmean"            
#> [467,] "area_sqmi"                                      "sum of counts"      
#>        calcweight      
#>   [1,] NA              
#>   [2,] NA              
#>   [3,] NA              
#>   [4,] NA              
#>   [5,] NA              
#>   [6,] ""              
#>   [7,] ""              
#>   [8,] ""              
#>   [9,] ""              
#>  [10,] ""              
#>  [11,] NA              
#>  [12,] ""              
#>  [13,] ""              
#>  [14,] ""              
#>  [15,] ""              
#>  [16,] ""              
#>  [17,] ""              
#>  [18,] ""              
#>  [19,] ""              
#>  [20,] ""              
#>  [21,] ""              
#>  [22,] ""              
#>  [23,] ""              
#>  [24,] ""              
#>  [25,] ""              
#>  [26,] ""              
#>  [27,] ""              
#>  [28,] ""              
#>  [29,] ""              
#>  [30,] ""              
#>  [31,] ""              
#>  [32,] ""              
#>  [33,] ""              
#>  [34,] ""              
#>  [35,] ""              
#>  [36,] ""              
#>  [37,] ""              
#>  [38,] ""              
#>  [39,] ""              
#>  [40,] ""              
#>  [41,] ""              
#>  [42,] ""              
#>  [43,] ""              
#>  [44,] ""              
#>  [45,] ""              
#>  [46,] ""              
#>  [47,] ""              
#>  [48,] ""              
#>  [49,] ""              
#>  [50,] ""              
#>  [51,] ""              
#>  [52,] ""              
#>  [53,] ""              
#>  [54,] ""              
#>  [55,] ""              
#>  [56,] ""              
#>  [57,] ""              
#>  [58,] ""              
#>  [59,] ""              
#>  [60,] ""              
#>  [61,] ""              
#>  [62,] ""              
#>  [63,] ""              
#>  [64,] ""              
#>  [65,] ""              
#>  [66,] ""              
#>  [67,] ""              
#>  [68,] ""              
#>  [69,] ""              
#>  [70,] ""              
#>  [71,] ""              
#>  [72,] ""              
#>  [73,] ""              
#>  [74,] ""              
#>  [75,] ""              
#>  [76,] ""              
#>  [77,] "pop"           
#>  [78,] "pop"           
#>  [79,] "povknownratio" 
#>  [80,] "hhlds"         
#>  [81,] "unemployedbase"
#>  [82,] "age25up"       
#>  [83,] "pop"           
#>  [84,] "pop"           
#>  [85,] "pop"           
#>  [86,] "pop"           
#>  [87,] "pop"           
#>  [88,] "pop"           
#>  [89,] "pop"           
#>  [90,] "pop"           
#>  [91,] "pop"           
#>  [92,] "pop"           
#>  [93,] "pop"           
#>  [94,] "pop"           
#>  [95,] "pop"           
#>  [96,] "hhlds"         
#>  [97,] ""              
#>  [98,] "pop"           
#>  [99,] "pop"           
#> [100,] "pop"           
#> [101,] "pop"           
#> [102,] "occupiedunits" 
#> [103,] "pop"           
#> [104,] "pop"           
#> [105,] "lan_universe"  
#> [106,] "lan_universe"  
#> [107,] "lan_universe"  
#> [108,] "lan_universe"  
#> [109,] "lan_universe"  
#> [110,] "lan_universe"  
#> [111,] "lan_universe"  
#> [112,] "lan_universe"  
#> [113,] "lan_universe"  
#> [114,] "lan_universe"  
#> [115,] "lan_universe"  
#> [116,] "lan_universe"  
#> [117,] "lingiso"       
#> [118,] "lingiso"       
#> [119,] "lingiso"       
#> [120,] "lingiso"       
#> [121,] "disab_universe"
#> [122,] "pop"           
#> [123,] "pop"           
#> [124,] "pop"           
#> [125,] "pop"           
#> [126,] ""              
#> [127,] ""              
#> [128,] "hhlds"         
#> [129,] "hhlds"         
#> [130,] ""              
#> [131,] "pop"           
#> [132,] "pop"           
#> [133,] "pop"           
#> [134,] "pop"           
#> [135,] "pop"           
#> [136,] "pop"           
#> [137,] "pop"           
#> [138,] "pop"           
#> [139,] "builtunits"    
#> [140,] "pop"           
#> [141,] "pop"           
#> [142,] "pop"           
#> [143,] "pop"           
#> [144,] "pop"           
#> [145,] "pop"           
#> [146,] ""              
#> [147,] ""              
#> [148,] ""              
#> [149,] ""              
#> [150,] ""              
#> [151,] ""              
#> [152,] ""              
#> [153,] ""              
#> [154,] ""              
#> [155,] ""              
#> [156,] ""              
#> [157,] ""              
#> [158,] ""              
#> [159,] ""              
#> [160,] ""              
#> [161,] ""              
#> [162,] "pop"           
#> [163,] ""              
#> [164,] ""              
#> [165,] ""              
#> [166,] ""              
#> [167,] ""              
#> [168,] ""              
#> [169,] ""              
#> [170,] ""              
#> [171,] ""              
#> [172,] ""              
#> [173,] ""              
#> [174,] ""              
#> [175,] ""              
#> [176,] ""              
#> [177,] ""              
#> [178,] ""              
#> [179,] ""              
#> [180,] ""              
#> [181,] ""              
#> [182,] ""              
#> [183,] ""              
#> [184,] ""              
#> [185,] ""              
#> [186,] ""              
#> [187,] ""              
#> [188,] ""              
#> [189,] ""              
#> [190,] ""              
#> [191,] ""              
#> [192,] ""              
#> [193,] ""              
#> [194,] ""              
#> [195,] ""              
#> [196,] ""              
#> [197,] ""              
#> [198,] ""              
#> [199,] ""              
#> [200,] ""              
#> [201,] ""              
#> [202,] ""              
#> [203,] ""              
#> [204,] ""              
#> [205,] ""              
#> [206,] ""              
#> [207,] ""              
#> [208,] ""              
#> [209,] ""              
#> [210,] ""              
#> [211,] ""              
#> [212,] ""              
#> [213,] ""              
#> [214,] ""              
#> [215,] ""              
#> [216,] ""              
#> [217,] ""              
#> [218,] ""              
#> [219,] ""              
#> [220,] ""              
#> [221,] ""              
#> [222,] ""              
#> [223,] ""              
#> [224,] ""              
#> [225,] ""              
#> [226,] ""              
#> [227,] ""              
#> [228,] ""              
#> [229,] ""              
#> [230,] ""              
#> [231,] ""              
#> [232,] ""              
#> [233,] ""              
#> [234,] ""              
#> [235,] ""              
#> [236,] ""              
#> [237,] ""              
#> [238,] ""              
#> [239,] ""              
#> [240,] ""              
#> [241,] ""              
#> [242,] ""              
#> [243,] ""              
#> [244,] ""              
#> [245,] ""              
#> [246,] ""              
#> [247,] ""              
#> [248,] ""              
#> [249,] ""              
#> [250,] ""              
#> [251,] ""              
#> [252,] ""              
#> [253,] ""              
#> [254,] ""              
#> [255,] ""              
#> [256,] ""              
#> [257,] ""              
#> [258,] ""              
#> [259,] ""              
#> [260,] ""              
#> [261,] ""              
#> [262,] ""              
#> [263,] ""              
#> [264,] ""              
#> [265,] ""              
#> [266,] ""              
#> [267,] ""              
#> [268,] ""              
#> [269,] ""              
#> [270,] ""              
#> [271,] ""              
#> [272,] ""              
#> [273,] ""              
#> [274,] ""              
#> [275,] ""              
#> [276,] ""              
#> [277,] ""              
#> [278,] ""              
#> [279,] ""              
#> [280,] ""              
#> [281,] ""              
#> [282,] ""              
#> [283,] ""              
#> [284,] ""              
#> [285,] ""              
#> [286,] ""              
#> [287,] ""              
#> [288,] ""              
#> [289,] ""              
#> [290,] ""              
#> [291,] ""              
#> [292,] ""              
#> [293,] ""              
#> [294,] ""              
#> [295,] ""              
#> [296,] ""              
#> [297,] ""              
#> [298,] ""              
#> [299,] ""              
#> [300,] ""              
#> [301,] ""              
#> [302,] ""              
#> [303,] ""              
#> [304,] ""              
#> [305,] ""              
#> [306,] ""              
#> [307,] ""              
#> [308,] ""              
#> [309,] ""              
#> [310,] ""              
#> [311,] ""              
#> [312,] ""              
#> [313,] ""              
#> [314,] ""              
#> [315,] ""              
#> [316,] ""              
#> [317,] ""              
#> [318,] ""              
#> [319,] ""              
#> [320,] ""              
#> [321,] ""              
#> [322,] ""              
#> [323,] ""              
#> [324,] ""              
#> [325,] ""              
#> [326,] ""              
#> [327,] ""              
#> [328,] ""              
#> [329,] ""              
#> [330,] ""              
#> [331,] ""              
#> [332,] ""              
#> [333,] ""              
#> [334,] ""              
#> [335,] ""              
#> [336,] ""              
#> [337,] ""              
#> [338,] ""              
#> [339,] ""              
#> [340,] ""              
#> [341,] ""              
#> [342,] ""              
#> [343,] ""              
#> [344,] ""              
#> [345,] ""              
#> [346,] ""              
#> [347,] ""              
#> [348,] ""              
#> [349,] ""              
#> [350,] ""              
#> [351,] ""              
#> [352,] ""              
#> [353,] ""              
#> [354,] ""              
#> [355,] ""              
#> [356,] ""              
#> [357,] ""              
#> [358,] ""              
#> [359,] ""              
#> [360,] ""              
#> [361,] ""              
#> [362,] ""              
#> [363,] ""              
#> [364,] ""              
#> [365,] ""              
#> [366,] ""              
#> [367,] ""              
#> [368,] ""              
#> [369,] ""              
#> [370,] ""              
#> [371,] ""              
#> [372,] ""              
#> [373,] ""              
#> [374,] ""              
#> [375,] ""              
#> [376,] ""              
#> [377,] ""              
#> [378,] ""              
#> [379,] ""              
#> [380,] ""              
#> [381,] ""              
#> [382,] ""              
#> [383,] ""              
#> [384,] ""              
#> [385,] ""              
#> [386,] ""              
#> [387,] ""              
#> [388,] ""              
#> [389,] ""              
#> [390,] ""              
#> [391,] ""              
#> [392,] ""              
#> [393,] ""              
#> [394,] ""              
#> [395,] ""              
#> [396,] ""              
#> [397,] ""              
#> [398,] ""              
#> [399,] ""              
#> [400,] ""              
#> [401,] ""              
#> [402,] ""              
#> [403,] ""              
#> [404,] ""              
#> [405,] ""              
#> [406,] ""              
#> [407,] ""              
#> [408,] "pop"           
#> [409,] "pop"           
#> [410,] "pop"           
#> [411,] "pop"           
#> [412,] "pop"           
#> [413,] "pop"           
#> [414,] "pop"           
#> [415,] "pop"           
#> [416,] "pop"           
#> [417,] "pop"           
#> [418,] "pop"           
#> [419,] "pop"           
#> [420,] "pop"           
#> [421,] "pop"           
#> [422,] "pop"           
#> [423,] "pop"           
#> [424,] "pop"           
#> [425,] "pop"           
#> [426,] "pop"           
#> [427,] "pop"           
#> [428,] "pop"           
#> [429,] "pop"           
#> [430,] "pop"           
#> [431,] "pop"           
#> [432,] "pop"           
#> [433,] "pop"           
#> [434,] "pop"           
#> [435,] "pop"           
#> [436,] "pop"           
#> [437,] "pop"           
#> [438,] "pop"           
#> [439,] "pop"           
#> [440,] "pop"           
#> [441,] "pop"           
#> [442,] "pop"           
#> [443,] "pop"           
#> [444,] "pop"           
#> [445,] "pop"           
#> [446,] "pop"           
#> [447,] "pop"           
#> [448,] "pop"           
#> [449,] "pop"           
#> [450,] "pop"           
#> [451,] "pop"           
#> [452,] "pop"           
#> [453,] "pop"           
#> [454,] "pop"           
#> [455,] "pop"           
#> [456,] "pop"           
#> [457,] "pop"           
#> [458,] "pop"           
#> [459,] "pop"           
#> [460,] "pop"           
#> [461,] "pop"           
#> [462,] "pop"           
#> [463,] "pop"           
#> [464,] "pop"           
#> [465,] "pop"           
#> [466,] "pop"           
#> [467,] ""