Summarize count (and percent) of rows with exactly (and at least) N cols >= various thresholds

This wraps 4 functions to return 4 tables: using colcounter_summary(), colcounter_summary_pct(), colcounter_summary_cum(), colcounter_summary_cum_pct() For another view and text explanations of the findings, see count_sites_with_n_high_scores()

Usage

colcounter_summary_all(x, thresholdlist, ...)

Arguments

x

Data.frame or matrix of numbers to be compared to threshold value, like percentiles for example.

thresholdlist

vector of numeric threshold values to compare to

...

passed to the 4 functions like or.tied=TRUE, na.rm=TRUE, below=FALSE, one.cut.per.col=FALSE

Value

A table of cumulative frequency counts etc., including count, cum, pct, cum_pct

See also

Examples

    # df <-  bg22[ , names.ej.pctile]
 df <- data.frame(a=rep(80,4),b=rep(93,4), col3=c(49,98,100,100))
 bench <- 5 * (0:20)
 a3 <- colcounter_summary_all(df, bench)
 a3[,'95',]
#>      stat
#> count count cum pct cum_pct
#>     0     1   4  25     100
#>     1     3   3  75      75
#>     2     0   0   0       0
#>     3     0   0   0       0
 a3[,,'cum_pct']
#>      cut
#> count   0   5  10  15  20  25  30  35  40  45  50  55  60  65  70  75  80  85  90  95 100
#>     0 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100
#>     1 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100  75  50
#>     2 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100 100  75  75   0   0
#>     3 100 100 100 100 100 100 100 100 100 100  75  75  75  75  75  75  75   0   0   0   0
 a3['0',,]; a3[1,,]
#>      stat
#> cut   count cum pct cum_pct
#>   0       0   4   0     100
#>   5       0   4   0     100
#>   10      0   4   0     100
#>   15      0   4   0     100
#>   20      0   4   0     100
#>   25      0   4   0     100
#>   30      0   4   0     100
#>   35      0   4   0     100
#>   40      0   4   0     100
#>   45      0   4   0     100
#>   50      0   4   0     100
#>   55      0   4   0     100
#>   60      0   4   0     100
#>   65      0   4   0     100
#>   70      0   4   0     100
#>   75      0   4   0     100
#>   80      0   4   0     100
#>   85      0   4   0     100
#>   90      0   4   0     100
#>   95      1   4  25     100
#>   100     2   4  50     100
#>      stat
#> cut   count cum pct cum_pct
#>   0       0   4   0     100
#>   5       0   4   0     100
#>   10      0   4   0     100
#>   15      0   4   0     100
#>   20      0   4   0     100
#>   25      0   4   0     100
#>   30      0   4   0     100
#>   35      0   4   0     100
#>   40      0   4   0     100
#>   45      0   4   0     100
#>   50      0   4   0     100
#>   55      0   4   0     100
#>   60      0   4   0     100
#>   65      0   4   0     100
#>   70      0   4   0     100
#>   75      0   4   0     100
#>   80      0   4   0     100
#>   85      0   4   0     100
#>   90      0   4   0     100
#>   95      1   4  25     100
#>   100     2   4  50     100
 a3[dim(a3)[1],,]
#>      stat
#> cut   count cum pct cum_pct
#>   0       4   4 100     100
#>   5       4   4 100     100
#>   10      4   4 100     100
#>   15      4   4 100     100
#>   20      4   4 100     100
#>   25      4   4 100     100
#>   30      4   4 100     100
#>   35      4   4 100     100
#>   40      4   4 100     100
#>   45      4   4 100     100
#>   50      3   3  75      75
#>   55      3   3  75      75
#>   60      3   3  75      75
#>   65      3   3  75      75
#>   70      3   3  75      75
#>   75      3   3  75      75
#>   80      3   3  75      75
#>   85      0   0   0       0
#>   90      0   0   0       0
#>   95      0   0   0       0
#>   100     0   0   0       0
 # a3['12',,]; a3[13,,]
 
 if (FALSE) {
 barplot(colcounter_summary_cum_pct(pdata, pcuts)[ , '80'],
    ylab='% of places', xlab='# of indicators at/above threshold',
    main='% of places with at least N/12 indicators >=80th percentile')

 barplot(colcounter_summary(pdata, pcuts)[2:13 , '95'],
    ylab='# of places', xlab='# of indicators at/above threshold',
    main='# of places with exactly N/12 indicators >=95th percentile')

  # pdata <- ejscreen package file bg22[ , names.e.pctile]
  colcounter_summary_cum_pct(pdata,c(50,80,90,95))
  xs <- 1:12
  plot(x=xs, y=colcounter_summary_cum_pct(pdata, 50)[xs+1],
   type='b', col='gray', ylim=c(0, 100),
    main='% of places with at least x/12 indicators >=Nth percentile',
     ylab='% of places', xlab='# of indicators')
  points(xs, colcounter_summary_cum_pct(pdata, 80)[xs+1], type='b', col='blue')
  points(xs, colcounter_summary_cum_pct(pdata, 90)[xs+1], type='b', col='orange')
  points(xs, colcounter_summary_cum_pct(pdata, 95)[xs+1], type='b', col='red')
  legend(x = 'topright', legend = paste0('>= ', c(50, 80, 90, 95),'th percentile'),
   fill = c('gray', 'blue', 'orange', 'red'))

  # pdata <- bg22[ ,  names.ej.pctile]
  colcounter_summary_cum_pct(pdata,c(50,80,90,95))
  xs <- 1:12
  plot(x=xs, y=colcounter_summary_cum_pct(pdata, 50)[xs+1], 
    type='b', col='gray', ylim=c(0, 40),
    main='% of places with at least x/12 indicators >=Nth percentile', ylab='% of places', 
    xlab='# of indicators')
  points(xs, colcounter_summary_cum_pct(pdata, 80)[xs+1], type='b', col='blue')
  points(xs, colcounter_summary_cum_pct(pdata, 90)[xs+1], type='b', col='orange')
  points(xs, colcounter_summary_cum_pct(pdata, 95)[xs+1], type='b', col='red')
  legend(x = 'topright', legend = paste0('>= ', c(50, 80, 90, 95),'th percentile'), 
    fill = c('gray', 'blue', 'orange', 'red'))
}