Summarize what percent of rows have N columns at or above (or below) various thresholds

See colcounter_summary() for details and examples

Usage

colcounter_summary_pct(x, thresholdlist, ...)

Arguments

x

Data.frame or matrix of numbers to be compared to threshold value, like percentiles for example.

thresholdlist

vector of numeric threshold values to compare to

...

passed to colcounter_summary() like or.tied=TRUE, na.rm=TRUE, below=FALSE, one.cut.per.col=FALSE

See also