create_filename - construct custom filename for downloads in EJAM app

Builds filename from file description, analysis title, buffer distance, and site selection method. Values are pulled from Shiny app if used there.

Usage

create_filename(
  file_desc = "",
  title = "",
  buffer_dist = 0,
  site_method = "",
  with_datetime = TRUE,
  ext = NULL
)

Arguments

file_desc

file description, such as "short report", "long report", "results_table"

title

analysis title

buffer_dist

buffer distance

site_method

site selection method, such as NAICS, FRS, SHP, latlon

with_datetime

boolean to include date and time

ext

optional file extension. will check for '.' and add if not provided

Value

Returns string of pasted filename with specified components

Examples

# specify all arguments
create_filename(file_desc = 'results_table', title = 'My Title', buffer_dist = 1, site_method = 'NAICS', with_datetime=TRUE, ext = '.xlsx')
#> Error in create_filename(file_desc = "results_table", title = "My Title",     buffer_dist = 1, site_method = "NAICS", with_datetime = TRUE,     ext = ".xlsx"): could not find function "create_filename"
# specify title only
create_filename(title = 'Summary of EJ Analysis')
#> Error in create_filename(title = "Summary of EJ Analysis"): could not find function "create_filename"