Load datasets from local disk folder

Utility for analysts / developers to store large block / other data locally instead of redownloading

Usage

dataload_from_local(
  varnames = .arrow_ds_names[1:3],
  ext = c(".arrow", ".rda")[1],
  fun = c("arrow::read_ipc_file", "load")[1],
  envir = globalenv(),
  folder_local_source = NULL,
  justchecking = FALSE,
  testing = FALSE,
  silent = FALSE
)

Arguments

varnames

use defaults, or vector of names like "bgej" or use "all" to get all available

ext

use defaults

fun

use defaults

envir

use defaults. see dataload_from_pins()

folder_local_source

Your local folder path. see dataload_from_pins()

justchecking

use defaults. see dataload_from_pins()

testing

use defaults

silent

set to TRUE to stop cat() printing to console like when running tests

Value

vector of paths to files (as derived from varnames) that were actually found in folder_local_source, but only for those not already in memory, so it is just the ones loaded from disk because not already in memory and found on disk locally.

Details

See dataload_from_pins() also.

rm(bgid2fips, blockid2fips, blockpoints, blockwts, quaddata)

dataload_from_local(folder_local_source = '.')