Utility to download / load datasets from pin board in pkgdown github pages of public repo

Utility to download / load datasets from pin board in pkgdown github pages of public repo

Usage

dataload_from_urlpins(
  varnames = .arrow_ds_names[1:3],
  server = "https://ejanalysis.github.io/ejscreendata/pins/",
  envir = globalenv(),
  justchecking = FALSE
)

Arguments

varnames

vector of names of data objects stored in the pins board

server

do not change

envir

do not change default of globalenv()

justchecking

set to TRUE if you just need to see what is stored in the pins board

Value

subset of varnames that is a vector of the ones successfully assigned to specified environment