UTILITY - See names and size of data sets in package(s) - internal utility function

Wrapper for data() and can get memory size of objects

Usage

datapack(pkg = ejampackages, len = 30, sortbysize = TRUE, simple = TRUE)

Arguments

pkg

a character vector giving the package(s) to look in for data sets

len

Only affects what is printed to console - specifies the number of characters to limit Title to, making it easier to see in the console.

sortbysize

if TRUE (and simple=F), sort by increasing size of object, within each package, not alpha.

simple

FALSE to get object sizes, etc., or TRUE to just get names in each package, like data(package = "EJAM")$results[, c("Package", 'Item')]

Value

If simple = TRUE, data.frame with colnames Package and Item.

If simple = FALSE, data.frame with colnames Package, Item, size, Title.Short

Details

do not rely on this much - it was a quick utility. It may create and leave objects in global envt - not careful about that.

Examples

 # see just a vector of the data object names
 data(package = "EJAM")$results[, 'Item']
#>   [1] "NAICS"                                              
#>   [2] "SIC"                                                
#>   [3] "avg.in.us"                                          
#>   [4] "bg_cenpop2020"                                      
#>   [5] "bgpts"                                              
#>   [6] "blockgroupstats"                                    
#>   [7] "censusplaces"                                       
#>   [8] "ejampackages"                                       
#>   [9] "epa_programs"                                       
#>  [10] "formulas_d"                                         
#>  [11] "frsprogramcodes"                                    
#>  [12] "high_pctiles_tied_with_min"                         
#>  [13] "islandareas"                                        
#>  [14] "lat_alias"                                          
#>  [15] "lon_alias"                                          
#>  [16] "mact_table"                                         
#>  [17] "meters_per_mile"                                    
#>  [18] "naics_counts"                                       
#>  [19] "naicstable"                                         
#>  [20] "names_all"                                          
#>  [21] "names_all_r"                                        
#>  [22] "names_avg_these"                                    
#>  [23] "names_d"                                            
#>  [24] "names_d_avg"                                        
#>  [25] "names_d_avg_friendly"                               
#>  [26] "names_d_count"                                      
#>  [27] "names_d_count_friendly"                             
#>  [28] "names_d_friendly"                                   
#>  [29] "names_d_other_count"                                
#>  [30] "names_d_other_count_friendly"                       
#>  [31] "names_d_pctile"                                     
#>  [32] "names_d_pctile_friendly"                            
#>  [33] "names_d_ratio_to_avg"                               
#>  [34] "names_d_ratio_to_avg_friendly"                      
#>  [35] "names_d_ratio_to_state_avg"                         
#>  [36] "names_d_ratio_to_state_avg_friendly"                
#>  [37] "names_d_state_avg"                                  
#>  [38] "names_d_state_avg_friendly"                         
#>  [39] "names_d_state_pctile"                               
#>  [40] "names_d_state_pctile_friendly"                      
#>  [41] "names_d_subgroups"                                  
#>  [42] "names_d_subgroups_alone"                            
#>  [43] "names_d_subgroups_alone_avg"                        
#>  [44] "names_d_subgroups_alone_avg_friendly"               
#>  [45] "names_d_subgroups_alone_count"                      
#>  [46] "names_d_subgroups_alone_friendly"                   
#>  [47] "names_d_subgroups_alone_pctile"                     
#>  [48] "names_d_subgroups_alone_pctile_friendly"            
#>  [49] "names_d_subgroups_alone_ratio_to_avg"               
#>  [50] "names_d_subgroups_alone_ratio_to_avg_friendly"      
#>  [51] "names_d_subgroups_alone_ratio_to_state_avg"         
#>  [52] "names_d_subgroups_alone_ratio_to_state_avg_friendly"
#>  [53] "names_d_subgroups_alone_state_avg"                  
#>  [54] "names_d_subgroups_alone_state_avg_friendly"         
#>  [55] "names_d_subgroups_alone_state_pctile"               
#>  [56] "names_d_subgroups_alone_state_pctile_friendly"      
#>  [57] "names_d_subgroups_avg"                              
#>  [58] "names_d_subgroups_avg_friendly"                     
#>  [59] "names_d_subgroups_count"                            
#>  [60] "names_d_subgroups_friendly"                         
#>  [61] "names_d_subgroups_nh"                               
#>  [62] "names_d_subgroups_nh_avg"                           
#>  [63] "names_d_subgroups_nh_avg_friendly"                  
#>  [64] "names_d_subgroups_nh_count"                         
#>  [65] "names_d_subgroups_nh_friendly"                      
#>  [66] "names_d_subgroups_nh_pctile"                        
#>  [67] "names_d_subgroups_nh_pctile_friendly"               
#>  [68] "names_d_subgroups_nh_ratio_to_avg"                  
#>  [69] "names_d_subgroups_nh_ratio_to_avg_friendly"         
#>  [70] "names_d_subgroups_nh_ratio_to_state_avg"            
#>  [71] "names_d_subgroups_nh_ratio_to_state_avg_friendly"   
#>  [72] "names_d_subgroups_nh_state_avg"                     
#>  [73] "names_d_subgroups_nh_state_avg_friendly"            
#>  [74] "names_d_subgroups_nh_state_pctile"                  
#>  [75] "names_d_subgroups_nh_state_pctile_friendly"         
#>  [76] "names_d_subgroups_pctile"                           
#>  [77] "names_d_subgroups_pctile_friendly"                  
#>  [78] "names_d_subgroups_ratio_to_avg"                     
#>  [79] "names_d_subgroups_ratio_to_avg_friendly"            
#>  [80] "names_d_subgroups_ratio_to_state_avg"               
#>  [81] "names_d_subgroups_ratio_to_state_avg_friendly"      
#>  [82] "names_d_subgroups_state_avg"                        
#>  [83] "names_d_subgroups_state_avg_friendly"               
#>  [84] "names_d_subgroups_state_pctile"                     
#>  [85] "names_d_subgroups_state_pctile_friendly"            
#>  [86] "names_e"                                            
#>  [87] "names_e_avg"                                        
#>  [88] "names_e_avg_friendly"                               
#>  [89] "names_e_friendly"                                   
#>  [90] "names_e_pctile"                                     
#>  [91] "names_e_pctile_friendly"                            
#>  [92] "names_e_ratio_to_avg"                               
#>  [93] "names_e_ratio_to_avg_friendly"                      
#>  [94] "names_e_ratio_to_state_avg"                         
#>  [95] "names_e_ratio_to_state_avg_friendly"                
#>  [96] "names_e_state_avg"                                  
#>  [97] "names_e_state_avg_friendly"                         
#>  [98] "names_e_state_pctile"                               
#>  [99] "names_e_state_pctile_friendly"                      
#> [100] "names_ej"                                           
#> [101] "names_ej_friendly"                                  
#> [102] "names_ej_pctile"                                    
#> [103] "names_ej_pctile_friendly"                           
#> [104] "names_ej_state"                                     
#> [105] "names_ej_state_friendly"                            
#> [106] "names_ej_state_pctile"                              
#> [107] "names_ej_state_pctile_friendly"                     
#> [108] "names_ej_supp"                                      
#> [109] "names_ej_supp_friendly"                             
#> [110] "names_ej_supp_pctile"                               
#> [111] "names_ej_supp_pctile_friendly"                      
#> [112] "names_ej_supp_state"                                
#> [113] "names_ej_supp_state_friendly"                       
#> [114] "names_ej_supp_state_pctile"                         
#> [115] "names_ej_supp_state_pctile_friendly"                
#> [116] "names_need_pctile"                                  
#> [117] "names_need_state_pctile"                            
#> [118] "names_other"                                        
#> [119] "names_pct_as_fraction_blockgroupstats"              
#> [120] "names_pct_as_fraction_ejamit"                       
#> [121] "names_pct_as_fraction_ejscreenit"                   
#> [122] "names_pctile"                                       
#> [123] "names_ratio_to_avg_these"                           
#> [124] "names_ratio_to_state_avg_these"                     
#> [125] "names_state_avg_these"                              
#> [126] "names_state_pctile"                                 
#> [127] "names_these"                                        
#> [128] "names_these_avg"                                    
#> [129] "names_these_ratio_to_avg"                           
#> [130] "names_these_ratio_to_state_avg"                     
#> [131] "names_these_state_avg"                              
#> [132] "names_wts"                                          
#> [133] "namez"                                              
#> [134] "regionstats"                                        
#> [135] "sictable"                                           
#> [136] "sites2blocks_example1000pts_1miles"                 
#> [137] "sites2blocks_example100pts_1miles"                  
#> [138] "sites2blocks_example10pts_1miles"                   
#> [139] "stateinfo"                                          
#> [140] "stateinfo2"                                         
#> [141] "states_shapefile"                                   
#> [142] "statestats"                                         
#> [143] "test_address_parts1"                                
#> [144] "test_address_table"                                 
#> [145] "test_address_table_9"                               
#> [146] "test_address_table_goodnames"                       
#> [147] "test_address_table_withfull"                        
#> [148] "test_addresses2"                                    
#> [149] "test_addresses_9"                                   
#> [150] "test_regid"                                         
#> [151] "test_xtrac"                                         
#> [152] "testoutput_doaggregate_1000pts_1miles"              
#> [153] "testoutput_doaggregate_100pts_1miles"               
#> [154] "testoutput_doaggregate_10pts_1miles"                
#> [155] "testoutput_ejamit_1000pts_1miles"                   
#> [156] "testoutput_ejamit_100pts_1miles"                    
#> [157] "testoutput_ejamit_10pts_1miles"                     
#> [158] "testoutput_ejscreenit_10pts_1miles"                 
#> [159] "testoutput_getblocksnearby_1000pts_1miles"          
#> [160] "testoutput_getblocksnearby_100pts_1miles"           
#> [161] "testoutput_getblocksnearby_10pts_1miles"            
#> [162] "testpoints_10"                                      
#> [163] "testpoints_100"                                     
#> [164] "testpoints_1000"                                    
#> [165] "testpoints_10000"                                   
#> [166] "testpoints_100_dt"                                  
#> [167] "testpoints_conus5"                                  
#> [168] "usastats"                                           
 
 # not actually sorted within each pkg by default
 datapack()
#> Error in datapack(): could not find function "datapack"
 # not actually sorted by default
 datapack("EJAM")$Item
#> Error in datapack("EJAM"): could not find function "datapack"
 ##datapack("MASS", simple=T)
 
 # sorted by size if simple=F
 ##datapack("datasets", simple=F)
 x <- datapack(simple = F)
#> Error in datapack(simple = F): could not find function "datapack"
 # sorted by size already, to see largest ones among all these pkgs:
 tail(x[, 1:3], 20)
#> Error in tail(x[, 1:3], 20): object 'x' not found
 
 # sorted alphabetically within each pkg
 x[order(x$Package, x$Item), 1:2]
#> Error in eval(expr, envir, enclos): object 'x' not found
 # sorted alphabetically across all the pkgs
 x[order(x$Item), 1:2]
#> Error in eval(expr, envir, enclos): object 'x' not found