Utility NOT YET WORKING - AccessDenied

Utility to write object(s) like a dataset to DMAP Data Commons, formatted as .arrow or .rda

Usage

datawrite_to_aws(
  varnames = c("bgid2fips", "blockid2fips", "blockpoints", "blockwts", "quaddata"),
  ext = c(".arrow", ".rda")[2],
  fun = c("arrow::write_ipc_file", "save")[2],
  mybucket = "dmap-data-commons-oa",
  mybucketfolder = "EJAM",
  justchecking = TRUE
)

Arguments

varnames

vector of object names to upload

ext

file .extension appropriate to the format and fun, like ".rda" or ".arrow"

fun

function to use, but as a character string, like "arrow::write_ipc_file" but fun is ignored if ext=".rda" since it then just uses s3save()

mybucket

do not need to change

mybucketfolder

do not need to change

justchecking

set this to FALSE to actually upload instead of just viewing in console the commands to be used, to test/check this

Value

the paths of the objects on server

Details

mybucket <- 'dmap-data-commons-oa' #

bucket_contents <- data.table::rbindlist(

get_bucket(bucket = mybucket, prefix = "EJAM"),

fill = TRUE

)

bucket_contents