compare ejamit results for more than one radius run ejamit() once per radius, get a table with a row per radius

compare ejamit results for more than one radius run ejamit() once per radius, get a table with a row per radius

Usage

distance_comparison_ejamit(
  sitepoints,
  radii = c(1, 2, 3),
  quiet = TRUE,
  silentinteractive = TRUE,
  plot = TRUE,
  ...
)

Arguments

sitepoints

like for ejamit()

radii

vector of radius values like 1:3 for ejamit()

quiet

passed to ejamit()

silentinteractive

passed to ejamit()

...

passed to ejamit()

Value

data.table like ejamit()$results_overall but with one row per radius.

See also

Examples

  radii <- c(1,2,3,6,10)
  radii <- c(1, 10)
  pts <- testpoints_100
  pts <- testpoints_10
  x <- distance_comparison_ejamit(pts, radii = radii)
  EJAM:::ejamit_compare_distances2plot(x)