EJAM/EJSCREEN comparisons - see summary stats after using ejscreen_vs_ejam()

EJAM/EJSCREEN comparisons - see summary stats after using ejscreen_vs_ejam()

ejscreen_vs_ejam_summary(
  z = ejscreen_vs_ejam(),
  myvars = colnames(z$EJAM),
  tol = 0.01,
  prob = 0.95,
  na.rm = TRUE
)

Arguments

z

output of ejscreen_vs_ejam()

myvars

optional to check just a subset of the colnames found in z$EJAM and z$EJSCREEN, such as myvars = c(names_d, names_d_subgroups) or myvars = grep("pctile", colnames(z$EJAM), value = T)

tol

optional, set this so that results can be said to agree with this tolerance if they differ by less than tol percent where tol is expressed as a fraction 0 to 1.

prob

optional fraction of 1 representing percentile p to check for absolute percentage differences. See within.x.pct.at.p.pct.of.sites value that is returned.

na.rm

needs testing, optional

Value

A data.frame of summary stats showing counts and percents of analyzed sites (or those with valid data that are found in both EJAM and EJSCREEN outputs), indicating how many of the sites agree between EJSCREEN and EJAM estimates, exactly as reported or within some tolerance. Columns include

"indicator" (variable name)

"sites.with.data.ejam" How many of the sites had a value from EJAM for the given indicator?

"sites.with.data.neither" How many sites had NA from both EJAM and EJSCREEN?

"sites.with.data.both"

"sites.agree.rounded" How many sites agree (EJSCREEN vs EJAM) in the value shown on reports? i.e., the reported, rounded value.

"sites.agree.within.tol" How many sites agree within tol? (i.e., with tol x 100 percent)

"pct.of.sites.agree.rounded" as a percent 0-100% of sites with data

"pct.of.sites.agree.within.tol" as a percent 0-100% of sites with data

"median.abs.diff" Median over sites with data, of the absolute differences, EJAM - EJSCREEN

"max.abs.diff"

"mean.pct.diff" Percent difference 0-100% is absolute value of 100*(ratio - 1), and ratio is EJAM/EJSCREEN

"median.pct.diff" 0-100%

"max.pct.diff" 0-100%

"within.x.pct.at.p.pct.of.sites" X, where EJAM and EJSCREEN agree within X percent 0-100% or better at prob share of sites. Prob share as used in this last stat should mean prob (e.g. 0.95) share of sites have an absolute percentage difference in estimated indicator values that is less than or equal to x where x is one of the actual values of abspctdiff found * 100. It uses 100 * quantile(y, probs = prob, type = 1)

Examples