EJAM/EJSCREEN comparisons - for interactive RStudio use

EJAM/EJSCREEN comparisons - for interactive RStudio use

Usage

ejscreen_vscript(
  defdir = ".",
  n,
  newpts,
  pts,
  radius,
  savedejscreentableoutput,
  x100fix = TRUE,
  x100varnames = names_pct_as_fraction_ejamit
)

Arguments

defdir

folder

n

how many points

newpts

logical, if need new set of random locations

pts

data.frame of points with columns lat,lon

radius

miles

savedejscreentableoutput

is a data.table from ejscreenit()$table

Value

a list of data frames, with names EJSCREEN, EJAM, EJSCREEN_shown, EJAM_shown, same_shown, ratio, diff, absdiff, pctdiff, abspctdiff

diff is EJAM - EJSCREEN

ratio is EJAM / EJSCREEN

pctdiff is ratio - 1

abs is absolute value

For each data.frame, colnames are indicators like pop, blockcount_near_site, etc. and rows represent sites analyzed.

Details

THIS IS A SCRIPT FOR INTERACTIVE RSTUDIO CONSOLE USE TO RUN A SET OF POINTS THROUGH BOTH EJSCREEN AND EJAM AND SAVE STATS ON THE DIFFERENCES – including letting you use saved ejscreenapi results and input points so you can iterate and rerun just the EJAM portion and compare to the saved benchmark data.

Examples

vs = ejscreen_vscript(pts = testpoints_100, radius = 3)

## look only at places where blockcount was identical, to exclude that as source of difference
vs_blocksmatch = lapply(vs, function(df) df[vs$absdiff[, "blockcount_near_site"] == 0, ])
# vss = ejscreen_vs_ejam_summary(vs )
vssb = ejscreen_vs_ejam_summary(vs_blocksmatch)
# vss[vss$indicator %in% names_these, c(1,7,8)]
vssb[vssb$indicator %in% names_these, c(1,7,8)]
x = vssb[vssb$indicator %in% c('pop', names_these), c(1,7,8)] 
x[order(x$pct.of.sites.agree.within.tol), ]

xx = cbind(indicator = vssb$indicator[vssb$indicator %in% c('pop', names_these)],
 round(vssb[vssb$indicator %in% c('pop', names_these),
  c('pct.of.sites.agree.within.tol', 'max.abs.diff', 'median.abs.diff')], 0))
xx[xx$pct.of.sites.agree.within.tol < 100, ]