Utility to get quick view in RStudio viewer of datatable of ejscreenapi results

Utility to get quick view in RStudio viewer of datatable of ejscreenapi results

Usage

ejscreenit_see_table(x)

Arguments

x

output from ejscreenit(), not just the table element of that output, so it easier to do something like in the example.

Value

output of DT::datatable()

Details

Also see functions in the EJAM package.

Examples

  # out <- ejscreenit(testpoints_5, radius = 1)
  out <- testoutput_ejscreenit_5
  ejscreenit_see_table(out)
#> Warning: lat or lon or radius is not numeric
#> Warning: lat or lon or radius missing entirely (length of a vector is zero
#> Warning: lat or lon or radius is NULL

  out$map