Get county or state names from county or state FIPS codes

Get county or state names from county or state FIPS codes

fips2name(fips, ...)

Arguments

fips

vector of US Census FIPS codes for Counties (5 digits each) or States (2 digits). Can be string or numeric, with or without leading zeroes.

...

passed to fips2countyname() to control whether it appends something like , NY or , New York after county name

Value

vector of state and/or county names, where county names optionally have comma and 2-character abbreviation or full state name.

Examples

  fips2name(fips_counties_from_state_abbrev(c("AK", "LA"))  )
#>  [1] "Aleutians East Borough, AK"           
#>  [2] "Aleutians West Census Area, AK"       
#>  [3] "Anchorage Municipality, AK"           
#>  [4] "Bethel Census Area, AK"               
#>  [5] "Bristol Bay Borough, AK"              
#>  [6] "Chugach Census Area, AK"              
#>  [7] "Copper River Census Area, AK"         
#>  [8] "Denali Borough, AK"                   
#>  [9] "Dillingham Census Area, AK"           
#> [10] "Fairbanks North Star Borough, AK"     
#> [11] "Haines Borough, AK"                   
#> [12] "Hoonah-Angoon Census Area, AK"        
#> [13] "Juneau City and Borough, AK"          
#> [14] "Kenai Peninsula Borough, AK"          
#> [15] "Ketchikan Gateway Borough, AK"        
#> [16] "Kodiak Island Borough, AK"            
#> [17] "Kusilvak Census Area, AK"             
#> [18] "Lake and Peninsula Borough, AK"       
#> [19] "Matanuska-Susitna Borough, AK"        
#> [20] "Nome Census Area, AK"                 
#> [21] "North Slope Borough, AK"              
#> [22] "Northwest Arctic Borough, AK"         
#> [23] "Petersburg Borough, AK"               
#> [24] "Prince of Wales-Hyder Census Area, AK"
#> [25] "Sitka City and Borough, AK"           
#> [26] "Skagway Municipality, AK"             
#> [27] "Southeast Fairbanks Census Area, AK"  
#> [28] "Wrangell City and Borough, AK"        
#> [29] "Yakutat City and Borough, AK"         
#> [30] "Yukon-Koyukuk Census Area, AK"        
#> [31] "Acadia Parish, LA"                    
#> [32] "Allen Parish, LA"                     
#> [33] "Ascension Parish, LA"                 
#> [34] "Assumption Parish, LA"                
#> [35] "Avoyelles Parish, LA"                 
#> [36] "Beauregard Parish, LA"                
#> [37] "Bienville Parish, LA"                 
#> [38] "Bossier Parish, LA"                   
#> [39] "Caddo Parish, LA"                     
#> [40] "Calcasieu Parish, LA"                 
#> [41] "Caldwell Parish, LA"                  
#> [42] "Cameron Parish, LA"                   
#> [43] "Catahoula Parish, LA"                 
#> [44] "Claiborne Parish, LA"                 
#> [45] "Concordia Parish, LA"                 
#> [46] "De Soto Parish, LA"                   
#> [47] "East Baton Rouge Parish, LA"          
#> [48] "East Carroll Parish, LA"              
#> [49] "East Feliciana Parish, LA"            
#> [50] "Evangeline Parish, LA"                
#> [51] "Franklin Parish, LA"                  
#> [52] "Grant Parish, LA"                     
#> [53] "Iberia Parish, LA"                    
#> [54] "Iberville Parish, LA"                 
#> [55] "Jackson Parish, LA"                   
#> [56] "Jefferson Parish, LA"                 
#> [57] "Jefferson Davis Parish, LA"           
#> [58] "Lafayette Parish, LA"                 
#> [59] "Lafourche Parish, LA"                 
#> [60] "LaSalle Parish, LA"                   
#> [61] "Lincoln Parish, LA"                   
#> [62] "Livingston Parish, LA"                
#> [63] "Madison Parish, LA"                   
#> [64] "Morehouse Parish, LA"                 
#> [65] "Natchitoches Parish, LA"              
#> [66] "Orleans Parish, LA"                   
#> [67] "Ouachita Parish, LA"                  
#> [68] "Plaquemines Parish, LA"               
#> [69] "Pointe Coupee Parish, LA"             
#> [70] "Rapides Parish, LA"                   
#> [71] "Red River Parish, LA"                 
#> [72] "Richland Parish, LA"                  
#> [73] "Sabine Parish, LA"                    
#> [74] "St. Bernard Parish, LA"               
#> [75] "St. Charles Parish, LA"               
#> [76] "St. Helena Parish, LA"                
#> [77] "St. James Parish, LA"                 
#> [78] "St. John the Baptist Parish, LA"      
#> [79] "St. Landry Parish, LA"                
#> [80] "St. Martin Parish, LA"                
#> [81] "St. Mary Parish, LA"                  
#> [82] "St. Tammany Parish, LA"               
#> [83] "Tangipahoa Parish, LA"                
#> [84] "Tensas Parish, LA"                    
#> [85] "Terrebonne Parish, LA"                
#> [86] "Union Parish, LA"                     
#> [87] "Vermilion Parish, LA"                 
#> [88] "Vernon Parish, LA"                    
#> [89] "Washington Parish, LA"                
#> [90] "Webster Parish, LA"                   
#> [91] "West Baton Rouge Parish, LA"          
#> [92] "West Carroll Parish, LA"              
#> [93] "West Feliciana Parish, LA"            
#> [94] "Winn Parish, LA"                      
  fips2name(c(22, 02013))  # can have mix where some are a whole state and others are a county.
#> [1] "Louisiana"                  "Aleutians East Borough, AK"