utility to multiply certain percentage columns by 100 to convert 0-1.00 into 0-100

multiplies some data to rescale percentages stored as 0 to 1, into 0-100

fix_pctcols_x100(df, cnames = NULL)

Arguments

df

data.frame but can be data.table

cnames

colnames in df of indicators to multiply by 100, like those in

names_pct_as_fraction_ejamit,

names_pct_as_fraction_blockgroupstats, or

names_pct_as_fraction_ejscreenit

Value

df with data in specified columns multiplied by 100

Examples

 y = data.frame(pctlowinc = 1:2, pctpre1960 = 1:2, avg.pctunemployed = 1:2, avg.pctpre1960 = 1:2)
 fix_pctcols_x100(y)
#> Error in fix_pctcols_x100(y): could not find function "fix_pctcols_x100"
 fix_pctcols_x100(y, names_pct_as_fraction_ejscreenit)
#> Error in fix_pctcols_x100(y, names_pct_as_fraction_ejscreenit): could not find function "fix_pctcols_x100"
 fix_pctcols_x100(y, names_pct_as_fraction_blockgroupstats)
#> Error in fix_pctcols_x100(y, names_pct_as_fraction_blockgroupstats): could not find function "fix_pctcols_x100"
 fix_pctcols_x100(y, names_pct_as_fraction_ejamit)
#> Error in fix_pctcols_x100(y, names_pct_as_fraction_ejamit): could not find function "fix_pctcols_x100"
 cat("\n\n")
#> 
#> 
 names_pct_as_fraction_ejscreenit
#> [1] "pctpre1960"           "avg.pctpre1960"       "state.avg.pctpre1960"
 names_pct_as_fraction_blockgroupstats
#>  [1] "Demog.Index"      "Demog.Index.Supp" "pctlowinc"        "pctlingiso"      
#>  [5] "pctunemployed"    "pctlths"          "lowlifex"         "pctunder5"       
#>  [9] "pctover64"        "pctmin"          
 names_pct_as_fraction_ejamit
#>   [1] "Demog.Index"                "Demog.Index.Supp"          
#>   [3] "pctlowinc"                  "pctlingiso"                
#>   [5] "pctunemployed"              "pctlths"                   
#>   [7] "lowlifex"                   "pctunder5"                 
#>   [9] "pctover64"                  "pctmin"                    
#>  [11] "avg.Demog.Index"            "avg.Demog.Index.Supp"      
#>  [13] "avg.pctlowinc"              "avg.pctlingiso"            
#>  [15] "avg.pctunemployed"          "avg.pctlths"               
#>  [17] "avg.lowlifex"               "avg.pctunder5"             
#>  [19] "avg.pctover64"              "avg.pctmin"                
#>  [21] "state.avg.Demog.Index"      "state.avg.Demog.Index.Supp"
#>  [23] "state.avg.pctlowinc"        "state.avg.pctlingiso"      
#>  [25] "state.avg.pctunemployed"    "state.avg.pctlths"         
#>  [27] "state.avg.lowlifex"         "state.avg.pctunder5"       
#>  [29] "state.avg.pctover64"        "state.avg.pctmin"          
#>  [31] "pcthisp"                    "pctnhba"                   
#>  [33] "pctnhaa"                    "pctnhaiana"                
#>  [35] "pctnhnhpia"                 "pctnhotheralone"           
#>  [37] "pctnhmulti"                 "pctnhwa"                   
#>  [39] "avg.pcthisp"                "avg.pctnhba"               
#>  [41] "avg.pctnhaa"                "avg.pctnhaiana"            
#>  [43] "avg.pctnhnhpia"             "avg.pctnhotheralone"       
#>  [45] "avg.pctnhmulti"             "avg.pctnhwa"               
#>  [47] "state.avg.pcthisp"          "state.avg.pctnhba"         
#>  [49] "state.avg.pctnhaa"          "state.avg.pctnhaiana"      
#>  [51] "state.avg.pctnhnhpia"       "state.avg.pctnhotheralone" 
#>  [53] "state.avg.pctnhmulti"       "state.avg.pctnhwa"         
#>  [55] "pcthisp"                    "pctba"                     
#>  [57] "pctaa"                      "pctaiana"                  
#>  [59] "pctnhpia"                   "pctotheralone"             
#>  [61] "pctmulti"                   "pctwa"                     
#>  [63] "avg.pcthisp"                "avg.pctba"                 
#>  [65] "avg.pctaa"                  "avg.pctaiana"              
#>  [67] "avg.pctnhpia"               "avg.pctotheralone"         
#>  [69] "avg.pctmulti"               "avg.pctwa"                 
#>  [71] "state.avg.pcthisp"          "state.avg.pctba"           
#>  [73] "state.avg.pctaa"            "state.avg.pctaiana"        
#>  [75] "state.avg.pctnhpia"         "state.avg.pctotheralone"   
#>  [77] "state.avg.pctmulti"         "state.avg.pctwa"           
#>  [79] "pcthisp"                    "pctnhba"                   
#>  [81] "pctnhaa"                    "pctnhaiana"                
#>  [83] "pctnhnhpia"                 "pctnhotheralone"           
#>  [85] "pctnhmulti"                 "pctnhwa"                   
#>  [87] "avg.pcthisp"                "avg.pctnhba"               
#>  [89] "avg.pctnhaa"                "avg.pctnhaiana"            
#>  [91] "avg.pctnhnhpia"             "avg.pctnhotheralone"       
#>  [93] "avg.pctnhmulti"             "avg.pctnhwa"               
#>  [95] "state.avg.pcthisp"          "state.avg.pctnhba"         
#>  [97] "state.avg.pctnhaa"          "state.avg.pctnhaiana"      
#>  [99] "state.avg.pctnhnhpia"       "state.avg.pctnhotheralone" 
#> [101] "state.avg.pctnhmulti"       "state.avg.pctnhwa"         
#> [103] "pctdisability"              "pctmale"                   
#> [105] "pctfemale"                  "p_own_occupied"            
#> [107] "pctunder18"                 "pctover17"                 
 cat("\n\n")
#> 
#> 
 ytable = data.table(pctlowinc = 1:2, pctpre1960 = 1:2, avg.pctunemployed = 1:2, avg.pctpre1960 = 1:2)
#> Error in data.table(pctlowinc = 1:2, pctpre1960 = 1:2, avg.pctunemployed = 1:2,     avg.pctpre1960 = 1:2): could not find function "data.table"
 fix_pctcols_x100(ytable)
#> Error in fix_pctcols_x100(ytable): could not find function "fix_pctcols_x100"
 fix_pctcols_x100(ytable, names_pct_as_fraction_blockgroupstats) 
#> Error in fix_pctcols_x100(ytable, names_pct_as_fraction_blockgroupstats): could not find function "fix_pctcols_x100"
 fix_pctcols_x100(ytable, names_pct_as_fraction_ejamit)
#> Error in fix_pctcols_x100(ytable, names_pct_as_fraction_ejamit): could not find function "fix_pctcols_x100"
 cat("\n\n")
#> 
#> 
 y
#>   pctlowinc pctpre1960 avg.pctunemployed avg.pctpre1960
#> 1         1          1                 1              1
#> 2         2          2                 2              2
 ytable
#> Error in eval(expr, envir, enclos): object 'ytable' not found