Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites

Use NAICS code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites

Usage

frs_from_naics(naics_code_or_name, childrenForNAICS = TRUE, ...)

Arguments

naics_code_or_name

passed to naics_from_any() as the query

childrenForNAICS

passed to naics_from_any() as the children param of that function

...

passed to naics_from_any()

Value

relevant rows of the data.table called frs, which has column names that are "lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"

Details

The EPA also provides a FRS Facility Industrial Classification Search tool where you can find facilities based on NAICS or SIC.

EPA's ECHO query tools also provide search by NAICS or SIC, and by MACT subpart.

See also

Examples

  frs_from_naics("uranium")
#> 
#> Failed trying to connect to pins board server.
#> 
#> Arrow-format datasets (blocks, etc.) are up-to-date -- locally-installed and latest-released data repository versions match.
#> frs          - is NOT in memory. Checking local disk... frs         is loading from /home/runner/work/_temp/Library/EJAM/data/frs.arrow ...done.
#> frs          - was loaded from local folder, so server copy was not sought.
#> 
#> 
#> Failed trying to connect to pins board server.
#> 
#> Arrow-format datasets (blocks, etc.) are up-to-date -- locally-installed and latest-released data repository versions match.
#> frs_by_naics   - is NOT in memory. Checking local disk... frs_by_naics  is loading from /home/runner/work/_temp/Library/EJAM/data/frs_by_naics.arrow ...done.
#> frs_by_naics   - was loaded from local folder, so server copy was not sought.
#> 
#> Key: <REGISTRY_ID>
#>          lat        lon  REGISTRY_ID
#>        <num>      <num>       <char>
#>  1: 34.58119  -92.51189 110001710890
#>  2: 42.64365 -103.36436 110002358788
#>  3: 28.00390  -82.09650 110002473966
#>  4: 38.40361 -106.29725 110010054092
#>  5: 38.06416 -109.20513 110010134416
#>  6: 38.57326 -109.55075 110010668599
#>  7: 38.36444 -108.73028 110010732332
#>  8: 42.20970 -120.36680 110014208730
#>  9: 47.93129 -118.06677 110018918527
#> 10: 34.46428  -92.94709 110024954334
#> 11: 35.36147 -107.68931 110033148626
#> 12: 27.60451  -98.40847 110033861603
#> 13: 32.98081 -102.82339 110034198491
#> 14: 26.89270  -98.31050 110034202994
#> 15: 28.90778  -98.09889 110034291255
#> 16: 38.31249 -109.30900 110035858765
#> 17: 38.30965 -109.23542 110037102079
#> 18: 38.64200 -109.04900 110037576058
#> 19: 38.08012 -108.82188 110038438125
#> 20: 38.11944 -109.15707 110039201477
#> 21: 37.59833 -110.19306 110046339119
#> 22: 32.71957 -103.33963 110046349572
#> 23: 34.58123  -92.51848 110056511421
#> 24: 41.13629 -104.81906 110058359486
#> 25: 42.33270 -120.52080 110064803075
#> 26: 40.07147  -75.42028 110070326975
#> 27: 37.69743 -122.40613 110070443812
#> 28: 43.04778 -105.70602 110070647328
#> 29: 42.85969 -106.33265 110070984780
#> 30: 41.67111 -107.97139 110070987873
#> 31: 42.85548 -106.33531 110070988026
#> 32: 42.39837 -107.84901 110070989587
#> 33: 43.87184 -106.13106 110070989597
#> 34: 42.84492 -106.23747 110070990432
#> 35: 42.33212 -107.79405 110071007259
#> 36: 41.31203 -105.59451 110071014333
#> 37: 38.36444 -108.73028 110071722424
#>          lat        lon  REGISTRY_ID
#>                                           PRIMARY_NAME
#>                                                 <char>
#>  1:           SAINT-GOBAIN PROPPANTS PROPPANTS PLANT 3
#>  2:                               CROW BUTTE RESOURCES
#>  3:   IMC FERTILIZER INC., PLANT CITY URANIUM RECOVERY
#>  4:             HOMESTAKE MINING COMPANY OF CALIFORNIA
#>  5:                                           RIM MINE
#>  6:                                  ATLAS MINERALS CO
#>  7:                               UMETCO MINERALS CORP
#>  8:                          RENEWABLE RESOURCE SYSTEM
#>  9:                           DAWN MINING MIDNITE MINE
#> 10:                               UMETCO MINERALS CORP
#> 11:                            ROCA HONDA URANIUM MINE
#> 12:      LA PALANGANA URANIUM IN-SITU RECOVERY PROJECT
#> 13:                                RHODE RANCH PROJECT
#> 14:                                  ALTA MESA PROJECT
#> 15:                                  CONQUISTA PROJECT
#> 16: ENERGY FUEL RESOURCES (USA) INC.-ENERGY QUEEN MINE
#> 17:                                       LA SAL MINES
#> 18:             ENERGY FUELS RESOURCES- WHIRLWIND MINE
#> 19:                       DENISON MINES - SUNDAY MINES
#> 20:                                   VELVET WOOD MINE
#> 21:                                       DANEROS MINE
#> 22:           INTERNATIONAL ISOTOPES FLUORINE PRODUCTS
#> 23:                        SAINT-GOBAIN PROPPANTS (LR)
#> 24:                  CAMECO RESOURCES CORPORATE OFFICE
#> 25:            WHITE KING AND LUCKY LASS URANIUM MINES
#> 26:                                      METALLURG,INC
#> 27:                   L&D TRUCK & EQUIPMENT REPAIR INC
#> 28:       POWER RESOURCES INC. D/B/A/ CAMECO RESOURCES
#> 29:                      URANIUM ONE USA INC WILLOW CR
#> 30:                      LOST CREEK ISR, LLC: LOST CRE
#> 31:                      CAMECO RESOURCES: REYNOLDS RA
#> 32:                      US ENERGY CORPORATION: SIMS S
#> 33:                      URANIUM ONE USA, INC.: SEPTIC
#> 34:                      LOST CREEK ISR, LLC: LOST CRE
#> 35:                      KENNECOTT URANIUM COMPANY: SE
#> 36:                       SOLUTION MINING INC.: H SAND
#> 37:                               UMETCO MINERALS CORP
#>                                           PRIMARY_NAME
#>                              NAICS                                SIC
#>                             <char>                             <char>
#>  1: 212291, 212399, 327124, 327992                               3295
#>  2:                 212290, 212291 1081, 1094, 1429, 1499, 4952, 4953
#>  3:                         212291                               1094
#>  4:                 212290, 212291                         1094, 3341
#>  5:                         212291                               1094
#>  6:                         212291                               1094
#>  7:                 212290, 212291                         1094, 4959
#>  8:                         212291                               1094
#>  9:                         212291                               1094
#> 10:                         212291                                   
#> 11:                         212291                               1094
#> 12:                         212291                               1094
#> 13:                         212291                               1094
#> 14:                         212291                               1094
#> 15:                         212291                         1094, 1099
#> 16:                         212291                               1094
#> 17:                         212291                               1094
#> 18:                 212290, 212291                               1094
#> 19:                 212290, 212291                               1094
#> 20:                         212291                               1094
#> 21:                         212291                               1094
#> 22:                         212291                               1094
#> 23:                 212291, 327992                                   
#> 24:                 212290, 212291                                   
#> 25:                         212291                               1094
#> 26:                         212291                                   
#> 27:                         212291                                   
#> 28:                         212291                                   
#> 29:                         212291                                   
#> 30:                         212291                                   
#> 31:                         212291                                   
#> 32:                         212291                                   
#> 33:                         212291                                   
#> 34:                         212291                                   
#> 35:                         212291                                   
#> 36:                         212291                                   
#> 37:                         212291                         1094, 4959
#>                              NAICS                                SIC
#>                                                                                                           PGM_SYS_ACRNMS
#>                                                                                                                   <char>
#>  1:                              AIR:AR0000000512500011, AIRS/AFS:0512500011, EIS:17077111, NPDES:ARR00B810, PDS:6300169
#>  2: AIR:NE0000003104500016, AIRS/AFS:3104500016, NDEQ:63416, NPDES:NE0130613, RCRAINFO:NED000830596, SWIPR:SWIPR10003622
#>  3:                                                                                                 TRIS:33566MCFRTHWY39
#>  4:                                                             AIR:CO0000000810900002, AIRS/AFS:0810900002, EIS:4378211
#>  5:                                                                                                      NPDES:UT0023922
#>  6:                                                                                    ICIS:37092, RCRAINFO:UTD980717607
#>  7:                        AIRS/AFS:0808500032, EIS:3887511, ICIS:37602, NCDB:D08#P-VIII-92-10, NCDB:I08#199109112295  1
#>  8:                                                                                                          OR-DEQ:7004
#>  9:                                                                       ICIS:2586012, NPDES:WA0025721, NPDES:WA0026841
#> 10:                                                                                                          PDS:2600277
#> 11:                                                                                                       NM-TEMPO:27123
#> 12:                                                                                              TX-TCEQ ACR:RN105231872
#> 13:                                                                                              TX-TCEQ ACR:RN101520385
#> 14:                                                                                              TX-TCEQ ACR:RN102360039
#> 15:                                                                                              TX-TCEQ ACR:RN105386866
#> 16:                                                                                                      NPDES:UT0025712
#> 17:                                                                                                      NPDES:UTR000829
#> 18:                                                            AIR:CO0000000807700450, AIRS/AFS:0807700450, EIS:12856611
#> 19:                                                            AIR:CO0000000811300100, AIRS/AFS:0811300100, EIS:12866511
#> 20:                                                                                                      NPDES:UT0025810
#> 21:                                                                                                      NPDES:UTR260661
#> 22:                                                          AIR:NM0000003502500377, AIRS/AFS:3502500377, NM-TEMPO:30806
#> 23:                                                           CEDRI:CEDRI10183385, E-GGRT:1011316, RCRAINFO:ARR000024661
#> 24:                                                                   EIS:16465111, WDEQ:113811, WDEQ:113812, WDEQ:70267
#> 25:                                                         ICIS:23200, ICIS:5869493, OR-DEQ:39986, RADINFO:RAD200000047
#> 26:                                                                                                   OSHA-OIS:338335011
#> 27:                                                                                                RCRAINFO:CAL000045563
#> 28:                                                                                                RCRAINFO:WYR000000356
#> 29:                               WDEQ:107024, WDEQ:107026, WDEQ:107028, WDEQ:114402, WDEQ:71762, WDEQ:71763, WDEQ:71765
#> 30:                                                                                             WDEQ:114451, WDEQ:114805
#> 31:                                                                       WDEQ:70263, WDEQ:70264, WDEQ:70265, WDEQ:70266
#> 32:                                                                                               WDEQ:71719, WDEQ:71720
#> 33:                                                                                                           WDEQ:71766
#> 34:                                                                                               WDEQ:70987, WDEQ:70988
#> 35:                                                                                                           WDEQ:71725
#> 36:                                                                                                           WDEQ:71523
#> 37:                                                                                               AIR:CO0000000808500032
#>                                                                                                           PGM_SYS_ACRNMS
  mapfast(frs_from_naics(naics_from_any("nuclear")$code))

  naics_from_any("silver")
#>      code     n2     n3     n4     n5     n6
#>     <num> <char> <char> <char> <char> <char>
#> 1:  21222     21    212   2122  21222  21222
#> 2: 212222     21    212   2122  21222 212222
#> 3: 332812     33    332   3328  33281 332812
#> 4:  33991     33    339   3399  33991  33991
#> 5: 339910     33    339   3399  33991 339910
#>                                                                                              name
#>                                                                                            <char>
#> 1:                                                                 Gold Ore and Silver Ore Mining
#> 2:                                                                              Silver Ore Mining
#> 3: Metal Coating, Engraving (except Jewelry and Silverware), and Allied Services to Manufacturers
#> 4:                                                           Jewelry and Silverware Manufacturing
#> 5:                                                           Jewelry and Silverware Manufacturing
#>                                                                                                   num_name
#>                                                                                                     <char>
#> 1:                                                                  21222 - Gold Ore and Silver Ore Mining
#> 2:                                                                              212222 - Silver Ore Mining
#> 3: 332812 - Metal Coating, Engraving (except Jewelry and Silverware), and Allied Services to Manufacturers
#> 4:                                                            33991 - Jewelry and Silverware Manufacturing
#> 5:                                                           339910 - Jewelry and Silverware Manufacturing
  naics_from_name("silver")
#> Error in naics_from_name("silver"): could not find function "naics_from_name"
  naics_from_any(212222 )
#>      code     n2     n3     n4     n5     n6              name
#>     <num> <char> <char> <char> <char> <char>            <char>
#> 1: 212222     21    212   2122  21222 212222 Silver Ore Mining
#>                      num_name
#>                        <char>
#> 1: 212222 - Silver Ore Mining
  frs_from_naics(21222)
#> Key: <REGISTRY_ID>
#>           lat       lon  REGISTRY_ID
#>         <num>     <num>       <char>
#>   1: 39.66667 -119.4167 110000472998
#>   2: 32.96667 -114.6417 110000479125
#>   3: 46.33400 -111.4300 110000595213
#>   4: 39.67715 -115.5249 110000601082
#>   5: 40.65334 -117.2236 110000607353
#>  ---                                
#> 317: 64.48490 -165.3487 110071514683
#> 318: 65.21900 -150.0701 110071515323
#> 319: 43.73127 -115.9396 110071643460
#> 320: 40.94694 -116.4447 110071711435
#> 321: 45.53239 -116.2836 110071816137
#>                                          PRIMARY_NAME  NAICS        SIC
#>                                                <char> <char>     <char>
#>   1:             ALTA GOLD COMPANY OLINGHOUSE PROJECT 212221       1041
#>   2:                              GLAMIS PICACHO MINE 212221       1041
#>   3:                          DIAMOND HILL MINING INC 212221 1041, 1044
#>   4:                                      YANKEE MINE 212221       1041
#>   5:          NEWMONT MINING CORP TRENTON CANYON MINE 212221       1041
#>  ---                                                                   
#> 317: CORDELIA DUNE MINING NORTON SOUND SUCTION DREDGE 212221       1041
#> 318:                   GRAFF DEADWOOD CREEK MINE SITE 212221       1041
#> 319:                                   DANIEL J WOODS 212221       1041
#> 320:                               BOULDER VALLEY WTF 212221       1041
#> 321:                               MCKINLEY MINE, INC 212221       1041
#>                                           PGM_SYS_ACRNMS
#>                                                   <char>
#>   1:                                TRIS:89442LTGLD6000S
#>   2:                                TRIS:92283GLMSP3481P
#>   3:           AIRS/AFS:3000700006, TRIS:59644DMNDH8MILE
#>   4:                                TRIS:89803YNKMN95MIL
#>   5: ICIS:6357299, NPDES:NVU000738, TRIS:89438NWMNTEIGHT
#>  ---                                                    
#> 317:                                     NPDES:AKG371675
#> 318:                                     NPDES:AKG370D7P
#> 319:                                     NPDES:IDU100016
#> 320:                                     NPDES:NV0022675
#> 321:                                     NPDES:IDR05J10Z
  regid_from_naics(21222)
#>   [1] "110036140306" "110000472998" "110000479125" "110000595213" "110000601082"
#>   [6] "110000607353" "110000608334" "110000868106" "110001075149" "110001128208"
#>  [11] "110001157872" "110001971778" "110002041610" "110004942108" "110006852478"
#>  [16] "110007013739" "110007978581" "110007980569" "110007980809" "110008062470"
#>  [21] "110008283696" "110009037069" "110009037103" "110009306535" "110009306562"
#>  [26] "110009306571" "110009343656" "110009816878" "110010114303" "110010340149"
#>  [31] "110010952308" "110013669075" "110013798480" "110013798569" "110013798685"
#>  [36] "110013799014" "110014075981" "110014083106" "110014083124" "110014083179"
#>  [41] "110014083213" "110014083231" "110014092640" "110014115019" "110014146129"
#>  [46] "110014149019" "110014149457" "110014151201" "110014151247" "110014151265"
#>  [51] "110014152095" "110014152148" "110014152807" "110014157526" "110014157553"
#>  [56] "110014169620" "110014169639" "110014169737" "110014169764" "110014169817"
#>  [61] "110014169844" "110014192523" "110014286726" "110014298955" "110014299721"
#>  [66] "110014300041" "110014300201" "110014300461" "110014301380" "110014302851"
#>  [71] "110014303137" "110014377021" "110015753905" "110015784392" "110015844176"
#>  [76] "110015944282" "110016621761" "110016622797" "110016622877" "110016622886"
#>  [81] "110016622895" "110016622902" "110016622920" "110016622939" "110016622948"
#>  [86] "110016623572" "110016623830" "110016624321" "110018499828" "110018500139"
#>  [91] "110018884670" "110020221909" "110020222631" "110021023667" "110022285034"
#>  [96] "110024568751" "110032827145" "110036140501" "110037094792" "110037246245"
#> [101] "110037263181" "110037430973" "110037591193" "110037703278" "110037703287"
#> [106] "110037703456" "110037704847" "110037707096" "110037719430" "110037725183"
#> [111] "110037729928" "110037737884" "110037738311" "110037740335" "110037745866"
#> [116] "110037750519" "110037755952" "110037761008" "110037782138" "110037786241"
#> [121] "110037786919" "110037790030" "110037795703" "110037800369" "110037801126"
#> [126] "110037802358" "110037808842" "110037811142" "110037811972" "110037817468"
#> [131] "110037819652" "110037820962" "110037828018" "110037943143" "110037966261"
#> [136] "110038050916" "110038798637" "110039020467" "110039396800" "110039598600"
#> [141] "110040054204" "110040054311" "110040064596" "110040071926" "110040078803"
#> [146] "110040078812" "110040132825" "110040482625" "110040488022" "110040863517"
#> [151] "110040944903" "110041028134" "110041243856" "110041258420" "110041360470"
#> [156] "110041618666" "110041905016" "110041971337" "110043210325" "110043210780"
#> [161] "110043211636" "110043697005" "110043697078" "110044227592" "110044258969"
#> [166] "110044797547" "110045554066" "110046301248" "110046338673" "110046386530"
#> [171] "110046530367" "110054080766" "110054111279" "110054121428" "110054121446"
#> [176] "110054121455" "110054122971" "110055370673" "110055487977" "110055522331"
#> [181] "110055625756" "110055625765" "110055625774" "110055625783" "110055625792"
#> [186] "110055625809" "110055625818" "110055625827" "110055625836" "110055625845"
#> [191] "110055625854" "110055691087" "110058298854" "110058300565" "110058300814"
#> [196] "110058931250" "110059724375" "110059794085" "110059814820" "110059815464"
#> [201] "110059821849" "110059832873" "110059838216" "110059841471" "110059845333"
#> [206] "110060358839" "110060358848" "110060358857" "110060358866" "110060358875"
#> [211] "110060361558" "110060361567" "110062857582" "110063914572" "110063989938"
#> [216] "110064031603" "110064031658" "110064415173" "110064415208" "110064415397"
#> [221] "110064415976" "110064416047" "110064529032" "110064533875" "110064589253"
#> [226] "110064694987" "110064875719" "110065401559" "110066308187" "110067448275"
#> [231] "110067560730" "110069227232" "110069346700" "110070033868" "110070098032"
#> [236] "110070101607" "110070151883" "110070222908" "110070244799" "110070262419"
#> [241] "110070262434" "110070262486" "110070262501" "110070262577" "110070262647"
#> [246] "110070289914" "110070294934" "110070321492" "110070380058" "110070380096"
#> [251] "110070380229" "110070384063" "110070387800" "110070393688" "110070502220"
#> [256] "110070508229" "110070523617" "110070524753" "110070537859" "110070557695"
#> [261] "110070566037" "110070566209" "110070566210" "110070566211" "110070566830"
#> [266] "110070597032" "110070599053" "110070814159" "110070830519" "110070862515"
#> [271] "110070909250" "110070916461" "110071074582" "110071094637" "110071159263"
#> [276] "110071159266" "110071161145" "110071203367" "110071239289" "110071362194"
#> [281] "110071439431" "110071442613" "110071444026" "110071444199" "110071445479"
#> [286] "110071445965" "110071446259" "110071448022" "110071499928" "110071500329"
#> [291] "110071500336" "110071501090" "110071501817" "110071503144" "110071504205"
#> [296] "110071505849" "110071506904" "110071509129" "110071513134" "110071514683"
#> [301] "110071515323" "110071643460" "110071711435" "110071816137" "110003339986"
#> [306] "110006026790" "110007516018" "110011682731" "110020303633" "110035210232"
#> [311] "110039612461" "110043214688" "110055042234" "110055487904" "110055487913"
#> [316] "110055487922" "110059869520" "110067027130" "110070376780" "110070385158"
#> [321] "110071284349"
  latlon_from_naics(21222)
#>           lat       lon  REGISTRY_ID  NAICS naics_found naics_query
#>         <num>     <num>       <char>  <num>       <num>       <num>
#>   1: 48.88601 -118.0281 110036140306  21222       21222       21222
#>   2: 39.66667 -119.4167 110000472998 212221      212221       21222
#>   3: 32.96667 -114.6417 110000479125 212221      212221       21222
#>   4: 46.33400 -111.4300 110000595213 212221      212221       21222
#>   5: 39.67715 -115.5249 110000601082 212221      212221       21222
#>  ---                                                               
#> 339: 29.81083 -104.3158 110070376780 212222      212222       21222
#> 340: 47.47662 -115.9432 110070385158 212222      212222       21222
#> 341: 39.97276 -105.5617 110070909250 212222      212222       21222
#> 342: 58.37149 -134.5975 110071159263 212222      212222       21222
#> 343: 37.97546 -107.7476 110071284349 212222      212222       21222