Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites

Use SIC code or industry title text search to see FRS Facility Registry Service data on those EPA-regulated sites

Usage

frs_from_sic(sic_code_or_name, ...)

Arguments

sic_code_or_name

passed to sic_from_any()

...

passed to sic_from_any()

Value

relevant rows of the data.table called frs, which has column names that are "lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "SIC" "PGM_SYS_ACRNMS"

The EPA also provides a FRS Facility Industrial Classification Search tool where you can find facilities based on NAICS or SIC.

See also

Examples

  frs_from_sic("glass")
#> 
#> Failed trying to connect to pins board server.
#> 
#> Arrow-format datasets (blocks, etc.) are up-to-date -- locally-installed and latest-released data repository versions match.
#> frs_by_sic   - is NOT in memory. Checking local disk... frs_by_sic  is loading from /home/runner/work/_temp/Library/EJAM/data/frs_by_sic.arrow ...done.
#> frs_by_sic   - was loaded from local folder, so server copy was not sought.
#> 
#> Key: <REGISTRY_ID>
#>           lat        lon  REGISTRY_ID
#>         <num>      <num>       <char>
#>   1: 39.39124  -75.03496 110000321553
#>   2: 37.39273  -76.80001 110000341657
#>   3: 39.03713  -80.45324 110000344887
#>   4: 35.75525  -80.28116 110000346082
#>   5: 39.57878  -82.95292 110000381925
#>  ---                                 
#> 456: 39.71799 -121.82084 110071762235
#> 457: 33.99720 -118.21720 110071772456
#> 458: 33.88343 -118.21739 110071786937
#> 459: 36.77210 -119.68650 110071804275
#> 460: 39.96806  -85.99816 110071816563
#>                                    PRIMARY_NAME  NAICS        SIC
#>                                          <char> <char>     <char>
#>   1:         BALL-FOSTER GLASS CONTAINER CO LLC 327213       3221
#>   2: OWENS-BROCKWAY GLASS CONTAINER INC - TOANO 327213       3221
#>   3:                         GLASS WORKS WV LLC 327212       3229
#>   4:          ELECTRIC GLASS FIBER AMERICA, LLC 327212 3229, 9999
#>   5:                                THOMSON INC 327212       3229
#>  ---                                                             
#> 456:                   ORIENT & FLUME ART GLASS 327212       3229
#> 457:         OWENS-BROCKWAY GLASS CONTAINER INC 327213       3221
#> 458:                PERFORMANCE COMPOSITES, INC 325211       3229
#> 459:                           FRESNO NEON SIGN 327212       3229
#> 460:                            STEVANATO GROUP              3221
#>                                                                                                                                                                                                                                   PGM_SYS_ACRNMS
#>                                                                                                                                                                                                                                           <char>
#>   1:                                                                                                                                                                                                                        TRIS:08332MRCNN328SO
#>   2:                                                                          AIR:VA0000005109500022, AIRS/AFS:5109500022, CEDRI:CEDRI10171441, CEDS:200000091498, NPDES:VAR050374, RCRAINFO:VAD000765503, TRIS:23168WNSLL150IN, TSCA:TSCA123931
#>   3:                                                                                                                                                              ICIS:25572, NCDB:D03#EPCRA-062, NCDB:I03#19910312A3003 1, TRIS:26452LGLSS631EA
#>   4: AIR:NC0000003702900109, AIRS/AFS:3705700109, BR:NCD043712298, CEDRI:CEDRI10045193, E-GGRT:1002322, EIA-860:58046, EIS:7377911, EPS:EPS10139559, NC-FITS:3325, NPDES:NC0004626, NPDES:NCG070030, RCRAINFO:NCD043712298, TRIS:27292PPGNDNEWJE
#>   5:                                                                                                                                                                                  OH-CORE:34652, RCRAINFO:OHD043736644, TRIS:43113THMSN24200
#>  ---                                                                                                                                                                                                                                            
#> 456:                                                                                                                                                                                                               CARB-TCH:4SVBUT30, EIS:875311
#> 457:                                                                                                                                                                                                            CARB-TCH:19SCSC7427, EIS:5797211
#> 458:                                                                                                                                                                                                         CARB-TCH:19SCSC140552, EIS:10316411
#> 459:                                                                                                                                                                                                         CARB-TCH:10SJVSJU2207, EIS:10377811
#> 460:                                                                                                                                                                                                                             NPDES:INP000733
  mapfast(frs_from_sic(sic_from_any("silver")$code))

  sic_from_any("silver")
#>      code     n2     n3     n4                                 name
#>    <char> <char> <char> <char>                               <char>
#> 1:   1040     10    104   1040                 Gold and silver ores
#> 2:   1044     10    104   1044                          Silver ores
#> 3:   3910     39    391   3910 Jewelry, silverware, and plated ware
#> 4:   3914     39    391   3914           Silverware and plated ware
#>                                       num_name
#>                                         <char>
#> 1:                 1040 - Gold and silver ores
#> 2:                          1044 - Silver ores
#> 3: 3910 - Jewelry, silverware, and plated ware
#> 4:           3914 - Silverware and plated ware
  sic_from_name("silver")
#>      code     n2     n3     n4                                 name
#>    <char> <char> <char> <char>                               <char>
#> 1:   1040     10    104   1040                 Gold and silver ores
#> 2:   1044     10    104   1044                          Silver ores
#> 3:   3910     39    391   3910 Jewelry, silverware, and plated ware
#> 4:   3914     39    391   3914           Silverware and plated ware
#>                                       num_name
#>                                         <char>
#> 1:                 1040 - Gold and silver ores
#> 2:                          1044 - Silver ores
#> 3: 3910 - Jewelry, silverware, and plated ware
#> 4:           3914 - Silverware and plated ware
  sic_from_any('0780')
#>      code     n2     n3     n4                                 name
#>    <char> <char> <char> <char>                               <char>
#> 1:   0780     07    078   0780 Landscape and horticultural services
#>                                       num_name
#>                                         <char>
#> 1: 0780 - Landscape and horticultural services
  frs_from_sic('0780')
#> Key: <REGISTRY_ID>
#>         lat        lon  REGISTRY_ID                   PRIMARY_NAME  NAICS
#>       <num>      <num>       <char>                         <char> <char>
#> 1: 34.01928 -118.06651 110002942059    WHITTIER FERTILIZER COMPANY       
#> 2: 39.46249  -75.31125 110032015067                 ARC GREENHOUSE       
#> 3: 40.56092  -74.55954 110032369568      BRIDGEWATER TRANSPORT INC       
#> 4: 39.76020  -75.38356 110032373036 GREEN ACRES LAWNS & LANDSCAPES       
#> 5: 38.51118 -121.35022 110065267570      NATURE CARE LANDSCAPE INC       
#> 6: 40.15182 -124.08750 110065281572        A123 LAWN AND LANDSCAPE       
#>                 SIC
#>              <char>
#> 1:             0780
#> 2:             0780
#> 3: 0780, 0782, 5032
#> 4:       0780, 0782
#> 5:             0780
#> 6:             0780
#>                                                                              PGM_SYS_ACRNMS
#>                                                                                      <char>
#> 1: CA-CERS:10281400, CA-ENVIROVIEW:85922, HWTS-DATAMART:CAR000080473, RCRAINFO:CAR000080473
#> 2:                                                                          NJ-NJEMS:148036
#> 3:                                                                            NJ-NJEMS:1781
#> 4:                                                                          NJ-NJEMS:197107
#> 5:                                                                      CA-ENVIROVIEW:51231
#> 6:                                               CA-ENVIROVIEW:89083, RCRAINFO:CAL000363297
  regid_from_sic('0780')
#> Key: <SIC, REGISTRY_ID>
#>         lat        lon  REGISTRY_ID    SIC
#>       <num>      <num>       <char> <char>
#> 1: 34.01928 -118.06651 110002942059   0780
#> 2: 39.46249  -75.31125 110032015067   0780
#> 3: 40.56092  -74.55954 110032369568   0780
#> 4: 39.76020  -75.38356 110032373036   0780
#> 5: 38.51118 -121.35022 110065267570   0780
#> 6: 40.15182 -124.08750 110065281572   0780
  latlon_from_sic('0780')
#> Key: <SIC, REGISTRY_ID>
#>         lat        lon  REGISTRY_ID    SIC
#>       <num>      <num>       <char> <char>
#> 1: 34.01928 -118.06651 110002942059   0780
#> 2: 39.46249  -75.31125 110032015067   0780
#> 3: 40.56092  -74.55954 110032369568   0780
#> 4: 39.76020  -75.38356 110032373036   0780
#> 5: 38.51118 -121.35022 110065267570   0780
#> 6: 40.15182 -124.08750 110065281572   0780