utility - How many blocks and many other stats about blocks and sites

utility - How many blocks and many other stats about blocks and sites

getblocks_diagnostics(
  x,
  detailed = FALSE,
  see_pctiles = FALSE,
  see_distanceplot = FALSE
)

Arguments

x

The output of getblocksnearby() like testoutput_getblocksnearby_10pts_1miles

detailed

if TRUE, also shows in console a long table of frequencies via getblocks_summarize_blocks_per_site()

see_pctiles

set to TRUE to see 20 percentiles of distance in a table

see_distanceplot

if TRUE, also draws scatter plot of adjusted vs unadj distances

Value

A list of stats

See also

Examples

  getblocks_diagnostics(testoutput_getblocksnearby_10pts_1miles)
#> 
#>    Summary stats on distances reported from any sites to any nearby blocks
#> 
#> 0.9998549 miles is max. distance to block internal point (distance_unadjusted)   
#> 0.9998549 miles is max. distance to average resident in block (distance reported)   
#> 0.008486795 miles is shortest distance to block internal point (distance_unadjusted)   
#> 0.018 miles is shortest distance to average resident in block (distance reported)   
#> 7 block distances were adjusted (these stats may count some blocks twice if adjusted at 2+ sites)
#>   6 block distances were adjusted up (reported dist to avg resident is > dist to block internal point)
#>   1 block distances were adjusted down (reported < unadjusted)
#> 6 unique sites had one or more block distances adjusted due to large block and short distance to block point
#> 
#> 9 unique output sites
#> 128 blocks are near the avg site or in avg buffer
#>              (based on their block internal point, like a centroid)
#> 1,152 blocks including doublecounting in overlaps, 
#>              in final row count (block-to-site pairs table)
#> 1,152 actual unique blocks total
#> 1,152 blocks (and their residents) have exactly 1 site nearby 
#> 0 blocks (and their residents) have exactly 2 sites nearby 
#> 0 blocks (and their residents) have exactly 3 sites nearby 
#> 1 is ratio of blocks including multicounting / actual count of unique blocks
#> 0% of unique blocks could get counted more than once 
#>              because those residents are near two or more sites 
#>              (assuming they live at the block internal point
  # library(data.table)
  x <- data.table::copy(testpoints_10)
  setDT(x)
#> Error in setDT(x): could not find function "setDT"
  pts <- rbind(data.table(lat = 40.3, lon = -96.23),
    x[ , .(lat, lon)])
#> Error in data.table(lat = 40.3, lon = -96.23): could not find function "data.table"
 z <- getblocksnearbyviaQuadTree(pts, 1, quadtree = localtree, quiet = T)
#> Error in is.data.frame(sitepoints): object 'pts' not found
 z[ , .(blocks = .N) , keyby = 'ejam_uniq_id']
#> Error in eval(expr, envir, enclos): object 'z' not found
 plotblocksnearby(pts, radius = 1, sites2blocks = z)
#> Error in c("blockid", "distance", siteidvarname) %in% names(sitepoints): object 'pts' not found
 zz <- getblocks_diagnostics(z, detailed = T, see_pctiles = T)
#> Error in NROW(x): object 'z' not found
cbind(stats = zz)
#> Error in cbind(stats = zz): object 'zz' not found

  getblocks_diagostics(testoutput_getblocksnearby_1000pts_1miles, see_distanceplot = TRUE)
#> Error in getblocks_diagostics(testoutput_getblocksnearby_1000pts_1miles,     see_distanceplot = TRUE): could not find function "getblocks_diagostics"