Find all blocks within each of the FIPS codes provided

Allows EJAM to analyze and compare Counties, for example

getblocksnearby_from_fips(fips, inshiny = FALSE, need_blockwt = TRUE)

Arguments

fips

vector of FIPS codes identifying blockgroups, tracts, counties, or states. This is useful if -- instead of gettings stats on and comparing circular buffers or polygons -- one will be getting stats on one or more tracts, or analyzing and comparing blockgroups in a county, or comparing whole counties to each other, within a State.

inshiny

used by shiny app server code to handle errors via validate() instead of stop()

need_blockwt

set to FALSE to speed it up if you do not need blockwt

Value

same as for getblocksnearby but one row per FIPS, and the distance column is irrelevant

See also

Examples

  x <- getblocksnearby_from_fips(fips_counties_from_state_abbrev("DE"))
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> blockid2fips   - is NOT in memory. Checking local disk... blockid2fips  is NOT found locally on disk at ~/../Downloads/EJAMbigfiles/blockid2fips.arrow 
#> blockid2fips   - is NOT in local folder.
#> blockid2fips  - was read from pins board server, into memory (to specified envir).
#> 
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> bgid2fips    - is NOT in memory. Checking local disk... bgid2fips   is NOT found locally on disk at ~/../Downloads/EJAMbigfiles/bgid2fips.arrow 
#> bgid2fips    - is NOT in local folder.
#> bgid2fips   - was read from pins board server, into memory (to specified envir).
#> 
#> Joining with `by = join_by(blockid)`
  counties_ej <- doaggregate(x)
  #cannot use mapfast(counties_ej$results_bysite) since no lat lon.  mapfastej_counties() should work...
  y =  ejamit(fips=fips_counties_from_statename("Delaware"))
#> Finding blocks in each FIPS Census unit.
#> Joining with `by = join_by(blockid)`
#> Aggregating at each FIPS Census unit and overall.
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
  
  # x=getblocksnearby_from_fips("482011000011") # one blockgroup only
  # y=doaggregate(x)