Find all blocks within each of the FIPS codes provided

Allows EJAM to analyze and compare Counties, for example

Usage

getblocksnearby_from_fips2(fips, inshiny = FALSE, need_blockwt = TRUE)

Arguments

fips

vector of FIPS codes identifying blockgroups, tracts, counties, or states. This is useful if – instead of gettings stats on and comparing circular buffers or polygons – one will be getting stats on one or more tracts, or analyzing and comparing blockgroups in a county, or comparing whole counties to each other, within a State.

inshiny

used by shiny app server code to handle errors via validate() instead of stop()

need_blockwt

ignored now

Value

same as for getblocksnearby but one row per FIPS, and the distance column is irrelevant

See also

Examples

  x <- getblocksnearby_from_fips(fips_counties_from_state_abbrev("DE"))
  counties_ej <- doaggregate(x)
  #cannot use mapfast(counties_ej$results_bysite) since no lat lon.  mapfastej_counties() should work...
  y =  ejamit(fips=fips_counties_from_statename("Delaware"))
  
  # x=getblocksnearby_from_fips("482011000011") # one blockgroup only
  # y=doaggregate(x)