DRAFT / WAS WORK IN PROGRESS - find nearby blocks using Quad Tree data structure for speed, CLUSTERED FOR PARALLEL PROCESSING

Uses packages parallel and snow. parallel::makePSOCKcluster is an enhanced version of snow::makeSOCKcluster in package snow. It runs Rscript on the specified host(s) to set up a worker process which listens on a socket for expressions to evaluate, and returns the results (as serialized objects).

Usage

getblocksnearbyviaQuadTree_Clustered(
  sitepoints,
  radius,
  maxradius,
  avoidorphans,
  CountCPU = 1,
  quadtree
)

Arguments

sitepoints

data.table with columns LAT, LONG

radius

in miles

maxradius

miles distance

avoidorphans

logical

CountCPU

for parallel processing via makeCluster() and doSNOW::registerDoSNOW()

quadtree

index of all US blocks like localtree

Details

For all examples, see getblocksnearbyviaQuadTree()

Uses indexgridsize and quaddata variables that come from global environment (but should pass to this function rather than assume in global env?)

See also