Get lat/lon flexibly - from file, data.frame, data.table, or lat/lon vectors

Try to figure out if user provided latitude / longitude as vectors, data.frame, file, or interactively pick file.

Usage

latlon_from_anything(x, y)

Arguments

x

If missing and interactive mode in RStudio, prompts user for file. Otherwise, this can be a filename (csv or xlsx, with path), or data.frame/ data.table/ matrix, or vector of longitudes (in which case y must be the latitudes). Note that even though it is called latlon_etc the lon is x and comes before the lat among parameters x,y (unlike in most other functions here using lat,lon) File or data.frame/data.table/matrix must have columns called lat and lon, or names that can be inferred to be that by latlon_infer()

y

If x is a vector of longitudes, y must be the latitudes. Ignored otherwise.

Value

A data.frame that has at least columns lon and lat (and others if they were in x)

Details

This function relies on

read_csv_or_xl() and

latlon_df_clean() which in turn uses latlon_infer() latlon_as.numeric() latlon_is.valid()

read_and_clean_points() from EJAMejscreenapi

would be the most general / flexible broadest way to get points, but is still work in progress

is similar to what is done by latlon_from_anything()

except it also uses these functions:

latlon_from_regid()

latlon_from_programid() but not _from_naics() ?

See also

Examples

  if (FALSE) {
 if (interactive()) {
 pts <- latlon_from_anything()
 }}
 latlon_from_anything(testpoints_100[1:6,] )
#>        lat        lon sitenumber       sitename valid
#> 1 40.66478  -74.89137          1 Example Site 1  TRUE
#> 2 38.39657  -98.70086          2 Example Site 2  TRUE
#> 3 40.34971  -79.84964          3 Example Site 3  TRUE
#> 4 34.23836 -118.25343          4 Example Site 4  TRUE
#> 5 41.35163  -85.13566          5 Example Site 5  TRUE
#> 6 42.21061  -93.21535          6 Example Site 6  TRUE
 latlon_from_anything(testpoints_100[1:6, c('lat','lon')] )
#>        lat        lon valid
#> 1 40.66478  -74.89137  TRUE
#> 2 38.39657  -98.70086  TRUE
#> 3 40.34971  -79.84964  TRUE
#> 4 34.23836 -118.25343  TRUE
#> 5 41.35163  -85.13566  TRUE
#> 6 42.21061  -93.21535  TRUE
 latlon_from_anything(x=testpoints_100$lon[1:6], y=testpoints_100$lat[1:6] )
#>          lon      lat valid
#> 1  -74.89137 40.66478  TRUE
#> 2  -98.70086 38.39657  TRUE
#> 3  -79.84964 40.34971  TRUE
#> 4 -118.25343 34.23836  TRUE
#> 5  -85.13566 41.35163  TRUE
#> 6  -93.21535 42.21061  TRUE