Get point locations for US EPA-regulated facilities that have sources subject to Maximum Achievable Control Technology (MACT) standards under the Clean Air Act.

Get point locations for US EPA-regulated facilities that have sources subject to Maximum Achievable Control Technology (MACT) standards under the Clean Air Act.

Usage

latlon_from_mactsubpart(subpart = "JJJ", include_if_no_latlon = FALSE)

Arguments

subpart

vector of one or more strings indicating the Subpart of CFR Title 40 Part 63 that covers the source category of interest, such as "FFFF" - see for example, https://www.ecfr.gov/current/title-40/part-63/subpart-FFFF

include_if_no_latlon

logical - many in the database lack lat lon values but have a MACT code

Value

a data.table with columns named

programid, subpart, title, lat, lon, REGISTRY_ID, program

for US EPA FRS sites with that MACT code. Or NA if none found.

Details

See https://www.epa.gov/stationary-sources-air-pollution/national-emission-standards-hazardous-air-pollutants-neshap-8

Examples

  mact_table
#>     subpart
#> 1         A
#> 2        AA
#> 3      AAAA
#> 4     AAAAA
#> 5   AAAAAAA
#> 6         B
#> 7        BB
#> 8     BBBBB
#> 9    BBBBBB
#> 10  BBBBBBB
#> 11        C
#> 12       CC
#> 13      CCC
#> 14     CCCC
#> 15    CCCCC
#> 16   CCCCCC
#> 17  CCCCCCC
#> 18        D
#> 19       DD
#> 20      DDD
#> 21     DDDD
#> 22    DDDDD
#> 23   DDDDDD
#> 24  DDDDDDD
#> 25        E
#> 26       EE
#> 27      EEE
#> 28     EEEE
#> 29    EEEEE
#> 30  EEEEEEE
#> 31        F
#> 32     FFFF
#> 33    FFFFF
#> 34   FFFFFF
#> 35        G
#> 36       GG
#> 37      GGG
#> 38     GGGG
#> 39    GGGGG
#> 40        H
#> 41       HH
#> 42      HHH
#> 43     HHHH
#> 44    HHHHH
#> 45   HHHHHH
#> 46  HHHHHHH
#> 47        I
#> 48       II
#> 49      III
#> 50     IIII
#> 51    IIIII
#> 52        J
#> 53       JJ
#> 54      JJJ
#> 55     JJJJ
#> 56    JJJJJ
#> 57   JJJJJJ
#> 58       KK
#> 59     KKKK
#> 60    KKKKK
#> 61        L
#> 62       LL
#> 63      LLL
#> 64    LLLLL
#> 65   LLLLLL
#> 66        M
#> 67       MM
#> 68      MMM
#> 69     MMMM
#> 70    MMMMM
#> 71   MMMMMM
#> 72        N
#> 73      NNN
#> 74     NNNN
#> 75    NNNNN
#> 76   NNNNNN
#> 77        O
#> 78       OO
#> 79      OOO
#> 80     OOOO
#> 81   OOOOOO
#> 82       PP
#> 83      PPP
#> 84     PPPP
#> 85    PPPPP
#> 86   PPPPPP
#> 87        Q
#> 88       QQ
#> 89     QQQQ
#> 90    QQQQQ
#> 91   QQQQQQ
#> 92        R
#> 93       RR
#> 94      RRR
#> 95     RRRR
#> 96    RRRRR
#> 97   RRRRRR
#> 98        S
#> 99       SS
#> 100    SSSS
#> 101   SSSSS
#> 102  SSSSSS
#> 103       T
#> 104      TT
#> 105     TTT
#> 106    TTTT
#> 107   TTTTT
#> 108  TTTTTT
#> 109       U
#> 110      UU
#> 111     UUU
#> 112    UUUU
#> 113   UUUUU
#> 114      VV
#> 115     VVV
#> 116    VVVV
#> 117  VVVVVV
#> 118       W
#> 119      WW
#> 120    WWWW
#> 121   WWWWW
#> 122  WWWWWW
#> 123       X
#> 124      XX
#> 125     XXX
#> 126    XXXX
#> 127  XXXXXX
#> 128       Y
#> 129      YY
#> 130    YYYY
#> 131   YYYYY
#> 132  YYYYYY
#> 133     ZZZ
#> 134    ZZZZ
#> 135   ZZZZZ
#> 136  ZZZZZZ
#>                                                                            title
#> 1                                                             GENERAL PROVISIONS
#> 2                                            PHOSPERIC ACID MANUFACTURING PLANTS
#> 3                                                MUNICIPAL SOLID WASTE LANDFILLS
#> 4                                                      LIME MANUFACTURING PLANTS
#> 5              AREA SOURCE: ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 6                    REQ FOR CONTROL TECHNOLOGY DETERMINATIONS FOR MAJOR SOURCES
#> 7                                                          PHOSPHATE FERTILIZERS
#> 8                                                    SEMICONDUCTOR MANUFACTURING
#> 9     GASOLINE DISTRIBUTION BULK TERMINALS, BULK PLANTS, AND PIPELINE FACILITIES
#> 10                                   AREA SOURCE: CHEMICAL PREPARATIONS INDUSTRY
#> 11                                                            DELISTINGS OF HAPS
#> 12                                                          PETROLEUM REFINERIES
#> 13                      STEEL PICKLING-HCI PROCESS FACILITIES AND HCI REG PLANTS
#> 14                                            MANUFACTURING OF NUTRITIONAL YEAST
#> 15                             COKE OVENS: PUSHING, QUENCHING AND BATTERY STACKS
#> 16                                                GASOLINE DISPENSING FACILITIES
#> 17                          AREA SOURCE: PAINT AND ALLIED PRODUCTS MANUFACTURING
#> 18                                       COMPLIANCE EXTENSION, EARLY HAP REDUCTN
#> 19                                        OFF-SITE WASTE AND RECOVERY OPERATIONS
#> 20                                                       MINERAL WOOL PRODUCTION
#> 21                                           PLYWOOD AND COMPOSITE WOOD PRODUCTS
#> 22   MAJOR SOURCES: INDUSTRIAL/COMMERCIAL/INSTITUTIONAL BOILERS & PROCESS HEATER
#> 23                     POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION AREA SOURCES
#> 24                                               AREA SOURCE: PREPARED FEEDS MFG
#> 25                         APPROVAL OF STATE PROGRAMS &  DELEGATION OF AUTHORITY
#> 26                                                 MAGNETIC TAPE MFG. OPERATIONS
#> 27                                                    HAZARDOUS WASTE COMBUSTORS
#> 28                                   ORGANIC LIQUIDS DISTRIBUTION (NON-GASOLINE)
#> 29                                                      IRON AND STEEL FOUNDRIES
#> 30                          AREA SOURCE: GOLD MINE ORE PROCESSING AND PRODUCTION
#> 31                                SYNTHETIC HAZARDOUS ORGANICS (SOCMI) STANDARDS
#> 32                            MISCELLANEOUS ORGANIC CHEMICAL MANUFACTURING (MON)
#> 33                            INTEGRATED IRON AND STEEL MANUFACTURING FACILITIES
#> 34                                        SECONDARY COPPER SMELTING AREA SOURCES
#> 35         SOCMI PROCESS VENTS, STORAGE VESSELS, TRANSFER OPERATIONS, WASTEWATER
#> 36                                                       AEROSPACE MFG. & REWORK
#> 37                                                  PHARMACEUTICAL MANUFACTURING
#> 38                               SOLVENT EXTRACTION FOR VEGETABLE OIL PRODUCTION
#> 39                                                              SITE REMEDIATION
#> 40                                         EQUIPMENT LEAKS OF HAZARDOUS ORGANICS
#> 41                                     OIL AND NATURAL GAS PRODUCTION FACILITIES
#> 42                                 NATURAL GAS TRANSMISSION & STORAGE FACILITIES
#> 43                                          WET FORMED FIBERGLASS MAT PRODUCTION
#> 44                                           MISCELLANEOUS COATING MANUFACTURING
#> 45                 PAINT STRIP & MISC SURFACE COATING OPERATIONS AT AREA SOURCES
#> 46                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 47                                      NEGOTIATED REGS: EQUIP LEAKS, HAZ-ORGNCS
#> 48                                                    SHIPBUILDING & SHIP REPAIR
#> 49                                         FLEXIBLE POLYURETHANE FOAM PRODUCTION
#> 50                                 SURFACE COATING OF AUTO AND LIGHT DUTY TRUCKS
#> 51                       MERCURY EMISSIONS FROM MERCURY CELL CHLOR-ALKALI PLANTS
#> 52                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 53                                       WOOD FURNITURE MANUFACTURING OPERATIONS
#> 54                                                  GROUP IV POLYMERS AND RESINS
#> 55                                                    PAPER & OTHER WEB COATINGS
#> 56                                BRICK & STRUCTURAL CLAY PRODUCTS MANUFACTURING
#> 57                INDUSTRIAL, COMMERCIAL, AND INSTITUTIONAL BOILERS AREA SOURCES
#> 58                                                         PRINTING & PUBLISHING
#> 59                                                 SURFACE COATING OF METAL CANS
#> 60                                                   CLAY CERAMICS MANUFACTURING
#> 61                                                           COKE OVEN BATTERIES
#> 62                                             PRIMARY ALUMINUM REDUCTION PLANTS
#> 63                                                        PORTLAND CEMENT PLANTS
#> 64                          ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 65                         ACRYLIC AND MODACRYLIC FIBERS PRODUCTION AREA SOURCES
#> 66                                                DRY CLEANERS PERCHLOROETHYLENE
#> 67                            CHEMICAL RECOVERY COMBUSTION SOURCES AT PULP MILLS
#> 68                                        PESTICIDE ACTIVE INGREDIENT PRODUCTION
#> 69                             SURFACE COATING OF MISC. METAL PARTS AND PRODUCTS
#> 70                             FLEXIBLE POLYURETHANE FOAM FABRICATION OPERATIONS
#> 71                                          CARBON BLACK PRODUCTION AREA SOURCES
#> 72                                                       CHROMIUM ELECTROPLATING
#> 73                                                 WOOL FIBERGLASS MANUFACTURING
#> 74                                           SURFACE COATING OF LARGE APPLIANCES
#> 75                                                  HYDROCHLORIC ACID PRODUCTION
#> 76                       CHEMICAL MANUFACTURING AREA SOURCES: CHROMIUM COMPOUNDS
#> 77                                                    ETHYLENE OXIDE STERILIZERS
#> 78                                                      TANK STANDARDS - LEVEL 1
#> 79                                          MANUFACTURE OF AMINO/PHENOLIC RESINS
#> 80                    PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 81                            FLEXIBLE POLYURETHANE FOAM PROD & FAB AREA SOURCES
#> 82                                                           CONTAINER STANDARDS
#> 83                                                  POLYETHER POLYOLS PRODUCTION
#> 84                                 SURFACE COATING OF PLASTIC PARTS AND PRODUCTS
#> 85                                                      ENGINE TEST CELLS/STANDS
#> 86                                  LEAD ACID BATTERY MANUFACTURING AREA SOURCES
#> 87                                             INDUSTRIAL PROCESS COOLING TOWERS
#> 88                                                           SURFACE IMPOUNDMENT
#> 89                                     SURFACE COATING OF WOOD BUILDING PRODUCTS
#> 90                                   FRICTION MATERIALS MANUFACTURING FACILITIES
#> 91                                                  WOOD PERSERVING AREA SOURCES
#> 92                                                         GASOLINE DISTRIBUTION
#> 93                                                      INDIVIDUAL DRAIN SYSTEMS
#> 94                                                 SECONDARY ALUMINUM PRODUCTION
#> 95                                            SURFACE COATING OF METAL FURNITURE
#> 96                                                  TACONITE IRON ORE PROCESSING
#> 97                                                CLAY CERAMICS MFG AREA SOURCES
#> 98                                                                PULP AND PAPER
#> 99  CLOSED VENT SYS/CTRL DEVICES, RECOVERY DEVICES, ROUTING TO FUEL GAS SYS/PROC
#> 100                                                SURFACE COATING OF METAL COIL
#> 101                                            REFRACTORY PRODUCTS MANUFACTURING
#> 102                                                       GLASS MFG AREA SOURCES
#> 103                                                 HALOGENATED SOLVENT CLEANING
#> 104                                            EQUIPMENT LEAKS - CONTROL LEVEL 1
#> 105                                                        PRIMARY LEAD SMELTING
#> 106                                                 LEATHER FINISHING OPERATIONS
#> 107                                                   PRIMARY MAGNESIUM REFINING
#> 108                          SECONDARY NONFERROUS METALS PROCESSING AREA SOURCES
#> 109                                                  GROUP 1 POLYMERS AND RESINS
#> 110                                                  EQUIP LEAKS-CONTROL LEVEL 2
#> 111                                   PETROLEUM REFINERIES: CCU'S, CRU'S & SRU'S
#> 112                                              CELLULOSE PRODUCT MANUFACTURING
#> 113                                                     COAL & OIL-FIRED EUSGU'S
#> 114                                         OIL-WATER & ORGANIC-WATER SEPARATORS
#> 115                                               PUBLICLY OWNED TREATMENT WORKS
#> 116                                                           BOAT MANUFACTURING
#> 117                                          CHEMICAL MANUFACTURING AREA SOURCES
#> 118                               EPOXY RESINS & NON-NYLON POLYAMIDES PRODUCTION
#> 119                                     STORAGE VESSELS (TANKS)- CONTROL LEVEL 2
#> 120                                     REINFORCED PLASTIC COMPOSITES PRODUCTION
#> 121                                          HOSPITAL ETHYLENE OXIDE STERILIZERS
#> 122                                AREA SOURCE: PLATING AND POLISHING OPERATIONS
#> 123                                                      SECONDARY LEAD SMELTERS
#> 124                       ETHYLENE MPF: HEAT EXCHANGE SYSTEMS & WASTE OPERATIONS
#> 125                                                        FERROALLOY PRODUCTION
#> 126                                                    RUBBER TIRE MANUFACTURING
#> 127          AREA SOURCE: NINE METAL FABRICATION AND FINISHING SOURCE CATEGORIES
#> 128                            MARINE TANK VESSEL LOADING & UNLOADING OPERATIONS
#> 129                                                       GENERIC MACT STANDARDS
#> 130                                               STATIONARY COMBUSTION TURBINES
#> 131                     AREA SOURCE: ELECTRIC ARC FURNACE STEELMAKING FACILITIES
#> 132                               AREA SOURCE: FERROALLOYS PRODUCTION FACILITIES
#> 133                                                         CHEMICAL SECTOR RULE
#> 134                  STATIONARY RECIPROCATING INTERNAL COMBUSTION ENGINES (RICE)
#> 135                                        IRON AND STEEL FOUNDRIES AREA SOURCES
#> 136                   AREA SOURCE: ALUMINUM, COPPER & OTHER NONFERROUS FOUNDRIES
#>                                                                                  dropdown_label
#> 1                                                                A - General Provisions (5,052)
#> 2                                                 AA - Phosperic Acid Manufacturing Plants (15)
#> 3                                                  AAAA - Municipal Solid Waste Landfills (711)
#> 4                                                        AAAAA - Lime Manufacturing Plants (45)
#> 5              AAAAAAA - Area Source: Asphalt Processing And Asphalt Roofing Manufacturing (34)
#> 6                          B - Req For Control Technology Determinations For Major Sources (81)
#> 7                                                               BB - Phosphate Fertilizers (15)
#> 8                                                      BBBBB - Semiconductor Manufacturing (45)
#> 9     BBBBBB - Gasoline Distribution Bulk Terminals, Bulk Plants, And Pipeline Facilities (943)
#> 10                                    BBBBBBB - Area Source: Chemical Preparations Industry (3)
#> 11                                                                   C - Delistings Of Haps (1)
#> 12                                                              CC - Petroleum Refineries (186)
#> 13                          CCC - Steel Pickling-Hci Process Facilities And Hci Reg Plants (51)
#> 14                                                CCCC - Manufacturing Of Nutritional Yeast (6)
#> 15                               CCCCC - Coke Ovens: Pushing, Quenching And Battery Stacks (27)
#> 16                                              CCCCCC - Gasoline Dispensing Facilities (1,700)
#> 17                          CCCCCCC - Area Source: Paint And Allied Products Manufacturing (40)
#> 18                                              D - Compliance Extension, Early Hap Reductn (5)
#> 19                                             DD - Off-Site Waste And Recovery Operations (97)
#> 20                                                           DDD - Mineral Wool Production (13)
#> 21                                             DDDD - Plywood And Composite Wood Products (392)
#> 22  DDDDD - Major Sources: Industrial/Commercial/Institutional Boilers & Process Heater (2,737)
#> 23                       DDDDDD - Polyvinyl Chloride And Copolymers Production Area Sources (7)
#> 24                                              DDDDDDD - Area Source: Prepared Feeds Mfg (123)
#> 25                                E - Approval Of State Programs &  Delegation Of Authority (1)
#> 26                                                      EE - Magnetic Tape Mfg. Operations (11)
#> 27                                                       EEE - Hazardous Waste Combustors (139)
#> 28                                     EEEE - Organic Liquids Distribution (Non-Gasoline) (322)
#> 29                                                        EEEEE - Iron And Steel Foundries (92)
#> 30                           EEEEEEE - Area Source: Gold Mine Ore Processing And Production (3)
#> 31                                     F - Synthetic Hazardous Organics (Socmi) Standards (269)
#> 32                              FFFF - Miscellaneous Organic Chemical Manufacturing (Mon) (394)
#> 33                              FFFFF - Integrated Iron And Steel Manufacturing Facilities (19)
#> 34                                          FFFFFF - Secondary Copper Smelting Area Sources (1)
#> 35              G - Socmi Process Vents, Storage Vessels, Transfer Operations, Wastewater (304)
#> 36                                                           GG - Aerospace Mfg. & Rework (159)
#> 37                                                      GGG - Pharmaceutical Manufacturing (76)
#> 38                                 GGGG - Solvent Extraction For Vegetable Oil Production (102)
#> 39                                                               GGGGG - Site Remediation (146)
#> 40                                              H - Equipment Leaks Of Hazardous Organics (342)
#> 41                                       HH - Oil And Natural Gas Production Facilities (3,656)
#> 42                                    HHH - Natural Gas Transmission & Storage Facilities (107)
#> 43                                              HHHH - Wet Formed Fiberglass Mat Production (8)
#> 44                                            HHHHH - Miscellaneous Coating Manufacturing (182)
#> 45               HHHHHH - Paint Strip & Misc Surface Coating Operations At Area Sources (3,169)
#> 46                                  HHHHHHH - Polyvinyl Chloride And Copolymers Production (21)
#> 47                                            I - Negotiated Regs: Equip Leaks, Haz-Orgncs (47)
#> 48                                                         II - Shipbuilding & Ship Repair (96)
#> 49                                             III - Flexible Polyurethane Foam Production (42)
#> 50                                   IIII - Surface Coating Of Auto And Light Duty Trucks (116)
#> 51                         IIIII - Mercury Emissions From Mercury Cell Chlor-Alkali Plants (12)
#> 52                                         J - Polyvinyl Chloride And Copolymers Production (9)
#> 53                                           JJ - Wood Furniture Manufacturing Operations (522)
#> 54                                                      JJJ - Group Iv Polymers And Resins (63)
#> 55                                                      JJJJ - Paper & Other Web Coatings (311)
#> 56                                  JJJJJ - Brick & Structural Clay Products Manufacturing (98)
#> 57                JJJJJJ - Industrial, Commercial, And Institutional Boilers Area Sources (740)
#> 58                                                             KK - Printing & Publishing (395)
#> 59                                                    KKKK - Surface Coating Of Metal Cans (60)
#> 60                                                      KKKKK - Clay Ceramics Manufacturing (7)
#> 61                                                                 L - Coke Oven Batteries (28)
#> 62                                                  LL - Primary Aluminum Reduction Plants (16)
#> 63                                                           LLL - Portland Cement Plants (121)
#> 64                            LLLLL - Asphalt Processing And Asphalt Roofing Manufacturing (19)
#> 65                           LLLLLL - Acrylic And Modacrylic Fibers Production Area Sources (2)
#> 66                                                   M - Dry Cleaners Perchloroethylene (8,763)
#> 67                                MM - Chemical Recovery Combustion Sources At Pulp Mills (117)
#> 68                                            MMM - Pesticide Active Ingredient Production (30)
#> 69                               MMMM - Surface Coating Of Misc. Metal Parts And Products (775)
#> 70                                MMMMM - Flexible Polyurethane Foam Fabrication Operations (9)
#> 71                                            MMMMMM - Carbon Black Production Area Sources (4)
#> 72                                                          N - Chromium Electroplating (1,185)
#> 73                                                     NNN - Wool Fiberglass Manufacturing (24)
#> 74                                              NNNN - Surface Coating Of Large Appliances (20)
#> 75                                                    NNNNN - Hydrochloric Acid Production (29)
#> 76                         NNNNNN - Chemical Manufacturing Area Sources: Chromium Compounds (3)
#> 77                                                         O - Ethylene Oxide Sterilizers (111)
#> 78                                                           OO - Tank Standards - Level 1 (16)
#> 79                                              OOO - Manufacture Of Amino/Phenolic Resins (39)
#> 80                      OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles (101)
#> 81                             OOOOOO - Flexible Polyurethane Foam Prod & Fab Area Sources (43)
#> 82                                                                PP - Container Standards (16)
#> 83                                                      PPP - Polyether Polyols Production (31)
#> 84                                   PPPP - Surface Coating Of Plastic Parts And Products (396)
#> 85                                                        PPPPP - Engine Test Cells/Stands (30)
#> 86                                   PPPPPP - Lead Acid Battery Manufacturing Area Sources (29)
#> 87                                                  Q - Industrial Process Cooling Towers (115)
#> 88                                                                 QQ - Surface Impoundment (2)
#> 89                                        QQQQ - Surface Coating Of Wood Building Products (80)
#> 90                                      QQQQQ - Friction Materials Manufacturing Facilities (3)
#> 91                                                   QQQQQQ - Wood Perserving Area Sources (37)
#> 92                                                              R - Gasoline Distribution (562)
#> 93                                                           RR - Individual Drain Systems (18)
#> 94                                                    RRR - Secondary Aluminum Production (340)
#> 95                                               RRRR - Surface Coating Of Metal Furniture (31)
#> 96                                                     RRRRR - Taconite Iron Ore Processing (6)
#> 97                                                  RRRRRR - Clay Ceramics Mfg Area Sources (4)
#> 98                                                                     S - Pulp And Paper (134)
#> 99       SS - Closed Vent Sys/Ctrl Devices, Recovery Devices, Routing To Fuel Gas Sys/Proc (94)
#> 100                                                   SSSS - Surface Coating Of Metal Coil (66)
#> 101                                               SSSSS - Refractory Products Manufacturing (4)
#> 102                                                        SSSSSS - Glass Mfg Area Sources (16)
#> 103                                                    T - Halogenated Solvent Cleaning (1,255)
#> 104                                                 TT - Equipment Leaks - Control Level 1 (13)
#> 105                                                             TTT - Primary Lead Smelting (2)
#> 106                                                    TTTT - Leather Finishing Operations (14)
#> 107                                                      TTTTT - Primary Magnesium Refining (5)
#> 108                           TTTTTT - Secondary Nonferrous Metals Processing Area Sources (10)
#> 109                                                        U - Group 1 Polymers And Resins (28)
#> 110                                                       UU - Equip Leaks-Control Level 2 (86)
#> 111                                      UUU - Petroleum Refineries: Ccu's, Cru's & Sru's (126)
#> 112                                                 UUUU - Cellulose Product Manufacturing (15)
#> 113                                                      UUUUU - Coal & Oil-Fired Eusgu's (249)
#> 114                                              VV - Oil-Water & Organic-Water Separators (16)
#> 115                                                   VVV - Publicly Owned Treatment Works (18)
#> 116                                                             VVVV - Boat Manufacturing (237)
#> 117                                          VVVVVV -  Chemical Manufacturing Area Sources (81)
#> 118                                     W - Epoxy Resins & Non-Nylon Polyamides Production (10)
#> 119                                          WW - Storage Vessels (Tanks)- Control Level 2 (54)
#> 120                                       WWWW - Reinforced Plastic Composites Production (729)
#> 121                                           WWWWW - Hospital Ethylene Oxide Sterilizers (111)
#> 122                                WWWWWW - Area Source: Plating And Polishing Operations (222)
#> 123                                                            X - Secondary Lead Smelters (22)
#> 124                            XX - Ethylene Mpf: Heat Exchange Systems & Waste Operations (23)
#> 125                                                             XXX - Ferroalloy Production (4)
#> 126                                                       XXXX - Rubber Tire Manufacturing (33)
#> 127          XXXXXX - Area Source: Nine Metal Fabrication And Finishing Source Categories (228)
#> 128                                 Y - Marine Tank Vessel Loading & Unloading Operations (185)
#> 129                                                            YY - Generic Mact Standards (90)
#> 130                                                 YYYY - Stationary Combustion Turbines (248)
#> 131                       YYYYY - Area Source: Electric Arc Furnace Steelmaking Facilities (48)
#> 132                                 YYYYYY - Area Source: Ferroalloys Production Facilities (6)
#> 133                                                              ZZZ - Chemical Sector Rule (4)
#> 134                 ZZZZ - Stationary Reciprocating Internal Combustion Engines (Rice) (25,996)
#> 135                                         ZZZZZ - Iron And Steel Foundries Area Sources (149)
#> 136                    ZZZZZZ - Area Source: Aluminum, Copper & Other Nonferrous Foundries (29)
#>         N
#> 1    5052
#> 2      15
#> 3     711
#> 4      45
#> 5      34
#> 6      81
#> 7      15
#> 8      45
#> 9     943
#> 10      3
#> 11      1
#> 12    186
#> 13     51
#> 14      6
#> 15     27
#> 16   1700
#> 17     40
#> 18      5
#> 19     97
#> 20     13
#> 21    392
#> 22   2737
#> 23      7
#> 24    123
#> 25      1
#> 26     11
#> 27    139
#> 28    322
#> 29     92
#> 30      3
#> 31    269
#> 32    394
#> 33     19
#> 34      1
#> 35    304
#> 36    159
#> 37     76
#> 38    102
#> 39    146
#> 40    342
#> 41   3656
#> 42    107
#> 43      8
#> 44    182
#> 45   3169
#> 46     21
#> 47     47
#> 48     96
#> 49     42
#> 50    116
#> 51     12
#> 52      9
#> 53    522
#> 54     63
#> 55    311
#> 56     98
#> 57    740
#> 58    395
#> 59     60
#> 60      7
#> 61     28
#> 62     16
#> 63    121
#> 64     19
#> 65      2
#> 66   8763
#> 67    117
#> 68     30
#> 69    775
#> 70      9
#> 71      4
#> 72   1185
#> 73     24
#> 74     20
#> 75     29
#> 76      3
#> 77    111
#> 78     16
#> 79     39
#> 80    101
#> 81     43
#> 82     16
#> 83     31
#> 84    396
#> 85     30
#> 86     29
#> 87    115
#> 88      2
#> 89     80
#> 90      3
#> 91     37
#> 92    562
#> 93     18
#> 94    340
#> 95     31
#> 96      6
#> 97      4
#> 98    134
#> 99     94
#> 100    66
#> 101     4
#> 102    16
#> 103  1255
#> 104    13
#> 105     2
#> 106    14
#> 107     5
#> 108    10
#> 109    28
#> 110    86
#> 111   126
#> 112    15
#> 113   249
#> 114    16
#> 115    18
#> 116   237
#> 117    81
#> 118    10
#> 119    54
#> 120   729
#> 121   111
#> 122   222
#> 123    22
#> 124    23
#> 125     4
#> 126    33
#> 127   228
#> 128   185
#> 129    90
#> 130   248
#> 131    48
#> 132     6
#> 133     4
#> 134 25996
#> 135   149
#> 136    29
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = FALSE) # default
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> frs_by_mact   - is NOT in memory. Checking local disk... frs_by_mact  is NOT found locally on disk at ~/../Downloads/EJAMbigfiles/frs_by_mact.arrow 
#> frs_by_mact   - is NOT in local folder.
#> frs_by_mact  - was read from pins board server, into memory (to specified envir).
#> 
#> Key: <subpart, programid>
#>              programid subpart
#>                 <char>  <char>
#>  1: 080000003800500001    OOOO
#>  2: 0900000006111R9011    OOOO
#>  3: AR0000000500100022    OOOO
#>  4: CO0000000812300098    OOOO
#>  5: CT0000000900500003    OOOO
#>  6: CT0000000901501138    OOOO
#>  7: GA0000001318900002    OOOO
#>  8: KS0000002013300027    OOOO
#>  9: MA0000002512000290    OOOO
#> 10: MA0000002512100102    OOOO
#> 11: MI00000000000P0634    OOOO
#> 12: MO0000002918901012    OOOO
#> 13: MS0000002808100088    OOOO
#> 14: MT0000003007500005    OOOO
#> 15: NC0000003701800493    OOOO
#> 16: NC0000003705100014    OOOO
#> 17: NC0000003708100164    OOOO
#> 18: NH0000003301100129    OOOO
#> 19: NY0000005532600004    OOOO
#> 20: NY0000006301300065    OOOO
#> 21: OH0000000616010036    OOOO
#> 22: OH0000001667060088    OOOO
#> 23:        PA000492043    OOOO
#> 24: RI0000004400900010    OOOO
#> 25:      SC00000800027    OOOO
#> 26:      SC00002000032    OOOO
#> 27:      SC00012000029    OOOO
#> 28:      SC00018800040    OOOO
#> 29: TN0000004706300275    OOOO
#> 30: TN0000004716300007    OOOO
#> 31: VA0000005115500016    OOOO
#> 32: WI0000005502500072    OOOO
#>              programid subpart
#>                                                          title      lat
#>                                                         <char>    <num>
#>  1: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 47.97940
#>  2: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.22082
#>  3: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.27496
#>  4: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 40.27542
#>  5: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 41.61256
#>  6: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 41.66805
#>  7: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 33.44801
#>  8: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 37.68370
#>  9: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 41.60689
#> 10: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 42.72361
#> 11: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 42.88929
#> 12: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 38.76637
#> 13: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.27973
#> 14: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 45.11306
#> 15: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 35.72438
#> 16: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 35.36954
#> 17: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 35.25610
#> 18: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 42.75050
#> 19: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 43.11147
#> 20: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 43.21786
#> 21: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 40.26201
#> 22: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 41.16331
#> 23: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 40.20903
#> 24: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 41.44594
#> 25: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 33.56057
#> 26: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.61164
#> 27: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 35.01979
#> 28: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.78693
#> 29: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 36.16847
#> 30: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 36.56050
#> 31: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 37.04957
#> 32: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 43.11747
#>                                                          title      lat
#>            lon  REGISTRY_ID program
#>          <num>       <char>  <char>
#>  1:  -99.02064 110000825867     AIR
#>  2: -119.02272 110043415578     AIR
#>  3:  -91.34179 110025082924     AIR
#>  4: -104.80199 110001718053     AIR
#>  5:  -73.08986 110039211064     AIR
#>  6:  -71.91848 110001732554     AIR
#>  7:  -82.53785 110000359406     AIR
#>  8:  -95.46628 110000445938     AIR
#>  9:  -70.90407 110000739675     AIR
#> 10:  -70.96296 110000309773     AIR
#> 11:  -85.63853 110067012878     AIR
#> 12:  -90.47491 110000440381     AIR
#> 13:  -88.76446 110013841432     AIR
#> 14: -105.08569 110046594146     AIR
#> 15:  -81.37813 110000350584     AIR
#> 16:  -78.55103 110001478730     AIR
#> 17:  -81.77920 110013792921     AIR
#> 18:  -71.82836 110001522184     AIR
#> 19:  -73.56985 110000582619     AIR
#> 20:  -75.46524 110000325871     AIR
#> 21:  -81.86949 110000384753     AIR
#> 22:  -81.25312 110000766644     AIR
#> 23:  -77.02432 110012697812     AIR
#> 24:  -71.62625 110000312117     AIR
#> 25:  -81.80997 110031367036     AIR
#> 26:  -82.47996 110001664823     AIR
#> 27:  -82.56309 110041625989     AIR
#> 28:  -82.68964 110001667802     AIR
#> 29:  -83.39178 110038045139     AIR
#> 30:  -82.17101 110000913548     AIR
#> 31:  -80.75514 110000343469     AIR
#> 32:  -89.50733 110000419878     AIR
#>            lon  REGISTRY_ID program
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = TRUE)
#> Key: <subpart, programid>
#>               programid subpart
#>                  <char>  <char>
#>   1: 080000003800500001    OOOO
#>   2: 0900000006111R9011    OOOO
#>   3: AR0000000500100022    OOOO
#>   4: AZMCA004013X138761    OOOO
#>   5: CO0000000812300098    OOOO
#>  ---                           
#>  97: VA0000005115500016    OOOO
#>  98: VA0000005116300001    OOOO
#>  99: VA0000005159000002    OOOO
#> 100: VA0000005170000091    OOOO
#> 101: WI0000005502500072    OOOO
#>                                                           title      lat
#>                                                          <char>    <num>
#>   1: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 47.97940
#>   2: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.22082
#>   3: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 34.27496
#>   4: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES       NA
#>   5: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 40.27542
#>  ---                                                                    
#>  97: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 37.04957
#>  98: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES       NA
#>  99: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES       NA
#> 100: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES       NA
#> 101: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES 43.11747
#>             lon  REGISTRY_ID program
#>           <num>       <char>  <char>
#>   1:  -99.02064 110000825867     AIR
#>   2: -119.02272 110043415578     AIR
#>   3:  -91.34179 110025082924     AIR
#>   4:         NA         <NA>    <NA>
#>   5: -104.80199 110001718053     AIR
#>  ---                                
#>  97:  -80.75514 110000343469     AIR
#>  98:         NA         <NA>    <NA>
#>  99:         NA         <NA>    <NA>
#> 100:         NA         <NA>    <NA>
#> 101:  -89.50733 110000419878     AIR