guess which columns have lat and lon based on aliases like latitude, FacLat, etc.

guess which columns have lat and lon based on aliases like latitude, FacLat, etc.

latlon_infer(mycolnames)

Arguments

mycolnames

e.g., colnames(x) where x is a data.frame from read.csv

Value

returns all of mycolnames except replacing the best candidates with lat and lon

See also

Examples

  latlon_infer(c('trilat', 'belong', 'belong')) # warns if no alias found, 
#> Warning: lat missing and no synonyms found
#> Warning: lon missing and no synonyms found
#> [1] "trilat" "belong" "belong"
    #  but doesnt warn of dupes in other terms, just preferred term.
  latlon_infer(c('a', 'LONG', 'Longitude', 'lat')) # only the best alias is converted/used
#> Warning: Replaced column names that were inferred to be and therefore renamed as the lat and/or lon columns!
#> [1] "a"    "LONG" "lon"  "lat" 
  latlon_infer(c('a', 'LONGITUDE', 'Long', 'Lat')) # only the best alias is converted/used
#> Warning: Replaced column names that were inferred to be and therefore renamed as the lat and/or lon columns!
#> [1] "a"    "lon"  "Long" "lat" 
  latlon_infer(c('a', 'longing', 'Lat', 'lat', 'LAT')) # case variants of preferred are 
#> Warning: lon missing and no synonyms found
#> [1] "a"       "longing" "Lat"     "lat"     "LAT"    
      # left alone only if lowercase one is found
  latlon_infer(c('LONG', 'long', 'lat')) # case variants of a single alias are 
#> Warning: DUPLICATED lon
#> Warning: Replaced column names that were inferred to be and therefore renamed as the lat and/or lon columns!
#> [1] "lon" "lon" "lat"
      # converted to preferred word (if pref not found), creating dupes!  warn!
  latlon_infer(c('LONG', 'LONG')) # dupes of an alias are renamed and still are dupes! warn!
#> Warning: lat missing and no synonyms found
#> Warning: DUPLICATED lon
#> Warning: Replaced column names that were inferred to be and therefore renamed as the lat and/or lon columns!
#> [1] "lon" "lon"
  latlon_infer(c('lat', 'lat', 'Lon')) # dupes left as dupes but warn!
#> Warning: DUPLICATED lat
#> Warning: Replaced column names that were inferred to be and therefore renamed as the lat and/or lon columns!
#> [1] "lat" "lat" "lon"