Validate latitudes and longitudes

Check each latitude and longitude value to see if they are valid.

latlon_is.valid(lat, lon, quiet = TRUE)

Arguments

lat

vector of latitudes

lon

vector of longitudes

quiet

optional logical, if TRUE, show list of bad values in console

Value

logical vector, one element per lat lon pair (location)

Details

NA or outside expected numeric ranges

(based on approx ranges of lat lon seen among block internal points dataset)

But note Guam, American Samoa, Northern Mariana Islands, and U.S. Virgin Islands are outside these ranges! EJScreen 2.2 does not provide demographic data in those locations anyway, but can map sites there.

lat must be between 17.5 and 71.5, and

lon must be ( between -180 and -64) OR (between 172 and 180)

See also

Examples

 if (FALSE) {
 # this would only work using the EJAM package datasets frs and blockpoints:
   if (!exists("frs")) dataload_from_pins("frs")
 table(latlon_is.valid(lat =  frs$lat, lon =  frs$lon))
 # blockpoints may need to be downloaded using dataload_from_pins()
 table(latlon_is.valid(lat =  blockpoints$lat, lon =  blockpoints$lon))
  }