Map - Blockgroup polygons / boundaries near 1 site - Create leaflet map

Overlay blockgroups near 1 site, after plotblocksnearby(returnmap = TRUE)

Usage

map_blockgroups_over_blocks(y)

Arguments

y

output of plotblocksnearby() but with returnmap = TRUE

Value

leaflet map widget

See also

Examples

 y <- plotblocksnearby(testpoints_10[5,],
        radius = 3,
        returnmap = TRUE)
#> Analyzing 1 points, radius of 3 miles around each.
#> Finding Census blocks with internal point within  3  miles of the site (point), for each of 1  sites (points)...
#> Stats via getblocks_diagnostics(), but NOT ADJUSTING UP FOR VERY SHORT DISTANCES: 
#> min distance before adjustment:  0.2532525 
#> max distance before adjustment:  5.165464 
#> 
 map_blockgroups_over_blocks(y)
#> Loading arrow datasets: bgid2fips
#> looking for bgid2fips_arrow in memory...
#> bgid2fips_arrow  is loading from local folder ...done.
#> 
#> Reading layer `OGRGeoJSON' from data source 
#>   `https://services.arcgis.com/P3ePLMYs2RVChkJx/ArcGIS/rest/services/USA_Boundaries_2022/FeatureServer/5/query?where=FIPS%3D%27340010114011%27%20OR%20FIPS%3D%27340010114012%27%20OR%20FIPS%3D%27340010114013%27%20OR%20FIPS%3D%27340010114033%27%20OR%20FIPS%3D%27340010114041%27%20OR%20FIPS%3D%27340010114042%27%20OR%20FIPS%3D%27340010115002%27%20OR%20FIPS%3D%27340010115003%27%20OR%20FIPS%3D%27340010116001%27%20OR%20FIPS%3D%27340010116002%27%20OR%20FIPS%3D%27340010116003%27%20OR%20FIPS%3D%27340010118051%27&outFields=FIPS%2CSTATE_ABBR%2CSQMI&returnGeometry=true&f=geojson' 
#>   using driver `GeoJSON'
#> Simple feature collection with 12 features and 3 fields
#> Geometry type: MULTIPOLYGON
#> Dimension:     XY
#> Bounding box:  xmin: -74.86494 ymin: 39.28943 xmax: -74.61692 ymax: 39.60117
#> Geodetic CRS:  WGS 84
#> Registered S3 method overwritten by 'geojsonsf':
#>   method        from   
#>   print.geojson geojson
#> Turn off the blockgroup boundaries layer using the map layer control button, to enable popup info for each block point.