Map - points - Create leaflet map of points, in shiny app

make a leaflet map of uploaded points such as facilities

Usage

map_facilities(mypoints, rad = 3, highlight = FALSE, clustered)

Arguments

mypoints,

data frame of uploaded points

rad,

a size for drawing each circle (buffer search radius)

highlight,

a logicial for whether to highlight overlapping points (defaults to FALSE)

clustered,

a vector of T/F values for each point, indicating if they overlap with another

Value

a leaflet map with circles, circleMarkers, and basic popup

See also