helper function in updating the package metadata, used by scripts in /data-raw/

Quick and dirty helper during development, to check all the attributes of all the data files in relevant packages. It loads unloaded packages as needed, which you might not want it to do, but it is not coded to be able to check attributes without doing that.

Usage

metadata_check(
  packages = EJAM::ejampackages,
  datasets = "all",
  which = c("ejam_package_version", "date_saved_in_package", "date_downloaded",
    "ejscreen_version", "ejscreen_releasedate", "acs_releasedate", "acs_version",
    "census_version"),
  grepdatasets = FALSE,
  loadifnotloaded = TRUE
)

Arguments

packages

Optional. e.g. 'EJAMejscreendata', or can be a vector of character strings, and if not specified, default is to report on EJAM::ejampackages. If set to NULL, it only reports on objects already attached.

datasets

optional, "all" means all data objects exported. Can be a vector of character names of the ones to check like c("bgpts", "blockpoints")

which

Optional vector (not list) of strings, the attributes. Default is some typical ones used in EJAM-related packages currently.

grepdatasets

optional, if set to TRUE, datasets should be a query to use via grep to identify which datasets to check. It always uses ignore.case=TRUE for this.

loadifnotloaded

Optional to control if func should temporarily attach packages not already loaded.

See also

Examples

x = metadata_check( which = "ejam_package_version")
#> Error in metadata_check(which = "ejam_package_version"): could not find function "metadata_check"
x[!x$ejam_package_version %in% "2.32.6", ]
#> Error: object 'x' not found

  # tail(EJAM:::metadata_check( ))
  EJAM:::metadata_check(packages = NULL)
#>    package                    item ejam_package_version date_saved_in_package
#> 1       NA               avg.in.us               2.32.6            2024-10-12
#> 2       NA                    bgej               2.32.5            2025-06-23
#> 3       NA               bgid2fips                2.3.0            2024-08-14
#> 4       NA         bgid2fips_arrow                 <NA>                  <NA>
#> 5       NA            blockid2fips                2.3.0            2024-08-14
#> 6       NA             blockpoints                2.3.0            2024-08-14
#> 7       NA                blockwts                2.3.0            2024-08-14
#> 8       NA                     frs                 <NA>                  <NA>
#> 9       NA               frs_arrow                 <NA>                  <NA>
#> 10      NA       frs_by_mact_arrow                 <NA>                  <NA>
#> 11      NA            frs_by_naics                 <NA>                  <NA>
#> 12      NA        frs_by_programid                 <NA>                  <NA>
#> 13      NA  frs_by_programid_arrow                 <NA>                  <NA>
#> 14      NA              frs_by_sic                 <NA>                  <NA>
#> 15      NA global_defaults_package                 <NA>                  <NA>
#> 16      NA               localtree                 <NA>                  <NA>
#> 17      NA                quaddata                2.3.0            2024-08-14
#>    date_downloaded ejscreen_version ejscreen_releasedate acs_releasedate
#> 1               NA             2.32           2024-08-12      2023-12-07
#> 2               NA             2.32           2024-08-12      2023-12-07
#> 3               NA              2.3            July 2024      2023-12-07
#> 4               NA             <NA>                 <NA>            <NA>
#> 5               NA              2.3            July 2024      2023-12-07
#> 6               NA              2.3            July 2024      2023-12-07
#> 7               NA              2.3            July 2024      2023-12-07
#> 8               NA             <NA>                 <NA>            <NA>
#> 9               NA             <NA>                 <NA>            <NA>
#> 10              NA             <NA>                 <NA>            <NA>
#> 11              NA             <NA>                 <NA>            <NA>
#> 12              NA             <NA>                 <NA>            <NA>
#> 13              NA             <NA>                 <NA>            <NA>
#> 14              NA             <NA>                 <NA>            <NA>
#> 15              NA             <NA>                 <NA>            <NA>
#> 16              NA             <NA>                 <NA>            <NA>
#> 17              NA              2.3            July 2024      2023-12-07
#>    acs_version census_version has_metadata
#> 1    2018-2022           2020         TRUE
#> 2    2018-2022           2020         TRUE
#> 3    2018-2022           2020         TRUE
#> 4         <NA>           <NA>        FALSE
#> 5    2018-2022           2020         TRUE
#> 6    2018-2022           2020         TRUE
#> 7    2018-2022           2020         TRUE
#> 8         <NA>           <NA>        FALSE
#> 9         <NA>           <NA>        FALSE
#> 10        <NA>           <NA>        FALSE
#> 11        <NA>           <NA>        FALSE
#> 12        <NA>           <NA>        FALSE
#> 13        <NA>           <NA>        FALSE
#> 14        <NA>           <NA>        FALSE
#> 15        <NA>           <NA>        FALSE
#> 16        <NA>           <NA>        FALSE
#> 17   2018-2022           2020         TRUE

  x <- EJAM:::metadata_check_print("EJAM")
#> 
#> 
#> Also see ?metadata_check_print() ?metadata_check() ?metadata_add() ?metadata_update_attr() ?metadata_add_and_use_this() ?dataset_documenter()
#> 
#> Also see
#> 
#>     x = EJAM:::pkg_functions_and_data(pkg = "EJAM",
#>       internal_included = FALSE, exportedfuncs_included = FALSE, data_included = TRUE)$object
#>     x[!grepl("^name", x)] 
#> 
#> Also see 
#> 
#>     y = pkg_data(pkg = "EJAM", simple = F)
#>     rdafiles =  y$Item   # same thing as data(package = pkg)$results[ , "Item"]
#>     
#> 
#> 
#>  See what metadata is stored as attributes 
#> 
#>                       1       
#> package               "EJAM"  
#> item                  "NAICS" 
#> ejam_package_version  "2.32.6"
#> date_saved_in_package NA      
#> date_downloaded       NA      
#> ejscreen_version      NA      
#> ejscreen_releasedate  NA      
#> acs_releasedate       NA      
#> acs_version           NA      
#> census_version        NA      
#> has_metadata          "TRUE"  
#> 
#> 
#> 
#> 
#>  See which data objects have outdated or missing metadata about ejam_package_version, etc. 
#> 
#> but omit atomic vectors since adding attributes to those makes them nonvectors and messier to print 
#>                                         item is.atomic ejam_package_version
#> 2                                  avg.in.us     FALSE               2.32.6
#> 3                              bg_cenpop2020     FALSE               2.32.6
#> 4                                      bgpts     FALSE               2.32.6
#> 5                            blockgroupstats     FALSE               2.32.6
#> 6                               censusplaces     FALSE               2.32.6
#> 8         ejscreenRESTbroker2table_na_filler     FALSE               2.32.6
#> 10                      epa_programs_defined     FALSE               2.32.6
#> 13                           frsprogramcodes     FALSE               2.32.6
#> 14                high_pctiles_tied_with_min     FALSE               2.32.6
#> 15                                mact_table     FALSE               2.32.6
#> 16                           map_headernames     FALSE               2.32.6
#> 17                              naics_counts     FALSE               2.32.6
#> 18                                naicstable     FALSE               2.32.6
#> 19                                     namez     FALSE               2.32.6
#> 20                                 stateinfo     FALSE               2.32.6
#> 21                                stateinfo2     FALSE               2.32.6
#> 22                          states_shapefile     FALSE               2.32.6
#> 23                                statestats     FALSE               2.32.6
#> 24                   testinput_address_table     FALSE               2.32.6
#> 25                 testinput_address_table_9     FALSE               2.32.6
#> 26         testinput_address_table_goodnames     FALSE               2.32.6
#> 27          testinput_address_table_withfull     FALSE               2.32.6
#> 28                  testinput_program_sys_id     FALSE               2.32.6
#> 29                        testinput_shapes_2     FALSE               2.32.6
#> 30     testoutput_doaggregate_1000pts_1miles     FALSE               2.32.6
#> 31      testoutput_doaggregate_100pts_1miles     FALSE               2.32.6
#> 32       testoutput_doaggregate_10pts_1miles     FALSE               2.32.6
#> 33          testoutput_ejamit_1000pts_1miles     FALSE               2.32.6
#> 34           testoutput_ejamit_100pts_1miles     FALSE               2.32.6
#> 35            testoutput_ejamit_10pts_1miles     FALSE               2.32.6
#> 36             testoutput_ejamit_fips_cities     FALSE               2.32.6
#> 37           testoutput_ejamit_fips_counties     FALSE               2.32.6
#> 38                testoutput_ejamit_shapes_2     FALSE               2.32.6
#> 39 testoutput_ejscreenRESTbroker_1pts_1miles     FALSE               2.32.6
#> 40        testoutput_ejscreenapi_1pts_1miles     FALSE               2.32.6
#> 41             testoutput_ejscreenapi_plus_5     FALSE               2.32.6
#> 42                   testoutput_ejscreenit_5     FALSE               2.32.6
#> 43                  testoutput_ejscreenit_50     FALSE               2.32.6
#> 44                 testoutput_ejscreenit_500     FALSE               2.32.6
#> 45 testoutput_getblocksnearby_1000pts_1miles     FALSE               2.32.6
#> 46  testoutput_getblocksnearby_100pts_1miles     FALSE               2.32.6
#> 47   testoutput_getblocksnearby_10pts_1miles     FALSE               2.32.6
#> 48                             testpoints_10     FALSE               2.32.6
#> 49                            testpoints_100     FALSE               2.32.6
#> 50                           testpoints_1000     FALSE               2.32.6
#> 51                          testpoints_10000     FALSE               2.32.6
#> 52                         testpoints_100_dt     FALSE               2.32.6
#> 53                              testpoints_5     FALSE               2.32.6
#> 54                             testpoints_50     FALSE               2.32.6
#> 55                            testpoints_500     FALSE               2.32.6
#> 56                            testpoints_bad     FALSE               2.32.6
#> 57                       testpoints_overlap3     FALSE               2.32.6
#> 58                              testshapes_2     FALSE               2.32.6
#> 59                                  usastats     FALSE               2.32.6
#> 61           default_points_shown_at_startup     FALSE               2.32.6
#> 62                               islandareas     FALSE               2.32.6
#> 63                             metadata4pins     FALSE               2.32.6
#> 64                          modelDoaggregate     FALSE               2.32.6
#> 65                               modelEjamit     FALSE               2.32.6
#> 66                                  sictable     FALSE               2.32.6
#> 68                          ejamdata_version     FALSE                 <NA>
#>    date_saved_in_package has_metadata
#> 2             2024-10-12         TRUE
#> 3             2024-08-22         TRUE
#> 4             2024-08-22         TRUE
#> 5             2024-12-13         TRUE
#> 6             2024-09-26         TRUE
#> 8             2024-12-11         TRUE
#> 10            2024-11-15         TRUE
#> 13            2025-02-12         TRUE
#> 14            2024-10-12         TRUE
#> 15            2025-02-12         TRUE
#> 16            2025-08-26         TRUE
#> 17            2025-02-12         TRUE
#> 18            2024-10-12         TRUE
#> 19            2025-08-26         TRUE
#> 20            2024-08-22         TRUE
#> 21            2024-08-22         TRUE
#> 22            2024-08-22         TRUE
#> 23            2024-08-22         TRUE
#> 24            2025-05-01         TRUE
#> 25            2025-05-01         TRUE
#> 26            2025-05-01         TRUE
#> 27            2025-05-01         TRUE
#> 28            2025-02-12         TRUE
#> 29            2025-05-30         TRUE
#> 30            2025-09-22         TRUE
#> 31            2025-09-22         TRUE
#> 32            2025-09-22         TRUE
#> 33            2025-09-19         TRUE
#> 34            2025-09-19         TRUE
#> 35            2025-09-19         TRUE
#> 36            2025-09-23         TRUE
#> 37            2025-09-23         TRUE
#> 38            2025-09-23         TRUE
#> 39            2024-12-11         TRUE
#> 40            2024-12-11         TRUE
#> 41            2024-12-11         TRUE
#> 42            2024-12-11         TRUE
#> 43            2024-12-11         TRUE
#> 44            2024-12-11         TRUE
#> 45            2025-09-19         TRUE
#> 46            2025-09-19         TRUE
#> 47            2025-09-19         TRUE
#> 48            2025-09-19         TRUE
#> 49            2025-09-19         TRUE
#> 50            2025-09-19         TRUE
#> 51            2025-09-19         TRUE
#> 52            2025-02-14         TRUE
#> 53            2024-09-20         TRUE
#> 54            2024-09-20         TRUE
#> 55            2024-09-20         TRUE
#> 56            2025-09-19         TRUE
#> 57            2024-09-05         TRUE
#> 58            2025-05-30         TRUE
#> 59            2024-08-22         TRUE
#> 61                  <NA>         TRUE
#> 62                  <NA>         TRUE
#> 63                  <NA>         TRUE
#> 64                  <NA>         TRUE
#> 65                  <NA>         TRUE
#> 66                  <NA>         TRUE
#> 68                  <NA>        FALSE
#> 
#> 
#> atomic: 
#> 
#>   [1] "SIC"                                       
#>   [2] "custom"                                    
#>   [3] "epa_programs"                              
#>   [4] "formulas_all"                              
#>   [5] "formulas_d"                                
#>   [6] "NAICS"                                     
#>   [7] "x_anyother"                                
#>   [8] "ejampackages"                              
#>   [9] "lat_alias"                                 
#>  [10] "lon_alias"                                 
#>  [11] "meters_per_mile"                           
#>  [12] "names_age"                                 
#>  [13] "names_age_count"                           
#>  [14] "names_all"                                 
#>  [15] "names_all_r"                               
#>  [16] "names_climate"                             
#>  [17] "names_climate_avg"                         
#>  [18] "names_climate_pctile"                      
#>  [19] "names_climate_state_avg"                   
#>  [20] "names_climate_state_pctile"                
#>  [21] "names_community"                           
#>  [22] "names_community_count"                     
#>  [23] "names_countabove"                          
#>  [24] "names_criticalservice"                     
#>  [25] "names_criticalservice_avg"                 
#>  [26] "names_criticalservice_pctile"              
#>  [27] "names_criticalservice_state_avg"           
#>  [28] "names_criticalservice_state_pctile"        
#>  [29] "names_d"                                   
#>  [30] "names_d_avg"                               
#>  [31] "names_d_count"                             
#>  [32] "names_d_demogindexstate"                   
#>  [33] "names_d_extra"                             
#>  [34] "names_d_extra_count"                       
#>  [35] "names_d_language"                          
#>  [36] "names_d_language_count"                    
#>  [37] "names_d_languageli"                        
#>  [38] "names_d_languageli_count"                  
#>  [39] "names_d_other_count"                       
#>  [40] "names_d_pctile"                            
#>  [41] "names_d_ratio_to_avg"                      
#>  [42] "names_d_ratio_to_state_avg"                
#>  [43] "names_d_state_avg"                         
#>  [44] "names_d_state_pctile"                      
#>  [45] "names_d_subgroups"                         
#>  [46] "names_d_subgroups_alone"                   
#>  [47] "names_d_subgroups_alone_avg"               
#>  [48] "names_d_subgroups_alone_count"             
#>  [49] "names_d_subgroups_alone_pctile"            
#>  [50] "names_d_subgroups_alone_ratio_to_avg"      
#>  [51] "names_d_subgroups_alone_ratio_to_state_avg"
#>  [52] "names_d_subgroups_alone_state_avg"         
#>  [53] "names_d_subgroups_alone_state_pctile"      
#>  [54] "names_d_subgroups_avg"                     
#>  [55] "names_d_subgroups_count"                   
#>  [56] "names_d_subgroups_nh"                      
#>  [57] "names_d_subgroups_nh_avg"                  
#>  [58] "names_d_subgroups_nh_count"                
#>  [59] "names_d_subgroups_nh_pctile"               
#>  [60] "names_d_subgroups_nh_ratio_to_avg"         
#>  [61] "names_d_subgroups_nh_ratio_to_state_avg"   
#>  [62] "names_d_subgroups_nh_state_avg"            
#>  [63] "names_d_subgroups_nh_state_pctile"         
#>  [64] "names_d_subgroups_pctile"                  
#>  [65] "names_d_subgroups_ratio_to_avg"            
#>  [66] "names_d_subgroups_ratio_to_state_avg"      
#>  [67] "names_d_subgroups_state_avg"               
#>  [68] "names_d_subgroups_state_pctile"            
#>  [69] "names_e"                                   
#>  [70] "names_e_avg"                               
#>  [71] "names_e_other"                             
#>  [72] "names_e_pctile"                            
#>  [73] "names_e_ratio_to_avg"                      
#>  [74] "names_e_ratio_to_state_avg"                
#>  [75] "names_e_state_avg"                         
#>  [76] "names_e_state_pctile"                      
#>  [77] "names_ej"                                  
#>  [78] "names_ej_pctile"                           
#>  [79] "names_ej_state"                            
#>  [80] "names_ej_state_pctile"                     
#>  [81] "names_ej_supp"                             
#>  [82] "names_ej_supp_pctile"                      
#>  [83] "names_ej_supp_state"                       
#>  [84] "names_ej_supp_state_pctile"                
#>  [85] "names_featuresinarea"                      
#>  [86] "names_flag"                                
#>  [87] "names_geo"                                 
#>  [88] "names_health"                              
#>  [89] "names_health_avg"                          
#>  [90] "names_health_count"                        
#>  [91] "names_health_pctile"                       
#>  [92] "names_health_ratio_to_avg"                 
#>  [93] "names_health_ratio_to_state_avg"           
#>  [94] "names_health_state_avg"                    
#>  [95] "names_health_state_pctile"                 
#>  [96] "names_misc"                                
#>  [97] "names_pct_as_fraction_blockgroupstats"     
#>  [98] "names_pct_as_fraction_ejamit"              
#>  [99] "names_pct_as_fraction_ejscreenit"          
#> [100] "names_sitesinarea"                         
#> [101] "names_these"                               
#> [102] "names_these_avg"                           
#> [103] "names_these_ratio_to_avg"                  
#> [104] "names_these_ratio_to_state_avg"            
#> [105] "names_these_state_avg"                     
#> [106] "names_wts"                                 
#> [107] "testinput_address_2"                       
#> [108] "testinput_address_9"                       
#> [109] "testinput_address_parts"                   
#> [110] "testinput_fips_blockgroups"                
#> [111] "testinput_fips_cities"                     
#> [112] "testinput_fips_counties"                   
#> [113] "testinput_fips_mix"                        
#> [114] "testinput_fips_states"                     
#> [115] "testinput_fips_tracts"                     
#> [116] "testinput_mact"                            
#> [117] "testinput_naics"                           
#> [118] "testinput_program_name"                    
#> [119] "testinput_regid"                           
#> [120] "testinput_registry_id"                     
#> [121] "testinput_sic"                             
#> 
#> 
#> How.many.have.ejscreen_info
#> 2.32 <NA> 
#>   42  140 
#> 
#> 
#> How.many.have.acs_version
#> 2018-2022      <NA> 
#>        41       141 
#> 
#> 
#>  see where is ACS version info missing, and date saved 
#> 
#>                                           item date_saved_in_package
#> 1                                        NAICS                  <NA>
#> 2                                          SIC            2025-02-12
#> 4                                bg_cenpop2020            2024-08-22
#> 5                                        bgpts            2024-08-22
#> 7                                 censusplaces            2024-09-26
#> 9              default_points_shown_at_startup                  <NA>
#> 10                            ejamdata_version                  <NA>
#> 11                                ejampackages                  <NA>
#> 13                                epa_programs            2025-02-12
#> 17                             frsprogramcodes            2025-02-12
#> 19                                 islandareas                  <NA>
#> 20                                   lat_alias                  <NA>
#> 21                                   lon_alias                  <NA>
#> 22                                  mact_table            2025-02-12
#> 23                             map_headernames            2025-08-26
#> 24                               metadata4pins                  <NA>
#> 25                             meters_per_mile                  <NA>
#> 26                            modelDoaggregate                  <NA>
#> 27                                 modelEjamit                  <NA>
#> 28                                naics_counts            2025-02-12
#> 29                                  naicstable            2024-10-12
#> 30                                   names_age                  <NA>
#> 31                             names_age_count                  <NA>
#> 32                                   names_all                  <NA>
#> 33                                 names_all_r                  <NA>
#> 34                               names_climate                  <NA>
#> 35                           names_climate_avg                  <NA>
#> 36                        names_climate_pctile                  <NA>
#> 37                     names_climate_state_avg                  <NA>
#> 38                  names_climate_state_pctile                  <NA>
#> 39                             names_community                  <NA>
#> 40                       names_community_count                  <NA>
#> 41                            names_countabove                  <NA>
#> 42                       names_criticalservice                  <NA>
#> 43                   names_criticalservice_avg                  <NA>
#> 44                names_criticalservice_pctile                  <NA>
#> 45             names_criticalservice_state_avg                  <NA>
#> 46          names_criticalservice_state_pctile                  <NA>
#> 47                                     names_d                  <NA>
#> 48                                 names_d_avg                  <NA>
#> 49                               names_d_count                  <NA>
#> 50                     names_d_demogindexstate                  <NA>
#> 51                               names_d_extra                  <NA>
#> 52                         names_d_extra_count                  <NA>
#> 53                            names_d_language                  <NA>
#> 54                      names_d_language_count                  <NA>
#> 55                          names_d_languageli                  <NA>
#> 56                    names_d_languageli_count                  <NA>
#> 57                         names_d_other_count                  <NA>
#> 58                              names_d_pctile                  <NA>
#> 59                        names_d_ratio_to_avg                  <NA>
#> 60                  names_d_ratio_to_state_avg                  <NA>
#> 61                           names_d_state_avg                  <NA>
#> 62                        names_d_state_pctile                  <NA>
#> 63                           names_d_subgroups                  <NA>
#> 64                     names_d_subgroups_alone                  <NA>
#> 65                 names_d_subgroups_alone_avg                  <NA>
#> 66               names_d_subgroups_alone_count                  <NA>
#> 67              names_d_subgroups_alone_pctile                  <NA>
#> 68        names_d_subgroups_alone_ratio_to_avg                  <NA>
#> 69  names_d_subgroups_alone_ratio_to_state_avg                  <NA>
#> 70           names_d_subgroups_alone_state_avg                  <NA>
#> 71        names_d_subgroups_alone_state_pctile                  <NA>
#> 72                       names_d_subgroups_avg                  <NA>
#> 73                     names_d_subgroups_count                  <NA>
#> 74                        names_d_subgroups_nh                  <NA>
#> 75                    names_d_subgroups_nh_avg                  <NA>
#> 76                  names_d_subgroups_nh_count                  <NA>
#> 77                 names_d_subgroups_nh_pctile                  <NA>
#> 78           names_d_subgroups_nh_ratio_to_avg                  <NA>
#> 79     names_d_subgroups_nh_ratio_to_state_avg                  <NA>
#> 80              names_d_subgroups_nh_state_avg                  <NA>
#> 81           names_d_subgroups_nh_state_pctile                  <NA>
#> 82                    names_d_subgroups_pctile                  <NA>
#> 83              names_d_subgroups_ratio_to_avg                  <NA>
#> 84        names_d_subgroups_ratio_to_state_avg                  <NA>
#> 85                 names_d_subgroups_state_avg                  <NA>
#> 86              names_d_subgroups_state_pctile                  <NA>
#> 87                                     names_e                  <NA>
#> 88                                 names_e_avg                  <NA>
#> 89                               names_e_other                  <NA>
#> 90                              names_e_pctile                  <NA>
#> 91                        names_e_ratio_to_avg                  <NA>
#> 92                  names_e_ratio_to_state_avg                  <NA>
#> 93                           names_e_state_avg                  <NA>
#> 94                        names_e_state_pctile                  <NA>
#> 95                                    names_ej                  <NA>
#> 96                             names_ej_pctile                  <NA>
#> 97                              names_ej_state                  <NA>
#> 98                       names_ej_state_pctile                  <NA>
#> 99                               names_ej_supp                  <NA>
#> 100                       names_ej_supp_pctile                  <NA>
#> 101                        names_ej_supp_state                  <NA>
#> 102                 names_ej_supp_state_pctile                  <NA>
#> 103                       names_featuresinarea                  <NA>
#> 104                                 names_flag                  <NA>
#> 105                                  names_geo                  <NA>
#> 106                               names_health                  <NA>
#> 107                           names_health_avg                  <NA>
#> 108                         names_health_count                  <NA>
#> 109                        names_health_pctile                  <NA>
#> 110                  names_health_ratio_to_avg                  <NA>
#> 111            names_health_ratio_to_state_avg                  <NA>
#> 112                     names_health_state_avg                  <NA>
#> 113                  names_health_state_pctile                  <NA>
#> 114                                 names_misc                  <NA>
#> 115      names_pct_as_fraction_blockgroupstats                  <NA>
#> 116               names_pct_as_fraction_ejamit                  <NA>
#> 117           names_pct_as_fraction_ejscreenit                  <NA>
#> 118                          names_sitesinarea                  <NA>
#> 119                                names_these                  <NA>
#> 120                            names_these_avg                  <NA>
#> 121                   names_these_ratio_to_avg                  <NA>
#> 122             names_these_ratio_to_state_avg                  <NA>
#> 123                      names_these_state_avg                  <NA>
#> 124                                  names_wts                  <NA>
#> 126                                   sictable                  <NA>
#> 129                           states_shapefile            2024-08-22
#> 131                        testinput_address_2                  <NA>
#> 132                        testinput_address_9                  <NA>
#> 133                    testinput_address_parts                  <NA>
#> 138                 testinput_fips_blockgroups                  <NA>
#> 139                      testinput_fips_cities                  <NA>
#> 140                    testinput_fips_counties                  <NA>
#> 141                         testinput_fips_mix                  <NA>
#> 142                      testinput_fips_states                  <NA>
#> 143                      testinput_fips_tracts                  <NA>
#> 144                             testinput_mact                  <NA>
#> 145                            testinput_naics                  <NA>
#> 146                     testinput_program_name                  <NA>
#> 147                   testinput_program_sys_id            2025-02-12
#> 148                            testinput_regid                  <NA>
#> 149                      testinput_registry_id                  <NA>
#> 151                              testinput_sic                  <NA>
#> 170                              testpoints_10            2025-09-19
#> 171                             testpoints_100            2025-09-19
#> 172                            testpoints_1000            2025-09-19
#> 173                           testpoints_10000            2025-09-19
#> 174                          testpoints_100_dt            2025-02-14
#> 180                               testshapes_2            2025-05-30
#> 182                                 x_anyother                  <NA>
#>     acs_version
#> 1          <NA>
#> 2          <NA>
#> 4          <NA>
#> 5          <NA>
#> 7          <NA>
#> 9          <NA>
#> 10         <NA>
#> 11         <NA>
#> 13         <NA>
#> 17         <NA>
#> 19         <NA>
#> 20         <NA>
#> 21         <NA>
#> 22         <NA>
#> 23         <NA>
#> 24         <NA>
#> 25         <NA>
#> 26         <NA>
#> 27         <NA>
#> 28         <NA>
#> 29         <NA>
#> 30         <NA>
#> 31         <NA>
#> 32         <NA>
#> 33         <NA>
#> 34         <NA>
#> 35         <NA>
#> 36         <NA>
#> 37         <NA>
#> 38         <NA>
#> 39         <NA>
#> 40         <NA>
#> 41         <NA>
#> 42         <NA>
#> 43         <NA>
#> 44         <NA>
#> 45         <NA>
#> 46         <NA>
#> 47         <NA>
#> 48         <NA>
#> 49         <NA>
#> 50         <NA>
#> 51         <NA>
#> 52         <NA>
#> 53         <NA>
#> 54         <NA>
#> 55         <NA>
#> 56         <NA>
#> 57         <NA>
#> 58         <NA>
#> 59         <NA>
#> 60         <NA>
#> 61         <NA>
#> 62         <NA>
#> 63         <NA>
#> 64         <NA>
#> 65         <NA>
#> 66         <NA>
#> 67         <NA>
#> 68         <NA>
#> 69         <NA>
#> 70         <NA>
#> 71         <NA>
#> 72         <NA>
#> 73         <NA>
#> 74         <NA>
#> 75         <NA>
#> 76         <NA>
#> 77         <NA>
#> 78         <NA>
#> 79         <NA>
#> 80         <NA>
#> 81         <NA>
#> 82         <NA>
#> 83         <NA>
#> 84         <NA>
#> 85         <NA>
#> 86         <NA>
#> 87         <NA>
#> 88         <NA>
#> 89         <NA>
#> 90         <NA>
#> 91         <NA>
#> 92         <NA>
#> 93         <NA>
#> 94         <NA>
#> 95         <NA>
#> 96         <NA>
#> 97         <NA>
#> 98         <NA>
#> 99         <NA>
#> 100        <NA>
#> 101        <NA>
#> 102        <NA>
#> 103        <NA>
#> 104        <NA>
#> 105        <NA>
#> 106        <NA>
#> 107        <NA>
#> 108        <NA>
#> 109        <NA>
#> 110        <NA>
#> 111        <NA>
#> 112        <NA>
#> 113        <NA>
#> 114        <NA>
#> 115        <NA>
#> 116        <NA>
#> 117        <NA>
#> 118        <NA>
#> 119        <NA>
#> 120        <NA>
#> 121        <NA>
#> 122        <NA>
#> 123        <NA>
#> 124        <NA>
#> 126        <NA>
#> 129        <NA>
#> 131        <NA>
#> 132        <NA>
#> 133        <NA>
#> 138        <NA>
#> 139        <NA>
#> 140        <NA>
#> 141        <NA>
#> 142        <NA>
#> 143        <NA>
#> 144        <NA>
#> 145        <NA>
#> 146        <NA>
#> 147        <NA>
#> 148        <NA>
#> 149        <NA>
#> 151        <NA>
#> 170        <NA>
#> 171        <NA>
#> 172        <NA>
#> 173        <NA>
#> 174        <NA>
#> 180        <NA>
#> 182        <NA>
  x[x$has_metadata == TRUE, ]
#>     package                                      item ejam_package_version
#> 1      EJAM                                     NAICS               2.32.6
#> 2      EJAM                                       SIC               2.32.6
#> 3      EJAM                                 avg.in.us               2.32.6
#> 4      EJAM                             bg_cenpop2020               2.32.6
#> 5      EJAM                                     bgpts               2.32.6
#> 6      EJAM                           blockgroupstats               2.32.6
#> 7      EJAM                              censusplaces               2.32.6
#> 8      EJAM                                    custom               2.32.6
#> 9      EJAM           default_points_shown_at_startup               2.32.6
#> 12     EJAM        ejscreenRESTbroker2table_na_filler               2.32.6
#> 13     EJAM                              epa_programs               2.32.6
#> 14     EJAM                      epa_programs_defined               2.32.6
#> 15     EJAM                              formulas_all               2.32.6
#> 16     EJAM                                formulas_d               2.32.6
#> 17     EJAM                           frsprogramcodes               2.32.6
#> 18     EJAM                high_pctiles_tied_with_min               2.32.6
#> 19     EJAM                               islandareas               2.32.6
#> 22     EJAM                                mact_table               2.32.6
#> 23     EJAM                           map_headernames               2.32.6
#> 24     EJAM                             metadata4pins               2.32.6
#> 26     EJAM                          modelDoaggregate               2.32.6
#> 27     EJAM                               modelEjamit               2.32.6
#> 28     EJAM                              naics_counts               2.32.6
#> 29     EJAM                                naicstable               2.32.6
#> 125    EJAM                                     namez               2.32.6
#> 126    EJAM                                  sictable               2.32.6
#> 127    EJAM                                 stateinfo               2.32.6
#> 128    EJAM                                stateinfo2               2.32.6
#> 129    EJAM                          states_shapefile               2.32.6
#> 130    EJAM                                statestats               2.32.6
#> 134    EJAM                   testinput_address_table               2.32.6
#> 135    EJAM                 testinput_address_table_9               2.32.6
#> 136    EJAM         testinput_address_table_goodnames               2.32.6
#> 137    EJAM          testinput_address_table_withfull               2.32.6
#> 147    EJAM                  testinput_program_sys_id               2.32.6
#> 150    EJAM                        testinput_shapes_2               2.32.6
#> 152    EJAM     testoutput_doaggregate_1000pts_1miles               2.32.6
#> 153    EJAM      testoutput_doaggregate_100pts_1miles               2.32.6
#> 154    EJAM       testoutput_doaggregate_10pts_1miles               2.32.6
#> 155    EJAM          testoutput_ejamit_1000pts_1miles               2.32.6
#> 156    EJAM           testoutput_ejamit_100pts_1miles               2.32.6
#> 157    EJAM            testoutput_ejamit_10pts_1miles               2.32.6
#> 158    EJAM             testoutput_ejamit_fips_cities               2.32.6
#> 159    EJAM           testoutput_ejamit_fips_counties               2.32.6
#> 160    EJAM                testoutput_ejamit_shapes_2               2.32.6
#> 161    EJAM testoutput_ejscreenRESTbroker_1pts_1miles               2.32.6
#> 162    EJAM        testoutput_ejscreenapi_1pts_1miles               2.32.6
#> 163    EJAM             testoutput_ejscreenapi_plus_5               2.32.6
#> 164    EJAM                   testoutput_ejscreenit_5               2.32.6
#> 165    EJAM                  testoutput_ejscreenit_50               2.32.6
#> 166    EJAM                 testoutput_ejscreenit_500               2.32.6
#> 167    EJAM testoutput_getblocksnearby_1000pts_1miles               2.32.6
#> 168    EJAM  testoutput_getblocksnearby_100pts_1miles               2.32.6
#> 169    EJAM   testoutput_getblocksnearby_10pts_1miles               2.32.6
#> 170    EJAM                             testpoints_10               2.32.6
#> 171    EJAM                            testpoints_100               2.32.6
#> 172    EJAM                           testpoints_1000               2.32.6
#> 173    EJAM                          testpoints_10000               2.32.6
#> 174    EJAM                         testpoints_100_dt               2.32.6
#> 175    EJAM                              testpoints_5               2.32.6
#> 176    EJAM                             testpoints_50               2.32.6
#> 177    EJAM                            testpoints_500               2.32.6
#> 178    EJAM                            testpoints_bad               2.32.6
#> 179    EJAM                       testpoints_overlap3               2.32.6
#> 180    EJAM                              testshapes_2               2.32.6
#> 181    EJAM                                  usastats               2.32.6
#> 182    EJAM                                x_anyother               2.32.6
#>     date_saved_in_package date_downloaded ejscreen_version ejscreen_releasedate
#> 1                    <NA>            <NA>             <NA>                 <NA>
#> 2              2025-02-12            <NA>             <NA>                 <NA>
#> 3              2024-10-12            <NA>             2.32           2024-08-12
#> 4              2024-08-22            <NA>             <NA>                 <NA>
#> 5              2024-08-22            <NA>             <NA>                 <NA>
#> 6              2024-12-13            <NA>             2.32           2024-08-12
#> 7              2024-09-26            <NA>             <NA>                 <NA>
#> 8              2025-08-26            <NA>             2.32           2024-08-12
#> 9                    <NA>            <NA>             <NA>                 <NA>
#> 12             2024-12-11            <NA>             2.32           2024-08-12
#> 13             2025-02-12            <NA>             <NA>                 <NA>
#> 14             2024-11-15            <NA>             2.32           2024-08-12
#> 15             2025-05-06            <NA>             2.32           2024-08-12
#> 16             2025-05-06            <NA>             2.32           2024-08-12
#> 17             2025-02-12            <NA>             <NA>                 <NA>
#> 18             2024-10-12            <NA>             2.32           2024-08-12
#> 19                   <NA>            <NA>             <NA>                 <NA>
#> 22             2025-02-12            <NA>             <NA>                 <NA>
#> 23             2025-08-26            <NA>             2.32           2024-08-12
#> 24                   <NA>            <NA>             <NA>                 <NA>
#> 26                   <NA>            <NA>             <NA>                 <NA>
#> 27                   <NA>            <NA>             <NA>                 <NA>
#> 28             2025-02-12            <NA>             <NA>                 <NA>
#> 29             2024-10-12            <NA>             <NA>                 <NA>
#> 125            2025-08-26            <NA>             2.32           2024-08-12
#> 126                  <NA>            <NA>             <NA>                 <NA>
#> 127            2024-08-22            <NA>             2.32           2024-08-12
#> 128            2024-08-22            <NA>             2.32           2024-08-12
#> 129            2024-08-22      2024-08-22             <NA>                 <NA>
#> 130            2024-08-22            <NA>             2.32           2024-08-12
#> 134            2025-05-01            <NA>             2.32           2024-08-12
#> 135            2025-05-01            <NA>             2.32           2024-08-12
#> 136            2025-05-01            <NA>             2.32           2024-08-12
#> 137            2025-05-01            <NA>             2.32           2024-08-12
#> 147            2025-02-12            <NA>             <NA>                 <NA>
#> 150            2025-05-30            <NA>             2.32           2024-08-12
#> 152            2025-09-22            <NA>             2.32           2024-08-12
#> 153            2025-09-22            <NA>             2.32           2024-08-12
#> 154            2025-09-22            <NA>             2.32           2024-08-12
#> 155            2025-09-19            <NA>             2.32           2024-08-12
#> 156            2025-09-19            <NA>             2.32           2024-08-12
#> 157            2025-09-19            <NA>             2.32           2024-08-12
#> 158            2025-09-23            <NA>             2.32           2024-08-12
#> 159            2025-09-23            <NA>             2.32           2024-08-12
#> 160            2025-09-23            <NA>             2.32           2024-08-12
#> 161            2024-12-11            <NA>             2.32           2024-08-12
#> 162            2024-12-11            <NA>             2.32           2024-08-12
#> 163            2024-12-11            <NA>             2.32           2024-08-12
#> 164            2024-12-11            <NA>             2.32           2024-08-12
#> 165            2024-12-11            <NA>             2.32           2024-08-12
#> 166            2024-12-11            <NA>             2.32           2024-08-12
#> 167            2025-09-19            <NA>             2.32           2024-08-12
#> 168            2025-09-19            <NA>             2.32           2024-08-12
#> 169            2025-09-19            <NA>             2.32           2024-08-12
#> 170            2025-09-19            <NA>             <NA>                 <NA>
#> 171            2025-09-19            <NA>             <NA>                 <NA>
#> 172            2025-09-19            <NA>             <NA>                 <NA>
#> 173            2025-09-19            <NA>             <NA>                 <NA>
#> 174            2025-02-14            <NA>             <NA>                 <NA>
#> 175            2024-09-20            <NA>             2.32           2024-08-12
#> 176            2024-09-20            <NA>             2.32           2024-08-12
#> 177            2024-09-20            <NA>             2.32           2024-08-12
#> 178            2025-09-19            <NA>             2.32           2024-08-12
#> 179            2024-09-05            <NA>             2.32           2024-08-12
#> 180            2025-05-30            <NA>             <NA>                 <NA>
#> 181            2024-08-22            <NA>             2.32           2024-08-12
#> 182                  <NA>            <NA>             <NA>                 <NA>
#>     acs_releasedate acs_version census_version has_metadata
#> 1              <NA>        <NA>           <NA>         TRUE
#> 2              <NA>        <NA>           <NA>         TRUE
#> 3        2023-12-07   2018-2022           2020         TRUE
#> 4              <NA>        <NA>           2020         TRUE
#> 5              <NA>        <NA>           <NA>         TRUE
#> 6        2023-12-07   2018-2022           2020         TRUE
#> 7              <NA>        <NA>           <NA>         TRUE
#> 8        2023-12-07   2018-2022           2020         TRUE
#> 9              <NA>        <NA>           <NA>         TRUE
#> 12       2023-12-07   2018-2022           2020         TRUE
#> 13             <NA>        <NA>           <NA>         TRUE
#> 14       2023-12-07   2018-2022           2020         TRUE
#> 15       2023-12-07   2018-2022           2020         TRUE
#> 16       2023-12-07   2018-2022           2020         TRUE
#> 17             <NA>        <NA>           <NA>         TRUE
#> 18       2023-12-07   2018-2022           2020         TRUE
#> 19             <NA>        <NA>           <NA>         TRUE
#> 22             <NA>        <NA>           <NA>         TRUE
#> 23             <NA>        <NA>           <NA>         TRUE
#> 24             <NA>        <NA>           <NA>         TRUE
#> 26             <NA>        <NA>           <NA>         TRUE
#> 27             <NA>        <NA>           <NA>         TRUE
#> 28             <NA>        <NA>           <NA>         TRUE
#> 29             <NA>        <NA>           <NA>         TRUE
#> 125      2023-12-07   2018-2022           2020         TRUE
#> 126            <NA>        <NA>           <NA>         TRUE
#> 127      2023-12-07   2018-2022           2020         TRUE
#> 128      2023-12-07   2018-2022           2020         TRUE
#> 129            <NA>        <NA>           <NA>         TRUE
#> 130      2023-12-07   2018-2022           2020         TRUE
#> 134      2023-12-07   2018-2022           2020         TRUE
#> 135      2023-12-07   2018-2022           2020         TRUE
#> 136      2023-12-07   2018-2022           2020         TRUE
#> 137      2023-12-07   2018-2022           2020         TRUE
#> 147            <NA>        <NA>           <NA>         TRUE
#> 150      2023-12-07   2018-2022           2020         TRUE
#> 152      2023-12-07   2018-2022           2020         TRUE
#> 153      2023-12-07   2018-2022           2020         TRUE
#> 154      2023-12-07   2018-2022           2020         TRUE
#> 155      2023-12-07   2018-2022           2020         TRUE
#> 156      2023-12-07   2018-2022           2020         TRUE
#> 157      2023-12-07   2018-2022           2020         TRUE
#> 158      2023-12-07   2018-2022           2020         TRUE
#> 159      2023-12-07   2018-2022           2020         TRUE
#> 160      2023-12-07   2018-2022           2020         TRUE
#> 161      2023-12-07   2018-2022           2020         TRUE
#> 162      2023-12-07   2018-2022           2020         TRUE
#> 163      2023-12-07   2018-2022           2020         TRUE
#> 164      2023-12-07   2018-2022           2020         TRUE
#> 165      2023-12-07   2018-2022           2020         TRUE
#> 166      2023-12-07   2018-2022           2020         TRUE
#> 167      2023-12-07   2018-2022           2020         TRUE
#> 168      2023-12-07   2018-2022           2020         TRUE
#> 169      2023-12-07   2018-2022           2020         TRUE
#> 170            <NA>        <NA>           <NA>         TRUE
#> 171            <NA>        <NA>           <NA>         TRUE
#> 172            <NA>        <NA>           <NA>         TRUE
#> 173            <NA>        <NA>           <NA>         TRUE
#> 174            <NA>        <NA>           <NA>         TRUE
#> 175      2023-12-07   2018-2022           2020         TRUE
#> 176      2023-12-07   2018-2022           2020         TRUE
#> 177      2023-12-07   2018-2022           2020         TRUE
#> 178      2023-12-07   2018-2022           2020         TRUE
#> 179      2023-12-07   2018-2022           2020         TRUE
#> 180            <NA>        <NA>           <NA>         TRUE
#> 181      2023-12-07   2018-2022           2020         TRUE
#> 182            <NA>        <NA>           <NA>         TRUE
  table(x$has_metadata)
#> 
#> FALSE  TRUE 
#>   115    67