for query term, show list of roughly matching NAICS, scraped from web

This finds more than just naics_from_any() does, since that needs an exact match but this looks at naics.com website which lists various aliases for a sector.

Usage

naics_findwebscrape(query)

Arguments

query

text like "gasoline" or "copper smelting"

Value

data.frame of info on what was found, naics and title

See also

Examples

 # naics_from_any("copper smelting")
 # naics_from_any("copper smelting", website_scrape=TRUE)
 # browseURL(naics_from_any("copper smelting", website_url=TRUE) )

  url_naics.com("copper smelting")
#> [1] "https://www.naics.com/code-search/?trms=copper+smelting&v=2017&styp=naics"
  # \donttest{
  naics_findwebscrape("copper smelting")
#>      code
#> 1  331492
#> 2  331410
#> 3  331314
#> 4  331420
#> 5  331491
#> 6  212230
#> 7  423510
#> 8  331318
#> 9  331313
#> 10 332420
#> 11 331315
#> 12 333994
#> 13 423830
#> 14 331523
#> 15 331529
#> 16 325180
#> 17 238160
#> 18 325612
#> 19 332999
#> 20 325130
#> 21 332112
#>                                                                                           name
#> 1  Secondary Smelting, Refining, and Alloying of Nonferrous Metal (except Copper and Aluminum)
#> 2                                     Nonferrous Metal (except Aluminum) Smelting and Refining
#> 3                                                  Secondary Smelting and Alloying of Aluminum
#> 4                                             Copper Rolling, Drawing, Extruding, and Alloying
#> 5                Nonferrous Metal (except Copper and Aluminum) Rolling, Drawing, and Extruding
#> 6                                                        Copper, Nickel, Lead, and Zinc Mining
#> 7                                   Metal Service Centers and Other Metal Merchant Wholesalers
#> 8                                               Other Aluminum Rolling, Drawing, and Extruding
#> 9                                             Alumina Refining and Primary Aluminum Production
#> 10                                                      Metal Tank (Heavy Gauge) Manufacturing
#> 11                                               Aluminum Sheet, Plate, and Foil Manufacturing
#> 12                                           Industrial Process Furnace and Oven Manufacturing
#> 13                                     Industrial Machinery and Equipment Merchant Wholesalers
#> 14                                                      Nonferrous Metal Die-Casting Foundries
#> 15                                       Other Nonferrous Metal Foundries (except Die-Casting)
#> 16                                                Other Basic Inorganic Chemical Manufacturing
#> 17                                                                         Roofing Contractors
#> 18                                              Polish and Other Sanitation Good Manufacturing
#> 19                              All Other Miscellaneous Fabricated Metal Product Manufacturing
#> 20                                                     Synthetic Dye and Pigment Manufacturing
#> 21                                                                          Nonferrous Forging
  browseURL(url_naics.com("copper smelting"))
  browseURL(naics_url_of_code(326))
  # }