search for industry names and NAICS codes by query string

query by parts of words, etc. in the industry name.

naics_from_name(mynames, children = FALSE, ignore.case = TRUE, fixed = FALSE)

Arguments

mynames

query string, vector of NAICS industry names or any regular expression or partial words. See https://naics.com

children

logical, if TRUE, also return all the subcategories - where NAICS starts with the same digits

ignore.case

see grepl()

fixed

should it be an exact match? see grepl()

search_on_naics_website

whether to query on naics website for more hits than just search for text in industry title

Value

a subset of the naicstable data.table (not just the codes column)

Details

See naics_from_any() which uses this

See also

Examples

 data.table::fintersect(naics_from_any( "manufac"), naics_from_any("chem"))
#> Warning: NAs introduced by coercion
#> Warning: mynames should be non-NA character vector of text to look for in industry title(s) like "concrete"
#> Warning: NAs introduced by coercion
#> Warning: mynames should be non-NA character vector of text to look for in industry title(s) like "concrete"
#>       code n2  n3   n4    n5     n6
#>  1:    325 32 325  325   325    325
#>  2:   3251 32 325 3251  3251   3251
#>  3:  32511 32 325 3251 32511  32511
#>  4: 325110 32 325 3251 32511 325110
#>  5:  32518 32 325 3251 32518  32518
#>  6: 325180 32 325 3251 32518 325180
#>  7:  32519 32 325 3251 32519  32519
#>  8: 325194 32 325 3251 32519 325194
#>  9: 325199 32 325 3251 32519 325199
#> 10:   3253 32 325 3253  3253   3253
#> 11:  32532 32 325 3253 32532  32532
#> 12: 325320 32 325 3253 32532 325320
#> 13:   3259 32 325 3259  3259   3259
#> 14:  32599 32 325 3259 32599  32599
#> 15: 325992 32 325 3259 32599 325992
#> 16: 325998 32 325 3259 32599 325998
#>                                                                       name
#>  1:                                                 Chemical Manufacturing
#>  2:                                           Basic Chemical Manufacturing
#>  3:                                            Petrochemical Manufacturing
#>  4:                                            Petrochemical Manufacturing
#>  5:                           Other Basic Inorganic Chemical Manufacturing
#>  6:                           Other Basic Inorganic Chemical Manufacturing
#>  7:                             Other Basic Organic Chemical Manufacturing
#>  8:    Cyclic Crude, Intermediate, and Gum and Wood Chemical Manufacturing
#>  9:                         All Other Basic Organic Chemical Manufacturing
#> 10:   Pesticide, Fertilizer, and Other Agricultural Chemical Manufacturing
#> 11:                Pesticide and Other Agricultural Chemical Manufacturing
#> 12:                Pesticide and Other Agricultural Chemical Manufacturing
#> 13:                   Other Chemical Product and Preparation Manufacturing
#> 14:               All Other Chemical Product and Preparation Manufacturing
#> 15:            Photographic Film, Paper, Plate, and Chemical Manufacturing
#> 16: All Other Miscellaneous Chemical Product and Preparation Manufacturing
#>                                                                            num_name
#>  1:                                                    325 - Chemical Manufacturing
#>  2:                                             3251 - Basic Chemical Manufacturing
#>  3:                                             32511 - Petrochemical Manufacturing
#>  4:                                            325110 - Petrochemical Manufacturing
#>  5:                            32518 - Other Basic Inorganic Chemical Manufacturing
#>  6:                           325180 - Other Basic Inorganic Chemical Manufacturing
#>  7:                              32519 - Other Basic Organic Chemical Manufacturing
#>  8:    325194 - Cyclic Crude, Intermediate, and Gum and Wood Chemical Manufacturing
#>  9:                         325199 - All Other Basic Organic Chemical Manufacturing
#> 10:     3253 - Pesticide, Fertilizer, and Other Agricultural Chemical Manufacturing
#> 11:                 32532 - Pesticide and Other Agricultural Chemical Manufacturing
#> 12:                325320 - Pesticide and Other Agricultural Chemical Manufacturing
#> 13:                     3259 - Other Chemical Product and Preparation Manufacturing
#> 14:                32599 - All Other Chemical Product and Preparation Manufacturing
#> 15:            325992 - Photographic Film, Paper, Plate, and Chemical Manufacturing
#> 16: 325998 - All Other Miscellaneous Chemical Product and Preparation Manufacturing