See which of the lists of names a single term appears in

See which of the lists of names a single term appears in

Usage

names_whichlist(
  x,
  exact = T,
  grepmatching = T,
  ignore.case.exact = FALSE,
  ignore.case.grep = FALSE,
  keylists = F,
  exactonly = FALSE
)

Arguments

x

term, like part or all of a variable name, such as state.avg

exact

whether to look for exact matches

grepmatching

whether to look for matches via grep (partial match)

ignore.case.exact

whether to ignore capitalization in exact matches

ignore.case.grep

passed to grep as ignore.case param

keylists

if true, only report for the key lists not friendly, all, these, need types.

exactonly

to limit output to rows with exact matches

Value

a data.frame of whichlist, exactmatch, grepmatch, and grephits (examples)

Details

EJAM::namez has a list of lists of names used for indicators or variables, such as namez$d_friendly which is a vector of terms like "Demog.Ind.", "Suppl Demog Index", "% Low-inc.", etc.

Examples

 
   x <- names_whichlist("rsei", ignore.case.exact = T, ignore.case.grep = T)
   subset(x, !grepl("friendly", x$whichlist))
#>                                whichlist exactmatch grepmatch
#> 1                                  all_r        yes       yes
#> 2                                    wts                     
#> 3                                      d                     
#> 5                                  d_avg                     
#> 6                            d_state_avg                     
#> 9                               d_pctile                     
#> 10                        d_state_pctile                     
#> 13                        d_ratio_to_avg                     
#> 14                  d_ratio_to_state_avg                     
#> 17                               d_count                     
#> 19                         d_other_count                     
#> 21                           d_subgroups                     
#> 23                       d_subgroups_avg                     
#> 24                 d_subgroups_state_avg                     
#> 27                    d_subgroups_pctile                     
#> 28              d_subgroups_state_pctile                     
#> 31              d_subgroups_ratio_to_avg                     
#> 32        d_subgroups_ratio_to_state_avg                     
#> 35                     d_subgroups_count                     
#> 36                        d_subgroups_nh                     
#> 38                    d_subgroups_nh_avg                     
#> 39              d_subgroups_nh_state_avg                     
#> 42                 d_subgroups_nh_pctile                     
#> 43           d_subgroups_nh_state_pctile                     
#> 46           d_subgroups_nh_ratio_to_avg                     
#> 47     d_subgroups_nh_ratio_to_state_avg                     
#> 50                  d_subgroups_nh_count                     
#> 51                     d_subgroups_alone                     
#> 53                 d_subgroups_alone_avg                     
#> 54           d_subgroups_alone_state_avg                     
#> 57              d_subgroups_alone_pctile                     
#> 58        d_subgroups_alone_state_pctile                     
#> 61        d_subgroups_alone_ratio_to_avg                     
#> 62  d_subgroups_alone_ratio_to_state_avg                     
#> 65               d_subgroups_alone_count                     
#> 66                                     e        yes       yes
#> 68                                 e_avg                  yes
#> 69                           e_state_avg                  yes
#> 72                              e_pctile                  yes
#> 73                        e_state_pctile                  yes
#> 76                        e_ratio_to_avg                  yes
#> 77                  e_ratio_to_state_avg                  yes
#> 80                                    ej                  yes
#> 81                              ej_state                  yes
#> 84                               ej_supp                  yes
#> 85                         ej_supp_state                  yes
#> 88                             ej_pctile                  yes
#> 89                       ej_state_pctile                  yes
#> 92                        ej_supp_pctile                  yes
#> 93                  ej_supp_state_pctile                  yes
#> 96                           need_pctile        yes       yes
#> 97                     need_state_pctile        yes       yes
#> 98                                 these        yes       yes
#> 99                             these_avg                  yes
#> 100                      these_state_avg                  yes
#> 101                   these_ratio_to_avg                  yes
#> 102             these_ratio_to_state_avg                  yes
#>                                               grephits
#> 1                 rsei, avg.rsei, state.avg.rsei, etc.
#> 2                                                     
#> 3                                                     
#> 5                                                     
#> 6                                                     
#> 9                                                     
#> 10                                                    
#> 13                                                    
#> 14                                                    
#> 17                                                    
#> 19                                                    
#> 21                                                    
#> 23                                                    
#> 24                                                    
#> 27                                                    
#> 28                                                    
#> 31                                                    
#> 32                                                    
#> 35                                                    
#> 36                                                    
#> 38                                                    
#> 39                                                    
#> 42                                                    
#> 43                                                    
#> 46                                                    
#> 47                                                    
#> 50                                                    
#> 51                                                    
#> 53                                                    
#> 54                                                    
#> 57                                                    
#> 58                                                    
#> 61                                                    
#> 62                                                    
#> 65                                                    
#> 66                                                rsei
#> 68                                            avg.rsei
#> 69                                      state.avg.rsei
#> 72                                         pctile.rsei
#> 73                                   state.pctile.rsei
#> 76                                   ratio.to.avg.rsei
#> 77                             ratio.to.state.avg.rsei
#> 80                                EJ.DISPARITY.rsei.eo
#> 81                          state.EJ.DISPARITY.rsei.eo
#> 84                              EJ.DISPARITY.rsei.supp
#> 85                        state.EJ.DISPARITY.rsei.supp
#> 88                         pctile.EJ.DISPARITY.rsei.eo
#> 89                   state.pctile.EJ.DISPARITY.rsei.eo
#> 92                       pctile.EJ.DISPARITY.rsei.supp
#> 93                 state.pctile.EJ.DISPARITY.rsei.supp
#> 96  rsei, EJ.DISPARITY.rsei.eo, EJ.DISPARITY.rsei.supp
#> 97  rsei, EJ.DISPARITY.rsei.eo, EJ.DISPARITY.rsei.supp
#> 98                                                rsei
#> 99                                            avg.rsei
#> 100                                     state.avg.rsei
#> 101                                  ratio.to.avg.rsei
#> 102                            ratio.to.state.avg.rsei
   
   subset(x,  grepl("friendly", x$whichlist))
#>                                        whichlist exactmatch grepmatch
#> 4                                     d_friendly                     
#> 7                                 d_avg_friendly                     
#> 8                           d_state_avg_friendly                     
#> 11                             d_pctile_friendly                     
#> 12                       d_state_pctile_friendly                     
#> 15                       d_ratio_to_avg_friendly                     
#> 16                 d_ratio_to_state_avg_friendly                     
#> 18                              d_count_friendly                     
#> 20                        d_other_count_friendly                     
#> 22                          d_subgroups_friendly                     
#> 25                      d_subgroups_avg_friendly                     
#> 26                d_subgroups_state_avg_friendly                     
#> 29                   d_subgroups_pctile_friendly                     
#> 30             d_subgroups_state_pctile_friendly                     
#> 33             d_subgroups_ratio_to_avg_friendly                     
#> 34       d_subgroups_ratio_to_state_avg_friendly                     
#> 37                       d_subgroups_nh_friendly                     
#> 40                   d_subgroups_nh_avg_friendly                     
#> 41             d_subgroups_nh_state_avg_friendly                     
#> 44                d_subgroups_nh_pctile_friendly                     
#> 45          d_subgroups_nh_state_pctile_friendly                     
#> 48          d_subgroups_nh_ratio_to_avg_friendly                     
#> 49    d_subgroups_nh_ratio_to_state_avg_friendly                     
#> 52                    d_subgroups_alone_friendly                     
#> 55                d_subgroups_alone_avg_friendly                     
#> 56          d_subgroups_alone_state_avg_friendly                     
#> 59             d_subgroups_alone_pctile_friendly                     
#> 60       d_subgroups_alone_state_pctile_friendly                     
#> 63       d_subgroups_alone_ratio_to_avg_friendly                     
#> 64 d_subgroups_alone_ratio_to_state_avg_friendly                     
#> 67                                    e_friendly                  yes
#> 70                                e_avg_friendly                  yes
#> 71                          e_state_avg_friendly                  yes
#> 74                             e_pctile_friendly                  yes
#> 75                       e_state_pctile_friendly                  yes
#> 78                       e_ratio_to_avg_friendly                  yes
#> 79                 e_ratio_to_state_avg_friendly                  yes
#> 82                                   ej_friendly                  yes
#> 83                             ej_state_friendly                  yes
#> 86                              ej_supp_friendly                  yes
#> 87                        ej_supp_state_friendly                  yes
#> 90                            ej_pctile_friendly                  yes
#> 91                      ej_state_pctile_friendly                  yes
#> 94                       ej_supp_pctile_friendly                  yes
#> 95                 ej_supp_state_pctile_friendly                  yes
#>                               grephits
#> 4                                     
#> 7                                     
#> 8                                     
#> 11                                    
#> 12                                    
#> 15                                    
#> 16                                    
#> 18                                    
#> 20                                    
#> 22                                    
#> 25                                    
#> 26                                    
#> 29                                    
#> 30                                    
#> 33                                    
#> 34                                    
#> 37                                    
#> 40                                    
#> 41                                    
#> 44                                    
#> 45                                    
#> 48                                    
#> 49                                    
#> 52                                    
#> 55                                    
#> 56                                    
#> 59                                    
#> 60                                    
#> 63                                    
#> 64                                    
#> 67                            RSEI Air
#> 70                     US Avg RSEI Air
#> 71                  State Avg RSEI Air
#> 74          US percentile for RSEI Air
#> 75       State percentile for RSEI Air
#> 78            Ratio to US Avg RSEI Air
#> 79         Ratio to State Avg RSEI Air
#> 82                            EJ: RSEI
#> 83                      State EJ: RSEI
#> 86                      Supp. EJ: RSEI
#> 87                State Supp. EJ: RSEI
#> 90          US percentile for EJ: RSEI
#> 91       State percentile for EJ: RSEI
#> 94    US percentile for Supp. EJ: RSEI
#> 95 State percentile for Supp. EJ: RSEI
   subset(namez, names(namez) != "all_r" & names(namez) %in%
      subset(x, x$grepmatch == "yes" & !grepl("friendly", x$whichlist))$whichlist  )
#> $e
#>  [1] "pm"              "o3"              "cancer"          "resp"           
#>  [5] "dpm"             "pctpre1960"      "traffic.score"   "proximity.npl"  
#>  [9] "proximity.rmp"   "proximity.tsdf"  "proximity.npdes" "ust"            
#> [13] "rsei"           
#> 
#> $e_avg
#>  [1] "avg.pm"              "avg.o3"              "avg.cancer"         
#>  [4] "avg.resp"            "avg.dpm"             "avg.pctpre1960"     
#>  [7] "avg.traffic.score"   "avg.proximity.npl"   "avg.proximity.rmp"  
#> [10] "avg.proximity.tsdf"  "avg.proximity.npdes" "avg.ust"            
#> [13] "avg.rsei"           
#> 
#> $e_state_avg
#>  [1] "state.avg.pm"              "state.avg.o3"             
#>  [3] "state.avg.cancer"          "state.avg.resp"           
#>  [5] "state.avg.dpm"             "state.avg.pctpre1960"     
#>  [7] "state.avg.traffic.score"   "state.avg.proximity.npl"  
#>  [9] "state.avg.proximity.rmp"   "state.avg.proximity.tsdf" 
#> [11] "state.avg.proximity.npdes" "state.avg.ust"            
#> [13] "state.avg.rsei"           
#> 
#> $e_pctile
#>  [1] "pctile.pm"              "pctile.o3"              "pctile.cancer"         
#>  [4] "pctile.resp"            "pctile.dpm"             "pctile.pctpre1960"     
#>  [7] "pctile.traffic.score"   "pctile.proximity.npl"   "pctile.proximity.rmp"  
#> [10] "pctile.proximity.tsdf"  "pctile.proximity.npdes" "pctile.ust"            
#> [13] "pctile.rsei"           
#> 
#> $e_state_pctile
#>  [1] "state.pctile.pm"              "state.pctile.o3"             
#>  [3] "state.pctile.cancer"          "state.pctile.resp"           
#>  [5] "state.pctile.dpm"             "state.pctile.pctpre1960"     
#>  [7] "state.pctile.traffic.score"   "state.pctile.proximity.npl"  
#>  [9] "state.pctile.proximity.rmp"   "state.pctile.proximity.tsdf" 
#> [11] "state.pctile.proximity.npdes" "state.pctile.ust"            
#> [13] "state.pctile.rsei"           
#> 
#> $e_ratio_to_avg
#>  [1] "ratio.to.avg.pm"              "ratio.to.avg.o3"             
#>  [3] "ratio.to.avg.cancer"          "ratio.to.avg.resp"           
#>  [5] "ratio.to.avg.dpm"             "ratio.to.avg.pctpre1960"     
#>  [7] "ratio.to.avg.traffic.score"   "ratio.to.avg.proximity.npl"  
#>  [9] "ratio.to.avg.proximity.rmp"   "ratio.to.avg.proximity.tsdf" 
#> [11] "ratio.to.avg.proximity.npdes" "ratio.to.avg.ust"            
#> [13] "ratio.to.avg.rsei"           
#> 
#> $e_ratio_to_state_avg
#>  [1] "ratio.to.state.avg.pm"              "ratio.to.state.avg.o3"             
#>  [3] "ratio.to.state.avg.cancer"          "ratio.to.state.avg.resp"           
#>  [5] "ratio.to.state.avg.dpm"             "ratio.to.state.avg.pctpre1960"     
#>  [7] "ratio.to.state.avg.traffic.score"   "ratio.to.state.avg.proximity.npl"  
#>  [9] "ratio.to.state.avg.proximity.rmp"   "ratio.to.state.avg.proximity.tsdf" 
#> [11] "ratio.to.state.avg.proximity.npdes" "ratio.to.state.avg.ust"            
#> [13] "ratio.to.state.avg.rsei"           
#> 
#> $ej
#>  [1] "EJ.DISPARITY.pm.eo"              "EJ.DISPARITY.o3.eo"             
#>  [3] "EJ.DISPARITY.cancer.eo"          "EJ.DISPARITY.resp.eo"           
#>  [5] "EJ.DISPARITY.dpm.eo"             "EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "EJ.DISPARITY.traffic.score.eo"   "EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "EJ.DISPARITY.proximity.rmp.eo"   "EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "EJ.DISPARITY.proximity.npdes.eo" "EJ.DISPARITY.ust.eo"            
#> [13] "EJ.DISPARITY.rsei.eo"           
#> 
#> $ej_state
#>  [1] "state.EJ.DISPARITY.pm.eo"             
#>  [2] "state.EJ.DISPARITY.o3.eo"             
#>  [3] "state.EJ.DISPARITY.cancer.eo"         
#>  [4] "state.EJ.DISPARITY.resp.eo"           
#>  [5] "state.EJ.DISPARITY.dpm.eo"            
#>  [6] "state.EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "state.EJ.DISPARITY.traffic.score.eo"  
#>  [8] "state.EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "state.EJ.DISPARITY.proximity.rmp.eo"  
#> [10] "state.EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "state.EJ.DISPARITY.proximity.npdes.eo"
#> [12] "state.EJ.DISPARITY.ust.eo"            
#> [13] "state.EJ.DISPARITY.rsei.eo"           
#> 
#> $ej_supp
#>  [1] "EJ.DISPARITY.pm.supp"              "EJ.DISPARITY.o3.supp"             
#>  [3] "EJ.DISPARITY.cancer.supp"          "EJ.DISPARITY.resp.supp"           
#>  [5] "EJ.DISPARITY.dpm.supp"             "EJ.DISPARITY.pctpre1960.supp"     
#>  [7] "EJ.DISPARITY.traffic.score.supp"   "EJ.DISPARITY.proximity.npl.supp"  
#>  [9] "EJ.DISPARITY.proximity.rmp.supp"   "EJ.DISPARITY.proximity.tsdf.supp" 
#> [11] "EJ.DISPARITY.proximity.npdes.supp" "EJ.DISPARITY.ust.supp"            
#> [13] "EJ.DISPARITY.rsei.supp"           
#> 
#> $ej_supp_state
#>  [1] "state.EJ.DISPARITY.pm.supp"             
#>  [2] "state.EJ.DISPARITY.o3.supp"             
#>  [3] "state.EJ.DISPARITY.cancer.supp"         
#>  [4] "state.EJ.DISPARITY.resp.supp"           
#>  [5] "state.EJ.DISPARITY.dpm.supp"            
#>  [6] "state.EJ.DISPARITY.pctpre1960.supp"     
#>  [7] "state.EJ.DISPARITY.traffic.score.supp"  
#>  [8] "state.EJ.DISPARITY.proximity.npl.supp"  
#>  [9] "state.EJ.DISPARITY.proximity.rmp.supp"  
#> [10] "state.EJ.DISPARITY.proximity.tsdf.supp" 
#> [11] "state.EJ.DISPARITY.proximity.npdes.supp"
#> [12] "state.EJ.DISPARITY.ust.supp"            
#> [13] "state.EJ.DISPARITY.rsei.supp"           
#> 
#> $ej_pctile
#>  [1] "pctile.EJ.DISPARITY.pm.eo"             
#>  [2] "pctile.EJ.DISPARITY.o3.eo"             
#>  [3] "pctile.EJ.DISPARITY.cancer.eo"         
#>  [4] "pctile.EJ.DISPARITY.resp.eo"           
#>  [5] "pctile.EJ.DISPARITY.dpm.eo"            
#>  [6] "pctile.EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "pctile.EJ.DISPARITY.traffic.score.eo"  
#>  [8] "pctile.EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "pctile.EJ.DISPARITY.proximity.rmp.eo"  
#> [10] "pctile.EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "pctile.EJ.DISPARITY.proximity.npdes.eo"
#> [12] "pctile.EJ.DISPARITY.ust.eo"            
#> [13] "pctile.EJ.DISPARITY.rsei.eo"           
#> 
#> $ej_state_pctile
#>  [1] "state.pctile.EJ.DISPARITY.pm.eo"             
#>  [2] "state.pctile.EJ.DISPARITY.o3.eo"             
#>  [3] "state.pctile.EJ.DISPARITY.cancer.eo"         
#>  [4] "state.pctile.EJ.DISPARITY.resp.eo"           
#>  [5] "state.pctile.EJ.DISPARITY.dpm.eo"            
#>  [6] "state.pctile.EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "state.pctile.EJ.DISPARITY.traffic.score.eo"  
#>  [8] "state.pctile.EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "state.pctile.EJ.DISPARITY.proximity.rmp.eo"  
#> [10] "state.pctile.EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "state.pctile.EJ.DISPARITY.proximity.npdes.eo"
#> [12] "state.pctile.EJ.DISPARITY.ust.eo"            
#> [13] "state.pctile.EJ.DISPARITY.rsei.eo"           
#> 
#> $ej_supp_pctile
#>  [1] "pctile.EJ.DISPARITY.pm.supp"             
#>  [2] "pctile.EJ.DISPARITY.o3.supp"             
#>  [3] "pctile.EJ.DISPARITY.cancer.supp"         
#>  [4] "pctile.EJ.DISPARITY.resp.supp"           
#>  [5] "pctile.EJ.DISPARITY.dpm.supp"            
#>  [6] "pctile.EJ.DISPARITY.pctpre1960.supp"     
#>  [7] "pctile.EJ.DISPARITY.traffic.score.supp"  
#>  [8] "pctile.EJ.DISPARITY.proximity.npl.supp"  
#>  [9] "pctile.EJ.DISPARITY.proximity.rmp.supp"  
#> [10] "pctile.EJ.DISPARITY.proximity.tsdf.supp" 
#> [11] "pctile.EJ.DISPARITY.proximity.npdes.supp"
#> [12] "pctile.EJ.DISPARITY.ust.supp"            
#> [13] "pctile.EJ.DISPARITY.rsei.supp"           
#> 
#> $ej_supp_state_pctile
#>  [1] "state.pctile.EJ.DISPARITY.pm.supp"             
#>  [2] "state.pctile.EJ.DISPARITY.o3.supp"             
#>  [3] "state.pctile.EJ.DISPARITY.cancer.supp"         
#>  [4] "state.pctile.EJ.DISPARITY.resp.supp"           
#>  [5] "state.pctile.EJ.DISPARITY.dpm.supp"            
#>  [6] "state.pctile.EJ.DISPARITY.pctpre1960.supp"     
#>  [7] "state.pctile.EJ.DISPARITY.traffic.score.supp"  
#>  [8] "state.pctile.EJ.DISPARITY.proximity.npl.supp"  
#>  [9] "state.pctile.EJ.DISPARITY.proximity.rmp.supp"  
#> [10] "state.pctile.EJ.DISPARITY.proximity.tsdf.supp" 
#> [11] "state.pctile.EJ.DISPARITY.proximity.npdes.supp"
#> [12] "state.pctile.EJ.DISPARITY.ust.supp"            
#> [13] "state.pctile.EJ.DISPARITY.rsei.supp"           
#> 
#> $need_pctile
#>  [1] "Demog.Index"                       "Demog.Index.Supp"                 
#>  [3] "pctlowinc"                         "pctlingiso"                       
#>  [5] "pctunemployed"                     "pctlths"                          
#>  [7] "lowlifex"                          "pctunder5"                        
#>  [9] "pctover64"                         "pctmin"                           
#> [11] "pcthisp"                           "pctnhba"                          
#> [13] "pctnhaa"                           "pctnhaiana"                       
#> [15] "pctnhnhpia"                        "pctnhotheralone"                  
#> [17] "pctnhmulti"                        "pctnhwa"                          
#> [19] "pm"                                "o3"                               
#> [21] "cancer"                            "resp"                             
#> [23] "dpm"                               "pctpre1960"                       
#> [25] "traffic.score"                     "proximity.npl"                    
#> [27] "proximity.rmp"                     "proximity.tsdf"                   
#> [29] "proximity.npdes"                   "ust"                              
#> [31] "rsei"                              "EJ.DISPARITY.pm.eo"               
#> [33] "EJ.DISPARITY.o3.eo"                "EJ.DISPARITY.cancer.eo"           
#> [35] "EJ.DISPARITY.resp.eo"              "EJ.DISPARITY.dpm.eo"              
#> [37] "EJ.DISPARITY.pctpre1960.eo"        "EJ.DISPARITY.traffic.score.eo"    
#> [39] "EJ.DISPARITY.proximity.npl.eo"     "EJ.DISPARITY.proximity.rmp.eo"    
#> [41] "EJ.DISPARITY.proximity.tsdf.eo"    "EJ.DISPARITY.proximity.npdes.eo"  
#> [43] "EJ.DISPARITY.ust.eo"               "EJ.DISPARITY.rsei.eo"             
#> [45] "EJ.DISPARITY.pm.supp"              "EJ.DISPARITY.o3.supp"             
#> [47] "EJ.DISPARITY.cancer.supp"          "EJ.DISPARITY.resp.supp"           
#> [49] "EJ.DISPARITY.dpm.supp"             "EJ.DISPARITY.pctpre1960.supp"     
#> [51] "EJ.DISPARITY.traffic.score.supp"   "EJ.DISPARITY.proximity.npl.supp"  
#> [53] "EJ.DISPARITY.proximity.rmp.supp"   "EJ.DISPARITY.proximity.tsdf.supp" 
#> [55] "EJ.DISPARITY.proximity.npdes.supp" "EJ.DISPARITY.ust.supp"            
#> [57] "EJ.DISPARITY.rsei.supp"           
#> 
#> $need_state_pctile
#>  [1] "Demog.Index"                       "Demog.Index.Supp"                 
#>  [3] "pctlowinc"                         "pctlingiso"                       
#>  [5] "pctunemployed"                     "pctlths"                          
#>  [7] "lowlifex"                          "pctunder5"                        
#>  [9] "pctover64"                         "pctmin"                           
#> [11] "pcthisp"                           "pctnhba"                          
#> [13] "pctnhaa"                           "pctnhaiana"                       
#> [15] "pctnhnhpia"                        "pctnhotheralone"                  
#> [17] "pctnhmulti"                        "pctnhwa"                          
#> [19] "pm"                                "o3"                               
#> [21] "cancer"                            "resp"                             
#> [23] "dpm"                               "pctpre1960"                       
#> [25] "traffic.score"                     "proximity.npl"                    
#> [27] "proximity.rmp"                     "proximity.tsdf"                   
#> [29] "proximity.npdes"                   "ust"                              
#> [31] "rsei"                              "EJ.DISPARITY.pm.eo"               
#> [33] "EJ.DISPARITY.o3.eo"                "EJ.DISPARITY.cancer.eo"           
#> [35] "EJ.DISPARITY.resp.eo"              "EJ.DISPARITY.dpm.eo"              
#> [37] "EJ.DISPARITY.pctpre1960.eo"        "EJ.DISPARITY.traffic.score.eo"    
#> [39] "EJ.DISPARITY.proximity.npl.eo"     "EJ.DISPARITY.proximity.rmp.eo"    
#> [41] "EJ.DISPARITY.proximity.tsdf.eo"    "EJ.DISPARITY.proximity.npdes.eo"  
#> [43] "EJ.DISPARITY.ust.eo"               "EJ.DISPARITY.rsei.eo"             
#> [45] "EJ.DISPARITY.pm.supp"              "EJ.DISPARITY.o3.supp"             
#> [47] "EJ.DISPARITY.cancer.supp"          "EJ.DISPARITY.resp.supp"           
#> [49] "EJ.DISPARITY.dpm.supp"             "EJ.DISPARITY.pctpre1960.supp"     
#> [51] "EJ.DISPARITY.traffic.score.supp"   "EJ.DISPARITY.proximity.npl.supp"  
#> [53] "EJ.DISPARITY.proximity.rmp.supp"   "EJ.DISPARITY.proximity.tsdf.supp" 
#> [55] "EJ.DISPARITY.proximity.npdes.supp" "EJ.DISPARITY.ust.supp"            
#> [57] "EJ.DISPARITY.rsei.supp"           
#> 
#> $these
#>  [1] "Demog.Index"      "Demog.Index.Supp" "pctlowinc"        "pctlingiso"      
#>  [5] "pctunemployed"    "pctlths"          "lowlifex"         "pctunder5"       
#>  [9] "pctover64"        "pctmin"           "pcthisp"          "pctnhba"         
#> [13] "pctnhaa"          "pctnhaiana"       "pctnhnhpia"       "pctnhotheralone" 
#> [17] "pctnhmulti"       "pctnhwa"          "pm"               "o3"              
#> [21] "cancer"           "resp"             "dpm"              "pctpre1960"      
#> [25] "traffic.score"    "proximity.npl"    "proximity.rmp"    "proximity.tsdf"  
#> [29] "proximity.npdes"  "ust"              "rsei"            
#> 
#> $these_avg
#>  [1] "avg.Demog.Index"      "avg.Demog.Index.Supp" "avg.pctlowinc"       
#>  [4] "avg.pctlingiso"       "avg.pctunemployed"    "avg.pctlths"         
#>  [7] "avg.lowlifex"         "avg.pctunder5"        "avg.pctover64"       
#> [10] "avg.pctmin"           "avg.pcthisp"          "avg.pctnhba"         
#> [13] "avg.pctnhaa"          "avg.pctnhaiana"       "avg.pctnhnhpia"      
#> [16] "avg.pctnhotheralone"  "avg.pctnhmulti"       "avg.pctnhwa"         
#> [19] "avg.pm"               "avg.o3"               "avg.cancer"          
#> [22] "avg.resp"             "avg.dpm"              "avg.pctpre1960"      
#> [25] "avg.traffic.score"    "avg.proximity.npl"    "avg.proximity.rmp"   
#> [28] "avg.proximity.tsdf"   "avg.proximity.npdes"  "avg.ust"             
#> [31] "avg.rsei"            
#> 
#> $these_state_avg
#>  [1] "state.avg.Demog.Index"      "state.avg.Demog.Index.Supp"
#>  [3] "state.avg.pctlowinc"        "state.avg.pctlingiso"      
#>  [5] "state.avg.pctunemployed"    "state.avg.pctlths"         
#>  [7] "state.avg.lowlifex"         "state.avg.pctunder5"       
#>  [9] "state.avg.pctover64"        "state.avg.pctmin"          
#> [11] "state.avg.pcthisp"          "state.avg.pctnhba"         
#> [13] "state.avg.pctnhaa"          "state.avg.pctnhaiana"      
#> [15] "state.avg.pctnhnhpia"       "state.avg.pctnhotheralone" 
#> [17] "state.avg.pctnhmulti"       "state.avg.pctnhwa"         
#> [19] "state.avg.pm"               "state.avg.o3"              
#> [21] "state.avg.cancer"           "state.avg.resp"            
#> [23] "state.avg.dpm"              "state.avg.pctpre1960"      
#> [25] "state.avg.traffic.score"    "state.avg.proximity.npl"   
#> [27] "state.avg.proximity.rmp"    "state.avg.proximity.tsdf"  
#> [29] "state.avg.proximity.npdes"  "state.avg.ust"             
#> [31] "state.avg.rsei"            
#> 
#> $these_ratio_to_avg
#>  [1] "ratio.to.avg.Demog.Index"      "ratio.to.avg.Demog.Index.Supp"
#>  [3] "ratio.to.avg.pctlowinc"        "ratio.to.avg.pctlingiso"      
#>  [5] "ratio.to.avg.pctunemployed"    "ratio.to.avg.pctlths"         
#>  [7] "ratio.to.avg.lowlifex"         "ratio.to.avg.pctunder5"       
#>  [9] "ratio.to.avg.pctover64"        "ratio.to.avg.pctmin"          
#> [11] "ratio.to.avg.pcthisp"          "ratio.to.avg.pctnhba"         
#> [13] "ratio.to.avg.pctnhaa"          "ratio.to.avg.pctnhaiana"      
#> [15] "ratio.to.avg.pctnhnhpia"       "ratio.to.avg.pctnhotheralone" 
#> [17] "ratio.to.avg.pctnhmulti"       "ratio.to.avg.pctnhwa"         
#> [19] "ratio.to.avg.pm"               "ratio.to.avg.o3"              
#> [21] "ratio.to.avg.cancer"           "ratio.to.avg.resp"            
#> [23] "ratio.to.avg.dpm"              "ratio.to.avg.pctpre1960"      
#> [25] "ratio.to.avg.traffic.score"    "ratio.to.avg.proximity.npl"   
#> [27] "ratio.to.avg.proximity.rmp"    "ratio.to.avg.proximity.tsdf"  
#> [29] "ratio.to.avg.proximity.npdes"  "ratio.to.avg.ust"             
#> [31] "ratio.to.avg.rsei"            
#> 
#> $these_ratio_to_state_avg
#>  [1] "ratio.to.state.avg.Demog.Index"      "ratio.to.state.avg.Demog.Index.Supp"
#>  [3] "ratio.to.state.avg.pctlowinc"        "ratio.to.state.avg.pctlingiso"      
#>  [5] "ratio.to.state.avg.pctunemployed"    "ratio.to.state.avg.pctlths"         
#>  [7] "ratio.to.state.avg.lowlifex"         "ratio.to.state.avg.pctunder5"       
#>  [9] "ratio.to.state.avg.pctover64"        "ratio.to.state.avg.pctmin"          
#> [11] "ratio.to.state.avg.pcthisp"          "ratio.to.state.avg.pctnhba"         
#> [13] "ratio.to.state.avg.pctnhaa"          "ratio.to.state.avg.pctnhaiana"      
#> [15] "ratio.to.state.avg.pctnhnhpia"       "ratio.to.state.avg.pctnhotheralone" 
#> [17] "ratio.to.state.avg.pctnhmulti"       "ratio.to.state.avg.pctnhwa"         
#> [19] "ratio.to.state.avg.pm"               "ratio.to.state.avg.o3"              
#> [21] "ratio.to.state.avg.cancer"           "ratio.to.state.avg.resp"            
#> [23] "ratio.to.state.avg.dpm"              "ratio.to.state.avg.pctpre1960"      
#> [25] "ratio.to.state.avg.traffic.score"    "ratio.to.state.avg.proximity.npl"   
#> [27] "ratio.to.state.avg.proximity.rmp"    "ratio.to.state.avg.proximity.tsdf"  
#> [29] "ratio.to.state.avg.proximity.npdes"  "ratio.to.state.avg.ust"             
#> [31] "ratio.to.state.avg.rsei"            
#> 
   grep("\\.eo$", namez$ej, value = T)
#>  [1] "EJ.DISPARITY.pm.eo"              "EJ.DISPARITY.o3.eo"             
#>  [3] "EJ.DISPARITY.cancer.eo"          "EJ.DISPARITY.resp.eo"           
#>  [5] "EJ.DISPARITY.dpm.eo"             "EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "EJ.DISPARITY.traffic.score.eo"   "EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "EJ.DISPARITY.proximity.rmp.eo"   "EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "EJ.DISPARITY.proximity.npdes.eo" "EJ.DISPARITY.ust.eo"            
#> [13] "EJ.DISPARITY.rsei.eo"