Plot distribution of data among residents at sites vs reference area (US, etc.)

Visualize indicator values (scores) as distributed across places

Usage

plot_vs_us(
  bysite = NULL,
  varname = "pctlowinc",
  type = "ggplot",
  refarealabel = "All Blockgroups Nationwide",
  siteslabel = "At Sites Analyzed",
  siteidlabel = NULL,
  refdata = NULL,
  nsample = 5000,
  fix_pctcols = TRUE,
  colorfills = c("lightblue", "yellow"),
  box.cex.ref = 0.6,
  box.cex.here = 2.2,
  box.pch.ref = 20,
  box.pch.here = 2,
  ...
)

Arguments

bysite

table of results from ejamit()$results_bysite, like testoutput_ejamit_1000pts_1miles$results_bysite

varname

name of column in bysite, like "Demog.Index"

type

"box", "plotly", or "ggplot"

refarealabel

e.g., "All blockgroups in this State"

siteslabel

e.g., "At Avg. Site Analyzed"

siteidlabel

vector of text one per site to show if type 'box'

refdata

reference area dataset, like blockgroupstats, but must have columns named 'pop' and varname. e.g., refdata = blockgroupstatsST %in% "DE", .(pop, pcthisp)

nsample

to limit dots on plot of ref area like all bg in US

colorfills

two colors for boxplot

box.cex.ref

use default

box.cex.here

use default

box.pch.ref

use default

box.pch.here

description

...

passed to boxplot()

Value

plots

Details

Not population weighted, so it is the distribution across sites not residents.

Could be edited to allow multiple types of sites and/or reference zones more generally like for ejamit_compare_types_of_places()

See also

Examples

if (FALSE) { # \dontrun{
  out <- testoutput_ejamit_1000pts_1miles
  # ejam2boxplot(out)
  # plot_boxplot_vs_ref(out$results_bysite)
  plot_vs_us(out$results_bysite, type = 'box')
  plot_vs_us(out$results_bysite, varname = "pctlingiso", type =  'box', ylim=c(0, 20))
  plot_vs_us(out$results_bysite, varname = "pctlingiso", type =  'ggplot')
  plot_vs_us(out$results_bysite, varname = "pctnhaa", type =  'ggplot')
  plot_vs_us(out$results_bysite, varname = "pctnhaa", type = 'box', ylim = c(0, 20))
  
 # td = testoutput_ejamit_1000pts_1miles$results_bysite
 # plot_vs_us(, type = 'box')
 # plot_vs_us(td, varname = "pctlingiso", type =  'box', ylim=c(0,20))
 # plot_vs_us(td, varname = "pctlingiso", type =  'ggplot')
 # plot_vs_us(td, varname = "pctnhaa", type =  'ggplot')
 # plot_vs_us(td, varname = "pctnhaa", type = 'box', ylim = c(0,20))
 # plot_vs_us(td[td$ST %in% "DE", ], 'pcthisp', refdata = blockgroupstats[ST %in% "DE", .(pop, pcthisp)])
  } # }