what percent of sites account for P percent of residents?

what percent of sites account for P percent of residents?

Usage

popshare_p_lives_at_what_pct(pop, p, astext = FALSE, dig = 0)

Arguments

pop

vector of population totals across places, like out$results_bysite$pop where out is the output of ejamit()

p

share of population (0-1, fraction), vector of one or more

astext

if TRUE, return text of description of results

dig

rounding digits for text output

Value

vector of fractions 0-1 of all sites, or text about that

Examples

 x <- testdata_ejamit_output_100pts_1miles$results_bysite
 popshare_p_lives_at_what_pct(x$pop, p = 0.50, astext=TRUE)
 popshare_p_lives_at_what_n(  x$pop, p = c(0.50, 0.67, 0.80, 0.95))
 popshare_at_top_x_pct(       x$pop, x = c(0.25, 0.50, .90))
 popshare_at_top_n(           x$pop, n = c(1, 5, 10))