Find EPA-regulated facilities in FRS by NAICS code (industrial category)

Get lat lon, Registry ID, given NAICS industry code(s) Find all EPA Facility Registry Service (FRS) sites with this exact NAICS code (not subcategories)

regid_from_naics(naics, id_only = FALSE)

Arguments

naics

a vector of naics codes, or a data.table with column named code, as with output of naics_from_any()

Value

A data.table (not just data.frame) with columns called lat, lon, REGISTRY_ID, NAICS (but see the id_only parameter)

Details

NOTE: many FRS sites lack NAICS code!

Also, this function does not find the sites identified by FRS data as being in a child NAICS (subcategory of your exact query)!

Relies on frs_by_naics (a data.table)

See info about NAICS industry codes at https://www.naics.com/search

Examples

  regid_from_naics(321114)
#>           lat        lon  REGISTRY_ID  NAICS
#>   1: 18.41565  -66.20266 110000307837 321114
#>   2: 42.69898  -73.82309 110000324505 321114
#>   3: 41.71228  -73.93705 110000324809 321114
#>   4: 40.08626  -76.18172 110000333826 321114
#>   5: 40.18434  -74.79960 110000336253 321114
#>  ---                                        
#> 493: 35.59275  -84.46818 110071307905 321114
#> 494: 35.42391  -87.28222 110071334308 321114
#> 495: 37.50455 -122.21792 110071380069 321114
#> 496: 39.79828 -104.99275 110071414306 321114
#> 497: 35.89974  -88.75981 110071434078 321114
  latlon_from_naics(321114)
#>           lat        lon  REGISTRY_ID  NAICS
#>   1: 18.41565  -66.20266 110000307837 321114
#>   2: 42.69898  -73.82309 110000324505 321114
#>   3: 41.71228  -73.93705 110000324809 321114
#>   4: 40.08626  -76.18172 110000333826 321114
#>   5: 40.18434  -74.79960 110000336253 321114
#>  ---                                        
#> 493: 35.59275  -84.46818 110071307905 321114
#> 494: 35.42391  -87.28222 110071334308 321114
#> 495: 37.50455 -122.21792 110071380069 321114
#> 496: 39.79828 -104.99275 110071414306 321114
#> 497: 35.89974  -88.75981 110071434078 321114
  latlon_from_naics(EJAM::naics_from_any("cheese")[,code] )
#> Warning: NAs introduced by coercion
#> Warning: mynames should be non-NA character vector of text to look for in industry title(s) like "concrete"
#>           lat        lon  REGISTRY_ID  NAICS
#>   1: 44.18750  -75.95694 110000326102 311513
#>   2: 42.87593  -78.86949 110000326825 311513
#>   3: 42.83404  -78.82388 110000327183 311513
#>   4: 41.47924  -81.07851 110000385798 311513
#>   5: 43.59796  -85.14557 110000410207 311513
#>  ---                                        
#> 397: 44.91011  -72.79787 110071292784 311513
#> 398: 42.84509  -90.69292 110071433364 311513
#> 399: 42.84416  -73.98835 110071433604 311513
#> 400: 32.87831 -111.78364 110071434204 311513
#> 401: 44.75453  -90.29649 110071498797 311513
  head(latlon_from_naics(c(3366, 33661, 336611), id_only=TRUE))
#> [1] "110000897290" "110000332621" "110000360564" "110000369869" "110000377333"
#> [6] "110000448962"
  # mapfast(frs_from_naics(336611)) # simple map