Generate Single-site or Multi-site Summary Report rendered as an HTML file for download in shiny app

Saves a 2 page report (on overall results or for just one site), with residential population and environmental indicators, Summary Indexes if needed, etc.

Usage

report_community_download(
  file,
  row_index = NULL,
  react_cur_button,
  inshiny = FALSE,
  input_analysis_title,
  input_include_ejindexes,
  input_show_ratios_in_report,
  input_extratable_show_ratios_in_report,
  input_extratable_title,
  input_extratable_title_top_row,
  extratable_list_of_sections,
  extratable_hide_missing_rows_for,
  input_plotkind_1pager,
  input_Custom_title_for_bar_plot_of_indicators,
  input_circleweight_in,
  react_sanitized_analysis_title,
  react_sanitized_bt_rad_buff,
  react_total_pop,
  react_submitted_upload_method,
  react_data_uploaded,
  react_data_processed,
  react_ratio.to.us.d,
  react_report_map,
  react_v1_summary_plot,
  react_v1_summary_plot_state
)

Arguments

file

for download

row_index

which row of results (which location), related to react_cur_button

react_cur_button

like "button_2", event button asking for a 1-site report, a shiny app reactive

inshiny

logical

input_analysis_title

a shiny app input$

input_include_ejindexes

a shiny app input$

input_show_ratios_in_report

a shiny app input$

input_extratable_show_ratios_in_report

a shiny app input$

input_extratable_title

a shiny app input$

input_extratable_title_top_row

a shiny app input$

extratable_list_of_sections

see build_community_report()

extratable_hide_missing_rows_for

see build_community_report()

input_plotkind_1pager

a shiny app input$, like "bar", "box", "ridgeline"

input_Custom_title_for_bar_plot_of_indicators

a shiny app input$

input_circleweight_in

a shiny app input$

react_sanitized_analysis_title

a shiny app reactive

react_sanitized_bt_rad_buff

a shiny app reactive

react_total_pop

a shiny app reactive

react_submitted_upload_method

a shiny app reactive

react_data_uploaded

a shiny app reactive, like input to ejamit()

react_data_processed

a shiny app reactive, like output from ejamit()

react_ratio.to.us.d

a shiny app reactive

react_report_map

a shiny app reactive, # map of all sites Overall

react_v1_summary_plot

a shiny app reactive

react_v1_summary_plot_state

a shiny app reactive

Value

Renders HTML report

Details

This function report_community_download() is similar to build_community_report() (and could be merged with it ideally) but was used by the shiny app to render the report as an HTML file for download instead of returning HTML for display in a browser.

report_community_download() relies on helpers report_setup_temp_files() to copy files, map_single_location() to draw map, v1_summary_plot_report() to draw plot, and rmarkdown::render() to do the parameterized render from .Rmd template into HTML

See also