Build text for report: Residents within( X miles of)( any of) the (N) point(s)/polygon(s)/Census unit(s)

Build text for report: Residents within( X miles of)( any of) the (N) point(s)/polygon(s)/Census unit(s)

Usage

report_residents_within_xyz(
  text1 = "Residents within ",
  radius = NULL,
  unitsingular = "mile",
  nsites = 1,
  sitetype = c("latlon", "fips", "shp", "frs", "epa_program_up", "echo", "naics", "sic",
    "mact", "epa_program_sel")[1]
)

Arguments

text1

text to start the phrase, like "Residents within "

radius

The distance from each place, normally in miles (which can be 0), or custom text like "seven kilometers from" in which case it should end with words like "a safe distance from" or "the vicinity of" or "proximity to" or "near" – but may need to specify custom text1 also.

unitsingular

'mile' by default, but can use 'kilometer' etc. Ignored if radius is not a number.

nsites

number of places or text in lieu of number

sitetype

can be 'latlon', 'fips', 'shp', or some singular custom text like "Georgia location" but should be something that can be made plural by just adding "s" so ending with "site" works better than ending with "... facility" since that would print as "facilitys" here.