given vector of blockids, get state abbreviation of each unused. Not needed if you have sites2blocks table that includes a bgid column

given vector of blockids, get state abbreviation of each unused. Not needed if you have sites2blocks table that includes a bgid column

Usage

state_from_blockid(blockid)

Arguments

blockid

vector of blockid values as from EJAM in a table called blockpoints

Value

vector of ST info like AK, CA, DE, etc.

See also

Examples

x = sample(blockpoints$blockid, 3)
state_from_blockid(x)[]
#> Error in state_from_blockid(x): could not find function "state_from_blockid"
mapfast(blockpoints[blockid %in% x, ])

all.equal(state_from_blockid(x), state_from_blockid_table(blockpoints[blockid %in% x, ]))
#> Error in state_from_blockid(x): could not find function "state_from_blockid"