Find what state is where each point is located

Takes 3 seconds to find state for 1k points, so a faster alternative would be useful

Usage

state_from_latlon(lat, lon, states_shapefile = EJAM::states_shapefile)

Arguments

lat

latitudes vector

lon

longitudes vector

shapefile

shapefile of US States, in package already

Value

Returns data.frame: ST, statename, FIPS.ST, REGION, n as many rows as elements in lat or lon

See also

Examples

 myprogram <- "CAMDBS" # 739 sites
 pts <- frs_from_program(myprogram)[ , .(lat, lon, REGISTRY_ID,  PRIMARY_NAME)]
 # add a column with State abbreviation
 pts[, ST := state_from_latlon(lat=lat, lon = lon)$ST]
#>           lat       lon  REGISTRY_ID                                    PRIMARY_NAME ST
#>   1: 17.94930 -66.23170 110000307800                                   PREPA AGUIRRE PR
#>   2: 42.15698 -72.52237 110000308444                 MASSPOWER COGENERATION FACILITY MA
#>   3: 42.45488 -73.21734 110000308480                PITTSFIELD GENERATING COMPANY LP MA
#>   4: 43.16180 -78.75460 110000326530                  LOCKPORT COGENERATION FACILITY NY
#>   5: 41.49085 -78.67752 110000330400                       DOMTAR - JOHNSONBURG MILL PA
#>  ---                                                                                   
#> 710: 29.70204 -95.07088 110070100752               EXELON LAPORTE GENERATING STATION TX
#> 711: 40.95444 -72.88500 110070696981                             SHOREHAM ENERGY LLC NY
#> 712: 41.03170 -81.87140 110070828207                               OMEGA JV2 SEVILLE OH
#> 713: 40.75195 -82.79389 110070828208                   OMEGA JV2-GALION (0317030059) OH
#> 714: 38.37147 -87.74668 110070834810 GIBSON GENERATING STATION TYPE I SOUTH LANDFILL IN
 #map these points
 mapfast(pts[ST == 'TX',], radius = 1) # 1 miles radius circles