Create a formatted table of results for 1 site from EJAM

Uses 1 row from the results_bysite part of ejamit() output

Usage

table_gt_from_ejamit_1site(...)

Arguments

...

passed to table_gt_from_ejamit_overall()

Examples

 table_gt_from_ejamit_1site(testoutput_ejamit_100pts_1miles$results_bysite[ 1, ])

  

Socioeconomic Indicators
      	Value
      	Average
 in State1
      	Percentile
 in State1
      	Average
 in USA
      	Percentile
 in USA
      	Ratio to
 State Avg.
      	Ratio to
 USA Avg.
    

Demographic Index USA
	205.99%
	153.79%
	75.00
	134.07%
	79
	1.40
	1.5

Supplemental Demographic Index USA
	238.49%
	138.12%
	86.00
	164.23%
	86
	1.51
	1.5

% People of Color
	51%
	46%
	58
	40%
	66
	1.1
	1.3

% Low Income
	53%
	32%
	83
	30%
	84
	1.7
	1.8

% Unemployed
	11%
	5%
	87
	6%
	85
	2.2
	2.0

% in limited English-speaking Households
	1%
	7%
	45
	5%
	59
	0.2
	0.2

% with Less Than High School Education
	12%
	11%
	65
	11%
	65
	1.1
	1.1

% under Age 5
	6%
	5%
	70
	5%
	64
	1.3
	1.2

% over Age 64
	15%
	23%
	36
	18%
	44
	0.6
	0.8

1 Avg. in state means the average indicator value, among all the residents at these sites, using the statewide value in each resident's state. Percentile in state means the same, but using the site-specific value (expressed as a percentile) where each resident lives.