Get order of variable names to sort by, as seen in EJSCREEN Community Report

Get order of variable names to sort by, as seen in EJSCREEN Community Report

Usage

table_order_variables(varnames, s1 = "newsort", s2 = "ejscreensort", s3 = "n")

Arguments

varnames

vector of indicator variables names from blockgroupstats, bgej, etc., such as "pm", "pctlowinc", "pctile.EJ.DISPARITY.traffic.score.eo" etc. and others as found in names_all_r, or specific subsets of those like in c(names_d, names_d_subgroups, names_e) and c(names_ej_pctile, names_ej_state_pctile, names_ej_supp_pctile, names_ej_supp_state_pctile)

s1

name of column in map_headernames to get sort info from

s2

optional like s1 but secondary to s1

s3

optional tertiary

Value

vector as from order(), to be used in sorting a data.frame for example

Examples

  cbind(table_order_variables(c(names_d, names_d_subgroups, names_e)))
#>       [,1]
#>  [1,]    1
#>  [2,]    2
#>  [3,]    3
#>  [4,]    4
#>  [5,]    5
#>  [6,]    6
#>  [7,]    7
#>  [8,]    8
#>  [9,]    9
#> [10,]   10
#> [11,]   11
#> [12,]   12
#> [13,]   13
#> [14,]   14
#> [15,]   15
#> [16,]   16
#> [17,]   17
#> [18,]   18
#> [19,]   19
#> [20,]   20
#> [21,]   21
#> [22,]   22
#> [23,]   23
#> [24,]   24
#> [25,]   25
#> [26,]   26
#> [27,]   27
#> [28,]   28
#> [29,]   29
#> [30,]   30

  out <- testoutput_ejamit_10pts_1miles
  vars <- out$formatted[ , 'indicator']
  vars <- fixcolnames(vars, 'long', 'r')
  out$formatted[table_order_variables(vars), ]
#>           value
#> 1          <NA>
#> 2          <NA>
#> 3          <NA>
#> 4          <NA>
#> 5          <NA>
#> 11            7
#> 77         1.35
#> 78         1.28
#> 79           23
#> 80            6
#> 81            6
#> 82            9
#> 83            6
#> 84           13
#> 85           53
#> 94            1
#> 95            1
#> 13            1
#> 14          0.8
#> 15          0.7
#> 16          1.2
#> 17          1.2
#> 18          0.8
#> 19          1.2
#> 20          0.7
#> 21          1.3
#> 45            1
#> 46          0.8
#> 47          0.8
#> 48          0.8
#> 49          1.1
#> 50          0.7
#> 51          1.2
#> 52          0.7
#> 53          1.2
#> 165          58
#> 166          33
#> 167          43
#> 168          77
#> 169          68
#> 170          54
#> 171          65
#> 172          35
#> 173          67
#> 202          57
#> 203          41
#> 204          50
#> 205          67
#> 206          63
#> 207          49
#> 208          66
#> 209          36
#> 210          62
#> 239        1.34
#> 240        1.64
#> 241          30
#> 242           5
#> 243           6
#> 244          11
#> 245           5
#> 246          18
#> 247          40
#> 271        1.35
#> 272        1.47
#> 273          27
#> 274           7
#> 275           6
#> 276          12
#> 277           5
#> 278          17
#> 279          44
#> 303       46000
#> 304        5000
#> 305        7700
#> 306       13000
#> 307       14000
#> 308       27000
#> 309      110000
#> 408          17
#> 409          12
#> 410           0
#> 411           0
#> 412           7
#> 413          11
#> 414          53
#> 318       40000
#> 319       30000
#> 320         800
#> 321         700
#> 322       10000
#> 323       20000
#> 324       1e+05
#> 86           19
#> 87           16
#> 88           12
#> 89            0
#> 90            0
#> 91            1
#> 92            5
#> 93           47
#> 22            1
#> 23          1.3
#> 24          2.4
#> 25          0.3
#> 26          2.2
#> 27          1.6
#> 28          1.5
#> 29          0.8
#> 54          0.9
#> 55          1.4
#> 56          1.4
#> 57          0.8
#> 58          4.9
#> 59            1
#> 60          1.6
#> 61          0.9
#> 174          70
#> 175          77
#> 176          87
#> 177          84
#> 178          95
#> 179          88
#> 180          77
#> 181          33
#> 211          61
#> 212          72
#> 213          74
#> 214          88
#> 215          72
#> 216          81
#> 217          77
#> 218          39
#> 248          18
#> 249          12
#> 250           5
#> 251           1
#> 252           0
#> 253           0
#> 254           3
#> 255          60
#> 280          20
#> 281          12
#> 282           9
#> 283           0
#> 284           0
#> 285           1
#> 286           3
#> 287          54
#> 310       40000
#> 311       30000
#> 312       30000
#> 313         400
#> 314         700
#> 315        1000
#> 316       10000
#> 317       1e+05
#> 105          28
#> 106           7
#> 107           6
#> 108           2
#> 109           1
#> 110          13
#> 111           0
#> 112           0
#> 113           0
#> 114           0
#> 115           0
#> 116           0
#> 337      160000
#> 338       42874
#> 339       35830
#> 340        9800
#> 341      575168
#> 342        3633
#> 343       72060
#> 117          47
#> 118          25
#> 119          27
#> 120           1
#> 344        2343
#> 345        1257
#> 346        1330
#> 347          68
#> 6        212317
#> 325      151161
#> 326       87417
#> 327      175609
#> 328       52699
#> 329       95293
#> 330      202460
#> 96           12
#> 331       33361
#> 332       10660
#> 121           9
#> 122          17
#> 123           4
#> 124         9.9
#> 125         4.9
#> 43          0.7
#> 44          0.9
#> 75          0.8
#> 76            1
#> 182          26
#> 183          25
#> 184        <NA>
#> 185        <NA>
#> 186        <NA>
#> 219          35
#> 221        <NA>
#> 222        <NA>
#> 223        <NA>
#> 256        13.6
#> 257          20
#> 288        11.9
#> 289          18
#> 403       20000
#> 404      210565
#> 126           1
#> 127           1
#> 128           8
#> 129           1
#> 130           1
#> 100          78
#> 220          43
#> 133        8.41
#> 134        62.7
#> 135        15.8
#> 136       0.591
#> 137         540
#> 138     5400000
#> 139        55.3
#> 140        1.61
#> 141        0.13
#> 142        18.7
#> 143        14.4
#> 144       10000
#> 145           6
#> 30            1
#> 31            1
#> 32            2
#> 33          3.1
#> 34          0.1
#> 35          3.2
#> 36          1.8
#> 37          4.1
#> 38          0.2
#> 39          5.3
#> 40          3.9
#> 41            0
#> 42          2.9
#> 62            1
#> 63            1
#> 64          1.6
#> 65          1.9
#> 66          0.7
#> 67          1.8
#> 68          1.2
#> 69          1.8
#> 70          0.4
#> 71          2.5
#> 72            2
#> 73          0.9
#> 74          1.6
#> 189          59
#> 190          61
#> 191          97
#> 192          97
#> 193          48
#> 194          92
#> 195          77
#> 196          95
#> 197          38
#> 198          96
#> 199          93
#> 200          88
#> 201          91
#> 226          69
#> 227          44
#> 228          81
#> 229          76
#> 230          69
#> 231          77
#> 232          55
#> 233          72
#> 234          53
#> 235          80
#> 236          69
#> 237          69
#> 238          37
#> 258        8.45
#> 259        61.8
#> 260         7.8
#> 261       0.191
#> 262        4600
#> 263     1700000
#> 264         0.3
#> 265        0.39
#> 266        0.57
#> 267         3.5
#> 268         3.6
#> 269       7e+05
#> 270         2.2
#> 290        8.13
#> 291        63.3
#> 292         9.7
#> 293       0.318
#> 294         810
#> 295     2900000
#> 296        0.48
#> 297        0.89
#> 298        0.34
#> 299         7.5
#> 300         7.2
#> 301       11000
#> 302        3.97
#> 131           6
#> 132           2
#> 187        <NA>
#> 188        <NA>
#> 224        <NA>
#> 225        <NA>
#> 161          10
#> 164         0.7
#> 162           1
#> 160           1
#> 163           0
#> 415      72.662
#> 416      83.142
#> 417     117.641
#> 418     116.821
#> 419      59.572
#> 420     118.693
#> 421      94.692
#> 422     108.065
#> 423      36.502
#> 424     123.057
#> 425      99.994
#> 426     105.869
#> 427      80.294
#> 428      91.953
#> 429      58.857
#> 430     104.733
#> 431      97.954
#> 432       91.73
#> 433     101.953
#> 434      76.136
#> 435      92.378
#> 436      54.627
#> 437     106.043
#> 438      83.854
#> 439      91.479
#> 440       46.56
#> 348          63
#> 349          69
#> 350          76
#> 351          77
#> 352          58
#> 353          77
#> 354          73
#> 355          80
#> 356          45
#> 357          78
#> 358          74
#> 359          76
#> 360          83
#> 374          66
#> 375          53
#> 376          68
#> 377          65
#> 378          66
#> 379          68
#> 380          60
#> 381          64
#> 382          58
#> 383          68
#> 384          59
#> 385          66
#> 386          45
#> 441      66.552
#> 442      78.034
#> 443     110.354
#> 444     109.369
#> 445      56.357
#> 446     110.815
#> 447      91.293
#> 448     101.699
#> 449      35.509
#> 450     114.085
#> 451      94.901
#> 452      99.413
#> 453       78.09
#> 454      85.146
#> 455      55.113
#> 456      97.476
#> 457      92.088
#> 458      85.865
#> 459      93.448
#> 460      71.343
#> 461      87.488
#> 462      53.506
#> 463      97.693
#> 464      80.199
#> 465      84.606
#> 466      47.318
#> 361          47
#> 362          57
#> 363          72
#> 364          73
#> 365          41
#> 366          74
#> 367          60
#> 368          74
#> 369          32
#> 370          75
#> 371          64
#> 372          67
#> 373          76
#> 387          64
#> 388          45
#> 389          67
#> 390          65
#> 391          61
#> 392          67
#> 393          54
#> 394          64
#> 395          51
#> 396          67
#> 397          56
#> 398          63
#> 399          43
#> 146          69
#> 147          72
#> 148           9
#> 103          19
#> 104          81
#> 335      171920
#> 336       40397
#> 97        87417
#> 98           48
#> 99           52
#> 101       78766
#> 102          38
#> 333      110517
#> 334      101800
#> 149           0
#> 154           1
#> 155          18
#> 156         163
#> 157          28
#> 158         112
#> 159           5
#> 150           1
#> 151           1
#> 152           1
#> 153           1
#> 467        31.4
#> 7          <NA>
#> 10         <NA>
#> 402           1
#> 401        1640
#> 400         211
#> 405  11335817.4
#> 9          <NA>
#> 12         <NA>
#> 8          <NA>
#> 406 205467995.8
#> 407           0
#>                                                                                          indicator
#> 1                                                                                      EJAM Report
#> 2                                                                                     EJSCREEN Map
#> 3                                                                                     ejam_uniq_id
#> 4                                                                                            valid
#> 5                                                                                      invalid_msg
#> 11                                                                              in_how_many_states
#> 77                                                                           Demographic Index USA
#> 78                                                              Supplemental Demographic Index USA
#> 79                                                                                    % Low Income
#> 80                                                        % in limited English-speaking Households
#> 81                                                                                    % Unemployed
#> 82                                                          % with Less Than High School Education
#> 83                                                                                   % under Age 5
#> 84                                                                                   % over Age 64
#> 85                                                                               % People of Color
#> 94                                                                         Demographic Index State
#> 95                                                            Supplemental Demographic Index State
#> 13                                                               Ratio to US avg Demographic Index
#> 14                   Ratio to US avg Supplemental Summary Index for Supplemental Demographic Index
#> 15                                                                    Ratio to US avg % Low Income
#> 16                                        Ratio to US avg % in limited English-speaking Households
#> 17                                                                    Ratio to US avg % Unemployed
#> 18                                          Ratio to US avg % with Less Than High School Education
#> 19                                                                   Ratio to US avg % under Age 5
#> 20                                                                   Ratio to US avg % over Age 64
#> 21                                                               Ratio to US avg % People of Color
#> 45                                                            Ratio to State avg Demographic Index
#> 46                Ratio to State avg Supplemental Summary Index for Supplemental Demographic Index
#> 47                                                                 Ratio to State avg % Low Income
#> 48                                     Ratio to State avg % in limited English-speaking Households
#> 49                                                                 Ratio to State avg % Unemployed
#> 50                                       Ratio to State avg % with Less Than High School Education
#> 51                                                                Ratio to State avg % under Age 5
#> 52                                                                Ratio to State avg % over Age 64
#> 53                                                            Ratio to State avg % People of Color
#> 165                                                            US percentile for Demographic Index
#> 166                US percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 167                                                                 US percentile for % Low Income
#> 168                                     US percentile for % in limited English-speaking Households
#> 169                                                                 US percentile for % Unemployed
#> 170                                       US percentile for % with Less Than High School Education
#> 171                                                                US percentile for % under Age 5
#> 172                                                                US percentile for % over Age 64
#> 173                                                            US percentile for % People of Color
#> 202                                                         State percentile for Demographic Index
#> 203             State percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 204                                                              State percentile for % Low Income
#> 205                                  State percentile for % in limited English-speaking Households
#> 206                                                              State percentile for % Unemployed
#> 207                                    State percentile for % with Less Than High School Education
#> 208                                                             State percentile for % under Age 5
#> 209                                                             State percentile for % over Age 64
#> 210                                                         State percentile for % People of Color
#> 239                                                               US Average for Demographic Index
#> 240                   US Average for Supplemental Summary Index for Supplemental Demographic Index
#> 241                                                                    US Average for % Low Income
#> 242                                        US Average for % in limited English-speaking Households
#> 243                                                                    US Average for % Unemployed
#> 244                                          US Average for % with Less Than High School Education
#> 245                                                                   US Average for % under Age 5
#> 246                                                                   US Average for % over Age 64
#> 247                                                               US Average for % People of Color
#> 271                                                            State Average for Demographic Index
#> 272                State Average for Supplemental Summary Index for Supplemental Demographic Index
#> 273                                                                 State Average for % Low Income
#> 274                                     State Average for % in limited English-speaking Households
#> 275                                                                 State Average for % Unemployed
#> 276                                       State Average for % with Less Than High School Education
#> 277                                                                State Average for % under Age 5
#> 278                                                                State Average for % over Age 64
#> 279                                                            State Average for % People of Color
#> 303                                                                      Low income resident count
#> 304                                                            Limited English-speaking Households
#> 305                                                                      Unemployed resident count
#> 306                                                 Less Than High School Education resident count
#> 307                                                                     Under Age 5 resident count
#> 308                                                                     Over Age 64 resident count
#> 309                                                                 People of Color resident count
#> 408                                   % Black or African American (single race, includes Hispanic)
#> 409                                                       % Asian (single race, includes Hispanic)
#> 410                           % American Indian and Alaska Native (single race, includes Hispanic)
#> 411                  % Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 412                                                  % Other race (single race, includes Hispanic)
#> 413                                                        % Two or more races (includes Hispanic)
#> 414                                                       % White (single race, includes Hispanic)
#> 318                            Count of Black or African American (single race, includes Hispanic)
#> 319                                                Count of Asian (single race, includes Hispanic)
#> 320                    Count of American Indian and Alaska Native (single race, includes Hispanic)
#> 321           Count of Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 322                                           Count of Other race (single race, includes Hispanic)
#> 323                                                 Count of Two or more races (includes Hispanic)
#> 324                                                Count of White (single race, includes Hispanic)
#> 86                                                                            % Hispanic or Latino
#> 87                                         % Black or African American (non-Hispanic, single race)
#> 88                                                             % Asian (non-Hispanic, single race)
#> 89                                 % American Indian and Alaska Native (non-Hispanic, single race)
#> 90                        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 91                                                        % Other race (non-Hispanic, single race)
#> 92                                                              % Two or more races (non-Hispanic)
#> 93                                                             % White (non-Hispanic, single race)
#> 22                                                            Ratio to US avg % Hispanic or Latino
#> 23                         Ratio to US avg % Black or African American (non-Hispanic, single race)
#> 24                                             Ratio to US avg % Asian (non-Hispanic, single race)
#> 25                 Ratio to US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 26        Ratio to US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 27                                        Ratio to US avg % Other race (non-Hispanic, single race)
#> 28                                              Ratio to US avg % Two or more races (non-Hispanic)
#> 29                                             Ratio to US avg % White (non-Hispanic, single race)
#> 54                                                         Ratio to State avg % Hispanic or Latino
#> 55                      Ratio to State avg % Black or African American (non-Hispanic, single race)
#> 56                                          Ratio to State avg % Asian (non-Hispanic, single race)
#> 57              Ratio to State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 58     Ratio to State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 59                                     Ratio to State avg % Other race (non-Hispanic, single race)
#> 60                                           Ratio to State avg % Two or more races (non-Hispanic)
#> 61                                          Ratio to State avg % White (non-Hispanic, single race)
#> 174                                                         US percentile for % Hispanic or Latino
#> 175                      US percentile for % Black or African American (non-Hispanic, single race)
#> 176                                          US percentile for % Asian (non-Hispanic, single race)
#> 177              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 178     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 179                                     US percentile for % Other race (non-Hispanic, single race)
#> 180                                           US percentile for % Two or more races (non-Hispanic)
#> 181                                          US percentile for % White (non-Hispanic, single race)
#> 211                                                      State percentile for % Hispanic or Latino
#> 212                   State percentile for % Black or African American (non-Hispanic, single race)
#> 213                                       State percentile for % Asian (non-Hispanic, single race)
#> 214           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 215  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 216                                  State percentile for % Other race (non-Hispanic, single race)
#> 217                                        State percentile for % Two or more races (non-Hispanic)
#> 218                                       State percentile for % White (non-Hispanic, single race)
#> 248                                                            US Average for % Hispanic or Latino
#> 249                         US Average for % Black or African American (non-Hispanic, single race)
#> 250                                             US Average for % Asian (non-Hispanic, single race)
#> 251                 US Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 252        US Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 253                                        US Average for % Other race (non-Hispanic, single race)
#> 254                                              US Average for % Two or more races (non-Hispanic)
#> 255                                             US Average for % White (non-Hispanic, single race)
#> 280                                                         State Average for % Hispanic or Latino
#> 281                      State Average for % Black or African American (non-Hispanic, single race)
#> 282                                          State Average for % Asian (non-Hispanic, single race)
#> 283              State Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 284     State Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 285                                     State Average for % Other race (non-Hispanic, single race)
#> 286                                           State Average for % Two or more races (non-Hispanic)
#> 287                                          State Average for % White (non-Hispanic, single race)
#> 310                                                                    Count of Hispanic or Latino
#> 311                                 Count of Black or African American (non-Hispanic, single race)
#> 312                                                     Count of Asian (non-Hispanic, single race)
#> 313                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 314                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 315                                                Count of Other race (non-Hispanic, single race)
#> 316                                                      Count of Two or more races (non-Hispanic)
#> 317                                                     Count of White (non-Hispanic, single race)
#> 105                                                       % speaking Non English languages at home
#> 106                                                         % speaking Other Indo-European at home
#> 107                                          % speaking Asian and Pacific Island languages at home
#> 108                                             % speaking Other and Unspecified languages at home
#> 109                                                                     % speaking English at home
#> 110                                                                     % speaking Spanish at home
#> 111                                                                      % speaking French at home
#> 112                                             % speaking Russian, Polish or Other Slavic at home
#> 113                                                               % speaking Indo-European at home
#> 114                                                                  % speaking Vietnamese at home
#> 115                                    % speaking Other Asian and Pacific Island languages at home
#> 116                                                                      % speaking Arabic at home
#> 337                                                            Number speaking non-English at home
#> 338                                                    Number speaking Other Indo-European at Home
#> 339                                          Number speaking Asian-Pacific Island language at Home
#> 340                                        Number speaking Other and Unspecified languages at home
#> 341                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 342                                                             Number speaking English Not at All
#> 343                                                                Number speaking Spanish at Home
#> 117                                        % speaking Spanish (as % of limited English households)
#> 118                  % speaking Other Indo-European languages (as % of limited English households)
#> 119                 % speaking Asian-Pacific Island languages (as % of limited English households)
#> 120                                % speaking Other languages (as % of limited English households)
#> 344                                         Number speaking Spanish (in limited English household)
#> 345                             Number speaking Other Indo-European (in limited English household)
#> 346                            Number speaking Asian-Pacific Island (in limited English household)
#> 347                                           Number speaking Other (in limited English household)
#> 6                                                                                 Total Population
#> 325                                                                  Count of Population Age 25 up
#> 326                                                                            Count of Households
#> 327                                                 Universe for % unemployed (denominator, count)
#> 328                                                          Count of Housing Units Built Pre 1960
#> 329                                         Built housing units count (denominator for % pre 1960)
#> 330                                      Count of Population for whom Poverty Status is Determined
#> 96                                                             % of Households below Poverty Level
#> 331                                                          Count of Owner Occupied Housing Units
#> 332                                                                 Households below Poverty Level
#> 121                                                                            % with Disabilities
#> 122                                                                     % with Low Life Expectancy
#> 123                                                                          % with Heart Diseases
#> 124                                                                        % of Adults with Asthma
#> 125                                                % of Adults with Cancer (excluding skin cancer)
#> 43                                                             Ratio to US avg % with Disabilities
#> 44                                                      Ratio to US avg % with Low Life Expectancy
#> 75                                                          Ratio to State avg % with Disabilities
#> 76                                                   Ratio to State avg % with Low Life Expectancy
#> 182                                                          US percentile for % with Disabilities
#> 183                                                   US percentile for % with Low Life Expectancy
#> 184                                                        US percentile for % with Heart Diseases
#> 185                                                      US percentile for % of Adults with Asthma
#> 186                              US percentile for % of Adults with Cancer (excluding skin cancer)
#> 219                                                       State percentile for % with Disabilities
#> 221                                                     State percentile for % with Heart Diseases
#> 222                                                   State percentile for % of Adults with Asthma
#> 223                           State percentile for % of Adults with Cancer (excluding skin cancer)
#> 256                                                             US Average for % with Disabilities
#> 257                                                      US Average for % with Low Life Expectancy
#> 288                                                          State Average for % with Disabilities
#> 289                                                   State Average for % with Low Life Expectancy
#> 403                                                               Count of Persons with Disability
#> 404                        Civilian noninstitutionalized population (denominator for % disability)
#> 126                                           Flag for Overlapping with Housing Burden Communities
#> 127                                                    Flag for Overlapping with Food Desert Areas
#> 128                                                        % Households without Broadband Internet
#> 129                                                          % Households without Health Insurance
#> 130                             Flag for Overlapping with Transportation Disadvantaged Communities
#> 100                                                                       Life Expectancy in Years
#> 220                                                State percentile for % with Low Life Expectancy
#> 133                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 134                                                                                    Ozone (ppb)
#> 135                                                                         Nitrogen Dioxide (NO2)
#> 136                                                              Diesel Particulate Matter (ug/m3)
#> 137                                                                          Toxic Releases to Air
#> 138                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 139                                                     Lead Paint Indicator (% pre-1960s housing)
#> 140                                                   Superfund Proximity (site count/km distance)
#> 141                                                     RMP Proximity (facility count/km distance)
#> 142                                         Hazardous Waste Proximity (facility count/km distance)
#> 143                                                      Underground Storage Tanks (UST) indicator
#> 144                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 145                                                                  Drinking Water Non-Compliance
#> 30                                                              Ratio to US avg Particulate Matter
#> 31                                                                           Ratio to US avg Ozone
#> 32                                                          Ratio to US avg Nitrogen Dioxide (NO2)
#> 33                                                       Ratio to US avg Diesel Particulate Matter
#> 34                                                           Ratio to US avg Toxic Releases to Air
#> 35                                                    Ratio to US avg Traffic Proximity and Volume
#> 36                                                            Ratio to US avg Lead Paint Indicator
#> 37                                                             Ratio to US avg Superfund Proximity
#> 38                                                                   Ratio to US avg RMP Proximity
#> 39                                                       Ratio to US avg Hazardous Waste Proximity
#> 40                                             Ratio to US avg Underground Storage Tanks indicator
#> 41                                                  Ratio to US avg Wastewater Discharge Indicator
#> 42                                                   Ratio to US avg Drinking Water Non-Compliance
#> 62                                                           Ratio to State avg Particulate Matter
#> 63                                                                        Ratio to State avg Ozone
#> 64                                                       Ratio to State avg Nitrogen Dioxide (NO2)
#> 65                                                    Ratio to State avg Diesel Particulate Matter
#> 66                                                        Ratio to State avg Toxic Releases to Air
#> 67                                                 Ratio to State avg Traffic Proximity and Volume
#> 68                                                         Ratio to State avg Lead Paint Indicator
#> 69                                                          Ratio to State avg Superfund Proximity
#> 70                                                                Ratio to State avg RMP Proximity
#> 71                                                    Ratio to State avg Hazardous Waste Proximity
#> 72                                          Ratio to State avg Underground Storage Tanks indicator
#> 73                                               Ratio to State avg Wastewater Discharge Indicator
#> 74                                                Ratio to State avg Drinking Water Non-Compliance
#> 189                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 190                                                                  US percentile for Ozone (ppb)
#> 191                                                       US percentile for Nitrogen Dioxide (NO2)
#> 192                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 193                                                        US percentile for Toxic Releases to Air
#> 194          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 195                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 196                                 US percentile for Superfund Proximity (site count/km distance)
#> 197                                   US percentile for RMP Proximity (facility count/km distance)
#> 198                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 199                                    US percentile for Underground Storage Tanks (UST) indicator
#> 200    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 201                                                US percentile for Drinking Water Non-Compliance
#> 226                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 227                                                               State percentile for Ozone (ppb)
#> 228                                                    State percentile for Nitrogen Dioxide (NO2)
#> 229                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 230                                                     State percentile for Toxic Releases to Air
#> 231       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 232                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 233                              State percentile for Superfund Proximity (site count/km distance)
#> 234                                State percentile for RMP Proximity (facility count/km distance)
#> 235                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 236                                 State percentile for Underground Storage Tanks (UST) indicator
#> 237 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 238                                             State percentile for Drinking Water Non-Compliance
#> 258                                            US Average for Particulate Matter (PM 2.5 in ug/m3)
#> 259                                                                     US Average for Ozone (ppb)
#> 260                                                          US Average for Nitrogen Dioxide (NO2)
#> 261                                               US Average for Diesel Particulate Matter (ug/m3)
#> 262                                                           US Average for Toxic Releases to Air
#> 263             US Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 264                                      US Average for Lead Paint Indicator (% pre-1960s housing)
#> 265                                    US Average for Superfund Proximity (site count/km distance)
#> 266                                      US Average for RMP Proximity (facility count/km distance)
#> 267                          US Average for Hazardous Waste Proximity (facility count/km distance)
#> 268                                       US Average for Underground Storage Tanks (UST) indicator
#> 269       US Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 270                                                   US Average for Drinking Water Non-Compliance
#> 290                                         State Average for Particulate Matter (PM 2.5 in ug/m3)
#> 291                                                                  State Average for Ozone (ppb)
#> 292                                                       State Average for Nitrogen Dioxide (NO2)
#> 293                                            State Average for Diesel Particulate Matter (ug/m3)
#> 294                                                        State Average for Toxic Releases to Air
#> 295          State Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 296                                   State Average for Lead Paint Indicator (% pre-1960s housing)
#> 297                                 State Average for Superfund Proximity (site count/km distance)
#> 298                                   State Average for RMP Proximity (facility count/km distance)
#> 299                       State Average for Hazardous Waste Proximity (facility count/km distance)
#> 300                                    State Average for Underground Storage Tanks (UST) indicator
#> 301    State Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 302                                                State Average for Drinking Water Non-Compliance
#> 131                                                                Estimated Fire Risk in 30 Years
#> 132                                                               Estimated Flood Risk in 30 Years
#> 187                                              US percentile for Estimated Fire Risk in 30 Years
#> 188                                             US percentile for Estimated Flood Risk in 30 Years
#> 224                                           State percentile for Estimated Fire Risk in 30 Years
#> 225                                          State percentile for Estimated Flood Risk in 30 Years
#> 161                                                          Number of Sites Nearby (total unique)
#> 164                                                        Distance to Closest Site for Avg Person
#> 162                                                                   Number of Sites Nearby (avg)
#> 160                                                                   Number of Sites Nearby (max)
#> 163                                                                       Distance to Closest Site
#> 415                                      US type of raw score for Particulate Matter Summary Index
#> 416                                                   US type of raw score for Ozone Summary Index
#> 417                                  US type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 418                               US type of raw score for Diesel Particulate Matter Summary Index
#> 419                                   US type of raw score for Toxic Releases to Air Summary Index
#> 420                            US type of raw score for Traffic Proximity and Volume Summary Index
#> 421                                              US type of raw score for Lead Paint Summary Index
#> 422                                     US type of raw score for Superfund Proximity Summary Index
#> 423                                           US type of raw score for RMP Proximity Summary Index
#> 424                               US type of raw score for Hazardous Waste Proximity Summary Index
#> 425                               US type of raw score for Underground Storage Tanks Summary Index
#> 426                                    US type of raw score for Wastewater Discharge Summary Index
#> 427                           US type of raw score for Drinking Water Non-Compliance Summary Index
#> 428                                   State type of raw score for Particulate Matter Summary Index
#> 429                                                State type of raw score for Ozone Summary Index
#> 430                               State type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 431                            State type of raw score for Diesel Particulate Matter Summary Index
#> 432                                State type of raw score for Toxic Releases to Air Summary Index
#> 433                         State type of raw score for Traffic Proximity and Volume Summary Index
#> 434                                           State type of raw score for Lead Paint Summary Index
#> 435                                  State type of raw score for Superfund Proximity Summary Index
#> 436                                        State type of raw score for RMP Proximity Summary Index
#> 437                            State type of raw score for Hazardous Waste Proximity Summary Index
#> 438                            State type of raw score for Underground Storage Tanks Summary Index
#> 439                                 State type of raw score for Wastewater Discharge Summary Index
#> 440                        State type of raw score for Drinking Water Non-Compliance Summary Index
#> 348                                         US percentile for Summary Index for Particulate Matter
#> 349                                                      US percentile for Summary Index for Ozone
#> 350                                     US percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 351                                  US percentile for Summary Index for Diesel Particulate Matter
#> 352                                      US percentile for Summary Index for Toxic Releases to Air
#> 353                               US percentile for Summary Index for Traffic Proximity and Volume
#> 354                                       US percentile for Summary Index for Lead Paint Indicator
#> 355                                        US percentile for Summary Index for Superfund Proximity
#> 356                                              US percentile for Summary Index for RMP Proximity
#> 357                                  US percentile for Summary Index for Hazardous Waste Proximity
#> 358                  US percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 359                             US percentile for Summary Index for Wastewater Discharge Indicator
#> 360                              US percentile for Summary Index for Drinking Water Non-Compliance
#> 374                                      State percentile for Summary Index for Particulate Matter
#> 375                                                   State percentile for Summary Index for Ozone
#> 376                                  State percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 377                               State percentile for Summary Index for Diesel Particulate Matter
#> 378                                   State percentile for Summary Index for Toxic Releases to Air
#> 379                            State percentile for Summary Index for Traffic Proximity and Volume
#> 380                                    State percentile for Summary Index for Lead Paint Indicator
#> 381                                     State percentile for Summary Index for Superfund Proximity
#> 382                                           State percentile for Summary Index for RMP Proximity
#> 383                               State percentile for Summary Index for Hazardous Waste Proximity
#> 384               State percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 385                          State percentile for Summary Index for Wastewater Discharge Indicator
#> 386                           State percentile for Summary Index for Drinking Water Non-Compliance
#> 441                         US type of raw score for Particulate Matter Supplemental Summary Index
#> 442                                      US type of raw score for Ozone Supplemental Summary Index
#> 443                     US type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 444                  US type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 445                      US type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 446               US type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 447                                 US type of raw score for Lead Paint Supplemental Summary Index
#> 448                        US type of raw score for Superfund Proximity Supplemental Summary Index
#> 449                              US type of raw score for RMP Proximity Supplemental Summary Index
#> 450                  US type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 451                  US type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 452                       US type of raw score for Wastewater Discharge Supplemental Summary Index
#> 453              US type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 454                      State type of raw score for Particulate Matter Supplemental Summary Index
#> 455                                   State type of raw score for Ozone Supplemental Summary Index
#> 456                  State type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 457               State type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 458                   State type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 459            State type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 460                              State type of raw score for Lead Paint Supplemental Summary Index
#> 461                     State type of raw score for Superfund Proximity Supplemental Summary Index
#> 462                           State type of raw score for RMP Proximity Supplemental Summary Index
#> 463               State type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 464               State type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 465                    State type of raw score for Wastewater Discharge Supplemental Summary Index
#> 466           State type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 361                            US percentile for Supplemental Summary Index for Particulate Matter
#> 362                                         US percentile for Supplemental Summary Index for Ozone
#> 363                        US percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 364                     US percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 365                         US percentile for Supplemental Summary Index for Toxic Releases to Air
#> 366                  US percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 367                          US percentile for Supplemental Summary Index for Lead Paint Indicator
#> 368                           US percentile for Supplemental Summary Index for Superfund Proximity
#> 369                                 US percentile for Supplemental Summary Index for RMP Proximity
#> 370                     US percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 371     US percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 372                US percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 373                 US percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 387                         State percentile for Supplemental Summary Index for Particulate Matter
#> 388                                      State percentile for Supplemental Summary Index for Ozone
#> 389                     State percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 390                  State percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 391                      State percentile for Supplemental Summary Index for Toxic Releases to Air
#> 392               State percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 393                       State percentile for Supplemental Summary Index for Lead Paint Indicator
#> 394                        State percentile for Supplemental Summary Index for Superfund Proximity
#> 395                              State percentile for Supplemental Summary Index for RMP Proximity
#> 396                  State percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 397  State percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 398             State percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 399              State percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 146                                                                              Number of Schools
#> 147                                                                       Number of Worship Places
#> 148                                                                            Number of Hospitals
#> 103                                                                                 % under age 18
#> 104                                                                                 % above age 17
#> 335                                                                         Population Over Age 17
#> 336                                                                        Population Under Age 18
#> 97                                                                          Occupied Housing Units
#> 98                                                                                         % Males
#> 99                                                                                       % Females
#> 101                                                                              Per Capita Income
#> 102                                                                    % Owner Occupied households
#> 333                                                                              Female Population
#> 334                                                                                Male Population
#> 149                                                               Flag for Overlapping with Tribes
#> 154                                                Count of National Priority List Superfund sites
#> 155                                          Count of Treatment Storage Disposal Facilities (TSDF)
#> 156                                                             Number of Air Pollution Facilities
#> 157                                                             Number of Toxic Release Facilities
#> 158                                                           Number of Water Discharge Facilities
#> 159                                                                          Number of Brownfields
#> 150                                                      Flag for Overlapping with Impaired Waters
#> 151                                                 Flag for Overlapping with Non-Attainment Areas
#> 152                                      Flag for Overlapping with CEJST Disadvantaged Communities
#> 153                                    Flag for Overlapping with EPA IRA Disadvantaged Communities
#> 467                                                   Area of Analyzed Location(s) in Square Miles
#> 7                                                                                        Longitude
#> 10                                                                                      State Name
#> 402                                                                        Buffer distance (miles)
#> 401                                                                          Census blocks (count)
#> 400                                                                     Census blockgroups (count)
#> 405                                                       Water area of block group in geodatabase
#> 9                                                                               State Abbreviation
#> 12                                                                                      EPA Region
#> 8                                                                                         Latitude
#> 406                                                        Land area of block group in geodatabase
#> 407                                                             Area of block group in geodatabase