How many decimal places to round to for given variable(s)

How many decimal places to round to for given variable(s)

Usage

table_rounding_info(var, varnametype = "rname")

Arguments

var

vector of variable names such as c("pctlowinc", "cancer") or c(names_d, names_d_subgroups)

varnametype

which column of map_headernames to use when looking for var, like "rname" or "api" or "long"

Value

named vector same size as var, with var as names.

See also

Examples

  table_rounding_info("pm")
#> [1] 2
  table_round(8.252345, "pm")
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> [1] 8.25
  table_round(8, "pm")
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> [1] 8
  
  cbind(table_rounding_info(names_all_r), fixcolnames(names_all_r, "r", "long"))
#>        [,1] [,2]                                                                                            
#>   [1,] "0"  "Total Population"                                                                              
#>   [2,] "0"  "Non-POC resident count"                                                                        
#>   [3,] "0"  "Count of Population for whom Poverty Status is Determined"                                     
#>   [4,] "0"  "Count of Population Age 25 up"                                                                 
#>   [5,] "0"  "Count of Households"                                                                           
#>   [6,] "0"  "Universe for percent unemployed (denominator, count)"                                          
#>   [7,] "0"  "Count of Housing Units Built Pre 1960"                                                         
#>   [8,] "0"  "Built housing units count (denominator for percent pre 1960)"                                  
#>   [9,] "0"  "Low income resident count"                                                                     
#>  [10,] "0"  "Limited English-speaking Households"                                                           
#>  [11,] "0"  "Unemployed resident count"                                                                     
#>  [12,] "0"  "Less Than High School Education resident count"                                                
#>  [13,] "0"  "Under Age 5 resident count"                                                                    
#>  [14,] "0"  "Over Age 64 resident count"                                                                    
#>  [15,] "0"  "People of Color resident count"                                                                
#>  [16,] "0"  "Demographic Index"                                                                             
#>  [17,] "0"  "Supplemental Demographic Index"                                                                
#>  [18,] "0"  "% Low Income"                                                                                  
#>  [19,] "0"  "% in limited English-speaking Households"                                                      
#>  [20,] "0"  "% Unemployed"                                                                                  
#>  [21,] "0"  "% with Less Than High School Education"                                                        
#>  [22,] "0"  "Low life expectancy"                                                                           
#>  [23,] "0"  "% under Age 5"                                                                                 
#>  [24,] "0"  "% over Age 64"                                                                                 
#>  [25,] "0"  "% People of Color"                                                                             
#>  [26,] "0"  "US Average of Demographic Index"                                                               
#>  [27,] "0"  "US Average of Supplemental Demographic Index"                                                  
#>  [28,] "0"  "US average for % Low Income"                                                                   
#>  [29,] "0"  "US average for % in limited English-speaking Households"                                       
#>  [30,] "0"  "US average for % Unemployed"                                                                   
#>  [31,] "0"  "US average for % with Less Than High School Education"                                         
#>  [32,] "0"  "US average for Low life expectancy"                                                            
#>  [33,] "0"  "US average for % under Age 5"                                                                  
#>  [34,] "0"  "US average for % over Age 64"                                                                  
#>  [35,] "0"  "US average for % People of Color"                                                              
#>  [36,] "0"  "State Average of Demographic Index"                                                            
#>  [37,] "0"  "State Average of Supplemental Demographic Index"                                               
#>  [38,] "0"  "State average for % Low Income"                                                                
#>  [39,] "0"  "State average for % in limited English-speaking Households"                                    
#>  [40,] "0"  "State average for % Unemployed"                                                                
#>  [41,] "0"  "State average for % with Less Than High School Education"                                      
#>  [42,] "0"  "State average for Low life expectancy"                                                         
#>  [43,] "0"  "State average for % under Age 5"                                                               
#>  [44,] "0"  "State average for % over Age 64"                                                               
#>  [45,] "0"  "State average for % People of Color"                                                           
#>  [46,] "0"  "US percentile for Demographic Index"                                                           
#>  [47,] "0"  "US percentile for Supplemental Demographic Index"                                              
#>  [48,] "0"  "US percentile for % Low Income"                                                                
#>  [49,] "0"  "US percentile for % in limited English-speaking Households"                                    
#>  [50,] "0"  "US percentile for % Unemployed"                                                                
#>  [51,] "0"  "US percentile for % with Less Than High School Education"                                      
#>  [52,] "0"  "US percentile for Low life expectancy"                                                         
#>  [53,] "0"  "US percentile for % under Age 5"                                                               
#>  [54,] "0"  "US percentile for % over Age 64"                                                               
#>  [55,] "0"  "US percentile for % People of Color"                                                           
#>  [56,] "0"  "State percentile for Demographic Index"                                                        
#>  [57,] "0"  "State percentile for Supplemental Demographic Index"                                           
#>  [58,] "0"  "State percentile for % Low Income"                                                             
#>  [59,] "0"  "State percentile for % in limited English-speaking Households"                                 
#>  [60,] "0"  "State percentile for % Unemployed"                                                             
#>  [61,] "0"  "State percentile for % with Less Than High School Education"                                   
#>  [62,] "0"  "State percentile for Low life expectancy"                                                      
#>  [63,] "0"  "State percentile for % under Age 5"                                                            
#>  [64,] "0"  "State percentile for % over Age 64"                                                            
#>  [65,] "0"  "State percentile for % People of Color"                                                        
#>  [66,] "1"  "Ratio to US avg Demog Index"                                                                   
#>  [67,] "1"  "Ratio to US avg Suppl Demog Index"                                                             
#>  [68,] "1"  "Ratio to US avg % Low-inc."                                                                    
#>  [69,] "1"  "Ratio to US avg % Limited English"                                                             
#>  [70,] "1"  "Ratio to US avg % Unemployed"                                                                  
#>  [71,] "1"  "Ratio to US avg % < High School"                                                               
#>  [72,] "1"  "Ratio to US avg Low life expectancy"                                                           
#>  [73,] "1"  "Ratio to US avg % < age 5"                                                                     
#>  [74,] "1"  "Ratio to US avg % > age 64"                                                                    
#>  [75,] "1"  "Ratio to US avg % People of Color"                                                             
#>  [76,] "1"  "Ratio to State avg Demog.Ind."                                                                 
#>  [77,] "1"  "Ratio to State avg Suppl Demog Index"                                                          
#>  [78,] "1"  "Ratio to State avg % Low-inc."                                                                 
#>  [79,] "1"  "Ratio to State avg % Limited English"                                                          
#>  [80,] "1"  "Ratio to State avg % Unemployed"                                                               
#>  [81,] "1"  "Ratio to State avg % < High School"                                                            
#>  [82,] "1"  "Ratio to State avg Low life expectancy"                                                        
#>  [83,] "1"  "Ratio to State avg % < age 5"                                                                  
#>  [84,] "1"  "Ratio to State avg % > age 64"                                                                 
#>  [85,] "1"  "Ratio to State avg % People of Color"                                                          
#>  [86,] "0"  "% Hispanic or Latino"                                                                          
#>  [87,] "0"  "% Black or African American (non-Hispanic, single race)"                                       
#>  [88,] "0"  "% Asian (non-Hispanic, single race)"                                                           
#>  [89,] "0"  "% American Indian and Alaska Native (non-Hispanic, single race)"                               
#>  [90,] "0"  "% Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"                      
#>  [91,] "0"  "% Other race (non-Hispanic, single race)"                                                      
#>  [92,] "0"  "% Two or more races (non-Hispanic)"                                                            
#>  [93,] "0"  "% White (non-Hispanic, single race)"                                                           
#>  [94,] "0"  "US avg % Hispanic or Latino"                                                                   
#>  [95,] "0"  "US avg % Black or African American (non-Hispanic, single race)"                                
#>  [96,] "0"  "US avg % Asian (non-Hispanic, single race)"                                                    
#>  [97,] "0"  "US avg % American Indian and Alaska Native (non-Hispanic, single race)"                        
#>  [98,] "0"  "US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"               
#>  [99,] "0"  "US avg % Other race (non-Hispanic, single race)"                                               
#> [100,] "0"  "US avg % Two or more races (non-Hispanic)"                                                     
#> [101,] "0"  "US avg % White (non-Hispanic, single race)"                                                    
#> [102,] "0"  "State avg % Hispanic or Latino"                                                                
#> [103,] "0"  "State avg % Black or African American (non-Hispanic, single race)"                             
#> [104,] "0"  "State avg % Asian (non-Hispanic, single race)"                                                 
#> [105,] "0"  "State avg % American Indian and Alaska Native (non-Hispanic, single race)"                     
#> [106,] "0"  "State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"            
#> [107,] "0"  "State avg % Other race (non-Hispanic, single race)"                                            
#> [108,] "0"  "State avg % Two or more races (non-Hispanic)"                                                  
#> [109,] "0"  "State avg % White (non-Hispanic, single race)"                                                 
#> [110,] "0"  "US percentile for % Hispanic or Latino"                                                        
#> [111,] "0"  "US percentile for % Black or African American (non-Hispanic, single race)"                     
#> [112,] "0"  "US percentile for % Asian (non-Hispanic, single race)"                                         
#> [113,] "0"  "US percentile for % American Indian and Alaska Native (non-Hispanic, single race)"             
#> [114,] "0"  "US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"    
#> [115,] "0"  "US percentile for % Other race (non-Hispanic, single race)"                                    
#> [116,] "0"  "US percentile for % Two or more races (non-Hispanic)"                                          
#> [117,] "0"  "US percentile for % White (non-Hispanic, single race)"                                         
#> [118,] "0"  "State percentile for % Hispanic or Latino"                                                     
#> [119,] "0"  "State percentile for % Black or African American (non-Hispanic, single race)"                  
#> [120,] "0"  "State percentile for % Asian (non-Hispanic, single race)"                                      
#> [121,] "0"  "State percentile for % American Indian and Alaska Native (non-Hispanic, single race)"          
#> [122,] "0"  "State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)" 
#> [123,] "0"  "State percentile for % Other race (non-Hispanic, single race)"                                 
#> [124,] "0"  "State percentile for % Two or more races (non-Hispanic)"                                       
#> [125,] "0"  "State percentile for % White (non-Hispanic, single race)"                                      
#> [126,] "1"  "Ratio to US avg % Hispanic"                                                                    
#> [127,] "1"  "Ratio to US avg % Black (NHA)"                                                                 
#> [128,] "1"  "Ratio to US avg % Asian (NHA)"                                                                 
#> [129,] "1"  "Ratio to US avg % American Indian Alaska Native (NHA)"                                         
#> [130,] "1"  "Ratio to US avg % Native Hawaiian Pacific Islander (NHA)"                                      
#> [131,] "1"  "Ratio to US avg % Other single race (NHA)"                                                     
#> [132,] "1"  "Ratio to US avg % Two or more race (NH)"                                                       
#> [133,] "1"  "Ratio to US avg % White nonHispanic single race (NHA)"                                         
#> [134,] "1"  "Ratio to State avg % Hispanic"                                                                 
#> [135,] "1"  "Ratio to State avg % American Indian Alaska Native (NHA)"                                      
#> [136,] "1"  "Ratio to State avg % Black (NHA)"                                                              
#> [137,] "1"  "Ratio to State avg % Asian (NHA)"                                                              
#> [138,] "1"  "Ratio to State avg % Other single race (NHA)"                                                  
#> [139,] "1"  "Ratio to State avg % Two or more race (NH)"                                                    
#> [140,] "1"  "Ratio to State avg % Native Hawaiian Pacific Islander (NHA)"                                   
#> [141,] "1"  "Ratio to State avg % White nonHispanic single race"                                            
#> [142,] "0"  "Count of Hispanic or Latino"                                                                   
#> [143,] "0"  "Count of Black or African American (non-Hispanic, single race)"                                
#> [144,] "0"  "Count of Asian (non-Hispanic, single race)"                                                    
#> [145,] "0"  "Count of American Indian and Alaska Native (non-Hispanic, single race)"                        
#> [146,] "0"  "Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"               
#> [147,] "0"  "Count of Other race (non-Hispanic, single race)"                                               
#> [148,] "0"  "Count of Two or more races (non-Hispanic)"                                                     
#> [149,] "0"  "Count of White (non-Hispanic, single race)"                                                    
#> [150,] "2"  "Particulate Matter (PM 2.5 in ug/m3)"                                                          
#> [151,] "1"  "Ozone (ppb)"                                                                                   
#> [152,] "0"  "Air Toxics Cancer Risk (risk per million)"                                                     
#> [153,] "2"  "Air Toxics Respiratory Hazard Index"                                                           
#> [154,] "2"  "Air Toxics Diesel Particulate Matter (ug/m3)"                                                  
#> [155,] "1"  "Lead Paint Indicator (% pre-1960s housing)"                                                    
#> [156,] "0"  "Traffic Proximity and Volume (daily traffic count/distance to road)"                           
#> [157,] "2"  "Superfund Proximity (site count/km distance)"                                                  
#> [158,] "2"  "RMP Proximity (facility count/km distance)"                                                    
#> [159,] "1"  "Indicator for Hazardous Waste Proximity (facility count/km distance)"                          
#> [160,] "0"  "Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"                     
#> [161,] "1"  "Underground Storage Tanks (UST) indicator"                                                     
#> [162,] "0"  "Toxic Releases to Air"                                                                         
#> [163,] "2"  "US average for Particulate Matter (PM 2.5 in ug/m3)"                                           
#> [164,] "1"  "US average for Ozone (ppb)"                                                                    
#> [165,] "0"  "US average for Air Toxics Cancer Risk (risk per million)"                                      
#> [166,] "2"  "US average for Air Toxics Respiratory Hazard Index"                                            
#> [167,] "2"  "US average for Air Toxics Diesel Particulate Matter (ug/m3)"                                   
#> [168,] "1"  "US average for Lead Paint Indicator (% pre-1960s housing)"                                     
#> [169,] "0"  "US average for Traffic Proximity and Volume (daily traffic count/distance to road)"            
#> [170,] "2"  "US average for Superfund Proximity (site count/km distance)"                                   
#> [171,] "2"  "US average for RMP Proximity (facility count/km distance)"                                     
#> [172,] "1"  "US average for Hazardous Waste Proximity (facility count/km distance)"                         
#> [173,] "0"  "US average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"      
#> [174,] "1"  "US average for Underground Storage Tanks (UST) indicator"                                      
#> [175,] "0"  "US avg Toxic Releases to Air"                                                                  
#> [176,] "2"  "State average for Particulate Matter (PM 2.5 in ug/m3)"                                        
#> [177,] "1"  "State average for Ozone (ppb)"                                                                 
#> [178,] "0"  "State average for Air Toxics Cancer Risk (risk per million)"                                   
#> [179,] "2"  "State average for Air Toxics Respiratory Hazard Index"                                         
#> [180,] "2"  "State average for Air Toxics Diesel Particulate Matter (ug/m3)"                                
#> [181,] "1"  "State average for Lead Paint Indicator (% pre-1960s housing)"                                  
#> [182,] "0"  "State average for Traffic Proximity and Volume (daily traffic count/distance to road)"         
#> [183,] "2"  "State average for Superfund Proximity (site count/km distance)"                                
#> [184,] "2"  "State average for RMP Proximity (facility count/km distance)"                                  
#> [185,] "1"  "State average for Hazardous Waste Proximity (facility count/km distance)"                      
#> [186,] "0"  "State average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"   
#> [187,] "1"  "State average for Underground Storage Tanks (UST) indicator"                                   
#> [188,] "0"  "State avg Toxic Releases to Air"                                                               
#> [189,] "0"  "US percentile for Particulate Matter (PM 2.5 in ug/m3)"                                        
#> [190,] "0"  "US percentile for Ozone (ppb)"                                                                 
#> [191,] "0"  "US percentile for Air Toxics Cancer Risk (risk per million)"                                   
#> [192,] "0"  "US percentile for Air Toxics Respiratory Hazard Index"                                         
#> [193,] "0"  "US percentile for Air Toxics Diesel Particulate Matter (ug/m3)"                                
#> [194,] "0"  "US percentile for Lead Paint Indicator (% pre-1960s housing)"                                  
#> [195,] "0"  "US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)"         
#> [196,] "0"  "US percentile for Superfund Proximity (site count/km distance)"                                
#> [197,] "0"  "US percentile for RMP Proximity (facility count/km distance)"                                  
#> [198,] "0"  "US percentile for Hazardous Waste Proximity (facility count/km distance)"                      
#> [199,] "0"  "US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"   
#> [200,] "0"  "US percentile for Underground Storage Tanks (UST) indicator"                                   
#> [201,] "0"  "US percentile for Toxic Releases to Air"                                                       
#> [202,] "0"  "State percentile for Particulate Matter (PM 2.5 in ug/m3)"                                     
#> [203,] "0"  "State percentile for Ozone (ppb)"                                                              
#> [204,] "0"  "State percentile for Air Toxics Cancer Risk (risk per million)"                                
#> [205,] "0"  "State percentile for Air Toxics Respiratory Hazard Index"                                      
#> [206,] "0"  "State percentile for Air Toxics Diesel Particulate Matter (ug/m3)"                             
#> [207,] "0"  "State percentile for Lead Paint Indicator (% pre-1960s housing)"                               
#> [208,] "0"  "State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)"      
#> [209,] "0"  "State percentile for Superfund Proximity (site count/km distance)"                             
#> [210,] "0"  "State percentile for RMP Proximity (facility count/km distance)"                               
#> [211,] "0"  "State percentile for Hazardous Waste Proximity (facility count/km distance)"                   
#> [212,] "0"  "State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"
#> [213,] "0"  "State percentile for Underground Storage Tanks (UST) indicator"                                
#> [214,] "0"  "State percentile for Toxic Releases to Air"                                                    
#> [215,] "1"  "Ratio to US avg PM2.5"                                                                         
#> [216,] "1"  "Ratio to US avg Ozone"                                                                         
#> [217,] "1"  "Ratio to US avg Cancer risk"                                                                   
#> [218,] "1"  "Ratio to US avg Respiratory"                                                                   
#> [219,] "1"  "Ratio to US avg Diesel PM"                                                                     
#> [220,] "1"  "Ratio to US avg % built pre-1960"                                                              
#> [221,] "1"  "Ratio to US avg Traffic"                                                                       
#> [222,] "1"  "Ratio to US avg NPL"                                                                           
#> [223,] "1"  "Ratio to US avg RMP"                                                                           
#> [224,] "1"  "Ratio to US avg TSDF"                                                                          
#> [225,] "1"  "Ratio to US avg NPDES"                                                                         
#> [226,] "1"  "Ratio to US avg UST"                                                                           
#> [227,] "1"  "Ratio to US avg Toxic Releases to Air"                                                         
#> [228,] "1"  "Ratio to State avg PM2.5"                                                                      
#> [229,] "1"  "Ratio to State avg Ozone"                                                                      
#> [230,] "1"  "Ratio to State avg Cancer risk"                                                                
#> [231,] "1"  "Ratio to State avg Respiratory"                                                                
#> [232,] "1"  "Ratio to State avg Diesel PM"                                                                  
#> [233,] "1"  "Ratio to State avg % built pre-1960"                                                           
#> [234,] "1"  "Ratio to State avg Traffic"                                                                    
#> [235,] "1"  "Ratio to State avg NPL"                                                                        
#> [236,] "1"  "Ratio to State avg RMP"                                                                        
#> [237,] "1"  "Ratio to State avg TSDF"                                                                       
#> [238,] "1"  "Ratio to State avg NPDES"                                                                      
#> [239,] "1"  "Ratio to State avg UST"                                                                        
#> [240,] "1"  "Ratio to State avg Toxic Releases to Air"                                                      
#> [241,] "3"  "Particulate Matter 2.5 EJ Index"                                                               
#> [242,] "3"  "Ozone EJ Index"                                                                                
#> [243,] "3"  "Air toxics cancer risk EJ Index"                                                               
#> [244,] "3"  "Air toxics respiratory HI EJ Index"                                                            
#> [245,] "3"  "Diesel particulate matter EJ Index"                                                            
#> [246,] "3"  "Lead paint EJ Index"                                                                           
#> [247,] "3"  "US raw Traffic proximity EJ Index"                                                             
#> [248,] "3"  "US raw Superfund Proximity EJ Index"                                                           
#> [249,] "3"  "RMP Facility Proximity EJ Index"                                                               
#> [250,] "3"  "Hazardous waste proximity EJ Index"                                                            
#> [251,] "3"  "US raw Wastewater discharge EJ Index"                                                          
#> [252,] "3"  "US raw Underground storage tanks EJ Index"                                                     
#> [253,] "3"  "Toxic Releases to Air EJ Index"                                                                
#> [254,] "3"  "State raw Particulate Matter 2.5 EJ Index"                                                     
#> [255,] "3"  "State raw Ozone EJ Index"                                                                      
#> [256,] "3"  "State Air toxics cancer risk EJ Index"                                                         
#> [257,] "3"  "State Air toxics respiratory HI EJ Index"                                                      
#> [258,] "3"  "State Diesel particulate matter EJ Index"                                                      
#> [259,] "3"  "State Lead paint EJ Index"                                                                     
#> [260,] "3"  "State raw Traffic proximity EJ Index"                                                          
#> [261,] "3"  "State raw Superfund Proximity EJ Index"                                                        
#> [262,] "3"  "State raw RMP Facility Proximity EJ Index"                                                     
#> [263,] "3"  "State Hazardous waste proximity EJ Index"                                                      
#> [264,] "3"  "State raw Wastewater discharge EJ Index"                                                       
#> [265,] "3"  "State raw Underground storage tanks EJ Index"                                                  
#> [266,] "3"  "State Toxic Releases to Air EJ Index"                                                          
#> [267,] "3"  "Particulate Matter 2.5 Supplemental Index"                                                     
#> [268,] "3"  "Ozone Supplemental Index"                                                                      
#> [269,] "3"  "Air Toxics Cancer Risk Supplemental Index"                                                     
#> [270,] "3"  "Air Toxics Respiratory HI Supplemental Index"                                                  
#> [271,] "3"  "Diesel Particulate Matter Supplemental Index"                                                  
#> [272,] "3"  "Lead Paint Supplemental Index"                                                                 
#> [273,] "3"  "Traffic Proximity Supplemental Index"                                                          
#> [274,] "3"  "Superfund Proximity Supplemental Index"                                                        
#> [275,] "3"  "RMP Facility Proximity Supplemental Index"                                                     
#> [276,] "3"  "Hazardous Waste Proximity Supplemental Index"                                                  
#> [277,] "3"  "Wastewater Discharge Supplemental Index"                                                       
#> [278,] "3"  "Underground Storage Tanks Supplemental Index"                                                  
#> [279,] "3"  "EJ Supp: Toxic Releases to Air (raw)"                                                          
#> [280,] "3"  "State raw Particulate Matter 2.5 Supplemental Index"                                           
#> [281,] "3"  "State raw Ozone Supplemental Index"                                                            
#> [282,] "3"  "State Air toxics cancer risk Supplemental Index"                                               
#> [283,] "3"  "State Air toxics respiratory HI Supplemental Index"                                            
#> [284,] "3"  "State Diesel particulate matter Supplemental Index"                                            
#> [285,] "3"  "State Lead paint Supplemental Index"                                                           
#> [286,] "3"  "State raw Traffic proximity Supplemental Index"                                                
#> [287,] "3"  "State raw Superfund Proximity Supplemental Index"                                              
#> [288,] "3"  "State raw RMP Facility Proximity Supplemental Index"                                           
#> [289,] "3"  "State Hazardous waste proximity Supplemental Index"                                            
#> [290,] "3"  "State raw Wastewater discharge Supplemental Index"                                             
#> [291,] "3"  "State raw Underground storage tanks Supplemental Index"                                        
#> [292,] "3"  "State Toxic Releases to Air Supplemental Index"                                                
#> [293,] "0"  "US percentile for EJ Index for Particulate Matter (PM 2.5)"                                    
#> [294,] "0"  "US percentile for EJ Index for Ozone"                                                          
#> [295,] "0"  "US percentile for EJ Index for Air Toxics Cancer Risk"                                         
#> [296,] "0"  "US percentile for EJ Index for Air Toxics Respiratory Hazard Index"                            
#> [297,] "0"  "US percentile for EJ Index for Air Toxics Diesel Particulate Matter"                           
#> [298,] "0"  "US percentile for EJ Index for Lead Paint Indicator"                                           
#> [299,] "0"  "US percentile for EJ Index for Traffic Proximity and Volume"                                   
#> [300,] "0"  "US percentile for EJ Index for Superfund Proximity"                                            
#> [301,] "0"  "US percentile for EJ Index for RMP Proximity"                                                  
#> [302,] "0"  "US percentile for EJ Index for Hazardous Waste Proximity"                                      
#> [303,] "0"  "US percentile for EJ Index for Wastewater Discharge Indicator"                                 
#> [304,] "0"  "US percentile for EJ Index for Underground Storage Tanks (UST) indicator"                      
#> [305,] "0"  "US percentile for EJ Index for Toxic Releases to Air"                                          
#> [306,] "0"  "State percentile for EJ Index for Particulate Matter (PM 2.5)"                                 
#> [307,] "0"  "State percentile for EJ Index for Ozone"                                                       
#> [308,] "0"  "State percentile for EJ Index for Air Toxics Cancer Risk"                                      
#> [309,] "0"  "State percentile for EJ Index for Air Toxics Respiratory Hazard Index"                         
#> [310,] "0"  "State percentile for EJ Index for Air Toxics Diesel Particulate Matter"                        
#> [311,] "0"  "State percentile for EJ Index for Lead Paint Indicator"                                        
#> [312,] "0"  "State percentile for EJ Index for Traffic Proximity and Volume"                                
#> [313,] "0"  "State percentile for EJ Index for Superfund Proximity"                                         
#> [314,] "0"  "State percentile for EJ Index for RMP Proximity"                                               
#> [315,] "0"  "State percentile for EJ Index for Hazardous Waste Proximity"                                   
#> [316,] "0"  "State percentile for EJ Index for Wastewater Discharge Indicator"                              
#> [317,] "0"  "State percentile for EJ Index for Underground Storage Tanks (UST) indicator"                   
#> [318,] "0"  "State percentile for EJ Index for Toxic Releases to Air"                                       
#> [319,] "0"  "US percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)"                       
#> [320,] "0"  "US percentile for EJ Supplemental Index for Ozone"                                             
#> [321,] "0"  "US percentile for EJ Supplemental Index for Air Toxics Cancer Risk"                            
#> [322,] "0"  "US percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index"               
#> [323,] "0"  "US percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter"              
#> [324,] "0"  "US percentile for EJ Supplemental Index for Lead Paint Indicator"                              
#> [325,] "0"  "US percentile for EJ Supplemental Index for Traffic Proximity and Volume"                      
#> [326,] "0"  "US percentile for EJ Supplemental Index for Superfund Proximity"                               
#> [327,] "0"  "US percentile for EJ Supplemental Index for RMP Proximity"                                     
#> [328,] "0"  "US percentile for EJ Supplemental Index for Hazardous Waste Proximity"                         
#> [329,] "0"  "US percentile for EJ Supplemental Index for Wastewater Discharge Indicator"                    
#> [330,] "0"  "US percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator"         
#> [331,] "0"  "US Percentile for Toxic Releases to Air Supplemental Index"                                    
#> [332,] "0"  "State percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)"                    
#> [333,] "0"  "State percentile for EJ Supplemental Index for Ozone"                                          
#> [334,] "0"  "State percentile for EJ Supplemental Index for Air Toxics Cancer Risk"                         
#> [335,] "0"  "State percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index"            
#> [336,] "0"  "State percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter"           
#> [337,] "0"  "State percentile for EJ Supplemental Index for Lead Paint Indicator"                           
#> [338,] "0"  "State percentile for EJ Supplemental Index for Traffic Proximity and Volume"                   
#> [339,] "0"  "State percentile for EJ Supplemental Index for Superfund Proximity"                            
#> [340,] "0"  "State percentile for EJ Supplemental Index for RMP Proximity"                                  
#> [341,] "0"  "State percentile for EJ Supplemental Index for Hazardous Waste Proximity"                      
#> [342,] "0"  "State percentile for EJ Supplemental Index for Wastewater Discharge Indicator"                 
#> [343,] "0"  "State percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator"      
#> [344,] "0"  "State Percentile for Toxic Releases to Air Supplemental Index"