Format the results_overall part of the output of ejamit() or doaggregate()

Take a quick look at results in the RStudio console

Usage

table_tall_from_overall(results_overall, longnames)

Arguments

results_overall

data.table of 1 row, from output of ejamit() or doaggregate()

longnames

vector of names of variables in results_overall, from output of ejamit() or doaggregate()

Value

data.table that is one row per indicator

Examples

 table_tall_from_overall(testoutput_ejamit_10pts_1miles$results_overall)
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#>                  value
#> 1                 <NA>
#> 2                 <NA>
#> 3                 <NA>
#> 4                 <NA>
#> 5                60740
#> 6                 <NA>
#> 7                 <NA>
#> 8                 <NA>
#> 9                 <NA>
#> 10                 1.2
#> 11                 1.2
#> 12                 1.3
#> 13                 1.4
#> 14                   1
#> 15                 1.3
#> 16                 1.1
#> 17                 1.2
#> 18                 0.6
#> 19                   1
#> 20                 1.4
#> 21                 0.7
#> 22                 0.6
#> 23                 1.4
#> 24                   0
#> 25                 0.3
#> 26                 0.9
#> 27                   1
#> 28                   1
#> 29                   1
#> 30                 1.1
#> 31                 0.9
#> 32                 1.4
#> 33                 1.4
#> 34                 0.8
#> 35                 2.9
#> 36                 1.4
#> 37                 0.7
#> 38                   0
#> 39                 1.4
#> 40                 1.4
#> 41                 1.3
#> 42                 1.3
#> 43                 1.3
#> 44                 1.9
#> 45                 1.1
#> 46                 1.5
#> 47                 1.1
#> 48                 1.1
#> 49                 0.6
#> 50                 1.3
#> 51                 1.8
#> 52                 0.8
#> 53                 1.3
#> 54                 0.8
#> 55                   0
#> 56                 0.3
#> 57                   1
#> 58                 0.9
#> 59                   1
#> 60                   1
#> 61                 1.2
#> 62                 1.1
#> 63                 1.5
#> 64                 1.4
#> 65                 1.4
#> 66                 4.3
#> 67                 1.2
#> 68                 1.6
#> 69                   0
#> 70                 1.4
#> 71                 1.1
#> 72                   0
#> 73                   0
#> 74                   0
#> 75                   0
#> 76                   0
#> 77                   0
#> 78                   0
#> 79                   0
#> 80                   0
#> 81                   0
#> 82                   0
#> 83                   0
#> 84                   0
#> 85                   0
#> 86                   0
#> 87                   0
#> 88                   0
#> 89                   1
#> 90                  64
#> 91                  72
#> 92                  69
#> 93                  80
#> 94                  63
#> 95                  71
#> 96                  70
#> 97                  64
#> 98                  30
#> 99                  59
#> 100                 76
#> 101                 68
#> 102                 68
#> 103                 90
#> 104                  0
#> 105                 87
#> 106                 64
#> 107                 40
#> 108                  0
#> 109                  0
#> 110                  0
#> 111                  0
#> 112                  0
#> 113                  0
#> 114                  0
#> 115                  0
#> 116                  0
#> 117                  0
#> 118                  0
#> 119                  0
#> 120                  0
#> 121                  0
#> 122                  0
#> 123                  0
#> 124                  0
#> 125                  1
#> 126                  0
#> 127                  0
#> 128                  0
#> 129                  0
#> 130                  0
#> 131                  0
#> 132                  0
#> 133                  1
#> 134                 69
#> 135                 71
#> 136                 66
#> 137                 81
#> 138                 65
#> 139                 74
#> 140                 67
#> 141                 63
#> 142                 28
#> 143                 67
#> 144                 76
#> 145                 62
#> 146                 73
#> 147                 52
#> 148                  0
#> 149                 25
#> 150                 59
#> 151                 32
#> 152                  0
#> 153                  0
#> 154                  0
#> 155                  0
#> 156                  0
#> 157                  0
#> 158                  0
#> 159                  0
#> 160                  0
#> 161                  0
#> 162               8.03
#> 163               60.6
#> 164                 27
#> 165               0.29
#> 166               0.36
#> 167               41.3
#> 168                179
#> 169               0.39
#> 170                0.6
#> 171                1.4
#> 172                  0
#> 173                5.5
#> 174               6493
#> 175                 45
#> 176                 46
#> 177                  5
#> 178                  4
#> 179                 79
#> 180                 67
#> 181                 72
#> 182                 93
#> 183                 79
#> 184                 66
#> 185                 70
#> 186                 79
#> 187                 89
#> 188               8.08
#> 189               61.6
#> 190                 25
#> 191               0.31
#> 192               0.26
#> 193               30.3
#> 194                212
#> 195               0.13
#> 196               0.43
#> 197                1.9
#> 198                 22
#> 199                3.9
#> 200               4612
#> 201                 67
#> 202                 53
#> 203                 21
#> 204                 37
#> 205                 78
#> 206                 65
#> 207                 75
#> 208                 74
#> 209                 69
#> 210                 72
#> 211                 55
#> 212                 76
#> 213                 78
#> 214               7.89
#> 215               60.6
#> 216                 22
#> 217               0.26
#> 218               0.25
#> 219               29.1
#> 220                126
#> 221               0.09
#> 222               0.48
#> 223                0.8
#> 224                 25
#> 225                3.9
#> 226               5917
#> 227                 69
#> 228                 63
#> 229                 71
#> 230                 64
#> 231                 77
#> 232                 70
#> 233                 70
#> 234                 79
#> 235                 77
#> 236                 69
#> 237                 70
#> 238                 75
#> 239                 81
#> 240                 74
#> 241                 69
#> 242                 73
#> 243                 66
#> 244                 84
#> 245                 74
#> 246                 76
#> 247                 85
#> 248                 82
#> 249                 76
#> 250                 75
#> 251                 78
#> 252                 85
#> 253                 72
#> 254                 64
#> 255                 81
#> 256                 75
#> 257                 77
#> 258                 67
#> 259                 74
#> 260                 72
#> 261                 74
#> 262                 72
#> 263                 66
#> 264                 73
#> 265                 75
#> 266                 73
#> 267                 65
#> 268                 81
#> 269                 77
#> 270                 80
#> 271                 67
#> 272                 77
#> 273                 74
#> 274                 75
#> 275                 74
#> 276                 66
#> 277                 74
#> 278                 78
#> 279              21641
#> 280               1477
#> 281               1738
#> 282               5487
#> 283               3902
#> 284               6580
#> 285              24807
#> 286              15471
#> 287               5218
#> 288               1803
#> 289                537
#> 290                  1
#> 291                 55
#> 292               1723
#> 293              35933
#> 294              54648
#> 295              36889
#> 296              21660
#> 297              30403
#> 298              10184
#> 299              24671
#> 300              56838
#> 301              11331
#> 302                707
#> 303               1153
#> 304               1367
#> 305               1121
#> 306                179
#> 307                165
#> 308                 12
#> 309              13015
#> 310              47725
#> 311              30954
#> 312              29786
#> 313              11127
#> 314              21660
#> 315              58670
#> 316             6852.1
#> 317                 76
#> 318              27397
#> 319                  0
#> 320                  1
#> 321                  0
#> 322                  0
#> 323                  0
#> 324                  0
#> 325                  1
#> 326                  1
#> 327                  0
#> 328                 75
#> 329               1152
#> 330                  1
#> 331               4126
#> 332             21.749
#> 333              18.28
#> 334             16.944
#> 335             14.322
#> 336             31.306
#> 337             21.888
#> 338              25.68
#> 339             31.366
#> 340             30.822
#> 341             24.123
#> 342             22.239
#> 343             28.501
#> 344             30.878
#> 345             10.022
#> 346               8.39
#> 347              6.944
#> 348              5.949
#> 349             13.462
#> 350             10.484
#> 351             11.215
#> 352               13.5
#> 353             13.284
#> 354              10.79
#> 355             10.218
#> 356             12.078
#> 357             13.612
#> 358             27.806
#> 359             19.867
#> 360             21.399
#> 361             17.843
#> 362             32.732
#> 363             23.756
#> 364             27.033
#> 365             33.055
#> 366             30.084
#> 367             29.713
#> 368             21.378
#> 369             26.861
#> 370             31.691
#> 371             12.478
#> 372              9.358
#> 373              9.181
#> 374              8.069
#> 375             14.214
#> 376             11.008
#> 377              11.99
#> 378             14.316
#> 379             12.752
#> 380             13.015
#> 381              9.696
#> 382             11.524
#> 383             13.853
#> 384              0.018
#> 385  0.751551987007954
#> 386                  1
#> 387                  9
#> 388                  1
#> 389                  0
#> 390                  1
#> 391 0.0124440738023481
#> 392  0.199349734246098
#> 393  0.020282098514558
#> 394 0.0240542028878415
#>                                                                                          indicator
#> 1                                                                                  EJScreen Report
#> 2                                                                                     EJScreen Map
#> 3                                                                                      ECHO report
#> 4                                                                                     ejam_uniq_id
#> 5                                                                                 Total Population
#> 6                                                                                        Longitude
#> 7                                                                                         Latitude
#> 8                                                                               State Abbreviation
#> 9                                                                                       State Name
#> 10                                                                     Ratio to US avg Demog Index
#> 11                                                               Ratio to US avg Suppl Demog Index
#> 12                                                                      Ratio to US avg % Low-inc.
#> 13                                                               Ratio to US avg % Limited English
#> 14                                                                    Ratio to US avg % Unemployed
#> 15                                                                 Ratio to US avg % < High School
#> 16                                                             Ratio to US avg Low life expectancy
#> 17                                                                       Ratio to US avg % < age 5
#> 18                                                                      Ratio to US avg % > age 64
#> 19                                                               Ratio to US avg % People of Color
#> 20                                                                      Ratio to US avg % Hispanic
#> 21                                                                   Ratio to US avg % Black (NHA)
#> 22                                                                   Ratio to US avg % Asian (NHA)
#> 23                                           Ratio to US avg % American Indian Alaska Native (NHA)
#> 24                                        Ratio to US avg % Native Hawaiian Pacific Islander (NHA)
#> 25                                                       Ratio to US avg % Other single race (NHA)
#> 26                                                         Ratio to US avg % Two or more race (NH)
#> 27                                           Ratio to US avg % White nonHispanic single race (NHA)
#> 28                                                                           Ratio to US avg PM2.5
#> 29                                                                           Ratio to US avg Ozone
#> 30                                                                     Ratio to US avg Cancer risk
#> 31                                                                     Ratio to US avg Respiratory
#> 32                                                                       Ratio to US avg Diesel PM
#> 33                                                                Ratio to US avg % built pre-1960
#> 34                                                                         Ratio to US avg Traffic
#> 35                                                                             Ratio to US avg NPL
#> 36                                                                             Ratio to US avg RMP
#> 37                                                                            Ratio to US avg TSDF
#> 38                                                                           Ratio to US avg NPDES
#> 39                                                                             Ratio to US avg UST
#> 40                                                           Ratio to US avg Toxic Releases to Air
#> 41                                                                   Ratio to State avg Demog.Ind.
#> 42                                                            Ratio to State avg Suppl Demog Index
#> 43                                                                   Ratio to State avg % Low-inc.
#> 44                                                            Ratio to State avg % Limited English
#> 45                                                                 Ratio to State avg % Unemployed
#> 46                                                              Ratio to State avg % < High School
#> 47                                                          Ratio to State avg Low life expectancy
#> 48                                                                    Ratio to State avg % < age 5
#> 49                                                                   Ratio to State avg % > age 64
#> 50                                                            Ratio to State avg % People of Color
#> 51                                                                   Ratio to State avg % Hispanic
#> 52                                        Ratio to State avg % American Indian Alaska Native (NHA)
#> 53                                                                Ratio to State avg % Black (NHA)
#> 54                                                                Ratio to State avg % Asian (NHA)
#> 55                                                    Ratio to State avg % Other single race (NHA)
#> 56                                                      Ratio to State avg % Two or more race (NH)
#> 57                                     Ratio to State avg % Native Hawaiian Pacific Islander (NHA)
#> 58                                              Ratio to State avg % White nonHispanic single race
#> 59                                                                        Ratio to State avg PM2.5
#> 60                                                                        Ratio to State avg Ozone
#> 61                                                                  Ratio to State avg Cancer risk
#> 62                                                                  Ratio to State avg Respiratory
#> 63                                                                    Ratio to State avg Diesel PM
#> 64                                                             Ratio to State avg % built pre-1960
#> 65                                                                      Ratio to State avg Traffic
#> 66                                                                          Ratio to State avg NPL
#> 67                                                                          Ratio to State avg RMP
#> 68                                                                         Ratio to State avg TSDF
#> 69                                                                        Ratio to State avg NPDES
#> 70                                                                          Ratio to State avg UST
#> 71                                                        Ratio to State avg Toxic Releases to Air
#> 72                                                                               Demographic Index
#> 73                                                                  Supplemental Demographic Index
#> 74                                                                                    % Low Income
#> 75                                                        % in limited English-speaking Households
#> 76                                                                                    % Unemployed
#> 77                                                          % with Less Than High School Education
#> 78                                                                             Low life expectancy
#> 79                                                                                   % under Age 5
#> 80                                                                                   % over Age 64
#> 81                                                                               % People of Color
#> 82                                                                            % Hispanic or Latino
#> 83                                         % Black or African American (non-Hispanic, single race)
#> 84                                                             % Asian (non-Hispanic, single race)
#> 85                                 % American Indian and Alaska Native (non-Hispanic, single race)
#> 86                        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 87                                                        % Other race (non-Hispanic, single race)
#> 88                                                              % Two or more races (non-Hispanic)
#> 89                                                             % White (non-Hispanic, single race)
#> 90                                                             US percentile for Demographic Index
#> 91                                                US percentile for Supplemental Demographic Index
#> 92                                                                  US percentile for % Low Income
#> 93                                      US percentile for % in limited English-speaking Households
#> 94                                                                  US percentile for % Unemployed
#> 95                                        US percentile for % with Less Than High School Education
#> 96                                                           US percentile for Low life expectancy
#> 97                                                                 US percentile for % under Age 5
#> 98                                                                 US percentile for % over Age 64
#> 99                                                             US percentile for % People of Color
#> 100                                                         US percentile for % Hispanic or Latino
#> 101                      US percentile for % Black or African American (non-Hispanic, single race)
#> 102                                          US percentile for % Asian (non-Hispanic, single race)
#> 103              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 104     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 105                                     US percentile for % Other race (non-Hispanic, single race)
#> 106                                           US percentile for % Two or more races (non-Hispanic)
#> 107                                          US percentile for % White (non-Hispanic, single race)
#> 108                                                                US Average of Demographic Index
#> 109                                                   US Average of Supplemental Demographic Index
#> 110                                                                    US average for % Low Income
#> 111                                        US average for % in limited English-speaking Households
#> 112                                                                    US average for % Unemployed
#> 113                                          US average for % with Less Than High School Education
#> 114                                                             US average for Low life expectancy
#> 115                                                                   US average for % under Age 5
#> 116                                                                   US average for % over Age 64
#> 117                                                               US average for % People of Color
#> 118                                                                    US avg % Hispanic or Latino
#> 119                                 US avg % Black or African American (non-Hispanic, single race)
#> 120                                                     US avg % Asian (non-Hispanic, single race)
#> 121                         US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 122                US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 123                                                US avg % Other race (non-Hispanic, single race)
#> 124                                                      US avg % Two or more races (non-Hispanic)
#> 125                                                     US avg % White (non-Hispanic, single race)
#> 126                                                                 State avg % Hispanic or Latino
#> 127                              State avg % Black or African American (non-Hispanic, single race)
#> 128                                                  State avg % Asian (non-Hispanic, single race)
#> 129                      State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 130             State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 131                                             State avg % Other race (non-Hispanic, single race)
#> 132                                                   State avg % Two or more races (non-Hispanic)
#> 133                                                  State avg % White (non-Hispanic, single race)
#> 134                                                         State percentile for Demographic Index
#> 135                                            State percentile for Supplemental Demographic Index
#> 136                                                              State percentile for % Low Income
#> 137                                  State percentile for % in limited English-speaking Households
#> 138                                                              State percentile for % Unemployed
#> 139                                    State percentile for % with Less Than High School Education
#> 140                                                       State percentile for Low life expectancy
#> 141                                                             State percentile for % under Age 5
#> 142                                                             State percentile for % over Age 64
#> 143                                                         State percentile for % People of Color
#> 144                                                      State percentile for % Hispanic or Latino
#> 145                   State percentile for % Black or African American (non-Hispanic, single race)
#> 146                                       State percentile for % Asian (non-Hispanic, single race)
#> 147           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 148  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 149                                  State percentile for % Other race (non-Hispanic, single race)
#> 150                                        State percentile for % Two or more races (non-Hispanic)
#> 151                                       State percentile for % White (non-Hispanic, single race)
#> 152                                                             State Average of Demographic Index
#> 153                                                State Average of Supplemental Demographic Index
#> 154                                                                 State average for % Low Income
#> 155                                     State average for % in limited English-speaking Households
#> 156                                                                 State average for % Unemployed
#> 157                                       State average for % with Less Than High School Education
#> 158                                                          State average for Low life expectancy
#> 159                                                                State average for % under Age 5
#> 160                                                                State average for % over Age 64
#> 161                                                            State average for % People of Color
#> 162                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 163                                                                                    Ozone (ppb)
#> 164                                                      Air Toxics Cancer Risk (risk per million)
#> 165                                                            Air Toxics Respiratory Hazard Index
#> 166                                                   Air Toxics Diesel Particulate Matter (ug/m3)
#> 167                                                     Lead Paint Indicator (% pre-1960s housing)
#> 168                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 169                                                   Superfund Proximity (site count/km distance)
#> 170                                                     RMP Proximity (facility count/km distance)
#> 171                           Indicator for Hazardous Waste Proximity (facility count/km distance)
#> 172                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 173                                                      Underground Storage Tanks (UST) indicator
#> 174                                                                          Toxic Releases to Air
#> 175                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 176                                                                  US percentile for Ozone (ppb)
#> 177                                    US percentile for Air Toxics Cancer Risk (risk per million)
#> 178                                          US percentile for Air Toxics Respiratory Hazard Index
#> 179                                 US percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 180                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 181          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 182                                 US percentile for Superfund Proximity (site count/km distance)
#> 183                                   US percentile for RMP Proximity (facility count/km distance)
#> 184                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 185    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 186                                    US percentile for Underground Storage Tanks (UST) indicator
#> 187                                                        US percentile for Toxic Releases to Air
#> 188                                            US average for Particulate Matter (PM 2.5 in ug/m3)
#> 189                                                                     US average for Ozone (ppb)
#> 190                                       US average for Air Toxics Cancer Risk (risk per million)
#> 191                                             US average for Air Toxics Respiratory Hazard Index
#> 192                                    US average for Air Toxics Diesel Particulate Matter (ug/m3)
#> 193                                      US average for Lead Paint Indicator (% pre-1960s housing)
#> 194             US average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 195                                    US average for Superfund Proximity (site count/km distance)
#> 196                                      US average for RMP Proximity (facility count/km distance)
#> 197                          US average for Hazardous Waste Proximity (facility count/km distance)
#> 198       US average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 199                                       US average for Underground Storage Tanks (UST) indicator
#> 200                                                                   US avg Toxic Releases to Air
#> 201                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 202                                                               State percentile for Ozone (ppb)
#> 203                                 State percentile for Air Toxics Cancer Risk (risk per million)
#> 204                                       State percentile for Air Toxics Respiratory Hazard Index
#> 205                              State percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 206                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 207       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 208                              State percentile for Superfund Proximity (site count/km distance)
#> 209                                State percentile for RMP Proximity (facility count/km distance)
#> 210                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 211 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 212                                 State percentile for Underground Storage Tanks (UST) indicator
#> 213                                                     State percentile for Toxic Releases to Air
#> 214                                         State average for Particulate Matter (PM 2.5 in ug/m3)
#> 215                                                                  State average for Ozone (ppb)
#> 216                                    State average for Air Toxics Cancer Risk (risk per million)
#> 217                                          State average for Air Toxics Respiratory Hazard Index
#> 218                                 State average for Air Toxics Diesel Particulate Matter (ug/m3)
#> 219                                   State average for Lead Paint Indicator (% pre-1960s housing)
#> 220          State average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 221                                 State average for Superfund Proximity (site count/km distance)
#> 222                                   State average for RMP Proximity (facility count/km distance)
#> 223                       State average for Hazardous Waste Proximity (facility count/km distance)
#> 224    State average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 225                                    State average for Underground Storage Tanks (UST) indicator
#> 226                                                                State avg Toxic Releases to Air
#> 227                                     US percentile for EJ Index for Particulate Matter (PM 2.5)
#> 228                                                           US percentile for EJ Index for Ozone
#> 229                                          US percentile for EJ Index for Air Toxics Cancer Risk
#> 230                             US percentile for EJ Index for Air Toxics Respiratory Hazard Index
#> 231                            US percentile for EJ Index for Air Toxics Diesel Particulate Matter
#> 232                                            US percentile for EJ Index for Lead Paint Indicator
#> 233                                    US percentile for EJ Index for Traffic Proximity and Volume
#> 234                                             US percentile for EJ Index for Superfund Proximity
#> 235                                                   US percentile for EJ Index for RMP Proximity
#> 236                                       US percentile for EJ Index for Hazardous Waste Proximity
#> 237                                  US percentile for EJ Index for Wastewater Discharge Indicator
#> 238                       US percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 239                                           US percentile for EJ Index for Toxic Releases to Air
#> 240                        US percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)
#> 241                                              US percentile for EJ Supplemental Index for Ozone
#> 242                             US percentile for EJ Supplemental Index for Air Toxics Cancer Risk
#> 243                US percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index
#> 244               US percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter
#> 245                               US percentile for EJ Supplemental Index for Lead Paint Indicator
#> 246                       US percentile for EJ Supplemental Index for Traffic Proximity and Volume
#> 247                                US percentile for EJ Supplemental Index for Superfund Proximity
#> 248                                      US percentile for EJ Supplemental Index for RMP Proximity
#> 249                          US percentile for EJ Supplemental Index for Hazardous Waste Proximity
#> 250                     US percentile for EJ Supplemental Index for Wastewater Discharge Indicator
#> 251          US percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator
#> 252                                     US Percentile for Toxic Releases to Air Supplemental Index
#> 253                                  State percentile for EJ Index for Particulate Matter (PM 2.5)
#> 254                                                        State percentile for EJ Index for Ozone
#> 255                                       State percentile for EJ Index for Air Toxics Cancer Risk
#> 256                          State percentile for EJ Index for Air Toxics Respiratory Hazard Index
#> 257                         State percentile for EJ Index for Air Toxics Diesel Particulate Matter
#> 258                                         State percentile for EJ Index for Lead Paint Indicator
#> 259                                 State percentile for EJ Index for Traffic Proximity and Volume
#> 260                                          State percentile for EJ Index for Superfund Proximity
#> 261                                                State percentile for EJ Index for RMP Proximity
#> 262                                    State percentile for EJ Index for Hazardous Waste Proximity
#> 263                               State percentile for EJ Index for Wastewater Discharge Indicator
#> 264                    State percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 265                                        State percentile for EJ Index for Toxic Releases to Air
#> 266                     State percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)
#> 267                                           State percentile for EJ Supplemental Index for Ozone
#> 268                          State percentile for EJ Supplemental Index for Air Toxics Cancer Risk
#> 269             State percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index
#> 270            State percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter
#> 271                            State percentile for EJ Supplemental Index for Lead Paint Indicator
#> 272                    State percentile for EJ Supplemental Index for Traffic Proximity and Volume
#> 273                             State percentile for EJ Supplemental Index for Superfund Proximity
#> 274                                   State percentile for EJ Supplemental Index for RMP Proximity
#> 275                       State percentile for EJ Supplemental Index for Hazardous Waste Proximity
#> 276                  State percentile for EJ Supplemental Index for Wastewater Discharge Indicator
#> 277       State percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator
#> 278                                  State Percentile for Toxic Releases to Air Supplemental Index
#> 279                                                                      Low income resident count
#> 280                                                            Limited English-speaking Households
#> 281                                                                      Unemployed resident count
#> 282                                                 Less Than High School Education resident count
#> 283                                                                     Under Age 5 resident count
#> 284                                                                     Over Age 64 resident count
#> 285                                                                 People of Color resident count
#> 286                                                                    Count of Hispanic or Latino
#> 287                                 Count of Black or African American (non-Hispanic, single race)
#> 288                                                     Count of Asian (non-Hispanic, single race)
#> 289                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 290                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 291                                                Count of Other race (non-Hispanic, single race)
#> 292                                                      Count of Two or more races (non-Hispanic)
#> 293                                                     Count of White (non-Hispanic, single race)
#> 294                                      Count of Population for whom Poverty Status is Determined
#> 295                                                                  Count of Population Age 25 up
#> 296                                                                            Count of Households
#> 297                                           Universe for percent unemployed (denominator, count)
#> 298                                                          Count of Housing Units Built Pre 1960
#> 299                                   Built housing units count (denominator for percent pre 1960)
#> 300                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 301                                                                          Speak Spanish at Home
#> 302                                                           Persons who Speak English Not at All
#> 303                                                              Speak Other Indo-European at Home
#> 304                                                    Speak Asian-Pacific Island language at Home
#> 305                                                   Speak Spanish (in limited English household)
#> 306                                       Speak Other Indo-European (in limited English household)
#> 307                                      Speak Asian-Pacific Island (in limited English household)
#> 308                                                     Speak Other (in limited English household)
#> 309                                                                        Population Under Age 18
#> 310                                                                         Population Over Age 17
#> 311                                                                                Male Population
#> 312                                                                              Female Population
#> 313                                                                   Owner Occupied Housing Units
#> 314                                                                         Occupied Housing Units
#> 315                                                       Civilian noninstitutionalized population
#> 316                                                                                     DISABILITY
#> 317                                                                       Life Expectancy in Years
#> 318                                                                              Per Capita Income
#> 319                                                            Pct. Households below Poverty Level
#> 320                                  Percent Speaking Spanish (as % of limited English households)
#> 321            Percent Speaking Other Indo-European languages (as % of limited English households)
#> 322           Percent Speaking Asian-Pacific Island languages (as % of limited English households)
#> 323                          Percent Speaking Other languages (as % of limited English households)
#> 324                                                       Percent Break Down by Age below 18 Years
#> 325                                                       Percent Break Down by Age above 17 Years
#> 326                                                                                  Percent Males
#> 327                                                                                Percent Females
#> 328                                                                     Census blockgroups (count)
#> 329                                                                          Census blocks (count)
#> 330                                                                        Buffer distance (miles)
#> 331                                                                 Households below Poverty Level
#> 332                                                                Particulate Matter 2.5 EJ Index
#> 333                                                                                 Ozone EJ Index
#> 334                                                                Air toxics cancer risk EJ Index
#> 335                                                             Air toxics respiratory HI EJ Index
#> 336                                                             Diesel particulate matter EJ Index
#> 337                                                                            Lead paint EJ Index
#> 338                                                              US raw Traffic proximity EJ Index
#> 339                                                            US raw Superfund Proximity EJ Index
#> 340                                                                RMP Facility Proximity EJ Index
#> 341                                                             Hazardous waste proximity EJ Index
#> 342                                                           US raw Wastewater discharge EJ Index
#> 343                                                      US raw Underground storage tanks EJ Index
#> 344                                                                 Toxic Releases to Air EJ Index
#> 345                                                      Particulate Matter 2.5 Supplemental Index
#> 346                                                                       Ozone Supplemental Index
#> 347                                                      Air Toxics Cancer Risk Supplemental Index
#> 348                                                   Air Toxics Respiratory HI Supplemental Index
#> 349                                                   Diesel Particulate Matter Supplemental Index
#> 350                                                                  Lead Paint Supplemental Index
#> 351                                                           Traffic Proximity Supplemental Index
#> 352                                                         Superfund Proximity Supplemental Index
#> 353                                                      RMP Facility Proximity Supplemental Index
#> 354                                                   Hazardous Waste Proximity Supplemental Index
#> 355                                                        Wastewater Discharge Supplemental Index
#> 356                                                   Underground Storage Tanks Supplemental Index
#> 357                                                           EJ Supp: Toxic Releases to Air (raw)
#> 358                                                      State raw Particulate Matter 2.5 EJ Index
#> 359                                                                       State raw Ozone EJ Index
#> 360                                                          State Air toxics cancer risk EJ Index
#> 361                                                       State Air toxics respiratory HI EJ Index
#> 362                                                       State Diesel particulate matter EJ Index
#> 363                                                                      State Lead paint EJ Index
#> 364                                                           State raw Traffic proximity EJ Index
#> 365                                                         State raw Superfund Proximity EJ Index
#> 366                                                      State raw RMP Facility Proximity EJ Index
#> 367                                                       State Hazardous waste proximity EJ Index
#> 368                                                        State raw Wastewater discharge EJ Index
#> 369                                                   State raw Underground storage tanks EJ Index
#> 370                                                           State Toxic Releases to Air EJ Index
#> 371                                            State raw Particulate Matter 2.5 Supplemental Index
#> 372                                                             State raw Ozone Supplemental Index
#> 373                                                State Air toxics cancer risk Supplemental Index
#> 374                                             State Air toxics respiratory HI Supplemental Index
#> 375                                             State Diesel particulate matter Supplemental Index
#> 376                                                            State Lead paint Supplemental Index
#> 377                                                 State raw Traffic proximity Supplemental Index
#> 378                                               State raw Superfund Proximity Supplemental Index
#> 379                                            State raw RMP Facility Proximity Supplemental Index
#> 380                                             State Hazardous waste proximity Supplemental Index
#> 381                                              State raw Wastewater discharge Supplemental Index
#> 382                                         State raw Underground storage tanks Supplemental Index
#> 383                                                 State Toxic Releases to Air Supplemental Index
#> 384                                                                                   distance_min
#> 385                                                                         distance_min_avgperson
#> 386                                                                                  sitecount_max
#> 387                                                                               sitecount_unique
#> 388                                                                                  sitecount_avg
#> 389                                                           Percent population with Disabilities
#> 390                                                                      Owner Occupied households
#> 391                                                                                    pct_lan_eng
#> 392                                                                                pct_lan_spanish
#> 393                                                                                     pct_lan_ie
#> 394                                                                                    pct_lan_api
 table_tall_from_overall(x$results_bysite[1, ])
#> Error in oldnames %in% namesnow: object 'x' not found