URL functions - API WENT DOWN? Get URLs of EJSCREEN reports

Old function for pre-2025 EPA EJSCREEN report (see newer url_ejamapi() now)

Usage

url_ejscreen_report(
  sitepoints = NULL,
  lat = "",
  lon = "",
  radius = "",
  mobile = FALSE,
  areatype = "",
  areaid = "",
  namestr = "",
  shapefile = NULL,
  fips = NULL,
  wkid = 4326,
  unit = 9035,
  f = "report",
  interactiveprompt = FALSE,
  as_html = FALSE,
  linktext = "Report",
  ifna = "https://ejanalysis.com",
  baseurl = NULL,
  ...
)

Arguments

sitepoints

data.frame with colnames lat, lon (or lat, lon parameters can be provided separately)

lat

one or more latitudes (or a table with lat, lon columns, or filepath with that, or omit to interactively select file)

lon

one or more longitudes (or omitted – see lat parameter details)

radius

miles radius

mobile

If TRUE, provides URL for the mobile browser version, not desktop version

areatype

passed as areatype= in API, inferred if not provided but areaid is provided

areaid

fips codes if used, passed as areaid= in API, can be FIPS for blockgroups, tracts, counties. Not zip code! For example, "10001" will report on that county, not that zip code. Note that nearly half of all county fips codes are impossible to distinguish from 5-digit zipcodes because the same numbers are used for both purposes.

For county FIPS 10001, use url_ejscreen_report(areaid = '10001')

For zipcode 10001, you cannot use url_ejscreen_report() for zip codes since the API does not support zip codes. You can still at least try url_ejscreenmap(wherestr = '10001')

namestr

The character string text to show on the report as the name of the place

shapefile

not implemented for old EJSCREEN API, should work for newer EJAM-API

fips

not implemented for old EJSCREEN API, should work for newer EJAM-API

wkid

default is 4326 -WGS84 - World Geodetic System 1984, used in GPS - see (https://epsg.io/4326)

unit

default is 9035 which means miles; for kilometers use 9036

f

can be "report" or "pjson" or "json"

interactiveprompt

passed to sitepoints_from_any()

as_html

Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example

linktext

used as text for hyperlinks, if supplied and as_html=TRUE

Value

URL(s)

Details

If old EJSCREEN api is down, tries to use newer url_ejamapi()

For info on the old EJSCREEN API pre-Feb-2025, see ejscreenRESTbroker() and ejscreenapi1()

and archived EJAPIinstructions.pdf

and archived ejscreenapi1 page

See also