EJAM - Environmental Justice Analysis Multisite tool

EJAM provides a way to run hundreds or thousands of EJScreen-style reports all at once, very quickly. It is also an R package, with functions and data for very fast analysis of areas defined by proximity to points like regulated facilities, areas defined as shapefiles/polygons, or areas specified as Census units (e.g., Counties).

What is EJAM and How to do EJ Analysis with the EJAM R Package

EJAM package documentation provides an overview of EJAM, a complete reference document on all functions, and articles (vignettes) introducing key functions and how to use them.

Data Sources – Demographic and Environmental Indicators

EJScreen documentation provides information about the sources, definitions, and vintage of the demographic and environmental data and indicators used by EJScreen/EJAM.

Each year, EJScreen/EJAM are updated to use newer demographic and environmental data, and some indicators may be added, dropped, or revised during those updates. For details on these updates and changes to datasets and indicators, please see EJScreen and in particular note the information about updates and changes here: EJScreen Change Log. Demographics are from the American Community Survey 5-year summary file data that are updated by the United States Census Bureau annually and then used in EJScreen/EJAM.

See also