Generate HTML Page for Summary Barplot Report in shiny app

Creates header and footer of 1 page report to include a barplot on results for one site (to supplement the EJScreen Community Report)

Usage

build_barplot_report(
  analysis_title,
  totalpop,
  locationstr,
  in_shiny = FALSE,
  filename = NULL
)

Arguments

analysis_title,

title to use in header of report

totalpop,

total population included in location(s) analyzed

locationstr,

description of the location(s) analyzed

in_shiny,

whether the function is being called in or outside of shiny - affects location of header

filename,

path to file to save HTML content to; if null, returns as string (used in Shiny app)

Details

For a related function for use in RStudio, see ejam2report() which relies on build_community_report()