utility - what type of formula is used to aggregate this variable?

utility - what type of formula is used to aggregate this variable?

Usage

calctype(varnames)

Arguments

varnames

vector like names_d

Value

vector same length as varnames, like c("sum of counts", "wtdmean")

Examples

 calctype("pop")
#> [1] "sum of counts"
 calctype(names_d)
#> [1] "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean" "wtdmean"
#> [8] "wtdmean" "wtdmean"
 x = names(testoutput_ejamit_10pts_1miles$results_overall)
 cbind(x, calctype(x))
#>        x                                                                     
#>   [1,] "EJScreen Report"                                NA                   
#>   [2,] "EJScreen Map"                                   NA                   
#>   [3,] "ACS Report"                                     NA                   
#>   [4,] "ECHO Report"                                    NA                   
#>   [5,] "ejam_uniq_id"                                   NA                   
#>   [6,] "valid"                                          NA                   
#>   [7,] "invalid_msg"                                    NA                   
#>   [8,] "pop"                                            "sum of counts"      
#>   [9,] "lon"                                            "input"              
#>  [10,] "lat"                                            "input"              
#>  [11,] "ST"                                             "mode"               
#>  [12,] "statename"                                      "mode"               
#>  [13,] "in_how_many_states"                             NA                   
#>  [14,] "REGION"                                         "mode"               
#>  [15,] "ratio.to.avg.Demog.Index"                       "ratio to avg"       
#>  [16,] "ratio.to.avg.Demog.Index.Supp"                  "ratio to avg"       
#>  [17,] "ratio.to.avg.pctlowinc"                         "ratio to avg"       
#>  [18,] "ratio.to.avg.pctlingiso"                        "ratio to avg"       
#>  [19,] "ratio.to.avg.pctunemployed"                     "ratio to avg"       
#>  [20,] "ratio.to.avg.pctlths"                           "ratio to avg"       
#>  [21,] "ratio.to.avg.pctunder5"                         "ratio to avg"       
#>  [22,] "ratio.to.avg.pctover64"                         "ratio to avg"       
#>  [23,] "ratio.to.avg.pctmin"                            "ratio to avg"       
#>  [24,] "ratio.to.avg.pcthisp"                           "ratio to avg"       
#>  [25,] "ratio.to.avg.pctnhba"                           "ratio to avg"       
#>  [26,] "ratio.to.avg.pctnhaa"                           "ratio to avg"       
#>  [27,] "ratio.to.avg.pctnhaiana"                        "ratio to avg"       
#>  [28,] "ratio.to.avg.pctnhnhpia"                        "ratio to avg"       
#>  [29,] "ratio.to.avg.pctnhotheralone"                   "ratio to avg"       
#>  [30,] "ratio.to.avg.pctnhmulti"                        "ratio to avg"       
#>  [31,] "ratio.to.avg.pctnhwa"                           "ratio to avg"       
#>  [32,] "ratio.to.avg.pm"                                "ratio to avg"       
#>  [33,] "ratio.to.avg.o3"                                "ratio to avg"       
#>  [34,] "ratio.to.avg.no2"                               "ratio to avg"       
#>  [35,] "ratio.to.avg.dpm"                               "ratio to avg"       
#>  [36,] "ratio.to.avg.rsei"                              "ratio to avg"       
#>  [37,] "ratio.to.avg.traffic.score"                     "ratio to avg"       
#>  [38,] "ratio.to.avg.pctpre1960"                        "ratio to avg"       
#>  [39,] "ratio.to.avg.proximity.npl"                     "ratio to avg"       
#>  [40,] "ratio.to.avg.proximity.rmp"                     "ratio to avg"       
#>  [41,] "ratio.to.avg.proximity.tsdf"                    "ratio to avg"       
#>  [42,] "ratio.to.avg.ust"                               "ratio to avg"       
#>  [43,] "ratio.to.avg.proximity.npdes"                   "ratio to avg"       
#>  [44,] "ratio.to.avg.drinking"                          "ratio to avg"       
#>  [45,] "ratio.to.avg.pctdisability"                     "ratio to avg"       
#>  [46,] "ratio.to.avg.lowlifex"                          "ratio to avg"       
#>  [47,] "ratio.to.state.avg.Demog.Index"                 "ratio to avg"       
#>  [48,] "ratio.to.state.avg.Demog.Index.Supp"            "ratio to avg"       
#>  [49,] "ratio.to.state.avg.pctlowinc"                   "ratio to avg"       
#>  [50,] "ratio.to.state.avg.pctlingiso"                  "ratio to avg"       
#>  [51,] "ratio.to.state.avg.pctunemployed"               "ratio to avg"       
#>  [52,] "ratio.to.state.avg.pctlths"                     "ratio to avg"       
#>  [53,] "ratio.to.state.avg.pctunder5"                   "ratio to avg"       
#>  [54,] "ratio.to.state.avg.pctover64"                   "ratio to avg"       
#>  [55,] "ratio.to.state.avg.pctmin"                      "ratio to avg"       
#>  [56,] "ratio.to.state.avg.pcthisp"                     "ratio to avg"       
#>  [57,] "ratio.to.state.avg.pctnhba"                     "ratio to avg"       
#>  [58,] "ratio.to.state.avg.pctnhaa"                     "ratio to avg"       
#>  [59,] "ratio.to.state.avg.pctnhaiana"                  "ratio to avg"       
#>  [60,] "ratio.to.state.avg.pctnhnhpia"                  "ratio to avg"       
#>  [61,] "ratio.to.state.avg.pctnhotheralone"             "ratio to avg"       
#>  [62,] "ratio.to.state.avg.pctnhmulti"                  "ratio to avg"       
#>  [63,] "ratio.to.state.avg.pctnhwa"                     "ratio to avg"       
#>  [64,] "ratio.to.state.avg.pm"                          "ratio to avg"       
#>  [65,] "ratio.to.state.avg.o3"                          "ratio to avg"       
#>  [66,] "ratio.to.state.avg.no2"                         "ratio to avg"       
#>  [67,] "ratio.to.state.avg.dpm"                         "ratio to avg"       
#>  [68,] "ratio.to.state.avg.rsei"                        "ratio to avg"       
#>  [69,] "ratio.to.state.avg.traffic.score"               "ratio to avg"       
#>  [70,] "ratio.to.state.avg.pctpre1960"                  "ratio to avg"       
#>  [71,] "ratio.to.state.avg.proximity.npl"               "ratio to avg"       
#>  [72,] "ratio.to.state.avg.proximity.rmp"               "ratio to avg"       
#>  [73,] "ratio.to.state.avg.proximity.tsdf"              "ratio to avg"       
#>  [74,] "ratio.to.state.avg.ust"                         "ratio to avg"       
#>  [75,] "ratio.to.state.avg.proximity.npdes"             "ratio to avg"       
#>  [76,] "ratio.to.state.avg.drinking"                    "ratio to avg"       
#>  [77,] "ratio.to.state.avg.pctdisability"               "ratio to avg"       
#>  [78,] "ratio.to.state.avg.lowlifex"                    "ratio to avg"       
#>  [79,] "Demog.Index"                                    "wtdmean"            
#>  [80,] "Demog.Index.Supp"                               "wtdmean"            
#>  [81,] "pctlowinc"                                      "wtdmean"            
#>  [82,] "pctlingiso"                                     "wtdmean"            
#>  [83,] "pctunemployed"                                  "wtdmean"            
#>  [84,] "pctlths"                                        "wtdmean"            
#>  [85,] "pctunder5"                                      "wtdmean"            
#>  [86,] "pctover64"                                      "wtdmean"            
#>  [87,] "pctmin"                                         "wtdmean"            
#>  [88,] "pcthisp"                                        "wtdmean"            
#>  [89,] "pctnhba"                                        "wtdmean"            
#>  [90,] "pctnhaa"                                        "wtdmean"            
#>  [91,] "pctnhaiana"                                     "wtdmean"            
#>  [92,] "pctnhnhpia"                                     "wtdmean"            
#>  [93,] "pctnhotheralone"                                "wtdmean"            
#>  [94,] "pctnhmulti"                                     "wtdmean"            
#>  [95,] "pctnhwa"                                        "wtdmean"            
#>  [96,] "Demog.Index.State"                              "wtdmean"            
#>  [97,] "Demog.Index.Supp.State"                         "wtdmean"            
#>  [98,] "pctpoor"                                        "wtdmean"            
#>  [99,] "occupiedunits"                                  "sum of counts"      
#> [100,] "pctmale"                                        "wtdmean"            
#> [101,] "pctfemale"                                      "wtdmean"            
#> [102,] "lifexyears"                                     "wtdmean"            
#> [103,] "percapincome"                                   "wtdmean"            
#> [104,] "pctownedunits"                                  "wtdmean"            
#> [105,] "pctunder18"                                     "wtdmean"            
#> [106,] "pctover17"                                      "wtdmean"            
#> [107,] "pctlan_ie"                                      "wtdmean"            
#> [108,] "pctlan_api"                                     "wtdmean"            
#> [109,] "pctlan_other"                                   "wtdmean"            
#> [110,] "pctlan_spanish"                                 "wtdmean"            
#> [111,] "pctlan_nonenglish"                              "wtdmean"            
#> [112,] "pctspanish_li"                                  "wtdmean"            
#> [113,] "pctie_li"                                       "wtdmean"            
#> [114,] "pctapi_li"                                      "wtdmean"            
#> [115,] "pctother_li"                                    "wtdmean"            
#> [116,] "pctdisability"                                  "wtdmean"            
#> [117,] "lowlifex"                                       "wtdmean"            
#> [118,] "pm"                                             "wtdmean"            
#> [119,] "o3"                                             "wtdmean"            
#> [120,] "no2"                                            "wtdmean"            
#> [121,] "dpm"                                            "wtdmean"            
#> [122,] "rsei"                                           "wtdmean"            
#> [123,] "traffic.score"                                  "wtdmean"            
#> [124,] "pctpre1960"                                     "wtdmean"            
#> [125,] "proximity.npl"                                  "wtdmean"            
#> [126,] "proximity.rmp"                                  "wtdmean"            
#> [127,] "proximity.tsdf"                                 "wtdmean"            
#> [128,] "ust"                                            "wtdmean"            
#> [129,] "proximity.npdes"                                "wtdmean"            
#> [130,] "drinking"                                       "wtdmean"            
#> [131,] "sitecount_max"                                  "max"                
#> [132,] "sitecount_unique"                               "count of unique ids"
#> [133,] "sitecount_avg"                                  "wtdmean"            
#> [134,] "distance_min"                                   "min"                
#> [135,] "distance_min_avgperson"                         "min"                
#> [136,] "pctile.Demog.Index"                             "lookedup"           
#> [137,] "pctile.Demog.Index.Supp"                        "lookedup"           
#> [138,] "pctile.pctlowinc"                               "lookedup"           
#> [139,] "pctile.pctlingiso"                              "lookedup"           
#> [140,] "pctile.pctunemployed"                           "lookedup"           
#> [141,] "pctile.pctlths"                                 "lookedup"           
#> [142,] "pctile.pctunder5"                               "lookedup"           
#> [143,] "pctile.pctover64"                               "lookedup"           
#> [144,] "pctile.pctmin"                                  "lookedup"           
#> [145,] "pctile.pcthisp"                                 "lookedup"           
#> [146,] "pctile.pctnhba"                                 "lookedup"           
#> [147,] "pctile.pctnhaa"                                 "lookedup"           
#> [148,] "pctile.pctnhaiana"                              "lookedup"           
#> [149,] "pctile.pctnhnhpia"                              "lookedup"           
#> [150,] "pctile.pctnhotheralone"                         "lookedup"           
#> [151,] "pctile.pctnhmulti"                              "lookedup"           
#> [152,] "pctile.pctnhwa"                                 "lookedup"           
#> [153,] "pctile.pctdisability"                           "lookedup"           
#> [154,] "pctile.lowlifex"                                "lookedup"           
#> [155,] "pctile.pm"                                      "lookedup"           
#> [156,] "pctile.o3"                                      "lookedup"           
#> [157,] "pctile.no2"                                     "lookedup"           
#> [158,] "pctile.dpm"                                     "lookedup"           
#> [159,] "pctile.rsei"                                    "lookedup"           
#> [160,] "pctile.traffic.score"                           "lookedup"           
#> [161,] "pctile.pctpre1960"                              "lookedup"           
#> [162,] "pctile.proximity.npl"                           "lookedup"           
#> [163,] "pctile.proximity.rmp"                           "lookedup"           
#> [164,] "pctile.proximity.tsdf"                          "lookedup"           
#> [165,] "pctile.ust"                                     "lookedup"           
#> [166,] "pctile.proximity.npdes"                         "lookedup"           
#> [167,] "pctile.drinking"                                "lookedup"           
#> [168,] "state.pctile.Demog.Index"                       "lookedup"           
#> [169,] "state.pctile.Demog.Index.Supp"                  "lookedup"           
#> [170,] "state.pctile.pctlowinc"                         "lookedup"           
#> [171,] "state.pctile.pctlingiso"                        "lookedup"           
#> [172,] "state.pctile.pctunemployed"                     "lookedup"           
#> [173,] "state.pctile.pctlths"                           "lookedup"           
#> [174,] "state.pctile.pctunder5"                         "lookedup"           
#> [175,] "state.pctile.pctover64"                         "lookedup"           
#> [176,] "state.pctile.pctmin"                            "lookedup"           
#> [177,] "state.pctile.pcthisp"                           "lookedup"           
#> [178,] "state.pctile.pctnhba"                           "lookedup"           
#> [179,] "state.pctile.pctnhaa"                           "lookedup"           
#> [180,] "state.pctile.pctnhaiana"                        "lookedup"           
#> [181,] "state.pctile.pctnhnhpia"                        "lookedup"           
#> [182,] "state.pctile.pctnhotheralone"                   "lookedup"           
#> [183,] "state.pctile.pctnhmulti"                        "lookedup"           
#> [184,] "state.pctile.pctnhwa"                           "lookedup"           
#> [185,] "state.pctile.pctdisability"                     "lookedup"           
#> [186,] "state.pctile.lowlifex"                          "lookedup"           
#> [187,] "state.pctile.pm"                                "lookedup"           
#> [188,] "state.pctile.o3"                                "lookedup"           
#> [189,] "state.pctile.no2"                               "lookedup"           
#> [190,] "state.pctile.dpm"                               "lookedup"           
#> [191,] "state.pctile.rsei"                              "lookedup"           
#> [192,] "state.pctile.traffic.score"                     "lookedup"           
#> [193,] "state.pctile.pctpre1960"                        "lookedup"           
#> [194,] "state.pctile.proximity.npl"                     "lookedup"           
#> [195,] "state.pctile.proximity.rmp"                     "lookedup"           
#> [196,] "state.pctile.proximity.tsdf"                    "lookedup"           
#> [197,] "state.pctile.ust"                               "lookedup"           
#> [198,] "state.pctile.proximity.npdes"                   "lookedup"           
#> [199,] "state.pctile.drinking"                          "lookedup"           
#> [200,] "avg.Demog.Index"                                "constant"           
#> [201,] "avg.Demog.Index.Supp"                           "constant"           
#> [202,] "avg.pctlowinc"                                  "constant"           
#> [203,] "avg.pctlingiso"                                 "constant"           
#> [204,] "avg.pctunemployed"                              "constant"           
#> [205,] "avg.pctlths"                                    "constant"           
#> [206,] "avg.pctunder5"                                  "constant"           
#> [207,] "avg.pctover64"                                  "constant"           
#> [208,] "avg.pctmin"                                     "constant"           
#> [209,] "avg.pcthisp"                                    "constant"           
#> [210,] "avg.pctnhba"                                    "constant"           
#> [211,] "avg.pctnhaa"                                    "constant"           
#> [212,] "avg.pctnhaiana"                                 "constant"           
#> [213,] "avg.pctnhnhpia"                                 "constant"           
#> [214,] "avg.pctnhotheralone"                            "constant"           
#> [215,] "avg.pctnhmulti"                                 "constant"           
#> [216,] "avg.pctnhwa"                                    "constant"           
#> [217,] "avg.pctdisability"                              "constant"           
#> [218,] "avg.lowlifex"                                   "constant"           
#> [219,] "avg.pm"                                         "constant"           
#> [220,] "avg.o3"                                         "constant"           
#> [221,] "avg.no2"                                        "constant"           
#> [222,] "avg.dpm"                                        "constant"           
#> [223,] "avg.rsei"                                       "constant"           
#> [224,] "avg.traffic.score"                              "constant"           
#> [225,] "avg.pctpre1960"                                 "constant"           
#> [226,] "avg.proximity.npl"                              "constant"           
#> [227,] "avg.proximity.rmp"                              "constant"           
#> [228,] "avg.proximity.tsdf"                             "constant"           
#> [229,] "avg.ust"                                        "constant"           
#> [230,] "avg.proximity.npdes"                            "constant"           
#> [231,] "avg.drinking"                                   "constant"           
#> [232,] "state.avg.Demog.Index"                          "constant"           
#> [233,] "state.avg.Demog.Index.Supp"                     "constant"           
#> [234,] "state.avg.pctlowinc"                            "constant"           
#> [235,] "state.avg.pctlingiso"                           "constant"           
#> [236,] "state.avg.pctunemployed"                        "constant"           
#> [237,] "state.avg.pctlths"                              "constant"           
#> [238,] "state.avg.pctunder5"                            "constant"           
#> [239,] "state.avg.pctover64"                            "constant"           
#> [240,] "state.avg.pctmin"                               "constant"           
#> [241,] "state.avg.pcthisp"                              "constant"           
#> [242,] "state.avg.pctnhba"                              "constant"           
#> [243,] "state.avg.pctnhaa"                              "constant"           
#> [244,] "state.avg.pctnhaiana"                           "constant"           
#> [245,] "state.avg.pctnhnhpia"                           "constant"           
#> [246,] "state.avg.pctnhotheralone"                      "constant"           
#> [247,] "state.avg.pctnhmulti"                           "constant"           
#> [248,] "state.avg.pctnhwa"                              "constant"           
#> [249,] "state.avg.pctdisability"                        "constant"           
#> [250,] "state.avg.lowlifex"                             "constant"           
#> [251,] "state.avg.pm"                                   "constant"           
#> [252,] "state.avg.o3"                                   "constant"           
#> [253,] "state.avg.no2"                                  "constant"           
#> [254,] "state.avg.dpm"                                  "constant"           
#> [255,] "state.avg.rsei"                                 "constant"           
#> [256,] "state.avg.traffic.score"                        "constant"           
#> [257,] "state.avg.pctpre1960"                           "constant"           
#> [258,] "state.avg.proximity.npl"                        "constant"           
#> [259,] "state.avg.proximity.rmp"                        "constant"           
#> [260,] "state.avg.proximity.tsdf"                       "constant"           
#> [261,] "state.avg.ust"                                  "constant"           
#> [262,] "state.avg.proximity.npdes"                      "constant"           
#> [263,] "state.avg.drinking"                             "constant"           
#> [264,] "lowinc"                                         "sum of counts"      
#> [265,] "lingiso"                                        "sum of counts"      
#> [266,] "unemployed"                                     "sum of counts"      
#> [267,] "lths"                                           "sum of counts"      
#> [268,] "under5"                                         "sum of counts"      
#> [269,] "over64"                                         "sum of counts"      
#> [270,] "mins"                                           "sum of counts"      
#> [271,] "hisp"                                           "sum of counts"      
#> [272,] "nhba"                                           "sum of counts"      
#> [273,] "nhaa"                                           "sum of counts"      
#> [274,] "nhaiana"                                        "sum of counts"      
#> [275,] "nhnhpia"                                        "sum of counts"      
#> [276,] "nhotheralone"                                   "sum of counts"      
#> [277,] "nhmulti"                                        "sum of counts"      
#> [278,] "nhwa"                                           "sum of counts"      
#> [279,] "age25up"                                        "sum of counts"      
#> [280,] "hhlds"                                          "sum of counts"      
#> [281,] "unemployedbase"                                 "sum of counts"      
#> [282,] "pre1960"                                        "sum of counts"      
#> [283,] "builtunits"                                     "sum of counts"      
#> [284,] "povknownratio"                                  "sum of counts"      
#> [285,] "ownedunits"                                     "sum of counts"      
#> [286,] "poor"                                           "sum of counts"      
#> [287,] "female"                                         "sum of counts"      
#> [288,] "male"                                           "sum of counts"      
#> [289,] "over17"                                         "sum of counts"      
#> [290,] "under18"                                        "sum of counts"      
#> [291,] "lan_nonenglish"                                 "sum of counts"      
#> [292,] "lan_ie"                                         "sum of counts"      
#> [293,] "lan_api"                                        "sum of counts"      
#> [294,] "lan_other"                                      "sum of counts"      
#> [295,] "lan_universe"                                   "sum of counts"      
#> [296,] "lan_eng_na"                                     "sum of counts"      
#> [297,] "lan_spanish"                                    "sum of counts"      
#> [298,] "spanish_li"                                     "sum of counts"      
#> [299,] "ie_li"                                          "sum of counts"      
#> [300,] "api_li"                                         "sum of counts"      
#> [301,] "other_li"                                       "sum of counts"      
#> [302,] "pctile.EJ.DISPARITY.pm.eo"                      "lookedup"           
#> [303,] "pctile.EJ.DISPARITY.o3.eo"                      "lookedup"           
#> [304,] "pctile.EJ.DISPARITY.no2.eo"                     "lookedup"           
#> [305,] "pctile.EJ.DISPARITY.dpm.eo"                     "lookedup"           
#> [306,] "pctile.EJ.DISPARITY.rsei.eo"                    "lookedup"           
#> [307,] "pctile.EJ.DISPARITY.traffic.score.eo"           "lookedup"           
#> [308,] "pctile.EJ.DISPARITY.pctpre1960.eo"              "lookedup"           
#> [309,] "pctile.EJ.DISPARITY.proximity.npl.eo"           "lookedup"           
#> [310,] "pctile.EJ.DISPARITY.proximity.rmp.eo"           "lookedup"           
#> [311,] "pctile.EJ.DISPARITY.proximity.tsdf.eo"          "lookedup"           
#> [312,] "pctile.EJ.DISPARITY.ust.eo"                     "lookedup"           
#> [313,] "pctile.EJ.DISPARITY.proximity.npdes.eo"         "lookedup"           
#> [314,] "pctile.EJ.DISPARITY.drinking.eo"                "lookedup"           
#> [315,] "pctile.EJ.DISPARITY.pm.supp"                    "lookedup"           
#> [316,] "pctile.EJ.DISPARITY.o3.supp"                    "lookedup"           
#> [317,] "pctile.EJ.DISPARITY.no2.supp"                   "lookedup"           
#> [318,] "pctile.EJ.DISPARITY.dpm.supp"                   "lookedup"           
#> [319,] "pctile.EJ.DISPARITY.rsei.supp"                  "lookedup"           
#> [320,] "pctile.EJ.DISPARITY.traffic.score.supp"         "lookedup"           
#> [321,] "pctile.EJ.DISPARITY.pctpre1960.supp"            "lookedup"           
#> [322,] "pctile.EJ.DISPARITY.proximity.npl.supp"         "lookedup"           
#> [323,] "pctile.EJ.DISPARITY.proximity.rmp.supp"         "lookedup"           
#> [324,] "pctile.EJ.DISPARITY.proximity.tsdf.supp"        "lookedup"           
#> [325,] "pctile.EJ.DISPARITY.ust.supp"                   "lookedup"           
#> [326,] "pctile.EJ.DISPARITY.proximity.npdes.supp"       "lookedup"           
#> [327,] "pctile.EJ.DISPARITY.drinking.supp"              "lookedup"           
#> [328,] "state.pctile.EJ.DISPARITY.pm.eo"                "lookedup"           
#> [329,] "state.pctile.EJ.DISPARITY.o3.eo"                "lookedup"           
#> [330,] "state.pctile.EJ.DISPARITY.no2.eo"               "lookedup"           
#> [331,] "state.pctile.EJ.DISPARITY.dpm.eo"               "lookedup"           
#> [332,] "state.pctile.EJ.DISPARITY.rsei.eo"              "lookedup"           
#> [333,] "state.pctile.EJ.DISPARITY.traffic.score.eo"     "lookedup"           
#> [334,] "state.pctile.EJ.DISPARITY.pctpre1960.eo"        "lookedup"           
#> [335,] "state.pctile.EJ.DISPARITY.proximity.npl.eo"     "lookedup"           
#> [336,] "state.pctile.EJ.DISPARITY.proximity.rmp.eo"     "lookedup"           
#> [337,] "state.pctile.EJ.DISPARITY.proximity.tsdf.eo"    "lookedup"           
#> [338,] "state.pctile.EJ.DISPARITY.ust.eo"               "lookedup"           
#> [339,] "state.pctile.EJ.DISPARITY.proximity.npdes.eo"   "lookedup"           
#> [340,] "state.pctile.EJ.DISPARITY.drinking.eo"          "lookedup"           
#> [341,] "state.pctile.EJ.DISPARITY.pm.supp"              "lookedup"           
#> [342,] "state.pctile.EJ.DISPARITY.o3.supp"              "lookedup"           
#> [343,] "state.pctile.EJ.DISPARITY.no2.supp"             "lookedup"           
#> [344,] "state.pctile.EJ.DISPARITY.dpm.supp"             "lookedup"           
#> [345,] "state.pctile.EJ.DISPARITY.rsei.supp"            "lookedup"           
#> [346,] "state.pctile.EJ.DISPARITY.traffic.score.supp"   "lookedup"           
#> [347,] "state.pctile.EJ.DISPARITY.pctpre1960.supp"      "lookedup"           
#> [348,] "state.pctile.EJ.DISPARITY.proximity.npl.supp"   "lookedup"           
#> [349,] "state.pctile.EJ.DISPARITY.proximity.rmp.supp"   "lookedup"           
#> [350,] "state.pctile.EJ.DISPARITY.proximity.tsdf.supp"  "lookedup"           
#> [351,] "state.pctile.EJ.DISPARITY.ust.supp"             "lookedup"           
#> [352,] "state.pctile.EJ.DISPARITY.proximity.npdes.supp" "lookedup"           
#> [353,] "state.pctile.EJ.DISPARITY.drinking.supp"        "lookedup"           
#> [354,] "bgcount_near_site"                              "count of unique ids"
#> [355,] "blockcount_near_site"                           "count of unique ids"
#> [356,] "radius.miles"                                   "input"              
#> [357,] "disab_universe"                                 "sum of counts"      
#> [358,] "disability"                                     "sum of counts"      
#> [359,] "pctba"                                          "wtdmean"            
#> [360,] "pctaa"                                          "wtdmean"            
#> [361,] "pctaiana"                                       "wtdmean"            
#> [362,] "pctnhpia"                                       "wtdmean"            
#> [363,] "pctotheralone"                                  "wtdmean"            
#> [364,] "pctmulti"                                       "wtdmean"            
#> [365,] "pctwa"                                          "wtdmean"            
#> [366,] "EJ.DISPARITY.pm.eo"                             "wtdmean"            
#> [367,] "EJ.DISPARITY.o3.eo"                             "wtdmean"            
#> [368,] "EJ.DISPARITY.no2.eo"                            "wtdmean"            
#> [369,] "EJ.DISPARITY.dpm.eo"                            "wtdmean"            
#> [370,] "EJ.DISPARITY.rsei.eo"                           "wtdmean"            
#> [371,] "EJ.DISPARITY.traffic.score.eo"                  "wtdmean"            
#> [372,] "EJ.DISPARITY.pctpre1960.eo"                     "wtdmean"            
#> [373,] "EJ.DISPARITY.proximity.npl.eo"                  "wtdmean"            
#> [374,] "EJ.DISPARITY.proximity.rmp.eo"                  "wtdmean"            
#> [375,] "EJ.DISPARITY.proximity.tsdf.eo"                 "wtdmean"            
#> [376,] "EJ.DISPARITY.ust.eo"                            "wtdmean"            
#> [377,] "EJ.DISPARITY.proximity.npdes.eo"                "wtdmean"            
#> [378,] "EJ.DISPARITY.drinking.eo"                       "wtdmean"            
#> [379,] "state.EJ.DISPARITY.pm.eo"                       "wtdmean"            
#> [380,] "state.EJ.DISPARITY.o3.eo"                       "wtdmean"            
#> [381,] "state.EJ.DISPARITY.no2.eo"                      "wtdmean"            
#> [382,] "state.EJ.DISPARITY.dpm.eo"                      "wtdmean"            
#> [383,] "state.EJ.DISPARITY.rsei.eo"                     "wtdmean"            
#> [384,] "state.EJ.DISPARITY.traffic.score.eo"            "wtdmean"            
#> [385,] "state.EJ.DISPARITY.pctpre1960.eo"               "wtdmean"            
#> [386,] "state.EJ.DISPARITY.proximity.npl.eo"            "wtdmean"            
#> [387,] "state.EJ.DISPARITY.proximity.rmp.eo"            "wtdmean"            
#> [388,] "state.EJ.DISPARITY.proximity.tsdf.eo"           "wtdmean"            
#> [389,] "state.EJ.DISPARITY.ust.eo"                      "wtdmean"            
#> [390,] "state.EJ.DISPARITY.proximity.npdes.eo"          "wtdmean"            
#> [391,] "state.EJ.DISPARITY.drinking.eo"                 "wtdmean"            
#> [392,] "EJ.DISPARITY.pm.supp"                           "wtdmean"            
#> [393,] "EJ.DISPARITY.o3.supp"                           "wtdmean"            
#> [394,] "EJ.DISPARITY.no2.supp"                          "wtdmean"            
#> [395,] "EJ.DISPARITY.dpm.supp"                          "wtdmean"            
#> [396,] "EJ.DISPARITY.rsei.supp"                         "wtdmean"            
#> [397,] "EJ.DISPARITY.traffic.score.supp"                "wtdmean"            
#> [398,] "EJ.DISPARITY.pctpre1960.supp"                   "wtdmean"            
#> [399,] "EJ.DISPARITY.proximity.npl.supp"                "wtdmean"            
#> [400,] "EJ.DISPARITY.proximity.rmp.supp"                "wtdmean"            
#> [401,] "EJ.DISPARITY.proximity.tsdf.supp"               "wtdmean"            
#> [402,] "EJ.DISPARITY.ust.supp"                          "wtdmean"            
#> [403,] "EJ.DISPARITY.proximity.npdes.supp"              "wtdmean"            
#> [404,] "EJ.DISPARITY.drinking.supp"                     "wtdmean"            
#> [405,] "state.EJ.DISPARITY.pm.supp"                     "wtdmean"            
#> [406,] "state.EJ.DISPARITY.o3.supp"                     "wtdmean"            
#> [407,] "state.EJ.DISPARITY.no2.supp"                    "wtdmean"            
#> [408,] "state.EJ.DISPARITY.dpm.supp"                    "wtdmean"            
#> [409,] "state.EJ.DISPARITY.rsei.supp"                   "wtdmean"            
#> [410,] "state.EJ.DISPARITY.traffic.score.supp"          "wtdmean"            
#> [411,] "state.EJ.DISPARITY.pctpre1960.supp"             "wtdmean"            
#> [412,] "state.EJ.DISPARITY.proximity.npl.supp"          "wtdmean"            
#> [413,] "state.EJ.DISPARITY.proximity.rmp.supp"          "wtdmean"            
#> [414,] "state.EJ.DISPARITY.proximity.tsdf.supp"         "wtdmean"            
#> [415,] "state.EJ.DISPARITY.ust.supp"                    "wtdmean"            
#> [416,] "state.EJ.DISPARITY.proximity.npdes.supp"        "wtdmean"            
#> [417,] "state.EJ.DISPARITY.drinking.supp"               "wtdmean"