Answers questions like What % of sites have at least 1 demog. indicator >2x state avg?

This function provides tables of summary stats but also text that explains those findings in plain English. It relies on colcounter_summary_all()

count_sites_with_n_high_scores(
  scores,
  thresholds = c(1.01, 1.5, 2, 3, 5, 10),
  xwide = c("statewide", "nationwide")[1]
)

Arguments

scores

scores in a table with one row per place and one column per indicator

thresholds

thresholds vector of numbers as benchmarks. Assuming the indicators in the scores table are ratios to the average, then the thresholds could be for example, 1.50, 2, etc. which would represent ratios that are 1.5x or 2x etc.

xwide

must be "statewide" or "nationwide" -- used only in the text output that describes the findings.

Value

Returns a list with two named elements, "stats" and "text" where stats is a 3-dimensional array of numbers. See dimnames(output$stats).

Details

Helps provide stat summaries such as:

(x%) of these (sites) have

at least (N) of these (YTYPE )indicators

at least (R) times the (State/National average)

Examples

# x <- ejamit(testpoints_100, radius = 1)
x <- testoutput_ejamit_1000pts_1miles 
out <- x$results_bysite
out <- setDF(copy(out))
#> Error in setDF(copy(out)): could not find function "setDF"
ratio_benchmarks <- c(1.01, 1.50, 2, 3, 5, 10)
ratiodata <- out[, names_d_ratio_to_state_avg]
#> Error in `[.data.table`(out, , names_d_ratio_to_state_avg): j (the 2nd argument inside [...]) is a single symbol but column name 'names_d_ratio_to_state_avg' is not found. Perhaps you intended DT[, ..names_d_ratio_to_state_avg]. This difference to data.frame is deliberate and explained in FAQ 1.1.

findings = count_sites_with_n_high_scores(out)
#> Error in `[.data.table`(ratiodata, , ccc): j (the 2nd argument inside [...]) is a single symbol but column name 'ccc' is not found. Perhaps you intended DT[, ..ccc]. This difference to data.frame is deliberate and explained in FAQ 1.1.
dimnames(findings)
#> Error in eval(expr, envir, enclos): object 'findings' not found
findings$text[2]
#> Error in eval(expr, envir, enclos): object 'findings' not found
findings$stats[ , , 1]
#> Error in eval(expr, envir, enclos): object 'findings' not found
findings$stats[ , 1, ]
#> Error in eval(expr, envir, enclos): object 'findings' not found
findings$stats[ 1, , ]
#> Error in eval(expr, envir, enclos): object 'findings' not found