Analyze US Counties as if they were sites, to get EJ indicators summary for each county

Analyze US Counties as if they were sites, to get EJ indicators summary for each county

counties_as_sites(fips)

Arguments

fips

County FIPS vector (ideally as character not numeric values), like fips_counties_from_state_abbrev("DE")

Value

provides table similar to the output of getblocksnearby(), data.table with one row per blockgroup in these counties, or all pairs of county fips - bgid, and ejam_uniq_id (1 through N) assigned to each county but missing blockid and distance so not ready for doaggregate().

Details

This function provides one row per blockgroup. getblocksnearby_from_fips() provides one row per block. See more below under "Value"

See also

Examples

 counties_as_sites(c('01001','72153'))
#>     ejam_uniq_id countyfips   bgid
#>  1:            1      01001      1
#>  2:            1      01001      2
#>  3:            1      01001      3
#>  4:            1      01001      4
#>  5:            1      01001      5
#>  6:            1      01001      6
#>  7:            1      01001      7
#>  8:            1      01001      8
#>  9:            1      01001      9
#> 10:            1      01001     10
#> 11:            1      01001     11
#> 12:            1      01001     12
#> 13:            1      01001     13
#> 14:            1      01001     14
#> 15:            1      01001     15
#> 16:            1      01001     16
#> 17:            1      01001     17
#> 18:            1      01001     18
#> 19:            1      01001     19
#> 20:            1      01001     20
#> 21:            1      01001     21
#> 22:            1      01001     22
#> 23:            1      01001     23
#> 24:            1      01001     24
#> 25:            1      01001     25
#> 26:            1      01001     26
#> 27:            1      01001     27
#> 28:            1      01001     28
#> 29:            1      01001     29
#> 30:            1      01001     30
#> 31:            1      01001     31
#> 32:            1      01001     32
#> 33:            1      01001     33
#> 34:            1      01001     34
#> 35:            1      01001     35
#> 36:            1      01001     36
#> 37:            1      01001     37
#> 38:            1      01001     38
#> 39:            1      01001     39
#> 40:            1      01001     40
#> 41:            1      01001     41
#> 42:            1      01001     42
#> 43:            1      01001     43
#> 44:            1      01001     44
#> 45:            1      01001     45
#> 46:            2      72153 242311
#> 47:            2      72153 242312
#> 48:            2      72153 242313
#> 49:            2      72153 242314
#> 50:            2      72153 242315
#> 51:            2      72153 242316
#> 52:            2      72153 242317
#> 53:            2      72153 242318
#> 54:            2      72153 242319
#> 55:            2      72153 242320
#> 56:            2      72153 242321
#> 57:            2      72153 242322
#> 58:            2      72153 242323
#> 59:            2      72153 242324
#> 60:            2      72153 242325
#> 61:            2      72153 242326
#> 62:            2      72153 242327
#> 63:            2      72153 242328
#> 64:            2      72153 242329
#> 65:            2      72153 242330
#> 66:            2      72153 242331
#> 67:            2      72153 242332
#> 68:            2      72153 242333
#> 69:            2      72153 242334
#> 70:            2      72153 242335
#>     ejam_uniq_id countyfips   bgid
 # Largest US Counties by ACS Population Totals:
 blockgroupstats[ , .(ST = ST[1], countypop = sum(pop)),
  by=.(FIPS = substr(bgfips,1,5))][order(-countypop),][1:20, .(
  CountyPopulation = prettyNum(countypop, big.mark = ","), FIPS, ST)]
#>     CountyPopulation  FIPS ST
#>  1:       10,019,635 06037 CA
#>  2:        5,265,398 17031 IL
#>  3:        4,697,957 48201 TX
#>  4:        4,367,186 04013 AZ
#>  5:        3,296,317 06073 CA
#>  6:        3,182,923 06059 CA
#>  7:        2,712,360 36047 NY
#>  8:        2,690,113 12086 FL
#>  9:        2,604,722 48113 TX
#> 10:        2,409,331 06065 CA
#> 11:        2,393,104 36081 NY
#> 12:        2,240,876 53033 WA
#> 13:        2,231,147 32003 NV
#> 14:        2,171,071 06071 CA
#> 15:        2,091,953 48439 TX
#> 16:        1,990,522 48029 TX
#> 17:        1,934,551 12011 FL
#> 18:        1,932,022 06085 CA
#> 19:        1,789,781 26163 MI
#> 20:        1,673,133 06001 CA