Save EJAM results in a spreadsheet

ejam2excel() takes the output of something like ejamit() and creates a spreadsheet with an overall summary tab, a site by site table tab, as well as other tabs such as map, plot, notes, etc. It is an alias for table_xls_from_ejam()

ejam2excel(
  ejamitout,
  fname = NULL,
  save_now = TRUE,
  overwrite = TRUE,
  launchexcel = FALSE,
  interactive_console = TRUE,
  ok2plot = TRUE,
  in.testing = FALSE,
  in.analysis_title = "EJAM analysis",
  react.v1_summary_plot = NULL,
  radius_or_buffer_in_miles = NULL,
  buffer_desc = "Selected Locations",
  radius_or_buffer_description =
    "Miles radius of circular buffer (or distance used if buffering around polygons)",
  hyperlink_colnames = c("EJScreen Report", "EJScreen Map", "ECHO report"),
  ...
)

Arguments

ejamitout

output of ejamit()

fname

optional name or full path and name of file to save locally, like "out.xlsx"

save_now

optional logical, whether to save as a .xlsx file locally or just return workbook object that can later be written to .xlsx file using openxlsx::saveWorkbook()

overwrite

optional logical, passed to openxlsx::saveWorkbook()

launchexcel

optional logical, passed to table_xls_format(), whether to launch browser to see spreadsheet immediately

interactive_console

optional - should set to FALSE when used in code or server. If TRUE, prompts RStudio user interactively asking where to save the downloaded file

ok2plot

optional logical, passed to table_xls_format(), whether safe to try and plot or set FALSE if debugging plot problems

in.testing

optional logical

in.analysis_title

optional title as character string

react.v1_summary_plot

optional - a plot object

radius_or_buffer_in_miles

optional radius in miles

radius_or_buffer_description

optional text phrase describing places analyzed

hyperlink_colnames

optional names of columns with URLs

...

optional additional parameters passed to table_xls_format(), such as heatmap_colnames, heatmap_cuts, heatmap_colors, etc.