Show EJAM results as a map of points

Takes the output of ejamit() and uses mapfastej() to create a map of the points.

Usage

ejam2map(ejamitout, radius = NULL, column_names = "ej", browse = FALSE)

Arguments

ejamitout

output of ejamit()

radius

radius in miles

column_names

can be "ej", passed to mapfast()

browse

logical optional whether to open the web browser to view the map

Value

like what mapfastej() returns

Details

Gets radius by checking ejamitout$results_overall$radius.miles You can use browse=TRUE to save it as a shareable .html file and see it in your web browser.

Examples

pts = testpoints_100
mapfast(pts)

# out = ejamit(pts, radius = 1)
out = testoutput_ejamit_100pts_1miles

# See in RStudio viewer pane
ejam2map(out)
mapfastej(out$results_bysite[c(12,31),])
if (FALSE) { # \dontrun{

# See in local browser instead
ejam2map(out, browse = T)

# Open folder where interactive map
#  .html file is saved, so you can share it:
x = ejam2map(out)
fname = map2browser(x)
# browseURL(dirname(fname)) # to open the temp folder
# file.copy(fname, "./map.html") # to copy map file to working directory

out <- testoutput_ejscreenapi_plus_5
mapfastej(out)
} # }