View HTML Report on EJAM Results (Overall or at 1 Site)

Get URL for and view in browser a 2-page summary report similar to the EJScreen Community Report

Usage

ejam2report(
  ejamout = testoutput_ejamit_10pts_1miles,
  sitenumber = NULL,
  analysis_title = "EJAM Report",
  submitted_upload_method = c("latlon", "SHP", "FIPS")[1],
  data_up_shp = NA,
  return_html = FALSE,
  launch_browser = TRUE
)

Arguments

ejamout

output as from ejamit(), list with a data.table called results_bysite if sitenumber parameter is used, or a data.table called results_overall otherwise

sitenumber

If a number is provided, the report is about ejamout$results_bysite[sitenumber, ] and if no number is provided (param is NULL) then the report is about ejamout$results_overall

analysis_title

optional title of analysis

submitted_upload_method

"latlon" is the default and other options are not implemented yet.

data_up_shp

not implemented yet, for upload method SHP

return_html

set TRUE to have function return HTML object instead of URL of local file

launch_browser

set TRUE to have it launch browser and show report.

Value

URL of temp html file or object depending on return_html, and has side effect of launching browser to view it depending on return_html

Details

This relies on build_community_report() as used in web app for viewing report on 1 site from a list of sites.

Examples

#out <- ejamit(testpoints_10, radius = 3, include_ejindexes = T)
out <- testoutput_ejamit_10pts_1miles

ejam2report(out)
#> [1] "C:\\Users\\mcorrale\\AppData\\Local\\Temp\\RtmpyYmWDX/comm_report.html"
table_gt_from_ejamit_overall(out$results_overall)
table_tall_from_overall(out$results_overall)
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#>                  value                                                                                      indicator
#> 1                 <NA>                                                                                EJScreen Report
#> 2                 <NA>                                                                                   EJScreen Map
#> 3                 <NA>                                                                                    ECHO report
#> 4                 <NA>                                                                                   ejam_uniq_id
#> 5                60740                                                                               Total Population
#> 6                 <NA>                                                                                      Longitude
#> 7                 <NA>                                                                                       Latitude
#> 8                 <NA>                                                                             State Abbreviation
#> 9                 <NA>                                                                                     State Name
#> 10                 1.2                                                                    Ratio to US avg Demog Index
#> 11                 1.2                                                              Ratio to US avg Suppl Demog Index
#> 12                 1.3                                                                     Ratio to US avg % Low-inc.
#> 13                 1.4                                                              Ratio to US avg % Limited English
#> 14                   1                                                                   Ratio to US avg % Unemployed
#> 15                 1.3                                                                Ratio to US avg % < High School
#> 16                 1.1                                                            Ratio to US avg Low life expectancy
#> 17                 1.2                                                                      Ratio to US avg % < age 5
#> 18                 0.6                                                                     Ratio to US avg % > age 64
#> 19                   1                                                              Ratio to US avg % People of Color
#> 20                 1.4                                                                     Ratio to US avg % Hispanic
#> 21                 0.7                                                                  Ratio to US avg % Black (NHA)
#> 22                 0.6                                                                  Ratio to US avg % Asian (NHA)
#> 23                 1.4                                          Ratio to US avg % American Indian Alaska Native (NHA)
#> 24                   0                                       Ratio to US avg % Native Hawaiian Pacific Islander (NHA)
#> 25                 0.3                                                      Ratio to US avg % Other single race (NHA)
#> 26                 0.9                                                        Ratio to US avg % Two or more race (NH)
#> 27                   1                                          Ratio to US avg % White nonHispanic single race (NHA)
#> 28                   1                                                                          Ratio to US avg PM2.5
#> 29                   1                                                                          Ratio to US avg Ozone
#> 30                 1.1                                                                    Ratio to US avg Cancer risk
#> 31                 0.9                                                                    Ratio to US avg Respiratory
#> 32                 1.4                                                                      Ratio to US avg Diesel PM
#> 33                 1.4                                                               Ratio to US avg % built pre-1960
#> 34                 0.8                                                                        Ratio to US avg Traffic
#> 35                 2.9                                                                            Ratio to US avg NPL
#> 36                 1.4                                                                            Ratio to US avg RMP
#> 37                 0.7                                                                           Ratio to US avg TSDF
#> 38                   0                                                                          Ratio to US avg NPDES
#> 39                 1.4                                                                            Ratio to US avg UST
#> 40                 1.4                                                          Ratio to US avg Toxic Releases to Air
#> 41                 1.3                                                                  Ratio to State avg Demog.Ind.
#> 42                 1.3                                                           Ratio to State avg Suppl Demog Index
#> 43                 1.3                                                                  Ratio to State avg % Low-inc.
#> 44                 1.9                                                           Ratio to State avg % Limited English
#> 45                 1.1                                                                Ratio to State avg % Unemployed
#> 46                 1.5                                                             Ratio to State avg % < High School
#> 47                 1.1                                                         Ratio to State avg Low life expectancy
#> 48                 1.1                                                                   Ratio to State avg % < age 5
#> 49                 0.6                                                                  Ratio to State avg % > age 64
#> 50                 1.3                                                           Ratio to State avg % People of Color
#> 51                 1.8                                                                  Ratio to State avg % Hispanic
#> 52                 0.8                                       Ratio to State avg % American Indian Alaska Native (NHA)
#> 53                 1.3                                                               Ratio to State avg % Black (NHA)
#> 54                 0.8                                                               Ratio to State avg % Asian (NHA)
#> 55                   0                                                   Ratio to State avg % Other single race (NHA)
#> 56                 0.3                                                     Ratio to State avg % Two or more race (NH)
#> 57                   1                                    Ratio to State avg % Native Hawaiian Pacific Islander (NHA)
#> 58                 0.9                                             Ratio to State avg % White nonHispanic single race
#> 59                   1                                                                       Ratio to State avg PM2.5
#> 60                   1                                                                       Ratio to State avg Ozone
#> 61                 1.2                                                                 Ratio to State avg Cancer risk
#> 62                 1.1                                                                 Ratio to State avg Respiratory
#> 63                 1.5                                                                   Ratio to State avg Diesel PM
#> 64                 1.4                                                            Ratio to State avg % built pre-1960
#> 65                 1.4                                                                     Ratio to State avg Traffic
#> 66                 4.3                                                                         Ratio to State avg NPL
#> 67                 1.2                                                                         Ratio to State avg RMP
#> 68                 1.6                                                                        Ratio to State avg TSDF
#> 69                   0                                                                       Ratio to State avg NPDES
#> 70                 1.4                                                                         Ratio to State avg UST
#> 71                 1.1                                                       Ratio to State avg Toxic Releases to Air
#> 72                   0                                                                              Demographic Index
#> 73                   0                                                                 Supplemental Demographic Index
#> 74                   0                                                                                   % Low Income
#> 75                   0                                                       % in limited English-speaking Households
#> 76                   0                                                                                   % Unemployed
#> 77                   0                                                         % with Less Than High School Education
#> 78                   0                                                                            Low life expectancy
#> 79                   0                                                                                  % under Age 5
#> 80                   0                                                                                  % over Age 64
#> 81                   0                                                                              % People of Color
#> 82                   0                                                                           % Hispanic or Latino
#> 83                   0                                        % Black or African American (non-Hispanic, single race)
#> 84                   0                                                            % Asian (non-Hispanic, single race)
#> 85                   0                                % American Indian and Alaska Native (non-Hispanic, single race)
#> 86                   0                       % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 87                   0                                                       % Other race (non-Hispanic, single race)
#> 88                   0                                                             % Two or more races (non-Hispanic)
#> 89                   1                                                            % White (non-Hispanic, single race)
#> 90                  64                                                            US percentile for Demographic Index
#> 91                  72                                               US percentile for Supplemental Demographic Index
#> 92                  69                                                                 US percentile for % Low Income
#> 93                  80                                     US percentile for % in limited English-speaking Households
#> 94                  63                                                                 US percentile for % Unemployed
#> 95                  71                                       US percentile for % with Less Than High School Education
#> 96                  70                                                          US percentile for Low life expectancy
#> 97                  64                                                                US percentile for % under Age 5
#> 98                  30                                                                US percentile for % over Age 64
#> 99                  59                                                            US percentile for % People of Color
#> 100                 76                                                         US percentile for % Hispanic or Latino
#> 101                 68                      US percentile for % Black or African American (non-Hispanic, single race)
#> 102                 68                                          US percentile for % Asian (non-Hispanic, single race)
#> 103                 90              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 104                  0     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 105                 87                                     US percentile for % Other race (non-Hispanic, single race)
#> 106                 64                                           US percentile for % Two or more races (non-Hispanic)
#> 107                 40                                          US percentile for % White (non-Hispanic, single race)
#> 108                  0                                                                US Average of Demographic Index
#> 109                  0                                                   US Average of Supplemental Demographic Index
#> 110                  0                                                                    US average for % Low Income
#> 111                  0                                        US average for % in limited English-speaking Households
#> 112                  0                                                                    US average for % Unemployed
#> 113                  0                                          US average for % with Less Than High School Education
#> 114                  0                                                             US average for Low life expectancy
#> 115                  0                                                                   US average for % under Age 5
#> 116                  0                                                                   US average for % over Age 64
#> 117                  0                                                               US average for % People of Color
#> 118                  0                                                                    US avg % Hispanic or Latino
#> 119                  0                                 US avg % Black or African American (non-Hispanic, single race)
#> 120                  0                                                     US avg % Asian (non-Hispanic, single race)
#> 121                  0                         US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 122                  0                US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 123                  0                                                US avg % Other race (non-Hispanic, single race)
#> 124                  0                                                      US avg % Two or more races (non-Hispanic)
#> 125                  1                                                     US avg % White (non-Hispanic, single race)
#> 126                  0                                                                 State avg % Hispanic or Latino
#> 127                  0                              State avg % Black or African American (non-Hispanic, single race)
#> 128                  0                                                  State avg % Asian (non-Hispanic, single race)
#> 129                  0                      State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 130                  0             State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 131                  0                                             State avg % Other race (non-Hispanic, single race)
#> 132                  0                                                   State avg % Two or more races (non-Hispanic)
#> 133                  1                                                  State avg % White (non-Hispanic, single race)
#> 134                 69                                                         State percentile for Demographic Index
#> 135                 71                                            State percentile for Supplemental Demographic Index
#> 136                 66                                                              State percentile for % Low Income
#> 137                 81                                  State percentile for % in limited English-speaking Households
#> 138                 65                                                              State percentile for % Unemployed
#> 139                 74                                    State percentile for % with Less Than High School Education
#> 140                 67                                                       State percentile for Low life expectancy
#> 141                 63                                                             State percentile for % under Age 5
#> 142                 28                                                             State percentile for % over Age 64
#> 143                 67                                                         State percentile for % People of Color
#> 144                 76                                                      State percentile for % Hispanic or Latino
#> 145                 62                   State percentile for % Black or African American (non-Hispanic, single race)
#> 146                 73                                       State percentile for % Asian (non-Hispanic, single race)
#> 147                 52           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 148                  0  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 149                 25                                  State percentile for % Other race (non-Hispanic, single race)
#> 150                 59                                        State percentile for % Two or more races (non-Hispanic)
#> 151                 32                                       State percentile for % White (non-Hispanic, single race)
#> 152                  0                                                             State Average of Demographic Index
#> 153                  0                                                State Average of Supplemental Demographic Index
#> 154                  0                                                                 State average for % Low Income
#> 155                  0                                     State average for % in limited English-speaking Households
#> 156                  0                                                                 State average for % Unemployed
#> 157                  0                                       State average for % with Less Than High School Education
#> 158                  0                                                          State average for Low life expectancy
#> 159                  0                                                                State average for % under Age 5
#> 160                  0                                                                State average for % over Age 64
#> 161                  0                                                            State average for % People of Color
#> 162               8.03                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 163               60.6                                                                                    Ozone (ppb)
#> 164                 27                                                      Air Toxics Cancer Risk (risk per million)
#> 165               0.29                                                            Air Toxics Respiratory Hazard Index
#> 166               0.36                                                   Air Toxics Diesel Particulate Matter (ug/m3)
#> 167               41.3                                                     Lead Paint Indicator (% pre-1960s housing)
#> 168                179                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 169               0.39                                                   Superfund Proximity (site count/km distance)
#> 170                0.6                                                     RMP Proximity (facility count/km distance)
#> 171                1.4                           Indicator for Hazardous Waste Proximity (facility count/km distance)
#> 172                  0                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 173                5.5                                                      Underground Storage Tanks (UST) indicator
#> 174               6493                                                                          Toxic Releases to Air
#> 175                 45                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 176                 46                                                                  US percentile for Ozone (ppb)
#> 177                  5                                    US percentile for Air Toxics Cancer Risk (risk per million)
#> 178                  4                                          US percentile for Air Toxics Respiratory Hazard Index
#> 179                 79                                 US percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 180                 67                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 181                 72          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 182                 93                                 US percentile for Superfund Proximity (site count/km distance)
#> 183                 79                                   US percentile for RMP Proximity (facility count/km distance)
#> 184                 66                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 185                 70    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 186                 79                                    US percentile for Underground Storage Tanks (UST) indicator
#> 187                 89                                                        US percentile for Toxic Releases to Air
#> 188               8.08                                            US average for Particulate Matter (PM 2.5 in ug/m3)
#> 189               61.6                                                                     US average for Ozone (ppb)
#> 190                 25                                       US average for Air Toxics Cancer Risk (risk per million)
#> 191               0.31                                             US average for Air Toxics Respiratory Hazard Index
#> 192               0.26                                    US average for Air Toxics Diesel Particulate Matter (ug/m3)
#> 193               30.3                                      US average for Lead Paint Indicator (% pre-1960s housing)
#> 194                212             US average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 195               0.13                                    US average for Superfund Proximity (site count/km distance)
#> 196               0.43                                      US average for RMP Proximity (facility count/km distance)
#> 197                1.9                          US average for Hazardous Waste Proximity (facility count/km distance)
#> 198                 22       US average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 199                3.9                                       US average for Underground Storage Tanks (UST) indicator
#> 200               4612                                                                   US avg Toxic Releases to Air
#> 201                 67                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 202                 53                                                               State percentile for Ozone (ppb)
#> 203                 21                                 State percentile for Air Toxics Cancer Risk (risk per million)
#> 204                 37                                       State percentile for Air Toxics Respiratory Hazard Index
#> 205                 78                              State percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 206                 65                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 207                 75       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 208                 74                              State percentile for Superfund Proximity (site count/km distance)
#> 209                 69                                State percentile for RMP Proximity (facility count/km distance)
#> 210                 72                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 211                 55 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 212                 76                                 State percentile for Underground Storage Tanks (UST) indicator
#> 213                 78                                                     State percentile for Toxic Releases to Air
#> 214               7.89                                         State average for Particulate Matter (PM 2.5 in ug/m3)
#> 215               60.6                                                                  State average for Ozone (ppb)
#> 216                 22                                    State average for Air Toxics Cancer Risk (risk per million)
#> 217               0.26                                          State average for Air Toxics Respiratory Hazard Index
#> 218               0.25                                 State average for Air Toxics Diesel Particulate Matter (ug/m3)
#> 219               29.1                                   State average for Lead Paint Indicator (% pre-1960s housing)
#> 220                126          State average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 221               0.09                                 State average for Superfund Proximity (site count/km distance)
#> 222               0.48                                   State average for RMP Proximity (facility count/km distance)
#> 223                0.8                       State average for Hazardous Waste Proximity (facility count/km distance)
#> 224                 25    State average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 225                3.9                                    State average for Underground Storage Tanks (UST) indicator
#> 226               5917                                                                State avg Toxic Releases to Air
#> 227                 69                                     US percentile for EJ Index for Particulate Matter (PM 2.5)
#> 228                 63                                                           US percentile for EJ Index for Ozone
#> 229                 71                                          US percentile for EJ Index for Air Toxics Cancer Risk
#> 230                 64                             US percentile for EJ Index for Air Toxics Respiratory Hazard Index
#> 231                 77                            US percentile for EJ Index for Air Toxics Diesel Particulate Matter
#> 232                 70                                            US percentile for EJ Index for Lead Paint Indicator
#> 233                 70                                    US percentile for EJ Index for Traffic Proximity and Volume
#> 234                 79                                             US percentile for EJ Index for Superfund Proximity
#> 235                 77                                                   US percentile for EJ Index for RMP Proximity
#> 236                 69                                       US percentile for EJ Index for Hazardous Waste Proximity
#> 237                 70                                  US percentile for EJ Index for Wastewater Discharge Indicator
#> 238                 75                       US percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 239                 81                                           US percentile for EJ Index for Toxic Releases to Air
#> 240                 74                        US percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)
#> 241                 69                                              US percentile for EJ Supplemental Index for Ozone
#> 242                 73                             US percentile for EJ Supplemental Index for Air Toxics Cancer Risk
#> 243                 66                US percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index
#> 244                 84               US percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter
#> 245                 74                               US percentile for EJ Supplemental Index for Lead Paint Indicator
#> 246                 76                       US percentile for EJ Supplemental Index for Traffic Proximity and Volume
#> 247                 85                                US percentile for EJ Supplemental Index for Superfund Proximity
#> 248                 82                                      US percentile for EJ Supplemental Index for RMP Proximity
#> 249                 76                          US percentile for EJ Supplemental Index for Hazardous Waste Proximity
#> 250                 75                     US percentile for EJ Supplemental Index for Wastewater Discharge Indicator
#> 251                 78          US percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator
#> 252                 85                                     US Percentile for Toxic Releases to Air Supplemental Index
#> 253                 72                                  State percentile for EJ Index for Particulate Matter (PM 2.5)
#> 254                 64                                                        State percentile for EJ Index for Ozone
#> 255                 81                                       State percentile for EJ Index for Air Toxics Cancer Risk
#> 256                 75                          State percentile for EJ Index for Air Toxics Respiratory Hazard Index
#> 257                 77                         State percentile for EJ Index for Air Toxics Diesel Particulate Matter
#> 258                 67                                         State percentile for EJ Index for Lead Paint Indicator
#> 259                 74                                 State percentile for EJ Index for Traffic Proximity and Volume
#> 260                 72                                          State percentile for EJ Index for Superfund Proximity
#> 261                 74                                                State percentile for EJ Index for RMP Proximity
#> 262                 72                                    State percentile for EJ Index for Hazardous Waste Proximity
#> 263                 66                               State percentile for EJ Index for Wastewater Discharge Indicator
#> 264                 73                    State percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 265                 75                                        State percentile for EJ Index for Toxic Releases to Air
#> 266                 73                     State percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)
#> 267                 65                                           State percentile for EJ Supplemental Index for Ozone
#> 268                 81                          State percentile for EJ Supplemental Index for Air Toxics Cancer Risk
#> 269                 77             State percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index
#> 270                 80            State percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter
#> 271                 67                            State percentile for EJ Supplemental Index for Lead Paint Indicator
#> 272                 77                    State percentile for EJ Supplemental Index for Traffic Proximity and Volume
#> 273                 74                             State percentile for EJ Supplemental Index for Superfund Proximity
#> 274                 75                                   State percentile for EJ Supplemental Index for RMP Proximity
#> 275                 74                       State percentile for EJ Supplemental Index for Hazardous Waste Proximity
#> 276                 66                  State percentile for EJ Supplemental Index for Wastewater Discharge Indicator
#> 277                 74       State percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator
#> 278                 78                                  State Percentile for Toxic Releases to Air Supplemental Index
#> 279              21641                                                                      Low income resident count
#> 280               1477                                                            Limited English-speaking Households
#> 281               1738                                                                      Unemployed resident count
#> 282               5487                                                 Less Than High School Education resident count
#> 283               3902                                                                     Under Age 5 resident count
#> 284               6580                                                                     Over Age 64 resident count
#> 285              24807                                                                 People of Color resident count
#> 286              15471                                                                    Count of Hispanic or Latino
#> 287               5218                                 Count of Black or African American (non-Hispanic, single race)
#> 288               1803                                                     Count of Asian (non-Hispanic, single race)
#> 289                537                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 290                  1                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 291                 55                                                Count of Other race (non-Hispanic, single race)
#> 292               1723                                                      Count of Two or more races (non-Hispanic)
#> 293              35933                                                     Count of White (non-Hispanic, single race)
#> 294              54648                                      Count of Population for whom Poverty Status is Determined
#> 295              36889                                                                  Count of Population Age 25 up
#> 296              21660                                                                            Count of Households
#> 297              30403                                           Universe for percent unemployed (denominator, count)
#> 298              10184                                                          Count of Housing Units Built Pre 1960
#> 299              24671                                   Built housing units count (denominator for percent pre 1960)
#> 300              56838                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 301              11331                                                                          Speak Spanish at Home
#> 302                707                                                           Persons who Speak English Not at All
#> 303               1153                                                              Speak Other Indo-European at Home
#> 304               1367                                                    Speak Asian-Pacific Island language at Home
#> 305               1121                                                   Speak Spanish (in limited English household)
#> 306                179                                       Speak Other Indo-European (in limited English household)
#> 307                165                                      Speak Asian-Pacific Island (in limited English household)
#> 308                 12                                                     Speak Other (in limited English household)
#> 309              13015                                                                        Population Under Age 18
#> 310              47725                                                                         Population Over Age 17
#> 311              30954                                                                                Male Population
#> 312              29786                                                                              Female Population
#> 313              11127                                                                   Owner Occupied Housing Units
#> 314              21660                                                                         Occupied Housing Units
#> 315              58670                                                       Civilian noninstitutionalized population
#> 316             6852.1                                                                                     DISABILITY
#> 317                 76                                                                       Life Expectancy in Years
#> 318              27397                                                                              Per Capita Income
#> 319                  0                                                            Pct. Households below Poverty Level
#> 320                  1                                  Percent Speaking Spanish (as % of limited English households)
#> 321                  0            Percent Speaking Other Indo-European languages (as % of limited English households)
#> 322                  0           Percent Speaking Asian-Pacific Island languages (as % of limited English households)
#> 323                  0                          Percent Speaking Other languages (as % of limited English households)
#> 324                  0                                                       Percent Break Down by Age below 18 Years
#> 325                  1                                                       Percent Break Down by Age above 17 Years
#> 326                  1                                                                                  Percent Males
#> 327                  0                                                                                Percent Females
#> 328                 75                                                                     Census blockgroups (count)
#> 329               1152                                                                          Census blocks (count)
#> 330                  1                                                                        Buffer distance (miles)
#> 331               4126                                                                 Households below Poverty Level
#> 332             21.749                                                                Particulate Matter 2.5 EJ Index
#> 333              18.28                                                                                 Ozone EJ Index
#> 334             16.944                                                                Air toxics cancer risk EJ Index
#> 335             14.322                                                             Air toxics respiratory HI EJ Index
#> 336             31.306                                                             Diesel particulate matter EJ Index
#> 337             21.888                                                                            Lead paint EJ Index
#> 338              25.68                                                              US raw Traffic proximity EJ Index
#> 339             31.366                                                            US raw Superfund Proximity EJ Index
#> 340             30.822                                                                RMP Facility Proximity EJ Index
#> 341             24.123                                                             Hazardous waste proximity EJ Index
#> 342             22.239                                                           US raw Wastewater discharge EJ Index
#> 343             28.501                                                      US raw Underground storage tanks EJ Index
#> 344             30.878                                                                 Toxic Releases to Air EJ Index
#> 345             10.022                                                      Particulate Matter 2.5 Supplemental Index
#> 346               8.39                                                                       Ozone Supplemental Index
#> 347              6.944                                                      Air Toxics Cancer Risk Supplemental Index
#> 348              5.949                                                   Air Toxics Respiratory HI Supplemental Index
#> 349             13.462                                                   Diesel Particulate Matter Supplemental Index
#> 350             10.484                                                                  Lead Paint Supplemental Index
#> 351             11.215                                                           Traffic Proximity Supplemental Index
#> 352               13.5                                                         Superfund Proximity Supplemental Index
#> 353             13.284                                                      RMP Facility Proximity Supplemental Index
#> 354              10.79                                                   Hazardous Waste Proximity Supplemental Index
#> 355             10.218                                                        Wastewater Discharge Supplemental Index
#> 356             12.078                                                   Underground Storage Tanks Supplemental Index
#> 357             13.612                                                           EJ Supp: Toxic Releases to Air (raw)
#> 358             27.806                                                      State raw Particulate Matter 2.5 EJ Index
#> 359             19.867                                                                       State raw Ozone EJ Index
#> 360             21.399                                                          State Air toxics cancer risk EJ Index
#> 361             17.843                                                       State Air toxics respiratory HI EJ Index
#> 362             32.732                                                       State Diesel particulate matter EJ Index
#> 363             23.756                                                                      State Lead paint EJ Index
#> 364             27.033                                                           State raw Traffic proximity EJ Index
#> 365             33.055                                                         State raw Superfund Proximity EJ Index
#> 366             30.084                                                      State raw RMP Facility Proximity EJ Index
#> 367             29.713                                                       State Hazardous waste proximity EJ Index
#> 368             21.378                                                        State raw Wastewater discharge EJ Index
#> 369             26.861                                                   State raw Underground storage tanks EJ Index
#> 370             31.691                                                           State Toxic Releases to Air EJ Index
#> 371             12.478                                            State raw Particulate Matter 2.5 Supplemental Index
#> 372              9.358                                                             State raw Ozone Supplemental Index
#> 373              9.181                                                State Air toxics cancer risk Supplemental Index
#> 374              8.069                                             State Air toxics respiratory HI Supplemental Index
#> 375             14.214                                             State Diesel particulate matter Supplemental Index
#> 376             11.008                                                            State Lead paint Supplemental Index
#> 377              11.99                                                 State raw Traffic proximity Supplemental Index
#> 378             14.316                                               State raw Superfund Proximity Supplemental Index
#> 379             12.752                                            State raw RMP Facility Proximity Supplemental Index
#> 380             13.015                                             State Hazardous waste proximity Supplemental Index
#> 381              9.696                                              State raw Wastewater discharge Supplemental Index
#> 382             11.524                                         State raw Underground storage tanks Supplemental Index
#> 383             13.853                                                 State Toxic Releases to Air Supplemental Index
#> 384              0.018                                                                                   distance_min
#> 385  0.751551987007954                                                                         distance_min_avgperson
#> 386                  1                                                                                  sitecount_max
#> 387                  9                                                                               sitecount_unique
#> 388                  1                                                                                  sitecount_avg
#> 389                  0                                                           Percent population with Disabilities
#> 390                  1                                                                      Owner Occupied households
#> 391 0.0124440738023481                                                                                    pct_lan_eng
#> 392  0.199349734246098                                                                                pct_lan_spanish
#> 393  0.020282098514558                                                                                     pct_lan_ie
#> 394 0.0240542028878415                                                                                    pct_lan_api

x <- ejam2report(out, sitenumber = 1)
table_gt_from_ejamit_1site(out$results_bysite[1, ])
browseURL(x)