See ejamit()$results_bysite in interactive table in RStudio viewer pane

See ejamit()$results_bysite in interactive table in RStudio viewer pane

Usage

ejam2tableviewer(
  out,
  fname = "automatic",
  maxrows = 1000,
  launch_browser = TRUE
)

Arguments

out

output of ejamit(), or one table like ejamit()$results_overall, or subset like ejamit()$results_bysite7,

fname

optional. path and filename of the html file to save the table to, or it uses tempdir() if not specified. Set it to NULL to prevent saving a file.

maxrows

only load/ try to show this many rows max.

launch_browser

set TRUE to have it launch browser and show report. Ignored if not interactive() or if fname is set to NULL.

Value

a datatable object using DT::datatable() that can be printed to the console or shown in the RStudio viewer pane

Examples

ejam2tableviewer(testoutput_ejamit_10pts_1miles)
#> 
#> Interactive table of sites is saved here:  /tmp/RtmprpwYDs/EJAM_results_bysite_1_Miles_from_20250103_221209.html 
#> To open that folder: browseURL('/tmp/RtmprpwYDs')
#> To view that report in a browser: browseURL('/tmp/RtmprpwYDs/EJAM_results_bysite_1_Miles_from_20250103_221209.html')