Compare ejamit() full results for more than one radius Helper used by ejamit_compare_distances() to run ejamit() once per radius, get FULL ejamit() output list per radius

Compare ejamit() full results for more than one radius Helper used by ejamit_compare_distances() to run ejamit() once per radius, get FULL ejamit() output list per radius

Usage

ejamit_compare_distances_fulloutput(
  sitepoints,
  radii = c(1, 2, 3),
  donuts_not_cumulative = FALSE,
  quiet = TRUE,
  silentinteractive = TRUE,
  ...
)

Arguments

sitepoints

like for ejamit()

radii

vector of radius values like 1:3 for ejamit()

donuts_not_cumulative

set to TRUE to get results on each ring not each full circle

quiet

passed to ejamit()

silentinteractive

passed to ejamit()

...

passed to ejamit()

Value

list you can think of as "out_bydistance" where each element is the full output of ejamit() for 1 radius

Details

You typically only need ejamit_compare_distances(), which gives you just the summary overall at each distance, but if you want to retain the full outputs of ejamit() at each distance, such as results for every site at every distances, you can use ejamit_compare_distances_fulloutput() and then to extract a slice of results, use helper functions like

out_bydistance2results_bydistance()

out_bydistance2results_bydistance_bysite()

out_bydistance2results_bysite_bydistance()

See also

Examples

  radii <- c(1,2,3,6,10)
  pts <- testpoints_10
  if (FALSE) { # \dontrun{
  x <- ejamit_compare_distances_fulloutput(pts, radii = radii)
  } # }