EJAM/EJSCREEN comparisons

This is the main function that facilitates comparing EJScreen API vs EJAM stats near site(s)

Usage

ejscreen_vs_ejam(
  latlon,
  radius = 3,
  nadrop = FALSE,
  save_when_report = TRUE,
  report_every_n = 250,
  calculate_ratios = FALSE,
  include_ejindexes = TRUE,
  x100fix = TRUE,
  x100varnames = c(names_d, names_d_avg, names_d_state_avg, names_d_subgroups,
    names_d_subgroups_avg, names_d_subgroups_state_avg, "pctdisability",
    "p_own_occupied", "pctunder18", "pctover17", "pctmale", "pctfemale"),
  ...
)

Arguments

latlon

data.table or data.frame with colnames lat and lon, and one row per site

radius

in miles, used in ejamit() and ejscreenapi_plus()

save_when_report

see ejscreenapi_plus()

report_every_n

see ejscreenapi_plus()

calculate_ratios

passed to ejscreenapi_plus() and ejamit()

x100fix

whether to multiply x100 the names_d and names_d_subgroups indicator scores to convert fractions 0 to 1 into percentages of 0 to 100, prior to rounding and reporting EJAM results here.

x100varnames

optional, if x100fix=T, a vector of colnames of x$EJAM to convert from being scaled as 0 to 1 into rescaled values of 0 to 100, because some outputs of EJSCREEN were reported as percentages 0 to 100 but as 0 to 1 in EJAM.

...

passed to ejamit() as any additional parameters, like include_ejindexes = FALSE

Value

a list of data frames, with names EJSCREEN, EJAM, EJSCREEN_shown, EJAM_shown, same_shown, ratio, diff, absdiff, pctdiff, abspctdiff

diff is EJAM - EJSCREEN

ratio is EJAM / EJSCREEN

pctdiff is ratio - 1

abs is absolute value

For each data.frame, colnames are indicators like pop, blockcount_near_site, etc. and rows represent sites analyzed.

Details

Note that the EJAM tool/ function called ejamit() does not rely on EJScreen to do the calculations and instead tries to replicate what EJScreen would do. As a result, as of early 2024 at least, while

ejamit() is much, much faster than ejscreenit_for_ejam() and

provides additional information (distribution of distances by group, etc.)

features (histograms, spreadsheet with heatmaps, etc.)

flexibility (easy for analysts using R to customize analysis, etc.), ejamit() does not always exactly replicate EJScreen -- does not provide 100% identical results (percentiles, etc.) for every indicator in every analysis at every location. This is due to slight variations in

details of the spatial calculations (which blocks are nearby, sometimes counting 1 extra block as 2.99 miles away while EJScreen counts it as outside the 3 mile radius, e.g.)

rounding (how many digits are retained during calculations, and how many are shown in final reports)

percentile assignment method (how percentile lookup tables are used, how ties are treated in percentile lookup tables, etc.), or

weighted averages or other formulas used for aggregation of blockgroup scores (being updated in 2024 to more precisely match the formulas EJScreen uses)

See also

Examples