EJAM/EJSCREEN comparisons

Compare EJScreen API vs EJAM stats near site(s) (after results already run)

Usage

ejscreen_vs_ejam_alreadyrun(
  apisite,
  ejamsite,
  nadrop = FALSE,
  x100fix = TRUE,
  x100varnames = c(names_d, names_d_avg, names_d_state_avg, names_d_subgroups,
    names_d_subgroups_avg, names_d_subgroups_state_avg, "pctdisability",
    "p_own_occupied", "pctunder18", "pctover17", "pctmale", "pctfemale")
)

Arguments

apisite

table output of ejscreenit()$table, or ejscreenapi_plus(), and also see ejscreenit_for_ejam()

ejamsite

table output of ejamit()$results_bysite

nadrop

optional, whether to drop indicators for which EJScreen API returns NA

x100fix

optional, whether to multiply x100 the names_d and names_d_subgroups indicator scores to convert fractions 0 to 1 into percentages of 0 to 100, prior to rounding and reporting EJAM results here.

x100varnames

optional, if x100fix=T, a vector of colnames of x$EJAM to convert from being scaled as 0 to 1 into rescaled values of 0 to 100, because some outputs of EJSCREEN were reported as percentages 0 to 100 but as 0 to 1 in EJAM.

Value

prints summary to console, but returns invisible a list of data frames, with names EJSCREEN, EJAM, EJSCREEN_shown, EJAM_shown, same_shown, ratio, etc.

For each data.frame, colnames are indicators like pop, blockcount_near_site, etc. and rows represent sites analyzed.

See also

Examples

  blah = ejscreen_vs_ejam_alreadyrun(
    apisite = testoutput_ejscreenapi_plus_5,
    ejamsite = ejamit(testpoints_5, radius = 1, include_ejindexes = TRUE)$results_bysite)
#> Finding blocks nearby.
#> Analyzing 5 points, radius of 1 miles around each.
#> Aggregating at each buffer and overall.
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> bgej         - is NOT in memory. Checking local disk... bgej        is NOT found locally on disk at ~/../Downloads/EJAMbigfiles/bgej.arrow 
#> bgej         - is NOT in local folder.
#> bgej        - was read from pins board server, into memory (to specified envir).
#> 
#> loaded bgej data because include_ejindexes = TRUE
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for pm.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for pm.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for o3.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for o3.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for cancer.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for cancer.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for resp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for resp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for dpm.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for dpm.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for traffic.score.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for traffic.score.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for proximity.npl.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for proximity.npl.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for proximity.rmp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for proximity.rmp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for proximity.tsdf.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for proximity.tsdf.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for proximity.npdes.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for proximity.npdes.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for ust.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for ust.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for rsei.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for rsei.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for Demog.Index.Supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for Demog.Index.Supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for lowlifex.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for lowlifex.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.pm.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.pm.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.o3.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.o3.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.cancer.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.cancer.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.resp.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.resp.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.dpm.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.dpm.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.pctpre1960.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.pctpre1960.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.traffic.score.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.traffic.score.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.npl.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.npl.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.rmp.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.rmp.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.tsdf.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.tsdf.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.npdes.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.npdes.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.ust.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.ust.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.rsei.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.rsei.eo.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.pm.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.pm.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.o3.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.o3.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.cancer.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.cancer.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.resp.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.resp.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.dpm.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.dpm.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.pctpre1960.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.pctpre1960.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.traffic.score.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.traffic.score.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.npl.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.npl.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.rmp.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.rmp.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.tsdf.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.tsdf.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.proximity.npdes.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.proximity.npdes.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.ust.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.ust.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = TX for state.EJ.DISPARITY.rsei.supp.
#> Among these results, all raw scores were NA (so percentiles will be reported as NA) in zone = CA for state.EJ.DISPARITY.rsei.supp.
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> 
#> 
#>                                          indicator pct.of.sites.agree.rounded
#> 160                                           resp                  100.00000
#> 222                      pctile.EJ.DISPARITY.pm.eo                  100.00000
#> 223                      pctile.EJ.DISPARITY.o3.eo                  100.00000
#> 224                  pctile.EJ.DISPARITY.cancer.eo                  100.00000
#> 225                    pctile.EJ.DISPARITY.resp.eo                  100.00000
#> 226                     pctile.EJ.DISPARITY.dpm.eo                  100.00000
#> 228           pctile.EJ.DISPARITY.traffic.score.eo                  100.00000
#> 229           pctile.EJ.DISPARITY.proximity.npl.eo                  100.00000
#> 230           pctile.EJ.DISPARITY.proximity.rmp.eo                  100.00000
#> 231          pctile.EJ.DISPARITY.proximity.tsdf.eo                  100.00000
#> 232         pctile.EJ.DISPARITY.proximity.npdes.eo                  100.00000
#> 233                     pctile.EJ.DISPARITY.ust.eo                  100.00000
#> 234                    pctile.EJ.DISPARITY.rsei.eo                  100.00000
#> 235                    pctile.EJ.DISPARITY.pm.supp                  100.00000
#> 236                    pctile.EJ.DISPARITY.o3.supp                  100.00000
#> 237                pctile.EJ.DISPARITY.cancer.supp                  100.00000
#> 238                  pctile.EJ.DISPARITY.resp.supp                  100.00000
#> 239                   pctile.EJ.DISPARITY.dpm.supp                  100.00000
#> 241         pctile.EJ.DISPARITY.traffic.score.supp                  100.00000
#> 242         pctile.EJ.DISPARITY.proximity.npl.supp                  100.00000
#> 243         pctile.EJ.DISPARITY.proximity.rmp.supp                  100.00000
#> 244        pctile.EJ.DISPARITY.proximity.tsdf.supp                  100.00000
#> 245       pctile.EJ.DISPARITY.proximity.npdes.supp                  100.00000
#> 247                  pctile.EJ.DISPARITY.rsei.supp                  100.00000
#> 248                state.pctile.EJ.DISPARITY.pm.eo                  100.00000
#> 250            state.pctile.EJ.DISPARITY.cancer.eo                  100.00000
#> 251              state.pctile.EJ.DISPARITY.resp.eo                  100.00000
#> 252               state.pctile.EJ.DISPARITY.dpm.eo                  100.00000
#> 253        state.pctile.EJ.DISPARITY.pctpre1960.eo                  100.00000
#> 254     state.pctile.EJ.DISPARITY.traffic.score.eo                  100.00000
#> 255     state.pctile.EJ.DISPARITY.proximity.npl.eo                  100.00000
#> 259               state.pctile.EJ.DISPARITY.ust.eo                  100.00000
#> 260              state.pctile.EJ.DISPARITY.rsei.eo                  100.00000
#> 262              state.pctile.EJ.DISPARITY.o3.supp                  100.00000
#> 263          state.pctile.EJ.DISPARITY.cancer.supp                  100.00000
#> 264            state.pctile.EJ.DISPARITY.resp.supp                  100.00000
#> 265             state.pctile.EJ.DISPARITY.dpm.supp                  100.00000
#> 266      state.pctile.EJ.DISPARITY.pctpre1960.supp                  100.00000
#> 268   state.pctile.EJ.DISPARITY.proximity.npl.supp                  100.00000
#> 269   state.pctile.EJ.DISPARITY.proximity.rmp.supp                  100.00000
#> 270  state.pctile.EJ.DISPARITY.proximity.tsdf.supp                  100.00000
#> 271 state.pctile.EJ.DISPARITY.proximity.npdes.supp                  100.00000
#> 273            state.pctile.EJ.DISPARITY.rsei.supp                  100.00000
#> 157                                             pm                  100.00000
#> 158                                             o3                  100.00000
#> 161                                            dpm                  100.00000
#> 76                                          pctmin                  100.00000
#> 159                                         cancer                  100.00000
#> 166                                 proximity.tsdf                  100.00000
#> 162                                     pctpre1960                  100.00000
#> 67                                     Demog.Index                  100.00000
#> 164                                  proximity.npl                  100.00000
#> 167                                proximity.npdes                  100.00000
#> 75                                       pctover64                  100.00000
#> 69                                       pctlowinc                  100.00000
#> 165                                  proximity.rmp                  100.00000
#> 71                                   pctunemployed                  100.00000
#> 70                                      pctlingiso                  100.00000
#> 72                                         pctlths                  100.00000
#> 324                           blockcount_near_site                   80.00000
#> 240            pctile.EJ.DISPARITY.pctpre1960.supp                   66.66667
#> 256     state.pctile.EJ.DISPARITY.proximity.rmp.eo                   66.66667
#> 249                state.pctile.EJ.DISPARITY.o3.eo                   66.66667
#> 261              state.pctile.EJ.DISPARITY.pm.supp                   66.66667
#> 257    state.pctile.EJ.DISPARITY.proximity.tsdf.eo                   66.66667
#> 267   state.pctile.EJ.DISPARITY.traffic.score.supp                   66.66667
#> 168                                            ust                   66.66667
#> 258   state.pctile.EJ.DISPARITY.proximity.npdes.eo                   66.66667
#> 246                   pctile.EJ.DISPARITY.ust.supp                   66.66667
#> 68                                Demog.Index.Supp                   66.66667
#> 272             state.pctile.EJ.DISPARITY.ust.supp                   66.66667
#> 74                                       pctunder5                   66.66667
#> 2                                              pop                   60.00000
#> 169                                           rsei                   33.33333
#> 73                                        lowlifex                   33.33333
#> 227              pctile.EJ.DISPARITY.pctpre1960.eo                   33.33333
#> 163                                  traffic.score                    0.00000
#>     pct.of.sites.agree.within.tol median.pct.diff max.pct.diff
#> 160                     100.00000            0.00         0.00
#> 222                     100.00000            0.00         0.00
#> 223                     100.00000            0.00         0.00
#> 224                     100.00000            0.00         0.00
#> 225                     100.00000            0.00         0.00
#> 226                     100.00000            0.00         0.00
#> 228                     100.00000            0.00         0.00
#> 229                     100.00000            0.00         0.00
#> 230                     100.00000            0.00         0.00
#> 231                     100.00000            0.00         0.00
#> 232                     100.00000            0.00         0.00
#> 233                     100.00000            0.00         0.00
#> 234                     100.00000            0.00         0.00
#> 235                     100.00000            0.00         0.00
#> 236                     100.00000            0.00         0.00
#> 237                     100.00000            0.00         0.00
#> 238                     100.00000            0.00         0.00
#> 239                     100.00000            0.00         0.00
#> 241                     100.00000            0.00         0.00
#> 242                     100.00000            0.00         0.00
#> 243                     100.00000            0.00         0.00
#> 244                     100.00000            0.00         0.00
#> 245                     100.00000            0.00         0.00
#> 247                     100.00000            0.00         0.00
#> 248                     100.00000            0.00         0.00
#> 250                     100.00000            0.00         0.00
#> 251                     100.00000            0.00         0.00
#> 252                     100.00000            0.00         0.00
#> 253                      66.66667            0.00         0.00
#> 254                     100.00000            0.00         0.00
#> 255                     100.00000            0.00         0.00
#> 259                     100.00000            0.00         0.00
#> 260                     100.00000            0.00         0.00
#> 262                     100.00000            0.00         0.00
#> 263                     100.00000            0.00         0.00
#> 264                     100.00000            0.00         0.00
#> 265                     100.00000            0.00         0.00
#> 266                     100.00000            0.00         0.00
#> 268                     100.00000            0.00         0.00
#> 269                     100.00000            0.00         0.00
#> 270                     100.00000            0.00         0.00
#> 271                     100.00000            0.00         0.00
#> 273                     100.00000            0.00         0.00
#> 157                     100.00000            0.04         0.05
#> 158                     100.00000            0.03         0.05
#> 161                     100.00000            0.13         0.37
#> 76                      100.00000            0.42         0.52
#> 159                     100.00000            0.00         0.69
#> 166                     100.00000            0.66         0.73
#> 162                      66.66667            0.78         1.07
#> 67                       66.66667            0.46         1.45
#> 164                      66.66667            0.44         1.82
#> 167                      66.66667            0.51         2.08
#> 75                       66.66667            0.72         2.48
#> 69                       66.66667            0.22         2.49
#> 165                      66.66667            0.84         3.27
#> 71                       33.33333            1.34         3.75
#> 70                       33.33333            1.78        23.36
#> 72                       66.66667            0.63        42.58
#> 324                      80.00000            0.00         1.41
#> 240                      33.33333            0.53         1.06
#> 256                      66.66667            0.00         1.20
#> 249                      66.66667            0.00         1.28
#> 261                      66.66667            0.00         1.49
#> 257                      66.66667            0.00         1.79
#> 267                      66.66667            0.00         1.85
#> 168                       0.00000            2.13         2.20
#> 258                      66.66667            0.00         2.22
#> 246                      66.66667            0.00         4.00
#> 68                        0.00000            3.99         4.93
#> 272                      66.66667            0.00         5.26
#> 74                        0.00000            6.81        10.36
#> 2                        40.00000            0.07         2.63
#> 169                     100.00000            0.37         0.69
#> 73                       33.33333            3.25        67.80
#> 227                      33.33333            1.32       100.00
#> 163                      33.33333            1.23        12.44
#>     within.x.pct.at.p.pct.of.sites
#> 160                           0.00
#> 222                           0.00
#> 223                           0.00
#> 224                           0.00
#> 225                           0.00
#> 226                           0.00
#> 228                           0.00
#> 229                           0.00
#> 230                           0.00
#> 231                           0.00
#> 232                           0.00
#> 233                           0.00
#> 234                           0.00
#> 235                           0.00
#> 236                           0.00
#> 237                           0.00
#> 238                           0.00
#> 239                           0.00
#> 241                           0.00
#> 242                           0.00
#> 243                           0.00
#> 244                           0.00
#> 245                           0.00
#> 247                           0.00
#> 248                           0.00
#> 250                           0.00
#> 251                           0.00
#> 252                           0.00
#> 253                           0.00
#> 254                           0.00
#> 255                           0.00
#> 259                           0.00
#> 260                           0.00
#> 262                           0.00
#> 263                           0.00
#> 264                           0.00
#> 265                           0.00
#> 266                           0.00
#> 268                           0.00
#> 269                           0.00
#> 270                           0.00
#> 271                           0.00
#> 273                           0.00
#> 157                           0.05
#> 158                           0.05
#> 161                           0.37
#> 76                            0.52
#> 159                           0.69
#> 166                           0.73
#> 162                           1.07
#> 67                            1.45
#> 164                           1.82
#> 167                           2.08
#> 75                            2.48
#> 69                            2.49
#> 165                           3.27
#> 71                            3.75
#> 70                           23.36
#> 72                           42.58
#> 324                           1.41
#> 240                           1.06
#> 256                           1.20
#> 249                           1.28
#> 261                           1.49
#> 257                           1.79
#> 267                           1.85
#> 168                           2.20
#> 258                           2.22
#> 246                           4.00
#> 68                            4.93
#> 272                           5.26
#> 74                           10.36
#> 2                             2.63
#> 169                           0.69
#> 73                           67.80
#> 227                         100.00
#> 163                          12.44
#> 
#> 
#>  Tolerance of 0.01 was used, so difference of <1% is within tolerance.
#> 
#> Probs (p) used was 0.95, so 95% of sites are within absolute percentage difference reported in output of this function.
#> 
  ejscreen_vs_ejam_see1(blah, mysite = 1)
#>                  EJSCREEN EJAM      EJSCREEN_shown EJAM_shown same_shown
#> Demog.Index      18       17.73886  "18"           "18"       TRUE      
#> Demog.Index.Supp 8        8.319381  "8"            "8"        TRUE      
#> pctlowinc        19       18.52749  "19"           "19"       TRUE      
#> pctlingiso       1        0.7664381 "1"            "1"        TRUE      
#> pctunemployed    6        5.774969  "6"            "6"        TRUE      
#> pctlths          1        1.42578   "1"            "1"        TRUE      
#> lowlifex         9        15.10223  "9"            "15"       FALSE     
#> pctunder5        3        2.795675  "3"            "3"        TRUE      
#> pctover64        30       30.15342  "30"           "30"       TRUE      
#> pctmin           17       16.95023  "17"           "17"       TRUE      
#>                  ratio     diff        absdiff    pctdiff      abspctdiff 
#> Demog.Index      0.9854922 -0.2611395  0.2611395  -0.01450775  0.01450775 
#> Demog.Index.Supp 1.039923  0.3193813   0.3193813  0.03992266   0.03992266 
#> pctlowinc        0.9751312 -0.4725066  0.4725066  -0.02486877  0.02486877 
#> pctlingiso       0.7664381 -0.2335619  0.2335619  -0.2335619   0.2335619  
#> pctunemployed    0.9624948 -0.2250315  0.2250315  -0.03750525  0.03750525 
#> pctlths          1.42578   0.4257802   0.4257802  0.4257802    0.4257802  
#> lowlifex         1.678025  6.102226    6.102226   0.6780252    0.6780252  
#> pctunder5        0.9318917 -0.2043249  0.2043249  -0.0681083   0.0681083  
#> pctover64        1.005114  0.1534249   0.1534249  0.005114163  0.005114163
#> pctmin           0.9970722 -0.04977248 0.04977248 -0.002927793 0.002927793
 if (FALSE) {
 
 # requires data.table
 
 # analyze point(s) in both EJScreen and EJAM
  pts <- testpoints_100[1:5, ]
  #z <- ejscreen_vs_ejam(testpoints_100[27, ], radius = 3, include_ejindexes = TRUE)
  
  # just 1 point
  # z <- ejscreen_vs_ejam(pts[5, ], radius = 3, include_ejindexes = TRUE)
  
  # multiple points
  # This step can take a long time, almost 1 minute per 20 points, as it uses the EJScreen API:
  z <- ejscreen_vs_ejam(pts, radius = 3, include_ejindexes = TRUE)
    
   # same thing but step by step
   api1 <- ejscreenit(pts, radius = 3)
   api1 <- api1$table[5, ]
   ejam1 <- ejamit(pts, radius = 3, include_ejindexes = TRUE)
   ejam1 <- ejam1$results_bysite[5, ]
   zzz <- ejscreen_vs_ejam_alreadyrun(api1, ejam1)
   
   # to keep it as a list of data.frames but just 1 row each now:
   zlist1 <- lapply(z, function(x) x[myrow, ])
   names(zlist1); sapply(zlist1, dim)
   
   # look at a table about just 1 site among those analyzed, e.g., the first one:
   myrow <- 1
   z1 <- data.frame(sapply(z, function(x) x[myrow, ]))
   # but that creates a data.frame where each column is a list... 
   
   
   
   
   # **** WORK IN PROGRESS ... *** !!!!!  
   
   
   
   
   
   # Reported key indicators - which ones do or don't match
   # when comparing EJSCREEN and EJAM results?
   keyreportnames <- c('pop', names_these, names_pctile, names_state_pctile)
   rname <- colnames(z$EJAM)
   z1[rname %in% keyreportnames &  z1$same_shown, ]
   z1[rname %in% keyreportnames & !z1$same_shown & !is.na(z1$EJSCREEN), ]
  
   # Reported (rounded) numbers match:
   z[z$same_shown , ]
   # Reports disagree: 
   #  (and not just because of percentages being of 100 vs of 1.00)
   z[!z$same_shown & !is.na(z$EJSCREEN) & z$ratio != 0.01, ] 
   # Reports disagree if percentages reported as 0-100 vs fractions 0-1.00
   z[z$ratio == 0.01 & !is.na(z$ratio), ]
   
   
   # site number 1
   z1 <- sapply(z, function(x) (x[1,]))
   # Reported key indicators - which ones do or don't match
   # when comparing EJSCREEN and EJAM results?
   keyreportnames <- c('pop', names_these, names_pctile, names_state_pctile)
   z1[z1$rname %in% keyreportnames &  z1$same_shown, ]
   z1[z1$rname %in% keyreportnames & !z1$same_shown & !is.na(z1$EJSCREEN), ]
   
   # Reported (rounded) numbers match:
   z1[z1$same_shown , ]
   # Reports disagree: 
   #  (and not just because of percentages being of 100 vs of 1.00)
   z1[!z1$same_shown & !is.na(z1$EJSCREEN) & z1$ratio != 0.01, ] 
   # Reports disagree if percentages reported as 0-100 vs fractions 0-1.00
   z1[z1$ratio == 0.01 & !is.na(z1$ratio), ]
   
   # Comparing across sites:
   
   # .... to be continued ...
   
  }