FIPS - Get county or state names from county or state FIPS codes

FIPS - Get county or state names from county or state FIPS codes

Usage

fips2name(fips, ...)

Arguments

fips

vector of US Census FIPS codes for

States (2 digits once any required leading zeroes are included)

Counties (5)

City/town/CDP (7)

Tracts (11)

Blockgroups (12) Can be string or numeric, with or without leading zeroes.

...

passed to fips2countyname() to control whether it appends something like , NY or , New York after county name

Value

vector of state and/or county names, where county names optionally have comma and 2-character abbreviation or full state name.

See also

Examples

  fips2name(fips_counties_from_state_abbrev(c("AK", "LA"))  )
  fips2name(c(22, 02013))  # can have mix where some are a whole state and others are a county.