FIPS - Get unique blockgroup fips in or containing specified fips of any type

Convert any FIPS codes to the FIPS of all the blockgroups that are among or within or containing those FIPS

Usage

fips_bg_from_anyfips(fips)

Arguments

fips

vector of US FIPS codes, as character or numeric, with or without their leading zeroes, each with as many characters

Value

vector of blockgroup FIPS (or NA values) that may be much longer than the vector of fips passed to this function.

Details

This is a way to get a list of blockgroups, specified by state/county/tract or even block.

Takes a vector of one or more FIPS that could be State (2-digit), County (5-digit), Tract (11-digit), or blockgroup (12 digit), or even block (15-digit fips).

Returns unique vector of FIPS of all US blockgroups (including DC and Puerto Rico) that contain any specified blocks, are equal to any specified blockgroup fips, or are contained within any provided tract/county/state FIPS.

See also

Examples

  # all blockgroups in one state
  fips_counties_from_state_abbrev("DE")
#> [1] "10001" "10003" "10005"
  fips_bg_from_anyfips( fips_counties_from_state_abbrev("DE") )
#>   [1] "100010401001" "100010401002" "100010401003" "100010401004" "100010402011" "100010402012"
#>   [7] "100010402013" "100010402031" "100010402032" "100010402033" "100010402041" "100010402042"
#>  [13] "100010402043" "100010402051" "100010402052" "100010402053" "100010402061" "100010405011"
#>  [19] "100010405012" "100010405013" "100010405021" "100010405022" "100010407001" "100010407002"
#>  [25] "100010407003" "100010409001" "100010409002" "100010410001" "100010410002" "100010410003"
#>  [31] "100010411011" "100010411012" "100010411013" "100010412001" "100010412002" "100010412003"
#>  [37] "100010413001" "100010413002" "100010414001" "100010414002" "100010415001" "100010415002"
#>  [43] "100010416001" "100010417011" "100010417012" "100010417013" "100010417014" "100010417021"
#>  [49] "100010417022" "100010418021" "100010418022" "100010418023" "100010418031" "100010418032"
#>  [55] "100010418041" "100010418042" "100010418043" "100010419001" "100010419002" "100010419003"
#>  [61] "100010420001" "100010420002" "100010421001" "100010421002" "100010422031" "100010422032"
#>  [67] "100010422033" "100010422041" "100010422042" "100010422051" "100010422052" "100010422061"
#>  [73] "100010422062" "100010422071" "100010422081" "100010422082" "100010425001" "100010425002"
#>  [79] "100010425003" "100010428011" "100010428021" "100010428022" "100010428023" "100010429001"
#>  [85] "100010429002" "100010430001" "100010430002" "100010430003" "100010431001" "100010431002"
#>  [91] "100010432021" "100010432022" "100010432023" "100010432024" "100010433001" "100010433002"
#>  [97] "100010433003" "100010434001" "100010434002" "100010434003" "100019800001" "100019900000"
#> [103] "100030002001" "100030002002" "100030002003" "100030002004" "100030002005" "100030002006"
#> [109] "100030003001" "100030003002" "100030003003" "100030004001" "100030004002" "100030004003"
#> [115] "100030005001" "100030005002" "100030005003" "100030005004" "100030006011" "100030006012"
#> [121] "100030006013" "100030006021" "100030006022" "100030006023" "100030009001" "100030009002"
#> [127] "100030009003" "100030011001" "100030011002" "100030011003" "100030012001" "100030012002"
#> [133] "100030013001" "100030013002" "100030013003" "100030013004" "100030014001" "100030014002"
#> [139] "100030015001" "100030015002" "100030016001" "100030016002" "100030016003" "100030019021"
#> [145] "100030019022" "100030021001" "100030021002" "100030022001" "100030022002" "100030022003"
#> [151] "100030023001" "100030023002" "100030023003" "100030024001" "100030024002" "100030024003"
#> [157] "100030024004" "100030025001" "100030025002" "100030025003" "100030026001" "100030026002"
#> [163] "100030026003" "100030026004" "100030027001" "100030027002" "100030028001" "100030028002"
#> [169] "100030029001" "100030029002" "100030029003" "100030029004" "100030030021" "100030030022"
#> [175] "100030101041" "100030101042" "100030101051" "100030101052" "100030101061" "100030101062"
#> [181] "100030101063" "100030102001" "100030102002" "100030103001" "100030103002" "100030103003"
#> [187] "100030104001" "100030104002" "100030104003" "100030104004" "100030105021" "100030105022"
#> [193] "100030105023" "100030105024" "100030105025" "100030107031" "100030107032" "100030107041"
#> [199] "100030107042" "100030107043" "100030108001" "100030108002" "100030108003" "100030109001"
#> [205] "100030109002" "100030109003" "100030110001" "100030110002" "100030110003" "100030111001"
#> [211] "100030112011" "100030112012" "100030112021" "100030112022" "100030112031" "100030112032"
#> [217] "100030112033" "100030112041" "100030112042" "100030112051" "100030112052" "100030112061"
#> [223] "100030112062" "100030113001" "100030113002" "100030114001" "100030114002" "100030114003"
#> [229] "100030115001" "100030115002" "100030115003" "100030116001" "100030116002" "100030117001"
#> [235] "100030117002" "100030117003" "100030117004" "100030118001" "100030118002" "100030118003"
#> [241] "100030118004" "100030119001" "100030119002" "100030119003" "100030120001" "100030120002"
#> [247] "100030120003" "100030121001" "100030121002" "100030121003" "100030122001" "100030122002"
#> [253] "100030122003" "100030122004" "100030123001" "100030123002" "100030124001" "100030124002"
#> [259] "100030124003" "100030124004" "100030125001" "100030125002" "100030125003" "100030125004"
#> [265] "100030125005" "100030125006" "100030126001" "100030126002" "100030126003" "100030127001"
#> [271] "100030127002" "100030127003" "100030127004" "100030127005" "100030129001" "100030129002"
#> [277] "100030129003" "100030130001" "100030130002" "100030131001" "100030131002" "100030131003"
#> [283] "100030132001" "100030132002" "100030132003" "100030133001" "100030133002" "100030134001"
#> [289] "100030134002" "100030134003" "100030135011" "100030135012" "100030135013" "100030135014"
#> [295] "100030135015" "100030135051" "100030135052" "100030135061" "100030135062" "100030135063"
#> [301] "100030135071" "100030135081" "100030135082" "100030135083" "100030136041" "100030136042"
#> [307] "100030136043" "100030136071" "100030136072" "100030136073" "100030136074" "100030136081"
#> [313] "100030136082" "100030136101" "100030136102" "100030136103" "100030136104" "100030136111"
#> [319] "100030136112" "100030136113" "100030136114" "100030136121" "100030136122" "100030136123"
#> [325] "100030136131" "100030136132" "100030136133" "100030136141" "100030136142" "100030136151"
#> [331] "100030136152" "100030136153" "100030137001" "100030137002" "100030137003" "100030138001"
#> [337] "100030138002" "100030138003" "100030139011" "100030139012" "100030139013" "100030139031"
#> [343] "100030139032" "100030139033" "100030139051" "100030139052" "100030139053" "100030139061"
#> [349] "100030139062" "100030139063" "100030140001" "100030140002" "100030140003" "100030140004"
#> [355] "100030141001" "100030141002" "100030141003" "100030142001" "100030143011" "100030143021"
#> [361] "100030143022" "100030143023" "100030144021" "100030144022" "100030144031" "100030144032"
#> [367] "100030144033" "100030144041" "100030144042" "100030145011" "100030145021" "100030145022"
#> [373] "100030147021" "100030147031" "100030147032" "100030147033" "100030147034" "100030147051"
#> [379] "100030147052" "100030147053" "100030147061" "100030147062" "100030148031" "100030148032"
#> [385] "100030148033" "100030148081" "100030148082" "100030148083" "100030148084" "100030148091"
#> [391] "100030148092" "100030148093" "100030148094" "100030148095" "100030148101" "100030148102"
#> [397] "100030148103" "100030148111" "100030148112" "100030148113" "100030148121" "100030148122"
#> [403] "100030148123" "100030148131" "100030148132" "100030148133" "100030148141" "100030148142"
#> [409] "100030149031" "100030149032" "100030149033" "100030149034" "100030149035" "100030149036"
#> [415] "100030149041" "100030149042" "100030149061" "100030149062" "100030149063" "100030149071"
#> [421] "100030149072" "100030149073" "100030149081" "100030149082" "100030149091" "100030149092"
#> [427] "100030149093" "100030149094" "100030150001" "100030150002" "100030150003" "100030150004"
#> [433] "100030150005" "100030151001" "100030151002" "100030152001" "100030152002" "100030152003"
#> [439] "100030152004" "100030152005" "100030154001" "100030154002" "100030155021" "100030155022"
#> [445] "100030156001" "100030156002" "100030158021" "100030158022" "100030159001" "100030159002"
#> [451] "100030159003" "100030160001" "100030160002" "100030161001" "100030161002" "100030162001"
#> [457] "100030162002" "100030162003" "100030163011" "100030163012" "100030163013" "100030163014"
#> [463] "100030163061" "100030163062" "100030163063" "100030163071" "100030163081" "100030163082"
#> [469] "100030163091" "100030163092" "100030164011" "100030164012" "100030164013" "100030164014"
#> [475] "100030164041" "100030164042" "100030164043" "100030166081" "100030166082" "100030166091"
#> [481] "100030166092" "100030166093" "100030166101" "100030166102" "100030166103" "100030166104"
#> [487] "100030166111" "100030166112" "100030166113" "100030166121" "100030166122" "100030166123"
#> [493] "100030166131" "100030166132" "100030166133" "100030166134" "100030166141" "100030166142"
#> [499] "100030166143" "100030168051" "100030168061" "100030168062" "100030168063" "100030168071"
#> [505] "100030168081" "100030168082" "100030168083" "100030168084" "100030168085" "100030169011"
#> [511] "100030169012" "100030169041" "100039801001" "100039901000" "100050501011" "100050501012"
#> [517] "100050501013" "100050501031" "100050501032" "100050501033" "100050501034" "100050501041"
#> [523] "100050501042" "100050501043" "100050501051" "100050501052" "100050501053" "100050501054"
#> [529] "100050502001" "100050502002" "100050502003" "100050503021" "100050503022" "100050503023"
#> [535] "100050503031" "100050503032" "100050503041" "100050503042" "100050503043" "100050504011"
#> [541] "100050504012" "100050504031" "100050504032" "100050504051" "100050504052" "100050504061"
#> [547] "100050504062" "100050504063" "100050504064" "100050504071" "100050504072" "100050504073"
#> [553] "100050504081" "100050504082" "100050504083" "100050505011" "100050505012" "100050505013"
#> [559] "100050505031" "100050505032" "100050505033" "100050505051" "100050505061" "100050505062"
#> [565] "100050506011" "100050506012" "100050506013" "100050506031" "100050506032" "100050506041"
#> [571] "100050506042" "100050506043" "100050507011" "100050507012" "100050507013" "100050507031"
#> [577] "100050507032" "100050507061" "100050507062" "100050507063" "100050507071" "100050507072"
#> [583] "100050507081" "100050507091" "100050507092" "100050507101" "100050507102" "100050507111"
#> [589] "100050507112" "100050508011" "100050508012" "100050508013" "100050508041" "100050508042"
#> [595] "100050508051" "100050508052" "100050508061" "100050508062" "100050508063" "100050508064"
#> [601] "100050508071" "100050508072" "100050508081" "100050508082" "100050508083" "100050509031"
#> [607] "100050509032" "100050509041" "100050509042" "100050509051" "100050509052" "100050509053"
#> [613] "100050509054" "100050510081" "100050510082" "100050510083" "100050510091" "100050510092"
#> [619] "100050510093" "100050510101" "100050510102" "100050510103" "100050510111" "100050510112"
#> [625] "100050510113" "100050510121" "100050510122" "100050510131" "100050510132" "100050510141"
#> [631] "100050510142" "100050510151" "100050510152" "100050510161" "100050510162" "100050510171"
#> [637] "100050510172" "100050510173" "100050511011" "100050511012" "100050511013" "100050511021"
#> [643] "100050511022" "100050511023" "100050511031" "100050511032" "100050511033" "100050512011"
#> [649] "100050512012" "100050512021" "100050512022" "100050512031" "100050512032" "100050512033"
#> [655] "100050512034" "100050512041" "100050512042" "100050512043" "100050512044" "100050512051"
#> [661] "100050512052" "100050513021" "100050513022" "100050513023" "100050513071" "100050513081"
#> [667] "100050513082" "100050513083" "100050513091" "100050513092" "100050513101" "100050513111"
#> [673] "100050513112" "100050513121" "100050513122" "100050513131" "100050513132" "100050513133"
#> [679] "100050513141" "100050513142" "100050514001" "100050514002" "100050514003" "100050515011"
#> [685] "100050515012" "100050515021" "100050515022" "100050515023" "100050517011" "100050517012"
#> [691] "100050517013" "100050517021" "100050517022" "100050517023" "100050518011" "100050518012"
#> [697] "100050518013" "100050518014" "100050518021" "100050518022" "100050518023" "100050519001"
#> [703] "100050519002" "100050519003" "100059800001" "100059900000"
  
  blockgroupstats[,.N,by=substr(bgfips,1,2)]
#>     substr     N
#>  1:     60    77
#>  2:     66    58
#>  3:     69   135
#>  4:     78   416
#>  5:     01  3925
#>  6:     02   504
#>  7:     04  4773
#>  8:     05  2294
#>  9:     06 25607
#> 10:     08  4058
#> 11:     09  2716
#> 12:     10   706
#> 13:     11   571
#> 14:     12 13388
#> 15:     13  7446
#> 16:     15  1083
#> 17:     16  1284
#> 18:     17  9898
#> 19:     18  5290
#> 20:     19  2703
#> 21:     20  2461
#> 22:     21  3581
#> 23:     22  4294
#> 24:     23  1184
#> 25:     24  4079
#> 26:     25  5116
#> 27:     26  8386
#> 28:     27  4706
#> 29:     28  2445
#> 30:     29  5031
#> 31:     30   900
#> 32:     31  1648
#> 33:     32  1963
#> 34:     33   997
#> 35:     34  6599
#> 36:     35  1614
#> 37:     36 16070
#> 38:     37  7111
#> 39:     38   632
#> 40:     39  9472
#> 41:     40  3374
#> 42:     41  2970
#> 43:     42 10173
#> 44:     44   792
#> 45:     45  3408
#> 46:     46   694
#> 47:     47  4562
#> 48:     48 18638
#> 49:     49  2020
#> 50:     50   552
#> 51:     51  5963
#> 52:     53  5311
#> 53:     54  1639
#> 54:     55  4692
#> 55:     56   457
#> 56:     72  2555
#>     substr     N
  length(fips_bg_from_anyfips("72"))
#> [1] 2555
  
  # all blockgroups in this one county
  fips_bg_from_anyfips(30001)
#>  [1] "300010001001" "300010001002" "300010002001" "300010002002" "300010002003" "300010002004" "300010003001"
#>  [8] "300010003002" "300010003003" "300010003004"
  fips_bg_from_anyfips("30001")
#>  [1] "300010001001" "300010001002" "300010002001" "300010002002" "300010002003" "300010002004" "300010003001"
#>  [8] "300010003002" "300010003003" "300010003004"
  fips_bg_from_anyfips(fips_counties_from_statename("Rhode Island")[1])
#>  [1] "440010301001" "440010301002" "440010301003" "440010301004" "440010302001" "440010302002" "440010302003"
#>  [8] "440010303001" "440010303002" "440010303003" "440010304001" "440010304002" "440010304003" "440010305001"
#> [15] "440010305002" "440010305003" "440010306011" "440010306012" "440010306013" "440010306021" "440010306022"
#> [22] "440010306023" "440010306024" "440010307001" "440010307002" "440010307003" "440010307004" "440010308001"
#> [29] "440010308002" "440010308003" "440010309011" "440010309012" "440010309013" "440010309021" "440010309022"
#> [36] "440010309023" "440010309024"
  
  # all blockgroups that contain any of these 6 blocks (i.e., just one bg)
  fips_bg_from_anyfips( blockid2fips$blockfips[1:6])
#> Error in fips_lead_zero(fips): object 'blockid2fips' not found
  
  # 2 counties
  fips_bg_from_anyfips(c(36009,36011))
#>   [1] "360099400001" "360099400002" "360099402001" "360099403011" "360099403012" "360099403013"
#>   [7] "360099403021" "360099403022" "360099403023" "360099601001" "360099601002" "360099601003"
#>  [13] "360099602001" "360099602002" "360099602003" "360099602004" "360099603001" "360099603002"
#>  [19] "360099603003" "360099603004" "360099604001" "360099604002" "360099604003" "360099604004"
#>  [25] "360099605001" "360099605002" "360099605003" "360099605004" "360099605005" "360099606001"
#>  [31] "360099606002" "360099606003" "360099606004" "360099607031" "360099607032" "360099607041"
#>  [37] "360099607042" "360099608001" "360099608002" "360099608003" "360099608004" "360099610001"
#>  [43] "360099610002" "360099610003" "360099610004" "360099611001" "360099611002" "360099611003"
#>  [49] "360099611004" "360099612001" "360099612002" "360099612003" "360099612004" "360099613011"
#>  [55] "360099613012" "360099613021" "360099613022" "360099614001" "360099614002" "360099614003"
#>  [61] "360099614004" "360099615001" "360099615002" "360099616001" "360099616002" "360099616003"
#>  [67] "360099616004" "360099617001" "360099617002" "360099617003" "360099618001" "360099618002"
#>  [73] "360099618003" "360099618004" "360099618005" "360099618006" "360099622001" "360099622002"
#>  [79] "360099622003" "360099622004" "360110401011" "360110401012" "360110401021" "360110401022"
#>  [85] "360110401023" "360110402011" "360110402012" "360110402013" "360110402021" "360110402022"
#>  [91] "360110403001" "360110403002" "360110403003" "360110403004" "360110404001" "360110404002"
#>  [97] "360110404003" "360110405001" "360110405002" "360110405003" "360110406001" "360110406002"
#> [103] "360110406003" "360110407001" "360110407002" "360110407003" "360110407004" "360110408001"
#> [109] "360110408002" "360110408003" "360110408004" "360110409001" "360110409002" "360110409003"
#> [115] "360110409004" "360110410011" "360110410012" "360110410013" "360110410021" "360110410022"
#> [121] "360110411011" "360110411012" "360110411013" "360110411021" "360110411022" "360110412011"
#> [127] "360110412012" "360110412021" "360110412022" "360110412023" "360110413001" "360110413002"
#> [133] "360110413003" "360110413004" "360110413005" "360110414001" "360110414002" "360110414003"
#> [139] "360110414004" "360110415001" "360110415002" "360110415003" "360110415004" "360110416001"
#> [145] "360110416002" "360110416003" "360110417001" "360110417002" "360110417003" "360110417004"
#> [151] "360110417005" "360110418001" "360110418002" "360110418003" "360110418004" "360110418005"
#> [157] "360110421001" "360110421002" "360110421003" "360119902000"