frs_by_mact (DATA) MACT NESHAP subpart(s) that each EPA-regulated site is subject to

This is a data.table with one row per site – MACT subpart pair, so it has multiple rows for one site if the site is covered by multiple subparts. It has been joined with frs_by_programid to get latlons for matching facilities. @details

This file is not stored in the package, but is obtained via dataload_from_pins().

There are about 115k rows here but only about 86k unique program IDs in this table, which is from the ECHO data download of ICIS Air and AFS. The programid column here should be found in the pgm_sys_id column in frs_by_programid, but as of mid 2024 only 55k of them were found there.

 table(frs_by_mact$programid %in% frs_by_programid$pgm_sys_id)
FALSE  TRUE
60425 55467

Also note we have found some typos in downloaded data from ECHO/FRS, such as

"WOOD PERSERVING AREA SOURCES" instead of "WOOD PRESERVING AREA SOURCES"

See also

Examples

 
  mact_table
#>     subpart
#> 1         A
#> 2        AA
#> 3      AAAA
#> 4     AAAAA
#> 5   AAAAAAA
#> 6         B
#> 7        BB
#> 8     BBBBB
#> 9    BBBBBB
#> 10  BBBBBBB
#> 11        C
#> 12       CC
#> 13      CCC
#> 14     CCCC
#> 15    CCCCC
#> 16   CCCCCC
#> 17  CCCCCCC
#> 18        D
#> 19       DD
#> 20      DDD
#> 21     DDDD
#> 22    DDDDD
#> 23   DDDDDD
#> 24  DDDDDDD
#> 25        E
#> 26       EE
#> 27      EEE
#> 28     EEEE
#> 29    EEEEE
#> 30  EEEEEEE
#> 31        F
#> 32     FFFF
#> 33    FFFFF
#> 34   FFFFFF
#> 35        G
#> 36       GG
#> 37      GGG
#> 38     GGGG
#> 39    GGGGG
#> 40        H
#> 41       HH
#> 42      HHH
#> 43     HHHH
#> 44    HHHHH
#> 45   HHHHHH
#> 46  HHHHHHH
#> 47        I
#> 48       II
#> 49      III
#> 50     IIII
#> 51    IIIII
#> 52        J
#> 53       JJ
#> 54      JJJ
#> 55     JJJJ
#> 56    JJJJJ
#> 57   JJJJJJ
#> 58       KK
#> 59     KKKK
#> 60    KKKKK
#> 61        L
#> 62       LL
#> 63      LLL
#> 64    LLLLL
#> 65   LLLLLL
#> 66        M
#> 67       MM
#> 68      MMM
#> 69     MMMM
#> 70    MMMMM
#> 71   MMMMMM
#> 72        N
#> 73      NNN
#> 74     NNNN
#> 75    NNNNN
#> 76   NNNNNN
#> 77        O
#> 78       OO
#> 79      OOO
#> 80     OOOO
#> 81   OOOOOO
#> 82       PP
#> 83      PPP
#> 84     PPPP
#> 85    PPPPP
#> 86   PPPPPP
#> 87        Q
#> 88       QQ
#> 89     QQQQ
#> 90    QQQQQ
#> 91   QQQQQQ
#> 92        R
#> 93       RR
#> 94      RRR
#> 95     RRRR
#> 96    RRRRR
#> 97   RRRRRR
#> 98        S
#> 99       SS
#> 100    SSSS
#> 101   SSSSS
#> 102  SSSSSS
#> 103       T
#> 104      TT
#> 105     TTT
#> 106    TTTT
#> 107   TTTTT
#> 108  TTTTTT
#> 109       U
#> 110      UU
#> 111     UUU
#> 112    UUUU
#> 113   UUUUU
#> 114      VV
#> 115     VVV
#> 116    VVVV
#> 117  VVVVVV
#> 118       W
#> 119      WW
#> 120    WWWW
#> 121   WWWWW
#> 122  WWWWWW
#> 123       X
#> 124      XX
#> 125     XXX
#> 126    XXXX
#> 127  XXXXXX
#> 128       Y
#> 129      YY
#> 130    YYYY
#> 131   YYYYY
#> 132  YYYYYY
#> 133     ZZZ
#> 134    ZZZZ
#> 135   ZZZZZ
#> 136  ZZZZZZ
#>                                                                            title
#> 1                                                             GENERAL PROVISIONS
#> 2                                            PHOSPERIC ACID MANUFACTURING PLANTS
#> 3                                                MUNICIPAL SOLID WASTE LANDFILLS
#> 4                                                      LIME MANUFACTURING PLANTS
#> 5              AREA SOURCE: ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 6                    REQ FOR CONTROL TECHNOLOGY DETERMINATIONS FOR MAJOR SOURCES
#> 7                                                          PHOSPHATE FERTILIZERS
#> 8                                                    SEMICONDUCTOR MANUFACTURING
#> 9     GASOLINE DISTRIBUTION BULK TERMINALS, BULK PLANTS, AND PIPELINE FACILITIES
#> 10                                   AREA SOURCE: CHEMICAL PREPARATIONS INDUSTRY
#> 11                                                            DELISTINGS OF HAPS
#> 12                                                          PETROLEUM REFINERIES
#> 13                      STEEL PICKLING-HCI PROCESS FACILITIES AND HCI REG PLANTS
#> 14                                            MANUFACTURING OF NUTRITIONAL YEAST
#> 15                             COKE OVENS: PUSHING, QUENCHING AND BATTERY STACKS
#> 16                                                GASOLINE DISPENSING FACILITIES
#> 17                          AREA SOURCE: PAINT AND ALLIED PRODUCTS MANUFACTURING
#> 18                                       COMPLIANCE EXTENSION, EARLY HAP REDUCTN
#> 19                                        OFF-SITE WASTE AND RECOVERY OPERATIONS
#> 20                                                       MINERAL WOOL PRODUCTION
#> 21                                           PLYWOOD AND COMPOSITE WOOD PRODUCTS
#> 22   MAJOR SOURCES: INDUSTRIAL/COMMERCIAL/INSTITUTIONAL BOILERS & PROCESS HEATER
#> 23                     POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION AREA SOURCES
#> 24                                               AREA SOURCE: PREPARED FEEDS MFG
#> 25                         APPROVAL OF STATE PROGRAMS &  DELEGATION OF AUTHORITY
#> 26                                                 MAGNETIC TAPE MFG. OPERATIONS
#> 27                                                    HAZARDOUS WASTE COMBUSTORS
#> 28                                   ORGANIC LIQUIDS DISTRIBUTION (NON-GASOLINE)
#> 29                                                      IRON AND STEEL FOUNDRIES
#> 30                          AREA SOURCE: GOLD MINE ORE PROCESSING AND PRODUCTION
#> 31                                SYNTHETIC HAZARDOUS ORGANICS (SOCMI) STANDARDS
#> 32                            MISCELLANEOUS ORGANIC CHEMICAL MANUFACTURING (MON)
#> 33                            INTEGRATED IRON AND STEEL MANUFACTURING FACILITIES
#> 34                                        SECONDARY COPPER SMELTING AREA SOURCES
#> 35         SOCMI PROCESS VENTS, STORAGE VESSELS, TRANSFER OPERATIONS, WASTEWATER
#> 36                                                       AEROSPACE MFG. & REWORK
#> 37                                                  PHARMACEUTICAL MANUFACTURING
#> 38                               SOLVENT EXTRACTION FOR VEGETABLE OIL PRODUCTION
#> 39                                                              SITE REMEDIATION
#> 40                                         EQUIPMENT LEAKS OF HAZARDOUS ORGANICS
#> 41                                     OIL AND NATURAL GAS PRODUCTION FACILITIES
#> 42                                 NATURAL GAS TRANSMISSION & STORAGE FACILITIES
#> 43                                          WET FORMED FIBERGLASS MAT PRODUCTION
#> 44                                           MISCELLANEOUS COATING MANUFACTURING
#> 45                 PAINT STRIP & MISC SURFACE COATING OPERATIONS AT AREA SOURCES
#> 46                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 47                                      NEGOTIATED REGS: EQUIP LEAKS, HAZ-ORGNCS
#> 48                                                    SHIPBUILDING & SHIP REPAIR
#> 49                                         FLEXIBLE POLYURETHANE FOAM PRODUCTION
#> 50                                 SURFACE COATING OF AUTO AND LIGHT DUTY TRUCKS
#> 51                       MERCURY EMISSIONS FROM MERCURY CELL CHLOR-ALKALI PLANTS
#> 52                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 53                                       WOOD FURNITURE MANUFACTURING OPERATIONS
#> 54                                                  GROUP IV POLYMERS AND RESINS
#> 55                                                    PAPER & OTHER WEB COATINGS
#> 56                                BRICK & STRUCTURAL CLAY PRODUCTS MANUFACTURING
#> 57                INDUSTRIAL, COMMERCIAL, AND INSTITUTIONAL BOILERS AREA SOURCES
#> 58                                                         PRINTING & PUBLISHING
#> 59                                                 SURFACE COATING OF METAL CANS
#> 60                                                   CLAY CERAMICS MANUFACTURING
#> 61                                                           COKE OVEN BATTERIES
#> 62                                             PRIMARY ALUMINUM REDUCTION PLANTS
#> 63                                                        PORTLAND CEMENT PLANTS
#> 64                          ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 65                         ACRYLIC AND MODACRYLIC FIBERS PRODUCTION AREA SOURCES
#> 66                                                DRY CLEANERS PERCHLOROETHYLENE
#> 67                            CHEMICAL RECOVERY COMBUSTION SOURCES AT PULP MILLS
#> 68                                        PESTICIDE ACTIVE INGREDIENT PRODUCTION
#> 69                             SURFACE COATING OF MISC. METAL PARTS AND PRODUCTS
#> 70                             FLEXIBLE POLYURETHANE FOAM FABRICATION OPERATIONS
#> 71                                          CARBON BLACK PRODUCTION AREA SOURCES
#> 72                                                       CHROMIUM ELECTROPLATING
#> 73                                                 WOOL FIBERGLASS MANUFACTURING
#> 74                                           SURFACE COATING OF LARGE APPLIANCES
#> 75                                                  HYDROCHLORIC ACID PRODUCTION
#> 76                       CHEMICAL MANUFACTURING AREA SOURCES: CHROMIUM COMPOUNDS
#> 77                                                    ETHYLENE OXIDE STERILIZERS
#> 78                                                      TANK STANDARDS - LEVEL 1
#> 79                                          MANUFACTURE OF AMINO/PHENOLIC RESINS
#> 80                    PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 81                            FLEXIBLE POLYURETHANE FOAM PROD & FAB AREA SOURCES
#> 82                                                           CONTAINER STANDARDS
#> 83                                                  POLYETHER POLYOLS PRODUCTION
#> 84                                 SURFACE COATING OF PLASTIC PARTS AND PRODUCTS
#> 85                                                      ENGINE TEST CELLS/STANDS
#> 86                                  LEAD ACID BATTERY MANUFACTURING AREA SOURCES
#> 87                                             INDUSTRIAL PROCESS COOLING TOWERS
#> 88                                                           SURFACE IMPOUNDMENT
#> 89                                     SURFACE COATING OF WOOD BUILDING PRODUCTS
#> 90                                   FRICTION MATERIALS MANUFACTURING FACILITIES
#> 91                                                  WOOD PERSERVING AREA SOURCES
#> 92                                                         GASOLINE DISTRIBUTION
#> 93                                                      INDIVIDUAL DRAIN SYSTEMS
#> 94                                                 SECONDARY ALUMINUM PRODUCTION
#> 95                                            SURFACE COATING OF METAL FURNITURE
#> 96                                                  TACONITE IRON ORE PROCESSING
#> 97                                                CLAY CERAMICS MFG AREA SOURCES
#> 98                                                                PULP AND PAPER
#> 99  CLOSED VENT SYS/CTRL DEVICES, RECOVERY DEVICES, ROUTING TO FUEL GAS SYS/PROC
#> 100                                                SURFACE COATING OF METAL COIL
#> 101                                            REFRACTORY PRODUCTS MANUFACTURING
#> 102                                                       GLASS MFG AREA SOURCES
#> 103                                                 HALOGENATED SOLVENT CLEANING
#> 104                                            EQUIPMENT LEAKS - CONTROL LEVEL 1
#> 105                                                        PRIMARY LEAD SMELTING
#> 106                                                 LEATHER FINISHING OPERATIONS
#> 107                                                   PRIMARY MAGNESIUM REFINING
#> 108                          SECONDARY NONFERROUS METALS PROCESSING AREA SOURCES
#> 109                                                  GROUP 1 POLYMERS AND RESINS
#> 110                                                  EQUIP LEAKS-CONTROL LEVEL 2
#> 111                                   PETROLEUM REFINERIES: CCU'S, CRU'S & SRU'S
#> 112                                              CELLULOSE PRODUCT MANUFACTURING
#> 113                                                     COAL & OIL-FIRED EUSGU'S
#> 114                                         OIL-WATER & ORGANIC-WATER SEPARATORS
#> 115                                               PUBLICLY OWNED TREATMENT WORKS
#> 116                                                           BOAT MANUFACTURING
#> 117                                          CHEMICAL MANUFACTURING AREA SOURCES
#> 118                               EPOXY RESINS & NON-NYLON POLYAMIDES PRODUCTION
#> 119                                     STORAGE VESSELS (TANKS)- CONTROL LEVEL 2
#> 120                                     REINFORCED PLASTIC COMPOSITES PRODUCTION
#> 121                                          HOSPITAL ETHYLENE OXIDE STERILIZERS
#> 122                                AREA SOURCE: PLATING AND POLISHING OPERATIONS
#> 123                                                      SECONDARY LEAD SMELTERS
#> 124                       ETHYLENE MPF: HEAT EXCHANGE SYSTEMS & WASTE OPERATIONS
#> 125                                                        FERROALLOY PRODUCTION
#> 126                                                    RUBBER TIRE MANUFACTURING
#> 127          AREA SOURCE: NINE METAL FABRICATION AND FINISHING SOURCE CATEGORIES
#> 128                            MARINE TANK VESSEL LOADING & UNLOADING OPERATIONS
#> 129                                                       GENERIC MACT STANDARDS
#> 130                                               STATIONARY COMBUSTION TURBINES
#> 131                     AREA SOURCE: ELECTRIC ARC FURNACE STEELMAKING FACILITIES
#> 132                               AREA SOURCE: FERROALLOYS PRODUCTION FACILITIES
#> 133                                                         CHEMICAL SECTOR RULE
#> 134                  STATIONARY RECIPROCATING INTERNAL COMBUSTION ENGINES (RICE)
#> 135                                        IRON AND STEEL FOUNDRIES AREA SOURCES
#> 136                   AREA SOURCE: ALUMINUM, COPPER & OTHER NONFERROUS FOUNDRIES
#>                                                                                  dropdown_label
#> 1                                                                A - General Provisions (7,958)
#> 2                                                 AA - Phosperic Acid Manufacturing Plants (30)
#> 3                                                AAAA - Municipal Solid Waste Landfills (1,212)
#> 4                                                        AAAAA - Lime Manufacturing Plants (90)
#> 5              AAAAAAA - Area Source: Asphalt Processing And Asphalt Roofing Manufacturing (77)
#> 6                         B - Req For Control Technology Determinations For Major Sources (187)
#> 7                                                               BB - Phosphate Fertilizers (30)
#> 8                                                      BBBBB - Semiconductor Manufacturing (67)
#> 9   BBBBBB - Gasoline Distribution Bulk Terminals, Bulk Plants, And Pipeline Facilities (1,930)
#> 10                                    BBBBBBB - Area Source: Chemical Preparations Industry (7)
#> 11                                                                   C - Delistings Of Haps (1)
#> 12                                                              CC - Petroleum Refineries (340)
#> 13                          CCC - Steel Pickling-Hci Process Facilities And Hci Reg Plants (86)
#> 14                                                CCCC - Manufacturing Of Nutritional Yeast (8)
#> 15                               CCCCC - Coke Ovens: Pushing, Quenching And Battery Stacks (55)
#> 16                                              CCCCCC - Gasoline Dispensing Facilities (3,505)
#> 17                          CCCCCCC - Area Source: Paint And Allied Products Manufacturing (80)
#> 18                                             D - Compliance Extension, Early Hap Reductn (10)
#> 19                                            DD - Off-Site Waste And Recovery Operations (202)
#> 20                                                           DDD - Mineral Wool Production (17)
#> 21                                             DDDD - Plywood And Composite Wood Products (743)
#> 22  DDDDD - Major Sources: Industrial/Commercial/Institutional Boilers & Process Heater (4,324)
#> 23                      DDDDDD - Polyvinyl Chloride And Copolymers Production Area Sources (30)
#> 24                                              DDDDDDD - Area Source: Prepared Feeds Mfg (305)
#> 25                                E - Approval Of State Programs &  Delegation Of Authority (2)
#> 26                                                      EE - Magnetic Tape Mfg. Operations (15)
#> 27                                                       EEE - Hazardous Waste Combustors (263)
#> 28                                     EEEE - Organic Liquids Distribution (Non-Gasoline) (571)
#> 29                                                       EEEEE - Iron And Steel Foundries (174)
#> 30                           EEEEEEE - Area Source: Gold Mine Ore Processing And Production (7)
#> 31                                     F - Synthetic Hazardous Organics (Socmi) Standards (479)
#> 32                              FFFF - Miscellaneous Organic Chemical Manufacturing (Mon) (671)
#> 33                              FFFFF - Integrated Iron And Steel Manufacturing Facilities (35)
#> 34                                          FFFFFF - Secondary Copper Smelting Area Sources (2)
#> 35              G - Socmi Process Vents, Storage Vessels, Transfer Operations, Wastewater (539)
#> 36                                                           GG - Aerospace Mfg. & Rework (266)
#> 37                                                     GGG - Pharmaceutical Manufacturing (146)
#> 38                                 GGGG - Solvent Extraction For Vegetable Oil Production (202)
#> 39                                                               GGGGG - Site Remediation (259)
#> 40                                              H - Equipment Leaks Of Hazardous Organics (601)
#> 41                                       HH - Oil And Natural Gas Production Facilities (8,378)
#> 42                                    HHH - Natural Gas Transmission & Storage Facilities (205)
#> 43                                             HHHH - Wet Formed Fiberglass Mat Production (12)
#> 44                                            HHHHH - Miscellaneous Coating Manufacturing (284)
#> 45               HHHHHH - Paint Strip & Misc Surface Coating Operations At Area Sources (6,904)
#> 46                                  HHHHHHH - Polyvinyl Chloride And Copolymers Production (21)
#> 47                                           I - Negotiated Regs: Equip Leaks, Haz-Orgncs (107)
#> 48                                                        II - Shipbuilding & Ship Repair (181)
#> 49                                             III - Flexible Polyurethane Foam Production (61)
#> 50                                   IIII - Surface Coating Of Auto And Light Duty Trucks (183)
#> 51                         IIIII - Mercury Emissions From Mercury Cell Chlor-Alkali Plants (19)
#> 52                                        J - Polyvinyl Chloride And Copolymers Production (20)
#> 53                                         JJ - Wood Furniture Manufacturing Operations (1,047)
#> 54                                                     JJJ - Group Iv Polymers And Resins (138)
#> 55                                                      JJJJ - Paper & Other Web Coatings (543)
#> 56                                 JJJJJ - Brick & Structural Clay Products Manufacturing (191)
#> 57              JJJJJJ - Industrial, Commercial, And Institutional Boilers Area Sources (2,081)
#> 58                                                             KK - Printing & Publishing (711)
#> 59                                                    KKKK - Surface Coating Of Metal Cans (70)
#> 60                                                     KKKKK - Clay Ceramics Manufacturing (14)
#> 61                                                                 L - Coke Oven Batteries (58)
#> 62                                                  LL - Primary Aluminum Reduction Plants (33)
#> 63                                                           LLL - Portland Cement Plants (233)
#> 64                            LLLLL - Asphalt Processing And Asphalt Roofing Manufacturing (33)
#> 65                           LLLLLL - Acrylic And Modacrylic Fibers Production Area Sources (5)
#> 66                                                   M - Dry Cleaners Perchloroethylene (8,791)
#> 67                                MM - Chemical Recovery Combustion Sources At Pulp Mills (215)
#> 68                                            MMM - Pesticide Active Ingredient Production (53)
#> 69                             MMMM - Surface Coating Of Misc. Metal Parts And Products (1,375)
#> 70                               MMMMM - Flexible Polyurethane Foam Fabrication Operations (19)
#> 71                                            MMMMMM - Carbon Black Production Area Sources (9)
#> 72                                                          N - Chromium Electroplating (2,453)
#> 73                                                     NNN - Wool Fiberglass Manufacturing (50)
#> 74                                              NNNN - Surface Coating Of Large Appliances (36)
#> 75                                                    NNNNN - Hydrochloric Acid Production (50)
#> 76                         NNNNNN - Chemical Manufacturing Area Sources: Chromium Compounds (8)
#> 77                                                         O - Ethylene Oxide Sterilizers (225)
#> 78                                                           OO - Tank Standards - Level 1 (65)
#> 79                                              OOO - Manufacture Of Amino/Phenolic Resins (67)
#> 80                      OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles (226)
#> 81                            OOOOOO - Flexible Polyurethane Foam Prod & Fab Area Sources (148)
#> 82                                                                PP - Container Standards (29)
#> 83                                                      PPP - Polyether Polyols Production (54)
#> 84                                   PPPP - Surface Coating Of Plastic Parts And Products (663)
#> 85                                                        PPPPP - Engine Test Cells/Stands (58)
#> 86                                   PPPPPP - Lead Acid Battery Manufacturing Area Sources (60)
#> 87                                                  Q - Industrial Process Cooling Towers (340)
#> 88                                                                 QQ - Surface Impoundment (9)
#> 89                                       QQQQ - Surface Coating Of Wood Building Products (143)
#> 90                                      QQQQQ - Friction Materials Manufacturing Facilities (9)
#> 91                                                   QQQQQQ - Wood Perserving Area Sources (88)
#> 92                                                            R - Gasoline Distribution (1,121)
#> 93                                                           RR - Individual Drain Systems (32)
#> 94                                                    RRR - Secondary Aluminum Production (655)
#> 95                                               RRRR - Surface Coating Of Metal Furniture (57)
#> 96                                                    RRRRR - Taconite Iron Ore Processing (15)
#> 97                                                 RRRRRR - Clay Ceramics Mfg Area Sources (16)
#> 98                                                                     S - Pulp And Paper (249)
#> 99      SS - Closed Vent Sys/Ctrl Devices, Recovery Devices, Routing To Fuel Gas Sys/Proc (150)
#> 100                                                  SSSS - Surface Coating Of Metal Coil (126)
#> 101                                               SSSSS - Refractory Products Manufacturing (8)
#> 102                                                        SSSSSS - Glass Mfg Area Sources (35)
#> 103                                                    T - Halogenated Solvent Cleaning (2,355)
#> 104                                                 TT - Equipment Leaks - Control Level 1 (20)
#> 105                                                             TTT - Primary Lead Smelting (2)
#> 106                                                    TTTT - Leather Finishing Operations (26)
#> 107                                                      TTTTT - Primary Magnesium Refining (9)
#> 108                           TTTTTT - Secondary Nonferrous Metals Processing Area Sources (20)
#> 109                                                        U - Group 1 Polymers And Resins (49)
#> 110                                                      UU - Equip Leaks-Control Level 2 (131)
#> 111                                      UUU - Petroleum Refineries: Ccu's, Cru's & Sru's (217)
#> 112                                                 UUUU - Cellulose Product Manufacturing (26)
#> 113                                                      UUUUU - Coal & Oil-Fired Eusgu's (299)
#> 114                                              VV - Oil-Water & Organic-Water Separators (63)
#> 115                                                   VVV - Publicly Owned Treatment Works (27)
#> 116                                                             VVVV - Boat Manufacturing (440)
#> 117                                         VVVVVV -  Chemical Manufacturing Area Sources (185)
#> 118                                     W - Epoxy Resins & Non-Nylon Polyamides Production (22)
#> 119                                          WW - Storage Vessels (Tanks)- Control Level 2 (65)
#> 120                                     WWWW - Reinforced Plastic Composites Production (1,255)
#> 121                                           WWWWW - Hospital Ethylene Oxide Sterilizers (250)
#> 122                                WWWWWW - Area Source: Plating And Polishing Operations (524)
#> 123                                                            X - Secondary Lead Smelters (41)
#> 124                            XX - Ethylene Mpf: Heat Exchange Systems & Waste Operations (33)
#> 125                                                             XXX - Ferroalloy Production (7)
#> 126                                                       XXXX - Rubber Tire Manufacturing (68)
#> 127          XXXXXX - Area Source: Nine Metal Fabrication And Finishing Source Categories (582)
#> 128                                 Y - Marine Tank Vessel Loading & Unloading Operations (308)
#> 129                                                           YY - Generic Mact Standards (152)
#> 130                                                 YYYY - Stationary Combustion Turbines (443)
#> 131                      YYYYY - Area Source: Electric Arc Furnace Steelmaking Facilities (119)
#> 132                                YYYYYY - Area Source: Ferroalloys Production Facilities (12)
#> 133                                                              ZZZ - Chemical Sector Rule (3)
#> 134                 ZZZZ - Stationary Reciprocating Internal Combustion Engines (Rice) (42,984)
#> 135                                         ZZZZZ - Iron And Steel Foundries Area Sources (325)
#> 136                    ZZZZZZ - Area Source: Aluminum, Copper & Other Nonferrous Foundries (62)
#>         N
#> 1    7958
#> 2      30
#> 3    1212
#> 4      90
#> 5      77
#> 6     187
#> 7      30
#> 8      67
#> 9    1930
#> 10      7
#> 11      1
#> 12    340
#> 13     86
#> 14      8
#> 15     55
#> 16   3505
#> 17     80
#> 18     10
#> 19    202
#> 20     17
#> 21    743
#> 22   4324
#> 23     30
#> 24    305
#> 25      2
#> 26     15
#> 27    263
#> 28    571
#> 29    174
#> 30      7
#> 31    479
#> 32    671
#> 33     35
#> 34      2
#> 35    539
#> 36    266
#> 37    146
#> 38    202
#> 39    259
#> 40    601
#> 41   8378
#> 42    205
#> 43     12
#> 44    284
#> 45   6904
#> 46     21
#> 47    107
#> 48    181
#> 49     61
#> 50    183
#> 51     19
#> 52     20
#> 53   1047
#> 54    138
#> 55    543
#> 56    191
#> 57   2081
#> 58    711
#> 59     70
#> 60     14
#> 61     58
#> 62     33
#> 63    233
#> 64     33
#> 65      5
#> 66   8791
#> 67    215
#> 68     53
#> 69   1375
#> 70     19
#> 71      9
#> 72   2453
#> 73     50
#> 74     36
#> 75     50
#> 76      8
#> 77    225
#> 78     65
#> 79     67
#> 80    226
#> 81    148
#> 82     29
#> 83     54
#> 84    663
#> 85     58
#> 86     60
#> 87    340
#> 88      9
#> 89    143
#> 90      9
#> 91     88
#> 92   1121
#> 93     32
#> 94    655
#> 95     57
#> 96     15
#> 97     16
#> 98    249
#> 99    150
#> 100   126
#> 101     8
#> 102    35
#> 103  2355
#> 104    20
#> 105     2
#> 106    26
#> 107     9
#> 108    20
#> 109    49
#> 110   131
#> 111   217
#> 112    26
#> 113   299
#> 114    63
#> 115    27
#> 116   440
#> 117   185
#> 118    22
#> 119    65
#> 120  1255
#> 121   250
#> 122   524
#> 123    41
#> 124    33
#> 125     7
#> 126    68
#> 127   582
#> 128   308
#> 129   152
#> 130   443
#> 131   119
#> 132    12
#> 133     3
#> 134 42984
#> 135   325
#> 136    62
  mact_table[order(mact_table$title),]
#>     subpart
#> 117  VVVVVV
#> 65   LLLLLL
#> 36       GG
#> 25        E
#> 136  ZZZZZZ
#> 5   AAAAAAA
#> 10  BBBBBBB
#> 131   YYYYY
#> 132  YYYYYY
#> 30  EEEEEEE
#> 127  XXXXXX
#> 17  CCCCCCC
#> 122  WWWWWW
#> 24  DDDDDDD
#> 64    LLLLL
#> 116    VVVV
#> 56    JJJJJ
#> 71   MMMMMM
#> 112    UUUU
#> 76   NNNNNN
#> 67       MM
#> 133     ZZZ
#> 72        N
#> 60    KKKKK
#> 97   RRRRRR
#> 99       SS
#> 113   UUUUU
#> 61        L
#> 15    CCCCC
#> 18        D
#> 82       PP
#> 11        C
#> 66        M
#> 85    PPPPP
#> 118       W
#> 110      UU
#> 104      TT
#> 40        H
#> 124      XX
#> 77        O
#> 125     XXX
#> 70    MMMMM
#> 81   OOOOOO
#> 49      III
#> 90    QQQQQ
#> 16   CCCCCC
#> 92        R
#> 9    BBBBBB
#> 1         A
#> 129      YY
#> 102  SSSSSS
#> 109       U
#> 54      JJJ
#> 103       T
#> 27      EEE
#> 121   WWWWW
#> 75    NNNNN
#> 93       RR
#> 87        Q
#> 57   JJJJJJ
#> 33    FFFFF
#> 29    EEEEE
#> 135   ZZZZZ
#> 86   PPPPPP
#> 106    TTTT
#> 4     AAAAA
#> 26       EE
#> 22    DDDDD
#> 79      OOO
#> 14     CCCC
#> 128       Y
#> 51    IIIII
#> 20      DDD
#> 44    HHHHH
#> 32     FFFF
#> 3      AAAA
#> 42      HHH
#> 47        I
#> 19       DD
#> 41       HH
#> 114      VV
#> 28     EEEE
#> 45   HHHHHH
#> 55     JJJJ
#> 68      MMM
#> 12       CC
#> 111     UUU
#> 37      GGG
#> 2        AA
#> 7        BB
#> 21     DDDD
#> 83      PPP
#> 46  HHHHHHH
#> 52        J
#> 23   DDDDDD
#> 63      LLL
#> 62       LL
#> 105     TTT
#> 107   TTTTT
#> 58       KK
#> 80     OOOO
#> 115     VVV
#> 98        S
#> 101   SSSSS
#> 120    WWWW
#> 6         B
#> 126    XXXX
#> 94      RRR
#> 34   FFFFFF
#> 123       X
#> 108  TTTTTT
#> 8     BBBBB
#> 48       II
#> 39    GGGGG
#> 35        G
#> 38     GGGG
#> 130    YYYY
#> 134    ZZZZ
#> 13      CCC
#> 119      WW
#> 50     IIII
#> 74     NNNN
#> 59     KKKK
#> 100    SSSS
#> 95     RRRR
#> 69     MMMM
#> 84     PPPP
#> 89     QQQQ
#> 88       QQ
#> 31        F
#> 96    RRRRR
#> 78       OO
#> 43     HHHH
#> 53       JJ
#> 91   QQQQQQ
#> 73      NNN
#>                                                                            title
#> 117                                          CHEMICAL MANUFACTURING AREA SOURCES
#> 65                         ACRYLIC AND MODACRYLIC FIBERS PRODUCTION AREA SOURCES
#> 36                                                       AEROSPACE MFG. & REWORK
#> 25                         APPROVAL OF STATE PROGRAMS &  DELEGATION OF AUTHORITY
#> 136                   AREA SOURCE: ALUMINUM, COPPER & OTHER NONFERROUS FOUNDRIES
#> 5              AREA SOURCE: ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 10                                   AREA SOURCE: CHEMICAL PREPARATIONS INDUSTRY
#> 131                     AREA SOURCE: ELECTRIC ARC FURNACE STEELMAKING FACILITIES
#> 132                               AREA SOURCE: FERROALLOYS PRODUCTION FACILITIES
#> 30                          AREA SOURCE: GOLD MINE ORE PROCESSING AND PRODUCTION
#> 127          AREA SOURCE: NINE METAL FABRICATION AND FINISHING SOURCE CATEGORIES
#> 17                          AREA SOURCE: PAINT AND ALLIED PRODUCTS MANUFACTURING
#> 122                                AREA SOURCE: PLATING AND POLISHING OPERATIONS
#> 24                                               AREA SOURCE: PREPARED FEEDS MFG
#> 64                          ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 116                                                           BOAT MANUFACTURING
#> 56                                BRICK & STRUCTURAL CLAY PRODUCTS MANUFACTURING
#> 71                                          CARBON BLACK PRODUCTION AREA SOURCES
#> 112                                              CELLULOSE PRODUCT MANUFACTURING
#> 76                       CHEMICAL MANUFACTURING AREA SOURCES: CHROMIUM COMPOUNDS
#> 67                            CHEMICAL RECOVERY COMBUSTION SOURCES AT PULP MILLS
#> 133                                                         CHEMICAL SECTOR RULE
#> 72                                                       CHROMIUM ELECTROPLATING
#> 60                                                   CLAY CERAMICS MANUFACTURING
#> 97                                                CLAY CERAMICS MFG AREA SOURCES
#> 99  CLOSED VENT SYS/CTRL DEVICES, RECOVERY DEVICES, ROUTING TO FUEL GAS SYS/PROC
#> 113                                                     COAL & OIL-FIRED EUSGU'S
#> 61                                                           COKE OVEN BATTERIES
#> 15                             COKE OVENS: PUSHING, QUENCHING AND BATTERY STACKS
#> 18                                       COMPLIANCE EXTENSION, EARLY HAP REDUCTN
#> 82                                                           CONTAINER STANDARDS
#> 11                                                            DELISTINGS OF HAPS
#> 66                                                DRY CLEANERS PERCHLOROETHYLENE
#> 85                                                      ENGINE TEST CELLS/STANDS
#> 118                               EPOXY RESINS & NON-NYLON POLYAMIDES PRODUCTION
#> 110                                                  EQUIP LEAKS-CONTROL LEVEL 2
#> 104                                            EQUIPMENT LEAKS - CONTROL LEVEL 1
#> 40                                         EQUIPMENT LEAKS OF HAZARDOUS ORGANICS
#> 124                       ETHYLENE MPF: HEAT EXCHANGE SYSTEMS & WASTE OPERATIONS
#> 77                                                    ETHYLENE OXIDE STERILIZERS
#> 125                                                        FERROALLOY PRODUCTION
#> 70                             FLEXIBLE POLYURETHANE FOAM FABRICATION OPERATIONS
#> 81                            FLEXIBLE POLYURETHANE FOAM PROD & FAB AREA SOURCES
#> 49                                         FLEXIBLE POLYURETHANE FOAM PRODUCTION
#> 90                                   FRICTION MATERIALS MANUFACTURING FACILITIES
#> 16                                                GASOLINE DISPENSING FACILITIES
#> 92                                                         GASOLINE DISTRIBUTION
#> 9     GASOLINE DISTRIBUTION BULK TERMINALS, BULK PLANTS, AND PIPELINE FACILITIES
#> 1                                                             GENERAL PROVISIONS
#> 129                                                       GENERIC MACT STANDARDS
#> 102                                                       GLASS MFG AREA SOURCES
#> 109                                                  GROUP 1 POLYMERS AND RESINS
#> 54                                                  GROUP IV POLYMERS AND RESINS
#> 103                                                 HALOGENATED SOLVENT CLEANING
#> 27                                                    HAZARDOUS WASTE COMBUSTORS
#> 121                                          HOSPITAL ETHYLENE OXIDE STERILIZERS
#> 75                                                  HYDROCHLORIC ACID PRODUCTION
#> 93                                                      INDIVIDUAL DRAIN SYSTEMS
#> 87                                             INDUSTRIAL PROCESS COOLING TOWERS
#> 57                INDUSTRIAL, COMMERCIAL, AND INSTITUTIONAL BOILERS AREA SOURCES
#> 33                            INTEGRATED IRON AND STEEL MANUFACTURING FACILITIES
#> 29                                                      IRON AND STEEL FOUNDRIES
#> 135                                        IRON AND STEEL FOUNDRIES AREA SOURCES
#> 86                                  LEAD ACID BATTERY MANUFACTURING AREA SOURCES
#> 106                                                 LEATHER FINISHING OPERATIONS
#> 4                                                      LIME MANUFACTURING PLANTS
#> 26                                                 MAGNETIC TAPE MFG. OPERATIONS
#> 22   MAJOR SOURCES: INDUSTRIAL/COMMERCIAL/INSTITUTIONAL BOILERS & PROCESS HEATER
#> 79                                          MANUFACTURE OF AMINO/PHENOLIC RESINS
#> 14                                            MANUFACTURING OF NUTRITIONAL YEAST
#> 128                            MARINE TANK VESSEL LOADING & UNLOADING OPERATIONS
#> 51                       MERCURY EMISSIONS FROM MERCURY CELL CHLOR-ALKALI PLANTS
#> 20                                                       MINERAL WOOL PRODUCTION
#> 44                                           MISCELLANEOUS COATING MANUFACTURING
#> 32                            MISCELLANEOUS ORGANIC CHEMICAL MANUFACTURING (MON)
#> 3                                                MUNICIPAL SOLID WASTE LANDFILLS
#> 42                                 NATURAL GAS TRANSMISSION & STORAGE FACILITIES
#> 47                                      NEGOTIATED REGS: EQUIP LEAKS, HAZ-ORGNCS
#> 19                                        OFF-SITE WASTE AND RECOVERY OPERATIONS
#> 41                                     OIL AND NATURAL GAS PRODUCTION FACILITIES
#> 114                                         OIL-WATER & ORGANIC-WATER SEPARATORS
#> 28                                   ORGANIC LIQUIDS DISTRIBUTION (NON-GASOLINE)
#> 45                 PAINT STRIP & MISC SURFACE COATING OPERATIONS AT AREA SOURCES
#> 55                                                    PAPER & OTHER WEB COATINGS
#> 68                                        PESTICIDE ACTIVE INGREDIENT PRODUCTION
#> 12                                                          PETROLEUM REFINERIES
#> 111                                   PETROLEUM REFINERIES: CCU'S, CRU'S & SRU'S
#> 37                                                  PHARMACEUTICAL MANUFACTURING
#> 2                                            PHOSPERIC ACID MANUFACTURING PLANTS
#> 7                                                          PHOSPHATE FERTILIZERS
#> 21                                           PLYWOOD AND COMPOSITE WOOD PRODUCTS
#> 83                                                  POLYETHER POLYOLS PRODUCTION
#> 46                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 52                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 23                     POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION AREA SOURCES
#> 63                                                        PORTLAND CEMENT PLANTS
#> 62                                             PRIMARY ALUMINUM REDUCTION PLANTS
#> 105                                                        PRIMARY LEAD SMELTING
#> 107                                                   PRIMARY MAGNESIUM REFINING
#> 58                                                         PRINTING & PUBLISHING
#> 80                    PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 115                                               PUBLICLY OWNED TREATMENT WORKS
#> 98                                                                PULP AND PAPER
#> 101                                            REFRACTORY PRODUCTS MANUFACTURING
#> 120                                     REINFORCED PLASTIC COMPOSITES PRODUCTION
#> 6                    REQ FOR CONTROL TECHNOLOGY DETERMINATIONS FOR MAJOR SOURCES
#> 126                                                    RUBBER TIRE MANUFACTURING
#> 94                                                 SECONDARY ALUMINUM PRODUCTION
#> 34                                        SECONDARY COPPER SMELTING AREA SOURCES
#> 123                                                      SECONDARY LEAD SMELTERS
#> 108                          SECONDARY NONFERROUS METALS PROCESSING AREA SOURCES
#> 8                                                    SEMICONDUCTOR MANUFACTURING
#> 48                                                    SHIPBUILDING & SHIP REPAIR
#> 39                                                              SITE REMEDIATION
#> 35         SOCMI PROCESS VENTS, STORAGE VESSELS, TRANSFER OPERATIONS, WASTEWATER
#> 38                               SOLVENT EXTRACTION FOR VEGETABLE OIL PRODUCTION
#> 130                                               STATIONARY COMBUSTION TURBINES
#> 134                  STATIONARY RECIPROCATING INTERNAL COMBUSTION ENGINES (RICE)
#> 13                      STEEL PICKLING-HCI PROCESS FACILITIES AND HCI REG PLANTS
#> 119                                     STORAGE VESSELS (TANKS)- CONTROL LEVEL 2
#> 50                                 SURFACE COATING OF AUTO AND LIGHT DUTY TRUCKS
#> 74                                           SURFACE COATING OF LARGE APPLIANCES
#> 59                                                 SURFACE COATING OF METAL CANS
#> 100                                                SURFACE COATING OF METAL COIL
#> 95                                            SURFACE COATING OF METAL FURNITURE
#> 69                             SURFACE COATING OF MISC. METAL PARTS AND PRODUCTS
#> 84                                 SURFACE COATING OF PLASTIC PARTS AND PRODUCTS
#> 89                                     SURFACE COATING OF WOOD BUILDING PRODUCTS
#> 88                                                           SURFACE IMPOUNDMENT
#> 31                                SYNTHETIC HAZARDOUS ORGANICS (SOCMI) STANDARDS
#> 96                                                  TACONITE IRON ORE PROCESSING
#> 78                                                      TANK STANDARDS - LEVEL 1
#> 43                                          WET FORMED FIBERGLASS MAT PRODUCTION
#> 53                                       WOOD FURNITURE MANUFACTURING OPERATIONS
#> 91                                                  WOOD PERSERVING AREA SOURCES
#> 73                                                 WOOL FIBERGLASS MANUFACTURING
#>                                                                                  dropdown_label
#> 117                                         VVVVVV -  Chemical Manufacturing Area Sources (185)
#> 65                           LLLLLL - Acrylic And Modacrylic Fibers Production Area Sources (5)
#> 36                                                           GG - Aerospace Mfg. & Rework (266)
#> 25                                E - Approval Of State Programs &  Delegation Of Authority (2)
#> 136                    ZZZZZZ - Area Source: Aluminum, Copper & Other Nonferrous Foundries (62)
#> 5              AAAAAAA - Area Source: Asphalt Processing And Asphalt Roofing Manufacturing (77)
#> 10                                    BBBBBBB - Area Source: Chemical Preparations Industry (7)
#> 131                      YYYYY - Area Source: Electric Arc Furnace Steelmaking Facilities (119)
#> 132                                YYYYYY - Area Source: Ferroalloys Production Facilities (12)
#> 30                           EEEEEEE - Area Source: Gold Mine Ore Processing And Production (7)
#> 127          XXXXXX - Area Source: Nine Metal Fabrication And Finishing Source Categories (582)
#> 17                          CCCCCCC - Area Source: Paint And Allied Products Manufacturing (80)
#> 122                                WWWWWW - Area Source: Plating And Polishing Operations (524)
#> 24                                              DDDDDDD - Area Source: Prepared Feeds Mfg (305)
#> 64                            LLLLL - Asphalt Processing And Asphalt Roofing Manufacturing (33)
#> 116                                                             VVVV - Boat Manufacturing (440)
#> 56                                 JJJJJ - Brick & Structural Clay Products Manufacturing (191)
#> 71                                            MMMMMM - Carbon Black Production Area Sources (9)
#> 112                                                 UUUU - Cellulose Product Manufacturing (26)
#> 76                         NNNNNN - Chemical Manufacturing Area Sources: Chromium Compounds (8)
#> 67                                MM - Chemical Recovery Combustion Sources At Pulp Mills (215)
#> 133                                                              ZZZ - Chemical Sector Rule (3)
#> 72                                                          N - Chromium Electroplating (2,453)
#> 60                                                     KKKKK - Clay Ceramics Manufacturing (14)
#> 97                                                 RRRRRR - Clay Ceramics Mfg Area Sources (16)
#> 99      SS - Closed Vent Sys/Ctrl Devices, Recovery Devices, Routing To Fuel Gas Sys/Proc (150)
#> 113                                                      UUUUU - Coal & Oil-Fired Eusgu's (299)
#> 61                                                                 L - Coke Oven Batteries (58)
#> 15                               CCCCC - Coke Ovens: Pushing, Quenching And Battery Stacks (55)
#> 18                                             D - Compliance Extension, Early Hap Reductn (10)
#> 82                                                                PP - Container Standards (29)
#> 11                                                                   C - Delistings Of Haps (1)
#> 66                                                   M - Dry Cleaners Perchloroethylene (8,791)
#> 85                                                        PPPPP - Engine Test Cells/Stands (58)
#> 118                                     W - Epoxy Resins & Non-Nylon Polyamides Production (22)
#> 110                                                      UU - Equip Leaks-Control Level 2 (131)
#> 104                                                 TT - Equipment Leaks - Control Level 1 (20)
#> 40                                              H - Equipment Leaks Of Hazardous Organics (601)
#> 124                            XX - Ethylene Mpf: Heat Exchange Systems & Waste Operations (33)
#> 77                                                         O - Ethylene Oxide Sterilizers (225)
#> 125                                                             XXX - Ferroalloy Production (7)
#> 70                               MMMMM - Flexible Polyurethane Foam Fabrication Operations (19)
#> 81                            OOOOOO - Flexible Polyurethane Foam Prod & Fab Area Sources (148)
#> 49                                             III - Flexible Polyurethane Foam Production (61)
#> 90                                      QQQQQ - Friction Materials Manufacturing Facilities (9)
#> 16                                              CCCCCC - Gasoline Dispensing Facilities (3,505)
#> 92                                                            R - Gasoline Distribution (1,121)
#> 9   BBBBBB - Gasoline Distribution Bulk Terminals, Bulk Plants, And Pipeline Facilities (1,930)
#> 1                                                                A - General Provisions (7,958)
#> 129                                                           YY - Generic Mact Standards (152)
#> 102                                                        SSSSSS - Glass Mfg Area Sources (35)
#> 109                                                        U - Group 1 Polymers And Resins (49)
#> 54                                                     JJJ - Group Iv Polymers And Resins (138)
#> 103                                                    T - Halogenated Solvent Cleaning (2,355)
#> 27                                                       EEE - Hazardous Waste Combustors (263)
#> 121                                           WWWWW - Hospital Ethylene Oxide Sterilizers (250)
#> 75                                                    NNNNN - Hydrochloric Acid Production (50)
#> 93                                                           RR - Individual Drain Systems (32)
#> 87                                                  Q - Industrial Process Cooling Towers (340)
#> 57              JJJJJJ - Industrial, Commercial, And Institutional Boilers Area Sources (2,081)
#> 33                              FFFFF - Integrated Iron And Steel Manufacturing Facilities (35)
#> 29                                                       EEEEE - Iron And Steel Foundries (174)
#> 135                                         ZZZZZ - Iron And Steel Foundries Area Sources (325)
#> 86                                   PPPPPP - Lead Acid Battery Manufacturing Area Sources (60)
#> 106                                                    TTTT - Leather Finishing Operations (26)
#> 4                                                        AAAAA - Lime Manufacturing Plants (90)
#> 26                                                      EE - Magnetic Tape Mfg. Operations (15)
#> 22  DDDDD - Major Sources: Industrial/Commercial/Institutional Boilers & Process Heater (4,324)
#> 79                                              OOO - Manufacture Of Amino/Phenolic Resins (67)
#> 14                                                CCCC - Manufacturing Of Nutritional Yeast (8)
#> 128                                 Y - Marine Tank Vessel Loading & Unloading Operations (308)
#> 51                         IIIII - Mercury Emissions From Mercury Cell Chlor-Alkali Plants (19)
#> 20                                                           DDD - Mineral Wool Production (17)
#> 44                                            HHHHH - Miscellaneous Coating Manufacturing (284)
#> 32                              FFFF - Miscellaneous Organic Chemical Manufacturing (Mon) (671)
#> 3                                                AAAA - Municipal Solid Waste Landfills (1,212)
#> 42                                    HHH - Natural Gas Transmission & Storage Facilities (205)
#> 47                                           I - Negotiated Regs: Equip Leaks, Haz-Orgncs (107)
#> 19                                            DD - Off-Site Waste And Recovery Operations (202)
#> 41                                       HH - Oil And Natural Gas Production Facilities (8,378)
#> 114                                              VV - Oil-Water & Organic-Water Separators (63)
#> 28                                     EEEE - Organic Liquids Distribution (Non-Gasoline) (571)
#> 45               HHHHHH - Paint Strip & Misc Surface Coating Operations At Area Sources (6,904)
#> 55                                                      JJJJ - Paper & Other Web Coatings (543)
#> 68                                            MMM - Pesticide Active Ingredient Production (53)
#> 12                                                              CC - Petroleum Refineries (340)
#> 111                                      UUU - Petroleum Refineries: Ccu's, Cru's & Sru's (217)
#> 37                                                     GGG - Pharmaceutical Manufacturing (146)
#> 2                                                 AA - Phosperic Acid Manufacturing Plants (30)
#> 7                                                               BB - Phosphate Fertilizers (30)
#> 21                                             DDDD - Plywood And Composite Wood Products (743)
#> 83                                                      PPP - Polyether Polyols Production (54)
#> 46                                  HHHHHHH - Polyvinyl Chloride And Copolymers Production (21)
#> 52                                        J - Polyvinyl Chloride And Copolymers Production (20)
#> 23                      DDDDDD - Polyvinyl Chloride And Copolymers Production Area Sources (30)
#> 63                                                           LLL - Portland Cement Plants (233)
#> 62                                                  LL - Primary Aluminum Reduction Plants (33)
#> 105                                                             TTT - Primary Lead Smelting (2)
#> 107                                                      TTTTT - Primary Magnesium Refining (9)
#> 58                                                             KK - Printing & Publishing (711)
#> 80                      OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles (226)
#> 115                                                   VVV - Publicly Owned Treatment Works (27)
#> 98                                                                     S - Pulp And Paper (249)
#> 101                                               SSSSS - Refractory Products Manufacturing (8)
#> 120                                     WWWW - Reinforced Plastic Composites Production (1,255)
#> 6                         B - Req For Control Technology Determinations For Major Sources (187)
#> 126                                                       XXXX - Rubber Tire Manufacturing (68)
#> 94                                                    RRR - Secondary Aluminum Production (655)
#> 34                                          FFFFFF - Secondary Copper Smelting Area Sources (2)
#> 123                                                            X - Secondary Lead Smelters (41)
#> 108                           TTTTTT - Secondary Nonferrous Metals Processing Area Sources (20)
#> 8                                                      BBBBB - Semiconductor Manufacturing (67)
#> 48                                                        II - Shipbuilding & Ship Repair (181)
#> 39                                                               GGGGG - Site Remediation (259)
#> 35              G - Socmi Process Vents, Storage Vessels, Transfer Operations, Wastewater (539)
#> 38                                 GGGG - Solvent Extraction For Vegetable Oil Production (202)
#> 130                                                 YYYY - Stationary Combustion Turbines (443)
#> 134                 ZZZZ - Stationary Reciprocating Internal Combustion Engines (Rice) (42,984)
#> 13                          CCC - Steel Pickling-Hci Process Facilities And Hci Reg Plants (86)
#> 119                                          WW - Storage Vessels (Tanks)- Control Level 2 (65)
#> 50                                   IIII - Surface Coating Of Auto And Light Duty Trucks (183)
#> 74                                              NNNN - Surface Coating Of Large Appliances (36)
#> 59                                                    KKKK - Surface Coating Of Metal Cans (70)
#> 100                                                  SSSS - Surface Coating Of Metal Coil (126)
#> 95                                               RRRR - Surface Coating Of Metal Furniture (57)
#> 69                             MMMM - Surface Coating Of Misc. Metal Parts And Products (1,375)
#> 84                                   PPPP - Surface Coating Of Plastic Parts And Products (663)
#> 89                                       QQQQ - Surface Coating Of Wood Building Products (143)
#> 88                                                                 QQ - Surface Impoundment (9)
#> 31                                     F - Synthetic Hazardous Organics (Socmi) Standards (479)
#> 96                                                    RRRRR - Taconite Iron Ore Processing (15)
#> 78                                                           OO - Tank Standards - Level 1 (65)
#> 43                                             HHHH - Wet Formed Fiberglass Mat Production (12)
#> 53                                         JJ - Wood Furniture Manufacturing Operations (1,047)
#> 91                                                   QQQQQQ - Wood Perserving Area Sources (88)
#> 73                                                     NNN - Wool Fiberglass Manufacturing (50)
#>         N
#> 117   185
#> 65      5
#> 36    266
#> 25      2
#> 136    62
#> 5      77
#> 10      7
#> 131   119
#> 132    12
#> 30      7
#> 127   582
#> 17     80
#> 122   524
#> 24    305
#> 64     33
#> 116   440
#> 56    191
#> 71      9
#> 112    26
#> 76      8
#> 67    215
#> 133     3
#> 72   2453
#> 60     14
#> 97     16
#> 99    150
#> 113   299
#> 61     58
#> 15     55
#> 18     10
#> 82     29
#> 11      1
#> 66   8791
#> 85     58
#> 118    22
#> 110   131
#> 104    20
#> 40    601
#> 124    33
#> 77    225
#> 125     7
#> 70     19
#> 81    148
#> 49     61
#> 90      9
#> 16   3505
#> 92   1121
#> 9    1930
#> 1    7958
#> 129   152
#> 102    35
#> 109    49
#> 54    138
#> 103  2355
#> 27    263
#> 121   250
#> 75     50
#> 93     32
#> 87    340
#> 57   2081
#> 33     35
#> 29    174
#> 135   325
#> 86     60
#> 106    26
#> 4      90
#> 26     15
#> 22   4324
#> 79     67
#> 14      8
#> 128   308
#> 51     19
#> 20     17
#> 44    284
#> 32    671
#> 3    1212
#> 42    205
#> 47    107
#> 19    202
#> 41   8378
#> 114    63
#> 28    571
#> 45   6904
#> 55    543
#> 68     53
#> 12    340
#> 111   217
#> 37    146
#> 2      30
#> 7      30
#> 21    743
#> 83     54
#> 46     21
#> 52     20
#> 23     30
#> 63    233
#> 62     33
#> 105     2
#> 107     9
#> 58    711
#> 80    226
#> 115    27
#> 98    249
#> 101     8
#> 120  1255
#> 6     187
#> 126    68
#> 94    655
#> 34      2
#> 123    41
#> 108    20
#> 8      67
#> 48    181
#> 39    259
#> 35    539
#> 38    202
#> 130   443
#> 134 42984
#> 13     86
#> 119    65
#> 50    183
#> 74     36
#> 59     70
#> 100   126
#> 95     57
#> 69   1375
#> 84    663
#> 89    143
#> 88      9
#> 31    479
#> 96     15
#> 78     65
#> 43     12
#> 53   1047
#> 91     88
#> 73     50
  mycodes <- c("BBBBBB", "OOOO")
  frs_by_mact[subpart %in% mycodes, ]
#> Error: object 'frs_by_mact' not found
  mact_table[grepl("smelt", mact_table$title, ignore.case = T), ]
#>     subpart                                  title
#> 34   FFFFFF SECONDARY COPPER SMELTING AREA SOURCES
#> 105     TTT                  PRIMARY LEAD SMELTING
#> 123       X                SECONDARY LEAD SMELTERS
#>                                          dropdown_label  N
#> 34  FFFFFF - Secondary Copper Smelting Area Sources (2)  2
#> 105                     TTT - Primary Lead Smelting (2)  2
#> 123                    X - Secondary Lead Smelters (41) 41
  frs_by_mact[grepl("smelt", title, ignore.case = T), ]
#> Error: object 'frs_by_mact' not found
  # a single site can be covered by 19 categories
  frs_by_mact[, howmany := .N, by="programid"][order(howmany), ] 
#> Error: object 'frs_by_mact' not found
  table(frs_by_mact[, howmany := .N, by="programid"][order(howmany), howmany])
#> Error: object 'frs_by_mact' not found