frs_by_mact (DATA) MACT NESHAP subpart(s) that each EPA-regulated site is subject to

This is a data.table with one row per site -- MACT subpart pair, so it has multiple rows for one site if the site is covered by multiple subparts. It has been joined with frs_by_programid to get latlons for matching facilities. @details

There are about 112k rows here but only about 83k unique program IDs in this table,

which is from the ECHO data download of ICIS Air and AFS.

The programid column here should be found in the pgm_sys_id column in frs_by_programid,

but as of 6/14/23 only a little over half of them were found there, so this is work in progress

to be resolved.

table(frs_by_mact$programid %in% frs_by_programid$pgm_sys_id)

FALSE TRUE

56497 55411

Also there are some typos in the downloaded dataset from ECHO/FRS, such as

"WOOD PERSERVING AREA SOURCES"

See also

Examples

 
  mact_table
  mact_table[order(mact_table$title),]
  mycodes <- c("BBBBBB", "OOOO")
  frs_by_mact[subpart %in% mycodes, ]
  mact_table[grepl("smelt", mact_table$title, ignore.case = T), ]
  frs_by_mact[grepl("smelt", title, ignore.case = T), ]
  # a single site can be covered by 19 categories
  frs_by_mact[, howmany := .N, by="programid"][order(howmany), ] 
  table(frs_by_mact[, howmany := .N, by="programid"][order(howmany), howmany])