Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

Usage

frs_from_programid(programname, programid)

Arguments

programname

name of EPA program that the programid is from: "RCRAINFO" is the programname and "XJW000000174" is the programid if the full record was RCRAINFO:XJW000000174

programid

like "XJW000000174" "RCRAINFO" is the programname and "XJW000000174" is the programid if the full record was RCRAINFO:XJW000000174

Value

relevant rows of the data.table called frs, which has column names that are "lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"

Examples

 test <- data.frame(programname = c('STATE','FIS','FIS'),
                    programid = c('#5005','0-0000-01097','0-0000-01103'))
 x = frs_from_programid(test$programname, test$programid)
#> 
#> Failed trying to connect to pins board server.
#> 
#> Arrow-format datasets (blocks, etc.) are up-to-date -- locally-installed and latest-released data repository versions match.
#> frs_by_programid   - is NOT in memory. Checking local disk... frs_by_programid  is loading from /home/runner/work/_temp/Library/EJAM/data/frs_by_programid.arrow ...done.
#> frs_by_programid   - was loaded from local folder, so server copy was not sought.
#> 
 x
#>         lat        lon  REGISTRY_ID                           PRIMARY_NAME
#>       <num>      <num>       <char>                                 <char>
#> 1: 44.13415 -104.12563 110012799846                         SNOW ADVENTURE
#> 2: 40.71490  -74.00316 110019246163 SITE RECLAMATION SYSTEMS MTU (SUN R&M)
#> 3: 40.76395  -73.97037 110019163359                        ENVIROSOUND MTU
#>     NAICS    SIC            PGM_SYS_ACRNMS
#>    <char> <char>                    <char>
#> 1:          5541 ICIS:2656510, STATE:#5005
#> 2:                        FIS:0-0000-01097
#> 3:                        FIS:0-0000-01103
 mapfast(x)