Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

frs_from_programid(programid)

Arguments

programid

vector of one or more EPA Program ID codes used by FRS

Value

relevant rows of the data.table called frs, which has column names that are "lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"

Examples

 x = frs_from_programid(testids_program_sys_id)
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> frs_by_programid   - is NOT in memory. Checking local disk... frs_by_programid  is NOT found locally on disk at ~/../Downloads/EJAMbigfiles/frs_by_programid.arrow 
#> frs_by_programid   - is NOT in local folder.
#> frs_by_programid  - was read from pins board server, into memory (to specified envir).
#> 
 x
#>         lat        lon  REGISTRY_ID                  PRIMARY_NAME NAICS  SIC
#> 1: 43.00028  -76.69323 110019033810 HEJAMADA CAMPGROUND & RV PARK       8999
#> 2: 41.38786  -73.43935 110044340807            NEJAME & SONS INC.           
#> 3: 40.77499  -74.15929 110056111559      LA VEJAMA DISCOUNT STORE           
#> 4: 30.25468  -90.94284 110056982323           GRACEJAMES ENT, LLC       7542
#> 5: 38.22678 -121.02361 110066714951      CAMANCHE DAM POWER HOUSE       4911
#> 6: 32.80406 -117.02753 110070538057             BEJAMIN HAFERTEPE           
#> 7: 33.66110 -117.86959 110070874073                     EJAM INC.           
#> 8: 33.70369 -117.86675 110071293460                      EJAM INC           
#>           PGM_SYS_ACRNMS
#> 1:      FIS:7-0540-00003
#> 2:          SIMS:1513529
#> 3:       NJ-NJEMS:485659
#> 4:       NPDES:LAG750956
#> 5:  CA-ENVIROVIEW:354362
#> 6: RCRAINFO:CAC002995519
#> 7:       ICIS:3601252181
#> 8:       ICIS:3601439158
 mapfast(x)