Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

Use EPA Program ID to see FRS Facility Registry Service data on those EPA-regulated sites

Usage

frs_from_programid(programid)

Arguments

programid

vector of one or more EPA Program ID codes used by FRS

Value

relevant rows of the data.table called frs, which has column names that are "lat" "lon" "REGISTRY_ID" "PRIMARY_NAME" "NAICS" "PGM_SYS_ACRNMS"

Examples

 x = frs_from_programid(testids_program_sys_id)
#> Connecting to Posit Connect 2024.01.0 at <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> frs_by_programid   - is NOT in memory. Checking local disk... frs_by_programid  is loading from ~/../Downloads/EJAMbigfiles/frs_by_programid.arrow ...done.
#> frs_by_programid   - was loaded from local folder, so server copy was not sought.
#> 
 x
#> Key: <REGISTRY_ID>
#>         lat        lon  REGISTRY_ID                  PRIMARY_NAME  NAICS    SIC        PGM_SYS_ACRNMS
#>       <num>      <num>       <char>                        <char> <char> <char>                <char>
#> 1: 43.00028  -76.69323 110019033810 HEJAMADA CAMPGROUND & RV PARK          8999      FIS:7-0540-00003
#> 2: 41.38786  -73.43935 110044340807            NEJAME & SONS INC.                        SIMS:1513529
#> 3: 40.77499  -74.15929 110056111559      LA VEJAMA DISCOUNT STORE                     NJ-NJEMS:485659
#> 4: 30.25468  -90.94284 110056982323           GRACEJAMES ENT, LLC          7542       NPDES:LAG750956
#> 5: 38.22678 -121.02361 110066714951      CAMANCHE DAM POWER HOUSE          4911  CA-ENVIROVIEW:354362
#> 6: 32.80406 -117.02753 110070538057             BEJAMIN HAFERTEPE               RCRAINFO:CAC002995519
#> 7: 33.66110 -117.86959 110070874073                     EJAM INC.                     ICIS:3601252181
#> 8: 33.70369 -117.86675 110071293460                      EJAM INC                     ICIS:3601439158
 mapfast(x)