DRAFT / WAS WORK IN PROGRESS Key buffering function - wrapper redirecting to the right version of getblocksnearby()

DRAFT / WAS WORK IN PROGRESS Key buffering function - wrapper redirecting to the right version of getblocksnearby()

Usage

getblocksnearby2(
  sitepoints,
  radius = 3,
  maxradius = 31.07,
  avoidorphans = FALSE,
  quadtree = is.null,
  ...
)

Arguments

sitepoints

see getblocksnearbyviaQuadTree() or other such functions

radius

see getblocksnearbyviaQuadTree() or other such functions

maxradius

see getblocksnearbyviaQuadTree() or other such functions

avoidorphans

see getblocksnearbyviaQuadTree() or other such functions

quadtree

a large quadtree object created from the SearchTree package example: SearchTrees::createTree( quaddata, treeType = "quad", dataType = "point")

...

see getblocksnearbyviaQuadTree_Clustered() or other such functions

Details

For all examples, see ejamit()

Like getblocksnearby() but tries to handle localtree and quadtree parameter differently

not sure how to check if they are in the right environment.

See also