Find all blocks within each of the FIPS codes provided

Allows EJAM to analyze and compare Counties, for example

Usage

getblocksnearby_from_fips(fips, inshiny = FALSE, need_blockwt = TRUE)

Arguments

fips

vector of FIPS codes identifying blockgroups, tracts, counties, or states. This is useful if -- instead of gettings stats on and comparing circular buffers or polygons -- one will be getting stats on one or more tracts, or analyzing and comparing blockgroups in a county, or comparing whole counties to each other, within a State.

inshiny

used by shiny app server code to handle errors via validate() instead of stop()

need_blockwt

set to FALSE to speed it up if you do not need blockwt

Value

same as for getblocksnearby but one row per FIPS, and the distance column is irrelevant

See also

Examples

  x <- getblocksnearby_from_fips(fips_counties_from_state_abbrev("DE"))
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> blockid2fips   - is NOT in memory. Checking local disk... blockid2fips  is loading from ~/../Downloads/EJAMbigfiles/blockid2fips.arrow ...done.
#> blockid2fips   - was loaded from local folder, so server copy was not sought.
#> 
#> Connecting to Posit Connect 2024.01.0 at
#> <https://rstudio-connect.dmap-stage.aws.epa.gov>
#> Successfully connected to Posit Connect pins board.
#> 
#> bgid2fips    - is NOT in memory. Checking local disk... bgid2fips   is loading from ~/../Downloads/EJAMbigfiles/bgid2fips.arrow ...done.
#> bgid2fips    - was loaded from local folder, so server copy was not sought.
#> 
#> Joining with `by = join_by(blockid)`
#> Warning: Unknown argument 'on' has been passed.
  counties_ej <- doaggregate(x)
  #cannot use mapfast(counties_ej$results_bysite) since no lat lon.  mapfastej_counties() should work...
  y =  ejamit(fips=fips_counties_from_statename("Delaware"))
#> Finding blocks in each FIPS Census unit.
#> Joining with `by = join_by(blockid)`
#> Warning: Unknown argument 'on' has been passed.
#> Aggregating at each FIPS Census unit and overall.
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> Population Density: 
#> 
#>    0% of places account for 50% of the total population (approx.)
#>    1, 5, 10 places account for 58%, NA%, NA% of the total population
#> 
#> Output is a list with the following names:
#> 
#>  out is a list 
#> 
#>                                  Name     format rows cols vectorlength
#> 1                     results_overall data.table    1  290           NA
#> 2                      results_bysite data.table    3  290           NA
#> 3                 results_bybg_people data.table  706   85           NA
#> 4                           longnames     vector   NA   NA          290
#> 5 count_of_blocks_near_multiple_sites     vector   NA   NA            1
#> 6                  results_summarized     vector   NA   NA            2
#> 7                           formatted data.frame  290    2           NA
  
  # x=getblocksnearby_from_fips("482011000011") # one blockgroup only
  # y=doaggregate(x)