Get point locations for US EPA-regulated facilities that have sources subject to Maximum Achievable Control Technology (MACT) standards under the Clean Air Act.

Get point locations for US EPA-regulated facilities that have sources subject to Maximum Achievable Control Technology (MACT) standards under the Clean Air Act.

Usage

latlon_from_mactsubpart(subpart = "JJJ", include_if_no_latlon = FALSE)

Arguments

subpart

vector of one or more strings indicating the Subpart of CFR Title 40 Part 63 that covers the source category of interest, such as "FFFF" - see for example, https://www.ecfr.gov/current/title-40/part-63/subpart-FFFF

include_if_no_latlon

logical - many in the database lack lat lon values but have a MACT code

Value

a data.table with columns named

programid, subpart, title, lat, lon, REGISTRY_ID, program

for US EPA FRS sites with that MACT code. Or NA if none found.

Details

See EPA information about MACT subparts, NESHAPs, etc.

EPA's ECHO query tools also provide search by NAICS or SIC, and by MACT subpart.

Examples

  mact_table
#>     subpart
#> 1         A
#> 2        AA
#> 3      AAAA
#> 4     AAAAA
#> 5   AAAAAAA
#> 6         B
#> 7        BB
#> 8     BBBBB
#> 9    BBBBBB
#> 10  BBBBBBB
#> 11        C
#> 12       CC
#> 13      CCC
#> 14     CCCC
#> 15    CCCCC
#> 16   CCCCCC
#> 17  CCCCCCC
#> 18        D
#> 19       DD
#> 20      DDD
#> 21     DDDD
#> 22    DDDDD
#> 23   DDDDDD
#> 24  DDDDDDD
#> 25        E
#> 26       EE
#> 27      EEE
#> 28     EEEE
#> 29    EEEEE
#> 30  EEEEEEE
#> 31        F
#> 32     FFFF
#> 33    FFFFF
#> 34   FFFFFF
#> 35        G
#> 36       GG
#> 37      GGG
#> 38     GGGG
#> 39    GGGGG
#> 40        H
#> 41       HH
#> 42      HHH
#> 43     HHHH
#> 44    HHHHH
#> 45   HHHHHH
#> 46  HHHHHHH
#> 47        I
#> 48       II
#> 49      III
#> 50     IIII
#> 51    IIIII
#> 52        J
#> 53       JJ
#> 54      JJJ
#> 55     JJJJ
#> 56    JJJJJ
#> 57   JJJJJJ
#> 58       KK
#> 59     KKKK
#> 60    KKKKK
#> 61        L
#> 62       LL
#> 63      LLL
#> 64    LLLLL
#> 65   LLLLLL
#> 66        M
#> 67       MM
#> 68      MMM
#> 69     MMMM
#> 70    MMMMM
#> 71   MMMMMM
#> 72        N
#> 73      NNN
#> 74     NNNN
#> 75    NNNNN
#> 76   NNNNNN
#> 77        O
#> 78       OO
#> 79      OOO
#> 80     OOOO
#> 81   OOOOOO
#> 82       PP
#> 83      PPP
#> 84     PPPP
#> 85    PPPPP
#> 86   PPPPPP
#> 87        Q
#> 88       QQ
#> 89     QQQQ
#> 90    QQQQQ
#> 91   QQQQQQ
#> 92        R
#> 93       RR
#> 94      RRR
#> 95     RRRR
#> 96    RRRRR
#> 97   RRRRRR
#> 98        S
#> 99       SS
#> 100    SSSS
#> 101   SSSSS
#> 102  SSSSSS
#> 103       T
#> 104      TT
#> 105     TTT
#> 106    TTTT
#> 107   TTTTT
#> 108  TTTTTT
#> 109       U
#> 110      UU
#> 111     UUU
#> 112    UUUU
#> 113   UUUUU
#> 114      VV
#> 115     VVV
#> 116    VVVV
#> 117  VVVVVV
#> 118       W
#> 119      WW
#> 120    WWWW
#> 121   WWWWW
#> 122  WWWWWW
#> 123       X
#> 124      XX
#> 125     XXX
#> 126    XXXX
#> 127  XXXXXX
#> 128       Y
#> 129      YY
#> 130    YYYY
#> 131   YYYYY
#> 132  YYYYYY
#> 133     ZZZ
#> 134    ZZZZ
#> 135   ZZZZZ
#> 136  ZZZZZZ
#>                                                                            title
#> 1                                                             GENERAL PROVISIONS
#> 2                                            PHOSPERIC ACID MANUFACTURING PLANTS
#> 3                                                MUNICIPAL SOLID WASTE LANDFILLS
#> 4                                                      LIME MANUFACTURING PLANTS
#> 5              AREA SOURCE: ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 6                    REQ FOR CONTROL TECHNOLOGY DETERMINATIONS FOR MAJOR SOURCES
#> 7                                                          PHOSPHATE FERTILIZERS
#> 8                                                    SEMICONDUCTOR MANUFACTURING
#> 9     GASOLINE DISTRIBUTION BULK TERMINALS, BULK PLANTS, AND PIPELINE FACILITIES
#> 10                                   AREA SOURCE: CHEMICAL PREPARATIONS INDUSTRY
#> 11                                                            DELISTINGS OF HAPS
#> 12                                                          PETROLEUM REFINERIES
#> 13                      STEEL PICKLING-HCI PROCESS FACILITIES AND HCI REG PLANTS
#> 14                                            MANUFACTURING OF NUTRITIONAL YEAST
#> 15                             COKE OVENS: PUSHING, QUENCHING AND BATTERY STACKS
#> 16                                                GASOLINE DISPENSING FACILITIES
#> 17                          AREA SOURCE: PAINT AND ALLIED PRODUCTS MANUFACTURING
#> 18                                       COMPLIANCE EXTENSION, EARLY HAP REDUCTN
#> 19                                        OFF-SITE WASTE AND RECOVERY OPERATIONS
#> 20                                                       MINERAL WOOL PRODUCTION
#> 21                                           PLYWOOD AND COMPOSITE WOOD PRODUCTS
#> 22   MAJOR SOURCES: INDUSTRIAL/COMMERCIAL/INSTITUTIONAL BOILERS & PROCESS HEATER
#> 23                     POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION AREA SOURCES
#> 24                                               AREA SOURCE: PREPARED FEEDS MFG
#> 25                         APPROVAL OF STATE PROGRAMS &  DELEGATION OF AUTHORITY
#> 26                                                 MAGNETIC TAPE MFG. OPERATIONS
#> 27                                                    HAZARDOUS WASTE COMBUSTORS
#> 28                                   ORGANIC LIQUIDS DISTRIBUTION (NON-GASOLINE)
#> 29                                                      IRON AND STEEL FOUNDRIES
#> 30                          AREA SOURCE: GOLD MINE ORE PROCESSING AND PRODUCTION
#> 31                                SYNTHETIC HAZARDOUS ORGANICS (SOCMI) STANDARDS
#> 32                            MISCELLANEOUS ORGANIC CHEMICAL MANUFACTURING (MON)
#> 33                            INTEGRATED IRON AND STEEL MANUFACTURING FACILITIES
#> 34                                        SECONDARY COPPER SMELTING AREA SOURCES
#> 35         SOCMI PROCESS VENTS, STORAGE VESSELS, TRANSFER OPERATIONS, WASTEWATER
#> 36                                                       AEROSPACE MFG. & REWORK
#> 37                                                  PHARMACEUTICAL MANUFACTURING
#> 38                               SOLVENT EXTRACTION FOR VEGETABLE OIL PRODUCTION
#> 39                                                              SITE REMEDIATION
#> 40                                         EQUIPMENT LEAKS OF HAZARDOUS ORGANICS
#> 41                                     OIL AND NATURAL GAS PRODUCTION FACILITIES
#> 42                                 NATURAL GAS TRANSMISSION & STORAGE FACILITIES
#> 43                                          WET FORMED FIBERGLASS MAT PRODUCTION
#> 44                                           MISCELLANEOUS COATING MANUFACTURING
#> 45                 PAINT STRIP & MISC SURFACE COATING OPERATIONS AT AREA SOURCES
#> 46                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 47                                      NEGOTIATED REGS: EQUIP LEAKS, HAZ-ORGNCS
#> 48                                                    SHIPBUILDING & SHIP REPAIR
#> 49                                         FLEXIBLE POLYURETHANE FOAM PRODUCTION
#> 50                                 SURFACE COATING OF AUTO AND LIGHT DUTY TRUCKS
#> 51                       MERCURY EMISSIONS FROM MERCURY CELL CHLOR-ALKALI PLANTS
#> 52                                  POLYVINYL CHLORIDE AND COPOLYMERS PRODUCTION
#> 53                                       WOOD FURNITURE MANUFACTURING OPERATIONS
#> 54                                                  GROUP IV POLYMERS AND RESINS
#> 55                                                    PAPER & OTHER WEB COATINGS
#> 56                                BRICK & STRUCTURAL CLAY PRODUCTS MANUFACTURING
#> 57                INDUSTRIAL, COMMERCIAL, AND INSTITUTIONAL BOILERS AREA SOURCES
#> 58                                                         PRINTING & PUBLISHING
#> 59                                                 SURFACE COATING OF METAL CANS
#> 60                                                   CLAY CERAMICS MANUFACTURING
#> 61                                                           COKE OVEN BATTERIES
#> 62                                             PRIMARY ALUMINUM REDUCTION PLANTS
#> 63                                                        PORTLAND CEMENT PLANTS
#> 64                          ASPHALT PROCESSING AND ASPHALT ROOFING MANUFACTURING
#> 65                         ACRYLIC AND MODACRYLIC FIBERS PRODUCTION AREA SOURCES
#> 66                                                DRY CLEANERS PERCHLOROETHYLENE
#> 67                            CHEMICAL RECOVERY COMBUSTION SOURCES AT PULP MILLS
#> 68                                        PESTICIDE ACTIVE INGREDIENT PRODUCTION
#> 69                             SURFACE COATING OF MISC. METAL PARTS AND PRODUCTS
#> 70                             FLEXIBLE POLYURETHANE FOAM FABRICATION OPERATIONS
#> 71                                          CARBON BLACK PRODUCTION AREA SOURCES
#> 72                                                       CHROMIUM ELECTROPLATING
#> 73                                                 WOOL FIBERGLASS MANUFACTURING
#> 74                                           SURFACE COATING OF LARGE APPLIANCES
#> 75                                                  HYDROCHLORIC ACID PRODUCTION
#> 76                       CHEMICAL MANUFACTURING AREA SOURCES: CHROMIUM COMPOUNDS
#> 77                                                    ETHYLENE OXIDE STERILIZERS
#> 78                                                      TANK STANDARDS - LEVEL 1
#> 79                                          MANUFACTURE OF AMINO/PHENOLIC RESINS
#> 80                    PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 81                            FLEXIBLE POLYURETHANE FOAM PROD & FAB AREA SOURCES
#> 82                                                           CONTAINER STANDARDS
#> 83                                                  POLYETHER POLYOLS PRODUCTION
#> 84                                 SURFACE COATING OF PLASTIC PARTS AND PRODUCTS
#> 85                                                      ENGINE TEST CELLS/STANDS
#> 86                                  LEAD ACID BATTERY MANUFACTURING AREA SOURCES
#> 87                                             INDUSTRIAL PROCESS COOLING TOWERS
#> 88                                                           SURFACE IMPOUNDMENT
#> 89                                     SURFACE COATING OF WOOD BUILDING PRODUCTS
#> 90                                   FRICTION MATERIALS MANUFACTURING FACILITIES
#> 91                                                  WOOD PERSERVING AREA SOURCES
#> 92                                                         GASOLINE DISTRIBUTION
#> 93                                                      INDIVIDUAL DRAIN SYSTEMS
#> 94                                                 SECONDARY ALUMINUM PRODUCTION
#> 95                                            SURFACE COATING OF METAL FURNITURE
#> 96                                                  TACONITE IRON ORE PROCESSING
#> 97                                                CLAY CERAMICS MFG AREA SOURCES
#> 98                                                                PULP AND PAPER
#> 99  CLOSED VENT SYS/CTRL DEVICES, RECOVERY DEVICES, ROUTING TO FUEL GAS SYS/PROC
#> 100                                                SURFACE COATING OF METAL COIL
#> 101                                            REFRACTORY PRODUCTS MANUFACTURING
#> 102                                                       GLASS MFG AREA SOURCES
#> 103                                                 HALOGENATED SOLVENT CLEANING
#> 104                                            EQUIPMENT LEAKS - CONTROL LEVEL 1
#> 105                                                        PRIMARY LEAD SMELTING
#> 106                                                 LEATHER FINISHING OPERATIONS
#> 107                                                   PRIMARY MAGNESIUM REFINING
#> 108                          SECONDARY NONFERROUS METALS PROCESSING AREA SOURCES
#> 109                                                  GROUP 1 POLYMERS AND RESINS
#> 110                                                  EQUIP LEAKS-CONTROL LEVEL 2
#> 111                                   PETROLEUM REFINERIES: CCU'S, CRU'S & SRU'S
#> 112                                              CELLULOSE PRODUCT MANUFACTURING
#> 113                                                     COAL & OIL-FIRED EUSGU'S
#> 114                                         OIL-WATER & ORGANIC-WATER SEPARATORS
#> 115                                               PUBLICLY OWNED TREATMENT WORKS
#> 116                                                           BOAT MANUFACTURING
#> 117                                          CHEMICAL MANUFACTURING AREA SOURCES
#> 118                               EPOXY RESINS & NON-NYLON POLYAMIDES PRODUCTION
#> 119                                     STORAGE VESSELS (TANKS)- CONTROL LEVEL 2
#> 120                                     REINFORCED PLASTIC COMPOSITES PRODUCTION
#> 121                                          HOSPITAL ETHYLENE OXIDE STERILIZERS
#> 122                                AREA SOURCE: PLATING AND POLISHING OPERATIONS
#> 123                                                      SECONDARY LEAD SMELTERS
#> 124                       ETHYLENE MPF: HEAT EXCHANGE SYSTEMS & WASTE OPERATIONS
#> 125                                                        FERROALLOY PRODUCTION
#> 126                                                    RUBBER TIRE MANUFACTURING
#> 127          AREA SOURCE: NINE METAL FABRICATION AND FINISHING SOURCE CATEGORIES
#> 128                            MARINE TANK VESSEL LOADING & UNLOADING OPERATIONS
#> 129                                                       GENERIC MACT STANDARDS
#> 130                                               STATIONARY COMBUSTION TURBINES
#> 131                     AREA SOURCE: ELECTRIC ARC FURNACE STEELMAKING FACILITIES
#> 132                               AREA SOURCE: FERROALLOYS PRODUCTION FACILITIES
#> 133                                                         CHEMICAL SECTOR RULE
#> 134                  STATIONARY RECIPROCATING INTERNAL COMBUSTION ENGINES (RICE)
#> 135                                        IRON AND STEEL FOUNDRIES AREA SOURCES
#> 136                   AREA SOURCE: ALUMINUM, COPPER & OTHER NONFERROUS FOUNDRIES
#>                                                                                  dropdown_label
#> 1                                                                A - General Provisions (7,958)
#> 2                                                 AA - Phosperic Acid Manufacturing Plants (30)
#> 3                                                AAAA - Municipal Solid Waste Landfills (1,212)
#> 4                                                        AAAAA - Lime Manufacturing Plants (90)
#> 5              AAAAAAA - Area Source: Asphalt Processing And Asphalt Roofing Manufacturing (77)
#> 6                         B - Req For Control Technology Determinations For Major Sources (187)
#> 7                                                               BB - Phosphate Fertilizers (30)
#> 8                                                      BBBBB - Semiconductor Manufacturing (67)
#> 9   BBBBBB - Gasoline Distribution Bulk Terminals, Bulk Plants, And Pipeline Facilities (1,930)
#> 10                                    BBBBBBB - Area Source: Chemical Preparations Industry (7)
#> 11                                                                   C - Delistings Of Haps (1)
#> 12                                                              CC - Petroleum Refineries (340)
#> 13                          CCC - Steel Pickling-Hci Process Facilities And Hci Reg Plants (86)
#> 14                                                CCCC - Manufacturing Of Nutritional Yeast (8)
#> 15                               CCCCC - Coke Ovens: Pushing, Quenching And Battery Stacks (55)
#> 16                                              CCCCCC - Gasoline Dispensing Facilities (3,505)
#> 17                          CCCCCCC - Area Source: Paint And Allied Products Manufacturing (80)
#> 18                                             D - Compliance Extension, Early Hap Reductn (10)
#> 19                                            DD - Off-Site Waste And Recovery Operations (202)
#> 20                                                           DDD - Mineral Wool Production (17)
#> 21                                             DDDD - Plywood And Composite Wood Products (743)
#> 22  DDDDD - Major Sources: Industrial/Commercial/Institutional Boilers & Process Heater (4,324)
#> 23                      DDDDDD - Polyvinyl Chloride And Copolymers Production Area Sources (30)
#> 24                                              DDDDDDD - Area Source: Prepared Feeds Mfg (305)
#> 25                                E - Approval Of State Programs &  Delegation Of Authority (2)
#> 26                                                      EE - Magnetic Tape Mfg. Operations (15)
#> 27                                                       EEE - Hazardous Waste Combustors (263)
#> 28                                     EEEE - Organic Liquids Distribution (Non-Gasoline) (571)
#> 29                                                       EEEEE - Iron And Steel Foundries (174)
#> 30                           EEEEEEE - Area Source: Gold Mine Ore Processing And Production (7)
#> 31                                     F - Synthetic Hazardous Organics (Socmi) Standards (479)
#> 32                              FFFF - Miscellaneous Organic Chemical Manufacturing (Mon) (671)
#> 33                              FFFFF - Integrated Iron And Steel Manufacturing Facilities (35)
#> 34                                          FFFFFF - Secondary Copper Smelting Area Sources (2)
#> 35              G - Socmi Process Vents, Storage Vessels, Transfer Operations, Wastewater (539)
#> 36                                                           GG - Aerospace Mfg. & Rework (266)
#> 37                                                     GGG - Pharmaceutical Manufacturing (146)
#> 38                                 GGGG - Solvent Extraction For Vegetable Oil Production (202)
#> 39                                                               GGGGG - Site Remediation (259)
#> 40                                              H - Equipment Leaks Of Hazardous Organics (601)
#> 41                                       HH - Oil And Natural Gas Production Facilities (8,378)
#> 42                                    HHH - Natural Gas Transmission & Storage Facilities (205)
#> 43                                             HHHH - Wet Formed Fiberglass Mat Production (12)
#> 44                                            HHHHH - Miscellaneous Coating Manufacturing (284)
#> 45               HHHHHH - Paint Strip & Misc Surface Coating Operations At Area Sources (6,904)
#> 46                                  HHHHHHH - Polyvinyl Chloride And Copolymers Production (21)
#> 47                                           I - Negotiated Regs: Equip Leaks, Haz-Orgncs (107)
#> 48                                                        II - Shipbuilding & Ship Repair (181)
#> 49                                             III - Flexible Polyurethane Foam Production (61)
#> 50                                   IIII - Surface Coating Of Auto And Light Duty Trucks (183)
#> 51                         IIIII - Mercury Emissions From Mercury Cell Chlor-Alkali Plants (19)
#> 52                                        J - Polyvinyl Chloride And Copolymers Production (20)
#> 53                                         JJ - Wood Furniture Manufacturing Operations (1,047)
#> 54                                                     JJJ - Group Iv Polymers And Resins (138)
#> 55                                                      JJJJ - Paper & Other Web Coatings (543)
#> 56                                 JJJJJ - Brick & Structural Clay Products Manufacturing (191)
#> 57              JJJJJJ - Industrial, Commercial, And Institutional Boilers Area Sources (2,081)
#> 58                                                             KK - Printing & Publishing (711)
#> 59                                                    KKKK - Surface Coating Of Metal Cans (70)
#> 60                                                     KKKKK - Clay Ceramics Manufacturing (14)
#> 61                                                                 L - Coke Oven Batteries (58)
#> 62                                                  LL - Primary Aluminum Reduction Plants (33)
#> 63                                                           LLL - Portland Cement Plants (233)
#> 64                            LLLLL - Asphalt Processing And Asphalt Roofing Manufacturing (33)
#> 65                           LLLLLL - Acrylic And Modacrylic Fibers Production Area Sources (5)
#> 66                                                   M - Dry Cleaners Perchloroethylene (8,791)
#> 67                                MM - Chemical Recovery Combustion Sources At Pulp Mills (215)
#> 68                                            MMM - Pesticide Active Ingredient Production (53)
#> 69                             MMMM - Surface Coating Of Misc. Metal Parts And Products (1,375)
#> 70                               MMMMM - Flexible Polyurethane Foam Fabrication Operations (19)
#> 71                                            MMMMMM - Carbon Black Production Area Sources (9)
#> 72                                                          N - Chromium Electroplating (2,453)
#> 73                                                     NNN - Wool Fiberglass Manufacturing (50)
#> 74                                              NNNN - Surface Coating Of Large Appliances (36)
#> 75                                                    NNNNN - Hydrochloric Acid Production (50)
#> 76                         NNNNNN - Chemical Manufacturing Area Sources: Chromium Compounds (8)
#> 77                                                         O - Ethylene Oxide Sterilizers (225)
#> 78                                                           OO - Tank Standards - Level 1 (65)
#> 79                                              OOO - Manufacture Of Amino/Phenolic Resins (67)
#> 80                      OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles (226)
#> 81                            OOOOOO - Flexible Polyurethane Foam Prod & Fab Area Sources (148)
#> 82                                                                PP - Container Standards (29)
#> 83                                                      PPP - Polyether Polyols Production (54)
#> 84                                   PPPP - Surface Coating Of Plastic Parts And Products (663)
#> 85                                                        PPPPP - Engine Test Cells/Stands (58)
#> 86                                   PPPPPP - Lead Acid Battery Manufacturing Area Sources (60)
#> 87                                                  Q - Industrial Process Cooling Towers (340)
#> 88                                                                 QQ - Surface Impoundment (9)
#> 89                                       QQQQ - Surface Coating Of Wood Building Products (143)
#> 90                                      QQQQQ - Friction Materials Manufacturing Facilities (9)
#> 91                                                   QQQQQQ - Wood Perserving Area Sources (88)
#> 92                                                            R - Gasoline Distribution (1,121)
#> 93                                                           RR - Individual Drain Systems (32)
#> 94                                                    RRR - Secondary Aluminum Production (655)
#> 95                                               RRRR - Surface Coating Of Metal Furniture (57)
#> 96                                                    RRRRR - Taconite Iron Ore Processing (15)
#> 97                                                 RRRRRR - Clay Ceramics Mfg Area Sources (16)
#> 98                                                                     S - Pulp And Paper (249)
#> 99      SS - Closed Vent Sys/Ctrl Devices, Recovery Devices, Routing To Fuel Gas Sys/Proc (150)
#> 100                                                  SSSS - Surface Coating Of Metal Coil (126)
#> 101                                               SSSSS - Refractory Products Manufacturing (8)
#> 102                                                        SSSSSS - Glass Mfg Area Sources (35)
#> 103                                                    T - Halogenated Solvent Cleaning (2,355)
#> 104                                                 TT - Equipment Leaks - Control Level 1 (20)
#> 105                                                             TTT - Primary Lead Smelting (2)
#> 106                                                    TTTT - Leather Finishing Operations (26)
#> 107                                                      TTTTT - Primary Magnesium Refining (9)
#> 108                           TTTTTT - Secondary Nonferrous Metals Processing Area Sources (20)
#> 109                                                        U - Group 1 Polymers And Resins (49)
#> 110                                                      UU - Equip Leaks-Control Level 2 (131)
#> 111                                      UUU - Petroleum Refineries: Ccu's, Cru's & Sru's (217)
#> 112                                                 UUUU - Cellulose Product Manufacturing (26)
#> 113                                                      UUUUU - Coal & Oil-Fired Eusgu's (299)
#> 114                                              VV - Oil-Water & Organic-Water Separators (63)
#> 115                                                   VVV - Publicly Owned Treatment Works (27)
#> 116                                                             VVVV - Boat Manufacturing (440)
#> 117                                         VVVVVV -  Chemical Manufacturing Area Sources (185)
#> 118                                     W - Epoxy Resins & Non-Nylon Polyamides Production (22)
#> 119                                          WW - Storage Vessels (Tanks)- Control Level 2 (65)
#> 120                                     WWWW - Reinforced Plastic Composites Production (1,255)
#> 121                                           WWWWW - Hospital Ethylene Oxide Sterilizers (250)
#> 122                                WWWWWW - Area Source: Plating And Polishing Operations (524)
#> 123                                                            X - Secondary Lead Smelters (41)
#> 124                            XX - Ethylene Mpf: Heat Exchange Systems & Waste Operations (33)
#> 125                                                             XXX - Ferroalloy Production (7)
#> 126                                                       XXXX - Rubber Tire Manufacturing (68)
#> 127          XXXXXX - Area Source: Nine Metal Fabrication And Finishing Source Categories (582)
#> 128                                 Y - Marine Tank Vessel Loading & Unloading Operations (308)
#> 129                                                           YY - Generic Mact Standards (152)
#> 130                                                 YYYY - Stationary Combustion Turbines (443)
#> 131                      YYYYY - Area Source: Electric Arc Furnace Steelmaking Facilities (119)
#> 132                                YYYYYY - Area Source: Ferroalloys Production Facilities (12)
#> 133                                                              ZZZ - Chemical Sector Rule (3)
#> 134                 ZZZZ - Stationary Reciprocating Internal Combustion Engines (Rice) (42,984)
#> 135                                         ZZZZZ - Iron And Steel Foundries Area Sources (325)
#> 136                    ZZZZZZ - Area Source: Aluminum, Copper & Other Nonferrous Foundries (62)
#>         N
#> 1    7958
#> 2      30
#> 3    1212
#> 4      90
#> 5      77
#> 6     187
#> 7      30
#> 8      67
#> 9    1930
#> 10      7
#> 11      1
#> 12    340
#> 13     86
#> 14      8
#> 15     55
#> 16   3505
#> 17     80
#> 18     10
#> 19    202
#> 20     17
#> 21    743
#> 22   4324
#> 23     30
#> 24    305
#> 25      2
#> 26     15
#> 27    263
#> 28    571
#> 29    174
#> 30      7
#> 31    479
#> 32    671
#> 33     35
#> 34      2
#> 35    539
#> 36    266
#> 37    146
#> 38    202
#> 39    259
#> 40    601
#> 41   8378
#> 42    205
#> 43     12
#> 44    284
#> 45   6904
#> 46     21
#> 47    107
#> 48    181
#> 49     61
#> 50    183
#> 51     19
#> 52     20
#> 53   1047
#> 54    138
#> 55    543
#> 56    191
#> 57   2081
#> 58    711
#> 59     70
#> 60     14
#> 61     58
#> 62     33
#> 63    233
#> 64     33
#> 65      5
#> 66   8791
#> 67    215
#> 68     53
#> 69   1375
#> 70     19
#> 71      9
#> 72   2453
#> 73     50
#> 74     36
#> 75     50
#> 76      8
#> 77    225
#> 78     65
#> 79     67
#> 80    226
#> 81    148
#> 82     29
#> 83     54
#> 84    663
#> 85     58
#> 86     60
#> 87    340
#> 88      9
#> 89    143
#> 90      9
#> 91     88
#> 92   1121
#> 93     32
#> 94    655
#> 95     57
#> 96     15
#> 97     16
#> 98    249
#> 99    150
#> 100   126
#> 101     8
#> 102    35
#> 103  2355
#> 104    20
#> 105     2
#> 106    26
#> 107     9
#> 108    20
#> 109    49
#> 110   131
#> 111   217
#> 112    26
#> 113   299
#> 114    63
#> 115    27
#> 116   440
#> 117   185
#> 118    22
#> 119    65
#> 120  1255
#> 121   250
#> 122   524
#> 123    41
#> 124    33
#> 125     7
#> 126    68
#> 127   582
#> 128   308
#> 129   152
#> 130   443
#> 131   119
#> 132    12
#> 133     3
#> 134 42984
#> 135   325
#> 136    62
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = FALSE) # default
#> 
#> Failed trying to connect to pins board server.
#> 
#> Arrow-format datasets (blocks, etc.) are up-to-date -- locally-installed and latest-released data repository versions match.
#> frs_by_mact   - is NOT in memory. Checking local disk... frs_by_mact  is loading from /home/runner/work/_temp/Library/EJAM/data/frs_by_mact.arrow ...done.
#> frs_by_mact   - was loaded from local folder, so server copy was not sought.
#> 
#> Key: <subpart, programid>
#>              programid subpart
#>                 <char>  <char>
#>  1:         0500100022    OOOO
#>  2:         06111R9011    OOOO
#>  3: 080000003800500001    OOOO
#>  4: 0900000006111R9011    OOOO
#>  5:         0900500003    OOOO
#>  6:         0901501138    OOOO
#>  7:         1318900002    OOOO
#>  8:         2013300027    OOOO
#>  9:         2512000290    OOOO
#> 10:         2512100102    OOOO
#> 11:         2512100335    OOOO
#> 12:         2808100088    OOOO
#> 13:         2918901012    OOOO
#> 14:         3301191265    OOOO
#> 15:         3501500644    OOOO
#> 16:         3501500871    OOOO
#> 17:         3504500523    OOOO
#> 18:         3504500524    OOOO
#> 19:         3703500493    OOOO
#> 20:         3710100014    OOOO
#> 21:         3716100164    OOOO
#> 22:         3903100003    OOOO
#> 23:         4000300023    OOOO
#> 24:         4001700142    OOOO
#> 25:         4001900218    OOOO
#> 26:         4004300085    OOOO
#> 27:         4004500317    OOOO
#> 28:         4005100227    OOOO
#> 29:         4005300123    OOOO
#> 30:         4012900154    OOOO
#> 31:         4012900238    OOOO
#> 32:         4012900282    OOOO
#> 33:         4015100280    OOOO
#> 34:         4204100072    OOOO
#> 35:         4400900010    OOOO
#> 36:         4500300027    OOOO
#> 37:         4500700011    OOOO
#> 38:         4500700032    OOOO
#> 39:         4502500002    OOOO
#> 40:         4504500029    OOOO
#> 41:         4504500060    OOOO
#> 42:         4507700040    OOOO
#> 43:         4508500003    OOOO
#> 44:         4706300275    OOOO
#> 45:         4716300007    OOOO
#> 46:         5401700061    OOOO
#> 47:         5401700065    OOOO
#> 48:         5401700067    OOOO
#> 49:         5401700074    OOOO
#> 50:         5408500026    OOOO
#> 51:         5410300044    OOOO
#> 52:         5410300074    OOOO
#> 53: AR0000000500100022    OOOO
#> 54: CT0000000900500003    OOOO
#> 55: CT0000000901501138    OOOO
#> 56: GA0000001318900002    OOOO
#> 57: KS0000002013300027    OOOO
#> 58: MA0000002512000290    OOOO
#> 59: MA0000002512100102    OOOO
#> 60: MA0000002512100335    OOOO
#> 61: MI00000000000P0634    OOOO
#> 62: MO0000002918901012    OOOO
#> 63: MS0000002808100088    OOOO
#> 64: MT0000003007500005    OOOO
#> 65: NC0000003701800493    OOOO
#> 66: NC0000003705100014    OOOO
#> 67: NC0000003708100164    OOOO
#> 68: NH0000003301100129    OOOO
#> 69: NY0000006301300065    OOOO
#> 70: OH0000000616010036    OOOO
#> 71: OH0000001667060088    OOOO
#> 72:        PA000492043    OOOO
#> 73:        PA000743056    OOOO
#> 74: RI0000004400900010    OOOO
#> 75:      SC00000800027    OOOO
#> 76:      SC00002000032    OOOO
#> 77:      SC00012000029    OOOO
#> 78:      SC00018800040    OOOO
#> 79: TN0000004706300275    OOOO
#> 80: TN0000004716300007    OOOO
#> 81: VA0000005115500016    OOOO
#> 82: WI0000005502500072    OOOO
#>              programid subpart
#>                                                          title
#>                                                         <char>
#>  1: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  2: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  3: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  4: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  5: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  6: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  7: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  8: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  9: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 10: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 11: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 12: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 13: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 14: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 15: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 16: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 17: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 18: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 19: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 20: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 21: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 22: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 23: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 24: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 25: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 26: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 27: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 28: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 29: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 30: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 31: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 32: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 33: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 34: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 35: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 36: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 37: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 38: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 39: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 40: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 41: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 42: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 43: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 44: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 45: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 46: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 47: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 48: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 49: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 50: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 51: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 52: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 53: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 54: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 55: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 56: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 57: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 58: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 59: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 60: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 61: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 62: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 63: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 64: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 65: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 66: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 67: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 68: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 69: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 70: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 71: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 72: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 73: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 74: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 75: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 76: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 77: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 78: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 79: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 80: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 81: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 82: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>                                                          title
#>                                                        dropdown_label      lat
#>                                                                <char>    <num>
#>  1: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.27496
#>  2: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.22082
#>  3: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 47.97940
#>  4: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.22082
#>  5: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.61256
#>  6: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.66805
#>  7: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 33.44801
#>  8: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 37.68405
#>  9: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.60689
#> 10: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.72361
#> 11: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.59238
#> 12: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.27973
#> 13: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 38.76637
#> 14: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.79838
#> 15: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 32.23500
#> 16: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 32.21906
#> 17: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.72734
#> 18: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.72056
#> 19: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.72438
#> 20: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.36954
#> 21: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.25610
#> 22: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.26201
#> 23: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.85323
#> 24: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.45003
#> 25: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.25895
#> 26: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.85628
#> 27: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.14478
#> 28: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.96318
#> 29: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.79593
#> 30: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.79897
#> 31: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.61106
#> 32: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.85616
#> 33: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.94184
#> 34: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.20903
#> 35: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.44594
#> 36: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 33.56057
#> 37: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.65913
#> 38: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.61164
#> 39: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.69462
#> 40: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.01979
#> 41: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.02015
#> 42: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.78693
#> 43: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 33.91990
#> 44: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.16847
#> 45: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.56050
#> 46: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.27623
#> 47: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.17839
#> 48: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.31809
#> 49: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.15674
#> 50: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.18746
#> 51: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.44221
#> 52: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 39.27900
#> 53: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.27496
#> 54: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.61256
#> 55: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.66805
#> 56: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 33.44801
#> 57: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 37.68405
#> 58: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.60689
#> 59: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.72361
#> 60: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.59238
#> 61: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.88929
#> 62: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 38.76637
#> 63: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.27973
#> 64: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 45.11306
#> 65: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.72438
#> 66: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.36954
#> 67: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.25610
#> 68: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 42.75050
#> 69: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 43.21786
#> 70: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.26201
#> 71: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.16331
#> 72: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.21077
#> 73: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 40.74333
#> 74: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 41.44594
#> 75: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 33.56057
#> 76: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.61164
#> 77: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 35.01979
#> 78: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.78693
#> 79: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.16847
#> 80: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 36.56050
#> 81: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 37.04957
#> 82: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 43.11747
#>                                                        dropdown_label      lat
#>            lon  REGISTRY_ID         PGM_SYS_ACRNMS  program
#>          <num>       <char>                 <char>   <char>
#>  1:  -91.34179 110025082924    AIRS/AFS:0500100022 AIRS/AFS
#>  2: -119.02272 110043415578    AIRS/AFS:06111R9011 AIRS/AFS
#>  3:  -99.02064 110000825867 AIR:080000003800500001      AIR
#>  4: -119.02272 110043415578 AIR:0900000006111R9011      AIR
#>  5:  -73.08986 110039211064    AIRS/AFS:0900500003 AIRS/AFS
#>  6:  -71.91848 110001732554    AIRS/AFS:0901501138 AIRS/AFS
#>  7:  -82.53785 110000359406    AIRS/AFS:1318900002 AIRS/AFS
#>  8:  -95.46691 110000445938    AIRS/AFS:2013300027 AIRS/AFS
#>  9:  -70.90407 110000739675    AIRS/AFS:2512000290 AIRS/AFS
#> 10:  -70.96296 110000309773    AIRS/AFS:2512100102 AIRS/AFS
#> 11:  -71.28349 110000310020    AIRS/AFS:2512100335 AIRS/AFS
#> 12:  -88.76446 110013841432    AIRS/AFS:2808100088 AIRS/AFS
#> 13:  -90.47491 110000440381    AIRS/AFS:2918901012 AIRS/AFS
#> 14:  -71.53392 110023083973    AIRS/AFS:3301191265 AIRS/AFS
#> 15: -103.91583 110046351620    AIRS/AFS:3501500644 AIRS/AFS
#> 16: -103.81400 110056286184    AIRS/AFS:3501500871 AIRS/AFS
#> 17: -107.98155 110055609569    AIRS/AFS:3504500523 AIRS/AFS
#> 18: -107.97539 110055609578    AIRS/AFS:3504500524 AIRS/AFS
#> 19:  -81.37813 110000350584    AIRS/AFS:3703500493 AIRS/AFS
#> 20:  -78.55103 110001478730    AIRS/AFS:3710100014 AIRS/AFS
#> 21:  -81.77920 110013792921    AIRS/AFS:3716100164 AIRS/AFS
#> 22:  -81.86949 110000384753    AIRS/AFS:3903100003 AIRS/AFS
#> 23:  -98.35850 110007163577    AIRS/AFS:4000300023 AIRS/AFS
#> 24:  -98.05572 110054881277    AIRS/AFS:4001700142 AIRS/AFS
#> 25:  -96.95461 110058902166    AIRS/AFS:4001900218 AIRS/AFS
#> 26:  -99.10832 110046386665    AIRS/AFS:4004300085 AIRS/AFS
#> 27:  -99.83113 110056436735    AIRS/AFS:4004500317 AIRS/AFS
#> 28:  -97.78009 110058908357    AIRS/AFS:4005100227 AIRS/AFS
#> 29:  -98.03599 110046270352    AIRS/AFS:4005300123 AIRS/AFS
#> 30:  -99.93188 110044303198    AIRS/AFS:4012900154 AIRS/AFS
#> 31:  -99.72453 110054889698    AIRS/AFS:4012900238 AIRS/AFS
#> 32:  -99.47094 110055615927    AIRS/AFS:4012900282 AIRS/AFS
#> 33:  -98.91438 110046385498    AIRS/AFS:4015100280 AIRS/AFS
#> 34:  -77.02432 110012697812    AIRS/AFS:4204100072 AIRS/AFS
#> 35:  -71.62625 110000312117    AIRS/AFS:4400900010 AIRS/AFS
#> 36:  -81.80997 110031367036    AIRS/AFS:4500300027 AIRS/AFS
#> 37:  -82.80460 110000354811    AIRS/AFS:4500700011 AIRS/AFS
#> 38:  -82.47996 110001664823    AIRS/AFS:4500700032 AIRS/AFS
#> 39:  -79.91536 110002232922    AIRS/AFS:4502500002 AIRS/AFS
#> 40:  -82.56309 110041625989    AIRS/AFS:4504500029 AIRS/AFS
#> 41:  -82.56041 110001666554    AIRS/AFS:4504500060 AIRS/AFS
#> 42:  -82.68964 110001667802    AIRS/AFS:4507700040 AIRS/AFS
#> 43:  -80.31968 110000352305    AIRS/AFS:4508500003 AIRS/AFS
#> 44:  -83.39178 110038045139    AIRS/AFS:4706300275 AIRS/AFS
#> 45:  -82.17101 110000913548    AIRS/AFS:4716300007 AIRS/AFS
#> 46:  -80.69078 110061076455    AIRS/AFS:5401700061 AIRS/AFS
#> 47:  -80.70898 110061070889    AIRS/AFS:5401700065 AIRS/AFS
#> 48:  -80.69992 110061070898    AIRS/AFS:5401700067 AIRS/AFS
#> 49:  -80.69571 110061070905    AIRS/AFS:5401700074 AIRS/AFS
#> 50:  -80.95438 110056440506    AIRS/AFS:5408500026 AIRS/AFS
#> 51:  -80.00077 110043703258    AIRS/AFS:5410300044 AIRS/AFS
#> 52:  -80.33819 110058908847    AIRS/AFS:5410300074 AIRS/AFS
#> 53:  -91.34179 110025082924 AIR:AR0000000500100022      AIR
#> 54:  -73.08986 110039211064 AIR:CT0000000900500003      AIR
#> 55:  -71.91848 110001732554 AIR:CT0000000901501138      AIR
#> 56:  -82.53785 110000359406 AIR:GA0000001318900002      AIR
#> 57:  -95.46691 110000445938 AIR:KS0000002013300027      AIR
#> 58:  -70.90407 110000739675 AIR:MA0000002512000290      AIR
#> 59:  -70.96296 110000309773 AIR:MA0000002512100102      AIR
#> 60:  -71.28349 110000310020 AIR:MA0000002512100335      AIR
#> 61:  -85.63853 110067012878 AIR:MI00000000000P0634      AIR
#> 62:  -90.47491 110000440381 AIR:MO0000002918901012      AIR
#> 63:  -88.76446 110013841432 AIR:MS0000002808100088      AIR
#> 64: -105.08569 110046594146 AIR:MT0000003007500005      AIR
#> 65:  -81.37813 110000350584 AIR:NC0000003701800493      AIR
#> 66:  -78.55103 110001478730 AIR:NC0000003705100014      AIR
#> 67:  -81.77920 110013792921 AIR:NC0000003708100164      AIR
#> 68:  -71.82836 110001522184 AIR:NH0000003301100129      AIR
#> 69:  -75.46524 110000325871 AIR:NY0000006301300065      AIR
#> 70:  -81.86949 110000384753 AIR:OH0000000616010036      AIR
#> 71:  -81.25312 110000766644 AIR:OH0000001667060088      AIR
#> 72:  -77.02111 110071722797        AIR:PA000492043      AIR
#> 73:  -79.40278 110046273661        AIR:PA000743056      AIR
#> 74:  -71.62625 110000312117 AIR:RI0000004400900010      AIR
#> 75:  -81.80997 110031367036      AIR:SC00000800027      AIR
#> 76:  -82.47996 110001664823      AIR:SC00002000032      AIR
#> 77:  -82.56309 110041625989      AIR:SC00012000029      AIR
#> 78:  -82.68964 110001667802      AIR:SC00018800040      AIR
#> 79:  -83.39178 110038045139 AIR:TN0000004706300275      AIR
#> 80:  -82.17101 110000913548 AIR:TN0000004716300007      AIR
#> 81:  -80.75514 110000343469 AIR:VA0000005115500016      AIR
#> 82:  -89.50733 110000419878 AIR:WI0000005502500072      AIR
#>            lon  REGISTRY_ID         PGM_SYS_ACRNMS  program
  latlon_from_mactsubpart("OOOO", include_if_no_latlon = TRUE)
#> Key: <subpart, programid>
#>               programid subpart
#>                  <char>  <char>
#>   1:         0401300983    OOOO
#>   2:         0401301341    OOOO
#>   3:         0500100022    OOOO
#>   4:         0514100401    OOOO
#>   5:         06111R9011    OOOO
#>  ---                           
#> 222: VA0000005115500016    OOOO
#> 223: VA0000005116300001    OOOO
#> 224: VA0000005159000002    OOOO
#> 225: VA0000005170000091    OOOO
#> 226: WI0000005502500072    OOOO
#>                                                           title
#>                                                          <char>
#>   1: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>   2: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>   3: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>   4: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>   5: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>  ---                                                           
#> 222: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 223: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 224: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 225: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#> 226: PRINTING, COATING AND DYEING OF FABRICS AND OTHER TEXTILES
#>                                                         dropdown_label      lat
#>                                                                 <char>    <num>
#>   1: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#>   2: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#>   3: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.27496
#>   4: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#>   5: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 34.22082
#>  ---                                                                           
#> 222: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 37.04957
#> 223: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#> 224: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#> 225: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles       NA
#> 226: OOOO - Printing, Coating And Dyeing Of Fabrics And Other Textiles 43.11747
#>             lon  REGISTRY_ID         PGM_SYS_ACRNMS  program
#>           <num>       <char>                 <char>   <char>
#>   1:         NA         <NA>                   <NA>     <NA>
#>   2:         NA         <NA>                   <NA>     <NA>
#>   3:  -91.34179 110025082924    AIRS/AFS:0500100022 AIRS/AFS
#>   4:         NA         <NA>                   <NA>     <NA>
#>   5: -119.02272 110043415578    AIRS/AFS:06111R9011 AIRS/AFS
#>  ---                                                        
#> 222:  -80.75514 110000343469 AIR:VA0000005115500016      AIR
#> 223:         NA         <NA>                   <NA>     <NA>
#> 224:         NA         <NA>                   <NA>     <NA>
#> 225:         NA         <NA>                   <NA>     <NA>
#> 226:  -89.50733 110000419878 AIR:WI0000005502500072      AIR