Get lat lon, Registry ID, and NAICS, for given FRS Program System ID

Get lat lon, Registry ID, and NAICS, for given FRS Program System ID

Usage

latlon_from_programid(programid)

Arguments

programid

like "XJW000012435"

Value

data.table with lat lon REGISTRY_ID program pgm_sys_id

Details

The ID is the identification number, such as the permit number, assigned by an information management system that represents a facility site, waste site, operable unit, or other feature tracked by that Environmental Information System.

Also note the FRS API: https://www.epa.gov/frs/facility-registry-service-frs-api https://www.epa.gov/frs/frs-rest-services

Examples

 latlon_from_programid(c("XJW000012435", "00768SRTRSROAD1"))
#>         lat        lon  REGISTRY_ID  program      pgm_sys_id
#> 1: 48.39008 -122.50088 110071133925 RCRAINFO    XJW000012435
#> 2: 18.06229  -66.85967 110000307855     TRIS 00768SRTRSROAD1
 pids <- c("7-0540-00003", "354362", "1513529", "485659", "LAG750956", 
   "CAC002995519", "3601252181", "3601439158")
 latlon_from_regid(latlon_from_programid(pids)[,REGISTRY_ID])
#>         lat        lon  REGISTRY_ID                  PRIMARY_NAME NAICS  SIC        PGM_SYS_ACRNMS
#> 1: 43.00028  -76.69323 110019033810 HEJAMADA CAMPGROUND & RV PARK       8999      FIS:7-0540-00003
#> 2: 38.22678 -121.02361 110066714951      CAMANCHE DAM POWER HOUSE       4911  CA-ENVIROVIEW:354362
#> 3: 41.38786  -73.43935 110044340807            NEJAME & SONS INC.                     SIMS:1513529
#> 4: 40.77499  -74.15929 110056111559      LA VEJAMA DISCOUNT STORE                  NJ-NJEMS:485659
#> 5: 30.25468  -90.94284 110056982323           GRACEJAMES ENT, LLC       7542       NPDES:LAG750956
#> 6: 32.80406 -117.02753 110070538057             BEJAMIN HAFERTEPE            RCRAINFO:CAC002995519
#> 7: 33.66110 -117.86959 110070874073                     EJAM INC.                  ICIS:3601252181
#> 8: 33.70369 -117.86675 110071293460                      EJAM INC                  ICIS:3601439158
 latlon_from_programid(c("XJW000012435", "00768SRTRSROAD1", "asdfsdf"))[,.(lat,lon)]
#>         lat        lon
#> 1: 48.39008 -122.50088
#> 2: 18.06229  -66.85967
#> 3:       NA         NA