Map - points - Create leaflet html widget map of points using table with lat lon

Map - points - Create leaflet html widget map of points using table with lat lon

Usage

mapfast(mydf, radius = 3, column_names = "all", labels = column_names)

Arguments

mydf

data.frame or data.table with lat and lon columns or columns that latlon_infer() can infer to be that

radius

in miles, converted to meters and passed to leaflet::addCircles()

column_names

If "ej" then nice popup made based on just key EJScreen indicators. If "all" then every column in the entire mydf table is shown in the popup. If a vector of colnames, only those are shown in popups.

labels

The labels used before the column_names, for map popups, like label: column_name (ignored if column_names is ej or all)

Value

plots a leaflet map with popups with all the columns from mydf, and returns html widget

See also

Examples

 mapfast(testpoints_500)
 if (FALSE) {
 
 mapfastej(testoutput_ejamit_1000pts_1miles$results_bysite, radius = 1)
 
 ejam2map(testoutput_ejamit_1000pts_1miles)
 
 # see random sample of 1k FRS facilities
 if (!exists("frs")) {
   x <- try(require(EJAM)); if (!inherits(x, "try-error")) {
     dataload_from_pins("frs")
   }
   mydf <-  frs[sample(1:NROW( frs), 1000), 1:5]
   mapfast(mydf)
 }
 # out <- EJAMejscreenapi::testoutput_ejscreenit_50$table
 # names(out) <- fixcolnames(names(out), 'long', 'r')
 out <- testoutput_ejscreenapi_plus_50
 mapfastej(out)
 mapfast(out, column_names = 'ej')
 mapfast(out)

 # Save .html file and view it in your browser

 mytable <- testoutput_ejscreenit_5$table
 # names(mytable) <- fixcolnames(names(mytable), 'long', 'r')
 ## now, that renaming is done by mapfast function
 mymap <- mapfastej(mytable)
 tfile <- paste0(tempfile(), '.html')
 htmltools::save_html(mymap, file = tfile)
 browseURL(tfile)
  }