helper function in updating the package metadata

Quick and dirty helper during development, to check all the attributes of all the data files in relevant packages. It loads unloaded packages as needed, which you might not want it to do, but it is not coded to be able to check attributes without doing that.

metadata_check(
  packages = EJAM::ejampackages,
  which = c("census_version", "acs_version", "acs_releasedate", "ACS",
    "ejscreen_version", "ejscreen_releasedate", "ejscreen_pkg_data", "year", "released"),
  loadifnotloaded = TRUE
)

Arguments

packages

Optional. e.g. 'EJAMejscreendata', or can be a vector of character strings, and if not specified, default is to report on all packages with EJ as part of their name, like EJAMejscreenapi

which

Optional vector (not list) of strings, the attributes. Default is some typical ones used in EJAM-related packages currently.

loadifnotloaded

Optional to control if func should temporarily attach packages not already loaded.