NAICS - See the names of industrial categories and their NAICS code

Easy way to list the 2-digit NAICS (17 categories), or other level

Usage

naics_categories(digits = 2, dataset = EJAM::NAICS)

Arguments

digits

default is 2, for 2-digits NAICS, the top level, but could be up to 6.

dataset

Should default to the dataset called NAICS, installed with this package. see NAICS Check attr(NAICS, 'year')

Value

matrix with 1 column of 2-digit codes and rownames that look like "22 - Utilities" etc.

Details

Also see https://www.naics.com/search/

There are this many NAICS codes roughly by number of digits in the code:

table(nchar(NAICS))

2 3 4 5 6

17 99 311 709 1057

See https://www.census.gov/naics/

See also

Examples

 naics_categories()
#> Also see https://www.naics.com/search/ 
#>                                                                               [,1]
#> 11 - Agriculture, Forestry, Fishing and Hunting                                 11
#> 21 - Mining, Quarrying, and Oil and Gas Extraction                              21
#> 22 - Utilities                                                                  22
#> 23 - Construction                                                               23
#> 42 - Wholesale Trade                                                            42
#> 51 - Information                                                                51
#> 52 - Finance and Insurance                                                      52
#> 53 - Real Estate and Rental and Leasing                                         53
#> 54 - Professional, Scientific, and Technical Services                           54
#> 55 - Management of Companies and Enterprises                                    55
#> 56 - Administrative and Support and Waste Management and Remediation Services   56
#> 61 - Educational Services                                                       61
#> 62 - Health Care and Social Assistance                                          62
#> 71 - Arts, Entertainment, and Recreation                                        71
#> 72 - Accommodation and Food Services                                            72
#> 81 - Other Services (except Public Administration)                              81
#> 92 - Public Administration                                                      92
#> 31 - Manufacturing                                                              31
#> 32 - Manufacturing                                                              32
#> 33 - Manufacturing                                                              33
#> 44 - Retail Trade                                                               44
#> 45 - Retail Trade                                                               45
#> 48 - Transportation and Warehousing                                             48
#> 49 - Transportation and Warehousing                                             49