See which of the lists of names a single term appears in

See which of the lists of names a single term appears in

names_whichlist(
  x,
  exact = T,
  grepmatching = T,
  ignore.case.exact = FALSE,
  ignore.case.grep = FALSE,
  keylists = F,
  exactonly = FALSE
)

Arguments

x

term, like part or all of a variable name, such as state.avg

exact

whether to look for exact matches

grepmatching

whether to look for matches via grep (partial match)

ignore.case.exact

whether to ignore capitalization in exact matches

ignore.case.grep

passed to grep as ignore.case param

keylists

if true, only report for the key lists not friendly, all, these, need types.

exactonly

to limit output to rows with exact matches

Value

a data.frame of whichlist, exactmatch, grepmatch, and grephits (examples)

Details

EJAM::namez has a list of lists of names used for indicators or variables, such as namez$d_friendly which is a vector of terms like "Demog.Ind.", "Suppl Demog Index", "% Low-inc.", etc.

Examples

 
   x <- names_whichlist("rsei", ignore.case.exact = T, ignore.case.grep = T)
#> Error in names_whichlist("rsei", ignore.case.exact = T, ignore.case.grep = T): could not find function "names_whichlist"
   subset(x, !grepl("friendly", x$whichlist))
#> Error in subset(x, !grepl("friendly", x$whichlist)): object 'x' not found
   
   subset(x,  grepl("friendly", x$whichlist))
#> Error in subset(x, grepl("friendly", x$whichlist)): object 'x' not found
   subset(namez, names(namez) != "all_r" & names(namez) %in%
      subset(x, x$grepmatch == "yes" & !grepl("friendly", x$whichlist))$whichlist  )
#> Error in subset(x, x$grepmatch == "yes" & !grepl("friendly", x$whichlist)): object 'x' not found
   grep("\\.eo$", namez$ej, value = T)
#>  [1] "EJ.DISPARITY.pm.eo"              "EJ.DISPARITY.o3.eo"             
#>  [3] "EJ.DISPARITY.cancer.eo"          "EJ.DISPARITY.resp.eo"           
#>  [5] "EJ.DISPARITY.dpm.eo"             "EJ.DISPARITY.pctpre1960.eo"     
#>  [7] "EJ.DISPARITY.traffic.score.eo"   "EJ.DISPARITY.proximity.npl.eo"  
#>  [9] "EJ.DISPARITY.proximity.rmp.eo"   "EJ.DISPARITY.proximity.tsdf.eo" 
#> [11] "EJ.DISPARITY.proximity.npdes.eo" "EJ.DISPARITY.ust.eo"            
#> [13] "EJ.DISPARITY.rsei.eo"