Get indicator names within a varlist like names_d

Get indicator names within a varlist like names_d

Usage

namesbyvarlist(varlist, nametype = c("rname", "longname", "apiname")[1])

Arguments

varlist

one character string like "names_d", or a vector of them

nametype

vector of 1 or more names of columns in map_headernames, or a shortcut type that can be api, csv, r, original, long, shortlabel

Value

a data.frame one row per indicator, one col per nametype and a column identifying the varlist

Details

varlist2names() is a way to just get a vector of variable names even if the varlist is not stored as a separate data object and is only found in the map_headernames$varlist column:

varlist2names(c('names_d', 'names_d_subgroups'))

c(names_d, names_d_subgroups)

See also

Examples

 unique(map_headernames$varlist)
#>  [1] "names_d"                                   
#>  [2] "names_d_demogindexstate"                   
#>  [3] "names_d_ratio_to_avg"                      
#>  [4] "names_d_ratio_to_state_avg"                
#>  [5] "names_d_pctile"                            
#>  [6] "names_d_state_pctile"                      
#>  [7] "names_d_avg"                               
#>  [8] "names_d_state_avg"                         
#>  [9] "names_d_count"                             
#> [10] "names_d_subgroups"                         
#> [11] "names_d_subgroups_ratio_to_avg"            
#> [12] "names_d_subgroups_ratio_to_state_avg"      
#> [13] "names_d_subgroups_pctile"                  
#> [14] "names_d_subgroups_state_pctile"            
#> [15] "names_d_subgroups_avg"                     
#> [16] "names_d_subgroups_state_avg"               
#> [17] "names_d_subgroups_count"                   
#> [18] "names_d_subgroups_alone"                   
#> [19] "names_d_subgroups_alone_ratio_to_avg"      
#> [20] "names_d_subgroups_alone_ratio_to_state_avg"
#> [21] "names_d_subgroups_alone_pctile"            
#> [22] "names_d_subgroups_alone_state_pctile"      
#> [23] "names_d_subgroups_alone_avg"               
#> [24] "names_d_subgroups_alone_state_avg"         
#> [25] "names_d_subgroups_alone_count"             
#> [26] "names_d_subgroups_nh"                      
#> [27] "names_d_subgroups_nh_ratio_to_avg"         
#> [28] "names_d_subgroups_nh_ratio_to_state_avg"   
#> [29] "names_d_subgroups_nh_pctile"               
#> [30] "names_d_subgroups_nh_state_pctile"         
#> [31] "names_d_subgroups_nh_avg"                  
#> [32] "names_d_subgroups_nh_state_avg"            
#> [33] "names_d_subgroups_nh_count"                
#> [34] "names_d_language"                          
#> [35] "names_d_language_count"                    
#> [36] "names_d_languageli"                        
#> [37] "names_d_languageli_count"                  
#> [38] "names_d_other_count"                       
#> [39] "names_d_extra"                             
#> [40] "names_d_extra_count"                       
#> [41] "names_health"                              
#> [42] "names_health_ratio_to_avg"                 
#> [43] "names_health_ratio_to_state_avg"           
#> [44] "names_health_pctile"                       
#> [45] "names_health_state_pctile"                 
#> [46] "names_health_avg"                          
#> [47] "names_health_state_avg"                    
#> [48] "names_health_count"                        
#> [49] "names_criticalservice"                     
#> [50] "names_misc"                                
#> [51] "names_e"                                   
#> [52] "names_e_ratio_to_avg"                      
#> [53] "names_e_ratio_to_state_avg"                
#> [54] "names_e_pctile"                            
#> [55] "names_e_state_pctile"                      
#> [56] "names_e_avg"                               
#> [57] "names_e_state_avg"                         
#> [58] "names_climate"                             
#> [59] "names_climate_pctile"                      
#> [60] "names_climate_state_pctile"                
#> [61] "names_climate_avg"                         
#> [62] "names_climate_state_avg"                   
#> [63] "names_e_other"                             
#> [64] "names_ej"                                  
#> [65] "names_ej_state"                            
#> [66] "names_ej_pctile"                           
#> [67] "names_ej_state_pctile"                     
#> [68] "names_ej_supp"                             
#> [69] "names_ej_supp_state"                       
#> [70] "names_ej_supp_pctile"                      
#> [71] "names_ej_supp_state_pctile"                
#> [72] "names_countabove"                          
#> [73] "names_criticalservice_pctile"              
#> [74] "names_criticalservice_state_pctile"        
#> [75] "names_criticalservice_avg"                 
#> [76] "names_criticalservice_state_avg"           
#> [77] "names_featuresinarea"                      
#> [78] "names_age"                                 
#> [79] "names_age_count"                           
#> [80] "names_community"                           
#> [81] "names_community_count"                     
#> [82] "names_flag"                                
#> [83] "custom"                                    
#> [84] "names_sitesinarea"                         
#> [85] "names_geo"                                 
 
 namesbyvarlist('names_e_avg', 'rname')
#>         varlist               rname
#> 412 names_e_avg              avg.pm
#> 413 names_e_avg              avg.o3
#> 414 names_e_avg             avg.no2
#> 415 names_e_avg             avg.dpm
#> 416 names_e_avg            avg.rsei
#> 417 names_e_avg   avg.traffic.score
#> 418 names_e_avg      avg.pctpre1960
#> 419 names_e_avg   avg.proximity.npl
#> 420 names_e_avg   avg.proximity.rmp
#> 421 names_e_avg  avg.proximity.tsdf
#> 422 names_e_avg             avg.ust
#> 423 names_e_avg avg.proximity.npdes
#> 424 names_e_avg        avg.drinking
 namesbyvarlist('names_d')
#>   varlist            rname
#> 1 names_d      Demog.Index
#> 2 names_d Demog.Index.Supp
#> 3 names_d        pctlowinc
#> 4 names_d       pctlingiso
#> 5 names_d    pctunemployed
#> 6 names_d          pctlths
#> 7 names_d        pctunder5
#> 8 names_d        pctover64
#> 9 names_d           pctmin
 namesbyvarlist('names_d', 'r')
#>   varlist            rname
#> 1 names_d      Demog.Index
#> 2 names_d Demog.Index.Supp
#> 3 names_d        pctlowinc
#> 4 names_d       pctlingiso
#> 5 names_d    pctunemployed
#> 6 names_d          pctlths
#> 7 names_d        pctunder5
#> 8 names_d        pctover64
#> 9 names_d           pctmin
 namesbyvarlist('names_d', 'long')
#>   varlist                                 longname
#> 1 names_d                    Demographic Index USA
#> 2 names_d       Supplemental Demographic Index USA
#> 3 names_d                             % Low Income
#> 4 names_d % in limited English-speaking Households
#> 5 names_d                             % Unemployed
#> 6 names_d   % with Less Than High School Education
#> 7 names_d                            % under Age 5
#> 8 names_d                            % over Age 64
#> 9 names_d                        % People of Color
 namesbyvarlist('names_d', 'shortlabel')
#>   varlist       shortlabel
#> 1 names_d       Demog.Ind.
#> 2 names_d Suppl Demog.Ind.
#> 3 names_d        %Low-inc.
#> 4 names_d %Limited English
#> 5 names_d      %Unemployed
#> 6 names_d   %< High School
#> 7 names_d         %< age 5
#> 8 names_d        %> age 64
#> 9 names_d             %POC
 
 namesbyvarlist( 'names_e_pctile', c('r', 'longname'))    
#>            varlist                  rname
#> 386 names_e_pctile              pctile.pm
#> 387 names_e_pctile              pctile.o3
#> 388 names_e_pctile             pctile.no2
#> 389 names_e_pctile             pctile.dpm
#> 390 names_e_pctile            pctile.rsei
#> 391 names_e_pctile   pctile.traffic.score
#> 392 names_e_pctile      pctile.pctpre1960
#> 393 names_e_pctile   pctile.proximity.npl
#> 394 names_e_pctile   pctile.proximity.rmp
#> 395 names_e_pctile  pctile.proximity.tsdf
#> 396 names_e_pctile             pctile.ust
#> 397 names_e_pctile pctile.proximity.npdes
#> 398 names_e_pctile        pctile.drinking
#>                                                                                        longname
#> 386                                      US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 387                                                               US percentile for Ozone (ppb)
#> 388                                                    US percentile for Nitrogen Dioxide (NO2)
#> 389                                         US percentile for Diesel Particulate Matter (ug/m3)
#> 390                                                     US percentile for Toxic Releases to Air
#> 391       US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 392                                US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 393                              US percentile for Superfund Proximity (site count/km distance)
#> 394                                US percentile for RMP Proximity (facility count/km distance)
#> 395                    US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 396                                 US percentile for Underground Storage Tanks (UST) indicator
#> 397 US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 398                                             US percentile for Drinking Water Non-Compliance
 namesbyvarlist(c('names_e_pctile', 'names_e_state_pctile'), 
   c('varlist', 'rname', 'apiname', 'csvname', 'shortlabel', 'longname'))
#>                  varlist            varlist.1                        rname
#> 386       names_e_pctile       names_e_pctile                    pctile.pm
#> 387       names_e_pctile       names_e_pctile                    pctile.o3
#> 388       names_e_pctile       names_e_pctile                   pctile.no2
#> 389       names_e_pctile       names_e_pctile                   pctile.dpm
#> 390       names_e_pctile       names_e_pctile                  pctile.rsei
#> 391       names_e_pctile       names_e_pctile         pctile.traffic.score
#> 392       names_e_pctile       names_e_pctile            pctile.pctpre1960
#> 393       names_e_pctile       names_e_pctile         pctile.proximity.npl
#> 394       names_e_pctile       names_e_pctile         pctile.proximity.rmp
#> 395       names_e_pctile       names_e_pctile        pctile.proximity.tsdf
#> 396       names_e_pctile       names_e_pctile                   pctile.ust
#> 397       names_e_pctile       names_e_pctile       pctile.proximity.npdes
#> 398       names_e_pctile       names_e_pctile              pctile.drinking
#> 399 names_e_state_pctile names_e_state_pctile              state.pctile.pm
#> 400 names_e_state_pctile names_e_state_pctile              state.pctile.o3
#> 401 names_e_state_pctile names_e_state_pctile             state.pctile.no2
#> 402 names_e_state_pctile names_e_state_pctile             state.pctile.dpm
#> 403 names_e_state_pctile names_e_state_pctile            state.pctile.rsei
#> 404 names_e_state_pctile names_e_state_pctile   state.pctile.traffic.score
#> 405 names_e_state_pctile names_e_state_pctile      state.pctile.pctpre1960
#> 406 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.npl
#> 407 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.rmp
#> 408 names_e_state_pctile names_e_state_pctile  state.pctile.proximity.tsdf
#> 409 names_e_state_pctile names_e_state_pctile             state.pctile.ust
#> 410 names_e_state_pctile names_e_state_pctile state.pctile.proximity.npdes
#> 411 names_e_state_pctile names_e_state_pctile        state.pctile.drinking
#>              apiname      csvname                      shortlabel
#> 386     N_E_PM25_PER       P_PM25                    US%ile PM2.5
#> 387       N_E_O3_PER      P_OZONE                    US%ile Ozone
#> 388      N_E_NO2_PER        P_NO2                      US%ile NO2
#> 389   N_E_DIESEL_PER      P_DSLPM                US%ile Diesel PM
#> 390 N_E_RSEI_AIR_PER   P_RSEI_AIR    US%ile Toxic Releases to Air
#> 391  N_E_TRAFFIC_PER      P_PTRAF                  US%ile Traffic
#> 392     N_E_LEAD_PER      P_LDPNT                US%ile %pre-1960
#> 393      N_E_NPL_PER       P_PNPL                      US%ile NPL
#> 394      N_E_RMP_PER       P_PRMP                      US%ile RMP
#> 395     N_E_TSDF_PER      P_PTSDF                     US%ile TSDF
#> 396      N_E_UST_PER        P_UST                      US%ile UST
#> 397    N_E_NPDES_PER      P_PWDIS                    US%ile NPDES
#> 398   N_E_DWATER_PER     P_DWATER                 US%ile Drinking
#> 399     S_E_PM25_PER     S_P_PM25                 State%ile PM2.5
#> 400       S_E_O3_PER    S_P_OZONE                 State%ile Ozone
#> 401      S_E_NO2_PER      S_P_NO2                   State%ile NO2
#> 402   S_E_DIESEL_PER    S_P_DSLPM             State%ile Diesel PM
#> 403 S_E_RSEI_AIR_PER S_P_RSEI_AIR State%ile Toxic Releases to Air
#> 404  S_E_TRAFFIC_PER    S_P_PTRAF               State%ile Traffic
#> 405     S_E_LEAD_PER    S_P_LDPNT             State%ile %pre-1960
#> 406      S_E_NPL_PER     S_P_PNPL                   State%ile NPL
#> 407      S_E_RMP_PER     S_P_PRMP                   State%ile RMP
#> 408     S_E_TSDF_PER    S_P_PTSDF                  State%ile TSDF
#> 409      S_E_UST_PER      S_P_UST                   State%ile UST
#> 410    S_E_NPDES_PER    S_P_PWDIS                 State%ile NPDES
#> 411   S_E_DWATER_PER   S_P_DWATER              State%ile Drinking
#>                                                                                           longname
#> 386                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 387                                                                  US percentile for Ozone (ppb)
#> 388                                                       US percentile for Nitrogen Dioxide (NO2)
#> 389                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 390                                                        US percentile for Toxic Releases to Air
#> 391          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 392                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 393                                 US percentile for Superfund Proximity (site count/km distance)
#> 394                                   US percentile for RMP Proximity (facility count/km distance)
#> 395                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 396                                    US percentile for Underground Storage Tanks (UST) indicator
#> 397    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 398                                                US percentile for Drinking Water Non-Compliance
#> 399                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 400                                                               State percentile for Ozone (ppb)
#> 401                                                    State percentile for Nitrogen Dioxide (NO2)
#> 402                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 403                                                     State percentile for Toxic Releases to Air
#> 404       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 405                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 406                              State percentile for Superfund Proximity (site count/km distance)
#> 407                                State percentile for RMP Proximity (facility count/km distance)
#> 408                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 409                                 State percentile for Underground Storage Tanks (UST) indicator
#> 410 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 411                                             State percentile for Drinking Water Non-Compliance