Get indicator names within a varlist like names_d

Get indicator names within a varlist like names_d

Usage

namesbyvarlist(varlist, nametype = c("rname", "longname", "apiname")[1])

Arguments

varlist

one character string like "names_d", or a vector of them

nametype

vector of 1 or more names of columns in map_headernames, or a shortcut type that can be api, csv, r, original, long, shortlabel

Value

a data.frame one row per indicator, one col per nametype and a column identifying the varlist

Details

varlist2names() is a way to just get a vector of variable names even if the varlist is not stored as a separate data object and is only found in the map_headernames$varlist column:

varlist2names(c('names_d', 'names_d_subgroups'))

c(names_d, names_d_subgroups)

See also

Examples

 unique(map_headernames$varlist)
#>  [1] "names_d"                                   
#>  [2] "names_d_demogindexstate"                   
#>  [3] "names_d_ratio_to_avg"                      
#>  [4] "names_d_ratio_to_state_avg"                
#>  [5] "names_d_pctile"                            
#>  [6] "names_d_state_pctile"                      
#>  [7] "names_d_avg"                               
#>  [8] "names_d_state_avg"                         
#>  [9] "names_d_count"                             
#> [10] "names_criticalservice_pctile"              
#> [11] "names_criticalservice_state_pctile"        
#> [12] "names_criticalservice_avg"                 
#> [13] "names_d_subgroups"                         
#> [14] "names_criticalservice_state_avg"           
#> [15] "names_featuresinarea"                      
#> [16] "names_d_subgroups_ratio_to_avg"            
#> [17] "names_age"                                 
#> [18] "names_d_subgroups_ratio_to_state_avg"      
#> [19] "names_age_count"                           
#> [20] "names_community"                           
#> [21] "names_d_subgroups_pctile"                  
#> [22] "names_community_count"                     
#> [23] "names_flag"                                
#> [24] "custom"                                    
#> [25] "names_d_subgroups_state_pctile"            
#> [26] "names_d_subgroups_avg"                     
#> [27] "names_d_subgroups_state_avg"               
#> [28] "names_d_subgroups_count"                   
#> [29] "names_d_subgroups_alone"                   
#> [30] "names_d_subgroups_alone_ratio_to_avg"      
#> [31] "names_d_subgroups_alone_ratio_to_state_avg"
#> [32] "names_d_subgroups_alone_pctile"            
#> [33] "names_d_subgroups_alone_state_pctile"      
#> [34] "names_d_subgroups_alone_avg"               
#> [35] "names_d_subgroups_alone_state_avg"         
#> [36] "names_d_subgroups_alone_count"             
#> [37] "names_d_subgroups_nh"                      
#> [38] "names_d_subgroups_nh_ratio_to_avg"         
#> [39] "names_d_subgroups_nh_ratio_to_state_avg"   
#> [40] "names_d_subgroups_nh_pctile"               
#> [41] "names_d_subgroups_nh_state_pctile"         
#> [42] "names_d_subgroups_nh_avg"                  
#> [43] "names_d_subgroups_nh_state_avg"            
#> [44] "names_d_subgroups_nh_count"                
#> [45] "names_d_language"                          
#> [46] "names_d_language_count"                    
#> [47] "names_d_languageli"                        
#> [48] "names_d_languageli_count"                  
#> [49] "names_d_other_count"                       
#> [50] "names_d_extra"                             
#> [51] "names_d_extra_count"                       
#> [52] "names_health"                              
#> [53] "names_health_ratio_to_avg"                 
#> [54] "names_health_ratio_to_state_avg"           
#> [55] "names_health_pctile"                       
#> [56] "names_health_state_pctile"                 
#> [57] "names_health_avg"                          
#> [58] "names_health_state_avg"                    
#> [59] "names_health_count"                        
#> [60] "names_criticalservice"                     
#> [61] "names_misc"                                
#> [62] "names_sitesinarea"                         
#> [63] "names_e"                                   
#> [64] "names_e_ratio_to_avg"                      
#> [65] "names_e_ratio_to_state_avg"                
#> [66] "names_e_pctile"                            
#> [67] "names_e_state_pctile"                      
#> [68] "names_e_avg"                               
#> [69] "names_e_state_avg"                         
#> [70] "names_climate"                             
#> [71] "names_climate_pctile"                      
#> [72] "names_climate_state_pctile"                
#> [73] "names_climate_avg"                         
#> [74] "names_climate_state_avg"                   
#> [75] "names_e_other"                             
#> [76] "names_ej"                                  
#> [77] "names_ej_state"                            
#> [78] "names_ej_pctile"                           
#> [79] "names_ej_state_pctile"                     
#> [80] "names_ej_supp"                             
#> [81] "names_ej_supp_state"                       
#> [82] "names_ej_supp_pctile"                      
#> [83] "names_ej_supp_state_pctile"                
#> [84] "names_countabove"                          
#> [85] "names_geo"                                 
 
 namesbyvarlist('names_e_avg', 'rname')
#>         varlist               rname
#> 439 names_e_avg              avg.pm
#> 440 names_e_avg              avg.o3
#> 441 names_e_avg             avg.no2
#> 442 names_e_avg             avg.dpm
#> 443 names_e_avg            avg.rsei
#> 444 names_e_avg   avg.traffic.score
#> 445 names_e_avg      avg.pctpre1960
#> 446 names_e_avg   avg.proximity.npl
#> 447 names_e_avg   avg.proximity.rmp
#> 448 names_e_avg  avg.proximity.tsdf
#> 449 names_e_avg             avg.ust
#> 450 names_e_avg avg.proximity.npdes
#> 451 names_e_avg        avg.drinking
 namesbyvarlist('names_d')
#>   varlist            rname
#> 1 names_d      Demog.Index
#> 2 names_d Demog.Index.Supp
#> 3 names_d        pctlowinc
#> 4 names_d       pctlingiso
#> 5 names_d    pctunemployed
#> 6 names_d          pctlths
#> 7 names_d        pctunder5
#> 8 names_d        pctover64
#> 9 names_d           pctmin
 namesbyvarlist('names_d', 'r')
#>   varlist            rname
#> 1 names_d      Demog.Index
#> 2 names_d Demog.Index.Supp
#> 3 names_d        pctlowinc
#> 4 names_d       pctlingiso
#> 5 names_d    pctunemployed
#> 6 names_d          pctlths
#> 7 names_d        pctunder5
#> 8 names_d        pctover64
#> 9 names_d           pctmin
 namesbyvarlist('names_d', 'long')
#>   varlist                                 longname
#> 1 names_d                    Demographic Index USA
#> 2 names_d       Supplemental Demographic Index USA
#> 3 names_d                             % Low Income
#> 4 names_d % in limited English-speaking Households
#> 5 names_d                             % Unemployed
#> 6 names_d   % with Less Than High School Education
#> 7 names_d                            % under Age 5
#> 8 names_d                            % over Age 64
#> 9 names_d                        % People of Color
 namesbyvarlist('names_d', 'shortlabel')
#>   varlist       shortlabel
#> 1 names_d       Demog.Ind.
#> 2 names_d Suppl Demog.Ind.
#> 3 names_d        %Low-inc.
#> 4 names_d %Limited English
#> 5 names_d      %Unemployed
#> 6 names_d   %< High School
#> 7 names_d         %< age 5
#> 8 names_d        %> age 64
#> 9 names_d             %POC
 
 namesbyvarlist( 'names_e_pctile', c('r', 'longname'))    
#>            varlist                  rname
#> 413 names_e_pctile              pctile.pm
#> 414 names_e_pctile              pctile.o3
#> 415 names_e_pctile             pctile.no2
#> 416 names_e_pctile             pctile.dpm
#> 417 names_e_pctile            pctile.rsei
#> 418 names_e_pctile   pctile.traffic.score
#> 419 names_e_pctile      pctile.pctpre1960
#> 420 names_e_pctile   pctile.proximity.npl
#> 421 names_e_pctile   pctile.proximity.rmp
#> 422 names_e_pctile  pctile.proximity.tsdf
#> 423 names_e_pctile             pctile.ust
#> 424 names_e_pctile pctile.proximity.npdes
#> 425 names_e_pctile        pctile.drinking
#>                                                                                        longname
#> 413                                      US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 414                                                               US percentile for Ozone (ppb)
#> 415                                                    US percentile for Nitrogen Dioxide (NO2)
#> 416                                         US percentile for Diesel Particulate Matter (ug/m3)
#> 417                                                     US percentile for Toxic Releases to Air
#> 418       US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 419                                US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 420                              US percentile for Superfund Proximity (site count/km distance)
#> 421                                US percentile for RMP Proximity (facility count/km distance)
#> 422                    US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 423                                 US percentile for Underground Storage Tanks (UST) indicator
#> 424 US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 425                                             US percentile for Drinking Water Non-Compliance
 namesbyvarlist(c('names_e_pctile', 'names_e_state_pctile'), 
   c('varlist', 'rname', 'apiname', 'csvname', 'shortlabel', 'longname'))
#>                  varlist            varlist.1                        rname
#> 413       names_e_pctile       names_e_pctile                    pctile.pm
#> 414       names_e_pctile       names_e_pctile                    pctile.o3
#> 415       names_e_pctile       names_e_pctile                   pctile.no2
#> 416       names_e_pctile       names_e_pctile                   pctile.dpm
#> 417       names_e_pctile       names_e_pctile                  pctile.rsei
#> 418       names_e_pctile       names_e_pctile         pctile.traffic.score
#> 419       names_e_pctile       names_e_pctile            pctile.pctpre1960
#> 420       names_e_pctile       names_e_pctile         pctile.proximity.npl
#> 421       names_e_pctile       names_e_pctile         pctile.proximity.rmp
#> 422       names_e_pctile       names_e_pctile        pctile.proximity.tsdf
#> 423       names_e_pctile       names_e_pctile                   pctile.ust
#> 424       names_e_pctile       names_e_pctile       pctile.proximity.npdes
#> 425       names_e_pctile       names_e_pctile              pctile.drinking
#> 426 names_e_state_pctile names_e_state_pctile              state.pctile.pm
#> 427 names_e_state_pctile names_e_state_pctile              state.pctile.o3
#> 428 names_e_state_pctile names_e_state_pctile             state.pctile.no2
#> 429 names_e_state_pctile names_e_state_pctile             state.pctile.dpm
#> 430 names_e_state_pctile names_e_state_pctile            state.pctile.rsei
#> 431 names_e_state_pctile names_e_state_pctile   state.pctile.traffic.score
#> 432 names_e_state_pctile names_e_state_pctile      state.pctile.pctpre1960
#> 433 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.npl
#> 434 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.rmp
#> 435 names_e_state_pctile names_e_state_pctile  state.pctile.proximity.tsdf
#> 436 names_e_state_pctile names_e_state_pctile             state.pctile.ust
#> 437 names_e_state_pctile names_e_state_pctile state.pctile.proximity.npdes
#> 438 names_e_state_pctile names_e_state_pctile        state.pctile.drinking
#>              apiname      csvname                      shortlabel
#> 413     N_E_PM25_PER       P_PM25                    US%ile PM2.5
#> 414       N_E_O3_PER      P_OZONE                    US%ile Ozone
#> 415      N_E_NO2_PER        P_NO2                      US%ile NO2
#> 416   N_E_DIESEL_PER      P_DSLPM                US%ile Diesel PM
#> 417 N_E_RSEI_AIR_PER   P_RSEI_AIR    US%ile Toxic Releases to Air
#> 418  N_E_TRAFFIC_PER      P_PTRAF                  US%ile Traffic
#> 419     N_E_LEAD_PER      P_LDPNT                US%ile %pre-1960
#> 420      N_E_NPL_PER       P_PNPL                      US%ile NPL
#> 421      N_E_RMP_PER       P_PRMP                      US%ile RMP
#> 422     N_E_TSDF_PER      P_PTSDF                     US%ile TSDF
#> 423      N_E_UST_PER        P_UST                      US%ile UST
#> 424    N_E_NPDES_PER      P_PWDIS                    US%ile NPDES
#> 425   N_E_DWATER_PER     P_DWATER                 US%ile Drinking
#> 426     S_E_PM25_PER     S_P_PM25                 State%ile PM2.5
#> 427       S_E_O3_PER    S_P_OZONE                 State%ile Ozone
#> 428      S_E_NO2_PER      S_P_NO2                   State%ile NO2
#> 429   S_E_DIESEL_PER    S_P_DSLPM             State%ile Diesel PM
#> 430 S_E_RSEI_AIR_PER S_P_RSEI_AIR State%ile Toxic Releases to Air
#> 431  S_E_TRAFFIC_PER    S_P_PTRAF               State%ile Traffic
#> 432     S_E_LEAD_PER    S_P_LDPNT             State%ile %pre-1960
#> 433      S_E_NPL_PER     S_P_PNPL                   State%ile NPL
#> 434      S_E_RMP_PER     S_P_PRMP                   State%ile RMP
#> 435     S_E_TSDF_PER    S_P_PTSDF                  State%ile TSDF
#> 436      S_E_UST_PER      S_P_UST                   State%ile UST
#> 437    S_E_NPDES_PER    S_P_PWDIS                 State%ile NPDES
#> 438   S_E_DWATER_PER   S_P_DWATER              State%ile Drinking
#>                                                                                           longname
#> 413                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 414                                                                  US percentile for Ozone (ppb)
#> 415                                                       US percentile for Nitrogen Dioxide (NO2)
#> 416                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 417                                                        US percentile for Toxic Releases to Air
#> 418          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 419                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 420                                 US percentile for Superfund Proximity (site count/km distance)
#> 421                                   US percentile for RMP Proximity (facility count/km distance)
#> 422                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 423                                    US percentile for Underground Storage Tanks (UST) indicator
#> 424    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 425                                                US percentile for Drinking Water Non-Compliance
#> 426                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 427                                                               State percentile for Ozone (ppb)
#> 428                                                    State percentile for Nitrogen Dioxide (NO2)
#> 429                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 430                                                     State percentile for Toxic Releases to Air
#> 431       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 432                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 433                              State percentile for Superfund Proximity (site count/km distance)
#> 434                                State percentile for RMP Proximity (facility count/km distance)
#> 435                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 436                                 State percentile for Underground Storage Tanks (UST) indicator
#> 437 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 438                                             State percentile for Drinking Water Non-Compliance