Get indicator names within a varlist like names_d

Get indicator names within a varlist like names_d

Usage

namesbyvarlist(
  varlist,
  nametype = c("rname", "longname_tableheader", "apiname")[1]
)

Arguments

varlist

one character string like "names_d", or a vector of them

nametype

vector of 1 or more names of columns in map_headernames, or a shortcut type that can be api, csv, r, original, long, shortlabel

Value

a data.frame one row per indicator, one col per nametype and a column identifying the varlist

Details

Note names_d_friendly refers to the semi-long names to use in labels

Examples

 unique(map_headernames$varlist)
#>  [1] "names_d"                                   
#>  [2] "names_d_ratio_to_avg"                      
#>  [3] "names_d_ratio_to_state_avg"                
#>  [4] "names_d_pctile"                            
#>  [5] "names_d_state_pctile"                      
#>  [6] "names_d_avg"                               
#>  [7] "names_d_state_avg"                         
#>  [8] "names_d_count"                             
#>  [9] "names_d_subgroups"                         
#> [10] "names_d_subgroups_ratio_to_avg"            
#> [11] "names_d_subgroups_ratio_to_state_avg"      
#> [12] "names_d_subgroups_pctile"                  
#> [13] "names_d_subgroups_state_pctile"            
#> [14] "names_d_subgroups_avg"                     
#> [15] "names_d_subgroups_state_avg"               
#> [16] "names_d_subgroups_count"                   
#> [17] "names_d_subgroups_alone"                   
#> [18] "names_d_subgroups_alone_ratio_to_avg"      
#> [19] "names_d_subgroups_alone_ratio_to_state_avg"
#> [20] "names_d_subgroups_alone_pctile"            
#> [21] "names_d_subgroups_alone_state_pctile"      
#> [22] "names_d_subgroups_alone_avg"               
#> [23] "names_d_subgroups_alone_state_avg"         
#> [24] "names_d_subgroups_alone_count"             
#> [25] "names_d_subgroups_nh"                      
#> [26] "names_d_subgroups_nh_ratio_to_avg"         
#> [27] "names_d_subgroups_nh_ratio_to_state_avg"   
#> [28] "names_d_subgroups_nh_pctile"               
#> [29] "names_d_subgroups_nh_state_pctile"         
#> [30] "names_d_subgroups_nh_avg"                  
#> [31] "names_d_subgroups_nh_state_avg"            
#> [32] "names_d_subgroups_nh_count"                
#> [33] "names_d_other_count"                       
#> [34] "x_mapcolor"                                
#> [35] "x_maptextpopup"                            
#> [36] "x_anyother"                                
#> [37] "names_e"                                   
#> [38] "names_e_ratio_to_avg"                      
#> [39] "names_e_ratio_to_state_avg"                
#> [40] "names_e_pctile"                            
#> [41] "names_e_state_pctile"                      
#> [42] "names_e_avg"                               
#> [43] "names_e_state_avg"                         
#> [44] "names_ej"                                  
#> [45] "names_ej_state"                            
#> [46] "names_ej_supp"                             
#> [47] "names_ej_supp_state"                       
#> [48] "names_ej_pctile"                           
#> [49] "names_ej_state_pctile"                     
#> [50] "names_ej_supp_pctile"                      
#> [51] "names_ej_supp_state_pctile"                
 
 namesbyvarlist('names_e_avg', 'rname')
#> Error in namesbyvarlist("names_e_avg", "rname"): could not find function "namesbyvarlist"
 namesbyvarlist('names_d')
#> Error in namesbyvarlist("names_d"): could not find function "namesbyvarlist"
 namesbyvarlist('names_d', 'r')
#> Error in namesbyvarlist("names_d", "r"): could not find function "namesbyvarlist"
 namesbyvarlist('names_d', 'friendly')
#> Error in namesbyvarlist("names_d", "friendly"): could not find function "namesbyvarlist"
 namesbyvarlist('names_d', 'shortlabel')
#> Error in namesbyvarlist("names_d", "shortlabel"): could not find function "namesbyvarlist"
 
 namesbyvarlist( 'names_e_pctile', c('r', 'longname_tableheader'))    
#> Error in namesbyvarlist("names_e_pctile", c("r", "longname_tableheader")): could not find function "namesbyvarlist"
 namesbyvarlist(c('names_e_pctile', 'names_e_state_pctile'), 
   c('varlist', 'r', 'api',   'names_friendly', 'longname_tableheader'))
#> Error in namesbyvarlist(c("names_e_pctile", "names_e_state_pctile"), c("varlist",     "r", "api", "names_friendly", "longname_tableheader")): could not find function "namesbyvarlist"