Get indicator names within a varlist like names_d

Get indicator names within a varlist like names_d

Usage

namesbyvarlist(
  varlist,
  nametype = c("rname", "longname_tableheader", "apiname")[1]
)

Arguments

varlist

one character string like "names_d", or a vector of them

nametype

vector of 1 or more names of columns in map_headernames, or a shortcut type that can be api, csv, r, original, long, shortlabel

Value

a data.frame one row per indicator, one col per nametype and a column identifying the varlist

Details

Note names_d_friendly refers to the semi-long names to use in labels

Examples

 unique(map_headernames$varlist)
#>  [1] "names_d"                                    "names_d_ratio_to_avg"                      
#>  [3] "names_d_ratio_to_state_avg"                 "names_d_pctile"                            
#>  [5] "names_d_state_pctile"                       "names_d_avg"                               
#>  [7] "names_d_state_avg"                          "names_d_count"                             
#>  [9] "names_d_subgroups"                          "names_d_subgroups_ratio_to_avg"            
#> [11] "names_d_subgroups_ratio_to_state_avg"       "names_d_subgroups_pctile"                  
#> [13] "names_d_subgroups_state_pctile"             "names_d_subgroups_avg"                     
#> [15] "names_d_subgroups_state_avg"                "names_d_subgroups_count"                   
#> [17] "names_d_subgroups_alone"                    "names_d_subgroups_alone_ratio_to_avg"      
#> [19] "names_d_subgroups_alone_ratio_to_state_avg" "names_d_subgroups_alone_pctile"            
#> [21] "names_d_subgroups_alone_state_pctile"       "names_d_subgroups_alone_avg"               
#> [23] "names_d_subgroups_alone_state_avg"          "names_d_subgroups_alone_count"             
#> [25] "names_d_subgroups_nh"                       "names_d_subgroups_nh_ratio_to_avg"         
#> [27] "names_d_subgroups_nh_ratio_to_state_avg"    "names_d_subgroups_nh_pctile"               
#> [29] "names_d_subgroups_nh_state_pctile"          "names_d_subgroups_nh_avg"                  
#> [31] "names_d_subgroups_nh_state_avg"             "names_d_subgroups_nh_count"                
#> [33] "names_d_other_count"                        "x_mapcolor"                                
#> [35] "x_maptextpopup"                             "x_anyother"                                
#> [37] "names_e"                                    "names_e_ratio_to_avg"                      
#> [39] "names_e_ratio_to_state_avg"                 "names_e_pctile"                            
#> [41] "names_e_state_pctile"                       "names_e_avg"                               
#> [43] "names_e_state_avg"                          "names_ej"                                  
#> [45] "names_ej_state"                             "names_ej_supp"                             
#> [47] "names_ej_supp_state"                        "names_ej_pctile"                           
#> [49] "names_ej_state_pctile"                      "names_ej_supp_pctile"                      
#> [51] "names_ej_supp_state_pctile"                
 
 namesbyvarlist('names_e_avg', 'rname')
#>         varlist               rname
#> 408 names_e_avg              avg.pm
#> 409 names_e_avg              avg.o3
#> 410 names_e_avg          avg.cancer
#> 411 names_e_avg            avg.resp
#> 412 names_e_avg             avg.dpm
#> 413 names_e_avg      avg.pctpre1960
#> 414 names_e_avg   avg.traffic.score
#> 415 names_e_avg   avg.proximity.npl
#> 416 names_e_avg   avg.proximity.rmp
#> 417 names_e_avg  avg.proximity.tsdf
#> 418 names_e_avg avg.proximity.npdes
#> 419 names_e_avg             avg.ust
#> 420 names_e_avg            avg.rsei
 namesbyvarlist('names_d')
#>    varlist            rname
#> 1  names_d      Demog.Index
#> 2  names_d Demog.Index.Supp
#> 3  names_d        pctlowinc
#> 4  names_d       pctlingiso
#> 5  names_d    pctunemployed
#> 6  names_d          pctlths
#> 7  names_d         lowlifex
#> 8  names_d        pctunder5
#> 9  names_d        pctover64
#> 10 names_d           pctmin
 namesbyvarlist('names_d', 'r')
#>    varlist            rname
#> 1  names_d      Demog.Index
#> 2  names_d Demog.Index.Supp
#> 3  names_d        pctlowinc
#> 4  names_d       pctlingiso
#> 5  names_d    pctunemployed
#> 6  names_d          pctlths
#> 7  names_d         lowlifex
#> 8  names_d        pctunder5
#> 9  names_d        pctover64
#> 10 names_d           pctmin
 namesbyvarlist('names_d', 'friendly')
#> Error in namesbyvarlist("names_d", "friendly"): specified nametype not a known type and not found among colnames(map_headernames)
 namesbyvarlist('names_d', 'shortlabel')
#>    varlist           shortlabel
#> 1  names_d           Demog.Ind.
#> 2  names_d     Suppl Demog.Ind.
#> 3  names_d            %Low-inc.
#> 4  names_d     %Limited English
#> 5  names_d          %Unemployed
#> 6  names_d       %< High School
#> 7  names_d %Low life expectancy
#> 8  names_d             %< age 5
#> 9  names_d            %> age 64
#> 10 names_d                 %POC
 
 namesbyvarlist( 'names_e_pctile', c('r', 'longname_tableheader'))    
#>            varlist                  rname
#> 382 names_e_pctile              pctile.pm
#> 383 names_e_pctile              pctile.o3
#> 384 names_e_pctile          pctile.cancer
#> 385 names_e_pctile            pctile.resp
#> 386 names_e_pctile             pctile.dpm
#> 387 names_e_pctile      pctile.pctpre1960
#> 388 names_e_pctile   pctile.traffic.score
#> 389 names_e_pctile   pctile.proximity.npl
#> 390 names_e_pctile   pctile.proximity.rmp
#> 391 names_e_pctile  pctile.proximity.tsdf
#> 392 names_e_pctile pctile.proximity.npdes
#> 393 names_e_pctile             pctile.ust
#> 394 names_e_pctile            pctile.rsei
#>                                                                            longname_tableheader
#> 382                                      US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 383                                                               US percentile for Ozone (ppb)
#> 384                                 US percentile for Air Toxics Cancer Risk (risk per million)
#> 385                                       US percentile for Air Toxics Respiratory Hazard Index
#> 386                              US percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 387                                US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 388       US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 389                              US percentile for Superfund Proximity (site count/km distance)
#> 390                                US percentile for RMP Proximity (facility count/km distance)
#> 391                    US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 392 US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 393                                 US percentile for Underground Storage Tanks (UST) indicator
#> 394                                                     US percentile for Toxic Releases to Air
 namesbyvarlist(c('names_e_pctile', 'names_e_state_pctile'), 
   c('varlist', 'r', 'api',   'names_friendly', 'longname_tableheader'))
#>                  varlist            varlist.1                        rname          apiname
#> 382       names_e_pctile       names_e_pctile                    pctile.pm     N_E_PM25_PER
#> 383       names_e_pctile       names_e_pctile                    pctile.o3       N_E_O3_PER
#> 384       names_e_pctile       names_e_pctile                pctile.cancer   N_E_CANCER_PER
#> 385       names_e_pctile       names_e_pctile                  pctile.resp     N_E_RESP_PER
#> 386       names_e_pctile       names_e_pctile                   pctile.dpm   N_E_DIESEL_PER
#> 387       names_e_pctile       names_e_pctile            pctile.pctpre1960     N_E_LEAD_PER
#> 388       names_e_pctile       names_e_pctile         pctile.traffic.score  N_E_TRAFFIC_PER
#> 389       names_e_pctile       names_e_pctile         pctile.proximity.npl      N_E_NPL_PER
#> 390       names_e_pctile       names_e_pctile         pctile.proximity.rmp      N_E_RMP_PER
#> 391       names_e_pctile       names_e_pctile        pctile.proximity.tsdf     N_E_TSDF_PER
#> 392       names_e_pctile       names_e_pctile       pctile.proximity.npdes    N_E_NPDES_PER
#> 393       names_e_pctile       names_e_pctile                   pctile.ust      N_E_UST_PER
#> 394       names_e_pctile       names_e_pctile                  pctile.rsei N_E_RSEI_AIR_PER
#> 395 names_e_state_pctile names_e_state_pctile              state.pctile.pm     S_E_PM25_PER
#> 396 names_e_state_pctile names_e_state_pctile              state.pctile.o3       S_E_O3_PER
#> 397 names_e_state_pctile names_e_state_pctile          state.pctile.cancer   S_E_CANCER_PER
#> 398 names_e_state_pctile names_e_state_pctile            state.pctile.resp     S_E_RESP_PER
#> 399 names_e_state_pctile names_e_state_pctile             state.pctile.dpm   S_E_DIESEL_PER
#> 400 names_e_state_pctile names_e_state_pctile      state.pctile.pctpre1960     S_E_LEAD_PER
#> 401 names_e_state_pctile names_e_state_pctile   state.pctile.traffic.score  S_E_TRAFFIC_PER
#> 402 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.npl      S_E_NPL_PER
#> 403 names_e_state_pctile names_e_state_pctile   state.pctile.proximity.rmp      S_E_RMP_PER
#> 404 names_e_state_pctile names_e_state_pctile  state.pctile.proximity.tsdf     S_E_TSDF_PER
#> 405 names_e_state_pctile names_e_state_pctile state.pctile.proximity.npdes    S_E_NPDES_PER
#> 406 names_e_state_pctile names_e_state_pctile             state.pctile.ust      S_E_UST_PER
#> 407 names_e_state_pctile names_e_state_pctile            state.pctile.rsei S_E_RSEI_AIR_PER
#>                                                                                     names_friendly
#> 382                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 383                                                                  US percentile for Ozone (ppb)
#> 384                                    US percentile for Air Toxics Cancer Risk (risk per million)
#> 385                                          US percentile for Air Toxics Respiratory Hazard Index
#> 386                                 US percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 387                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 388          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 389                                 US percentile for Superfund Proximity (site count/km distance)
#> 390                                   US percentile for RMP Proximity (facility count/km distance)
#> 391                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 392    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 393                                    US percentile for Underground Storage Tanks (UST) indicator
#> 394                                                        US percentile for Toxic Releases to Air
#> 395                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 396                                                               State percentile for Ozone (ppb)
#> 397                                 State percentile for Air Toxics Cancer Risk (risk per million)
#> 398                                       State percentile for Air Toxics Respiratory Hazard Index
#> 399                              State percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 400                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 401       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 402                              State percentile for Superfund Proximity (site count/km distance)
#> 403                                State percentile for RMP Proximity (facility count/km distance)
#> 404                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 405 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 406                                 State percentile for Underground Storage Tanks (UST) indicator
#> 407                                                     State percentile for Toxic Releases to Air
#>                                                                               longname_tableheader
#> 382                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 383                                                                  US percentile for Ozone (ppb)
#> 384                                    US percentile for Air Toxics Cancer Risk (risk per million)
#> 385                                          US percentile for Air Toxics Respiratory Hazard Index
#> 386                                 US percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 387                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 388          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 389                                 US percentile for Superfund Proximity (site count/km distance)
#> 390                                   US percentile for RMP Proximity (facility count/km distance)
#> 391                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 392    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 393                                    US percentile for Underground Storage Tanks (UST) indicator
#> 394                                                        US percentile for Toxic Releases to Air
#> 395                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 396                                                               State percentile for Ozone (ppb)
#> 397                                 State percentile for Air Toxics Cancer Risk (risk per million)
#> 398                                       State percentile for Air Toxics Respiratory Hazard Index
#> 399                              State percentile for Air Toxics Diesel Particulate Matter (ug/m3)
#> 400                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 401       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 402                              State percentile for Superfund Proximity (site count/km distance)
#> 403                                State percentile for RMP Proximity (facility count/km distance)
#> 404                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 405 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 406                                 State percentile for Underground Storage Tanks (UST) indicator
#> 407                                                     State percentile for Toxic Releases to Air