Barplot of ratios of demographic (or other) scores to averages (or other references)

Barplot of ratios of demographic (or other) scores to averages (or other references)

Usage

plot_barplot_ratios(
  ratio.to.us.d.overall,
  shortlabels = NULL,
  mycolorsavailable = c("red", "orange", "yellow", "gray"),
  main = "Demographics at the Analyzed Locations Compared to US Overall"
)

Arguments

ratio.to.us.d.overall

named list of a few ratios to plot, but see ejam2barplot() for an easier way to specify which indicator to show.

shortlabels

names to use for plot - should be same length as named list ratio.to.us.d.overall

mycolorsavailable

leave as default

main

title for plot, like "Demographics at the Analyzed Locations Compared to US Overall"

Value

ggplot should be returned

See also

Examples

  
  ejam2barplot(testoutput_ejamit_100pts_1miles)
  
  plot_barplot_ratios(unlist(testoutput_ejamit_1000pts_1miles$results_overall[ , c(..names_d_ratio_to_avg , ..names_d_subgroups_ratio_to_avg) ]))