plot_distance_by_pctd - Plot percent demographics within X miles of a site

plot_distance_by_pctd - Plot percent demographics within X miles of a site

plot_distance_by_pctd(
  s2b = NULL,
  sitenum = NULL,
  myvars = c(names_d_count, names_d_subgroups_count)[1],
  dpctvar = paste0("pct", myvars)
)

Arguments

s2b

output of getblocksnearby()

sitenum

one number that is the unique ID (the row number of original list of points) to look at in s2b

myvars

a colname of a population count variable in blockgroupstats indicating which to plot, like "hisp" or "lowinc" and only works for one indicator at a time so far.

dpctvar

a colname of usastats and statestats that is the percentage version of myvars, like "pcthisp" or "pctlowinc"

Value

returns s2b but with more columns in it, like cumpop, cumdpop, pctdwithin

Examples

  plot_distance_by_pctd()
#> Analyzing 2 points, radius of 6.2 miles around each.
#> Finding Census blocks with internal point within  6.2  miles of the site (point), for each of 2  sites (points)...
#> Stats via getblocks_diagnostics(), but BEFORE ADJUSTING FOR SHORT DISTANCES: 
#> min distance before adjustment:  0.1061425 
#> max distance before adjustment:  6.199732 
#> 
#> 
#>    Summary stats on distances reported from any sites to any nearby blocks
#> 
#> Warning: no non-missing arguments to max; returning -Inf
#> -Inf miles is max. distance to block internal point (distance_unadjusted)   
#> 6.199732 miles is max. distance to average resident in block (distance reported)   
#> Warning: no non-missing arguments to min; returning Inf
#> Inf miles is shortest distance to block internal point (distance_unadjusted)   
#> 0.1061425 miles is shortest distance to average resident in block (distance reported)   
#> 
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to min; returning Inf
#> 2 unique output sites
#> 5,078 blocks are near the avg site or in avg buffer
#>              (based on their block internal point, like a centroid)
#> 10,157 blocks including doublecounting in overlaps, 
#>              in final row count (block-to-site pairs table)
#> 10,157 actual unique blocks total
#> 10,157 blocks (and their residents) have exactly 1 site nearby 
#> 0 blocks (and their residents) have exactly 2 sites nearby 
#> 0 blocks (and their residents) have exactly 3 sites nearby 
#> 1 is ratio of blocks including multicounting / actual count of unique blocks
#> 0% of unique blocks could get counted more than once 
#>              because those residents are near two or more sites 
#>              (assuming they live at the block internal point
#> 
#> 
#> Adjusting for very short distances now...
#>  Stats via getblocks_diagnostics(), AFTER ADJUSTING FOR SHORT DISTANCES: 
#> min distance AFTER adjustment:  0.1360451 
#> max distance AFTER adjustment:  6.199732 
#> 
#> 
#> Error: 'state_from_blockid' is not an exported object from 'namespace:EJAM'