Launch the Shiny Application in RStudio

Launches shiny web app from RStudio - app_run_EJAM() and run_app() are the same

Usage

run_app(
  onStart = NULL,
  options = list(),
  enableBookmarking = "server",
  uiPattern = "/",
  ...
)

Arguments

onStart

A function that will be called before the app is actually run. This is only needed for shinyAppObj, since in the shinyAppDir case, a global.R file can be used for this purpose.

options

Named options that should be passed to the runApp call (these can be any of the following: "port", "launch.browser", "host", "quiet", "display.mode" and "test.mode"). You can also specify width and height parameters which provide a hint to the embedding environment about the ideal height/width for the app.

enableBookmarking

Can be one of "url", "server", or "disable". The default value, NULL, will respect the setting from any previous calls to enableBookmarking(). See enableBookmarking() for more information on bookmarking your app.

uiPattern

A regular expression that will be applied to each GET request to determine whether the ui should be used to handle the request. Note that the entire request path must match the regular expression in order for the match to be considered successful.

...

arguments to pass to golem_opts. This can be used to pass some of the variables that otherwise would be set to defaults defined in global.R, and currently can include the following:

run_app(

default_default_miles=3.1, 
default_max_miles=31, 
default_max_mb_upload=100, 
default_default_miles_shapefile = 0.5, 
advanced=TRUE, 
testing=TRUE, 
shiny.testmode=TRUE, 

# and this will be implemented at some point: 

sitepoints = mytable such as [testpoints_100]
## or maybe
## sitepoints = "path\latlondata.xlsx"

)

This uses an internal function global_or_param()

Value

An object that represents the app. Printing the object or passing it to runApp() will run the app.

See also