shape_buffered_from_shapefile_points - add buffer around shape (points, here)

shape_buffered_from_shapefile_points - add buffer around shape (points, here)

Usage

shape_buffered_from_shapefile_points(
  shapefile_points,
  radius.miles,
  crs = 4269,
  ...
)

Arguments

shapefile_points

spatial object like areas at high risk or areas with facilities to be analyzed

radius.miles

width of buffer to add to shapefile_points (in case dist is a units object, it should be convertible to arc_degree if x has geographic coordinates, and to st_crs(x)$units otherwise)

crs

used in st_transform() default is crs = 4269 or Geodetic CRS NAD83

...

passed to st_buffer()

Value

same format as sf::st_buffer() returns

Details

Just a wrapper for sf::st_buffer()

See also

Examples

map_shapes_leaflet(
  shape_buffered_from_shapefile_points(
    shapefile_from_sitepoints(testpoints_100), 
    radius.miles = 3
  )
) 
#> Warning: sf layer has inconsistent datum (+proj=longlat +datum=NAD83 +no_defs).
#> Need '+proj=longlat +datum=WGS84'

# (ignoring projections for this example)
# compare to
mapfast(testpoints_100)