Find what state is where each point is located

Takes 3 seconds to find state for 1k points, so a faster alternative would be useful It can take approx. one minute for 2.5 million points as in state_from_latlon(frs$lat, frs$lon)

Usage

state_from_latlon(lat, lon)

Arguments

lat

latitudes vector

lon

longitudes vector

Value

Returns data.frame: ST, statename, FIPS.ST, REGION, n as many rows as elements in lat or lon

See also

Examples

 myprogram <- "CAMDBS" # 739 sites
 pts <- frs_from_program(myprogram)[ , .(lat, lon, REGISTRY_ID,  PRIMARY_NAME)]
 # add a column with State abbreviation
 pts[, ST := state_from_latlon(lat=lat, lon = lon)$ST]
#> Key: <REGISTRY_ID>
#>           lat       lon  REGISTRY_ID
#>         <num>     <num>       <char>
#>   1: 17.94930 -66.23170 110000307800
#>   2: 42.15698 -72.52237 110000308444
#>   3: 42.45488 -73.21734 110000308480
#>   4: 39.91822 -75.11814 110000321456
#>   5: 43.16100 -78.75450 110000326530
#>  ---                                
#> 704: 29.70204 -95.07088 110070100752
#> 705: 40.95444 -72.88500 110070696981
#> 706: 41.03141 -81.87759 110070828207
#> 707: 40.75168 -82.79338 110070828208
#> 708: 38.37008 -87.76438 110070834810
#>                                         PRIMARY_NAME     ST
#>                                               <char> <char>
#>   1:              PREPA - AGUIRRE GENERATING COMPLEX     PR
#>   2:                 MASSPOWER COGENERATION FACILITY     MA
#>   3:                PITTSFIELD GENERATING COMPANY LP     MA
#>   4:                             CAMDEN COGENERATION     NJ
#>   5:                  LOCKPORT COGENERATION FACILITY     NY
#>  ---                                                       
#> 704:               EXELON LAPORTE GENERATING STATION     TX
#> 705:                             SHOREHAM ENERGY LLC     NY
#> 706:                OMEGA JV2 - SEVILLE (1652070028)     OH
#> 707:   OMEGA JV2 GALION PEAKING STATION (0317030059)     OH
#> 708: GIBSON GENERATING STATION TYPE I SOUTH LANDFILL     IN
 #map these points
 mapfast(pts[ST == 'TX',], radius = 1) # 1 miles radius circles