Get order of variable names to sort by, as seen in EJScreen Community Report

Get order of variable names to sort by, as seen in EJScreen Community Report

Usage

table_order_variables(
  varnames,
  s1 = "ejscreensort",
  s2 = "sortvarlistEJSCREENREPORT",
  s3 = "sort_within_varlistEJSCREENREPORT"
)

Arguments

varnames

vector of indicator variables names from blockgroupstats, bgej, etc., such as "pm", "pctlowinc", "pctile.EJ.DISPARITY.traffic.score.eo" etc. and others as found in names_all_r, or specific subsets of those like in c(names_d, names_d_subgroups, names_e) and c(names_ej_pctile, names_ej_state_pctile, names_ej_supp_pctile, names_ej_supp_state_pctile)

s1

name of column in map_headernames to get sort info from

s2

optional like s1 but secondary to s1

s3

optional tertiary

Value

vector as from order(), to be used in sorting a data.frame for example

Examples

  cbind(table_order_variables(c(names_d, names_d_subgroups, names_e)))
#>       [,1]
#>  [1,]    1
#>  [2,]    2
#>  [3,]    9
#>  [4,]    3
#>  [5,]    5
#>  [6,]    4
#>  [7,]    6
#>  [8,]    7
#>  [9,]    8
#> [10,]   17
#> [11,]   11
#> [12,]   12
#> [13,]   10
#> [14,]   13
#> [15,]   14
#> [16,]   15
#> [17,]   16
#> [18,]   18
#> [19,]   19
#> [20,]   20
#> [21,]   21
#> [22,]   22
#> [23,]   23
#> [24,]   24
#> [25,]   25
#> [26,]   26
#> [27,]   27
#> [28,]   28
#> [29,]   29
#> [30,]   30
  
  out <- testoutput_ejamit_10pts_1miles
  vars <- out$formatted[ , 'indicator']
  vars <- fixcolnames(vars, 'long', 'r')
  out$formatted[table_order_variables(vars), ]
#>       value
#> 79     1.35
#> 80     1.28
#> 87       53
#> 81       23
#> 83        6
#> 82        6
#> 84        9
#> 85        6
#> 86       13
#> 96        1
#> 97        1
#> 15        1
#> 16      0.8
#> 23      1.3
#> 17      0.7
#> 19      1.2
#> 18      1.2
#> 20      0.8
#> 21      1.2
#> 22      0.7
#> 47        1
#> 48      0.8
#> 55      1.2
#> 49      0.8
#> 51      1.1
#> 50      0.8
#> 52      0.7
#> 53      1.2
#> 54      0.7
#> 136      58
#> 137      33
#> 144      67
#> 138      43
#> 140      68
#> 139      77
#> 141      54
#> 142      65
#> 143      35
#> 168      57
#> 169      41
#> 176      62
#> 170      50
#> 172      63
#> 171      67
#> 173      49
#> 174      66
#> 175      36
#> 200    1.34
#> 201    1.64
#> 208      40
#> 202      30
#> 204       6
#> 203       5
#> 205      11
#> 206       5
#> 207      18
#> 232    1.35
#> 233    1.47
#> 240      44
#> 234      27
#> 236       6
#> 235       7
#> 237      12
#> 238       5
#> 239      17
#> 270  110000
#> 264   46000
#> 266    7700
#> 265    5000
#> 267   13000
#> 268   14000
#> 269   27000
#> 365      53
#> 359      17
#> 360      12
#> 361       0
#> 362       0
#> 363       7
#> 364      11
#> 95       47
#> 89       16
#> 90       12
#> 88       19
#> 91        0
#> 92        0
#> 93        1
#> 94        5
#> 31      0.8
#> 25      1.3
#> 26      2.4
#> 24        1
#> 27      0.3
#> 28      2.2
#> 29      1.6
#> 30      1.5
#> 63      0.9
#> 57      1.4
#> 58      1.4
#> 56      0.9
#> 59      0.8
#> 60      4.9
#> 61        1
#> 62      1.6
#> 152      33
#> 146      77
#> 147      87
#> 145      70
#> 148      84
#> 149      95
#> 150      88
#> 151      77
#> 184      39
#> 178      72
#> 179      74
#> 177      61
#> 180      88
#> 181      72
#> 182      81
#> 183      77
#> 216      60
#> 210      12
#> 211       5
#> 209      18
#> 212       1
#> 213       0
#> 214       0
#> 215       3
#> 248      54
#> 242      12
#> 243       9
#> 241      20
#> 244       0
#> 245       0
#> 246       1
#> 247       3
#> 278   1e+05
#> 272   30000
#> 273   30000
#> 271   40000
#> 274     400
#> 275     700
#> 276    1000
#> 277   10000
#> 107       7
#> 108       6
#> 109       2
#> 111       0
#> 110      13
#> 292   42874
#> 293   35830
#> 294    9800
#> 291  160000
#> 295  575168
#> 296    3633
#> 297   72060
#> 112      47
#> 113      25
#> 114      27
#> 115       1
#> 298    2343
#> 299    1257
#> 300    1330
#> 301      68
#> 8    212317
#> 280   87417
#> 279  151161
#> 281  175609
#> 282   52699
#> 283   95293
#> 284  202460
#> 98       12
#> 285   33361
#> 286   10660
#> 116       9
#> 117      17
#> 45      0.7
#> 46      0.9
#> 77      0.8
#> 78        1
#> 153      26
#> 154      25
#> 185      35
#> 217    13.6
#> 218      20
#> 249    11.9
#> 250      18
#> 357  210565
#> 358   20000
#> 186      43
#> 118    8.41
#> 119    62.7
#> 120    15.8
#> 121   0.591
#> 122     540
#> 123 5400000
#> 124    0.55
#> 125    1.61
#> 126    0.13
#> 127    18.7
#> 128    14.4
#> 129   10000
#> 130       6
#> 32        1
#> 33        1
#> 34        2
#> 35      3.1
#> 36      0.1
#> 37      3.2
#> 38      1.8
#> 39      4.1
#> 40      0.2
#> 41      5.3
#> 42      3.9
#> 43        0
#> 44      2.9
#> 64        1
#> 65        1
#> 66      1.6
#> 67      1.9
#> 68      0.7
#> 69      1.8
#> 70      1.2
#> 71      1.8
#> 72      0.4
#> 73      2.5
#> 74        2
#> 75      0.9
#> 76      1.6
#> 155      59
#> 156      61
#> 157      97
#> 158      97
#> 159      48
#> 160      92
#> 161      77
#> 162      95
#> 163      38
#> 164      96
#> 165      93
#> 166      88
#> 167      91
#> 187      69
#> 188      44
#> 189      81
#> 190      76
#> 191      69
#> 192      77
#> 193      55
#> 194      72
#> 195      53
#> 196      80
#> 197      69
#> 198      69
#> 199      37
#> 219    8.45
#> 220    61.8
#> 221     7.8
#> 222   0.191
#> 223    4600
#> 224 1700000
#> 225     0.3
#> 226    0.39
#> 227    0.57
#> 228     3.5
#> 229     3.6
#> 230   7e+05
#> 231     2.2
#> 251    8.13
#> 252    63.3
#> 253     9.7
#> 254   0.318
#> 255     810
#> 256 2900000
#> 257    0.48
#> 258    0.89
#> 259    0.34
#> 260     7.5
#> 261     7.2
#> 262   11000
#> 263    3.97
#> 132      10
#> 135     0.7
#> 131       1
#> 133       1
#> 134       0
#> 366  72.662
#> 367  83.142
#> 368 117.641
#> 369 116.821
#> 370  59.572
#> 371 118.693
#> 372  94.692
#> 373 108.065
#> 374  36.502
#> 375 123.057
#> 376  99.994
#> 377 105.869
#> 378  80.294
#> 379  91.953
#> 380  58.857
#> 381 104.733
#> 382  97.954
#> 383   91.73
#> 384 101.953
#> 385  76.136
#> 386  92.378
#> 387  54.627
#> 388 106.043
#> 389  83.854
#> 390  91.479
#> 391   46.56
#> 302      63
#> 303      69
#> 304      76
#> 305      77
#> 306      58
#> 307      77
#> 308      73
#> 309      80
#> 310      45
#> 311      78
#> 312      74
#> 313      76
#> 314      83
#> 328      66
#> 329      53
#> 330      68
#> 331      65
#> 332      66
#> 333      68
#> 334      60
#> 335      64
#> 336      58
#> 337      68
#> 338      59
#> 339      66
#> 340      45
#> 392  66.552
#> 393  78.034
#> 394 110.354
#> 395 109.369
#> 396  56.357
#> 397 110.815
#> 398  91.293
#> 399 101.699
#> 400  35.509
#> 401 114.085
#> 402  94.901
#> 403  99.413
#> 404   78.09
#> 405  85.146
#> 406  55.113
#> 407  97.476
#> 408  92.088
#> 409  85.865
#> 410  93.448
#> 411  71.343
#> 412  87.488
#> 413  53.506
#> 414  97.693
#> 415  80.199
#> 416  84.606
#> 417  47.318
#> 315      47
#> 316      57
#> 317      72
#> 318      73
#> 319      41
#> 320      74
#> 321      60
#> 322      74
#> 323      32
#> 324      75
#> 325      64
#> 326      67
#> 327      76
#> 341      64
#> 342      45
#> 343      67
#> 344      65
#> 345      61
#> 346      67
#> 347      54
#> 348      64
#> 349      51
#> 350      67
#> 351      56
#> 352      63
#> 353      43
#> 105      19
#> 106      81
#> 289  171920
#> 290   40397
#> 99    87417
#> 100      48
#> 101      52
#> 102      78
#> 103   78766
#> 104      38
#> 287  110517
#> 288  101800
#> 9      <NA>
#> 10     <NA>
#> 355    1640
#> 356       1
#> 11     <NA>
#> 12     <NA>
#> 14     <NA>
#> 354     211
#> 1      <NA>
#> 2      <NA>
#> 3      <NA>
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 7      <NA>
#> 13        7
#>                                                                                          indicator
#> 79                                                                           Demographic Index USA
#> 80                                                              Supplemental Demographic Index USA
#> 87                                                                               % People of Color
#> 81                                                                                    % Low Income
#> 83                                                                                    % Unemployed
#> 82                                                        % in limited English-speaking Households
#> 84                                                          % with Less Than High School Education
#> 85                                                                                   % under Age 5
#> 86                                                                                   % over Age 64
#> 96                                                                         Demographic Index State
#> 97                                                            Supplemental Demographic Index State
#> 15                                                               Ratio to US avg Demographic Index
#> 16                   Ratio to US avg Supplemental Summary Index for Supplemental Demographic Index
#> 23                                                               Ratio to US avg % People of Color
#> 17                                                                    Ratio to US avg % Low Income
#> 19                                                                    Ratio to US avg % Unemployed
#> 18                                        Ratio to US avg % in limited English-speaking Households
#> 20                                          Ratio to US avg % with Less Than High School Education
#> 21                                                                   Ratio to US avg % under Age 5
#> 22                                                                   Ratio to US avg % over Age 64
#> 47                                                            Ratio to State avg Demographic Index
#> 48                Ratio to State avg Supplemental Summary Index for Supplemental Demographic Index
#> 55                                                            Ratio to State avg % People of Color
#> 49                                                                 Ratio to State avg % Low Income
#> 51                                                                 Ratio to State avg % Unemployed
#> 50                                     Ratio to State avg % in limited English-speaking Households
#> 52                                       Ratio to State avg % with Less Than High School Education
#> 53                                                                Ratio to State avg % under Age 5
#> 54                                                                Ratio to State avg % over Age 64
#> 136                                                            US percentile for Demographic Index
#> 137                US percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 144                                                            US percentile for % People of Color
#> 138                                                                 US percentile for % Low Income
#> 140                                                                 US percentile for % Unemployed
#> 139                                     US percentile for % in limited English-speaking Households
#> 141                                       US percentile for % with Less Than High School Education
#> 142                                                                US percentile for % under Age 5
#> 143                                                                US percentile for % over Age 64
#> 168                                                         State percentile for Demographic Index
#> 169             State percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 176                                                         State percentile for % People of Color
#> 170                                                              State percentile for % Low Income
#> 172                                                              State percentile for % Unemployed
#> 171                                  State percentile for % in limited English-speaking Households
#> 173                                    State percentile for % with Less Than High School Education
#> 174                                                             State percentile for % under Age 5
#> 175                                                             State percentile for % over Age 64
#> 200                                                               US Average for Demographic Index
#> 201                   US Average for Supplemental Summary Index for Supplemental Demographic Index
#> 208                                                               US Average for % People of Color
#> 202                                                                    US Average for % Low Income
#> 204                                                                    US Average for % Unemployed
#> 203                                        US Average for % in limited English-speaking Households
#> 205                                          US Average for % with Less Than High School Education
#> 206                                                                   US Average for % under Age 5
#> 207                                                                   US Average for % over Age 64
#> 232                                                            State Average for Demographic Index
#> 233                State Average for Supplemental Summary Index for Supplemental Demographic Index
#> 240                                                            State Average for % People of Color
#> 234                                                                 State Average for % Low Income
#> 236                                                                 State Average for % Unemployed
#> 235                                     State Average for % in limited English-speaking Households
#> 237                                       State Average for % with Less Than High School Education
#> 238                                                                State Average for % under Age 5
#> 239                                                                State Average for % over Age 64
#> 270                                                                 People of Color resident count
#> 264                                                                      Low income resident count
#> 266                                                                      Unemployed resident count
#> 265                                                            Limited English-speaking Households
#> 267                                                 Less Than High School Education resident count
#> 268                                                                     Under Age 5 resident count
#> 269                                                                     Over Age 64 resident count
#> 365                                                       % White (single race, includes Hispanic)
#> 359                                   % Black or African American (single race, includes Hispanic)
#> 360                                                       % Asian (single race, includes Hispanic)
#> 361                           % American Indian and Alaska Native (single race, includes Hispanic)
#> 362                  % Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 363                                                  % Other race (single race, includes Hispanic)
#> 364                                                        % Two or more races (includes Hispanic)
#> 95                                                             % White (non-Hispanic, single race)
#> 89                                         % Black or African American (non-Hispanic, single race)
#> 90                                                             % Asian (non-Hispanic, single race)
#> 88                                                                            % Hispanic or Latino
#> 91                                 % American Indian and Alaska Native (non-Hispanic, single race)
#> 92                        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 93                                                        % Other race (non-Hispanic, single race)
#> 94                                                              % Two or more races (non-Hispanic)
#> 31                                             Ratio to US avg % White (non-Hispanic, single race)
#> 25                         Ratio to US avg % Black or African American (non-Hispanic, single race)
#> 26                                             Ratio to US avg % Asian (non-Hispanic, single race)
#> 24                                                            Ratio to US avg % Hispanic or Latino
#> 27                 Ratio to US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 28        Ratio to US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 29                                        Ratio to US avg % Other race (non-Hispanic, single race)
#> 30                                              Ratio to US avg % Two or more races (non-Hispanic)
#> 63                                          Ratio to State avg % White (non-Hispanic, single race)
#> 57                      Ratio to State avg % Black or African American (non-Hispanic, single race)
#> 58                                          Ratio to State avg % Asian (non-Hispanic, single race)
#> 56                                                         Ratio to State avg % Hispanic or Latino
#> 59              Ratio to State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 60     Ratio to State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 61                                     Ratio to State avg % Other race (non-Hispanic, single race)
#> 62                                           Ratio to State avg % Two or more races (non-Hispanic)
#> 152                                          US percentile for % White (non-Hispanic, single race)
#> 146                      US percentile for % Black or African American (non-Hispanic, single race)
#> 147                                          US percentile for % Asian (non-Hispanic, single race)
#> 145                                                         US percentile for % Hispanic or Latino
#> 148              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 149     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 150                                     US percentile for % Other race (non-Hispanic, single race)
#> 151                                           US percentile for % Two or more races (non-Hispanic)
#> 184                                       State percentile for % White (non-Hispanic, single race)
#> 178                   State percentile for % Black or African American (non-Hispanic, single race)
#> 179                                       State percentile for % Asian (non-Hispanic, single race)
#> 177                                                      State percentile for % Hispanic or Latino
#> 180           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 181  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 182                                  State percentile for % Other race (non-Hispanic, single race)
#> 183                                        State percentile for % Two or more races (non-Hispanic)
#> 216                                             US Average for % White (non-Hispanic, single race)
#> 210                         US Average for % Black or African American (non-Hispanic, single race)
#> 211                                             US Average for % Asian (non-Hispanic, single race)
#> 209                                                            US Average for % Hispanic or Latino
#> 212                 US Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 213        US Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 214                                        US Average for % Other race (non-Hispanic, single race)
#> 215                                              US Average for % Two or more races (non-Hispanic)
#> 248                                          State Average for % White (non-Hispanic, single race)
#> 242                      State Average for % Black or African American (non-Hispanic, single race)
#> 243                                          State Average for % Asian (non-Hispanic, single race)
#> 241                                                         State Average for % Hispanic or Latino
#> 244              State Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 245     State Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 246                                     State Average for % Other race (non-Hispanic, single race)
#> 247                                           State Average for % Two or more races (non-Hispanic)
#> 278                                                     Count of White (non-Hispanic, single race)
#> 272                                 Count of Black or African American (non-Hispanic, single race)
#> 273                                                     Count of Asian (non-Hispanic, single race)
#> 271                                                                    Count of Hispanic or Latino
#> 274                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 275                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 276                                                Count of Other race (non-Hispanic, single race)
#> 277                                                      Count of Two or more races (non-Hispanic)
#> 107                                     Percent of population speaking Other Indo-European at home
#> 108                      Percent of population speaking Asian and Pacific Island languages at home
#> 109                         Percent of population speaking Other and Unspecified languages at home
#> 111                                   Percent of population speaking Non English languages at home
#> 110                                                 Percent of population speaking Spanish at home
#> 292                                                              Speak Other Indo-European at Home
#> 293                                                    Speak Asian-Pacific Island language at Home
#> 294                                        Number speaking Other and Unspecified languages at home
#> 291                                                            Number speaking non-English at home
#> 295                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 296                                                           Persons who Speak English Not at All
#> 297                                                                          Speak Spanish at Home
#> 112                                  Percent Speaking Spanish (as % of limited English households)
#> 113            Percent Speaking Other Indo-European languages (as % of limited English households)
#> 114           Percent Speaking Asian-Pacific Island languages (as % of limited English households)
#> 115                          Percent Speaking Other languages (as % of limited English households)
#> 298                                                   Speak Spanish (in limited English household)
#> 299                                       Speak Other Indo-European (in limited English household)
#> 300                                      Speak Asian-Pacific Island (in limited English household)
#> 301                                                     Speak Other (in limited English household)
#> 8                                                                                 Total Population
#> 280                                                                            Count of Households
#> 279                                                                  Count of Population Age 25 up
#> 281                                           Universe for percent unemployed (denominator, count)
#> 282                                                          Count of Housing Units Built Pre 1960
#> 283                                   Built housing units count (denominator for percent pre 1960)
#> 284                                      Count of Population for whom Poverty Status is Determined
#> 98                                                       Percent of Households below Poverty Level
#> 285                                                          Count of Owner Occupied Housing Units
#> 286                                                                 Households below Poverty Level
#> 116                                                                            % with Disabilities
#> 117                                                                           %Low life expectancy
#> 45                                                             Ratio to US avg % with Disabilities
#> 46                                                            Ratio to US avg %Low life expectancy
#> 77                                                          Ratio to State avg % with Disabilities
#> 78                                                         Ratio to State avg %Low life expectancy
#> 153                                                          US percentile for % with Disabilities
#> 154                                                         US percentile for %Low life expectancy
#> 185                                                       State percentile for % with Disabilities
#> 217                                                             US Average for % with Disabilities
#> 218                                                            US Average for %Low life expectancy
#> 249                                                          State Average for % with Disabilities
#> 250                                                         State Average for %Low life expectancy
#> 357                        Civilian noninstitutionalized population (denominator for % disability)
#> 358                                                               Count of Persons with Disability
#> 186                                                      State percentile for %Low life expectancy
#> 118                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 119                                                                                    Ozone (ppb)
#> 120                                                                         Nitrogen Dioxide (NO2)
#> 121                                                              Diesel Particulate Matter (ug/m3)
#> 122                                                                          Toxic Releases to Air
#> 123                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 124                                                     Lead Paint Indicator (% pre-1960s housing)
#> 125                                                   Superfund Proximity (site count/km distance)
#> 126                                                     RMP Proximity (facility count/km distance)
#> 127                                         Hazardous Waste Proximity (facility count/km distance)
#> 128                                                      Underground Storage Tanks (UST) indicator
#> 129                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 130                                                                  Drinking Water Non-Compliance
#> 32                                                              Ratio to US avg Particulate Matter
#> 33                                                                           Ratio to US avg Ozone
#> 34                                                          Ratio to US avg Nitrogen Dioxide (NO2)
#> 35                                                       Ratio to US avg Diesel Particulate Matter
#> 36                                                           Ratio to US avg Toxic Releases to Air
#> 37                                                    Ratio to US avg Traffic Proximity and Volume
#> 38                                                            Ratio to US avg Lead Paint Indicator
#> 39                                                             Ratio to US avg Superfund Proximity
#> 40                                                                   Ratio to US avg RMP Proximity
#> 41                                                       Ratio to US avg Hazardous Waste Proximity
#> 42                                             Ratio to US avg Underground Storage Tanks indicator
#> 43                                                  Ratio to US avg Wastewater Discharge Indicator
#> 44                                                   Ratio to US avg Drinking Water Non-Compliance
#> 64                                                           Ratio to State avg Particulate Matter
#> 65                                                                        Ratio to State avg Ozone
#> 66                                                       Ratio to State avg Nitrogen Dioxide (NO2)
#> 67                                                    Ratio to State avg Diesel Particulate Matter
#> 68                                                        Ratio to State avg Toxic Releases to Air
#> 69                                                 Ratio to State avg Traffic Proximity and Volume
#> 70                                                         Ratio to State avg Lead Paint Indicator
#> 71                                                          Ratio to State avg Superfund Proximity
#> 72                                                                Ratio to State avg RMP Proximity
#> 73                                                    Ratio to State avg Hazardous Waste Proximity
#> 74                                          Ratio to State avg Underground Storage Tanks indicator
#> 75                                               Ratio to State avg Wastewater Discharge Indicator
#> 76                                                Ratio to State avg Drinking Water Non-Compliance
#> 155                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 156                                                                  US percentile for Ozone (ppb)
#> 157                                                       US percentile for Nitrogen Dioxide (NO2)
#> 158                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 159                                                        US percentile for Toxic Releases to Air
#> 160          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 161                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 162                                 US percentile for Superfund Proximity (site count/km distance)
#> 163                                   US percentile for RMP Proximity (facility count/km distance)
#> 164                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 165                                    US percentile for Underground Storage Tanks (UST) indicator
#> 166    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 167                                                US percentile for Drinking Water Non-Compliance
#> 187                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 188                                                               State percentile for Ozone (ppb)
#> 189                                                    State percentile for Nitrogen Dioxide (NO2)
#> 190                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 191                                                     State percentile for Toxic Releases to Air
#> 192       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 193                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 194                              State percentile for Superfund Proximity (site count/km distance)
#> 195                                State percentile for RMP Proximity (facility count/km distance)
#> 196                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 197                                 State percentile for Underground Storage Tanks (UST) indicator
#> 198 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 199                                             State percentile for Drinking Water Non-Compliance
#> 219                                            US Average for Particulate Matter (PM 2.5 in ug/m3)
#> 220                                                                     US Average for Ozone (ppb)
#> 221                                                          US Average for Nitrogen Dioxide (NO2)
#> 222                                               US Average for Diesel Particulate Matter (ug/m3)
#> 223                                                           US Average for Toxic Releases to Air
#> 224             US Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 225                                      US Average for Lead Paint Indicator (% pre-1960s housing)
#> 226                                    US Average for Superfund Proximity (site count/km distance)
#> 227                                      US Average for RMP Proximity (facility count/km distance)
#> 228                          US Average for Hazardous Waste Proximity (facility count/km distance)
#> 229                                       US Average for Underground Storage Tanks (UST) indicator
#> 230       US Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 231                                                   US Average for Drinking Water Non-Compliance
#> 251                                         State Average for Particulate Matter (PM 2.5 in ug/m3)
#> 252                                                                  State Average for Ozone (ppb)
#> 253                                                       State Average for Nitrogen Dioxide (NO2)
#> 254                                            State Average for Diesel Particulate Matter (ug/m3)
#> 255                                                        State Average for Toxic Releases to Air
#> 256          State Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 257                                   State Average for Lead Paint Indicator (% pre-1960s housing)
#> 258                                 State Average for Superfund Proximity (site count/km distance)
#> 259                                   State Average for RMP Proximity (facility count/km distance)
#> 260                       State Average for Hazardous Waste Proximity (facility count/km distance)
#> 261                                    State Average for Underground Storage Tanks (UST) indicator
#> 262    State Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 263                                                State Average for Drinking Water Non-Compliance
#> 132                                                          Number of Sites Nearby (total unique)
#> 135                                                        Distance to Closest Site for Avg Person
#> 131                                                                   Number of Sites Nearby (max)
#> 133                                                                   Number of Sites Nearby (avg)
#> 134                                                                       Distance to Closest Site
#> 366                                      US type of raw score for Particulate Matter Summary Index
#> 367                                                   US type of raw score for Ozone Summary Index
#> 368                                  US type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 369                               US type of raw score for Diesel Particulate Matter Summary Index
#> 370                                   US type of raw score for Toxic Releases to Air Summary Index
#> 371                            US type of raw score for Traffic Proximity and Volume Summary Index
#> 372                                              US type of raw score for Lead Paint Summary Index
#> 373                                     US type of raw score for Superfund Proximity Summary Index
#> 374                                           US type of raw score for RMP Proximity Summary Index
#> 375                               US type of raw score for Hazardous Waste Proximity Summary Index
#> 376                               US type of raw score for Underground Storage Tanks Summary Index
#> 377                                    US type of raw score for Wastewater Discharge Summary Index
#> 378                           US type of raw score for Drinking Water Non-Compliance Summary Index
#> 379                                   State type of raw score for Particulate Matter Summary Index
#> 380                                                State type of raw score for Ozone Summary Index
#> 381                               State type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 382                            State type of raw score for Diesel Particulate Matter Summary Index
#> 383                                State type of raw score for Toxic Releases to Air Summary Index
#> 384                         State type of raw score for Traffic Proximity and Volume Summary Index
#> 385                                           State type of raw score for Lead Paint Summary Index
#> 386                                  State type of raw score for Superfund Proximity Summary Index
#> 387                                        State type of raw score for RMP Proximity Summary Index
#> 388                            State type of raw score for Hazardous Waste Proximity Summary Index
#> 389                            State type of raw score for Underground Storage Tanks Summary Index
#> 390                                 State type of raw score for Wastewater Discharge Summary Index
#> 391                        State type of raw score for Drinking Water Non-Compliance Summary Index
#> 302                                         US percentile for Summary Index for Particulate Matter
#> 303                                                      US percentile for Summary Index for Ozone
#> 304                                     US percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 305                                  US percentile for Summary Index for Diesel Particulate Matter
#> 306                                      US percentile for Summary Index for Toxic Releases to Air
#> 307                               US percentile for Summary Index for Traffic Proximity and Volume
#> 308                                       US percentile for Summary Index for Lead Paint Indicator
#> 309                                        US percentile for Summary Index for Superfund Proximity
#> 310                                              US percentile for Summary Index for RMP Proximity
#> 311                                  US percentile for Summary Index for Hazardous Waste Proximity
#> 312                  US percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 313                             US percentile for Summary Index for Wastewater Discharge Indicator
#> 314                              US percentile for Summary Index for Drinking Water Non-Compliance
#> 328                                      State percentile for Summary Index for Particulate Matter
#> 329                                                   State percentile for Summary Index for Ozone
#> 330                                  State percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 331                               State percentile for Summary Index for Diesel Particulate Matter
#> 332                                   State percentile for Summary Index for Toxic Releases to Air
#> 333                            State percentile for Summary Index for Traffic Proximity and Volume
#> 334                                    State percentile for Summary Index for Lead Paint Indicator
#> 335                                     State percentile for Summary Index for Superfund Proximity
#> 336                                           State percentile for Summary Index for RMP Proximity
#> 337                               State percentile for Summary Index for Hazardous Waste Proximity
#> 338               State percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 339                          State percentile for Summary Index for Wastewater Discharge Indicator
#> 340                           State percentile for Summary Index for Drinking Water Non-Compliance
#> 392                         US type of raw score for Particulate Matter Supplemental Summary Index
#> 393                                      US type of raw score for Ozone Supplemental Summary Index
#> 394                     US type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 395                  US type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 396                      US type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 397               US type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 398                                 US type of raw score for Lead Paint Supplemental Summary Index
#> 399                        US type of raw score for Superfund Proximity Supplemental Summary Index
#> 400                              US type of raw score for RMP Proximity Supplemental Summary Index
#> 401                  US type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 402                  US type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 403                       US type of raw score for Wastewater Discharge Supplemental Summary Index
#> 404              US type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 405                      State type of raw score for Particulate Matter Supplemental Summary Index
#> 406                                   State type of raw score for Ozone Supplemental Summary Index
#> 407                  State type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 408               State type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 409                   State type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 410            State type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 411                              State type of raw score for Lead Paint Supplemental Summary Index
#> 412                     State type of raw score for Superfund Proximity Supplemental Summary Index
#> 413                           State type of raw score for RMP Proximity Supplemental Summary Index
#> 414               State type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 415               State type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 416                    State type of raw score for Wastewater Discharge Supplemental Summary Index
#> 417           State type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 315                            US percentile for Supplemental Summary Index for Particulate Matter
#> 316                                         US percentile for Supplemental Summary Index for Ozone
#> 317                        US percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 318                     US percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 319                         US percentile for Supplemental Summary Index for Toxic Releases to Air
#> 320                  US percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 321                          US percentile for Supplemental Summary Index for Lead Paint Indicator
#> 322                           US percentile for Supplemental Summary Index for Superfund Proximity
#> 323                                 US percentile for Supplemental Summary Index for RMP Proximity
#> 324                     US percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 325     US percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 326                US percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 327                 US percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 341                         State percentile for Supplemental Summary Index for Particulate Matter
#> 342                                      State percentile for Supplemental Summary Index for Ozone
#> 343                     State percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 344                  State percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 345                      State percentile for Supplemental Summary Index for Toxic Releases to Air
#> 346               State percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 347                       State percentile for Supplemental Summary Index for Lead Paint Indicator
#> 348                        State percentile for Supplemental Summary Index for Superfund Proximity
#> 349                              State percentile for Supplemental Summary Index for RMP Proximity
#> 350                  State percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 351  State percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 352             State percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 353              State percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 105                                                                           Percent under age 18
#> 106                                                                           Percent above age 17
#> 289                                                                         Population Over Age 17
#> 290                                                                        Population Under Age 18
#> 99                                                                          Occupied Housing Units
#> 100                                                                                  Percent Males
#> 101                                                                                Percent Females
#> 102                                                                       Life Expectancy in Years
#> 103                                                                              Per Capita Income
#> 104                                                              Percent Owner Occupied households
#> 287                                                                              Female Population
#> 288                                                                                Male Population
#> 9                                                                                        Longitude
#> 10                                                                                        Latitude
#> 355                                                                          Census blocks (count)
#> 356                                                                        Buffer distance (miles)
#> 11                                                                              State Abbreviation
#> 12                                                                                      State Name
#> 14                                                                                      EPA Region
#> 354                                                                     Census blockgroups (count)
#> 1                                                                                  EJScreen Report
#> 2                                                                                     EJScreen Map
#> 3                                                                                       ACS Report
#> 4                                                                                      ECHO Report
#> 5                                                                                     ejam_uniq_id
#> 6                                                                                            valid
#> 7                                                                                      invalid_msg
#> 13                                                                              in_how_many_states