Get order of variable names to sort by, as seen in EJScreen Community Report

Get order of variable names to sort by, as seen in EJScreen Community Report

Usage

table_order_variables(
  varnames,
  s1 = "ejscreensort",
  s2 = "sortvarlistEJSCREENREPORT",
  s3 = "sort_within_varlistEJSCREENREPORT"
)

Arguments

varnames

vector of indicator variables names from blockgroupstats, bgej, etc., such as "pm", "pctlowinc", "pctile.EJ.DISPARITY.traffic.score.eo" etc. and others as found in names_all_r, or specific subsets of those like in c(names_d, names_d_subgroups, names_e) and c(names_ej_pctile, names_ej_state_pctile, names_ej_supp_pctile, names_ej_supp_state_pctile)

s1

name of column in map_headernames to get sort info from

s2

optional like s1 but secondary to s1

s3

optional tertiary

Value

vector as from order(), to be used in sorting a data.frame for example

Examples

  cbind(table_order_variables(c(names_d, names_d_subgroups, names_e)))
#>       [,1]
#>  [1,]    1
#>  [2,]    2
#>  [3,]    9
#>  [4,]    3
#>  [5,]    5
#>  [6,]    4
#>  [7,]    6
#>  [8,]    7
#>  [9,]    8
#> [10,]   17
#> [11,]   11
#> [12,]   12
#> [13,]   10
#> [14,]   13
#> [15,]   14
#> [16,]   15
#> [17,]   16
#> [18,]   18
#> [19,]   19
#> [20,]   20
#> [21,]   21
#> [22,]   22
#> [23,]   23
#> [24,]   24
#> [25,]   25
#> [26,]   26
#> [27,]   27
#> [28,]   28
#> [29,]   29
#> [30,]   30
  
  out <- testoutput_ejamit_10pts_1miles
  vars <- out$formatted[ , 'indicator']
  vars <- fixcolnames(vars, 'long', 'r')
  out$formatted[table_order_variables(vars), ]
#>           value
#> 79         1.35
#> 80         1.28
#> 87           53
#> 81           23
#> 83            6
#> 82            6
#> 84            9
#> 85            6
#> 86           13
#> 96            1
#> 97            1
#> 15            1
#> 16          0.8
#> 23          1.3
#> 17          0.7
#> 19          1.2
#> 18          1.2
#> 20          0.8
#> 21          1.2
#> 22          0.7
#> 47            1
#> 48          0.8
#> 55          1.2
#> 49          0.8
#> 51          1.1
#> 50          0.8
#> 52          0.7
#> 53          1.2
#> 54          0.7
#> 167          58
#> 168          33
#> 175          67
#> 169          43
#> 171          68
#> 170          77
#> 172          54
#> 173          65
#> 174          35
#> 204          57
#> 205          41
#> 212          62
#> 206          50
#> 208          63
#> 207          67
#> 209          49
#> 210          66
#> 211          36
#> 241        1.34
#> 242        1.64
#> 249          40
#> 243          30
#> 245           6
#> 244           5
#> 246          11
#> 247           5
#> 248          18
#> 273        1.35
#> 274        1.47
#> 281          44
#> 275          27
#> 277           6
#> 276           7
#> 278          12
#> 279           5
#> 280          17
#> 311      110000
#> 305       46000
#> 307        7700
#> 306        5000
#> 308       13000
#> 309       14000
#> 310       27000
#> 416          53
#> 410          17
#> 411          12
#> 412           0
#> 413           0
#> 414           7
#> 415          11
#> 326       1e+05
#> 320       40000
#> 321       30000
#> 322         800
#> 323         700
#> 324       10000
#> 325       20000
#> 95           47
#> 89           16
#> 90           12
#> 88           19
#> 91            0
#> 92            0
#> 93            1
#> 94            5
#> 31          0.8
#> 25          1.3
#> 26          2.4
#> 24            1
#> 27          0.3
#> 28          2.2
#> 29          1.6
#> 30          1.5
#> 63          0.9
#> 57          1.4
#> 58          1.4
#> 56          0.9
#> 59          0.8
#> 60          4.9
#> 61            1
#> 62          1.6
#> 183          33
#> 177          77
#> 178          87
#> 176          70
#> 179          84
#> 180          95
#> 181          88
#> 182          77
#> 220          39
#> 214          72
#> 215          74
#> 213          61
#> 216          88
#> 217          72
#> 218          81
#> 219          77
#> 257          60
#> 251          12
#> 252           5
#> 250          18
#> 253           1
#> 254           0
#> 255           0
#> 256           3
#> 289          54
#> 283          12
#> 284           9
#> 282          20
#> 285           0
#> 286           0
#> 287           1
#> 288           3
#> 319       1e+05
#> 313       30000
#> 314       30000
#> 312       40000
#> 315         400
#> 316         700
#> 317        1000
#> 318       10000
#> 107           7
#> 108           6
#> 109           2
#> 118          28
#> 110           1
#> 111          13
#> 112           0
#> 113           0
#> 114           0
#> 115           0
#> 116           0
#> 117           0
#> 340       42874
#> 341       35830
#> 342        9800
#> 339      160000
#> 343      575168
#> 344        3633
#> 345       72060
#> 119          47
#> 120          25
#> 121          27
#> 122           1
#> 346        2343
#> 347        1257
#> 348        1330
#> 349          68
#> 8        212317
#> 328       87417
#> 327      151161
#> 329      175609
#> 330       52699
#> 331       95293
#> 332      202460
#> 98           12
#> 333       33361
#> 334       10660
#> 123           9
#> 124          17
#> 125           4
#> 126         9.9
#> 127         4.9
#> 45          0.7
#> 46          0.9
#> 77          0.8
#> 78            1
#> 184          26
#> 185          25
#> 186        <NA>
#> 187        <NA>
#> 188        <NA>
#> 221          35
#> 223        <NA>
#> 224        <NA>
#> 225        <NA>
#> 258        13.6
#> 259          20
#> 290        11.9
#> 291          18
#> 405       20000
#> 406      210565
#> 130           1
#> 128           1
#> 129           1
#> 131           8
#> 132           1
#> 222          43
#> 135        8.41
#> 136        62.7
#> 137        15.8
#> 138       0.591
#> 139         540
#> 140     5400000
#> 141        55.3
#> 142        1.61
#> 143        0.13
#> 144        18.7
#> 145        14.4
#> 146       10000
#> 147           6
#> 32            1
#> 33            1
#> 34            2
#> 35          3.1
#> 36          0.1
#> 37          3.2
#> 38          1.8
#> 39          4.1
#> 40          0.2
#> 41          5.3
#> 42          3.9
#> 43            0
#> 44          2.9
#> 64            1
#> 65            1
#> 66          1.6
#> 67          1.9
#> 68          0.7
#> 69          1.8
#> 70          1.2
#> 71          1.8
#> 72          0.4
#> 73          2.5
#> 74            2
#> 75          0.9
#> 76          1.6
#> 191          59
#> 192          61
#> 193          97
#> 194          97
#> 195          48
#> 196          92
#> 197          77
#> 198          95
#> 199          38
#> 200          96
#> 201          93
#> 202          88
#> 203          91
#> 228          69
#> 229          44
#> 230          81
#> 231          76
#> 232          69
#> 233          77
#> 234          55
#> 235          72
#> 236          53
#> 237          80
#> 238          69
#> 239          69
#> 240          37
#> 260        8.45
#> 261        61.8
#> 262         7.8
#> 263       0.191
#> 264        4600
#> 265     1700000
#> 266         0.3
#> 267        0.39
#> 268        0.57
#> 269         3.5
#> 270         3.6
#> 271       7e+05
#> 272         2.2
#> 292        8.13
#> 293        63.3
#> 294         9.7
#> 295       0.318
#> 296         810
#> 297     2900000
#> 298        0.48
#> 299        0.89
#> 300        0.34
#> 301         7.5
#> 302         7.2
#> 303       11000
#> 304        3.97
#> 133           6
#> 134           2
#> 189        <NA>
#> 190        <NA>
#> 226        <NA>
#> 227        <NA>
#> 163          10
#> 166         0.7
#> 162           1
#> 164           1
#> 165           0
#> 417      72.662
#> 418      66.855
#> 419     117.641
#> 420     116.821
#> 421      59.572
#> 422     118.693
#> 423      94.667
#> 424     108.065
#> 425      36.502
#> 426     122.999
#> 427      99.994
#> 428     105.869
#> 429      80.283
#> 430      91.953
#> 431      48.636
#> 432     104.733
#> 433      97.954
#> 434       91.73
#> 435     101.953
#> 436      76.136
#> 437      92.378
#> 438      54.627
#> 439     106.014
#> 440      83.854
#> 441      91.479
#> 442       46.43
#> 350          63
#> 351          62
#> 352          76
#> 353          77
#> 354          58
#> 355          77
#> 356          73
#> 357          80
#> 358          45
#> 359          78
#> 360          74
#> 361          76
#> 362          83
#> 376          66
#> 377          48
#> 378          68
#> 379          65
#> 380          66
#> 381          68
#> 382          60
#> 383          64
#> 384          58
#> 385          68
#> 386          59
#> 387          66
#> 388          45
#> 443      66.552
#> 444      62.937
#> 445     110.354
#> 446     109.369
#> 447      56.357
#> 448     110.815
#> 449      91.252
#> 450     101.699
#> 451      35.509
#> 452     114.012
#> 453      94.901
#> 454      99.413
#> 455      78.077
#> 456      85.146
#> 457      45.359
#> 458      97.476
#> 459      92.088
#> 460      85.865
#> 461      93.448
#> 462      71.343
#> 463      87.488
#> 464      53.506
#> 465      97.655
#> 466      80.199
#> 467      84.606
#> 468      47.177
#> 363          47
#> 364          46
#> 365          72
#> 366          73
#> 367          41
#> 368          74
#> 369          60
#> 370          74
#> 371          32
#> 372          75
#> 373          64
#> 374          67
#> 375          76
#> 389          64
#> 390          38
#> 391          67
#> 392          65
#> 393          61
#> 394          67
#> 395          54
#> 396          64
#> 397          51
#> 398          67
#> 399          56
#> 400          63
#> 401          43
#> 148          69
#> 150          72
#> 149           9
#> 105          19
#> 106          81
#> 337      171920
#> 338       40397
#> 99        87417
#> 100          48
#> 101          52
#> 102          78
#> 103       78766
#> 104          38
#> 335      110517
#> 336      101800
#> 151           0
#> 156           1
#> 157          18
#> 158         112
#> 159         163
#> 161          28
#> 160           5
#> 152           1
#> 153           1
#> 155           1
#> 154           1
#> 9          <NA>
#> 403        1640
#> 404           1
#> 469        31.4
#> 10         <NA>
#> 11         <NA>
#> 12         <NA>
#> 14         <NA>
#> 402         211
#> 407  11335817.4
#> 408 205467995.8
#> 409           0
#> 1          <NA>
#> 2          <NA>
#> 3          <NA>
#> 4          <NA>
#> 5          <NA>
#> 6          <NA>
#> 7          <NA>
#> 13            7
#>                                                                                          indicator
#> 79                                                                           Demographic Index USA
#> 80                                                              Supplemental Demographic Index USA
#> 87                                                                               % People of Color
#> 81                                                                                    % Low Income
#> 83                                                                                    % Unemployed
#> 82                                                        % in limited English-speaking Households
#> 84                                                          % with Less Than High School Education
#> 85                                                                                   % under Age 5
#> 86                                                                                   % over Age 64
#> 96                                                                         Demographic Index State
#> 97                                                            Supplemental Demographic Index State
#> 15                                                               Ratio to US avg Demographic Index
#> 16                   Ratio to US avg Supplemental Summary Index for Supplemental Demographic Index
#> 23                                                               Ratio to US avg % People of Color
#> 17                                                                    Ratio to US avg % Low Income
#> 19                                                                    Ratio to US avg % Unemployed
#> 18                                        Ratio to US avg % in limited English-speaking Households
#> 20                                          Ratio to US avg % with Less Than High School Education
#> 21                                                                   Ratio to US avg % under Age 5
#> 22                                                                   Ratio to US avg % over Age 64
#> 47                                                            Ratio to State avg Demographic Index
#> 48                Ratio to State avg Supplemental Summary Index for Supplemental Demographic Index
#> 55                                                            Ratio to State avg % People of Color
#> 49                                                                 Ratio to State avg % Low Income
#> 51                                                                 Ratio to State avg % Unemployed
#> 50                                     Ratio to State avg % in limited English-speaking Households
#> 52                                       Ratio to State avg % with Less Than High School Education
#> 53                                                                Ratio to State avg % under Age 5
#> 54                                                                Ratio to State avg % over Age 64
#> 167                                                            US percentile for Demographic Index
#> 168                US percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 175                                                            US percentile for % People of Color
#> 169                                                                 US percentile for % Low Income
#> 171                                                                 US percentile for % Unemployed
#> 170                                     US percentile for % in limited English-speaking Households
#> 172                                       US percentile for % with Less Than High School Education
#> 173                                                                US percentile for % under Age 5
#> 174                                                                US percentile for % over Age 64
#> 204                                                         State percentile for Demographic Index
#> 205             State percentile for Supplemental Summary Index for Supplemental Demographic Index
#> 212                                                         State percentile for % People of Color
#> 206                                                              State percentile for % Low Income
#> 208                                                              State percentile for % Unemployed
#> 207                                  State percentile for % in limited English-speaking Households
#> 209                                    State percentile for % with Less Than High School Education
#> 210                                                             State percentile for % under Age 5
#> 211                                                             State percentile for % over Age 64
#> 241                                                               US Average for Demographic Index
#> 242                   US Average for Supplemental Summary Index for Supplemental Demographic Index
#> 249                                                               US Average for % People of Color
#> 243                                                                    US Average for % Low Income
#> 245                                                                    US Average for % Unemployed
#> 244                                        US Average for % in limited English-speaking Households
#> 246                                          US Average for % with Less Than High School Education
#> 247                                                                   US Average for % under Age 5
#> 248                                                                   US Average for % over Age 64
#> 273                                                            State Average for Demographic Index
#> 274                State Average for Supplemental Summary Index for Supplemental Demographic Index
#> 281                                                            State Average for % People of Color
#> 275                                                                 State Average for % Low Income
#> 277                                                                 State Average for % Unemployed
#> 276                                     State Average for % in limited English-speaking Households
#> 278                                       State Average for % with Less Than High School Education
#> 279                                                                State Average for % under Age 5
#> 280                                                                State Average for % over Age 64
#> 311                                                                 People of Color resident count
#> 305                                                                      Low income resident count
#> 307                                                                      Unemployed resident count
#> 306                                                            Limited English-speaking Households
#> 308                                                 Less Than High School Education resident count
#> 309                                                                     Under Age 5 resident count
#> 310                                                                     Over Age 64 resident count
#> 416                                                       % White (single race, includes Hispanic)
#> 410                                   % Black or African American (single race, includes Hispanic)
#> 411                                                       % Asian (single race, includes Hispanic)
#> 412                           % American Indian and Alaska Native (single race, includes Hispanic)
#> 413                  % Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 414                                                  % Other race (single race, includes Hispanic)
#> 415                                                        % Two or more races (includes Hispanic)
#> 326                                                Count of White (single race, includes Hispanic)
#> 320                            Count of Black or African American (single race, includes Hispanic)
#> 321                                                Count of Asian (single race, includes Hispanic)
#> 322                    Count of American Indian and Alaska Native (single race, includes Hispanic)
#> 323           Count of Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 324                                           Count of Other race (single race, includes Hispanic)
#> 325                                                 Count of Two or more races (includes Hispanic)
#> 95                                                             % White (non-Hispanic, single race)
#> 89                                         % Black or African American (non-Hispanic, single race)
#> 90                                                             % Asian (non-Hispanic, single race)
#> 88                                                                            % Hispanic or Latino
#> 91                                 % American Indian and Alaska Native (non-Hispanic, single race)
#> 92                        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 93                                                        % Other race (non-Hispanic, single race)
#> 94                                                              % Two or more races (non-Hispanic)
#> 31                                             Ratio to US avg % White (non-Hispanic, single race)
#> 25                         Ratio to US avg % Black or African American (non-Hispanic, single race)
#> 26                                             Ratio to US avg % Asian (non-Hispanic, single race)
#> 24                                                            Ratio to US avg % Hispanic or Latino
#> 27                 Ratio to US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 28        Ratio to US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 29                                        Ratio to US avg % Other race (non-Hispanic, single race)
#> 30                                              Ratio to US avg % Two or more races (non-Hispanic)
#> 63                                          Ratio to State avg % White (non-Hispanic, single race)
#> 57                      Ratio to State avg % Black or African American (non-Hispanic, single race)
#> 58                                          Ratio to State avg % Asian (non-Hispanic, single race)
#> 56                                                         Ratio to State avg % Hispanic or Latino
#> 59              Ratio to State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 60     Ratio to State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 61                                     Ratio to State avg % Other race (non-Hispanic, single race)
#> 62                                           Ratio to State avg % Two or more races (non-Hispanic)
#> 183                                          US percentile for % White (non-Hispanic, single race)
#> 177                      US percentile for % Black or African American (non-Hispanic, single race)
#> 178                                          US percentile for % Asian (non-Hispanic, single race)
#> 176                                                         US percentile for % Hispanic or Latino
#> 179              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 180     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 181                                     US percentile for % Other race (non-Hispanic, single race)
#> 182                                           US percentile for % Two or more races (non-Hispanic)
#> 220                                       State percentile for % White (non-Hispanic, single race)
#> 214                   State percentile for % Black or African American (non-Hispanic, single race)
#> 215                                       State percentile for % Asian (non-Hispanic, single race)
#> 213                                                      State percentile for % Hispanic or Latino
#> 216           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 217  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 218                                  State percentile for % Other race (non-Hispanic, single race)
#> 219                                        State percentile for % Two or more races (non-Hispanic)
#> 257                                             US Average for % White (non-Hispanic, single race)
#> 251                         US Average for % Black or African American (non-Hispanic, single race)
#> 252                                             US Average for % Asian (non-Hispanic, single race)
#> 250                                                            US Average for % Hispanic or Latino
#> 253                 US Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 254        US Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 255                                        US Average for % Other race (non-Hispanic, single race)
#> 256                                              US Average for % Two or more races (non-Hispanic)
#> 289                                          State Average for % White (non-Hispanic, single race)
#> 283                      State Average for % Black or African American (non-Hispanic, single race)
#> 284                                          State Average for % Asian (non-Hispanic, single race)
#> 282                                                         State Average for % Hispanic or Latino
#> 285              State Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 286     State Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 287                                     State Average for % Other race (non-Hispanic, single race)
#> 288                                           State Average for % Two or more races (non-Hispanic)
#> 319                                                     Count of White (non-Hispanic, single race)
#> 313                                 Count of Black or African American (non-Hispanic, single race)
#> 314                                                     Count of Asian (non-Hispanic, single race)
#> 312                                                                    Count of Hispanic or Latino
#> 315                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 316                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 317                                                Count of Other race (non-Hispanic, single race)
#> 318                                                      Count of Two or more races (non-Hispanic)
#> 107                                     Percent of population speaking Other Indo-European at home
#> 108                      Percent of population speaking Asian and Pacific Island languages at home
#> 109                         Percent of population speaking Other and Unspecified languages at home
#> 118                                   Percent of population speaking Non English languages at home
#> 110                                                 Percent of population speaking English at home
#> 111                                                 Percent of population speaking Spanish at home
#> 112                                                  Percent of population speaking French at home
#> 113                         Percent of population speaking Russian, Polish or Other Slavic at home
#> 114                                           Percent of population speaking Indo-European at home
#> 115                                              Percent of population speaking Vietnamese at home
#> 116                Percent of population speaking Other Asian and Pacific Island languages at home
#> 117                                                  Percent of population speaking Arabic at home
#> 340                                                              Speak Other Indo-European at Home
#> 341                                                    Speak Asian-Pacific Island language at Home
#> 342                                        Number speaking Other and Unspecified languages at home
#> 339                                                            Number speaking non-English at home
#> 343                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 344                                                           Persons who Speak English Not at All
#> 345                                                                          Speak Spanish at Home
#> 119                                  Percent Speaking Spanish (as % of limited English households)
#> 120            Percent Speaking Other Indo-European languages (as % of limited English households)
#> 121           Percent Speaking Asian-Pacific Island languages (as % of limited English households)
#> 122                          Percent Speaking Other languages (as % of limited English households)
#> 346                                                   Speak Spanish (in limited English household)
#> 347                                       Speak Other Indo-European (in limited English household)
#> 348                                      Speak Asian-Pacific Island (in limited English household)
#> 349                                                     Speak Other (in limited English household)
#> 8                                                                                 Total Population
#> 328                                                                            Count of Households
#> 327                                                                  Count of Population Age 25 up
#> 329                                           Universe for percent unemployed (denominator, count)
#> 330                                                          Count of Housing Units Built Pre 1960
#> 331                                   Built housing units count (denominator for percent pre 1960)
#> 332                                      Count of Population for whom Poverty Status is Determined
#> 98                                                       Percent of Households below Poverty Level
#> 333                                                          Count of Owner Occupied Housing Units
#> 334                                                                 Households below Poverty Level
#> 123                                                                            % with Disabilities
#> 124                                                                           %Low life expectancy
#> 125                                                                          % with Heart Diseases
#> 126                                                                        % of Adults with Asthma
#> 127                                                % of Adults with Cancer (excluding skin cancer)
#> 45                                                             Ratio to US avg % with Disabilities
#> 46                                                            Ratio to US avg %Low life expectancy
#> 77                                                          Ratio to State avg % with Disabilities
#> 78                                                         Ratio to State avg %Low life expectancy
#> 184                                                          US percentile for % with Disabilities
#> 185                                                         US percentile for %Low life expectancy
#> 186                                                        US percentile for % with Heart Diseases
#> 187                                                      US percentile for % of Adults with Asthma
#> 188                              US percentile for % of Adults with Cancer (excluding skin cancer)
#> 221                                                       State percentile for % with Disabilities
#> 223                                                     State percentile for % with Heart Diseases
#> 224                                                   State percentile for % of Adults with Asthma
#> 225                           State percentile for % of Adults with Cancer (excluding skin cancer)
#> 258                                                             US Average for % with Disabilities
#> 259                                                            US Average for %Low life expectancy
#> 290                                                          State Average for % with Disabilities
#> 291                                                         State Average for %Low life expectancy
#> 405                                                               Count of Persons with Disability
#> 406                        Civilian noninstitutionalized population (denominator for % disability)
#> 130                                                    Flag for Overlapping with Food Desert Areas
#> 128                                           Flag for Overlapping with Housing Burden Communities
#> 129                             Flag for Overlapping with Transportation Disadvantaged Communities
#> 131                                                        % Households without Broadband Internet
#> 132                                                          % Households without Health Insurance
#> 222                                                      State percentile for %Low life expectancy
#> 135                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 136                                                                                    Ozone (ppb)
#> 137                                                                         Nitrogen Dioxide (NO2)
#> 138                                                              Diesel Particulate Matter (ug/m3)
#> 139                                                                          Toxic Releases to Air
#> 140                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 141                                                     Lead Paint Indicator (% pre-1960s housing)
#> 142                                                   Superfund Proximity (site count/km distance)
#> 143                                                     RMP Proximity (facility count/km distance)
#> 144                                         Hazardous Waste Proximity (facility count/km distance)
#> 145                                                      Underground Storage Tanks (UST) indicator
#> 146                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 147                                                                  Drinking Water Non-Compliance
#> 32                                                              Ratio to US avg Particulate Matter
#> 33                                                                           Ratio to US avg Ozone
#> 34                                                          Ratio to US avg Nitrogen Dioxide (NO2)
#> 35                                                       Ratio to US avg Diesel Particulate Matter
#> 36                                                           Ratio to US avg Toxic Releases to Air
#> 37                                                    Ratio to US avg Traffic Proximity and Volume
#> 38                                                            Ratio to US avg Lead Paint Indicator
#> 39                                                             Ratio to US avg Superfund Proximity
#> 40                                                                   Ratio to US avg RMP Proximity
#> 41                                                       Ratio to US avg Hazardous Waste Proximity
#> 42                                             Ratio to US avg Underground Storage Tanks indicator
#> 43                                                  Ratio to US avg Wastewater Discharge Indicator
#> 44                                                   Ratio to US avg Drinking Water Non-Compliance
#> 64                                                           Ratio to State avg Particulate Matter
#> 65                                                                        Ratio to State avg Ozone
#> 66                                                       Ratio to State avg Nitrogen Dioxide (NO2)
#> 67                                                    Ratio to State avg Diesel Particulate Matter
#> 68                                                        Ratio to State avg Toxic Releases to Air
#> 69                                                 Ratio to State avg Traffic Proximity and Volume
#> 70                                                         Ratio to State avg Lead Paint Indicator
#> 71                                                          Ratio to State avg Superfund Proximity
#> 72                                                                Ratio to State avg RMP Proximity
#> 73                                                    Ratio to State avg Hazardous Waste Proximity
#> 74                                          Ratio to State avg Underground Storage Tanks indicator
#> 75                                               Ratio to State avg Wastewater Discharge Indicator
#> 76                                                Ratio to State avg Drinking Water Non-Compliance
#> 191                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 192                                                                  US percentile for Ozone (ppb)
#> 193                                                       US percentile for Nitrogen Dioxide (NO2)
#> 194                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 195                                                        US percentile for Toxic Releases to Air
#> 196          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 197                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 198                                 US percentile for Superfund Proximity (site count/km distance)
#> 199                                   US percentile for RMP Proximity (facility count/km distance)
#> 200                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 201                                    US percentile for Underground Storage Tanks (UST) indicator
#> 202    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 203                                                US percentile for Drinking Water Non-Compliance
#> 228                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 229                                                               State percentile for Ozone (ppb)
#> 230                                                    State percentile for Nitrogen Dioxide (NO2)
#> 231                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 232                                                     State percentile for Toxic Releases to Air
#> 233       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 234                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 235                              State percentile for Superfund Proximity (site count/km distance)
#> 236                                State percentile for RMP Proximity (facility count/km distance)
#> 237                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 238                                 State percentile for Underground Storage Tanks (UST) indicator
#> 239 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 240                                             State percentile for Drinking Water Non-Compliance
#> 260                                            US Average for Particulate Matter (PM 2.5 in ug/m3)
#> 261                                                                     US Average for Ozone (ppb)
#> 262                                                          US Average for Nitrogen Dioxide (NO2)
#> 263                                               US Average for Diesel Particulate Matter (ug/m3)
#> 264                                                           US Average for Toxic Releases to Air
#> 265             US Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 266                                      US Average for Lead Paint Indicator (% pre-1960s housing)
#> 267                                    US Average for Superfund Proximity (site count/km distance)
#> 268                                      US Average for RMP Proximity (facility count/km distance)
#> 269                          US Average for Hazardous Waste Proximity (facility count/km distance)
#> 270                                       US Average for Underground Storage Tanks (UST) indicator
#> 271       US Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 272                                                   US Average for Drinking Water Non-Compliance
#> 292                                         State Average for Particulate Matter (PM 2.5 in ug/m3)
#> 293                                                                  State Average for Ozone (ppb)
#> 294                                                       State Average for Nitrogen Dioxide (NO2)
#> 295                                            State Average for Diesel Particulate Matter (ug/m3)
#> 296                                                        State Average for Toxic Releases to Air
#> 297          State Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 298                                   State Average for Lead Paint Indicator (% pre-1960s housing)
#> 299                                 State Average for Superfund Proximity (site count/km distance)
#> 300                                   State Average for RMP Proximity (facility count/km distance)
#> 301                       State Average for Hazardous Waste Proximity (facility count/km distance)
#> 302                                    State Average for Underground Storage Tanks (UST) indicator
#> 303    State Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 304                                                State Average for Drinking Water Non-Compliance
#> 133                                                                Estimated Fire Risk in 30 Years
#> 134                                                               Estimated Flood Risk in 30 Years
#> 189                                              US percentile for Estimated Fire Risk in 30 Years
#> 190                                             US percentile for Estimated Flood Risk in 30 Years
#> 226                                           State percentile for Estimated Fire Risk in 30 Years
#> 227                                          State percentile for Estimated Flood Risk in 30 Years
#> 163                                                          Number of Sites Nearby (total unique)
#> 166                                                        Distance to Closest Site for Avg Person
#> 162                                                                   Number of Sites Nearby (max)
#> 164                                                                   Number of Sites Nearby (avg)
#> 165                                                                       Distance to Closest Site
#> 417                                      US type of raw score for Particulate Matter Summary Index
#> 418                                                   US type of raw score for Ozone Summary Index
#> 419                                  US type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 420                               US type of raw score for Diesel Particulate Matter Summary Index
#> 421                                   US type of raw score for Toxic Releases to Air Summary Index
#> 422                            US type of raw score for Traffic Proximity and Volume Summary Index
#> 423                                              US type of raw score for Lead Paint Summary Index
#> 424                                     US type of raw score for Superfund Proximity Summary Index
#> 425                                           US type of raw score for RMP Proximity Summary Index
#> 426                               US type of raw score for Hazardous Waste Proximity Summary Index
#> 427                               US type of raw score for Underground Storage Tanks Summary Index
#> 428                                    US type of raw score for Wastewater Discharge Summary Index
#> 429                           US type of raw score for Drinking Water Non-Compliance Summary Index
#> 430                                   State type of raw score for Particulate Matter Summary Index
#> 431                                                State type of raw score for Ozone Summary Index
#> 432                               State type of raw score for Nitrogen Dioxide (NO2) Summary Index
#> 433                            State type of raw score for Diesel Particulate Matter Summary Index
#> 434                                State type of raw score for Toxic Releases to Air Summary Index
#> 435                         State type of raw score for Traffic Proximity and Volume Summary Index
#> 436                                           State type of raw score for Lead Paint Summary Index
#> 437                                  State type of raw score for Superfund Proximity Summary Index
#> 438                                        State type of raw score for RMP Proximity Summary Index
#> 439                            State type of raw score for Hazardous Waste Proximity Summary Index
#> 440                            State type of raw score for Underground Storage Tanks Summary Index
#> 441                                 State type of raw score for Wastewater Discharge Summary Index
#> 442                        State type of raw score for Drinking Water Non-Compliance Summary Index
#> 350                                         US percentile for Summary Index for Particulate Matter
#> 351                                                      US percentile for Summary Index for Ozone
#> 352                                     US percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 353                                  US percentile for Summary Index for Diesel Particulate Matter
#> 354                                      US percentile for Summary Index for Toxic Releases to Air
#> 355                               US percentile for Summary Index for Traffic Proximity and Volume
#> 356                                       US percentile for Summary Index for Lead Paint Indicator
#> 357                                        US percentile for Summary Index for Superfund Proximity
#> 358                                              US percentile for Summary Index for RMP Proximity
#> 359                                  US percentile for Summary Index for Hazardous Waste Proximity
#> 360                  US percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 361                             US percentile for Summary Index for Wastewater Discharge Indicator
#> 362                              US percentile for Summary Index for Drinking Water Non-Compliance
#> 376                                      State percentile for Summary Index for Particulate Matter
#> 377                                                   State percentile for Summary Index for Ozone
#> 378                                  State percentile for Summary Index for Nitrogen Dioxide (NO2)
#> 379                               State percentile for Summary Index for Diesel Particulate Matter
#> 380                                   State percentile for Summary Index for Toxic Releases to Air
#> 381                            State percentile for Summary Index for Traffic Proximity and Volume
#> 382                                    State percentile for Summary Index for Lead Paint Indicator
#> 383                                     State percentile for Summary Index for Superfund Proximity
#> 384                                           State percentile for Summary Index for RMP Proximity
#> 385                               State percentile for Summary Index for Hazardous Waste Proximity
#> 386               State percentile for Summary Index for Underground Storage Tanks (UST) indicator
#> 387                          State percentile for Summary Index for Wastewater Discharge Indicator
#> 388                           State percentile for Summary Index for Drinking Water Non-Compliance
#> 443                         US type of raw score for Particulate Matter Supplemental Summary Index
#> 444                                      US type of raw score for Ozone Supplemental Summary Index
#> 445                     US type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 446                  US type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 447                      US type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 448               US type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 449                                 US type of raw score for Lead Paint Supplemental Summary Index
#> 450                        US type of raw score for Superfund Proximity Supplemental Summary Index
#> 451                              US type of raw score for RMP Proximity Supplemental Summary Index
#> 452                  US type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 453                  US type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 454                       US type of raw score for Wastewater Discharge Supplemental Summary Index
#> 455              US type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 456                      State type of raw score for Particulate Matter Supplemental Summary Index
#> 457                                   State type of raw score for Ozone Supplemental Summary Index
#> 458                  State type of raw score for Nitrogen Dioxide (NO2) Supplemental Summary Index
#> 459               State type of raw score for Diesel Particulate Matter Supplemental Summary Index
#> 460                   State type of raw score for Toxic Releases to Air Supplemental Summary Index
#> 461            State type of raw score for Traffic Proximity and Volume Supplemental Summary Index
#> 462                              State type of raw score for Lead Paint Supplemental Summary Index
#> 463                     State type of raw score for Superfund Proximity Supplemental Summary Index
#> 464                           State type of raw score for RMP Proximity Supplemental Summary Index
#> 465               State type of raw score for Hazardous Waste Proximity Supplemental Summary Index
#> 466               State type of raw score for Underground Storage Tanks Supplemental Summary Index
#> 467                    State type of raw score for Wastewater Discharge Supplemental Summary Index
#> 468           State type of raw score for Drinking Water Non-Compliance Supplemental Summary Index
#> 363                            US percentile for Supplemental Summary Index for Particulate Matter
#> 364                                         US percentile for Supplemental Summary Index for Ozone
#> 365                        US percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 366                     US percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 367                         US percentile for Supplemental Summary Index for Toxic Releases to Air
#> 368                  US percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 369                          US percentile for Supplemental Summary Index for Lead Paint Indicator
#> 370                           US percentile for Supplemental Summary Index for Superfund Proximity
#> 371                                 US percentile for Supplemental Summary Index for RMP Proximity
#> 372                     US percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 373     US percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 374                US percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 375                 US percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 389                         State percentile for Supplemental Summary Index for Particulate Matter
#> 390                                      State percentile for Supplemental Summary Index for Ozone
#> 391                     State percentile for Supplemental Summary Index for Nitrogen Dioxide (NO2)
#> 392                  State percentile for Supplemental Summary Index for Diesel Particulate Matter
#> 393                      State percentile for Supplemental Summary Index for Toxic Releases to Air
#> 394               State percentile for Supplemental Summary Index for Traffic Proximity and Volume
#> 395                       State percentile for Supplemental Summary Index for Lead Paint Indicator
#> 396                        State percentile for Supplemental Summary Index for Superfund Proximity
#> 397                              State percentile for Supplemental Summary Index for RMP Proximity
#> 398                  State percentile for Supplemental Summary Index for Hazardous Waste Proximity
#> 399  State percentile for Supplemental Summary Index for Underground Storage Tanks (UST) indicator
#> 400             State percentile for Supplemental Summary Index for Wastewater Discharge Indicator
#> 401              State percentile for Supplemental Summary Index for Drinking Water Non-Compliance
#> 148                                                                              Number of Schools
#> 150                                                                       Number of Worship Places
#> 149                                                                            Number of Hospitals
#> 105                                                                           Percent under age 18
#> 106                                                                           Percent above age 17
#> 337                                                                         Population Over Age 17
#> 338                                                                        Population Under Age 18
#> 99                                                                          Occupied Housing Units
#> 100                                                                                  Percent Males
#> 101                                                                                Percent Females
#> 102                                                                       Life Expectancy in Years
#> 103                                                                              Per Capita Income
#> 104                                                              Percent Owner Occupied households
#> 335                                                                              Female Population
#> 336                                                                                Male Population
#> 151                                                               Flag for Overlapping with Tribes
#> 156                                                Count of National Priority List Superfund sites
#> 157                                          Count of Treatment Storage Disposal Facilities (TSDF)
#> 158                                                           Number of Water Discharge Facilities
#> 159                                                             Number of Air Pollution Facilities
#> 161                                                             Number of Toxic Release Facilities
#> 160                                                                          Number of Brownfields
#> 152                                                 Flag for Overlapping with Non-Attainment Areas
#> 153                                                      Flag for Overlapping with Impaired Waters
#> 155                                    Flag for Overlapping with EPA IRA Disadvantaged Communities
#> 154                                      Flag for Overlapping with CJEST Disadvantaged Communities
#> 9                                                                                        Longitude
#> 403                                                                          Census blocks (count)
#> 404                                                                        Buffer distance (miles)
#> 469                                                   Area of Analyzed Location(s) in Square Miles
#> 10                                                                                        Latitude
#> 11                                                                              State Abbreviation
#> 12                                                                                      State Name
#> 14                                                                                      EPA Region
#> 402                                                                     Census blockgroups (count)
#> 407                                                       Water area of block group in geodatabase
#> 408                                                        Land area of block group in geodatabase
#> 409                                                             Area of block group in geodatabase
#> 1                                                                                  EJScreen Report
#> 2                                                                                     EJScreen Map
#> 3                                                                                       ACS Report
#> 4                                                                                      ECHO Report
#> 5                                                                                     ejam_uniq_id
#> 6                                                                                            valid
#> 7                                                                                      invalid_msg
#> 13                                                                              in_how_many_states