Get order of variable names to sort by, as seen in EJScreen Community Report

Get order of variable names to sort by, as seen in EJScreen Community Report

Usage

table_order_variables(
  varnames,
  s1 = "ejscreensort",
  s2 = "sortvarlistEJSCREENREPORT",
  s3 = "sort_within_varlistEJSCREENREPORT"
)

Arguments

varnames

vector of indicator variables names from blockgroupstats, bgej, etc., such as "pm", "pctlowinc", "pctile.EJ.DISPARITY.traffic.score.eo" etc. and others as found in names_all_r, or specific subsets of those like in c(names_d, names_d_subgroups, names_e) and c(names_ej_pctile, names_ej_state_pctile, names_ej_supp_pctile, names_ej_supp_state_pctile)

s1

name of column in map_headernames to get sort info from

s2

optional like s1 but secondary to s1

s3

optional tertiary

Value

vector as from order(), to be used in sorting a data.frame for example

Examples

  cbind(table_order_variables(c(names_d, names_d_subgroups, names_e)))
#>       [,1]
#>  [1,]    1
#>  [2,]    2
#>  [3,]    9
#>  [4,]    3
#>  [5,]    5
#>  [6,]    4
#>  [7,]    6
#>  [8,]    7
#>  [9,]    8
#> [10,]   17
#> [11,]   11
#> [12,]   12
#> [13,]   10
#> [14,]   13
#> [15,]   14
#> [16,]   15
#> [17,]   16
#> [18,]   18
#> [19,]   19
#> [20,]   20
#> [21,]   21
#> [22,]   22
#> [23,]   23
#> [24,]   24
#> [25,]   25
#> [26,]   26
#> [27,]   27
#> [28,]   28
#> [29,]   29
#> [30,]   30
  
  out <- testoutput_ejamit_10pts_1miles
  vars <- out$formatted[ , 'indicator']
  vars <- fixcolnames(vars, 'long', 'r')
  out$formatted[table_order_variables(vars), ]
#>       value
#> 78     1.71
#> 79     1.78
#> 86       61
#> 80       32
#> 82        7
#> 81       13
#> 83       17
#> 84        6
#> 85       12
#> 95        2
#> 96        2
#> 14      1.3
#> 15      1.1
#> 22      1.6
#> 16        1
#> 18      1.2
#> 17      2.8
#> 19      1.5
#> 20      1.2
#> 21      0.7
#> 46      1.2
#> 47      1.2
#> 54      1.2
#> 48      1.3
#> 50      1.1
#> 49        2
#> 51      1.4
#> 52      1.1
#> 53      0.7
#> 128      70
#> 129      63
#> 136      73
#> 130      58
#> 132      70
#> 131      89
#> 133      77
#> 134      64
#> 135      34
#> 166      60
#> 167      58
#> 174      62
#> 168      59
#> 170      64
#> 169      71
#> 171      64
#> 172      62
#> 173      39
#> 147    1.34
#> 148    1.64
#> 155      40
#> 149      30
#> 151       6
#> 150       5
#> 152      11
#> 153       5
#> 154      18
#> 185    1.54
#> 186    1.42
#> 193      50
#> 187      25
#> 189       6
#> 188       7
#> 190      12
#> 191       5
#> 192      17
#> 327  103075
#> 321   52030
#> 323    5886
#> 322    8258
#> 324   19014
#> 325   10528
#> 326   20676
#> 364      48
#> 358      12
#> 359       9
#> 360       1
#> 361       0
#> 362      18
#> 363      12
#> 94       39
#> 88       12
#> 89        9
#> 87       34
#> 90        0
#> 91        0
#> 92        2
#> 93        5
#> 30      0.6
#> 24        1
#> 25      1.7
#> 23      1.9
#> 26      0.4
#> 27      0.4
#> 28      5.6
#> 29      1.3
#> 62      0.8
#> 56      1.4
#> 57      0.8
#> 55      1.3
#> 58      0.9
#> 59      0.3
#> 60      4.5
#> 61      1.4
#> 144      27
#> 138      73
#> 139      83
#> 137      81
#> 140      85
#> 141      95
#> 142      94
#> 143      74
#> 182      38
#> 176      67
#> 177      58
#> 175      65
#> 178      85
#> 179      38
#> 180      86
#> 181      69
#> 163      60
#> 157      12
#> 158       5
#> 156      18
#> 159       1
#> 160       0
#> 161       0
#> 162       3
#> 201      50
#> 195       9
#> 196      10
#> 194      26
#> 197       0
#> 198       0
#> 199       1
#> 200       3
#> 335   64711
#> 329   20259
#> 330   14374
#> 328   56369
#> 331     436
#> 332      97
#> 333    3833
#> 334    7708
#> 97       45
#> 98       12
#> 99        4
#> 100       2
#> 101      27
#> 102   70061
#> 103   19059
#> 104    5824
#> 105    2457
#> 106  157258
#> 107    6362
#> 108   42722
#> 109      58
#> 110      35
#> 111       7
#> 112       1
#> 113    4748
#> 114    2883
#> 115     560
#> 116      67
#> 7    167787
#> 337   61844
#> 336  111683
#> 338  132931
#> 339   17968
#> 340   65004
#> 341  164324
#> 117      15
#> 354   26340
#> 357    9076
#> 118      11
#> 119      18
#> 44      0.8
#> 45      0.9
#> 76        1
#> 77        1
#> 145      38
#> 146      39
#> 183      50
#> 164    13.6
#> 165      20
#> 202    11.2
#> 203      18
#> 355  169610
#> 356   18905
#> 184      53
#> 204   10.34
#> 205    67.2
#> 206    12.1
#> 207   0.415
#> 208    1813
#> 209 4328185
#> 210    0.28
#> 211    1.57
#> 212       2
#> 213    16.9
#> 214    23.6
#> 215    1797
#> 216    <NA>
#> 31      1.2
#> 32      1.1
#> 33      1.5
#> 34      2.2
#> 35      0.4
#> 36      2.6
#> 37      0.9
#> 38        4
#> 39      3.5
#> 40      4.8
#> 41      6.5
#> 42        0
#> 43     <NA>
#> 63      1.1
#> 64        1
#> 65      1.4
#> 66      1.6
#> 67        1
#> 68      1.7
#> 69      0.8
#> 70      1.6
#> 71      3.2
#> 72      2.3
#> 73      3.9
#> 74      0.1
#> 75     <NA>
#> 217      89
#> 218      79
#> 219      88
#> 220      92
#> 221      70
#> 222      89
#> 223      56
#> 224      94
#> 225      94
#> 226      95
#> 227      97
#> 228      78
#> 229    <NA>
#> 243      70
#> 244      52
#> 245      77
#> 246      79
#> 247      77
#> 248      70
#> 249      46
#> 250      70
#> 251      82
#> 252      79
#> 253      47
#> 254      60
#> 255    <NA>
#> 230    8.45
#> 231    61.8
#> 232     7.8
#> 233   0.191
#> 234    4612
#> 235 1687230
#> 236     0.3
#> 237    0.39
#> 238    0.57
#> 239     3.5
#> 240     3.6
#> 241  701805
#> 242     2.2
#> 256    9.74
#> 257    67.3
#> 258     8.9
#> 259   0.264
#> 260    1885
#> 261 2603011
#> 262    0.34
#> 263    1.01
#> 264    0.62
#> 265     7.5
#> 266     6.1
#> 267   13392
#> 268    1.52
#> 344       0
#> 345     0.7
#> 346       1
#> 347      10
#> 348       1
#> 365 117.312
#> 366 121.441
#> 367  147.75
#> 368 153.513
#> 369 114.724
#> 370 141.447
#> 371   91.88
#> 372 145.606
#> 373 139.698
#> 374 152.021
#> 375  92.452
#> 376 109.501
#> 377  10.825
#> 378 142.892
#> 379 110.472
#> 380 154.348
#> 381 158.451
#> 382 149.919
#> 383 146.261
#> 384  85.508
#> 385 143.983
#> 386 162.213
#> 387 156.911
#> 388  94.797
#> 389  108.36
#> 390  12.691
#> 269      78
#> 270      81
#> 271      83
#> 272      85
#> 273      80
#> 274      82
#> 275      72
#> 276      86
#> 277      83
#> 278      85
#> 279      72
#> 280      77
#> 281      73
#> 295       1
#> 296       1
#> 297       1
#> 298       1
#> 299       1
#> 300       1
#> 301      12
#> 302      33
#> 303      24
#> 304       6
#> 305      54
#> 306      11
#> 307      77
#> 391 117.153
#> 392 122.523
#> 393   152.1
#> 394 157.202
#> 395 119.338
#> 396 144.478
#> 397  95.278
#> 398 147.302
#> 399 142.706
#> 400  155.08
#> 401 103.231
#> 402  114.04
#> 403  14.414
#> 404  131.09
#> 405 100.312
#> 406 139.762
#> 407 143.106
#> 408 135.386
#> 409 133.632
#> 410  74.151
#> 411 128.183
#> 412 144.944
#> 413 141.488
#> 414  98.275
#> 415 100.185
#> 416  14.286
#> 282      75
#> 283      79
#> 284      85
#> 285      87
#> 286      76
#> 287      84
#> 288      62
#> 289      86
#> 290      82
#> 291      86
#> 292      68
#> 293      74
#> 294      73
#> 308       1
#> 309       0
#> 310       1
#> 311       1
#> 312       1
#> 313       1
#> 314      12
#> 315      32
#> 316      23
#> 317       5
#> 318      53
#> 319      10
#> 320      77
#> 120      23
#> 121      77
#> 350  128643
#> 351   39143
#> 122   61844
#> 123      50
#> 124      50
#> 125      78
#> 126   43668
#> 127    <NA>
#> 352   84334
#> 353   83452
#> 8      <NA>
#> 9      <NA>
#> 10     <NA>
#> 11     <NA>
#> 13     <NA>
#> 342     170
#> 343    1914
#> 349       1
#> 1      <NA>
#> 2      <NA>
#> 3      <NA>
#> 4      <NA>
#> 5      <NA>
#> 6      <NA>
#> 12        7
#>                                                                                          indicator
#> 78                                                                           Demographic Index USA
#> 79                                                              Supplemental Demographic Index USA
#> 86                                                                               % People of Color
#> 80                                                                                    % Low Income
#> 82                                                                                    % Unemployed
#> 81                                                        % in limited English-speaking Households
#> 83                                                          % with Less Than High School Education
#> 84                                                                                   % under Age 5
#> 85                                                                                   % over Age 64
#> 95                                                                         Demographic Index State
#> 96                                                            Supplemental Demographic Index State
#> 14                                                               Ratio to US avg Demographic Index
#> 15                        Ratio to US avg Supplemental EJ Index for Supplemental Demographic Index
#> 22                                                               Ratio to US avg % People of Color
#> 16                                                                    Ratio to US avg % Low Income
#> 18                                                                    Ratio to US avg % Unemployed
#> 17                                        Ratio to US avg % in limited English-speaking Households
#> 19                                          Ratio to US avg % with Less Than High School Education
#> 20                                                                   Ratio to US avg % under Age 5
#> 21                                                                   Ratio to US avg % over Age 64
#> 46                                                            Ratio to State avg Demographic Index
#> 47                     Ratio to State avg Supplemental EJ Index for Supplemental Demographic Index
#> 54                                                            Ratio to State avg % People of Color
#> 48                                                                 Ratio to State avg % Low Income
#> 50                                                                 Ratio to State avg % Unemployed
#> 49                                     Ratio to State avg % in limited English-speaking Households
#> 51                                       Ratio to State avg % with Less Than High School Education
#> 52                                                                Ratio to State avg % under Age 5
#> 53                                                                Ratio to State avg % over Age 64
#> 128                                                            US percentile for Demographic Index
#> 129                     US percentile for Supplemental EJ Index for Supplemental Demographic Index
#> 136                                                            US percentile for % People of Color
#> 130                                                                 US percentile for % Low Income
#> 132                                                                 US percentile for % Unemployed
#> 131                                     US percentile for % in limited English-speaking Households
#> 133                                       US percentile for % with Less Than High School Education
#> 134                                                                US percentile for % under Age 5
#> 135                                                                US percentile for % over Age 64
#> 166                                                         State percentile for Demographic Index
#> 167                  State percentile for Supplemental EJ Index for Supplemental Demographic Index
#> 174                                                         State percentile for % People of Color
#> 168                                                              State percentile for % Low Income
#> 170                                                              State percentile for % Unemployed
#> 169                                  State percentile for % in limited English-speaking Households
#> 171                                    State percentile for % with Less Than High School Education
#> 172                                                             State percentile for % under Age 5
#> 173                                                             State percentile for % over Age 64
#> 147                                                               US Average for Demographic Index
#> 148                        US Average for Supplemental EJ Index for Supplemental Demographic Index
#> 155                                                               US Average for % People of Color
#> 149                                                                    US Average for % Low Income
#> 151                                                                    US Average for % Unemployed
#> 150                                        US Average for % in limited English-speaking Households
#> 152                                          US Average for % with Less Than High School Education
#> 153                                                                   US Average for % under Age 5
#> 154                                                                   US Average for % over Age 64
#> 185                                                            State Average for Demographic Index
#> 186                     State Average for Supplemental EJ Index for Supplemental Demographic Index
#> 193                                                            State Average for % People of Color
#> 187                                                                 State Average for % Low Income
#> 189                                                                 State Average for % Unemployed
#> 188                                     State Average for % in limited English-speaking Households
#> 190                                       State Average for % with Less Than High School Education
#> 191                                                                State Average for % under Age 5
#> 192                                                                State Average for % over Age 64
#> 327                                                                 People of Color resident count
#> 321                                                                      Low income resident count
#> 323                                                                      Unemployed resident count
#> 322                                                            Limited English-speaking Households
#> 324                                                 Less Than High School Education resident count
#> 325                                                                     Under Age 5 resident count
#> 326                                                                     Over Age 64 resident count
#> 364                                                       % White (single race, includes Hispanic)
#> 358                                   % Black or African American (single race, includes Hispanic)
#> 359                                                       % Asian (single race, includes Hispanic)
#> 360                           % American Indian and Alaska Native (single race, includes Hispanic)
#> 361                  % Native Hawaiian and Other Pacific Islander (single race, includes Hispanic)
#> 362                                                  % Other race (single race, includes Hispanic)
#> 363                                                        % Two or more races (includes Hispanic)
#> 94                                                             % White (non-Hispanic, single race)
#> 88                                         % Black or African American (non-Hispanic, single race)
#> 89                                                             % Asian (non-Hispanic, single race)
#> 87                                                                            % Hispanic or Latino
#> 90                                 % American Indian and Alaska Native (non-Hispanic, single race)
#> 91                        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 92                                                        % Other race (non-Hispanic, single race)
#> 93                                                              % Two or more races (non-Hispanic)
#> 30                                             Ratio to US avg % White (non-Hispanic, single race)
#> 24                         Ratio to US avg % Black or African American (non-Hispanic, single race)
#> 25                                             Ratio to US avg % Asian (non-Hispanic, single race)
#> 23                                                            Ratio to US avg % Hispanic or Latino
#> 26                 Ratio to US avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 27        Ratio to US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 28                                        Ratio to US avg % Other race (non-Hispanic, single race)
#> 29                                              Ratio to US avg % Two or more races (non-Hispanic)
#> 62                                          Ratio to State avg % White (non-Hispanic, single race)
#> 56                      Ratio to State avg % Black or African American (non-Hispanic, single race)
#> 57                                          Ratio to State avg % Asian (non-Hispanic, single race)
#> 55                                                         Ratio to State avg % Hispanic or Latino
#> 58              Ratio to State avg % American Indian and Alaska Native (non-Hispanic, single race)
#> 59     Ratio to State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 60                                     Ratio to State avg % Other race (non-Hispanic, single race)
#> 61                                           Ratio to State avg % Two or more races (non-Hispanic)
#> 144                                          US percentile for % White (non-Hispanic, single race)
#> 138                      US percentile for % Black or African American (non-Hispanic, single race)
#> 139                                          US percentile for % Asian (non-Hispanic, single race)
#> 137                                                         US percentile for % Hispanic or Latino
#> 140              US percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 141     US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 142                                     US percentile for % Other race (non-Hispanic, single race)
#> 143                                           US percentile for % Two or more races (non-Hispanic)
#> 182                                       State percentile for % White (non-Hispanic, single race)
#> 176                   State percentile for % Black or African American (non-Hispanic, single race)
#> 177                                       State percentile for % Asian (non-Hispanic, single race)
#> 175                                                      State percentile for % Hispanic or Latino
#> 178           State percentile for % American Indian and Alaska Native (non-Hispanic, single race)
#> 179  State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 180                                  State percentile for % Other race (non-Hispanic, single race)
#> 181                                        State percentile for % Two or more races (non-Hispanic)
#> 163                                             US Average for % White (non-Hispanic, single race)
#> 157                         US Average for % Black or African American (non-Hispanic, single race)
#> 158                                             US Average for % Asian (non-Hispanic, single race)
#> 156                                                            US Average for % Hispanic or Latino
#> 159                 US Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 160        US Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 161                                        US Average for % Other race (non-Hispanic, single race)
#> 162                                              US Average for % Two or more races (non-Hispanic)
#> 201                                          State Average for % White (non-Hispanic, single race)
#> 195                      State Average for % Black or African American (non-Hispanic, single race)
#> 196                                          State Average for % Asian (non-Hispanic, single race)
#> 194                                                         State Average for % Hispanic or Latino
#> 197              State Average for % American Indian and Alaska Native (non-Hispanic, single race)
#> 198     State Average for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 199                                     State Average for % Other race (non-Hispanic, single race)
#> 200                                           State Average for % Two or more races (non-Hispanic)
#> 335                                                     Count of White (non-Hispanic, single race)
#> 329                                 Count of Black or African American (non-Hispanic, single race)
#> 330                                                     Count of Asian (non-Hispanic, single race)
#> 328                                                                    Count of Hispanic or Latino
#> 331                         Count of American Indian and Alaska Native (non-Hispanic, single race)
#> 332                Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> 333                                                Count of Other race (non-Hispanic, single race)
#> 334                                                      Count of Two or more races (non-Hispanic)
#> 97                                              Percent of population speaking non-English at home
#> 98                                      Percent of population speaking Other Indo-European at home
#> 99                       Percent of population speaking Asian and Pacific Island languages at home
#> 100                         Percent of population speaking Other and Unspecified languages at home
#> 101                                                 Percent of population speaking Spanish at home
#> 102                                                            Number speaking non-English at home
#> 103                                                              Speak Other Indo-European at Home
#> 104                                                    Speak Asian-Pacific Island language at Home
#> 105                                        Number speaking Other and Unspecified languages at home
#> 106                    Number of Persons for whom Language Ability is Determined--age 5 and above.
#> 107                                                           Persons who Speak English Not at All
#> 108                                                                          Speak Spanish at Home
#> 109                                  Percent Speaking Spanish (as % of limited English households)
#> 110            Percent Speaking Other Indo-European languages (as % of limited English households)
#> 111           Percent Speaking Asian-Pacific Island languages (as % of limited English households)
#> 112                          Percent Speaking Other languages (as % of limited English households)
#> 113                                                   Speak Spanish (in limited English household)
#> 114                                       Speak Other Indo-European (in limited English household)
#> 115                                      Speak Asian-Pacific Island (in limited English household)
#> 116                                                     Speak Other (in limited English household)
#> 7                                                                                 Total Population
#> 337                                                                            Count of Households
#> 336                                                                  Count of Population Age 25 up
#> 338                                           Universe for percent unemployed (denominator, count)
#> 339                                                          Count of Housing Units Built Pre 1960
#> 340                                   Built housing units count (denominator for percent pre 1960)
#> 341                                      Count of Population for whom Poverty Status is Determined
#> 117                                                      Percent of Households below Poverty Level
#> 354                                                          Count of Owner Occupied Housing Units
#> 357                                                                 Households below Poverty Level
#> 118                                                                            % with Disabilities
#> 119                                                                           %Low life expectancy
#> 44                                                             Ratio to US avg % with Disabilities
#> 45                                                            Ratio to US avg %Low life expectancy
#> 76                                                          Ratio to State avg % with Disabilities
#> 77                                                         Ratio to State avg %Low life expectancy
#> 145                                                          US percentile for % with Disabilities
#> 146                                                         US percentile for %Low life expectancy
#> 183                                                       State percentile for % with Disabilities
#> 164                                                             US Average for % with Disabilities
#> 165                                                            US Average for %Low life expectancy
#> 202                                                          State Average for % with Disabilities
#> 203                                                         State Average for %Low life expectancy
#> 355                        Civilian noninstitutionalized population (denominator for % disability)
#> 356                                                               Count of Persons with Disability
#> 184                                                      State percentile for %Low life expectancy
#> 204                                                           Particulate Matter (PM 2.5 in ug/m3)
#> 205                                                                                    Ozone (ppb)
#> 206                                                                         Nitrogen Dioxide (NO2)
#> 207                                                              Diesel Particulate Matter (ug/m3)
#> 208                                                                          Toxic Releases to Air
#> 209                            Traffic Proximity and Volume (daily traffic count/distance to road)
#> 210                                                     Lead Paint Indicator (% pre-1960s housing)
#> 211                                                   Superfund Proximity (site count/km distance)
#> 212                                                     RMP Proximity (facility count/km distance)
#> 213                                         Hazardous Waste Proximity (facility count/km distance)
#> 214                                                      Underground Storage Tanks (UST) indicator
#> 215                      Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 216                                                                  Drinking Water Non-Compliance
#> 31                                                              Ratio to US avg Particulate Matter
#> 32                                                                           Ratio to US avg Ozone
#> 33                                                          Ratio to US avg Nitrogen Dioxide (NO2)
#> 34                                                       Ratio to US avg Diesel Particulate Matter
#> 35                                                           Ratio to US avg Toxic Releases to Air
#> 36                                                    Ratio to US avg Traffic Proximity and Volume
#> 37                                                            Ratio to US avg Lead Paint Indicator
#> 38                                                             Ratio to US avg Superfund Proximity
#> 39                                                                   Ratio to US avg RMP Proximity
#> 40                                                       Ratio to US avg Hazardous Waste Proximity
#> 41                                             Ratio to US avg Underground Storage Tanks indicator
#> 42                                                  Ratio to US avg Wastewater Discharge Indicator
#> 43                                                   Ratio to US avg Drinking Water Non-Compliance
#> 63                                                           Ratio to State avg Particulate Matter
#> 64                                                                        Ratio to State avg Ozone
#> 65                                                       Ratio to State avg Nitrogen Dioxide (NO2)
#> 66                                                    Ratio to State avg Diesel Particulate Matter
#> 67                                                        Ratio to State avg Toxic Releases to Air
#> 68                                                 Ratio to State avg Traffic Proximity and Volume
#> 69                                                         Ratio to State avg Lead Paint Indicator
#> 70                                                          Ratio to State avg Superfund Proximity
#> 71                                                                Ratio to State avg RMP Proximity
#> 72                                                    Ratio to State avg Hazardous Waste Proximity
#> 73                                          Ratio to State avg Underground Storage Tanks indicator
#> 74                                               Ratio to State avg Wastewater Discharge Indicator
#> 75                                                Ratio to State avg Drinking Water Non-Compliance
#> 217                                         US percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 218                                                                  US percentile for Ozone (ppb)
#> 219                                                       US percentile for Nitrogen Dioxide (NO2)
#> 220                                            US percentile for Diesel Particulate Matter (ug/m3)
#> 221                                                        US percentile for Toxic Releases to Air
#> 222          US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 223                                   US percentile for Lead Paint Indicator (% pre-1960s housing)
#> 224                                 US percentile for Superfund Proximity (site count/km distance)
#> 225                                   US percentile for RMP Proximity (facility count/km distance)
#> 226                       US percentile for Hazardous Waste Proximity (facility count/km distance)
#> 227                                    US percentile for Underground Storage Tanks (UST) indicator
#> 228    US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 229                                                US percentile for Drinking Water Non-Compliance
#> 243                                      State percentile for Particulate Matter (PM 2.5 in ug/m3)
#> 244                                                               State percentile for Ozone (ppb)
#> 245                                                    State percentile for Nitrogen Dioxide (NO2)
#> 246                                         State percentile for Diesel Particulate Matter (ug/m3)
#> 247                                                     State percentile for Toxic Releases to Air
#> 248       State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 249                                State percentile for Lead Paint Indicator (% pre-1960s housing)
#> 250                              State percentile for Superfund Proximity (site count/km distance)
#> 251                                State percentile for RMP Proximity (facility count/km distance)
#> 252                    State percentile for Hazardous Waste Proximity (facility count/km distance)
#> 253                                 State percentile for Underground Storage Tanks (UST) indicator
#> 254 State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 255                                             State percentile for Drinking Water Non-Compliance
#> 230                                            US Average for Particulate Matter (PM 2.5 in ug/m3)
#> 231                                                                     US Average for Ozone (ppb)
#> 232                                                          US Average for Nitrogen Dioxide (NO2)
#> 233                                               US Average for Diesel Particulate Matter (ug/m3)
#> 234                                                           US Average for Toxic Releases to Air
#> 235             US Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 236                                      US Average for Lead Paint Indicator (% pre-1960s housing)
#> 237                                    US Average for Superfund Proximity (site count/km distance)
#> 238                                      US Average for RMP Proximity (facility count/km distance)
#> 239                          US Average for Hazardous Waste Proximity (facility count/km distance)
#> 240                                       US Average for Underground Storage Tanks (UST) indicator
#> 241       US Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 242                                                   US Average for Drinking Water Non-Compliance
#> 256                                         State Average for Particulate Matter (PM 2.5 in ug/m3)
#> 257                                                                  State Average for Ozone (ppb)
#> 258                                                       State Average for Nitrogen Dioxide (NO2)
#> 259                                            State Average for Diesel Particulate Matter (ug/m3)
#> 260                                                        State Average for Toxic Releases to Air
#> 261          State Average for Traffic Proximity and Volume (daily traffic count/distance to road)
#> 262                                   State Average for Lead Paint Indicator (% pre-1960s housing)
#> 263                                 State Average for Superfund Proximity (site count/km distance)
#> 264                                   State Average for RMP Proximity (facility count/km distance)
#> 265                       State Average for Hazardous Waste Proximity (facility count/km distance)
#> 266                                    State Average for Underground Storage Tanks (UST) indicator
#> 267    State Average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> 268                                                State Average for Drinking Water Non-Compliance
#> 344                                                                       Distance to Closest Site
#> 345                                                        Distance to Closest Site for Avg Person
#> 346                                                                   Number of Sites Nearby (max)
#> 347                                                          Number of Sites Nearby (total unique)
#> 348                                                                   Number of Sites Nearby (avg)
#> 365                                           US type of raw score for Particulate Matter EJ Index
#> 366                                                        US type of raw score for Ozone EJ Index
#> 367                                       US type of raw score for Nitrogen Dioxide (NO2) EJ Index
#> 368                                    US type of raw score for Diesel Particulate Matter EJ Index
#> 369                                        US type of raw score for Toxic Releases to Air EJ Index
#> 370                                 US type of raw score for Traffic Proximity and Volume EJ Index
#> 371                                                   US type of raw score for Lead Paint EJ Index
#> 372                                          US type of raw score for Superfund Proximity EJ Index
#> 373                                                US type of raw score for RMP Proximity EJ Index
#> 374                                    US type of raw score for Hazardous Waste Proximity EJ Index
#> 375                                    US type of raw score for Underground Storage Tanks EJ Index
#> 376                                         US type of raw score for Wastewater Discharge EJ Index
#> 377                                US type of raw score for Drinking Water Non-Compliance EJ Index
#> 378                                        State type of raw score for Particulate Matter EJ Index
#> 379                                                     State type of raw score for Ozone EJ Index
#> 380                                    State type of raw score for Nitrogen Dioxide (NO2) EJ Index
#> 381                                 State type of raw score for Diesel Particulate Matter EJ Index
#> 382                                     State type of raw score for Toxic Releases to Air EJ Index
#> 383                              State type of raw score for Traffic Proximity and Volume EJ Index
#> 384                                                State type of raw score for Lead Paint EJ Index
#> 385                                       State type of raw score for Superfund Proximity EJ Index
#> 386                                             State type of raw score for RMP Proximity EJ Index
#> 387                                 State type of raw score for Hazardous Waste Proximity EJ Index
#> 388                                 State type of raw score for Underground Storage Tanks EJ Index
#> 389                                      State type of raw score for Wastewater Discharge EJ Index
#> 390                             State type of raw score for Drinking Water Non-Compliance EJ Index
#> 269                                              US percentile for EJ Index for Particulate Matter
#> 270                                                           US percentile for EJ Index for Ozone
#> 271                                          US percentile for EJ Index for Nitrogen Dioxide (NO2)
#> 272                                       US percentile for EJ Index for Diesel Particulate Matter
#> 273                                           US percentile for EJ Index for Toxic Releases to Air
#> 274                                    US percentile for EJ Index for Traffic Proximity and Volume
#> 275                                            US percentile for EJ Index for Lead Paint Indicator
#> 276                                             US percentile for EJ Index for Superfund Proximity
#> 277                                                   US percentile for EJ Index for RMP Proximity
#> 278                                       US percentile for EJ Index for Hazardous Waste Proximity
#> 279                       US percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 280                                  US percentile for EJ Index for Wastewater Discharge Indicator
#> 281                                   US percentile for EJ Index for Drinking Water Non-Compliance
#> 295                                           State percentile for EJ Index for Particulate Matter
#> 296                                                        State percentile for EJ Index for Ozone
#> 297                                       State percentile for EJ Index for Nitrogen Dioxide (NO2)
#> 298                                    State percentile for EJ Index for Diesel Particulate Matter
#> 299                                        State percentile for EJ Index for Toxic Releases to Air
#> 300                                 State percentile for EJ Index for Traffic Proximity and Volume
#> 301                                         State percentile for EJ Index for Lead Paint Indicator
#> 302                                          State percentile for EJ Index for Superfund Proximity
#> 303                                                State percentile for EJ Index for RMP Proximity
#> 304                                    State percentile for EJ Index for Hazardous Waste Proximity
#> 305                    State percentile for EJ Index for Underground Storage Tanks (UST) indicator
#> 306                               State percentile for EJ Index for Wastewater Discharge Indicator
#> 307                                State percentile for EJ Index for Drinking Water Non-Compliance
#> 391                              US type of raw score for Particulate Matter Supplemental EJ Index
#> 392                                           US type of raw score for Ozone Supplemental EJ Index
#> 393                          US type of raw score for Nitrogen Dioxide (NO2) Supplemental EJ Index
#> 394                       US type of raw score for Diesel Particulate Matter Supplemental EJ Index
#> 395                           US type of raw score for Toxic Releases to Air Supplemental EJ Index
#> 396                    US type of raw score for Traffic Proximity and Volume Supplemental EJ Index
#> 397                                      US type of raw score for Lead Paint Supplemental EJ Index
#> 398                             US type of raw score for Superfund Proximity Supplemental EJ Index
#> 399                                   US type of raw score for RMP Proximity Supplemental EJ Index
#> 400                       US type of raw score for Hazardous Waste Proximity Supplemental EJ Index
#> 401                       US type of raw score for Underground Storage Tanks Supplemental EJ Index
#> 402                            US type of raw score for Wastewater Discharge Supplemental EJ Index
#> 403                   US type of raw score for Drinking Water Non-Compliance Supplemental EJ Index
#> 404                           State type of raw score for Particulate Matter Supplemental EJ Index
#> 405                                        State type of raw score for Ozone Supplemental EJ Index
#> 406                       State type of raw score for Nitrogen Dioxide (NO2) Supplemental EJ Index
#> 407                    State type of raw score for Diesel Particulate Matter Supplemental EJ Index
#> 408                        State type of raw score for Toxic Releases to Air Supplemental EJ Index
#> 409                 State type of raw score for Traffic Proximity and Volume Supplemental EJ Index
#> 410                                   State type of raw score for Lead Paint Supplemental EJ Index
#> 411                          State type of raw score for Superfund Proximity Supplemental EJ Index
#> 412                                State type of raw score for RMP Proximity Supplemental EJ Index
#> 413                    State type of raw score for Hazardous Waste Proximity Supplemental EJ Index
#> 414                    State type of raw score for Underground Storage Tanks Supplemental EJ Index
#> 415                         State type of raw score for Wastewater Discharge Supplemental EJ Index
#> 416                State type of raw score for Drinking Water Non-Compliance Supplemental EJ Index
#> 282                                 US percentile for Supplemental EJ Index for Particulate Matter
#> 283                                              US percentile for Supplemental EJ Index for Ozone
#> 284                             US percentile for Supplemental EJ Index for Nitrogen Dioxide (NO2)
#> 285                          US percentile for Supplemental EJ Index for Diesel Particulate Matter
#> 286                              US percentile for Supplemental EJ Index for Toxic Releases to Air
#> 287                       US percentile for Supplemental EJ Index for Traffic Proximity and Volume
#> 288                               US percentile for Supplemental EJ Index for Lead Paint Indicator
#> 289                                US percentile for Supplemental EJ Index for Superfund Proximity
#> 290                                      US percentile for Supplemental EJ Index for RMP Proximity
#> 291                          US percentile for Supplemental EJ Index for Hazardous Waste Proximity
#> 292          US percentile for Supplemental EJ Index for Underground Storage Tanks (UST) indicator
#> 293                     US percentile for Supplemental EJ Index for Wastewater Discharge Indicator
#> 294                      US percentile for Supplemental EJ Index for Drinking Water Non-Compliance
#> 308                              State percentile for Supplemental EJ Index for Particulate Matter
#> 309                                           State percentile for Supplemental EJ Index for Ozone
#> 310                          State percentile for Supplemental EJ Index for Nitrogen Dioxide (NO2)
#> 311                       State percentile for Supplemental EJ Index for Diesel Particulate Matter
#> 312                           State percentile for Supplemental EJ Index for Toxic Releases to Air
#> 313                    State percentile for Supplemental EJ Index for Traffic Proximity and Volume
#> 314                            State percentile for Supplemental EJ Index for Lead Paint Indicator
#> 315                             State percentile for Supplemental EJ Index for Superfund Proximity
#> 316                                   State percentile for Supplemental EJ Index for RMP Proximity
#> 317                       State percentile for Supplemental EJ Index for Hazardous Waste Proximity
#> 318       State percentile for Supplemental EJ Index for Underground Storage Tanks (UST) indicator
#> 319                  State percentile for Supplemental EJ Index for Wastewater Discharge Indicator
#> 320                   State percentile for Supplemental EJ Index for Drinking Water Non-Compliance
#> 120                                                                           Percent under age 18
#> 121                                                                           Percent above age 17
#> 350                                                                         Population Over Age 17
#> 351                                                                        Population Under Age 18
#> 122                                                                         Occupied Housing Units
#> 123                                                                                  Percent Males
#> 124                                                                                Percent Females
#> 125                                                                       Life Expectancy in Years
#> 126                                                                              Per Capita Income
#> 127                                                              Percent Owner Occupied households
#> 352                                                                                Male Population
#> 353                                                                              Female Population
#> 8                                                                                        Longitude
#> 9                                                                                         Latitude
#> 10                                                                              State Abbreviation
#> 11                                                                                      State Name
#> 13                                                                                      EPA Region
#> 342                                                                     Census blockgroups (count)
#> 343                                                                          Census blocks (count)
#> 349                                                                        Buffer distance (miles)
#> 1                                                                                  EJScreen Report
#> 2                                                                                     EJScreen Map
#> 3                                                                                      ECHO report
#> 4                                                                                     ejam_uniq_id
#> 5                                                                                            valid
#> 6                                                                                      invalid_msg
#> 12                                                                              in_how_many_states