Round numbers in a table, each column to appropriate number of decimal places

Round numbers in a table, each column to appropriate number of decimal places

Usage

table_round(x, var = names(x), varnametype = "rname", ...)

Arguments

x

data.frame, data.table, or vector with at least some numerical columns, like the results of ejamit()$results_bysite

var

optional, but assumed to be names(x) by default, specifies colnames of table or names of vector elements, within x

varnametype

optional, name of column in map_headernames that is looked in for var

...

passed to var_is_numeric_ish()

Value

Returns the original x but with appropriate cells rounded off.

See also

Examples

 
  table_round(c(12.123456, 9, NA ), 'pm')
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> [1] 12.12  9.00    NA

 x <- testoutput_ejamit_10pts_1miles$results_bysite[
   1:2, c('lat','lon', 'pop', names_these, names_ratio_to_avg_these, names_e_pctile), 
   with = FALSE
 ]

 table_rounding_info(names(x))
#>  [1] NA NA  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  0  2  1  0  2
#> [26]  2  1  0  2  2  1  0  1  0  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1
#> [51]  1  1  1  1  1  1  1  1  1  1  1  1  1  1  1  0  0  0  0  0  0  0  0  0  0
#> [76]  0  0  0

 table_round(x)
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#>        lat       lon  pop Demog.Index Demog.Index.Supp pctlowinc pctlingiso
#> 1 36.77649 -114.0379  747           0                0         0          0
#> 2 39.78107 -104.9129 1941           1                0         0          0
#>   pctunemployed pctlths lowlifex pctunder5 pctover64 pctmin pcthisp pctnhba
#> 1             0       0        0         0         0      0       0       0
#> 2             0       0        0         0         0      1       0       0
#>   pctnhaa pctnhaiana pctnhnhpia pctnhotheralone pctnhmulti pctnhwa   pm   o3
#> 1       0          0          0               0          0       1 6.65 61.6
#> 2       0          0          0               0          0       0 8.61 65.9
#>   cancer resp  dpm pctpre1960 traffic.score proximity.npl proximity.rmp
#> 1     20 0.20 0.15        0.1            38          0.05          0.03
#> 2     36 0.39 0.53        0.6           157          0.30          4.32
#>   proximity.tsdf proximity.npdes ust rsei ratio.to.avg.Demog.Index
#> 1            0.3               0 2.0    0                      0.7
#> 2            2.9               0 3.6 2503                      1.5
#>   ratio.to.avg.Demog.Index.Supp ratio.to.avg.pctlowinc ratio.to.avg.pctlingiso
#> 1                           1.1                    1.4                     0.0
#> 2                           1.3                    1.3                     1.3
#>   ratio.to.avg.pctunemployed ratio.to.avg.pctlths ratio.to.avg.lowlifex
#> 1                        1.9                  0.3                     1
#> 2                        1.8                  1.4                     1
#>   ratio.to.avg.pctunder5 ratio.to.avg.pctover64 ratio.to.avg.pctmin
#> 1                    0.0                    1.0                 0.1
#> 2                    1.2                    0.7                 1.6
#>   ratio.to.avg.pcthisp ratio.to.avg.pctnhba ratio.to.avg.pctnhaa
#> 1                  0.1                  0.1                  0.0
#> 2                  1.7                  2.1                  0.4
#>   ratio.to.avg.pctnhaiana ratio.to.avg.pctnhnhpia ratio.to.avg.pctnhotheralone
#> 1                       0                     0.7                          0.0
#> 2                       0                     0.0                          1.6
#>   ratio.to.avg.pctnhmulti ratio.to.avg.pctnhwa ratio.to.avg.pm ratio.to.avg.o3
#> 1                     0.0                  1.6             0.8             1.0
#> 2                     1.4                  0.6             1.1             1.1
#>   ratio.to.avg.cancer ratio.to.avg.resp ratio.to.avg.dpm
#> 1                 0.8               0.7              0.6
#> 2                 1.4               1.3              2.0
#>   ratio.to.avg.pctpre1960 ratio.to.avg.traffic.score ratio.to.avg.proximity.npl
#> 1                     0.4                        0.2                        0.3
#> 2                     1.9                        0.7                        2.2
#>   ratio.to.avg.proximity.rmp ratio.to.avg.proximity.tsdf
#> 1                        0.1                         0.1
#> 2                       10.1                         1.5
#>   ratio.to.avg.proximity.npdes ratio.to.avg.ust ratio.to.avg.rsei pctile.pm
#> 1                            0              0.5               0.0        14
#> 2                            0              0.9               0.5        62
#>   pctile.o3 pctile.cancer pctile.resp pctile.dpm pctile.pctpre1960
#> 1        54             5           4         31                38
#> 2        80            52          31         92                78
#>   pctile.traffic.score pctile.proximity.npl pctile.proximity.rmp
#> 1                   34                   41                    4
#> 2                   68                   91                   99
#>   pctile.proximity.tsdf pctile.proximity.npdes pctile.ust pctile.rsei
#> 1                    40                     57         59           3
#> 2                    80                     52         70          76