How many decimal places to round to for given variable(s)

How many decimal places to round to for given variable(s)

table_rounding_info(var, varnametype = "rname")

Arguments

var

vector of variable names such as c("pctlowinc", "cancer") or c(names_d, names_d_subgroups)

varnametype

which column of map_headernames to use when looking for var, like "rname" or "api" or "long"

Value

named vector same size as var, with var as names.

See also

Examples

  table_rounding_info("pm")
#> [1] 2
  table_round(8.252345, "pm")
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> [1] 8.25
  table_round(8, "pm")
#> Warning: Percentages stored as 0 to 1 rather than 0 to 100 will not be shown correctly unless adjusted, 
#>           because rounding info says 0 digits when the intent is to show 0 digits after the 0-100 percent number.
#> [1] 8
  
  cbind(table_rounding_info(names_all_r), fixcolnames(names_all_r, "r", "long"))
#>        [,1]
#>   [1,] "0" 
#>   [2,] "0" 
#>   [3,] "0" 
#>   [4,] "0" 
#>   [5,] "0" 
#>   [6,] "0" 
#>   [7,] "0" 
#>   [8,] "0" 
#>   [9,] "0" 
#>  [10,] "0" 
#>  [11,] "0" 
#>  [12,] "0" 
#>  [13,] "0" 
#>  [14,] "0" 
#>  [15,] "0" 
#>  [16,] "0" 
#>  [17,] "0" 
#>  [18,] "0" 
#>  [19,] "0" 
#>  [20,] "0" 
#>  [21,] "0" 
#>  [22,] "0" 
#>  [23,] "0" 
#>  [24,] "0" 
#>  [25,] "0" 
#>  [26,] "0" 
#>  [27,] "0" 
#>  [28,] "0" 
#>  [29,] "0" 
#>  [30,] "0" 
#>  [31,] "0" 
#>  [32,] "0" 
#>  [33,] "0" 
#>  [34,] "0" 
#>  [35,] "0" 
#>  [36,] "0" 
#>  [37,] "0" 
#>  [38,] "0" 
#>  [39,] "0" 
#>  [40,] "0" 
#>  [41,] "0" 
#>  [42,] "0" 
#>  [43,] "0" 
#>  [44,] "0" 
#>  [45,] "0" 
#>  [46,] "0" 
#>  [47,] "0" 
#>  [48,] "0" 
#>  [49,] "0" 
#>  [50,] "0" 
#>  [51,] "0" 
#>  [52,] "0" 
#>  [53,] "0" 
#>  [54,] "0" 
#>  [55,] "0" 
#>  [56,] "0" 
#>  [57,] "0" 
#>  [58,] "0" 
#>  [59,] "0" 
#>  [60,] "0" 
#>  [61,] "0" 
#>  [62,] "0" 
#>  [63,] "0" 
#>  [64,] "0" 
#>  [65,] "0" 
#>  [66,] "1" 
#>  [67,] "1" 
#>  [68,] "1" 
#>  [69,] "1" 
#>  [70,] "1" 
#>  [71,] "1" 
#>  [72,] "1" 
#>  [73,] "1" 
#>  [74,] "1" 
#>  [75,] "1" 
#>  [76,] "1" 
#>  [77,] "1" 
#>  [78,] "1" 
#>  [79,] "1" 
#>  [80,] "1" 
#>  [81,] "1" 
#>  [82,] "1" 
#>  [83,] "1" 
#>  [84,] "1" 
#>  [85,] "1" 
#>  [86,] "0" 
#>  [87,] "0" 
#>  [88,] "0" 
#>  [89,] "0" 
#>  [90,] "0" 
#>  [91,] "0" 
#>  [92,] "0" 
#>  [93,] "0" 
#>  [94,] "0" 
#>  [95,] "0" 
#>  [96,] "0" 
#>  [97,] "0" 
#>  [98,] "0" 
#>  [99,] "0" 
#> [100,] "0" 
#> [101,] "0" 
#> [102,] "0" 
#> [103,] "0" 
#> [104,] "0" 
#> [105,] "0" 
#> [106,] "0" 
#> [107,] "0" 
#> [108,] "0" 
#> [109,] "0" 
#> [110,] "0" 
#> [111,] "0" 
#> [112,] "0" 
#> [113,] "0" 
#> [114,] "0" 
#> [115,] "0" 
#> [116,] "0" 
#> [117,] "0" 
#> [118,] "0" 
#> [119,] "0" 
#> [120,] "0" 
#> [121,] "0" 
#> [122,] "0" 
#> [123,] "0" 
#> [124,] "0" 
#> [125,] "0" 
#> [126,] "1" 
#> [127,] "1" 
#> [128,] "1" 
#> [129,] "1" 
#> [130,] "1" 
#> [131,] "1" 
#> [132,] "1" 
#> [133,] "1" 
#> [134,] "1" 
#> [135,] "1" 
#> [136,] "1" 
#> [137,] "1" 
#> [138,] "1" 
#> [139,] "1" 
#> [140,] "1" 
#> [141,] "1" 
#> [142,] "0" 
#> [143,] "0" 
#> [144,] "0" 
#> [145,] "0" 
#> [146,] "0" 
#> [147,] "0" 
#> [148,] "0" 
#> [149,] "0" 
#> [150,] "2" 
#> [151,] "1" 
#> [152,] "0" 
#> [153,] "2" 
#> [154,] "2" 
#> [155,] "1" 
#> [156,] "0" 
#> [157,] "2" 
#> [158,] "2" 
#> [159,] "1" 
#> [160,] "0" 
#> [161,] "1" 
#> [162,] "0" 
#> [163,] "2" 
#> [164,] "1" 
#> [165,] "0" 
#> [166,] "2" 
#> [167,] "2" 
#> [168,] "1" 
#> [169,] "0" 
#> [170,] "2" 
#> [171,] "2" 
#> [172,] "1" 
#> [173,] "0" 
#> [174,] "1" 
#> [175,] "0" 
#> [176,] "2" 
#> [177,] "1" 
#> [178,] "0" 
#> [179,] "2" 
#> [180,] "2" 
#> [181,] "1" 
#> [182,] "0" 
#> [183,] "2" 
#> [184,] "2" 
#> [185,] "1" 
#> [186,] "0" 
#> [187,] "1" 
#> [188,] "0" 
#> [189,] "0" 
#> [190,] "0" 
#> [191,] "0" 
#> [192,] "0" 
#> [193,] "0" 
#> [194,] "0" 
#> [195,] "0" 
#> [196,] "0" 
#> [197,] "0" 
#> [198,] "0" 
#> [199,] "0" 
#> [200,] "0" 
#> [201,] "0" 
#> [202,] "0" 
#> [203,] "0" 
#> [204,] "0" 
#> [205,] "0" 
#> [206,] "0" 
#> [207,] "0" 
#> [208,] "0" 
#> [209,] "0" 
#> [210,] "0" 
#> [211,] "0" 
#> [212,] "0" 
#> [213,] "0" 
#> [214,] "0" 
#> [215,] "1" 
#> [216,] "1" 
#> [217,] "1" 
#> [218,] "1" 
#> [219,] "1" 
#> [220,] "1" 
#> [221,] "1" 
#> [222,] "1" 
#> [223,] "1" 
#> [224,] "1" 
#> [225,] "1" 
#> [226,] "1" 
#> [227,] "1" 
#> [228,] "1" 
#> [229,] "1" 
#> [230,] "1" 
#> [231,] "1" 
#> [232,] "1" 
#> [233,] "1" 
#> [234,] "1" 
#> [235,] "1" 
#> [236,] "1" 
#> [237,] "1" 
#> [238,] "1" 
#> [239,] "1" 
#> [240,] "1" 
#> [241,] "3" 
#> [242,] "3" 
#> [243,] "3" 
#> [244,] "3" 
#> [245,] "3" 
#> [246,] "3" 
#> [247,] "3" 
#> [248,] "3" 
#> [249,] "3" 
#> [250,] "3" 
#> [251,] "3" 
#> [252,] "3" 
#> [253,] "3" 
#> [254,] "3" 
#> [255,] "3" 
#> [256,] "3" 
#> [257,] "3" 
#> [258,] "3" 
#> [259,] "3" 
#> [260,] "3" 
#> [261,] "3" 
#> [262,] "3" 
#> [263,] "3" 
#> [264,] "3" 
#> [265,] "3" 
#> [266,] "3" 
#> [267,] "3" 
#> [268,] "3" 
#> [269,] "3" 
#> [270,] "3" 
#> [271,] "3" 
#> [272,] "3" 
#> [273,] "3" 
#> [274,] "3" 
#> [275,] "3" 
#> [276,] "3" 
#> [277,] "3" 
#> [278,] "3" 
#> [279,] "3" 
#> [280,] "3" 
#> [281,] "3" 
#> [282,] "3" 
#> [283,] "3" 
#> [284,] "3" 
#> [285,] "3" 
#> [286,] "3" 
#> [287,] "3" 
#> [288,] "3" 
#> [289,] "3" 
#> [290,] "3" 
#> [291,] "3" 
#> [292,] "3" 
#> [293,] "0" 
#> [294,] "0" 
#> [295,] "0" 
#> [296,] "0" 
#> [297,] "0" 
#> [298,] "0" 
#> [299,] "0" 
#> [300,] "0" 
#> [301,] "0" 
#> [302,] "0" 
#> [303,] "0" 
#> [304,] "0" 
#> [305,] "0" 
#> [306,] "0" 
#> [307,] "0" 
#> [308,] "0" 
#> [309,] "0" 
#> [310,] "0" 
#> [311,] "0" 
#> [312,] "0" 
#> [313,] "0" 
#> [314,] "0" 
#> [315,] "0" 
#> [316,] "0" 
#> [317,] "0" 
#> [318,] "0" 
#> [319,] "0" 
#> [320,] "0" 
#> [321,] "0" 
#> [322,] "0" 
#> [323,] "0" 
#> [324,] "0" 
#> [325,] "0" 
#> [326,] "0" 
#> [327,] "0" 
#> [328,] "0" 
#> [329,] "0" 
#> [330,] "0" 
#> [331,] "0" 
#> [332,] "0" 
#> [333,] "0" 
#> [334,] "0" 
#> [335,] "0" 
#> [336,] "0" 
#> [337,] "0" 
#> [338,] "0" 
#> [339,] "0" 
#> [340,] "0" 
#> [341,] "0" 
#> [342,] "0" 
#> [343,] "0" 
#> [344,] "0" 
#>        [,2]                                                                                            
#>   [1,] "Total Population"                                                                              
#>   [2,] "Non-POC resident count"                                                                        
#>   [3,] "Count of Population for whom Poverty Status is Determined"                                     
#>   [4,] "Count of Population Age 25 up"                                                                 
#>   [5,] "Count of Households"                                                                           
#>   [6,] "Universe for percent unemployed (denominator, count)"                                          
#>   [7,] "Count of Housing Units Built Pre 1960"                                                         
#>   [8,] "Built housing units count (denominator for percent pre 1960)"                                  
#>   [9,] "Low income resident count"                                                                     
#>  [10,] "Limited English-speaking Households"                                                           
#>  [11,] "Unemployed resident count"                                                                     
#>  [12,] "Less Than High School Education resident count"                                                
#>  [13,] "Under Age 5 resident count"                                                                    
#>  [14,] "Over Age 64 resident count"                                                                    
#>  [15,] "People of Color resident count"                                                                
#>  [16,] "Demographic Index"                                                                             
#>  [17,] "Supplemental Demographic Index"                                                                
#>  [18,] "% Low Income"                                                                                  
#>  [19,] "% in limited English-speaking Households"                                                      
#>  [20,] "% Unemployed"                                                                                  
#>  [21,] "% with Less Than High School Education"                                                        
#>  [22,] "Low life expectancy"                                                                           
#>  [23,] "% under Age 5"                                                                                 
#>  [24,] "% over Age 64"                                                                                 
#>  [25,] "% People of Color"                                                                             
#>  [26,] "US Average of Demographic Index"                                                               
#>  [27,] "US Average of Supplemental Demographic Index"                                                  
#>  [28,] "US average for % Low Income"                                                                   
#>  [29,] "US average for % in limited English-speaking Households"                                       
#>  [30,] "US average for % Unemployed"                                                                   
#>  [31,] "US average for % with Less Than High School Education"                                         
#>  [32,] "US average for Low life expectancy"                                                            
#>  [33,] "US average for % under Age 5"                                                                  
#>  [34,] "US average for % over Age 64"                                                                  
#>  [35,] "US average for % People of Color"                                                              
#>  [36,] "State Average of Demographic Index"                                                            
#>  [37,] "State Average of Supplemental Demographic Index"                                               
#>  [38,] "State average for % Low Income"                                                                
#>  [39,] "State average for % in limited English-speaking Households"                                    
#>  [40,] "State average for % Unemployed"                                                                
#>  [41,] "State average for % with Less Than High School Education"                                      
#>  [42,] "State average for Low life expectancy"                                                         
#>  [43,] "State average for % under Age 5"                                                               
#>  [44,] "State average for % over Age 64"                                                               
#>  [45,] "State average for % People of Color"                                                           
#>  [46,] "US percentile for Demographic Index"                                                           
#>  [47,] "US percentile for Supplemental Demographic Index"                                              
#>  [48,] "US percentile for % Low Income"                                                                
#>  [49,] "US percentile for % in limited English-speaking Households"                                    
#>  [50,] "US percentile for % Unemployed"                                                                
#>  [51,] "US percentile for % with Less Than High School Education"                                      
#>  [52,] "US percentile for Low life expectancy"                                                         
#>  [53,] "US percentile for % under Age 5"                                                               
#>  [54,] "US percentile for % over Age 64"                                                               
#>  [55,] "US percentile for % People of Color"                                                           
#>  [56,] "State percentile for Demographic Index"                                                        
#>  [57,] "State percentile for Supplemental Demographic Index"                                           
#>  [58,] "State percentile for % Low Income"                                                             
#>  [59,] "State percentile for % in limited English-speaking Households"                                 
#>  [60,] "State percentile for % Unemployed"                                                             
#>  [61,] "State percentile for % with Less Than High School Education"                                   
#>  [62,] "State percentile for Low life expectancy"                                                      
#>  [63,] "State percentile for % under Age 5"                                                            
#>  [64,] "State percentile for % over Age 64"                                                            
#>  [65,] "State percentile for % People of Color"                                                        
#>  [66,] "Ratio to US avg Demog Index"                                                                   
#>  [67,] "Ratio to US avg Suppl Demog Index"                                                             
#>  [68,] "Ratio to US avg % Low-inc."                                                                    
#>  [69,] "Ratio to US avg % Limited English"                                                             
#>  [70,] "Ratio to US avg % Unemployed"                                                                  
#>  [71,] "Ratio to US avg % < High School"                                                               
#>  [72,] "Ratio to US avg Low life expectancy"                                                           
#>  [73,] "Ratio to US avg % < age 5"                                                                     
#>  [74,] "Ratio to US avg % > age 64"                                                                    
#>  [75,] "Ratio to US avg % People of Color"                                                             
#>  [76,] "Ratio to State avg Demog.Ind."                                                                 
#>  [77,] "Ratio to State avg Suppl Demog Index"                                                          
#>  [78,] "Ratio to State avg % Low-inc."                                                                 
#>  [79,] "Ratio to State avg % Limited English"                                                          
#>  [80,] "Ratio to State avg % Unemployed"                                                               
#>  [81,] "Ratio to State avg % < High School"                                                            
#>  [82,] "Ratio to State avg Low life expectancy"                                                        
#>  [83,] "Ratio to State avg % < age 5"                                                                  
#>  [84,] "Ratio to State avg % > age 64"                                                                 
#>  [85,] "Ratio to State avg % People of Color"                                                          
#>  [86,] "% Hispanic or Latino"                                                                          
#>  [87,] "% Black or African American (non-Hispanic, single race)"                                       
#>  [88,] "% Asian (non-Hispanic, single race)"                                                           
#>  [89,] "% American Indian and Alaska Native (non-Hispanic, single race)"                               
#>  [90,] "% Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"                      
#>  [91,] "% Other race (non-Hispanic, single race)"                                                      
#>  [92,] "% Two or more races (non-Hispanic)"                                                            
#>  [93,] "% White (non-Hispanic, single race)"                                                           
#>  [94,] "US avg % Hispanic or Latino"                                                                   
#>  [95,] "US avg % Black or African American (non-Hispanic, single race)"                                
#>  [96,] "US avg % Asian (non-Hispanic, single race)"                                                    
#>  [97,] "US avg % American Indian and Alaska Native (non-Hispanic, single race)"                        
#>  [98,] "US avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"               
#>  [99,] "US avg % Other race (non-Hispanic, single race)"                                               
#> [100,] "US avg % Two or more races (non-Hispanic)"                                                     
#> [101,] "US avg % White (non-Hispanic, single race)"                                                    
#> [102,] "State avg % Hispanic or Latino"                                                                
#> [103,] "State avg % Black or African American (non-Hispanic, single race)"                             
#> [104,] "State avg % Asian (non-Hispanic, single race)"                                                 
#> [105,] "State avg % American Indian and Alaska Native (non-Hispanic, single race)"                     
#> [106,] "State avg % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"            
#> [107,] "State avg % Other race (non-Hispanic, single race)"                                            
#> [108,] "State avg % Two or more races (non-Hispanic)"                                                  
#> [109,] "State avg % White (non-Hispanic, single race)"                                                 
#> [110,] "US percentile for % Hispanic or Latino"                                                        
#> [111,] "US percentile for % Black or African American (non-Hispanic, single race)"                     
#> [112,] "US percentile for % Asian (non-Hispanic, single race)"                                         
#> [113,] "US percentile for % American Indian and Alaska Native (non-Hispanic, single race)"             
#> [114,] "US percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"    
#> [115,] "US percentile for % Other race (non-Hispanic, single race)"                                    
#> [116,] "US percentile for % Two or more races (non-Hispanic)"                                          
#> [117,] "US percentile for % White (non-Hispanic, single race)"                                         
#> [118,] "State percentile for % Hispanic or Latino"                                                     
#> [119,] "State percentile for % Black or African American (non-Hispanic, single race)"                  
#> [120,] "State percentile for % Asian (non-Hispanic, single race)"                                      
#> [121,] "State percentile for % American Indian and Alaska Native (non-Hispanic, single race)"          
#> [122,] "State percentile for % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)" 
#> [123,] "State percentile for % Other race (non-Hispanic, single race)"                                 
#> [124,] "State percentile for % Two or more races (non-Hispanic)"                                       
#> [125,] "State percentile for % White (non-Hispanic, single race)"                                      
#> [126,] "Ratio to US avg % Hispanic"                                                                    
#> [127,] "Ratio to US avg % Black (NHA)"                                                                 
#> [128,] "Ratio to US avg % Asian (NHA)"                                                                 
#> [129,] "Ratio to US avg % American Indian Alaska Native (NHA)"                                         
#> [130,] "Ratio to US avg % Native Hawaiian Pacific Islander (NHA)"                                      
#> [131,] "Ratio to US avg % Other single race (NHA)"                                                     
#> [132,] "Ratio to US avg % Two or more race (NH)"                                                       
#> [133,] "Ratio to US avg % White nonHispanic single race (NHA)"                                         
#> [134,] "Ratio to State avg % Hispanic"                                                                 
#> [135,] "Ratio to State avg % American Indian Alaska Native (NHA)"                                      
#> [136,] "Ratio to State avg % Black (NHA)"                                                              
#> [137,] "Ratio to State avg % Asian (NHA)"                                                              
#> [138,] "Ratio to State avg % Other single race (NHA)"                                                  
#> [139,] "Ratio to State avg % Two or more race (NH)"                                                    
#> [140,] "Ratio to State avg % Native Hawaiian Pacific Islander (NHA)"                                   
#> [141,] "Ratio to State avg % White nonHispanic single race"                                            
#> [142,] "Count of Hispanic or Latino"                                                                   
#> [143,] "Count of Black or African American (non-Hispanic, single race)"                                
#> [144,] "Count of Asian (non-Hispanic, single race)"                                                    
#> [145,] "Count of American Indian and Alaska Native (non-Hispanic, single race)"                        
#> [146,] "Count of Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)"               
#> [147,] "Count of Other race (non-Hispanic, single race)"                                               
#> [148,] "Count of Two or more races (non-Hispanic)"                                                     
#> [149,] "Count of White (non-Hispanic, single race)"                                                    
#> [150,] "Particulate Matter (PM 2.5 in ug/m3)"                                                          
#> [151,] "Ozone (ppb)"                                                                                   
#> [152,] "Air Toxics Cancer Risk (risk per million)"                                                     
#> [153,] "Air Toxics Respiratory Hazard Index"                                                           
#> [154,] "Air Toxics Diesel Particulate Matter (ug/m3)"                                                  
#> [155,] "Lead Paint Indicator (% pre-1960s housing)"                                                    
#> [156,] "Traffic Proximity and Volume (daily traffic count/distance to road)"                           
#> [157,] "Superfund Proximity (site count/km distance)"                                                  
#> [158,] "RMP Proximity (facility count/km distance)"                                                    
#> [159,] "Indicator for Hazardous Waste Proximity (facility count/km distance)"                          
#> [160,] "Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"                     
#> [161,] "Underground Storage Tanks (UST) indicator"                                                     
#> [162,] "Toxic Releases to Air"                                                                         
#> [163,] "US average for Particulate Matter (PM 2.5 in ug/m3)"                                           
#> [164,] "US average for Ozone (ppb)"                                                                    
#> [165,] "US average for Air Toxics Cancer Risk (risk per million)"                                      
#> [166,] "US average for Air Toxics Respiratory Hazard Index"                                            
#> [167,] "US average for Air Toxics Diesel Particulate Matter (ug/m3)"                                   
#> [168,] "US average for Lead Paint Indicator (% pre-1960s housing)"                                     
#> [169,] "US average for Traffic Proximity and Volume (daily traffic count/distance to road)"            
#> [170,] "US average for Superfund Proximity (site count/km distance)"                                   
#> [171,] "US average for RMP Proximity (facility count/km distance)"                                     
#> [172,] "US average for Hazardous Waste Proximity (facility count/km distance)"                         
#> [173,] "US average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"      
#> [174,] "US average for Underground Storage Tanks (UST) indicator"                                      
#> [175,] "US avg Toxic Releases to Air"                                                                  
#> [176,] "State average for Particulate Matter (PM 2.5 in ug/m3)"                                        
#> [177,] "State average for Ozone (ppb)"                                                                 
#> [178,] "State average for Air Toxics Cancer Risk (risk per million)"                                   
#> [179,] "State average for Air Toxics Respiratory Hazard Index"                                         
#> [180,] "State average for Air Toxics Diesel Particulate Matter (ug/m3)"                                
#> [181,] "State average for Lead Paint Indicator (% pre-1960s housing)"                                  
#> [182,] "State average for Traffic Proximity and Volume (daily traffic count/distance to road)"         
#> [183,] "State average for Superfund Proximity (site count/km distance)"                                
#> [184,] "State average for RMP Proximity (facility count/km distance)"                                  
#> [185,] "State average for Hazardous Waste Proximity (facility count/km distance)"                      
#> [186,] "State average for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"   
#> [187,] "State average for Underground Storage Tanks (UST) indicator"                                   
#> [188,] "State avg Toxic Releases to Air"                                                               
#> [189,] "US percentile for Particulate Matter (PM 2.5 in ug/m3)"                                        
#> [190,] "US percentile for Ozone (ppb)"                                                                 
#> [191,] "US percentile for Air Toxics Cancer Risk (risk per million)"                                   
#> [192,] "US percentile for Air Toxics Respiratory Hazard Index"                                         
#> [193,] "US percentile for Air Toxics Diesel Particulate Matter (ug/m3)"                                
#> [194,] "US percentile for Lead Paint Indicator (% pre-1960s housing)"                                  
#> [195,] "US percentile for Traffic Proximity and Volume (daily traffic count/distance to road)"         
#> [196,] "US percentile for Superfund Proximity (site count/km distance)"                                
#> [197,] "US percentile for RMP Proximity (facility count/km distance)"                                  
#> [198,] "US percentile for Hazardous Waste Proximity (facility count/km distance)"                      
#> [199,] "US percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"   
#> [200,] "US percentile for Underground Storage Tanks (UST) indicator"                                   
#> [201,] "US percentile for Toxic Releases to Air"                                                       
#> [202,] "State percentile for Particulate Matter (PM 2.5 in ug/m3)"                                     
#> [203,] "State percentile for Ozone (ppb)"                                                              
#> [204,] "State percentile for Air Toxics Cancer Risk (risk per million)"                                
#> [205,] "State percentile for Air Toxics Respiratory Hazard Index"                                      
#> [206,] "State percentile for Air Toxics Diesel Particulate Matter (ug/m3)"                             
#> [207,] "State percentile for Lead Paint Indicator (% pre-1960s housing)"                               
#> [208,] "State percentile for Traffic Proximity and Volume (daily traffic count/distance to road)"      
#> [209,] "State percentile for Superfund Proximity (site count/km distance)"                             
#> [210,] "State percentile for RMP Proximity (facility count/km distance)"                               
#> [211,] "State percentile for Hazardous Waste Proximity (facility count/km distance)"                   
#> [212,] "State percentile for Wastewater Discharge Indicator (toxicity-weighted concentration/distance)"
#> [213,] "State percentile for Underground Storage Tanks (UST) indicator"                                
#> [214,] "State percentile for Toxic Releases to Air"                                                    
#> [215,] "Ratio to US avg PM2.5"                                                                         
#> [216,] "Ratio to US avg Ozone"                                                                         
#> [217,] "Ratio to US avg Cancer risk"                                                                   
#> [218,] "Ratio to US avg Respiratory"                                                                   
#> [219,] "Ratio to US avg Diesel PM"                                                                     
#> [220,] "Ratio to US avg % built pre-1960"                                                              
#> [221,] "Ratio to US avg Traffic"                                                                       
#> [222,] "Ratio to US avg NPL"                                                                           
#> [223,] "Ratio to US avg RMP"                                                                           
#> [224,] "Ratio to US avg TSDF"                                                                          
#> [225,] "Ratio to US avg NPDES"                                                                         
#> [226,] "Ratio to US avg UST"                                                                           
#> [227,] "Ratio to US avg Toxic Releases to Air"                                                         
#> [228,] "Ratio to State avg PM2.5"                                                                      
#> [229,] "Ratio to State avg Ozone"                                                                      
#> [230,] "Ratio to State avg Cancer risk"                                                                
#> [231,] "Ratio to State avg Respiratory"                                                                
#> [232,] "Ratio to State avg Diesel PM"                                                                  
#> [233,] "Ratio to State avg % built pre-1960"                                                           
#> [234,] "Ratio to State avg Traffic"                                                                    
#> [235,] "Ratio to State avg NPL"                                                                        
#> [236,] "Ratio to State avg RMP"                                                                        
#> [237,] "Ratio to State avg TSDF"                                                                       
#> [238,] "Ratio to State avg NPDES"                                                                      
#> [239,] "Ratio to State avg UST"                                                                        
#> [240,] "Ratio to State avg Toxic Releases to Air"                                                      
#> [241,] "Particulate Matter 2.5 EJ Index"                                                               
#> [242,] "Ozone EJ Index"                                                                                
#> [243,] "Air toxics cancer risk EJ Index"                                                               
#> [244,] "Air toxics respiratory HI EJ Index"                                                            
#> [245,] "Diesel particulate matter EJ Index"                                                            
#> [246,] "Lead paint EJ Index"                                                                           
#> [247,] "US raw Traffic proximity EJ Index"                                                             
#> [248,] "US raw Superfund Proximity EJ Index"                                                           
#> [249,] "RMP Facility Proximity EJ Index"                                                               
#> [250,] "Hazardous waste proximity EJ Index"                                                            
#> [251,] "US raw Wastewater discharge EJ Index"                                                          
#> [252,] "US raw Underground storage tanks EJ Index"                                                     
#> [253,] "Toxic Releases to Air EJ Index"                                                                
#> [254,] "State raw Particulate Matter 2.5 EJ Index"                                                     
#> [255,] "State raw Ozone EJ Index"                                                                      
#> [256,] "State Air toxics cancer risk EJ Index"                                                         
#> [257,] "State Air toxics respiratory HI EJ Index"                                                      
#> [258,] "State Diesel particulate matter EJ Index"                                                      
#> [259,] "State Lead paint EJ Index"                                                                     
#> [260,] "State raw Traffic proximity EJ Index"                                                          
#> [261,] "State raw Superfund Proximity EJ Index"                                                        
#> [262,] "State raw RMP Facility Proximity EJ Index"                                                     
#> [263,] "State Hazardous waste proximity EJ Index"                                                      
#> [264,] "State raw Wastewater discharge EJ Index"                                                       
#> [265,] "State raw Underground storage tanks EJ Index"                                                  
#> [266,] "State Toxic Releases to Air EJ Index"                                                          
#> [267,] "Particulate Matter 2.5 Supplemental Index"                                                     
#> [268,] "Ozone Supplemental Index"                                                                      
#> [269,] "Air Toxics Cancer Risk Supplemental Index"                                                     
#> [270,] "Air Toxics Respiratory HI Supplemental Index"                                                  
#> [271,] "Diesel Particulate Matter Supplemental Index"                                                  
#> [272,] "Lead Paint Supplemental Index"                                                                 
#> [273,] "Traffic Proximity Supplemental Index"                                                          
#> [274,] "Superfund Proximity Supplemental Index"                                                        
#> [275,] "RMP Facility Proximity Supplemental Index"                                                     
#> [276,] "Hazardous Waste Proximity Supplemental Index"                                                  
#> [277,] "Wastewater Discharge Supplemental Index"                                                       
#> [278,] "Underground Storage Tanks Supplemental Index"                                                  
#> [279,] "EJ Supp: Toxic Releases to Air (raw)"                                                          
#> [280,] "State raw Particulate Matter 2.5 Supplemental Index"                                           
#> [281,] "State raw Ozone Supplemental Index"                                                            
#> [282,] "State Air toxics cancer risk Supplemental Index"                                               
#> [283,] "State Air toxics respiratory HI Supplemental Index"                                            
#> [284,] "State Diesel particulate matter Supplemental Index"                                            
#> [285,] "State Lead paint Supplemental Index"                                                           
#> [286,] "State raw Traffic proximity Supplemental Index"                                                
#> [287,] "State raw Superfund Proximity Supplemental Index"                                              
#> [288,] "State raw RMP Facility Proximity Supplemental Index"                                           
#> [289,] "State Hazardous waste proximity Supplemental Index"                                            
#> [290,] "State raw Wastewater discharge Supplemental Index"                                             
#> [291,] "State raw Underground storage tanks Supplemental Index"                                        
#> [292,] "State Toxic Releases to Air Supplemental Index"                                                
#> [293,] "US percentile for EJ Index for Particulate Matter (PM 2.5)"                                    
#> [294,] "US percentile for EJ Index for Ozone"                                                          
#> [295,] "US percentile for EJ Index for Air Toxics Cancer Risk"                                         
#> [296,] "US percentile for EJ Index for Air Toxics Respiratory Hazard Index"                            
#> [297,] "US percentile for EJ Index for Air Toxics Diesel Particulate Matter"                           
#> [298,] "US percentile for EJ Index for Lead Paint Indicator"                                           
#> [299,] "US percentile for EJ Index for Traffic Proximity and Volume"                                   
#> [300,] "US percentile for EJ Index for Superfund Proximity"                                            
#> [301,] "US percentile for EJ Index for RMP Proximity"                                                  
#> [302,] "US percentile for EJ Index for Hazardous Waste Proximity"                                      
#> [303,] "US percentile for EJ Index for Wastewater Discharge Indicator"                                 
#> [304,] "US percentile for EJ Index for Underground Storage Tanks (UST) indicator"                      
#> [305,] "US percentile for EJ Index for Toxic Releases to Air"                                          
#> [306,] "State percentile for EJ Index for Particulate Matter (PM 2.5)"                                 
#> [307,] "State percentile for EJ Index for Ozone"                                                       
#> [308,] "State percentile for EJ Index for Air Toxics Cancer Risk"                                      
#> [309,] "State percentile for EJ Index for Air Toxics Respiratory Hazard Index"                         
#> [310,] "State percentile for EJ Index for Air Toxics Diesel Particulate Matter"                        
#> [311,] "State percentile for EJ Index for Lead Paint Indicator"                                        
#> [312,] "State percentile for EJ Index for Traffic Proximity and Volume"                                
#> [313,] "State percentile for EJ Index for Superfund Proximity"                                         
#> [314,] "State percentile for EJ Index for RMP Proximity"                                               
#> [315,] "State percentile for EJ Index for Hazardous Waste Proximity"                                   
#> [316,] "State percentile for EJ Index for Wastewater Discharge Indicator"                              
#> [317,] "State percentile for EJ Index for Underground Storage Tanks (UST) indicator"                   
#> [318,] "State percentile for EJ Index for Toxic Releases to Air"                                       
#> [319,] "US percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)"                       
#> [320,] "US percentile for EJ Supplemental Index for Ozone"                                             
#> [321,] "US percentile for EJ Supplemental Index for Air Toxics Cancer Risk"                            
#> [322,] "US percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index"               
#> [323,] "US percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter"              
#> [324,] "US percentile for EJ Supplemental Index for Lead Paint Indicator"                              
#> [325,] "US percentile for EJ Supplemental Index for Traffic Proximity and Volume"                      
#> [326,] "US percentile for EJ Supplemental Index for Superfund Proximity"                               
#> [327,] "US percentile for EJ Supplemental Index for RMP Proximity"                                     
#> [328,] "US percentile for EJ Supplemental Index for Hazardous Waste Proximity"                         
#> [329,] "US percentile for EJ Supplemental Index for Wastewater Discharge Indicator"                    
#> [330,] "US percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator"         
#> [331,] "US Percentile for Toxic Releases to Air Supplemental Index"                                    
#> [332,] "State percentile for EJ Supplemental Index for Particulate Matter (PM 2.5)"                    
#> [333,] "State percentile for EJ Supplemental Index for Ozone"                                          
#> [334,] "State percentile for EJ Supplemental Index for Air Toxics Cancer Risk"                         
#> [335,] "State percentile for EJ Supplemental Index for Air Toxics Respiratory Hazard Index"            
#> [336,] "State percentile for EJ Supplemental Index for Air Toxics Diesel Particulate Matter"           
#> [337,] "State percentile for EJ Supplemental Index for Lead Paint Indicator"                           
#> [338,] "State percentile for EJ Supplemental Index for Traffic Proximity and Volume"                   
#> [339,] "State percentile for EJ Supplemental Index for Superfund Proximity"                            
#> [340,] "State percentile for EJ Supplemental Index for RMP Proximity"                                  
#> [341,] "State percentile for EJ Supplemental Index for Hazardous Waste Proximity"                      
#> [342,] "State percentile for EJ Supplemental Index for Wastewater Discharge Indicator"                 
#> [343,] "State percentile for EJ Supplemental Index for Underground Storage Tanks (UST) indicator"      
#> [344,] "State Percentile for Toxic Releases to Air Supplemental Index"