Get URLs of EJScreen reports

Get URL(s) for EJScreen standard report on residents near given point(s)

Usage

url_ejscreen_report(
  lat = "",
  lon = "",
  radius = "",
  as_html = FALSE,
  linktext,
  mobile = FALSE,
  areatype = "",
  areaid = "",
  namestr = "",
  shapefile = NULL,
  wkid = 4326,
  unit = 9035,
  f = "report",
  interactiveprompt = FALSE
)

Arguments

lat

one or more latitudes (or a table with lat, lon columns, or filepath with that, or omit to interactively select file)

lon

one or more longitudes (or omitted – see lat parameter details)

radius

miles radius

as_html

Whether to return as just the urls or as html hyperlinks to use in a DT::datatable() for example

linktext

used as text for hyperlinks, if supplied and as_html=TRUE

mobile

If TRUE, provides URL for the mobile browser version, not desktop version

areatype

passed as areatype= in API, inferred if not provided but areaid is provided

areaid

fips codes if used, passed as areaid= in API, can be FIPS for blockgroups, tracts, counties.

namestr

The character string text to show on the report as the name of the place

shapefile

not implemented

wkid

default is 4326 -WGS84 - World Geodetic System 1984, used in GPS - see (https://epsg.io/4326)

unit

default is 9035 which means miles; for kilometers use 9036

f

can be "report" or "pjson" or "json"

interactiveprompt

passed to sitepoints_from_any()

Value

URL(s)

Details

Also see ejscreenRESTbroker() and ejscreenapi1()

and (https://ejscreen.epa.gov/mapper/EJAPIinstructions.pdf)

and (https://ejscreen.epa.gov/mapper/ejscreenapi1.html)

See also