Get metadata about a variable, like its type, short name, long definition, decimal places, etc.

This is just a way to query map_headernames, which has info about each indicator or variable used in EJAM.

Usage

varinfo(
  var = map_headernames$rname,
  info = colnames(map_headernames),
  varnametype = "rname"
)

Arguments

var

vector of variable names such as c("pctlowinc", "pm") or c(names_d, names_d_subgroups) (and must be found in the column of map_headernames indicated by varnametype parameter below).

info

types of metadata/info needed, such as "decimals", "long", etc. which should be among colnames of map_headernames, or alias like "long" as allowed by fixcolnames()

varnametype

optional. colname of map_headernames to use when looking for var, like "rname" or "api" or "long"

Value

data.frame of 1 or more rows, 1 or more columns, where

rowsnames are var (indicators like "pctmin")

colnames are info (metadata like "decimals")

Cells of table are metadata such as what type of indicator is that var, how many decimal places of rounding should be displayed for it in tables, etc.

Results can be character, numeric, etc. depending on what info is requested

Details

See map_headernames for what kind of information is available there. But if a variable appears twice+ in var or in map_headernames, info returned only for the 1st row of those.

Developers may wish to know about these functions, several of which are not exported:

varinfo()

varname2varcategory_ejam()

varname2vartype_ejam()

table_rounding_info()

EJAM table_signif_round_x100()

fixcolnames()

fixcolnames_infer()

EJAM fixcolnames2related()

EJAM fixmapheadernamescolname()

EJAM is.numericish()

EJAM varin_map_headernames()

EJAM namesbyvarlist()

EJAM names_whichlist()

EJAM names_whichlist_multi()

EJAM names_whichlist_multi_key()

EJAM formula_varname()

and the various related data objects like map_headernames and namez

See also

Examples

varinfo("traffic.score", "decimals")
#>               decimals
#> traffic.score        0
varinfo(names_d, "long")
#>                                                      long
#> Demog.Index                         Demographic Index USA
#> Demog.Index.Supp       Supplemental Demographic Index USA
#> pctlowinc                                    % Low Income
#> pctlingiso       % in limited English-speaking Households
#> pctunemployed                                % Unemployed
#> pctlths            % with Less Than High School Education
#> pctunder5                                   % under Age 5
#> pctover64                                   % over Age 64
#> pctmin                                  % People of Color
myvars <- c(names_d, names_d_subgroups, names_e)
myinfo <- "percentage"
cbind(  is.a.percentage = varinfo(myvars, myinfo) )
#>                  percentage
#> Demog.Index               1
#> Demog.Index.Supp          1
#> pctlowinc                 1
#> pctlingiso                1
#> pctunemployed             1
#> pctlths                   1
#> pctunder5                 1
#> pctover64                 1
#> pctmin                    1
#> pcthisp                   1
#> pctnhba                   1
#> pctnhaa                   1
#> pctnhaiana                1
#> pctnhnhpia                1
#> pctnhotheralone           1
#> pctnhmulti                1
#> pctnhwa                   1
#> pm                         
#> o3                         
#> no2                        
#> dpm                        
#> rsei                       
#> traffic.score              
#> pctpre1960                1
#> proximity.npl              
#> proximity.rmp              
#> proximity.tsdf             
#> ust                        
#> proximity.npdes            
#> drinking                   
cbind(varinfo(names_all_r, "pctile."))
#>                                                pctile.
#> aa                                                   0
#> age25up                                              0
#> aiana                                                0
#> api_li                                               0
#> area                                                 0
#> areaid                                               0
#> arealand                                             0
#> areatype                                             0
#> areawater                                            0
#> avg.Demog.Index                                      0
#> avg.Demog.Index.Supp                                 0
#> avg.dpm                                              0
#> avg.drinking                                         0
#> avg.lifexyears_synonym                               0
#> avg.lowlifex                                         0
#> avg.no2                                              0
#> avg.o3                                               0
#> avg.pctaa                                            0
#> avg.pctaiana                                         0
#> avg.pctba                                            0
#> avg.pctdisability                                    0
#> avg.pctfire                                          0
#> avg.pctfire30                                        0
#> avg.pctflood                                         0
#> avg.pctflood30                                       0
#> avg.pcthisp                                          0
#> avg.pctlingiso                                       0
#> avg.pctlowinc                                        0
#> avg.pctlths                                          0
#> avg.pctmin                                           0
#> avg.pctmulti                                         0
#> avg.pctnhaa                                          0
#> avg.pctnhaiana                                       0
#> avg.pctnhba                                          0
#> avg.pctnhmulti                                       0
#> avg.pctnhnhpia                                       0
#> avg.pctnhotheralone                                  0
#> avg.pctnhpia                                         0
#> avg.pctnhwa                                          0
#> avg.pctnobroadband                                   0
#> avg.pctnohealthinsurance                             0
#> avg.pctotheralone                                    0
#> avg.pctover64                                        0
#> avg.pctpre1960                                       0
#> avg.pctunder5                                        0
#> avg.pctunemployed                                    0
#> avg.pctwa                                            0
#> avg.pm                                               0
#> avg.proximity.npdes                                  0
#> avg.proximity.npl                                    0
#> avg.proximity.rmp                                    0
#> avg.proximity.tsdf                                   0
#> avg.rateasthma                                       0
#> avg.ratecancer                                       0
#> avg.rateheartdisease                                 0
#> avg.rsei                                             0
#> avg.traffic.score                                    0
#> avg.ust                                              0
#> ba                                                   0
#> bgcount_near_site                                    0
#> bgcount_zeropop_near_site                            0
#> blockcount_near_site                                 0
#> builtunits                                           0
#> count.ej.80up                                        0
#> count.ej.80up.supp                                   0
#> count.ej.80up2.eo                                    0
#> count.ej.80up2.supp                                  0
#> count.NPL                                            0
#> count.TSDF                                           0
#> countyname                                           0
#> Demog.Index                                          0
#> Demog.Index.State                                    0
#> Demog.Index.Supp                                     0
#> Demog.Index.Supp.State                               0
#> disab_universe                                       0
#> disability                                           0
#> distance_min                                         0
#> distance_min_avgperson                               0
#> dpm                                                  0
#> drinking                                             0
#> EJ.DISPARITY.dpm.eo                                  0
#> EJ.DISPARITY.dpm.supp                                0
#> EJ.DISPARITY.drinking.eo                             0
#> EJ.DISPARITY.drinking.supp                           0
#> EJ.DISPARITY.no2.eo                                  0
#> EJ.DISPARITY.no2.supp                                0
#> EJ.DISPARITY.o3.eo                                   0
#> EJ.DISPARITY.o3.supp                                 0
#> EJ.DISPARITY.pctpre1960.eo                           0
#> EJ.DISPARITY.pctpre1960.supp                         0
#> EJ.DISPARITY.pm.eo                                   0
#> EJ.DISPARITY.pm.supp                                 0
#> EJ.DISPARITY.proximity.npdes.eo                      0
#> EJ.DISPARITY.proximity.npdes.supp                    0
#> EJ.DISPARITY.proximity.npl.eo                        0
#> EJ.DISPARITY.proximity.npl.supp                      0
#> EJ.DISPARITY.proximity.rmp.eo                        0
#> EJ.DISPARITY.proximity.rmp.supp                      0
#> EJ.DISPARITY.proximity.tsdf.eo                       0
#> EJ.DISPARITY.proximity.tsdf.supp                     0
#> EJ.DISPARITY.rsei.eo                                 0
#> EJ.DISPARITY.rsei.supp                               0
#> EJ.DISPARITY.traffic.score.eo                        0
#> EJ.DISPARITY.traffic.score.supp                      0
#> EJ.DISPARITY.ust.eo                                  0
#> EJ.DISPARITY.ust.supp                                0
#> female                                               0
#> geometry                                             0
#> geometry.wkid                                        0
#> hhlds                                                0
#> hisp                                                 0
#> id                                                   0
#> ie_li                                                0
#> inputAreaMiles                                       0
#> lan_api                                              0
#> lan_eng_na                                           0
#> lan_ie                                               0
#> lan_nonenglish                                       0
#> lan_other                                            0
#> lan_spanish                                          0
#> lan_universe                                         0
#> lat                                                  0
#> lifexyears                                           0
#> lifexyears_synonym                                   0
#> lingiso                                              0
#> lon                                                  0
#> lowinc                                               0
#> lowlifex                                             0
#> lths                                                 0
#> male                                                 0
#> mins                                                 0
#> multi                                                0
#> nhaa                                                 0
#> nhaiana                                              0
#> nhba                                                 0
#> nhmulti                                              0
#> nhnhpia                                              0
#> nhotheralone                                         0
#> nhpia                                                0
#> nhwa                                                 0
#> no2                                                  0
#> nonmins                                              0
#> num_airpoll                                          0
#> num_brownfield                                       0
#> num_church                                           0
#> num_hospital                                         0
#> num_school                                           0
#> num_tri                                              0
#> num_waterdis                                         0
#> o3                                                   0
#> OBJECTID                                             0
#> occupiedunits                                        0
#> other_li                                             0
#> otheralone                                           0
#> over17                                               0
#> over64                                               0
#> ownedunits                                           0
#> p_chinese                                            0
#> p_korean                                             0
#> pctaa                                                0
#> pctaiana                                             0
#> pctapi_li                                            0
#> pctba                                                0
#> pctdisability                                        0
#> pctfemale                                            0
#> pctfire                                              0
#> pctfire30                                            0
#> pctflood                                             0
#> pctflood30                                           0
#> pcthisp                                              0
#> pctie_li                                             0
#> pctile.Demog.Index                                   1
#> pctile.Demog.Index.Supp                              1
#> pctile.dpm                                           1
#> pctile.drinking                                      1
#> pctile.EJ.DISPARITY.dpm.eo                           1
#> pctile.EJ.DISPARITY.dpm.supp                         1
#> pctile.EJ.DISPARITY.drinking.eo                      1
#> pctile.EJ.DISPARITY.drinking.supp                    1
#> pctile.EJ.DISPARITY.no2.eo                           1
#> pctile.EJ.DISPARITY.no2.supp                         1
#> pctile.EJ.DISPARITY.o3.eo                            1
#> pctile.EJ.DISPARITY.o3.supp                          1
#> pctile.EJ.DISPARITY.pctpre1960.eo                    1
#> pctile.EJ.DISPARITY.pctpre1960.supp                  1
#> pctile.EJ.DISPARITY.pm.eo                            1
#> pctile.EJ.DISPARITY.pm.supp                          1
#> pctile.EJ.DISPARITY.proximity.npdes.eo               1
#> pctile.EJ.DISPARITY.proximity.npdes.supp             1
#> pctile.EJ.DISPARITY.proximity.npl.eo                 1
#> pctile.EJ.DISPARITY.proximity.npl.supp               1
#> pctile.EJ.DISPARITY.proximity.rmp.eo                 1
#> pctile.EJ.DISPARITY.proximity.rmp.supp               1
#> pctile.EJ.DISPARITY.proximity.tsdf.eo                1
#> pctile.EJ.DISPARITY.proximity.tsdf.supp              1
#> pctile.EJ.DISPARITY.rsei.eo                          1
#> pctile.EJ.DISPARITY.rsei.supp                        1
#> pctile.EJ.DISPARITY.traffic.score.eo                 1
#> pctile.EJ.DISPARITY.traffic.score.supp               1
#> pctile.EJ.DISPARITY.ust.eo                           1
#> pctile.EJ.DISPARITY.ust.supp                         1
#> pctile.lifexyears_synonym                            1
#> pctile.lowlifex                                      1
#> pctile.no2                                           1
#> pctile.o3                                            1
#> pctile.pctaa                                         1
#> pctile.pctaiana                                      1
#> pctile.pctba                                         1
#> pctile.pctdisability                                 1
#> pctile.pctfire                                       1
#> pctile.pctfire30                                     1
#> pctile.pctflood                                      1
#> pctile.pctflood30                                    1
#> pctile.pcthisp                                       1
#> pctile.pctlingiso                                    1
#> pctile.pctlowinc                                     1
#> pctile.pctlths                                       1
#> pctile.pctmin                                        1
#> pctile.pctmulti                                      1
#> pctile.pctnhaa                                       1
#> pctile.pctnhaiana                                    1
#> pctile.pctnhba                                       1
#> pctile.pctnhmulti                                    1
#> pctile.pctnhnhpia                                    1
#> pctile.pctnhotheralone                               1
#> pctile.pctnhpia                                      1
#> pctile.pctnhwa                                       1
#> pctile.pctnobroadband                                1
#> pctile.pctnohealthinsurance                          1
#> pctile.pctotheralone                                 1
#> pctile.pctover64                                     1
#> pctile.pctpre1960                                    1
#> pctile.pctunder5                                     1
#> pctile.pctunemployed                                 1
#> pctile.pctwa                                         1
#> pctile.pm                                            1
#> pctile.proximity.npdes                               1
#> pctile.proximity.npl                                 1
#> pctile.proximity.rmp                                 1
#> pctile.proximity.tsdf                                1
#> pctile.rateasthma                                    1
#> pctile.ratecancer                                    1
#> pctile.rateheartdisease                              1
#> pctile.rsei                                          1
#> pctile.traffic.score                                 1
#> pctile.ust                                           1
#> pctlan_api                                           0
#> pctlan_arabic                                        0
#> pctlan_english                                       0
#> pctlan_french                                        0
#> pctlan_ie                                            0
#> pctlan_non_english                                  NA
#> pctlan_other                                         0
#> pctlan_other_asian                                   0
#> pctlan_other_ie                                      0
#> pctlan_rus_pol_slav                                  0
#> pctlan_spanish                                       0
#> pctlan_vietnamese                                    0
#> pctlingiso                                           0
#> pctlowinc                                            0
#> pctlths                                              0
#> pctmale                                              0
#> pctmin                                               0
#> pctmulti                                             0
#> pctnhaa                                              0
#> pctnhaiana                                           0
#> pctnhba                                              0
#> pctnhmulti                                           0
#> pctnhnhpia                                           0
#> pctnhotheralone                                      0
#> pctnhpia                                             0
#> pctnhwa                                              0
#> pctnobroadband                                       0
#> pctnohealthinsurance                                 0
#> pctother_li                                          0
#> pctotheralone                                        0
#> pctover17                                            0
#> pctover64                                            0
#> pctownedunits                                        0
#> pctpoor                                              0
#> pctpre1960                                           0
#> pctspanish_li                                        0
#> pctunder18                                           0
#> pctunder5                                            0
#> pctunemployed                                        0
#> pctwa                                                0
#> percapincome                                         0
#> placename                                            0
#> pm                                                   0
#> poor                                                 0
#> pop                                                  0
#> povknownratio                                        0
#> pre1960                                              0
#> proximity.npdes                                      0
#> proximity.npl                                        0
#> proximity.rmp                                        0
#> proximity.tsdf                                       0
#> radius.miles                                         0
#> rateasthma                                           0
#> ratecancer                                           0
#> rateheartdisease                                     0
#> ratio.to.avg.Demog.Index                             0
#> ratio.to.avg.Demog.Index.Supp                        0
#> ratio.to.avg.dpm                                     0
#> ratio.to.avg.drinking                                0
#> ratio.to.avg.lowlifex                                0
#> ratio.to.avg.no2                                     0
#> ratio.to.avg.o3                                      0
#> ratio.to.avg.pctaa                                   0
#> ratio.to.avg.pctaiana                                0
#> ratio.to.avg.pctba                                   0
#> ratio.to.avg.pctdisability                           0
#> ratio.to.avg.pcthisp                                 0
#> ratio.to.avg.pctlingiso                              0
#> ratio.to.avg.pctlowinc                               0
#> ratio.to.avg.pctlths                                 0
#> ratio.to.avg.pctmin                                  0
#> ratio.to.avg.pctmulti                                0
#> ratio.to.avg.pctnhaa                                 0
#> ratio.to.avg.pctnhaiana                              0
#> ratio.to.avg.pctnhba                                 0
#> ratio.to.avg.pctnhmulti                              0
#> ratio.to.avg.pctnhnhpia                              0
#> ratio.to.avg.pctnhotheralone                         0
#> ratio.to.avg.pctnhpia                                0
#> ratio.to.avg.pctnhwa                                 0
#> ratio.to.avg.pctotheralone                           0
#> ratio.to.avg.pctover64                               0
#> ratio.to.avg.pctpre1960                              0
#> ratio.to.avg.pctunder5                               0
#> ratio.to.avg.pctunemployed                           0
#> ratio.to.avg.pctwa                                   0
#> ratio.to.avg.pm                                      0
#> ratio.to.avg.proximity.npdes                         0
#> ratio.to.avg.proximity.npl                           0
#> ratio.to.avg.proximity.rmp                           0
#> ratio.to.avg.proximity.tsdf                          0
#> ratio.to.avg.rsei                                    0
#> ratio.to.avg.traffic.score                           0
#> ratio.to.avg.ust                                     0
#> ratio.to.state.avg.Demog.Index                       0
#> ratio.to.state.avg.Demog.Index.Supp                  0
#> ratio.to.state.avg.dpm                               0
#> ratio.to.state.avg.drinking                          0
#> ratio.to.state.avg.lowlifex                          0
#> ratio.to.state.avg.no2                               0
#> ratio.to.state.avg.o3                                0
#> ratio.to.state.avg.pctaa                             0
#> ratio.to.state.avg.pctaiana                          0
#> ratio.to.state.avg.pctba                             0
#> ratio.to.state.avg.pctdisability                     0
#> ratio.to.state.avg.pcthisp                           0
#> ratio.to.state.avg.pctlingiso                        0
#> ratio.to.state.avg.pctlowinc                         0
#> ratio.to.state.avg.pctlths                           0
#> ratio.to.state.avg.pctmin                            0
#> ratio.to.state.avg.pctmulti                          0
#> ratio.to.state.avg.pctnhaa                           0
#> ratio.to.state.avg.pctnhaiana                        0
#> ratio.to.state.avg.pctnhba                           0
#> ratio.to.state.avg.pctnhmulti                        0
#> ratio.to.state.avg.pctnhnhpia                        0
#> ratio.to.state.avg.pctnhotheralone                   0
#> ratio.to.state.avg.pctnhpia                          0
#> ratio.to.state.avg.pctnhwa                           0
#> ratio.to.state.avg.pctotheralone                     0
#> ratio.to.state.avg.pctover64                         0
#> ratio.to.state.avg.pctpre1960                        0
#> ratio.to.state.avg.pctunder5                         0
#> ratio.to.state.avg.pctunemployed                     0
#> ratio.to.state.avg.pctwa                             0
#> ratio.to.state.avg.pm                                0
#> ratio.to.state.avg.proximity.npdes                   0
#> ratio.to.state.avg.proximity.npl                     0
#> ratio.to.state.avg.proximity.rmp                     0
#> ratio.to.state.avg.proximity.tsdf                    0
#> ratio.to.state.avg.rsei                              0
#> ratio.to.state.avg.traffic.score                     0
#> ratio.to.state.avg.ust                               0
#> REGION                                               0
#> rsei                                                 0
#> Shape_Length                                         0
#> sitecount_avg                                        0
#> sitecount_max                                        0
#> sitecount_unique                                     0
#> spanish_li                                           0
#> ST                                                   0
#> state.avg.Demog.Index                                0
#> state.avg.Demog.Index.Supp                           0
#> state.avg.dpm                                        0
#> state.avg.drinking                                   0
#> state.avg.lifexyears_synonym                         0
#> state.avg.lowlifex                                   0
#> state.avg.no2                                        0
#> state.avg.o3                                         0
#> state.avg.pctaa                                      0
#> state.avg.pctaiana                                   0
#> state.avg.pctba                                      0
#> state.avg.pctdisability                              0
#> state.avg.pctfire                                    0
#> state.avg.pctfire30                                  0
#> state.avg.pctflood                                   0
#> state.avg.pctflood30                                 0
#> state.avg.pcthisp                                    0
#> state.avg.pctlingiso                                 0
#> state.avg.pctlowinc                                  0
#> state.avg.pctlths                                    0
#> state.avg.pctmin                                     0
#> state.avg.pctmulti                                   0
#> state.avg.pctnhaa                                    0
#> state.avg.pctnhaiana                                 0
#> state.avg.pctnhba                                    0
#> state.avg.pctnhmulti                                 0
#> state.avg.pctnhnhpia                                 0
#> state.avg.pctnhotheralone                            0
#> state.avg.pctnhpia                                   0
#> state.avg.pctnhwa                                    0
#> state.avg.pctnobroadband                             0
#> state.avg.pctnohealthinsurance                       0
#> state.avg.pctotheralone                              0
#> state.avg.pctover64                                  0
#> state.avg.pctpre1960                                 0
#> state.avg.pctunder5                                  0
#> state.avg.pctunemployed                              0
#> state.avg.pctwa                                      0
#> state.avg.pm                                         0
#> state.avg.proximity.npdes                            0
#> state.avg.proximity.npl                              0
#> state.avg.proximity.rmp                              0
#> state.avg.proximity.tsdf                             0
#> state.avg.rateasthma                                 0
#> state.avg.ratecancer                                 0
#> state.avg.rateheartdisease                           0
#> state.avg.rsei                                       0
#> state.avg.traffic.score                              0
#> state.avg.ust                                        0
#> state.count.ej.80up                                  0
#> state.count.ej.80up.supp                             0
#> state.EJ.DISPARITY.dpm.eo                            0
#> state.EJ.DISPARITY.dpm.supp                          0
#> state.EJ.DISPARITY.drinking.eo                       0
#> state.EJ.DISPARITY.drinking.supp                     0
#> state.EJ.DISPARITY.no2.eo                            0
#> state.EJ.DISPARITY.no2.supp                          0
#> state.EJ.DISPARITY.o3.eo                             0
#> state.EJ.DISPARITY.o3.supp                           0
#> state.EJ.DISPARITY.pctpre1960.eo                     0
#> state.EJ.DISPARITY.pctpre1960.supp                   0
#> state.EJ.DISPARITY.pm.eo                             0
#> state.EJ.DISPARITY.pm.supp                           0
#> state.EJ.DISPARITY.proximity.npdes.eo                0
#> state.EJ.DISPARITY.proximity.npdes.supp              0
#> state.EJ.DISPARITY.proximity.npl.eo                  0
#> state.EJ.DISPARITY.proximity.npl.supp                0
#> state.EJ.DISPARITY.proximity.rmp.eo                  0
#> state.EJ.DISPARITY.proximity.rmp.supp                0
#> state.EJ.DISPARITY.proximity.tsdf.eo                 0
#> state.EJ.DISPARITY.proximity.tsdf.supp               0
#> state.EJ.DISPARITY.rsei.eo                           0
#> state.EJ.DISPARITY.rsei.supp                         0
#> state.EJ.DISPARITY.traffic.score.eo                  0
#> state.EJ.DISPARITY.traffic.score.supp                0
#> state.EJ.DISPARITY.ust.eo                            0
#> state.EJ.DISPARITY.ust.supp                          0
#> state.pctile.Demog.Index                             1
#> state.pctile.Demog.Index.Supp                        1
#> state.pctile.dpm                                     1
#> state.pctile.drinking                                1
#> state.pctile.EJ.DISPARITY.dpm.eo                     1
#> state.pctile.EJ.DISPARITY.dpm.supp                   1
#> state.pctile.EJ.DISPARITY.drinking.eo                1
#> state.pctile.EJ.DISPARITY.drinking.supp              1
#> state.pctile.EJ.DISPARITY.no2.eo                     1
#> state.pctile.EJ.DISPARITY.no2.supp                   1
#> state.pctile.EJ.DISPARITY.o3.eo                      1
#> state.pctile.EJ.DISPARITY.o3.supp                    1
#> state.pctile.EJ.DISPARITY.pctpre1960.eo              1
#> state.pctile.EJ.DISPARITY.pctpre1960.supp            1
#> state.pctile.EJ.DISPARITY.pm.eo                      1
#> state.pctile.EJ.DISPARITY.pm.supp                    1
#> state.pctile.EJ.DISPARITY.proximity.npdes.eo         1
#> state.pctile.EJ.DISPARITY.proximity.npdes.supp       1
#> state.pctile.EJ.DISPARITY.proximity.npl.eo           1
#> state.pctile.EJ.DISPARITY.proximity.npl.supp         1
#> state.pctile.EJ.DISPARITY.proximity.rmp.eo           1
#> state.pctile.EJ.DISPARITY.proximity.rmp.supp         1
#> state.pctile.EJ.DISPARITY.proximity.tsdf.eo          1
#> state.pctile.EJ.DISPARITY.proximity.tsdf.supp        1
#> state.pctile.EJ.DISPARITY.rsei.eo                    1
#> state.pctile.EJ.DISPARITY.rsei.supp                  1
#> state.pctile.EJ.DISPARITY.traffic.score.eo           1
#> state.pctile.EJ.DISPARITY.traffic.score.supp         1
#> state.pctile.EJ.DISPARITY.ust.eo                     1
#> state.pctile.EJ.DISPARITY.ust.supp                   1
#> state.pctile.lifexyears_synonym                      1
#> state.pctile.lowlifex                                1
#> state.pctile.lowlifex_synonym                        1
#> state.pctile.no2                                     1
#> state.pctile.o3                                      1
#> state.pctile.pctaa                                   1
#> state.pctile.pctaiana                                1
#> state.pctile.pctba                                   1
#> state.pctile.pctdisability                           1
#> state.pctile.pctfire                                 1
#> state.pctile.pctfire30                               1
#> state.pctile.pctflood                                1
#> state.pctile.pctflood30                              1
#> state.pctile.pcthisp                                 1
#> state.pctile.pctlingiso                              1
#> state.pctile.pctlowinc                               1
#> state.pctile.pctlths                                 1
#> state.pctile.pctmin                                  1
#> state.pctile.pctmulti                                1
#> state.pctile.pctnhaa                                 1
#> state.pctile.pctnhaiana                              1
#> state.pctile.pctnhba                                 1
#> state.pctile.pctnhmulti                              1
#> state.pctile.pctnhnhpia                              1
#> state.pctile.pctnhotheralone                         1
#> state.pctile.pctnhpia                                1
#> state.pctile.pctnhwa                                 1
#> state.pctile.pctnobroadband                          1
#> state.pctile.pctnohealthinsurance                    1
#> state.pctile.pctotheralone                           1
#> state.pctile.pctover64                               1
#> state.pctile.pctpre1960                              1
#> state.pctile.pctunder5                               1
#> state.pctile.pctunemployed                           1
#> state.pctile.pctwa                                   1
#> state.pctile.pm                                      1
#> state.pctile.proximity.npdes                         1
#> state.pctile.proximity.npl                           1
#> state.pctile.proximity.rmp                           1
#> state.pctile.proximity.tsdf                          1
#> state.pctile.rateasthma                              1
#> state.pctile.ratecancer                              1
#> state.pctile.rateheartdisease                        1
#> state.pctile.rsei                                    1
#> state.pctile.traffic.score                           1
#> state.pctile.ust                                     1
#> statename                                            0
#> statlevel                                            0
#> timeSeconds                                          0
#> traffic.score                                        0
#> under18                                              0
#> under5                                               0
#> unemployed                                           0
#> unemployedbase                                       0
#> unit                                                 0
#> ust                                                  0
#> wa                                                   0
#> yesno_airnonatt                                      0
#> yesno_cejstdis                                       0
#> yesno_fooddesert                                     0
#> yesno_houseburden                                    0
#> yesno_impwaters                                      0
#> yesno_iradis                                         0
#> yesno_transdis                                       0
#> yesno_tribal                                         0
myinfo <- "long"
cbind(varinfo(myvars, myinfo) )
#>                                                                                       long
#> Demog.Index                                                          Demographic Index USA
#> Demog.Index.Supp                                        Supplemental Demographic Index USA
#> pctlowinc                                                                     % Low Income
#> pctlingiso                                        % in limited English-speaking Households
#> pctunemployed                                                                 % Unemployed
#> pctlths                                             % with Less Than High School Education
#> pctunder5                                                                    % under Age 5
#> pctover64                                                                    % over Age 64
#> pctmin                                                                   % People of Color
#> pcthisp                                                               % Hispanic or Latino
#> pctnhba                            % Black or African American (non-Hispanic, single race)
#> pctnhaa                                                % Asian (non-Hispanic, single race)
#> pctnhaiana                 % American Indian and Alaska Native (non-Hispanic, single race)
#> pctnhnhpia        % Native Hawaiian and Other Pacific Islander (non-Hispanic, single race)
#> pctnhotheralone                                   % Other race (non-Hispanic, single race)
#> pctnhmulti                                              % Two or more races (non-Hispanic)
#> pctnhwa                                                % White (non-Hispanic, single race)
#> pm                                                    Particulate Matter (PM 2.5 in ug/m3)
#> o3                                                                             Ozone (ppb)
#> no2                                                                 Nitrogen Dioxide (NO2)
#> dpm                                                      Diesel Particulate Matter (ug/m3)
#> rsei                                                                 Toxic Releases to Air
#> traffic.score          Traffic Proximity and Volume (daily traffic count/distance to road)
#> pctpre1960                                      Lead Paint Indicator (% pre-1960s housing)
#> proximity.npl                                 Superfund Proximity (site count/km distance)
#> proximity.rmp                                   RMP Proximity (facility count/km distance)
#> proximity.tsdf                      Hazardous Waste Proximity (facility count/km distance)
#> ust                                              Underground Storage Tanks (UST) indicator
#> proximity.npdes  Wastewater Discharge Indicator (toxicity-weighted concentration/distance)
#> drinking                                                     Drinking Water Non-Compliance
table_rounding_info(names_e)
#> Error in table_rounding_info(names_e): could not find function "table_rounding_info"

varinfo(
 var = c(names_these, names_d_pctile),
 info = c(
 "topic_root_term", "varcategory", "vartype", "percentage", "pctile.", "calculation_type"
))
#>                          topic_root_term   varcategory  vartype percentage
#> Demog.Index                  Demog.Index   Demographic      raw          1
#> Demog.Index.Supp        Demog.Index.Supp   Demographic      raw          1
#> pctlowinc                      pctlowinc   Demographic      raw          1
#> pctlingiso                    pctlingiso   Demographic      raw          1
#> pctunemployed              pctunemployed   Demographic      raw          1
#> pctlths                          pctlths   Demographic      raw          1
#> pctunder5                      pctunder5   Demographic      raw          1
#> pctover64                      pctover64   Demographic      raw          1
#> pctmin                            pctmin   Demographic      raw          1
#> pcthisp                          pcthisp   Demographic      raw          1
#> pctnhba                          pctnhba   Demographic      raw          1
#> pctnhaa                          pctnhaa   Demographic      raw          1
#> pctnhaiana                    pctnhaiana   Demographic      raw          1
#> pctnhnhpia                    pctnhnhpia   Demographic      raw          1
#> pctnhotheralone          pctnhotheralone   Demographic      raw          1
#> pctnhmulti                    pctnhmulti   Demographic      raw          1
#> pctnhwa                          pctnhwa   Demographic      raw          1
#> pm                                    pm Environmental      raw           
#> o3                                    o3 Environmental      raw           
#> no2                                  no2 Environmental      raw           
#> dpm                                  dpm Environmental      raw           
#> rsei                                rsei Environmental      raw           
#> traffic.score              traffic.score Environmental      raw           
#> pctpre1960                    pctpre1960 Environmental      raw          1
#> proximity.npl              proximity.npl Environmental      raw           
#> proximity.rmp              proximity.rmp Environmental      raw           
#> proximity.tsdf            proximity.tsdf Environmental      raw           
#> ust                                  ust Environmental      raw           
#> proximity.npdes          proximity.npdes Environmental      raw           
#> drinking                        drinking Environmental      raw           
#> pctile.Demog.Index           Demog.Index   Demographic uspctile           
#> pctile.Demog.Index.Supp Demog.Index.Supp   Demographic uspctile           
#> pctile.pctlowinc               pctlowinc   Demographic uspctile           
#> pctile.pctlingiso             pctlingiso   Demographic uspctile           
#> pctile.pctunemployed       pctunemployed   Demographic uspctile           
#> pctile.pctlths                   pctlths   Demographic uspctile           
#> pctile.pctunder5               pctunder5   Demographic uspctile           
#> pctile.pctover64               pctover64   Demographic uspctile           
#> pctile.pctmin                     pctmin   Demographic uspctile           
#>                         pctile. calculation_type
#> Demog.Index                   0          wtdmean
#> Demog.Index.Supp              0          wtdmean
#> pctlowinc                     0          wtdmean
#> pctlingiso                    0          wtdmean
#> pctunemployed                 0          wtdmean
#> pctlths                       0          wtdmean
#> pctunder5                     0          wtdmean
#> pctover64                     0          wtdmean
#> pctmin                        0          wtdmean
#> pcthisp                       0          wtdmean
#> pctnhba                       0          wtdmean
#> pctnhaa                       0          wtdmean
#> pctnhaiana                    0          wtdmean
#> pctnhnhpia                    0          wtdmean
#> pctnhotheralone               0          wtdmean
#> pctnhmulti                    0          wtdmean
#> pctnhwa                       0          wtdmean
#> pm                            0          wtdmean
#> o3                            0          wtdmean
#> no2                           0          wtdmean
#> dpm                           0          wtdmean
#> rsei                          0          wtdmean
#> traffic.score                 0          wtdmean
#> pctpre1960                    0          wtdmean
#> proximity.npl                 0          wtdmean
#> proximity.rmp                 0          wtdmean
#> proximity.tsdf                0          wtdmean
#> ust                           0          wtdmean
#> proximity.npdes               0          wtdmean
#> drinking                      0          wtdmean
#> pctile.Demog.Index            1         lookedup
#> pctile.Demog.Index.Supp       1         lookedup
#> pctile.pctlowinc              1         lookedup
#> pctile.pctlingiso             1         lookedup
#> pctile.pctunemployed          1         lookedup
#> pctile.pctlths                1         lookedup
#> pctile.pctunder5              1         lookedup
#> pctile.pctover64              1         lookedup
#> pctile.pctmin                 1         lookedup

varinfo(names_all_r, c("varcategory", "varlist", "in_api", "in_bgcsv"))
#> Error in `[.data.frame`(map_headernames, match(var, map_headernames[,     varnametype]), info_true_name, drop = FALSE): undefined columns selected

cbind(
  namez$d, 
  names_d, 
  varinfo(names_d, "varlist"), 
  usavg = unlist(avg.in.us[names_d]), 
  usavg_varname = EJAM:::fixcolnames2related(names_d, "usavg"), 
  categ = varname2varcategory_ejam(names_d)
)
#>                           namez$d          names_d varlist      usavg
#> Demog.Index           Demog.Index      Demog.Index names_d 1.34074408
#> Demog.Index.Supp Demog.Index.Supp Demog.Index.Supp names_d 1.64233560
#> pctlowinc               pctlowinc        pctlowinc names_d 0.30414988
#> pctlingiso             pctlingiso       pctlingiso names_d 0.04810017
#> pctunemployed       pctunemployed    pctunemployed names_d 0.05614969
#> pctlths                   pctlths          pctlths names_d 0.11306978
#> pctunder5               pctunder5        pctunder5 names_d 0.05390936
#> pctover64               pctover64        pctover64 names_d 0.17790662
#> pctmin                     pctmin           pctmin names_d 0.39584286
#>                         usavg_varname       categ
#> Demog.Index           avg.Demog.Index Demographic
#> Demog.Index.Supp avg.Demog.Index.Supp Demographic
#> pctlowinc               avg.pctlowinc Demographic
#> pctlingiso             avg.pctlingiso Demographic
#> pctunemployed       avg.pctunemployed Demographic
#> pctlths                   avg.pctlths Demographic
#> pctunder5               avg.pctunder5 Demographic
#> pctover64               avg.pctover64 Demographic
#> pctmin                     avg.pctmin Demographic

names(map_headernames)
#>  [1] "n"                                 "newsort"                          
#>  [3] "ejscreensort"                      "in_api"                           
#>  [5] "in_acs"                            "in_csv"                           
#>  [7] "oldname_is_what"                   "oldname"                          
#>  [9] "apiname"                           "ejscreen_api"                     
#> [11] "api_synonym"                       "acsname"                          
#> [13] "acs2017_2021v2.2"                  "csvname2.2"                       
#> [15] "csvname"                           "ejscreen_csv"                     
#> [17] "rname"                             "sort_within_varlistEJSCREENREPORT"
#> [19] "sort_within_varlist"               "topic_root_term"                  
#> [21] "basevarname"                       "sortvarlistEJSCREENREPORT"        
#> [23] "sortvarlist"                       "varlist"                          
#> [25] "ratio.to"                          "state."                           
#> [27] "pctile."                           ".text"                            
#> [29] "avg."                              "bin."                             
#> [31] "DISPARITY"                         ".eo"                              
#> [33] ".supp"                             "apitype"                          
#> [35] "apisection"                        "varcategory"                      
#> [37] "DEJ"                               "jsondoc_sort_DEJ"                 
#> [39] "zone"                              "jsondoc_zone"                     
#> [41] "jsondoc_shortzone"                 "jsondoc_sort_zone"                
#> [43] "vartype"                           "raw_pctile_avg_basedonrname"      
#> [45] "raw_pctile_avg"                    "agree"                            
#> [47] "jsondoc_vartype"                   "jsondoc_shortvartype"             
#> [49] "percentage"                        "dollar"                           
#> [51] "denominator"                       "is.wtdmean"                       
#> [53] "calculation_type"                  "sigfigs"                          
#> [55] "decimals"                          "pct_as_fraction_ejscreenit"       
#> [57] "pct_as_fraction_ejamit"            "pct_as_fraction_blockgroupstats"  
#> [59] "units"                             "shortmatch"                       
#> [61] "shortlabel"                        "longmatch"                        
#> [63] "longname"                          "names_friendly"                   
#> [65] "csv_description"                   "acs_description"                  
#> [67] "api_description"                   "reportlabel"                      
#> [69] "ejscreenreport"                    "reportsort"                       
#> [71] "errornote"                         "api_example"                      
#> [73] "csv_example"                       "csv_descriptions_name"            
#> [75] "EJAMejscreendata"                  "gdbfieldname in map_batch"        
t(varinfo(names_d[1]))
#>                                   Demog.Index               
#> n                                 "1"                       
#> newsort                           "010116162"               
#> ejscreensort                      "0101162"                 
#> in_api                            "1"                       
#> in_acs                            "0"                       
#> in_csv                            "1"                       
#> oldname_is_what                   "api"                     
#> oldname                           "RAW_D_DEMOGIDX2"         
#> apiname                           "RAW_D_DEMOGIDX2"         
#> ejscreen_api                      ""                        
#> api_synonym                       ""                        
#> acsname                           ""                        
#> acs2017_2021v2.2                  ""                        
#> csvname2.2                        "DEMOGIDX_2"              
#> csvname                           "DEMOGIDX_2"              
#> ejscreen_csv                      ""                        
#> rname                             "Demog.Index"             
#> sort_within_varlistEJSCREENREPORT "162"                     
#> sort_within_varlist               "16"                      
#> topic_root_term                   "Demog.Index"             
#> basevarname                       "Demog.Index"             
#> sortvarlistEJSCREENREPORT         "1"                       
#> sortvarlist                       "1"                       
#> varlist                           "names_d"                 
#> ratio.to                          "0"                       
#> state.                            "0"                       
#> pctile.                           "0"                       
#> .text                             "0"                       
#> avg.                              "0"                       
#> bin.                              "0"                       
#> DISPARITY                         "0"                       
#> .eo                               "0"                       
#> .supp                             "0"                       
#> apitype                           "main"                    
#> apisection                        "Socioeconomic Indicators"
#> varcategory                       "Demographic"             
#> DEJ                               "Demographic"             
#> jsondoc_sort_DEJ                  "1"                       
#> zone                              "Nation"                  
#> jsondoc_zone                      "Nation"                  
#> jsondoc_shortzone                 "us"                      
#> jsondoc_sort_zone                 "0"                       
#> vartype                           "raw"                     
#> raw_pctile_avg_basedonrname       "raw"                     
#> raw_pctile_avg                    "raw"                     
#> agree                             "TRUE"                    
#> jsondoc_vartype                   "raw data for indicator"  
#> jsondoc_shortvartype              "raw"                     
#> percentage                        "1"                       
#> dollar                            ""                        
#> denominator                       "pop"                     
#> is.wtdmean                        "TRUE"                    
#> calculation_type                  "wtdmean"                 
#> sigfigs                           "3"                       
#> decimals                          "2"                       
#> pct_as_fraction_ejscreenit        "FALSE"                   
#> pct_as_fraction_ejamit            "FALSE"                   
#> pct_as_fraction_blockgroupstats   "FALSE"                   
#> units                             ""                        
#> shortmatch                        "TRUE"                    
#> shortlabel                        "Demog.Ind."              
#> longmatch                         "TRUE"                    
#> longname                          "Demographic Index USA"   
#> names_friendly                    "Demographic Index USA"   
#> csv_description                   "Demographic Index"       
#> acs_description                   ""                        
#> api_description                   "Demographic Index"       
#> reportlabel                       "Demographic Index USA"   
#> ejscreenreport                    ""                        
#> reportsort                        "162"                     
#> errornote                         ""                        
#> api_example                       "13%"                     
#> csv_example                       "0.2662338"               
#> csv_descriptions_name             "DEMOGIDX_2"              
#> EJAMejscreendata                  ""                        
#> gdbfieldname in map_batch         ""